% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_aspect.R
\name{gdal_aspect}
\alias{gdal_aspect}
\title{QGIS algorithm Aspect}
\usage{
gdal_aspect(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  TRIG_ANGLE = qgisprocess::qgis_default_value(),
  ZERO_FLAT = qgisprocess::qgis_default_value(),
  COMPUTE_EDGES = qgisprocess::qgis_default_value(),
  ZEVENBERGEN = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{TRIG_ANGLE}{\code{boolean} - Return trigonometric angle instead of azimuth. 1 for true/yes. 0 for false/no.}

\item{ZERO_FLAT}{\code{boolean} - Return 0 for flat instead of -9999. 1 for true/yes. 0 for false/no.}

\item{COMPUTE_EDGES}{\code{boolean} - Compute edges. 1 for true/yes. 0 for false/no.}

\item{ZEVENBERGEN}{\code{boolean} - Use Zevenbergen&Thorne formula instead of the Horn's one. 1 for true/yes. 0 for false/no.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Aspect. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Aspect (gdal:aspect)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Aspect
}
}
}
