% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_buildvirtualvector.R
\name{gdal_buildvirtualvector}
\alias{gdal_buildvirtualvector}
\title{QGIS algorithm Build virtual vector}
\usage{
gdal_buildvirtualvector(
  INPUT = qgisprocess::qgis_default_value(),
  UNIONED = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input datasources. .}

\item{UNIONED}{\code{boolean} - Create "unioned" VRT. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{vectorDestination} - Virtual vector. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Build virtual vector (gdal:buildvirtualvector)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Virtual vector
\item VRT_STRING - outputString - Virtual string
}
}
}
