% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_cliprasterbyextent.R
\name{gdal_cliprasterbyextent}
\alias{gdal_cliprasterbyextent}
\title{QGIS algorithm Clip raster by extent}
\usage{
gdal_cliprasterbyextent(
  INPUT = qgisprocess::qgis_default_value(),
  PROJWIN = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{PROJWIN}{\code{extent} - Clipping extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{NODATA}{\code{number} - Assign a specified nodata value to output bands. A numeric value.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Use Input Layer Data Type", "Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Clipped (extent). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Clip raster by extent (gdal:cliprasterbyextent)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Clipped
}
}
}
