% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_colorrelief.R
\name{gdal_colorrelief}
\alias{gdal_colorrelief}
\title{QGIS algorithm Color relief}
\usage{
gdal_colorrelief(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  COMPUTE_EDGES = qgisprocess::qgis_default_value(),
  COLOR_TABLE = qgisprocess::qgis_default_value(),
  MATCH_MODE = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{COMPUTE_EDGES}{\code{boolean} - Compute edges. 1 for true/yes. 0 for false/no.}

\item{COLOR_TABLE}{\code{file} - Color configuration file. Path to a file.}

\item{MATCH_MODE}{\code{enum}  of \verb{("Use strict color matching", "Use closest RGBA quadruplet", "Use smoothly blended colors")} - Matching mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Color relief. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Color relief (gdal:colorrelief)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Color relief
}
}
}
