% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_contour_polygon.R
\name{gdal_contour_polygon}
\alias{gdal_contour_polygon}
\title{QGIS algorithm Contour Polygons}
\usage{
gdal_contour_polygon(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  INTERVAL = qgisprocess::qgis_default_value(),
  CREATE_3D = qgisprocess::qgis_default_value(),
  IGNORE_NODATA = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  OFFSET = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  FIELD_NAME_MIN = qgisprocess::qgis_default_value(),
  FIELD_NAME_MAX = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{INTERVAL}{\code{number} - Interval between contour lines. A numeric value.}

\item{CREATE_3D}{\code{boolean} - Produce 3D vector. 1 for true/yes. 0 for false/no.}

\item{IGNORE_NODATA}{\code{boolean} - Treat all raster values as valid. 1 for true/yes. 0 for false/no.}

\item{NODATA}{\code{number} - Input pixel value to treat as "nodata". A numeric value.}

\item{OFFSET}{\code{number} - Offset from zero relative to which to interpret intervals. A numeric value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{FIELD_NAME_MIN}{\code{string} - Attribute name for minimum elevation of contour polygon. String value.}

\item{FIELD_NAME_MAX}{\code{string} - Attribute name for maximum elevation of contour polygon. String value.}

\item{OUTPUT}{\code{vectorDestination} - Contours. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Contour Polygons (gdal:contour_polygon)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Contours
}
}
}
