% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_convertformat.R
\name{gdal_convertformat}
\alias{gdal_convertformat}
\title{QGIS algorithm Convert format}
\usage{
gdal_convertformat(
  INPUT = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{OUTPUT}{\code{vectorDestination} - Converted. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Convert format (gdal:convertformat)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Converted
}
}
}
