% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_executesql.R
\name{gdal_executesql}
\alias{gdal_executesql}
\title{QGIS algorithm Execute SQL}
\usage{
gdal_executesql(
  INPUT = qgisprocess::qgis_default_value(),
  SQL = qgisprocess::qgis_default_value(),
  DIALECT = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{SQL}{\code{string} - SQL expression. String value.}

\item{DIALECT}{\code{enum}  of \verb{("None", "OGR SQL", "SQLite")} - SQL dialect. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{OUTPUT}{\code{vectorDestination} - SQL result. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Execute SQL (gdal:executesql)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - SQL result
}
}
}
