% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_gdal2tiles.R
\name{gdal_gdal2tiles}
\alias{gdal_gdal2tiles}
\title{QGIS algorithm gdal2tiles}
\usage{
gdal_gdal2tiles(
  INPUT = qgisprocess::qgis_default_value(),
  PROFILE = qgisprocess::qgis_default_value(),
  ZOOM = qgisprocess::qgis_default_value(),
  VIEWER = qgisprocess::qgis_default_value(),
  TITLE = qgisprocess::qgis_default_value(),
  COPYRIGHT = qgisprocess::qgis_default_value(),
  RESAMPLING = qgisprocess::qgis_default_value(),
  SOURCE_CRS = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  URL = qgisprocess::qgis_default_value(),
  GOOGLE_KEY = qgisprocess::qgis_default_value(),
  BING_KEY = qgisprocess::qgis_default_value(),
  RESUME = qgisprocess::qgis_default_value(),
  KML = qgisprocess::qgis_default_value(),
  NO_KML = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{PROFILE}{\code{enum}  of \verb{("Mercator", "Geodetic", "Raster")} - Tile cutting profile. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ZOOM}{\code{string} - Zoom levels to render. String value.}

\item{VIEWER}{\code{enum}  of \verb{("All", "GoogleMaps", "OpenLayers", "Leaflet", "None")} - Web viewer to generate. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TITLE}{\code{string} - Title of the map. String value.}

\item{COPYRIGHT}{\code{string} - Copyright of the map. String value.}

\item{RESAMPLING}{\code{enum}  of \verb{("Average", "Nearest neighbour", "Bilinear", "Cubic", "Cubic spline", "Lanczos windowed sinc", "Antialias")} - Resampling method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SOURCE_CRS}{\code{crs} - The spatial reference system used for the source input data. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{NODATA}{\code{number} - Transparency value to assign to the input data. A numeric value.}

\item{URL}{\code{string} - URL address where the generated tiles are going to be published. String value.}

\item{GOOGLE_KEY}{\code{string} - Google Maps API key (http://code.google.com/apis/maps/signup.html). String value.}

\item{BING_KEY}{\code{string} - Bing Maps API key (https://www.bingmapsportal.com/). String value.}

\item{RESUME}{\code{boolean} - Generate only missing files. 1 for true/yes. 0 for false/no.}

\item{KML}{\code{boolean} - Generate KML for Google Earth. 1 for true/yes. 0 for false/no.}

\item{NO_KML}{\code{boolean} - Avoid automatic generation of KML files for EPSG:4326. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{folderDestination} - Output directory. Path for an existing or new folder.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL gdal2tiles (gdal:gdal2tiles)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFolder - Output directory
}
}
}
