% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_gdalinfo.R
\name{gdal_gdalinfo}
\alias{gdal_gdalinfo}
\title{QGIS algorithm Raster information}
\usage{
gdal_gdalinfo(
  INPUT = qgisprocess::qgis_default_value(),
  MIN_MAX = qgisprocess::qgis_default_value(),
  STATS = qgisprocess::qgis_default_value(),
  NOGCP = qgisprocess::qgis_default_value(),
  NO_METADATA = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{MIN_MAX}{\code{boolean} - Force computation of the actual min/max values for each band. 1 for true/yes. 0 for false/no.}

\item{STATS}{\code{boolean} - Read and display image statistics (force computation if necessary). 1 for true/yes. 0 for false/no.}

\item{NOGCP}{\code{boolean} - Suppress GCP info. 1 for true/yes. 0 for false/no.}

\item{NO_METADATA}{\code{boolean} - Suppress metadata info. 1 for true/yes. 0 for false/no.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{fileDestination} - Layer information. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Raster information (gdal:gdalinfo)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputHtml - Layer information
}
}
}
