% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_gridlinear.R
\name{gdal_gridlinear}
\alias{gdal_gridlinear}
\title{QGIS algorithm Grid (Linear)}
\usage{
gdal_gridlinear(
  INPUT = qgisprocess::qgis_default_value(),
  Z_FIELD = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Point layer. Path to a vector layer.}

\item{Z_FIELD}{\code{field} - Z value from field. The name of an existing field. ; delimited list of existing field names.}

\item{RADIUS}{\code{number} - Search distance . A numeric value.}

\item{NODATA}{\code{number} - NODATA marker to fill empty points. A numeric value.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{rasterDestination} - Interpolated (Linear). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Grid (Linear) (gdal:gridlinear)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Interpolated (Linear)
}
}
}
