% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_hillshade.R
\name{gdal_hillshade}
\alias{gdal_hillshade}
\title{QGIS algorithm Hillshade}
\usage{
gdal_hillshade(
  INPUT = qgisprocess:::qgis_default_value(),
  BAND = qgisprocess:::qgis_default_value(),
  Z_FACTOR = qgisprocess:::qgis_default_value(),
  SCALE = qgisprocess:::qgis_default_value(),
  AZIMUTH = qgisprocess:::qgis_default_value(),
  ALTITUDE = qgisprocess:::qgis_default_value(),
  COMPUTE_EDGES = qgisprocess:::qgis_default_value(),
  ZEVENBERGEN = qgisprocess:::qgis_default_value(),
  COMBINED = qgisprocess:::qgis_default_value(),
  MULTIDIRECTIONAL = qgisprocess:::qgis_default_value(),
  OPTIONS = qgisprocess:::qgis_default_value(),
  EXTRA = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{Z_FACTOR}{\code{number} - Z factor (vertical exaggeration). A numeric value.}

\item{SCALE}{\code{number} - Scale (ratio of vertical units to horizontal). A numeric value.}

\item{AZIMUTH}{\code{number} - Azimuth of the light. A numeric value.}

\item{ALTITUDE}{\code{number} - Altitude of the light. A numeric value.}

\item{COMPUTE_EDGES}{\code{boolean} - Compute edges. 1 for true/yes. 0 for false/no.}

\item{ZEVENBERGEN}{\code{boolean} - Use Zevenbergen&Thorne formula instead of the Horn's one. 1 for true/yes. 0 for false/no.}

\item{COMBINED}{\code{boolean} - Combined shading. 1 for true/yes. 0 for false/no.}

\item{MULTIDIRECTIONAL}{\code{boolean} - Multidirectional shading. 1 for true/yes. 0 for false/no.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Hillshade. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Hillshade (gdal:hillshade)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Hillshade
}
}
}
