% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gdal_importvectorintopostgisdatabaseavailableconnections.R
\name{gdal_importvectorintopostgisdatabaseavailableconnections}
\alias{gdal_importvectorintopostgisdatabaseavailableconnections}
\title{QGIS algorithm - Export to PostgreSQL (available connections)}
\usage{
gdal_importvectorintopostgisdatabaseavailableconnections(
  DATABASE = qgisprocess:::qgis_default_value(),
  INPUT = qgisprocess:::qgis_default_value(),
  SHAPE_ENCODING = qgisprocess:::qgis_default_value(),
  GTYPE = qgisprocess:::qgis_default_value(),
  A_SRS = qgisprocess:::qgis_default_value(),
  T_SRS = qgisprocess:::qgis_default_value(),
  S_SRS = qgisprocess:::qgis_default_value(),
  SCHEMA = qgisprocess:::qgis_default_value(),
  TABLE = qgisprocess:::qgis_default_value(),
  PK = qgisprocess:::qgis_default_value(),
  PRIMARY_KEY = qgisprocess:::qgis_default_value(),
  GEOCOLUMN = qgisprocess:::qgis_default_value(),
  DIM = qgisprocess:::qgis_default_value(),
  SIMPLIFY = qgisprocess:::qgis_default_value(),
  SEGMENTIZE = qgisprocess:::qgis_default_value(),
  SPAT = qgisprocess:::qgis_default_value(),
  CLIP = qgisprocess:::qgis_default_value(),
  WHERE = qgisprocess:::qgis_default_value(),
  GT = qgisprocess:::qgis_default_value(),
  OVERWRITE = qgisprocess:::qgis_default_value(),
  APPEND = qgisprocess:::qgis_default_value(),
  ADDFIELDS = qgisprocess:::qgis_default_value(),
  LAUNDER = qgisprocess:::qgis_default_value(),
  INDEX = qgisprocess:::qgis_default_value(),
  SKIPFAILURES = qgisprocess:::qgis_default_value(),
  PROMOTETOMULTI = qgisprocess:::qgis_default_value(),
  PRECISION = qgisprocess:::qgis_default_value(),
  OPTIONS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DATABASE}{\code{providerconnection} - Database (connection name). Name of registered database connection.}

\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{SHAPE_ENCODING}{\code{string} - Shape encoding. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GTYPE}{\code{enum}  of \verb{("", "NONE", "GEOMETRY", "POINT", "LINESTRING", "POLYGON", "GEOMETRYCOLLECTION", "MULTIPOINT", "MULTIPOLYGON", "MULTILINESTRING", "CIRCULARSTRING", "COMPOUNDCURVE", "CURVEPOLYGON", "MULTICURVE", "MULTISURFACE")} - Output geometry type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{A_SRS}{\code{crs} - Assign an output CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{T_SRS}{\code{crs} - Reproject to this CRS on output . CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{S_SRS}{\code{crs} - Override source CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{SCHEMA}{\code{databaseschema} - Schema (schema name). Name of existing database schema.}

\item{TABLE}{\code{databasetable} - Table to import to (leave blank to use layer name). Name of existing database table.}

\item{PK}{\code{string} - Primary key (new field). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PRIMARY_KEY}{\code{field} - Primary key (existing field, used if the above option is left empty). The name of an existing field. ; delimited list of existing field names.}

\item{GEOCOLUMN}{\code{string} - Geometry column name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DIM}{\code{enum}  of \verb{("2", "3", "4")} - Vector dimensions. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SIMPLIFY}{\code{string} - Distance tolerance for simplification. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SEGMENTIZE}{\code{string} - Maximum distance between 2 nodes (densification). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SPAT}{\code{extent} - Select features by extent (defined in input layer CRS). A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{CLIP}{\code{boolean} - Clip the input layer using the above (rectangle) extent. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WHERE}{\code{string} - Select features using a SQL "WHERE" statement (Ex: column='value'). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GT}{\code{string} - Group N features per transaction (Default: 20000). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OVERWRITE}{\code{boolean} - Overwrite existing table. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{APPEND}{\code{boolean} - Append to existing table. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ADDFIELDS}{\code{boolean} - Append and add new fields to existing table. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LAUNDER}{\code{boolean} - Do not launder columns\code{/}table names. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INDEX}{\code{boolean} - Do not create spatial index. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SKIPFAILURES}{\code{boolean} - Continue after a failure, skipping the failed feature. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PROMOTETOMULTI}{\code{boolean} - Promote to Multipart. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PRECISION}{\code{boolean} - Keep width and precision of input attributes. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OPTIONS}{\code{string} - Additional creation options. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Export to PostgreSQL (available connections) (gdal:importvectorintopostgisdatabaseavailableconnections). Exports a vector layer to an existing PostgreSQL database connection
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
