% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_nearblack.R
\name{gdal_nearblack}
\alias{gdal_nearblack}
\title{QGIS algorithm Near black}
\usage{
gdal_nearblack(
  INPUT = qgisprocess::qgis_default_value(),
  NEAR = qgisprocess::qgis_default_value(),
  WHITE = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{NEAR}{\code{number} - How far from black (white). A numeric value.}

\item{WHITE}{\code{boolean} - Search for nearly white pixels instead of nearly black. 1 for true/yes. 0 for false/no.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Nearblack. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Near black (gdal:nearblack)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Nearblack
}
}
}
