% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_overviews.R
\name{gdal_overviews}
\alias{gdal_overviews}
\title{QGIS algorithm - Build overviews (pyramids)}
\usage{
gdal_overviews(
  INPUT = qgisprocess:::qgis_default_value(),
  CLEAN = qgisprocess:::qgis_default_value(),
  LEVELS = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  FORMAT = qgisprocess:::qgis_default_value(),
  EXTRA = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{CLEAN}{\code{boolean} - Remove all existing overviews. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LEVELS}{\code{string} - Overview levels (e.g. 2 4 8 16). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RESAMPLING}{\code{enum}  of \verb{("Nearest Neighbour (default)", "Average", "Gaussian", "Cubic Convolution", "B-Spline Convolution", "Lanczos Windowed Sinc", "Average MP", "Average in Mag/Phase Space", "Mode")} - Resampling method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FORMAT}{\code{enum}  of \verb{("Internal (if possible)", "External (GTiff .ovr)", "External (ERDAS Imagine .aux)")} - Overviews format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Build overviews (pyramids) (gdal:overviews). ---------------- Arguments ----------------  INPUT: Input layer 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CLEAN: Remove all existing overviews 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LEVELS: Overview levels (e.g. 2 4 8 16) (optional) 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RESAMPLING: Resampling method (optional) 	Argument type:	enum 	Available values: 		- 0: Nearest Neighbour (default) 		- 1: Average 		- 2: Gaussian 		- 3: Cubic Convolution 		- 4: B-Spline Convolution 		- 5: Lanczos Windowed Sinc 		- 6: Average MP 		- 7: Average in Mag/Phase Space 		- 8: Mode 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FORMAT: Overviews format (optional) 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: Internal (if possible) 		- 1: External (GTiff .ovr) 		- 2: External (ERDAS Imagine .aux) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' EXTRA: Additional command-line parameters (optional) 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Pyramidized
}
}
}
