% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_pointsalonglines.R
\name{gdal_pointsalonglines}
\alias{gdal_pointsalonglines}
\title{QGIS algorithm Points along lines}
\usage{
gdal_pointsalonglines(
  INPUT = qgisprocess::qgis_default_value(),
  GEOMETRY = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{GEOMETRY}{\code{string} - Geometry column name. String value.}

\item{DISTANCE}{\code{number} - Distance from line start represented as fraction of line length. A numeric value.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{OUTPUT}{\code{vectorDestination} - Points along lines. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Points along lines (gdal:pointsalonglines)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Points along lines
}
}
}
