% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_rasterize_over.R
\name{gdal_rasterize_over}
\alias{gdal_rasterize_over}
\title{QGIS algorithm Rasterize (overwrite with attribute)}
\usage{
gdal_rasterize_over(
  INPUT = qgisprocess::qgis_default_value(),
  INPUT_RASTER = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  ADD = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input vector layer. Path to a vector layer.}

\item{INPUT_RASTER}{\code{raster} - Input raster layer. Path to a raster layer.}

\item{FIELD}{\code{field} - Field to use for burn in value. The name of an existing field. ; delimited list of existing field names.}

\item{ADD}{\code{boolean} - Add burn in values to existing raster values. 1 for true/yes. 0 for false/no.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Rasterize (overwrite with attribute) (gdal:rasterize_over)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Rasterized
}
}
}
