% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_rasterize_over_fixed_value.R
\name{gdal_rasterize_over_fixed_value}
\alias{gdal_rasterize_over_fixed_value}
\title{QGIS algorithm Rasterize (overwrite with fixed value)}
\usage{
gdal_rasterize_over_fixed_value(
  INPUT = qgisprocess::qgis_default_value(),
  INPUT_RASTER = qgisprocess::qgis_default_value(),
  BURN = qgisprocess::qgis_default_value(),
  ADD = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input vector layer. Path to a vector layer.}

\item{INPUT_RASTER}{\code{raster} - Input raster layer. Path to a raster layer.}

\item{BURN}{\code{number} - A fixed value to burn. A numeric value.}

\item{ADD}{\code{boolean} - Add burn in values to existing raster values. 1 for true/yes. 0 for false/no.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Rasterize (overwrite with fixed value) (gdal:rasterize_over_fixed_value)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Rasterized
}
}
}
