% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_retile.R
\name{gdal_retile}
\alias{gdal_retile}
\title{QGIS algorithm Retile}
\usage{
gdal_retile(
  INPUT = qgisprocess::qgis_default_value(),
  TILE_SIZE_X = qgisprocess::qgis_default_value(),
  TILE_SIZE_Y = qgisprocess::qgis_default_value(),
  OVERLAP = qgisprocess::qgis_default_value(),
  LEVELS = qgisprocess::qgis_default_value(),
  SOURCE_CRS = qgisprocess::qgis_default_value(),
  RESAMPLING = qgisprocess::qgis_default_value(),
  DELIMITER = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  ONLY_PYRAMIDS = qgisprocess::qgis_default_value(),
  DIR_FOR_ROW = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  OUTPUT_CSV = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input files. .}

\item{TILE_SIZE_X}{\code{number} - Tile width. A numeric value.}

\item{TILE_SIZE_Y}{\code{number} - Tile height. A numeric value.}

\item{OVERLAP}{\code{number} - Overlap in pixels between consecutive tiles. A numeric value.}

\item{LEVELS}{\code{number} - Number of pyramids levels to build. A numeric value.}

\item{SOURCE_CRS}{\code{crs} - Source coordinate reference system. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{RESAMPLING}{\code{enum}  of \verb{("Nearest Neighbour", "Bilinear", "Cubic", "Cubic Spline", "Lanczos Windowed Sinc")} - Resampling method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DELIMITER}{\code{string} - Column delimiter used in the CSV file. String value.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ONLY_PYRAMIDS}{\code{boolean} - Build only the pyramids. 1 for true/yes. 0 for false/no.}

\item{DIR_FOR_ROW}{\code{boolean} - Use separate directory for each tiles row. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{folderDestination} - Output directory. Path for an existing or new folder.}

\item{OUTPUT_CSV}{\code{fileDestination} - CSV file containing the tile(s) georeferencing information. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Retile (gdal:retile)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFolder - Output directory
\item OUTPUT_CSV - outputFile - CSV file containing the tile(s) georeferencing information
}
}
}
