% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_rgbtopct.R
\name{gdal_rgbtopct}
\alias{gdal_rgbtopct}
\title{QGIS algorithm RGB to PCT}
\usage{
gdal_rgbtopct(
  INPUT = qgisprocess::qgis_default_value(),
  NCOLORS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{NCOLORS}{\code{number} - Number of colors. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - RGB to PCT. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL RGB to PCT (gdal:rgbtopct)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - RGB to PCT
}
}
}
