% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_sieve.R
\name{gdal_sieve}
\alias{gdal_sieve}
\title{QGIS algorithm Sieve}
\usage{
gdal_sieve(
  INPUT = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  EIGHT_CONNECTEDNESS = qgisprocess::qgis_default_value(),
  NO_MASK = qgisprocess::qgis_default_value(),
  MASK_LAYER = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{THRESHOLD}{\code{number} - Threshold. A numeric value.}

\item{EIGHT_CONNECTEDNESS}{\code{boolean} - Use 8-connectedness. 1 for true/yes. 0 for false/no.}

\item{NO_MASK}{\code{boolean} - Do not use the default validity mask for the input band. 1 for true/yes. 0 for false/no.}

\item{MASK_LAYER}{\code{raster} - Validity mask. Path to a raster layer.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Sieved. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Sieve (gdal:sieve)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Sieved
}
}
}
