% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_triterrainruggednessindex.R
\name{gdal_triterrainruggednessindex}
\alias{gdal_triterrainruggednessindex}
\title{QGIS algorithm Terrain Ruggedness Index (TRI)}
\usage{
gdal_triterrainruggednessindex(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  COMPUTE_EDGES = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{COMPUTE_EDGES}{\code{boolean} - Compute edges. 1 for true/yes. 0 for false/no.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{OUTPUT}{\code{rasterDestination} - Terrain Ruggedness Index. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Terrain Ruggedness Index (TRI) (gdal:triterrainruggednessindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Terrain Ruggedness Index
}
}
}
