% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_warpreproject.R
\name{gdal_warpreproject}
\alias{gdal_warpreproject}
\title{QGIS algorithm Warp (reproject)}
\usage{
gdal_warpreproject(
  INPUT = qgisprocess::qgis_default_value(),
  SOURCE_CRS = qgisprocess::qgis_default_value(),
  TARGET_CRS = qgisprocess::qgis_default_value(),
  RESAMPLING = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  TARGET_RESOLUTION = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  TARGET_EXTENT = qgisprocess::qgis_default_value(),
  TARGET_EXTENT_CRS = qgisprocess::qgis_default_value(),
  MULTITHREADING = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{SOURCE_CRS}{\code{crs} - Source CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{TARGET_CRS}{\code{crs} - Target CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{RESAMPLING}{\code{enum}  of \verb{("Nearest Neighbour", "Bilinear", "Cubic", "Cubic Spline", "Lanczos Windowed Sinc", "Average", "Mode", "Maximum", "Minimum", "Median", "First Quartile", "Third Quartile")} - Resampling method to use. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NODATA}{\code{number} - Nodata value for output bands. A numeric value.}

\item{TARGET_RESOLUTION}{\code{number} - Output file resolution in target georeferenced units. A numeric value.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Use Input Layer Data Type", "Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_EXTENT}{\code{extent} - Georeferenced extents of output file to be created. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{TARGET_EXTENT_CRS}{\code{crs} - CRS of the target raster extent. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{MULTITHREADING}{\code{boolean} - Use multithreaded warping implementation. 1 for true/yes. 0 for false/no.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Reprojected. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Warp (reproject) (gdal:warpreproject)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Reprojected
}
}
}
