% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_albedo.R
\name{grass7_i_albedo}
\alias{grass7_i_albedo}
\title{QGIS algorithm i.albedo}
\usage{
grass7_i_albedo(
  input = qgisprocess::qgis_default_value(),
  .m = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  .l = qgisprocess::qgis_default_value(),
  .arg_8 = qgisprocess::qgis_default_value(),
  .a = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{multilayer} - Name of input raster maps. .}

\item{.m}{\code{boolean} - MODIS (7 input bands:1,2,3,4,5,6,7). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -m.}

\item{.n}{\code{boolean} - NOAA AVHRR (2 input bands:1,2). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.}

\item{.l}{\code{boolean} - Landsat 5+7 (6 input bands:1,2,3,4,5,7). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -l.}

\item{.arg_8}{\code{boolean} - Landsat 8 (7 input bands:1,2,3,4,5,6,7). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -8.}

\item{.a}{\code{boolean} - ASTER (6 input bands:1,3,5,6,8,9). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -a.}

\item{.c}{\code{boolean} - Aggressive mode (Landsat). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{.d}{\code{boolean} - Soft mode (MODIS). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -d.}

\item{output}{\code{rasterDestination} - Albedo. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS i.albedo (grass7:i.albedo)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Albedo
}
}
}
