% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_aster_toar.R
\name{grass7_i_aster_toar}
\alias{grass7_i_aster_toar}
\title{QGIS algorithm i.aster.toar}
\usage{
grass7_i_aster_toar(
  input = qgisprocess::qgis_default_value(),
  dayofyear = qgisprocess::qgis_default_value(),
  sun_elevation = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  .a = qgisprocess::qgis_default_value(),
  .b = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  .e = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{multilayer} - Names of ASTER DN layers (15 layers). .}

\item{dayofyear}{\code{number} - Day of Year of satellite overpass (0-366). A numeric value.}

\item{sun_elevation}{\code{number} - Sun elevation angle (degrees, < 90.0). A numeric value.}

\item{.r}{\code{boolean} - Output is radiance (W/m2). 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.a}{\code{boolean} - VNIR is High Gain. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.b}{\code{boolean} - SWIR is High Gain. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.c}{\code{boolean} - VNIR is Low Gain 1. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.d}{\code{boolean} - SWIR is Low Gain 1. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.e}{\code{boolean} - SWIR is Low Gain 2. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{folderDestination} - Output Directory. Path for an existing or new folder.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.aster.toar (grass7:i.aster.toar)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFolder - Output Directory
}
}
}
