% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_atcorr.R
\name{grass7_i_atcorr}
\alias{grass7_i_atcorr}
\title{QGIS algorithm i.atcorr}
\usage{
grass7_i_atcorr(
  input = qgisprocess::qgis_default_value(),
  range = qgisprocess::qgis_default_value(),
  elevation = qgisprocess::qgis_default_value(),
  visibility = qgisprocess::qgis_default_value(),
  parameters = qgisprocess::qgis_default_value(),
  rescale = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  .i = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  .a = qgisprocess::qgis_default_value(),
  .b = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{raster} - Name of input raster map. Path to a raster layer.}

\item{range}{\code{range} - Input imagery range (0,255). Two comma separated numeric values, e.g. '1,10'.}

\item{elevation}{\code{raster} - Input altitude raster map in m (optional). Path to a raster layer.}

\item{visibility}{\code{raster} - Input visibility raster map in km (optional). Path to a raster layer.}

\item{parameters}{\code{file} - Name of input text file. Path to a file.}

\item{rescale}{\code{range} - Rescale output raster map (0,255). Two comma separated numeric values, e.g. '1,10'.}

\item{output}{\code{rasterDestination} - Atmospheric correction. Path for new raster layer.}

\item{.i}{\code{boolean} - Output raster map as integer. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -i.}

\item{.r}{\code{boolean} - Input raster map converted to reflectance (default is radiance). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{.a}{\code{boolean} - Input from ETM+ image taken after July 1, 2000. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -a.}

\item{.b}{\code{boolean} - Input from ETM+ image taken before July 1, 2000. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -b.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS i.atcorr (grass7:i.atcorr)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Atmospheric correction
}
}
}
