% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_cluster.R
\name{grass7_i_cluster}
\alias{grass7_i_cluster}
\title{QGIS algorithm i.cluster}
\usage{
grass7_i_cluster(
  input = qgisprocess::qgis_default_value(),
  classes = qgisprocess::qgis_default_value(),
  seed = qgisprocess::qgis_default_value(),
  sample = qgisprocess::qgis_default_value(),
  iterations = qgisprocess::qgis_default_value(),
  convergence = qgisprocess::qgis_default_value(),
  separation = qgisprocess::qgis_default_value(),
  min_size = qgisprocess::qgis_default_value(),
  signaturefile = qgisprocess::qgis_default_value(),
  reportfile = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{multilayer} - Input rasters. .}

\item{classes}{\code{number} - Initial number of classes (1-255). A numeric value.}

\item{seed}{\code{file} - Name of file containing initial signatures. Path to a file.}

\item{sample}{\code{string} - Sampling intervals (by row and col). String value.}

\item{iterations}{\code{number} - Maximum number of iterations. A numeric value.}

\item{convergence}{\code{number} - Percent convergence. A numeric value.}

\item{separation}{\code{number} - Cluster separation. A numeric value.}

\item{min_size}{\code{number} - Minimum number of pixels in a class. A numeric value.}

\item{signaturefile}{\code{fileDestination} - Signature File. Path for new file.}

\item{reportfile}{\code{fileDestination} - Final Report File. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS i.cluster (grass7:i.cluster)
}
\details{
\subsection{Outputs description}{
\itemize{
\item reportfile - outputFile - Final Report File
\item signaturefile - outputFile - Signature File
}
}
}
