% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_colors_enhance.R
\name{grass7_i_colors_enhance}
\alias{grass7_i_colors_enhance}
\title{QGIS algorithm i.colors.enhance}
\usage{
grass7_i_colors_enhance(
  red = qgisprocess::qgis_default_value(),
  green = qgisprocess::qgis_default_value(),
  blue = qgisprocess::qgis_default_value(),
  strength = qgisprocess::qgis_default_value(),
  .f = qgisprocess::qgis_default_value(),
  .p = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  redoutput = qgisprocess::qgis_default_value(),
  greenoutput = qgisprocess::qgis_default_value(),
  blueoutput = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{red}{\code{raster} - Name of red channel. Path to a raster layer.}

\item{green}{\code{raster} - Name of green channel. Path to a raster layer.}

\item{blue}{\code{raster} - Name of blue channel. Path to a raster layer.}

\item{strength}{\code{number} - Cropping intensity (upper brightness level). A numeric value.}

\item{.f}{\code{boolean} - Extend colors to full range of data on each channel. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -f.}

\item{.p}{\code{boolean} - Preserve relative colors, adjust brightness only. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -p.}

\item{.r}{\code{boolean} - Reset to standard color range. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{.s}{\code{boolean} - Process bands serially (default: run in parallel). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -s.}

\item{redoutput}{\code{rasterDestination} - Enhanced Red. Path for new raster layer.}

\item{greenoutput}{\code{rasterDestination} - Enhanced Green. Path for new raster layer.}

\item{blueoutput}{\code{rasterDestination} - Enhanced Blue. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.colors.enhance (grass7:i.colors.enhance)
}
\details{
\subsection{Outputs description}{
\itemize{
\item blueoutput - outputRaster - Enhanced Blue
\item greenoutput - outputRaster - Enhanced Green
\item redoutput - outputRaster - Enhanced Red
}
}
}
