% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_eb_eta.R
\name{grass7_i_eb_eta}
\alias{grass7_i_eb_eta}
\title{QGIS algorithm i.eb.eta}
\usage{
grass7_i_eb_eta(
  netradiationdiurnal = qgisprocess::qgis_default_value(),
  evaporativefraction = qgisprocess::qgis_default_value(),
  temperature = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{netradiationdiurnal}{\code{raster} - Name of the diurnal net radiation map (W/m2). Path to a raster layer.}

\item{evaporativefraction}{\code{raster} - Name of the evaporative fraction map. Path to a raster layer.}

\item{temperature}{\code{raster} - Name of the surface skin temperature (K). Path to a raster layer.}

\item{output}{\code{rasterDestination} - Evapotranspiration. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.eb.eta (grass7:i.eb.eta)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Evapotranspiration
}
}
}
