% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_eb_evapfr.R
\name{grass7_i_eb_evapfr}
\alias{grass7_i_eb_evapfr}
\title{QGIS algorithm i.eb.evapfr}
\usage{
grass7_i_eb_evapfr(
  netradiation = qgisprocess::qgis_default_value(),
  soilheatflux = qgisprocess::qgis_default_value(),
  sensibleheatflux = qgisprocess::qgis_default_value(),
  evaporativefraction = qgisprocess::qgis_default_value(),
  soilmoisture = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{netradiation}{\code{raster} - Name of Net Radiation raster map (W/m2). Path to a raster layer.}

\item{soilheatflux}{\code{raster} - Name of soil heat flux raster map (W/m2). Path to a raster layer.}

\item{sensibleheatflux}{\code{raster} - Name of sensible heat flux raster map (W/m2). Path to a raster layer.}

\item{evaporativefraction}{\code{rasterDestination} - Evaporative Fraction. Path for new raster layer.}

\item{soilmoisture}{\code{rasterDestination} - Root Zone Soil Moisture. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.eb.evapfr (grass7:i.eb.evapfr)
}
\details{
\subsection{Outputs description}{
\itemize{
\item evaporativefraction - outputRaster - Evaporative Fraction
\item soilmoisture - outputRaster - Root Zone Soil Moisture
}
}
}
