% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_evapo_pm.R
\name{grass7_i_evapo_pm}
\alias{grass7_i_evapo_pm}
\title{QGIS algorithm i.evapo.pm}
\usage{
grass7_i_evapo_pm(
  elevation = qgisprocess::qgis_default_value(),
  temperature = qgisprocess::qgis_default_value(),
  relativehumidity = qgisprocess::qgis_default_value(),
  windspeed = qgisprocess::qgis_default_value(),
  netradiation = qgisprocess::qgis_default_value(),
  cropheight = qgisprocess::qgis_default_value(),
  .z = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{elevation}{\code{raster} - Name of input elevation raster map (m a.s.l.). Path to a raster layer.}

\item{temperature}{\code{raster} - Name of input temperature raster map (C). Path to a raster layer.}

\item{relativehumidity}{\code{raster} - Name of input relative humidity raster map (\%). Path to a raster layer.}

\item{windspeed}{\code{raster} - Name of input wind speed raster map (m/s). Path to a raster layer.}

\item{netradiation}{\code{raster} - Name of input net solar radiation raster map (MJ/m2/h). Path to a raster layer.}

\item{cropheight}{\code{raster} - Name of input crop height raster map (m). Path to a raster layer.}

\item{.z}{\code{boolean} - Set negative ETa to zero. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.n}{\code{boolean} - Use Night-time. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{rasterDestination} - Evapotranspiration. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.evapo.pm (grass7:i.evapo.pm)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Evapotranspiration
}
}
}
