% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_evapo_pt.R
\name{grass7_i_evapo_pt}
\alias{grass7_i_evapo_pt}
\title{QGIS algorithm i.evapo.pt}
\usage{
grass7_i_evapo_pt(
  net_radiation = qgisprocess::qgis_default_value(),
  soil_heatflux = qgisprocess::qgis_default_value(),
  air_temperature = qgisprocess::qgis_default_value(),
  atmospheric_pressure = qgisprocess::qgis_default_value(),
  priestley_taylor_coeff = qgisprocess::qgis_default_value(),
  .z = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{net_radiation}{\code{raster} - Name of input net radiation raster map (W/m2). Path to a raster layer.}

\item{soil_heatflux}{\code{raster} - Name of input soil heat flux raster map (W/m2). Path to a raster layer.}

\item{air_temperature}{\code{raster} - Name of input air temperature raster map (K). Path to a raster layer.}

\item{atmospheric_pressure}{\code{raster} - Name of input atmospheric pressure raster map (millibars). Path to a raster layer.}

\item{priestley_taylor_coeff}{\code{number} - Priestley-Taylor coefficient. A numeric value.}

\item{.z}{\code{boolean} - Set negative ETa to zero. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{rasterDestination} - Evapotranspiration. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.evapo.pt (grass7:i.evapo.pt)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Evapotranspiration
}
}
}
