% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_landsat_toar.R
\name{grass7_i_landsat_toar}
\alias{grass7_i_landsat_toar}
\title{QGIS algorithm i.landsat.toar}
\usage{
grass7_i_landsat_toar(
  rasters = qgisprocess::qgis_default_value(),
  metfile = qgisprocess::qgis_default_value(),
  sensor = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  date = qgisprocess::qgis_default_value(),
  sun_elevation = qgisprocess::qgis_default_value(),
  product_date = qgisprocess::qgis_default_value(),
  gain = qgisprocess::qgis_default_value(),
  percent = qgisprocess::qgis_default_value(),
  pixel = qgisprocess::qgis_default_value(),
  rayleigh = qgisprocess::qgis_default_value(),
  scale = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{rasters}{\code{multilayer} - Landsat input rasters. .}

\item{metfile}{\code{file} - Name of Landsat metadata file (.met or MTL.txt). Path to a file.}

\item{sensor}{\code{enum}  of \verb{("mss1", "mss2", "mss3", "mss4", "mss5", "tm4", "tm5", "tm7", "oli8")} - Spacecraft sensor. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{method}{\code{enum}  of \verb{("uncorrected", "dos1", "dos2", "dos2b", "dos3", "dos4")} - Atmospheric correction method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{date}{\code{string} - Image acquisition date (yyyy-mm-dd). String value.}

\item{sun_elevation}{\code{number} - Sun elevation in degrees. A numeric value.}

\item{product_date}{\code{string} - Image creation date (yyyy-mm-dd). String value.}

\item{gain}{\code{string} - Gain (H/L) of all Landsat ETM+ bands (1-5,61,62,7,8). String value.}

\item{percent}{\code{number} - Percent of solar radiance in path radiance. A numeric value.}

\item{pixel}{\code{number} - Minimum pixels to consider digital number as dark object. A numeric value.}

\item{rayleigh}{\code{number} - Rayleigh atmosphere (diffuse sky irradiance). A numeric value.}

\item{scale}{\code{number} - Scale factor for output. A numeric value.}

\item{.r}{\code{boolean} - Output at-sensor radiance instead of reflectance for all bands. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{.n}{\code{boolean} - Input raster maps use as extension the number of the band instead the code. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.}

\item{output}{\code{folderDestination} - Output Directory. Path for an existing or new folder.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.landsat.toar (grass7:i.landsat.toar)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFolder - Output Directory
}
}
}
