% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_oif.R
\name{grass7_i_oif}
\alias{grass7_i_oif}
\title{QGIS algorithm i.oif}
\usage{
grass7_i_oif(
  input = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{multilayer} - Name of input raster map(s). .}

\item{.g}{\code{boolean} - Print in shell script style. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -g.}

\item{.s}{\code{boolean} - Process bands serially (default: run in parallel). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -s.}

\item{output}{\code{fileDestination} - OIF File. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.oif (grass7:i.oif)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - OIF File
}
}
}
