% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_pansharpen.R
\name{grass7_i_pansharpen}
\alias{grass7_i_pansharpen}
\title{QGIS algorithm i.pansharpen}
\usage{
grass7_i_pansharpen(
  red = qgisprocess::qgis_default_value(),
  green = qgisprocess::qgis_default_value(),
  blue = qgisprocess::qgis_default_value(),
  pan = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  .l = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  redoutput = qgisprocess::qgis_default_value(),
  greenoutput = qgisprocess::qgis_default_value(),
  blueoutput = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{red}{\code{raster} - Name of red channel. Path to a raster layer.}

\item{green}{\code{raster} - Name of green channel. Path to a raster layer.}

\item{blue}{\code{raster} - Name of blue channel. Path to a raster layer.}

\item{pan}{\code{raster} - Name of raster map to be used for high resolution panchromatic channel. Path to a raster layer.}

\item{method}{\code{enum}  of \verb{("brovey", "ihs", "pca")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.l}{\code{boolean} - Rebalance blue channel for LANDSAT. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -l.}

\item{.s}{\code{boolean} - Process bands serially (default: run in parallel). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -s.}

\item{redoutput}{\code{rasterDestination} - Enhanced Red. Path for new raster layer.}

\item{greenoutput}{\code{rasterDestination} - Enhanced Green. Path for new raster layer.}

\item{blueoutput}{\code{rasterDestination} - Enhanced Blue. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.pansharpen (grass7:i.pansharpen)
}
\details{
\subsection{Outputs description}{
\itemize{
\item blueoutput - outputRaster - Enhanced Blue
\item greenoutput - outputRaster - Enhanced Green
\item redoutput - outputRaster - Enhanced Red
}
}
}
