% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_smap.R
\name{grass7_i_smap}
\alias{grass7_i_smap}
\title{QGIS algorithm i.smap}
\usage{
grass7_i_smap(
  input = qgisprocess::qgis_default_value(),
  signaturefile = qgisprocess::qgis_default_value(),
  blocksize = qgisprocess::qgis_default_value(),
  .m = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  goodness = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{multilayer} - Input rasters. .}

\item{signaturefile}{\code{file} - Name of input file containing signatures. Path to a file.}

\item{blocksize}{\code{number} - Size of submatrix to process at one time. A numeric value.}

\item{.m}{\code{boolean} - Use maximum likelihood estimation (instead of smap). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -m.}

\item{output}{\code{rasterDestination} - Classification. Path for new raster layer.}

\item{goodness}{\code{rasterDestination} - Goodness_of_fit. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS i.smap (grass7:i.smap)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Classification
\item goodness - outputRaster - Goodness_of_fit
}
}
}
