% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_topo_corr.R
\name{grass7_i_topo_corr}
\alias{grass7_i_topo_corr}
\title{QGIS algorithm i.topo.corr}
\usage{
grass7_i_topo_corr(
  input = qgisprocess::qgis_default_value(),
  basemap = qgisprocess::qgis_default_value(),
  zenith = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{multilayer} - Name of reflectance raster maps to be corrected topographically. .}

\item{basemap}{\code{raster} - Name of illumination input base raster map. Path to a raster layer.}

\item{zenith}{\code{number} - Solar zenith in degrees. A numeric value.}

\item{method}{\code{enum}  of \verb{("cosine", "minnaert", "c-factor", "percent")} - Topographic correction method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.s}{\code{boolean} - Scale output to input and copy color rules. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -s.}

\item{output}{\code{folderDestination} - Output Directory. Path for an existing or new folder.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS i.topo.corr (grass7:i.topo.corr)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFolder - Output Directory
}
}
}
