% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_i_vi.R
\name{grass7_i_vi}
\alias{grass7_i_vi}
\title{QGIS algorithm i.vi}
\usage{
grass7_i_vi(
  red = qgisprocess::qgis_default_value(),
  viname = qgisprocess::qgis_default_value(),
  nir = qgisprocess::qgis_default_value(),
  green = qgisprocess::qgis_default_value(),
  blue = qgisprocess::qgis_default_value(),
  band5 = qgisprocess::qgis_default_value(),
  band7 = qgisprocess::qgis_default_value(),
  soil_line_slope = qgisprocess::qgis_default_value(),
  soil_line_intercept = qgisprocess::qgis_default_value(),
  soil_noise_reduction = qgisprocess::qgis_default_value(),
  storage_bit = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{red}{\code{raster} - Name of input red channel surface reflectance map (0.0-1.0). Path to a raster layer.}

\item{viname}{\code{enum}  of \verb{("arvi", "dvi", "evi", "evi2", "gvi", "gari", "gemi", "ipvi", "msavi", "msavi2", "ndvi", "pvi", "savi", "sr", "vari", "wdvi")} - Type of vegetation index. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{nir}{\code{raster} - Name of input nir channel surface reflectance map (0.0-1.0). Path to a raster layer.}

\item{green}{\code{raster} - Name of input green channel surface reflectance map (0.0-1.0). Path to a raster layer.}

\item{blue}{\code{raster} - Name of input blue channel surface reflectance map (0.0-1.0). Path to a raster layer.}

\item{band5}{\code{raster} - Name of input 5th channel surface reflectance map (0.0-1.0). Path to a raster layer.}

\item{band7}{\code{raster} - Name of input 7th channel surface reflectance map (0.0-1.0). Path to a raster layer.}

\item{soil_line_slope}{\code{number} - Value of the slope of the soil line (MSAVI2 only). A numeric value.}

\item{soil_line_intercept}{\code{number} - Value of the factor of reduction of soil noise (MSAVI2 only). A numeric value.}

\item{soil_noise_reduction}{\code{number} - Value of the slope of the soil line (MSAVI2 only). A numeric value.}

\item{storage_bit}{\code{enum}  of \verb{("7", "8", "9", "10", "16")} - Maximum bits for digital numbers. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{output}{\code{rasterDestination} - Vegetation Index. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS i.vi (grass7:i.vi)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Vegetation Index
}
}
}
