% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_m_cogo.R
\name{grass7_m_cogo}
\alias{grass7_m_cogo}
\title{QGIS algorithm m.cogo}
\usage{
grass7_m_cogo(
  input = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  coordinates = qgisprocess:::qgis_default_value(),
  .l = qgisprocess:::qgis_default_value(),
  .q = qgisprocess:::qgis_default_value(),
  .r = qgisprocess:::qgis_default_value(),
  .c = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{file} - Name of input file. Path to a file.}

\item{output}{\code{fileDestination} - Output text file. Path for new file.}

\item{coordinates}{\code{point} - Starting coordinate pair. Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{.l}{\code{boolean} - Lines are labelled. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -l.}

\item{.q}{\code{boolean} - Suppress warnings. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -q.}

\item{.r}{\code{boolean} - Convert from coordinates to bearing and distance. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{.c}{\code{boolean} - Repeat the starting coordinate at the end to close a loop. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS m.cogo (grass7:m.cogo)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - Output text file
}
}
}
