% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_blend_rgb.R
\name{grass7_r_blend_rgb}
\alias{grass7_r_blend_rgb}
\title{QGIS algorithm r.blend.rgb}
\usage{
grass7_r_blend_rgb(
  first = qgisprocess::qgis_default_value(),
  second = qgisprocess::qgis_default_value(),
  percent = qgisprocess::qgis_default_value(),
  output_red = qgisprocess::qgis_default_value(),
  output_green = qgisprocess::qgis_default_value(),
  output_blue = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{first}{\code{raster} - Name of first raster map for blending. Path to a raster layer.}

\item{second}{\code{raster} - Name of second raster map for blending. Path to a raster layer.}

\item{percent}{\code{number} - Percentage weight of first map for color blending. A numeric value.}

\item{output_red}{\code{rasterDestination} - Blended Red. Path for new raster layer.}

\item{output_green}{\code{rasterDestination} - Blended Green. Path for new raster layer.}

\item{output_blue}{\code{rasterDestination} - Blended Blue. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.blend.rgb (grass7:r.blend.rgb)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output_blue - outputRaster - Blended Blue
\item output_green - outputRaster - Blended Green
\item output_red - outputRaster - Blended Red
}
}
}
