% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_category.R
\name{grass7_r_category}
\alias{grass7_r_category}
\title{QGIS algorithm r.category}
\usage{
grass7_r_category(
  map = qgisprocess::qgis_default_value(),
  separator = qgisprocess::qgis_default_value(),
  rules = qgisprocess::qgis_default_value(),
  txtrules = qgisprocess::qgis_default_value(),
  raster = qgisprocess::qgis_default_value(),
  format = qgisprocess::qgis_default_value(),
  coefficients = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{map}{\code{raster} - Name of raster map. Path to a raster layer.}

\item{separator}{\code{string} - Field separator (Special characters: pipe, comma, space, tab, newline). String value.}

\item{rules}{\code{file} - File containing category label rules. Path to a file.}

\item{txtrules}{\code{string} - Inline category label rules. String value.}

\item{raster}{\code{raster} - Raster map from which to copy category table. Path to a raster layer.}

\item{format}{\code{string} - Default label or format string for dynamic labeling. Used when no explicit label exists for the category. String value.}

\item{coefficients}{\code{string} - Dynamic label coefficients. Two pairs of category multiplier and offsets, for $1 and $2. String value.}

\item{output}{\code{rasterDestination} - Category. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.category (grass7:r.category)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Category
}
}
}
