% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_category_out.R
\name{grass7_r_category_out}
\alias{grass7_r_category_out}
\title{QGIS algorithm r.category.out}
\usage{
grass7_r_category_out(
  map = qgisprocess::qgis_default_value(),
  cats = qgisprocess::qgis_default_value(),
  values = qgisprocess::qgis_default_value(),
  separator = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{map}{\code{raster} - Name of raster map. Path to a raster layer.}

\item{cats}{\code{string} - Category values (for Integer rasters). Example: 1,3,7-9,13. String value.}

\item{values}{\code{string} - Comma separated value list (for float rasters). Example: 1.4,3.8,13. String value.}

\item{separator}{\code{string} - Field separator (Special characters: pipe, comma, space, tab, newline). String value.}

\item{html}{\code{fileDestination} - Category. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.category.out (grass7:r.category.out)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Category
}
}
}
