% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_clump.R
\name{grass7_r_clump}
\alias{grass7_r_clump}
\title{QGIS algorithm r.clump}
\usage{
grass7_r_clump(
  input = qgisprocess::qgis_default_value(),
  title = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  threshold = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Input layer. Path to a raster layer.}

\item{title}{\code{string} - Title for output raster map. String value.}

\item{.d}{\code{boolean} - Clump also diagonal cells. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -d.}

\item{output}{\code{rasterDestination} - Clumps. Path for new raster layer.}

\item{threshold}{\code{number} - Threshold to identify similar cells. A numeric value.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.clump (grass7:r.clump)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Clumps
}
}
}
