% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_coin.R
\name{grass7_r_coin}
\alias{grass7_r_coin}
\title{QGIS algorithm r.coin}
\usage{
grass7_r_coin(
  first = qgisprocess::qgis_default_value(),
  second = qgisprocess::qgis_default_value(),
  units = qgisprocess::qgis_default_value(),
  .w = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{first}{\code{raster} - Name of first raster map. Path to a raster layer.}

\item{second}{\code{raster} - Name of second raster map. Path to a raster layer.}

\item{units}{\code{enum}  of \verb{("c", "p", "x", "y", "a", "h", "k", "m")} - Unit of measure. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.w}{\code{boolean} - Wide report, 132 columns (default: 80). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -w.}

\item{html}{\code{fileDestination} - Coincidence report. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.coin (grass7:r.coin)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Coincidence report
}
}
}
