% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_colors.R
\name{grass7_r_colors}
\alias{grass7_r_colors}
\title{QGIS algorithm r.colors}
\usage{
grass7_r_colors(
  map = qgisprocess::qgis_default_value(),
  color = qgisprocess::qgis_default_value(),
  rules_txt = qgisprocess::qgis_default_value(),
  rules = qgisprocess::qgis_default_value(),
  raster = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  .w = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  .a = qgisprocess::qgis_default_value(),
  .e = qgisprocess::qgis_default_value(),
  output_dir = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{map}{\code{multilayer} - Name of raster maps(s). .}

\item{color}{\code{enum}  of \verb{("not selected", "aspect", "aspectcolr", "bcyr", "bgyr", "blues", "byg", "byr", "celsius", "corine", "curvature", "differences", "elevation", "etopo2", "evi", "fahrenheit", "gdd", "greens", "grey", "grey.eq", "grey.log", "grey1.0", "grey255", "gyr", "haxby", "kelvin", "ndvi", "ndwi", "oranges", "population", "population_dens", "precipitation", "precipitation_daily", "precipitation_monthly", "rainbow", "ramp", "random", "reds", "rstcurv", "ryb", "ryg", "sepia", "slope", "srtm", "srtm_plus", "terrain", "wave")} - Name of color table. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{rules_txt}{\code{string} - Color rules. String value.}

\item{rules}{\code{file} - Color rules file. Path to a file.}

\item{raster}{\code{raster} - Raster map from which to copy color table. Path to a raster layer.}

\item{.r}{\code{boolean} - Remove existing color table. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{.w}{\code{boolean} - Only write new color table if it does not already exist. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -w.}

\item{.n}{\code{boolean} - Invert colors. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.}

\item{.g}{\code{boolean} - Logarithmic scaling. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -g.}

\item{.a}{\code{boolean} - Logarithmic-absolute scaling. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -a.}

\item{.e}{\code{boolean} - Histogram equalization. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -e.}

\item{output_dir}{\code{folderDestination} - Output Directory. Path for an existing or new folder.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.colors (grass7:r.colors)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output_dir - outputFolder - Output Directory
}
}
}
