% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_composite.R
\name{grass7_r_composite}
\alias{grass7_r_composite}
\title{QGIS algorithm r.composite}
\usage{
grass7_r_composite(
  red = qgisprocess::qgis_default_value(),
  green = qgisprocess::qgis_default_value(),
  blue = qgisprocess::qgis_default_value(),
  levels = qgisprocess::qgis_default_value(),
  level_red = qgisprocess::qgis_default_value(),
  level_green = qgisprocess::qgis_default_value(),
  level_blue = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{red}{\code{raster} - Red. Path to a raster layer.}

\item{green}{\code{raster} - Green. Path to a raster layer.}

\item{blue}{\code{raster} - Blue. Path to a raster layer.}

\item{levels}{\code{number} - Number of levels to be used for each component. A numeric value.}

\item{level_red}{\code{number} - Number of levels to be used for red. A numeric value.}

\item{level_green}{\code{number} - Number of levels to be used for green. A numeric value.}

\item{level_blue}{\code{number} - Number of levels to be used for blue. A numeric value.}

\item{.d}{\code{boolean} - Dither. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -d.}

\item{.c}{\code{boolean} - Use closest color. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{output}{\code{rasterDestination} - Composite. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.composite (grass7:r.composite)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Composite
}
}
}
