% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_describe.R
\name{grass7_r_describe}
\alias{grass7_r_describe}
\title{QGIS algorithm r.describe}
\usage{
grass7_r_describe(
  map = qgisprocess:::qgis_default_value(),
  null_value = qgisprocess:::qgis_default_value(),
  nsteps = qgisprocess:::qgis_default_value(),
  .r = qgisprocess:::qgis_default_value(),
  .n = qgisprocess:::qgis_default_value(),
  .d = qgisprocess:::qgis_default_value(),
  .i = qgisprocess:::qgis_default_value(),
  html = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{map}{\code{raster} - input raster layer. Path to a raster layer.}

\item{null_value}{\code{string} - String representing NULL value. String value.}

\item{nsteps}{\code{number} - Number of quantization steps. A numeric value.}

\item{.r}{\code{boolean} - Only print the range of the data. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{.n}{\code{boolean} - Suppress reporting of any NULLs. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.}

\item{.d}{\code{boolean} - Use the current region. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -d.}

\item{.i}{\code{boolean} - Read floating-point map as integer. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -i.}

\item{html}{\code{fileDestination} - Categories. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.describe (grass7:r.describe)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Categories
}
}
}
