% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_distance.R
\name{grass7_r_distance}
\alias{grass7_r_distance}
\title{QGIS algorithm r.distance}
\usage{
grass7_r_distance(
  map = qgisprocess::qgis_default_value(),
  separator = qgisprocess::qgis_default_value(),
  sort = qgisprocess::qgis_default_value(),
  .l = qgisprocess::qgis_default_value(),
  .o = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{map}{\code{multilayer} - Name of two input raster for computing inter-class distances. .}

\item{separator}{\code{string} - Field separator (Special characters: pipe, comma, space, tab, newline). String value.}

\item{sort}{\code{enum}  of \verb{("asc", "desc")} - Sort output by distance. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.l}{\code{boolean} - Include category labels in the output. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.o}{\code{boolean} - Report zero distance if rasters are overlapping. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.n}{\code{boolean} - Report null objects as *. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{html}{\code{fileDestination} - Distance. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.distance (grass7:r.distance)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Distance
}
}
}
