% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_grow.R
\name{grass7_r_grow}
\alias{grass7_r_grow}
\title{QGIS algorithm r.grow}
\usage{
grass7_r_grow(
  input = qgisprocess::qgis_default_value(),
  radius = qgisprocess::qgis_default_value(),
  metric = qgisprocess::qgis_default_value(),
  old = qgisprocess::qgis_default_value(),
  new = qgisprocess::qgis_default_value(),
  .m = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{raster} - input raster layer. Path to a raster layer.}

\item{radius}{\code{number} - Radius of buffer in raster cells. A numeric value.}

\item{metric}{\code{enum}  of \verb{("euclidean", "maximum", "manhattan")} - Metric. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{old}{\code{number} - Value to write for input cells which are non-NULL (-1 = NULL). A numeric value.}

\item{new}{\code{number} - Value to write for "grown" cells. A numeric value.}

\item{.m}{\code{boolean} - Radius is in map units rather than cells. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -m.}

\item{output}{\code{rasterDestination} - Expanded. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.grow (grass7:r.grow)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Expanded
}
}
}
