% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_grow_distance.R
\name{grass7_r_grow_distance}
\alias{grass7_r_grow_distance}
\title{QGIS algorithm r.grow.distance}
\usage{
grass7_r_grow_distance(
  input = qgisprocess::qgis_default_value(),
  metric = qgisprocess::qgis_default_value(),
  .m = qgisprocess::qgis_default_value(),
  . = qgisprocess::qgis_default_value(),
  distance = qgisprocess::qgis_default_value(),
  value = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Input input raster layer. Path to a raster layer.}

\item{metric}{\code{enum}  of \verb{("euclidean", "squared", "maximum", "manhattan", "geodesic")} - Metric. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.m}{\code{boolean} - Output distances in meters instead of map units. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -m.}

\item{.}{\code{boolean} - Calculate distance to nearest NULL cell. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -.}

\item{distance}{\code{rasterDestination} - Distance. Path for new raster layer.}

\item{value}{\code{rasterDestination} - Value of nearest cell. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.grow.distance (grass7:r.grow.distance)
}
\details{
\subsection{Outputs description}{
\itemize{
\item distance - outputRaster - Distance
\item value - outputRaster - Value of nearest cell
}
}
}
