% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_his.R
\name{grass7_r_his}
\alias{grass7_r_his}
\title{QGIS algorithm r.his}
\usage{
grass7_r_his(
  hue = qgisprocess:::qgis_default_value(),
  intensity = qgisprocess:::qgis_default_value(),
  saturation = qgisprocess:::qgis_default_value(),
  bgcolor = qgisprocess:::qgis_default_value(),
  .c = qgisprocess:::qgis_default_value(),
  red = qgisprocess:::qgis_default_value(),
  green = qgisprocess:::qgis_default_value(),
  blue = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{hue}{\code{raster} - Hue. Path to a raster layer.}

\item{intensity}{\code{raster} - Intensity. Path to a raster layer.}

\item{saturation}{\code{raster} - Saturation. Path to a raster layer.}

\item{bgcolor}{\code{string} - Color to use instead of NULL values. Either a standard color name, R:G:B triplet, or "none". String value.}

\item{.c}{\code{boolean} - Use colors from color tables for NULL values. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{red}{\code{rasterDestination} - Red. Path for new raster layer.}

\item{green}{\code{rasterDestination} - Green. Path for new raster layer.}

\item{blue}{\code{rasterDestination} - Blue. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.his (grass7:r.his)
}
\details{
\subsection{Outputs description}{
\itemize{
\item blue - outputRaster - Blue
\item green - outputRaster - Green
\item red - outputRaster - Red
}
}
}
