% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_horizon.R
\name{grass7_r_horizon}
\alias{grass7_r_horizon}
\title{QGIS algorithm r.horizon}
\usage{
grass7_r_horizon(
  elevation = qgisprocess::qgis_default_value(),
  direction = qgisprocess::qgis_default_value(),
  step = qgisprocess::qgis_default_value(),
  start = qgisprocess::qgis_default_value(),
  end = qgisprocess::qgis_default_value(),
  bufferzone = qgisprocess::qgis_default_value(),
  e_buff = qgisprocess::qgis_default_value(),
  w_buff = qgisprocess::qgis_default_value(),
  n_buff = qgisprocess::qgis_default_value(),
  s_buff = qgisprocess::qgis_default_value(),
  maxdistance = qgisprocess::qgis_default_value(),
  distance = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{elevation}{\code{raster} - Name of input elevation raster map. Path to a raster layer.}

\item{direction}{\code{number} - Direction in which you want to know the horizon height. A numeric value.}

\item{step}{\code{number} - Angle step size for multidirectional horizon. A numeric value.}

\item{start}{\code{number} - Start angle for multidirectional horizon. A numeric value.}

\item{end}{\code{number} - End angle for multidirectional horizon. A numeric value.}

\item{bufferzone}{\code{number} - For horizon rasters, read from the DEM an extra buffer around the present region. A numeric value.}

\item{e_buff}{\code{number} - For horizon rasters, read from the DEM an extra buffer eastward the present region. A numeric value.}

\item{w_buff}{\code{number} - For horizon rasters, read from the DEM an extra buffer westward the present region. A numeric value.}

\item{n_buff}{\code{number} - For horizon rasters, read from the DEM an extra buffer northward the present region. A numeric value.}

\item{s_buff}{\code{number} - For horizon rasters, read from the DEM an extra buffer southward the present region. A numeric value.}

\item{maxdistance}{\code{number} - The maximum distance to consider when finding the horizon height. A numeric value.}

\item{distance}{\code{number} - Sampling distance step coefficient. A numeric value.}

\item{.d}{\code{boolean} - Write output in degrees (default is radians). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -d.}

\item{.c}{\code{boolean} - Write output in compass orientation (default is CCW, East=0). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{output}{\code{folderDestination} - Folder to get horizon rasters. Path for an existing or new folder.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.horizon (grass7:r.horizon)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFolder - Folder to get horizon rasters
}
}
}
