% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_in_lidar.R
\name{grass7_r_in_lidar}
\alias{grass7_r_in_lidar}
\title{QGIS algorithm r.in.lidar}
\usage{
grass7_r_in_lidar(
  input = qgisprocess:::qgis_default_value(),
  method = qgisprocess:::qgis_default_value(),
  type = qgisprocess:::qgis_default_value(),
  base_raster = qgisprocess:::qgis_default_value(),
  zrange = qgisprocess:::qgis_default_value(),
  zscale = qgisprocess:::qgis_default_value(),
  intensity_range = qgisprocess:::qgis_default_value(),
  intensity_scale = qgisprocess:::qgis_default_value(),
  percent = qgisprocess:::qgis_default_value(),
  pth = qgisprocess:::qgis_default_value(),
  trim = qgisprocess:::qgis_default_value(),
  resolution = qgisprocess:::qgis_default_value(),
  return_filter = qgisprocess:::qgis_default_value(),
  class_filter = qgisprocess:::qgis_default_value(),
  .e = qgisprocess:::qgis_default_value(),
  .n = qgisprocess:::qgis_default_value(),
  .o = qgisprocess:::qgis_default_value(),
  .i = qgisprocess:::qgis_default_value(),
  .j = qgisprocess:::qgis_default_value(),
  .d = qgisprocess:::qgis_default_value(),
  .v = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{file} - LAS input file. Path to a file.}

\item{method}{\code{enum}  of \verb{("n", "min", "max", "range", "sum", "mean", "stddev", "variance", "coeff_var", "median", "percentile", "skewness", "trimmean")} - Statistic to use for raster values. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{type}{\code{enum}  of \verb{("CELL", "FCELL", "DCELL")} - Storage type for resultant raster map. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{base_raster}{\code{raster} - Subtract raster values from the Z coordinates. Path to a raster layer.}

\item{zrange}{\code{range} - Filter range for z data (min, max). Two comma separated numeric values, e.g. '1,10'.}

\item{zscale}{\code{number} - Scale to apply to z data. A numeric value.}

\item{intensity_range}{\code{range} - Filter range for intensity values (min, max). Two comma separated numeric values, e.g. '1,10'.}

\item{intensity_scale}{\code{number} - Scale to apply to intensity values. A numeric value.}

\item{percent}{\code{number} - Percent of map to keep in memory. A numeric value.}

\item{pth}{\code{number} - pth percentile of the values (between 1 and 100). A numeric value.}

\item{trim}{\code{number} - Discard trim percent of the smallest and trim percent of the largest observations (0-50). A numeric value.}

\item{resolution}{\code{number} - Output raster resolution. A numeric value.}

\item{return_filter}{\code{string} - Only import points of selected return type Options: first, last, mid. String value.}

\item{class_filter}{\code{string} - Only import points of selected class(es) (comma separated integers). String value.}

\item{.e}{\code{boolean} - Use the extent of the input for the raster extent. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -e.}

\item{.n}{\code{boolean} - Set computation region to match the new raster map. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.}

\item{.o}{\code{boolean} - Override projection check (use current location's projection). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -o.}

\item{.i}{\code{boolean} - Use intensity values rather than Z values. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -i.}

\item{.j}{\code{boolean} - Use Z values for filtering, but intensity values for statistics. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -j.}

\item{.d}{\code{boolean} - Use base raster resolution instead of computational region. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -d.}

\item{.v}{\code{boolean} - Use only valid points. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -v.}

\item{output}{\code{rasterDestination} - Lidar Raster. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.in.lidar (grass7:r.in.lidar)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Lidar Raster
}
}
}
