% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_in_lidar_info.R
\name{grass7_r_in_lidar_info}
\alias{grass7_r_in_lidar_info}
\title{QGIS algorithm r.in.lidar.info}
\usage{
grass7_r_in_lidar_info(
  input = qgisprocess:::qgis_default_value(),
  html = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{file} - LAS input file. Path to a file.}

\item{html}{\code{fileDestination} - LAS information. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.in.lidar.info (grass7:r.in.lidar.info)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - LAS information
}
}
}
