% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_info.R
\name{grass7_r_info}
\alias{grass7_r_info}
\title{QGIS algorithm r.info}
\usage{
grass7_r_info(
  map = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  .h = qgisprocess::qgis_default_value(),
  .e = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{map}{\code{raster} - Raster layer. Path to a raster layer.}

\item{.r}{\code{boolean} - Print range only. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.g}{\code{boolean} - Print raster array information in shell script style. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.h}{\code{boolean} - Print raster history instead of info. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.e}{\code{boolean} - Print extended metadata information in shell script style. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{html}{\code{fileDestination} - Basic information. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.info (grass7:r.info)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Basic information
}
}
}
