% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_kappa.R
\name{grass7_r_kappa}
\alias{grass7_r_kappa}
\title{QGIS algorithm r.kappa}
\usage{
grass7_r_kappa(
  classification = qgisprocess::qgis_default_value(),
  reference = qgisprocess::qgis_default_value(),
  title = qgisprocess::qgis_default_value(),
  .h = qgisprocess::qgis_default_value(),
  .w = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{classification}{\code{raster} - Raster layer containing classification result. Path to a raster layer.}

\item{reference}{\code{raster} - Raster layer containing reference classes. Path to a raster layer.}

\item{title}{\code{string} - Title for error matrix and kappa. String value.}

\item{.h}{\code{boolean} - No header in the report. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -h.}

\item{.w}{\code{boolean} - Wide report (132 columns). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -w.}

\item{output}{\code{fileDestination} - Error matrix and kappa. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.kappa (grass7:r.kappa)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - Error matrix and kappa
}
}
}
