% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_lake.R
\name{grass7_r_lake}
\alias{grass7_r_lake}
\title{QGIS algorithm r.lake}
\usage{
grass7_r_lake(
  elevation = qgisprocess::qgis_default_value(),
  water_level = qgisprocess::qgis_default_value(),
  coordinates = qgisprocess::qgis_default_value(),
  seed = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  lake = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{elevation}{\code{raster} - Elevation. Path to a raster layer.}

\item{water_level}{\code{number} - Water level. A numeric value.}

\item{coordinates}{\code{point} - Seed point coordinates. Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{seed}{\code{raster} - Raster layer with starting point(s) (at least 1 cell > 0). Path to a raster layer.}

\item{.n}{\code{boolean} - Use negative depth values for lake raster layer. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{lake}{\code{rasterDestination} - Lake. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.lake (grass7:r.lake)
}
\details{
\subsection{Outputs description}{
\itemize{
\item lake - outputRaster - Lake
}
}
}
