% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_li_renyi.R
\name{grass7_r_li_renyi}
\alias{grass7_r_li_renyi}
\title{QGIS algorithm r.li.renyi}
\usage{
grass7_r_li_renyi(
  input = qgisprocess::qgis_default_value(),
  config_txt = qgisprocess::qgis_default_value(),
  config = qgisprocess::qgis_default_value(),
  alpha = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Name of input raster map. Path to a raster layer.}

\item{config_txt}{\code{string} - Landscape structure configuration. String value.}

\item{config}{\code{file} - Landscape structure configuration file. Path to a file.}

\item{alpha}{\code{string} - Alpha value is the order of the generalized entropy. String value.}

\item{output}{\code{rasterDestination} - Renyi. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.li.renyi (grass7:r.li.renyi)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Renyi
}
}
}
