% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_mapcalc_simple.R
\name{grass7_r_mapcalc_simple}
\alias{grass7_r_mapcalc_simple}
\title{QGIS algorithm r.mapcalc.simple}
\usage{
grass7_r_mapcalc_simple(
  a = qgisprocess:::qgis_default_value(),
  b = qgisprocess:::qgis_default_value(),
  c = qgisprocess:::qgis_default_value(),
  d = qgisprocess:::qgis_default_value(),
  e = qgisprocess:::qgis_default_value(),
  f = qgisprocess:::qgis_default_value(),
  expression = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{a}{\code{raster} - Raster layer A. Path to a raster layer.}

\item{b}{\code{raster} - Raster layer B. Path to a raster layer.}

\item{c}{\code{raster} - Raster layer C. Path to a raster layer.}

\item{d}{\code{raster} - Raster layer D. Path to a raster layer.}

\item{e}{\code{raster} - Raster layer E. Path to a raster layer.}

\item{f}{\code{raster} - Raster layer F. Path to a raster layer.}

\item{expression}{\code{string} - Formula. String value.}

\item{output}{\code{rasterDestination} - Calculated. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.mapcalc.simple (grass7:r.mapcalc.simple)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Calculated
}
}
}
