% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_mask_rast.R
\name{grass7_r_mask_rast}
\alias{grass7_r_mask_rast}
\title{QGIS algorithm r.mask.rast}
\usage{
grass7_r_mask_rast(
  raster = qgisprocess::qgis_default_value(),
  input = qgisprocess::qgis_default_value(),
  maskcats = qgisprocess::qgis_default_value(),
  .i = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{raster}{\code{raster} - Name of raster map to use as mask. Path to a raster layer.}

\item{input}{\code{raster} - Name of raster map to which apply the mask. Path to a raster layer.}

\item{maskcats}{\code{string} - Raster values to use for mask. Format: 1 2 3 thru 7 *. String value.}

\item{.i}{\code{boolean} - Create inverse mask. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -i.}

\item{output}{\code{rasterDestination} - Masked. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.mask.rast (grass7:r.mask.rast)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Masked
}
}
}
