% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_neighbors.R
\name{grass7_r_neighbors}
\alias{grass7_r_neighbors}
\title{QGIS algorithm r.neighbors}
\usage{
grass7_r_neighbors(
  input = qgisprocess::qgis_default_value(),
  selection = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  size = qgisprocess::qgis_default_value(),
  gauss = qgisprocess::qgis_default_value(),
  quantile = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  .a = qgisprocess::qgis_default_value(),
  weight = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Input raster layer. Path to a raster layer.}

\item{selection}{\code{raster} - Raster layer to select the cells which should be processed. Path to a raster layer.}

\item{method}{\code{enum}  of \verb{("average", "median", "mode", "minimum", "maximum", "range", "stddev", "sum", "count", "variance", "diversity", "interspersion", "quart1", "quart3", "perc90", "quantile")} - Neighborhood operation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{size}{\code{number} - Neighborhood size (must be odd). A numeric value.}

\item{gauss}{\code{number} - Sigma (in cells) for Gaussian filter. A numeric value.}

\item{quantile}{\code{string} - Quantile to calculate for method=quantile. String value.}

\item{.c}{\code{boolean} - Use circular neighborhood. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{.a}{\code{boolean} - Do not align output with the input. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -a.}

\item{weight}{\code{file} - File containing weights. Path to a file.}

\item{output}{\code{rasterDestination} - Neighbors. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.neighbors (grass7:r.neighbors)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Neighbors
}
}
}
