% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_out_ascii.R
\name{grass7_r_out_ascii}
\alias{grass7_r_out_ascii}
\title{QGIS algorithm r.out.ascii}
\usage{
grass7_r_out_ascii(
  input = qgisprocess::qgis_default_value(),
  precision = qgisprocess::qgis_default_value(),
  width = qgisprocess::qgis_default_value(),
  null_value = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  .m = qgisprocess::qgis_default_value(),
  .i = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Input raster. Path to a raster layer.}

\item{precision}{\code{number} - Number of significant digits. A numeric value.}

\item{width}{\code{number} - Number of values printed before wrapping a line (only SURFER or MODFLOW format). A numeric value.}

\item{null_value}{\code{string} - String to represent null cell (GRASS grid only). String value.}

\item{.s}{\code{boolean} - Write SURFER (Golden Software) ASCII grid. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -s.}

\item{.m}{\code{boolean} - Write MODFLOW (USGS) ASCII array. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -m.}

\item{.i}{\code{boolean} - Force output of integer values. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -i.}

\item{output}{\code{fileDestination} - GRASS Ascii. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.out.ascii (grass7:r.out.ascii)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - GRASS Ascii
}
}
}
