% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_out_mpeg.R
\name{grass7_r_out_mpeg}
\alias{grass7_r_out_mpeg}
\title{QGIS algorithm r.out.mpeg}
\usage{
grass7_r_out_mpeg(
  view1 = qgisprocess::qgis_default_value(),
  view2 = qgisprocess::qgis_default_value(),
  view3 = qgisprocess::qgis_default_value(),
  view4 = qgisprocess::qgis_default_value(),
  quality = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{view1}{\code{multilayer} - Name of input raster map(s) for view no.1. .}

\item{view2}{\code{multilayer} - Name of input raster map(s) for view no.2. .}

\item{view3}{\code{multilayer} - Name of input raster map(s) for view no.3. .}

\item{view4}{\code{multilayer} - Name of input raster map(s) for view no.4. .}

\item{quality}{\code{number} - Quality factor (1 = highest quality, lowest compression). A numeric value.}

\item{output}{\code{fileDestination} - MPEG file. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.out.mpeg (grass7:r.out.mpeg)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - MPEG file
}
}
}
