% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_out_png.R
\name{grass7_r_out_png}
\alias{grass7_r_out_png}
\title{QGIS algorithm r.out.png}
\usage{
grass7_r_out_png(
  input = qgisprocess:::qgis_default_value(),
  compression = qgisprocess:::qgis_default_value(),
  .t = qgisprocess:::qgis_default_value(),
  .w = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{raster} - Input raster. Path to a raster layer.}

\item{compression}{\code{number} - Compression level of PNG file (0 = none, 1 = fastest, 9 = best). A numeric value.}

\item{.t}{\code{boolean} - Make NULL cells transparent. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{.w}{\code{boolean} - Output world file. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -w.}

\item{output}{\code{fileDestination} - PNG File. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.out.png (grass7:r.out.png)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - PNG File
}
}
}
