% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_out_ppm.R
\name{grass7_r_out_ppm}
\alias{grass7_r_out_ppm}
\title{QGIS algorithm r.out.ppm}
\usage{
grass7_r_out_ppm(
  input = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Input raster layer. Path to a raster layer.}

\item{.g}{\code{boolean} - Output greyscale instead of color. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -g.}

\item{output}{\code{fileDestination} - PPM. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.out.ppm (grass7:r.out.ppm)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - PPM
}
}
}
