% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_out_ppm3.R
\name{grass7_r_out_ppm3}
\alias{grass7_r_out_ppm3}
\title{QGIS algorithm r.out.ppm3}
\usage{
grass7_r_out_ppm3(
  red = qgisprocess::qgis_default_value(),
  green = qgisprocess::qgis_default_value(),
  blue = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{red}{\code{raster} - Name of raster map to be used for red. Path to a raster layer.}

\item{green}{\code{raster} - Name of raster map to be used for green. Path to a raster layer.}

\item{blue}{\code{raster} - Name of raster map to be used for blue. Path to a raster layer.}

\item{.c}{\code{boolean} - Add comments to describe the region. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{output}{\code{fileDestination} - Name for new PPM file. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.out.ppm3 (grass7:r.out.ppm3)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - Name for new PPM file
}
}
}
