% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_profile.R
\name{grass7_r_profile}
\alias{grass7_r_profile}
\title{QGIS algorithm r.profile}
\usage{
grass7_r_profile(
  input = qgisprocess::qgis_default_value(),
  coordinates = qgisprocess::qgis_default_value(),
  resolution = qgisprocess::qgis_default_value(),
  null_value = qgisprocess::qgis_default_value(),
  file = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Input raster layer. Path to a raster layer.}

\item{coordinates}{\code{string} - Profile coordinate pairs. String value.}

\item{resolution}{\code{number} - Resolution along profile. A numeric value.}

\item{null_value}{\code{string} - Character to represent no data cell. String value.}

\item{file}{\code{file} - Name of input file containing coordinate pairs. Path to a file.}

\item{.g}{\code{boolean} - Output easting and northing in first two columns of four column output. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.c}{\code{boolean} - Output RRR:GGG:BBB color values for each profile point. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{fileDestination} - Profile. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.profile (grass7:r.profile)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - Profile
}
}
}
