% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_quant.R
\name{grass7_r_quant}
\alias{grass7_r_quant}
\title{QGIS algorithm r.quant}
\usage{
grass7_r_quant(
  input = qgisprocess::qgis_default_value(),
  basemap = qgisprocess::qgis_default_value(),
  fprange = qgisprocess::qgis_default_value(),
  range = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{multilayer} - Raster layer(s) to be quantized. .}

\item{basemap}{\code{raster} - Base layer to take quant rules from. Path to a raster layer.}

\item{fprange}{\code{range} - Floating point range: dmin,dmax. Two comma separated numeric values, e.g. '1,10'.}

\item{range}{\code{range} - Integer range: min,max. Two comma separated numeric values, e.g. '1,10'.}

\item{.t}{\code{boolean} - Truncate floating point data. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{.r}{\code{boolean} - Round floating point data. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{output}{\code{folderDestination} - Quantized raster(s). Path for an existing or new folder.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.quant (grass7:r.quant)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFolder - Quantized raster(s)
}
}
}
