% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_random.R
\name{grass7_r_random}
\alias{grass7_r_random}
\title{QGIS algorithm r.random}
\usage{
grass7_r_random(
  input = qgisprocess::qgis_default_value(),
  cover = qgisprocess::qgis_default_value(),
  npoints = qgisprocess::qgis_default_value(),
  .z = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  .b = qgisprocess::qgis_default_value(),
  raster = qgisprocess::qgis_default_value(),
  vector = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Input raster layer. Path to a raster layer.}

\item{cover}{\code{raster} - Input cover raster layer. Path to a raster layer.}

\item{npoints}{\code{number} - The number of points to allocate. A numeric value.}

\item{.z}{\code{boolean} - Generate points also for NULL category. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.d}{\code{boolean} - Generate vector points as 3D points. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.b}{\code{boolean} - Do not build topology. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{raster}{\code{rasterDestination} - Random raster. Path for new raster layer.}

\item{vector}{\code{vectorDestination} - Random vector. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.random (grass7:r.random)
}
\details{
\subsection{Outputs description}{
\itemize{
\item raster - outputRaster - Random raster
\item vector - outputVector - Random vector
}
}
}
