% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_reclass_area.R
\name{grass7_r_reclass_area}
\alias{grass7_r_reclass_area}
\title{QGIS algorithm r.reclass.area}
\usage{
grass7_r_reclass_area(
  input = qgisprocess::qgis_default_value(),
  value = qgisprocess::qgis_default_value(),
  mode = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Input raster layer. Path to a raster layer.}

\item{value}{\code{number} - Value option that sets the area size limit (hectares). A numeric value.}

\item{mode}{\code{enum}  of \verb{("lesser", "greater")} - Lesser or greater than specified value. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{method}{\code{enum}  of \verb{("reclass", "rmarea")} - Method used for reclassification. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.c}{\code{boolean} - Input map is clumped. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.d}{\code{boolean} - Clumps including diagonal neighbors. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{rasterDestination} - Reclassified. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.reclass.area (grass7:r.reclass.area)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Reclassified
}
}
}
