% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_regression_line.R
\name{grass7_r_regression_line}
\alias{grass7_r_regression_line}
\title{QGIS algorithm r.regression.line}
\usage{
grass7_r_regression_line(
  mapx = qgisprocess::qgis_default_value(),
  mapy = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{mapx}{\code{raster} - Layer for x coefficient. Path to a raster layer.}

\item{mapy}{\code{raster} - Layer for y coefficient. Path to a raster layer.}

\item{html}{\code{fileDestination} - Regression coefficients. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.regression.line (grass7:r.regression.line)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Regression coefficients
}
}
}
