% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_resamp_filter.R
\name{grass7_r_resamp_filter}
\alias{grass7_r_resamp_filter}
\title{QGIS algorithm r.resamp.filter}
\usage{
grass7_r_resamp_filter(
  input = qgisprocess::qgis_default_value(),
  filter = qgisprocess::qgis_default_value(),
  radius = qgisprocess::qgis_default_value(),
  x_radius = qgisprocess::qgis_default_value(),
  y_radius = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Input raster layer. Path to a raster layer.}

\item{filter}{\code{enum}  of \verb{("box", "bartlett", "gauss", "normal", "hermite", "sinc", "lanczos1", "lanczos2", "lanczos3", "hann", "hamming", "blackman")} - Filter kernel(s). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{radius}{\code{string} - Filter radius for each filter (comma separated list of float if multiple). String value.}

\item{x_radius}{\code{string} - Filter radius (horizontal) for each filter (comma separated list of float if multiple). String value.}

\item{y_radius}{\code{string} - Filter radius (vertical) for each filter (comma separated list of float if multiple). String value.}

\item{.n}{\code{boolean} - Propagate NULLs. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{rasterDestination} - Resampled Filter. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.resamp.filter (grass7:r.resamp.filter)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Resampled Filter
}
}
}
