% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_resamp_rst.R
\name{grass7_r_resamp_rst}
\alias{grass7_r_resamp_rst}
\title{QGIS algorithm r.resamp.rst}
\usage{
grass7_r_resamp_rst(
  input = qgisprocess::qgis_default_value(),
  smooth = qgisprocess::qgis_default_value(),
  maskmap = qgisprocess::qgis_default_value(),
  ew_res = qgisprocess::qgis_default_value(),
  ns_res = qgisprocess::qgis_default_value(),
  overlap = qgisprocess::qgis_default_value(),
  zscale = qgisprocess::qgis_default_value(),
  tension = qgisprocess::qgis_default_value(),
  theta = qgisprocess::qgis_default_value(),
  scalex = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  elevation = qgisprocess::qgis_default_value(),
  slope = qgisprocess::qgis_default_value(),
  aspect = qgisprocess::qgis_default_value(),
  pcurvature = qgisprocess::qgis_default_value(),
  tcurvature = qgisprocess::qgis_default_value(),
  mcurvature = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{raster} - Raster layer. Path to a raster layer.}

\item{smooth}{\code{raster} - Input raster map containing smoothing. Path to a raster layer.}

\item{maskmap}{\code{raster} - Input raster map to be used as mask. Path to a raster layer.}

\item{ew_res}{\code{number} - Desired east-west resolution. A numeric value.}

\item{ns_res}{\code{number} - Desired north-south resolution. A numeric value.}

\item{overlap}{\code{number} - Rows/columns overlap for segmentation. A numeric value.}

\item{zscale}{\code{number} - Multiplier for z-values. A numeric value.}

\item{tension}{\code{number} - Spline tension value. A numeric value.}

\item{theta}{\code{number} - Anisotropy angle (in degrees counterclockwise from East). A numeric value.}

\item{scalex}{\code{number} - Anisotropy scaling factor. A numeric value.}

\item{.t}{\code{boolean} - Use dnorm independent tension. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{.d}{\code{boolean} - Output partial derivatives instead of topographic parameters. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -d.}

\item{elevation}{\code{rasterDestination} - Resampled RST. Path for new raster layer.}

\item{slope}{\code{rasterDestination} - Slope raster. Path for new raster layer.}

\item{aspect}{\code{rasterDestination} - Aspect raster. Path for new raster layer.}

\item{pcurvature}{\code{rasterDestination} - Profile curvature raster. Path for new raster layer.}

\item{tcurvature}{\code{rasterDestination} - Tangential curvature raster. Path for new raster layer.}

\item{mcurvature}{\code{rasterDestination} - Mean curvature raster. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.resamp.rst (grass7:r.resamp.rst)
}
\details{
\subsection{Outputs description}{
\itemize{
\item aspect - outputRaster - Aspect raster
\item elevation - outputRaster - Resampled RST
\item mcurvature - outputRaster - Mean curvature raster
\item pcurvature - outputRaster - Profile curvature raster
\item slope - outputRaster - Slope raster
\item tcurvature - outputRaster - Tangential curvature raster
}
}
}
