% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_ros.R
\name{grass7_r_ros}
\alias{grass7_r_ros}
\title{QGIS algorithm r.ros}
\usage{
grass7_r_ros(
  model = qgisprocess:::qgis_default_value(),
  moisture_1h = qgisprocess:::qgis_default_value(),
  moisture_10h = qgisprocess:::qgis_default_value(),
  moisture_100h = qgisprocess:::qgis_default_value(),
  moisture_live = qgisprocess:::qgis_default_value(),
  velocity = qgisprocess:::qgis_default_value(),
  direction = qgisprocess:::qgis_default_value(),
  slope = qgisprocess:::qgis_default_value(),
  aspect = qgisprocess:::qgis_default_value(),
  elevation = qgisprocess:::qgis_default_value(),
  base_ros = qgisprocess:::qgis_default_value(),
  max_ros = qgisprocess:::qgis_default_value(),
  direction_ros = qgisprocess:::qgis_default_value(),
  spotting_distance = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{model}{\code{raster} - Raster map containing fuel models. Path to a raster layer.}

\item{moisture_1h}{\code{raster} - Raster map containing the 1-hour fuel moisture (\%). Path to a raster layer.}

\item{moisture_10h}{\code{raster} - Raster map containing the 10-hour fuel moisture (\%). Path to a raster layer.}

\item{moisture_100h}{\code{raster} - Raster map containing the 100-hour fuel moisture (\%). Path to a raster layer.}

\item{moisture_live}{\code{raster} - Raster map containing live fuel moisture (\%). Path to a raster layer.}

\item{velocity}{\code{raster} - Raster map containing midflame wind velocities (ft/min). Path to a raster layer.}

\item{direction}{\code{raster} - Name of raster map containing wind directions (degree). Path to a raster layer.}

\item{slope}{\code{raster} - Name of raster map containing slope (degree). Path to a raster layer.}

\item{aspect}{\code{raster} - Raster map containing aspect (degree, CCW from E). Path to a raster layer.}

\item{elevation}{\code{raster} - Raster map containing elevation (m, required for spotting). Path to a raster layer.}

\item{base_ros}{\code{rasterDestination} - Base ROS. Path for new raster layer.}

\item{max_ros}{\code{rasterDestination} - Max ROS. Path for new raster layer.}

\item{direction_ros}{\code{rasterDestination} - Direction ROS. Path for new raster layer.}

\item{spotting_distance}{\code{rasterDestination} - Spotting Distance. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.ros (grass7:r.ros)
}
\details{
\subsection{Outputs description}{
\itemize{
\item base_ros - outputRaster - Base ROS
\item direction_ros - outputRaster - Direction ROS
\item max_ros - outputRaster - Max ROS
\item spotting_distance - outputRaster - Spotting Distance
}
}
}
