% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_series.R
\name{grass7_r_series}
\alias{grass7_r_series}
\title{QGIS algorithm r.series}
\usage{
grass7_r_series(
  input = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  quantile = qgisprocess::qgis_default_value(),
  weights = qgisprocess::qgis_default_value(),
  range = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{multilayer} - Input raster layer(s). .}

\item{.n}{\code{boolean} - Propagate NULLs. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.}

\item{method}{\code{enum}  of \verb{("average", "count", "median", "mode", "minimum", "min_raster", "maximum", "max_raster", "stddev", "range", "sum", "variance", "diversity", "slope", "offset", "detcoeff", "quart1", "quart3", "perc90", "skewness", "kurtosis", "quantile")} - Aggregate operation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{quantile}{\code{string} - Quantile to calculate for method=quantile. String value.}

\item{weights}{\code{string} - Weighting factor for each input map, default value is 1.0. String value.}

\item{range}{\code{range} - Ignore values outside this range (lo,hi). Two comma separated numeric values, e.g. '1,10'.}

\item{output}{\code{rasterDestination} - Aggregated. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.series (grass7:r.series)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Aggregated
}
}
}
