% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_shade.R
\name{grass7_r_shade}
\alias{grass7_r_shade}
\title{QGIS algorithm r.shade}
\usage{
grass7_r_shade(
  shade = qgisprocess::qgis_default_value(),
  color = qgisprocess::qgis_default_value(),
  brighten = qgisprocess::qgis_default_value(),
  bgcolor = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{shade}{\code{raster} - Name of shaded relief or aspect raster map. Path to a raster layer.}

\item{color}{\code{raster} - Name of raster to drape over relief raster map. Path to a raster layer.}

\item{brighten}{\code{number} - Percent to brighten. A numeric value.}

\item{bgcolor}{\code{string} - Color to use instead of NULL values. Either a standard color name, R:G:B triplet, or "none". String value.}

\item{.c}{\code{boolean} - Use colors from color tables for NULL values. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{rasterDestination} - Shaded. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.shade (grass7:r.shade)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Shaded
}
}
}
