% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_statistics.R
\name{grass7_r_statistics}
\alias{grass7_r_statistics}
\title{QGIS algorithm r.statistics}
\usage{
grass7_r_statistics(
  base = qgisprocess::qgis_default_value(),
  cover = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  routput = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{base}{\code{raster} - Base raster layer. Path to a raster layer.}

\item{cover}{\code{raster} - Cover raster layer. Path to a raster layer.}

\item{method}{\code{enum}  of \verb{("diversity", "average", "mode", "median", "avedev", "stddev", "variance", "skewness", "kurtosis", "min", "max", "sum")} - method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.c}{\code{boolean} - Cover values extracted from the category labels of the cover map. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{routput}{\code{rasterDestination} - Statistics. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.statistics (grass7:r.statistics)
}
\details{
\subsection{Outputs description}{
\itemize{
\item routput - outputRaster - Statistics
}
}
}
