% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_stats_quantile_out.R
\name{grass7_r_stats_quantile_out}
\alias{grass7_r_stats_quantile_out}
\title{QGIS algorithm r.stats.quantile.out}
\usage{
grass7_r_stats_quantile_out(
  base = qgisprocess::qgis_default_value(),
  cover = qgisprocess::qgis_default_value(),
  quantiles = qgisprocess::qgis_default_value(),
  percentiles = qgisprocess::qgis_default_value(),
  bins = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  file = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{base}{\code{raster} - Name of base raster map. Path to a raster layer.}

\item{cover}{\code{raster} - Name of cover raster map. Path to a raster layer.}

\item{quantiles}{\code{number} - Number of quantiles. A numeric value.}

\item{percentiles}{\code{string} - List of percentiles. String value.}

\item{bins}{\code{number} - Number of bins to use. A numeric value.}

\item{.r}{\code{boolean} - Create reclass map with statistics as category labels. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{file}{\code{fileDestination} - Statistics File. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.stats.quantile.out (grass7:r.stats.quantile.out)
}
\details{
\subsection{Outputs description}{
\itemize{
\item file - outputFile - Statistics File
}
}
}
