% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_sun_insoltime.R
\name{grass7_r_sun_insoltime}
\alias{grass7_r_sun_insoltime}
\title{QGIS algorithm r.sun.insoltime}
\usage{
grass7_r_sun_insoltime(
  elevation = qgisprocess::qgis_default_value(),
  aspect = qgisprocess::qgis_default_value(),
  aspect_value = qgisprocess::qgis_default_value(),
  slope = qgisprocess::qgis_default_value(),
  slope_value = qgisprocess::qgis_default_value(),
  linke = qgisprocess::qgis_default_value(),
  albedo = qgisprocess::qgis_default_value(),
  albedo_value = qgisprocess::qgis_default_value(),
  lat = qgisprocess::qgis_default_value(),
  long = qgisprocess::qgis_default_value(),
  coeff_bh = qgisprocess::qgis_default_value(),
  coeff_dh = qgisprocess::qgis_default_value(),
  horizon_basemap = qgisprocess::qgis_default_value(),
  horizon_step = qgisprocess::qgis_default_value(),
  day = qgisprocess::qgis_default_value(),
  step = qgisprocess::qgis_default_value(),
  declination = qgisprocess::qgis_default_value(),
  distance_step = qgisprocess::qgis_default_value(),
  npartitions = qgisprocess::qgis_default_value(),
  civil_time = qgisprocess::qgis_default_value(),
  .p = qgisprocess::qgis_default_value(),
  .m = qgisprocess::qgis_default_value(),
  insol_time = qgisprocess::qgis_default_value(),
  beam_rad = qgisprocess::qgis_default_value(),
  diff_rad = qgisprocess::qgis_default_value(),
  refl_rad = qgisprocess::qgis_default_value(),
  glob_rad = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{elevation}{\code{raster} - Elevation layer (meters). Path to a raster layer.}

\item{aspect}{\code{raster} - Aspect layer (decimal degrees). Path to a raster layer.}

\item{aspect_value}{\code{number} - A single value of the orientation (aspect), 270 is south. A numeric value.}

\item{slope}{\code{raster} - Name of the input slope raster map (terrain slope or solar panel inclination) (decimal degrees). Path to a raster layer.}

\item{slope_value}{\code{number} - A single value of inclination (slope). A numeric value.}

\item{linke}{\code{raster} - Name of the Linke atmospheric turbidity coefficient input raster map. Path to a raster layer.}

\item{albedo}{\code{raster} - Name of the ground albedo coefficient input raster map. Path to a raster layer.}

\item{albedo_value}{\code{number} - A single value of the ground albedo coefficient. A numeric value.}

\item{lat}{\code{raster} - Name of input raster map containing latitudes (decimal degrees). Path to a raster layer.}

\item{long}{\code{raster} - Name of input raster map containing longitudes (decimal degrees). Path to a raster layer.}

\item{coeff_bh}{\code{raster} - Name of real-sky beam radiation coefficient input raster map. Path to a raster layer.}

\item{coeff_dh}{\code{raster} - Name of real-sky diffuse radiation coefficient input raster map. Path to a raster layer.}

\item{horizon_basemap}{\code{raster} - The horizon information input map basename. Path to a raster layer.}

\item{horizon_step}{\code{number} - Angle step size for multidirectional horizon (degrees). A numeric value.}

\item{day}{\code{number} - No. of day of the year (1-365). A numeric value.}

\item{step}{\code{number} - Time step when computing all-day radiation sums (decimal hours). A numeric value.}

\item{declination}{\code{number} - Declination value (overriding the internally computed value) (radians). A numeric value.}

\item{distance_step}{\code{number} - Sampling distance step coefficient (0.5-1.5). A numeric value.}

\item{npartitions}{\code{number} - Read the input files in this number of chunks. A numeric value.}

\item{civil_time}{\code{number} - Civil time zone value, if none, the time will be local solar time. A numeric value.}

\item{.p}{\code{boolean} - Do not incorporate the shadowing effect of terrain. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.m}{\code{boolean} - Use the low-memory version of the program. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{insol_time}{\code{rasterDestination} - Insolation time (h). Path for new raster layer.}

\item{beam_rad}{\code{rasterDestination} - Irradiation raster map (Wh.m-2.day-1). Path for new raster layer.}

\item{diff_rad}{\code{rasterDestination} - Irradiation raster map (Wh.m-2.day-1). Path for new raster layer.}

\item{refl_rad}{\code{rasterDestination} - Irradiation raster map (Wh.m-2.day-1). Path for new raster layer.}

\item{glob_rad}{\code{rasterDestination} - Irradiance/irradiation raster map (Wh.m-2.day-1). Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.sun.insoltime (grass7:r.sun.insoltime)
}
\details{
\subsection{Outputs description}{
\itemize{
\item insol_time - outputRaster - Insolation time
\item beam_rad - outputRaster - Irradiation raster map
\item diff_rad - outputRaster - Irradiation raster map
\item refl_rad - outputRaster - Irradiation raster map
\item glob_rad - outputRaster - Irradiance
}
}
}
