% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_sunhours.R
\name{grass7_r_sunhours}
\alias{grass7_r_sunhours}
\title{QGIS algorithm r.sunhours}
\usage{
grass7_r_sunhours(
  year = qgisprocess::qgis_default_value(),
  month = qgisprocess::qgis_default_value(),
  day = qgisprocess::qgis_default_value(),
  hour = qgisprocess::qgis_default_value(),
  minute = qgisprocess::qgis_default_value(),
  second = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  elevation = qgisprocess::qgis_default_value(),
  azimuth = qgisprocess::qgis_default_value(),
  sunhour = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{year}{\code{number} - Year. A numeric value.}

\item{month}{\code{number} - Month. A numeric value.}

\item{day}{\code{number} - Day. A numeric value.}

\item{hour}{\code{number} - Hour. A numeric value.}

\item{minute}{\code{number} - Minutes. A numeric value.}

\item{second}{\code{number} - Seconds. A numeric value.}

\item{.t}{\code{boolean} - Time is local sidereal time, not Greenwich standard time. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{.s}{\code{boolean} - Do not use SOLPOS algorithm of NREL. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -s.}

\item{elevation}{\code{rasterDestination} - Solar Elevation Angle. Path for new raster layer.}

\item{azimuth}{\code{rasterDestination} - Solar Azimuth Angle. Path for new raster layer.}

\item{sunhour}{\code{rasterDestination} - Sunshine Hours. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.sunhours (grass7:r.sunhours)
}
\details{
\subsection{Outputs description}{
\itemize{
\item azimuth - outputRaster - Solar Azimuth Angle
\item elevation - outputRaster - Solar Elevation Angle
\item sunhour - outputRaster - Sunshine Hours
}
}
}
