% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_sunmask_datetime.R
\name{grass7_r_sunmask_datetime}
\alias{grass7_r_sunmask_datetime}
\title{QGIS algorithm r.sunmask.datetime}
\usage{
grass7_r_sunmask_datetime(
  elevation = qgisprocess::qgis_default_value(),
  year = qgisprocess::qgis_default_value(),
  month = qgisprocess::qgis_default_value(),
  day = qgisprocess::qgis_default_value(),
  hour = qgisprocess::qgis_default_value(),
  minute = qgisprocess::qgis_default_value(),
  second = qgisprocess::qgis_default_value(),
  timezone = qgisprocess::qgis_default_value(),
  east = qgisprocess::qgis_default_value(),
  north = qgisprocess::qgis_default_value(),
  .z = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{elevation}{\code{raster} - Elevation raster layer (meters). Path to a raster layer.}

\item{year}{\code{number} - year. A numeric value.}

\item{month}{\code{number} - month. A numeric value.}

\item{day}{\code{number} - day. A numeric value.}

\item{hour}{\code{number} - hour. A numeric value.}

\item{minute}{\code{number} - minute. A numeric value.}

\item{second}{\code{number} - second. A numeric value.}

\item{timezone}{\code{number} - East positive, offset from GMT. A numeric value.}

\item{east}{\code{string} - Easting coordinate (point of interest). String value.}

\item{north}{\code{string} - Northing coordinate (point of interest). String value.}

\item{.z}{\code{boolean} - Do not ignore zero elevation. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -z.}

\item{.s}{\code{boolean} - Calculate sun position only and exit. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -s.}

\item{output}{\code{rasterDestination} - Shadows. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.sunmask.datetime (grass7:r.sunmask.datetime)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Shadows
}
}
}
