% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_sunmask_position.R
\name{grass7_r_sunmask_position}
\alias{grass7_r_sunmask_position}
\title{QGIS algorithm r.sunmask.position}
\usage{
grass7_r_sunmask_position(
  elevation = qgisprocess::qgis_default_value(),
  altitude = qgisprocess::qgis_default_value(),
  azimuth = qgisprocess::qgis_default_value(),
  east = qgisprocess::qgis_default_value(),
  north = qgisprocess::qgis_default_value(),
  .z = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{elevation}{\code{raster} - Elevation raster layer (meters). Path to a raster layer.}

\item{altitude}{\code{number} - Altitude of the sun in degrees above the horizon. A numeric value.}

\item{azimuth}{\code{number} - Azimuth of the sun in degrees from north. A numeric value.}

\item{east}{\code{string} - Easting coordinate (point of interest). String value.}

\item{north}{\code{string} - Northing coordinate (point of interest). String value.}

\item{.z}{\code{boolean} - Do not ignore zero elevation. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -z.}

\item{.s}{\code{boolean} - Calculate sun position only and exit. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -s.}

\item{output}{\code{rasterDestination} - Shadows. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.sunmask.position (grass7:r.sunmask.position)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Shadows
}
}
}
