% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_terraflow.R
\name{grass7_r_terraflow}
\alias{grass7_r_terraflow}
\title{QGIS algorithm r.terraflow}
\usage{
grass7_r_terraflow(
  elevation = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  d8cut = qgisprocess::qgis_default_value(),
  memory = qgisprocess::qgis_default_value(),
  filled = qgisprocess::qgis_default_value(),
  direction = qgisprocess::qgis_default_value(),
  swatershed = qgisprocess::qgis_default_value(),
  accumulation = qgisprocess::qgis_default_value(),
  tci = qgisprocess::qgis_default_value(),
  stats = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{elevation}{\code{raster} - Name of elevation raster map. Path to a raster layer.}

\item{.s}{\code{boolean} - SFD (D8) flow (default is MFD). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -s.}

\item{d8cut}{\code{number} - Routing using SFD (D8) direction. A numeric value.}

\item{memory}{\code{number} - Maximum memory to be used (in MB). A numeric value.}

\item{filled}{\code{rasterDestination} - Filled (flooded) elevation. Path for new raster layer.}

\item{direction}{\code{rasterDestination} - Flow direction. Path for new raster layer.}

\item{swatershed}{\code{rasterDestination} - Sink-watershed. Path for new raster layer.}

\item{accumulation}{\code{rasterDestination} - Flow accumulation. Path for new raster layer.}

\item{tci}{\code{rasterDestination} - Topographic convergence index (tci). Path for new raster layer.}

\item{stats}{\code{fileDestination} - Runtime statistics. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.terraflow (grass7:r.terraflow)
}
\details{
\subsection{Outputs description}{
\itemize{
\item accumulation - outputRaster - Flow accumulation
\item direction - outputRaster - Flow direction
\item filled - outputRaster - Filled (flooded) elevation
\item stats - outputFile - Runtime statistics
\item swatershed - outputRaster - Sink-watershed
\item tci - outputRaster - Topographic convergence index (tci)
}
}
}
