% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_tile.R
\name{grass7_r_tile}
\alias{grass7_r_tile}
\title{QGIS algorithm r.tile}
\usage{
grass7_r_tile(
  input = qgisprocess::qgis_default_value(),
  width = qgisprocess::qgis_default_value(),
  height = qgisprocess::qgis_default_value(),
  overlap = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Name of input raster map. Path to a raster layer.}

\item{width}{\code{number} - Width of tiles (columns). A numeric value.}

\item{height}{\code{number} - Height of tiles (rows). A numeric value.}

\item{overlap}{\code{number} - Overlap of tiles. A numeric value.}

\item{output}{\code{folderDestination} - Tiles Directory. Path for an existing or new folder.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.tile (grass7:r.tile)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFolder - Tiles Directory
}
}
}
