% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_tileset.R
\name{grass7_r_tileset}
\alias{grass7_r_tileset}
\title{QGIS algorithm r.tileset}
\usage{
grass7_r_tileset(
  sourceproj = qgisprocess::qgis_default_value(),
  sourcescale = qgisprocess::qgis_default_value(),
  destproj = qgisprocess::qgis_default_value(),
  destscale = qgisprocess::qgis_default_value(),
  maxcols = qgisprocess::qgis_default_value(),
  maxrows = qgisprocess::qgis_default_value(),
  overlap = qgisprocess::qgis_default_value(),
  separator = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  .w = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{sourceproj}{\code{crs} - Source projection. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{sourcescale}{\code{number} - Conversion factor from units to meters in source projection. A numeric value.}

\item{destproj}{\code{crs} - Destination projection. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{destscale}{\code{number} - Conversion factor from units to meters in destination projection. A numeric value.}

\item{maxcols}{\code{number} - Maximum number of columns for a tile in the source projection. A numeric value.}

\item{maxrows}{\code{number} - Maximum number of rows for a tile in the source projection. A numeric value.}

\item{overlap}{\code{number} - Number of cells tiles should overlap in each direction. A numeric value.}

\item{separator}{\code{string} - Output field separator. String value.}

\item{.g}{\code{boolean} - Produces shell script output. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.w}{\code{boolean} - Produces web map server query string output. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{html}{\code{fileDestination} - Tileset. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.tileset (grass7:r.tileset)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Ti
}
}
}
