% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_topmodel.R
\name{grass7_r_topmodel}
\alias{grass7_r_topmodel}
\title{QGIS algorithm r.topmodel}
\usage{
grass7_r_topmodel(
  parameters = qgisprocess::qgis_default_value(),
  topidxstats = qgisprocess::qgis_default_value(),
  input = qgisprocess::qgis_default_value(),
  timestep = qgisprocess::qgis_default_value(),
  topidxclass = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{parameters}{\code{file} - Name of TOPMODEL parameters file. Path to a file.}

\item{topidxstats}{\code{file} - Name of topographic index statistics file. Path to a file.}

\item{input}{\code{file} - Name of rainfall and potential evapotranspiration data file. Path to a file.}

\item{timestep}{\code{number} - Time step. Generate output for this time step. A numeric value.}

\item{topidxclass}{\code{number} - Topographic index class. Generate output for this topographic index class. A numeric value.}

\item{output}{\code{fileDestination} - TOPMODEL output. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.topmodel (grass7:r.topmodel)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - TOPMODEL output
}
}
}
