% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_topmodel_topidxstats.R
\name{grass7_r_topmodel_topidxstats}
\alias{grass7_r_topmodel_topidxstats}
\title{QGIS algorithm r.topmodel.topidxstats}
\usage{
grass7_r_topmodel_topidxstats(
  topidx = qgisprocess::qgis_default_value(),
  ntopidxclasses = qgisprocess::qgis_default_value(),
  outtopidxstats = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{topidx}{\code{raster} - Name of input topographic index raster map. Path to a raster layer.}

\item{ntopidxclasses}{\code{number} - Number of topographic index classes. A numeric value.}

\item{outtopidxstats}{\code{fileDestination} - TOPMODEL topographic index statistics file. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.topmodel.topidxstats (grass7:r.topmodel.topidxstats)
}
\details{
\subsection{Outputs description}{
\itemize{
\item outtopidxstats - outputFile - TOPMODEL topographic index statistics file
}
}
}
