% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_transect.R
\name{grass7_r_transect}
\alias{grass7_r_transect}
\title{QGIS algorithm r.transect}
\usage{
grass7_r_transect(
  map = qgisprocess::qgis_default_value(),
  line = qgisprocess::qgis_default_value(),
  null_value = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{map}{\code{raster} - Raster map to be queried. Path to a raster layer.}

\item{line}{\code{string} - Transect definition: east,north,azimuth,distance(,east,north,azimuth,distance,...). String value.}

\item{null_value}{\code{string} - String representing NULL value. String value.}

\item{.g}{\code{boolean} - Output easting and northing in first two columns of four column output. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{html}{\code{fileDestination} - Transect file. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.transect (grass7:r.transect)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Transect file
}
}
}
