% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_uslek.R
\name{grass7_r_uslek}
\alias{grass7_r_uslek}
\title{QGIS algorithm r.uslek}
\usage{
grass7_r_uslek(
  psand = qgisprocess::qgis_default_value(),
  pclay = qgisprocess::qgis_default_value(),
  psilt = qgisprocess::qgis_default_value(),
  pomat = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{psand}{\code{raster} - Name of soil sand fraction raster map (0.0-1.0). Path to a raster layer.}

\item{pclay}{\code{raster} - Name of soil clay fraction raster map (0.0-1.0). Path to a raster layer.}

\item{psilt}{\code{raster} - Name of soil silt fraction raster map (0.0-1.0). Path to a raster layer.}

\item{pomat}{\code{raster} - Name of soil organic matter raster map (0.0-1.0). Path to a raster layer.}

\item{output}{\code{rasterDestination} - USLE R Raster. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.uslek (grass7:r.uslek)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - USLE R Raster
}
}
}
