% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_what_color.R
\name{grass7_r_what_color}
\alias{grass7_r_what_color}
\title{QGIS algorithm r.what.color}
\usage{
grass7_r_what_color(
  input = qgisprocess::qgis_default_value(),
  value = qgisprocess::qgis_default_value(),
  format = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{raster} - Raster map to query colors. Path to a raster layer.}

\item{value}{\code{string} - Values to query colors for (comma separated list). String value.}

\item{format}{\code{string} - Output format (printf-style). String value.}

\item{html}{\code{fileDestination} - Colors file. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.what.color (grass7:r.what.color)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Colors file
}
}
}
