% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_what_coords.R
\name{grass7_r_what_coords}
\alias{grass7_r_what_coords}
\title{QGIS algorithm r.what.coords}
\usage{
grass7_r_what_coords(
  map = qgisprocess::qgis_default_value(),
  coordinates = qgisprocess::qgis_default_value(),
  null_value = qgisprocess::qgis_default_value(),
  separator = qgisprocess::qgis_default_value(),
  cache = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  .f = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  .i = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{map}{\code{raster} - Name of raster map. Path to a raster layer.}

\item{coordinates}{\code{point} - Coordinates for query (east, north). Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{null_value}{\code{string} - String representing NULL value. String value.}

\item{separator}{\code{string} - Field separator. Special characters: pipe, comma, space, tab, newlineString representing NULL value. String value.}

\item{cache}{\code{number} - Size of point cache. A numeric value.}

\item{.n}{\code{boolean} - Output header row. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.}

\item{.f}{\code{boolean} - Show the category labels of the grid cell(s). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -f.}

\item{.r}{\code{boolean} - Output color values as RRR:GGG:BBB. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{.i}{\code{boolean} - Output integer category values, not cell values. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -i.}

\item{.c}{\code{boolean} - Turn on cache reporting. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{output}{\code{fileDestination} - Raster Value File. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.what.coords (grass7:r.what.coords)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - Raster Value File
}
}
}
