% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_class.R
\name{grass7_v_class}
\alias{grass7_v_class}
\title{QGIS algorithm v.class}
\usage{
grass7_v_class(
  map = qgisprocess::qgis_default_value(),
  column = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  algorithm = qgisprocess::qgis_default_value(),
  nbclasses = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{map}{\code{source} - Input vector layer. Path to a vector layer.}

\item{column}{\code{field} - Column name or expression. The name of an existing field. ; delimited list of existing field names.}

\item{where}{\code{string} - WHERE conditions of SQL statement without 'where' keyword. String value.}

\item{algorithm}{\code{enum}  of \verb{("int", "std", "qua", "equ")} - Algorithm to use for classification. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{nbclasses}{\code{number} - Number of classes to define. A numeric value.}

\item{.g}{\code{boolean} - Print only class breaks (without min and max). 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{html}{\code{fileDestination} - Classification. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.class (grass7:v.class)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Classification
}
}
}
