% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_cluster.R
\name{grass7_v_cluster}
\alias{grass7_v_cluster}
\title{QGIS algorithm v.cluster}
\usage{
grass7_v_cluster(
  input = qgisprocess::qgis_default_value(),
  distance = qgisprocess::qgis_default_value(),
  min = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  .arg_2 = qgisprocess::qgis_default_value(),
  .b = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{source} - Input layer. Path to a vector layer.}

\item{distance}{\code{number} - Maximum distance to neighbors. A numeric value.}

\item{min}{\code{number} - Minimum number of points to create a cluster. A numeric value.}

\item{method}{\code{enum}  of \verb{("dbscan", "dbscan2", "density", "optics", "optics2")} - Clustering method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.arg_2}{\code{boolean} - Force 2D clustering. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -2.}

\item{.b}{\code{boolean} - Do not build topology. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -b.}

\item{.t}{\code{boolean} - Do not create attribute table. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{output}{\code{vectorDestination} - Clustered. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.cluster (grass7:v.cluster)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Clustered
}
}
}
