% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_db_select.R
\name{grass7_v_db_select}
\alias{grass7_v_db_select}
\title{QGIS algorithm v.db.select}
\usage{
grass7_v_db_select(
  map = qgisprocess::qgis_default_value(),
  layer = qgisprocess::qgis_default_value(),
  columns = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  separator = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  group = qgisprocess::qgis_default_value(),
  vertical_separator = qgisprocess::qgis_default_value(),
  null_value = qgisprocess::qgis_default_value(),
  .v = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  file = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{map}{\code{source} - Input vector map. Path to a vector layer.}

\item{layer}{\code{number} - Layer Number. A numeric value.}

\item{columns}{\code{string} - Name of attribute column(s), comma separated. String value.}

\item{.c}{\code{boolean} - Do not include column names in output. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{separator}{\code{string} - Output field separator. String value.}

\item{where}{\code{string} - WHERE conditions of SQL statement without 'where' keyword. String value.}

\item{group}{\code{string} - GROUP BY conditions of SQL statement without 'group by' keyword. String value.}

\item{vertical_separator}{\code{string} - Output vertical record separator. String value.}

\item{null_value}{\code{string} - Null value indicator. String value.}

\item{.v}{\code{boolean} - Vertical output (instead of horizontal). 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.r}{\code{boolean} - Print minimal region extent of selected vector features instead of attributes. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{file}{\code{fileDestination} - Attributes. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.db.select (grass7:v.db.select)
}
\details{
\subsection{Outputs description}{
\itemize{
\item file - outputFile - Attributes
}
}
}
