% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_decimate.R
\name{grass7_v_decimate}
\alias{grass7_v_decimate}
\title{QGIS algorithm v.decimate}
\usage{
grass7_v_decimate(
  input = qgisprocess:::qgis_default_value(),
  zrange = qgisprocess:::qgis_default_value(),
  cats = qgisprocess:::qgis_default_value(),
  skip = qgisprocess:::qgis_default_value(),
  preserve = qgisprocess:::qgis_default_value(),
  offset = qgisprocess:::qgis_default_value(),
  limit = qgisprocess:::qgis_default_value(),
  zdiff = qgisprocess:::qgis_default_value(),
  cell_limit = qgisprocess:::qgis_default_value(),
  .g = qgisprocess:::qgis_default_value(),
  .f = qgisprocess:::qgis_default_value(),
  .c = qgisprocess:::qgis_default_value(),
  .z = qgisprocess:::qgis_default_value(),
  .x = qgisprocess:::qgis_default_value(),
  .b = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{source} - Input vector. Path to a vector layer.}

\item{zrange}{\code{range} - Filter range for z data (min,max). Two comma separated numeric values, e.g. '1,10'.}

\item{cats}{\code{string} - Category values. String value.}

\item{skip}{\code{number} - Throw away every n-th point. A numeric value.}

\item{preserve}{\code{number} - Preserve only every n-th point. A numeric value.}

\item{offset}{\code{number} - Skip first n points. A numeric value.}

\item{limit}{\code{number} - Copy only n points. A numeric value.}

\item{zdiff}{\code{number} - Minimal difference of z values. A numeric value.}

\item{cell_limit}{\code{number} - Preserve only n points per grid cell. A numeric value.}

\item{.g}{\code{boolean} - Apply grid-based decimation. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -g.}

\item{.f}{\code{boolean} - Use only first point in grid cell during grid-based decimation. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -f.}

\item{.c}{\code{boolean} - Only one point per cat in grid cell. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{.z}{\code{boolean} - Use z in grid decimation. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -z.}

\item{.x}{\code{boolean} - Store only the coordinates, throw away categories. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -x.}

\item{.b}{\code{boolean} - Do not build topology. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -b.}

\item{output}{\code{vectorDestination} - Output vector map. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.decimate (grass7:v.decimate)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Output vector map
}
}
}
