% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_extrude.R
\name{grass7_v_extrude}
\alias{grass7_v_extrude}
\title{QGIS algorithm v.extrude}
\usage{
grass7_v_extrude(
  input = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  type = qgisprocess::qgis_default_value(),
  zshift = qgisprocess::qgis_default_value(),
  height = qgisprocess::qgis_default_value(),
  height_column = qgisprocess::qgis_default_value(),
  elevation = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  scale = qgisprocess::qgis_default_value(),
  null_value = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{source} - Name of input 2D vector map. Path to a vector layer.}

\item{where}{\code{string} - WHERE conditions of SQL statement without 'where' keyword. String value.}

\item{type}{\code{enum}  of \verb{("point", "line", "area")} - Input feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{zshift}{\code{number} - Shifting value for z coordinates. A numeric value.}

\item{height}{\code{number} - Fixed height for 3D vector objects. A numeric value.}

\item{height_column}{\code{field} - Name of attribute column with object heights. The name of an existing field. ; delimited list of existing field names.}

\item{elevation}{\code{raster} - Elevation raster for height extraction. Path to a raster layer.}

\item{method}{\code{enum}  of \verb{("nearest", "bilinear", "bicubic")} - Sampling interpolation method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{scale}{\code{number} - Scale factor sampled raster values. A numeric value.}

\item{null_value}{\code{number} - Height for sampled raster NULL values. A numeric value.}

\item{.t}{\code{boolean} - Trace elevation. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{output}{\code{vectorDestination} - 3D Vector. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.extrude (grass7:v.extrude)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - 3D Vector
}
}
}
