% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_in_ascii.R
\name{grass7_v_in_ascii}
\alias{grass7_v_in_ascii}
\title{QGIS algorithm v.in.ascii}
\usage{
grass7_v_in_ascii(
  input = qgisprocess::qgis_default_value(),
  format = qgisprocess::qgis_default_value(),
  separator = qgisprocess::qgis_default_value(),
  text = qgisprocess::qgis_default_value(),
  skip = qgisprocess::qgis_default_value(),
  columns = qgisprocess::qgis_default_value(),
  x = qgisprocess::qgis_default_value(),
  y = qgisprocess::qgis_default_value(),
  z = qgisprocess::qgis_default_value(),
  cat = qgisprocess::qgis_default_value(),
  .z = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  .b = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  .i = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{file} - ASCII file to be imported. Path to a file.}

\item{format}{\code{enum}  of \verb{("point", "standard")} - Input file format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{separator}{\code{string} - Field separator. String value.}

\item{text}{\code{string} - Text delimiter. String value.}

\item{skip}{\code{number} - Number of header lines to skip at top of input file. A numeric value.}

\item{columns}{\code{string} - Column definition in SQL style (example: 'x double precision, y double precision, cat int, name varchar(10)'). String value.}

\item{x}{\code{number} - Number of column used as x coordinate. A numeric value.}

\item{y}{\code{number} - Number of column used as y coordinate. A numeric value.}

\item{z}{\code{number} - Number of column used as z coordinate. A numeric value.}

\item{cat}{\code{number} - Number of column used as category. A numeric value.}

\item{.z}{\code{boolean} - Create 3D vector map. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -z.}

\item{.n}{\code{boolean} - Do not expect a header when reading in standard format. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.}

\item{.t}{\code{boolean} - Do not create table in points mode. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{.b}{\code{boolean} - Do not build topology in points mode. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -b.}

\item{.r}{\code{boolean} - Only import points falling within current region (points mode). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{.i}{\code{boolean} - Ignore broken line(s) in points mode. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -i.}

\item{output}{\code{vectorDestination} - ASCII. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.in.ascii (grass7:v.in.ascii)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - ASCII
}
}
}
