% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_in_lidar.R
\name{grass7_v_in_lidar}
\alias{grass7_v_in_lidar}
\title{QGIS algorithm v.in.lidar}
\usage{
grass7_v_in_lidar(
  input = qgisprocess::qgis_default_value(),
  spatial = qgisprocess::qgis_default_value(),
  zrange = qgisprocess::qgis_default_value(),
  return_filter = qgisprocess::qgis_default_value(),
  class_filter = qgisprocess::qgis_default_value(),
  skip = qgisprocess::qgis_default_value(),
  preserve = qgisprocess::qgis_default_value(),
  offset = qgisprocess::qgis_default_value(),
  limit = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  .b = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{file} - LiDAR input files in LAS format (*.las or *.laz). Path to a file.}

\item{spatial}{\code{extent} - Import subregion only. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{zrange}{\code{range} - Filter range for z data. Two comma separated numeric values, e.g. '1,10'.}

\item{return_filter}{\code{enum}  of \verb{("first", "last", "mid")} - Only import points of selected return type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{class_filter}{\code{string} - Only import points of selected class(es) (comma separated integers). String value.}

\item{skip}{\code{number} - Do not import every n-th point. A numeric value.}

\item{preserve}{\code{number} - Import only every n-th point. A numeric value.}

\item{offset}{\code{number} - Skip first n points. A numeric value.}

\item{limit}{\code{number} - Import only n points. A numeric value.}

\item{.t}{\code{boolean} - Do not create attribute table. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.c}{\code{boolean} - Do not automatically add unique ID as category to each point. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.b}{\code{boolean} - Do not build topology. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{vectorDestination} - Lidar. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.in.lidar (grass7:v.in.lidar)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Lidar
}
}
}
