% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_info.R
\name{grass7_v_info}
\alias{grass7_v_info}
\title{QGIS algorithm v.info}
\usage{
grass7_v_info(
  map = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  .e = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{map}{\code{source} - Name of input vector map. Path to a vector layer.}

\item{.c}{\code{boolean} - Print types/names of table columns for specified layer instead of info. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{.g}{\code{boolean} - Print map region only. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -g.}

\item{.e}{\code{boolean} - Print extended metadata info in shell script style. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -e.}

\item{.t}{\code{boolean} - Print topology information only. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{html}{\code{fileDestination} - Information report. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.info (grass7:v.info)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Information report
}
}
}
