% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_kernel_rast.R
\name{grass7_v_kernel_rast}
\alias{grass7_v_kernel_rast}
\title{QGIS algorithm v.kernel.rast}
\usage{
grass7_v_kernel_rast(
  input = qgisprocess::qgis_default_value(),
  radius = qgisprocess::qgis_default_value(),
  dsize = qgisprocess::qgis_default_value(),
  segmax = qgisprocess::qgis_default_value(),
  distmax = qgisprocess::qgis_default_value(),
  multiplier = qgisprocess::qgis_default_value(),
  node = qgisprocess::qgis_default_value(),
  kernel = qgisprocess::qgis_default_value(),
  .o = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{source} - Name of input vector map with training points. Path to a vector layer.}

\item{radius}{\code{number} - Kernel radius in map units. A numeric value.}

\item{dsize}{\code{number} - Discretization error in map units. A numeric value.}

\item{segmax}{\code{number} - Maximum length of segment on network. A numeric value.}

\item{distmax}{\code{number} - Maximum distance from point to network. A numeric value.}

\item{multiplier}{\code{number} - Multiply the density result by this number. A numeric value.}

\item{node}{\code{enum}  of \verb{("none", "split")} - Node method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{kernel}{\code{enum}  of \verb{("uniform", "triangular", "epanechnikov", "quartic", "triweight", "gaussian", "cosine")} - Kernel function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.o}{\code{boolean} - Try to calculate an optimal radius with given 'radius' taken as maximum (experimental). 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{rasterDestination} - Kernel. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.kernel.rast (grass7:v.kernel.rast)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Kernel
}
}
}
