% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_lidar_correction.R
\name{grass7_v_lidar_correction}
\alias{grass7_v_lidar_correction}
\title{QGIS algorithm v.lidar.correction}
\usage{
grass7_v_lidar_correction(
  input = qgisprocess::qgis_default_value(),
  ew_step = qgisprocess::qgis_default_value(),
  ns_step = qgisprocess::qgis_default_value(),
  lambda_c = qgisprocess::qgis_default_value(),
  tch = qgisprocess::qgis_default_value(),
  tcl = qgisprocess::qgis_default_value(),
  .e = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  terrain = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{source} - Input vector layer (v.lidar.growing output). Path to a vector layer.}

\item{ew_step}{\code{number} - Length of each spline step in the east-west direction. A numeric value.}

\item{ns_step}{\code{number} - Length of each spline step in the north-south direction. A numeric value.}

\item{lambda_c}{\code{number} - Regularization weight in reclassification evaluation. A numeric value.}

\item{tch}{\code{number} - High threshold for object to terrain reclassification. A numeric value.}

\item{tcl}{\code{number} - Low threshold for terrain to object reclassification. A numeric value.}

\item{.e}{\code{boolean} - Estimate point density and distance. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -e.}

\item{output}{\code{vectorDestination} - Classified. Path for new vector layer.}

\item{terrain}{\code{vectorDestination} - Only 'terrain' points. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.lidar.correction (grass7:v.lidar.correction)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Classified
\item terrain - outputVector - Only 'terrain' points
}
}
}
