% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_lidar_edgedetection.R
\name{grass7_v_lidar_edgedetection}
\alias{grass7_v_lidar_edgedetection}
\title{QGIS algorithm v.lidar.edgedetection}
\usage{
grass7_v_lidar_edgedetection(
  input = qgisprocess::qgis_default_value(),
  ew_step = qgisprocess::qgis_default_value(),
  ns_step = qgisprocess::qgis_default_value(),
  lambda_g = qgisprocess::qgis_default_value(),
  tgh = qgisprocess::qgis_default_value(),
  tgl = qgisprocess::qgis_default_value(),
  theta_g = qgisprocess::qgis_default_value(),
  lambda_r = qgisprocess::qgis_default_value(),
  .e = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{source} - Input vector layer. Path to a vector layer.}

\item{ew_step}{\code{number} - Length of each spline step in the east-west direction. A numeric value.}

\item{ns_step}{\code{number} - Length of each spline step in the north-south direction. A numeric value.}

\item{lambda_g}{\code{number} - Regularization weight in gradient evaluation. A numeric value.}

\item{tgh}{\code{number} - High gradient threshold for edge classification. A numeric value.}

\item{tgl}{\code{number} - Low gradient threshold for edge classification. A numeric value.}

\item{theta_g}{\code{number} - Angle range for same direction detection. A numeric value.}

\item{lambda_r}{\code{number} - Regularization weight in residual evaluation. A numeric value.}

\item{.e}{\code{boolean} - Estimate point density and distance. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{vectorDestination} - Edges. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.lidar.edgedetection (grass7:v.lidar.edgedetection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Edges
}
}
}
