% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_net_allpairs.R
\name{grass7_v_net_allpairs}
\alias{grass7_v_net_allpairs}
\title{QGIS algorithm v.net.allpairs}
\usage{
grass7_v_net_allpairs(
  input = qgisprocess::qgis_default_value(),
  points = qgisprocess::qgis_default_value(),
  threshold = qgisprocess::qgis_default_value(),
  cats = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  arc_column = qgisprocess::qgis_default_value(),
  arc_backward_column = qgisprocess::qgis_default_value(),
  node_column = qgisprocess::qgis_default_value(),
  .g = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{source} - Input vector line layer (arcs). Path to a vector layer.}

\item{points}{\code{source} - Centers point layer (nodes). Path to a vector layer.}

\item{threshold}{\code{number} - Threshold for connecting centers to the network (in map unit). A numeric value.}

\item{cats}{\code{string} - Category values. String value.}

\item{where}{\code{string} - WHERE condition of SQL statement without 'where' keyword'. String value.}

\item{arc_column}{\code{field} - Arc forward/both direction(s) cost column (number). The name of an existing field. ; delimited list of existing field names.}

\item{arc_backward_column}{\code{field} - Arc backward direction cost column (number). The name of an existing field. ; delimited list of existing field names.}

\item{node_column}{\code{field} - Node cost column (number). The name of an existing field. ; delimited list of existing field names.}

\item{.g}{\code{boolean} - Use geodesic calculation for longitude-latitude locations. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{vectorDestination} - Network_Allpairs. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.net.allpairs (grass7:v.net.allpairs)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Network_Allpairs
}
}
}
