% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_normal.R
\name{grass7_v_normal}
\alias{grass7_v_normal}
\title{QGIS algorithm v.normal}
\usage{
grass7_v_normal(
  map = qgisprocess::qgis_default_value(),
  tests = qgisprocess::qgis_default_value(),
  column = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  .l = qgisprocess::qgis_default_value(),
  html = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{map}{\code{source} - point vector defining sample points. Path to a vector layer.}

\item{tests}{\code{string} - Lists of tests (1-15): e.g. 1,3-8,13. String value.}

\item{column}{\code{field} - Attribute column. The name of an existing field. ; delimited list of existing field names.}

\item{.r}{\code{boolean} - Use only points in current region. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{.l}{\code{boolean} - lognormal. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -l.}

\item{html}{\code{fileDestination} - Normality. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.normal (grass7:v.normal)
}
\details{
\subsection{Outputs description}{
\itemize{
\item html - outputHtml - Normality
}
}
}
