% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_out_ascii.R
\name{grass7_v_out_ascii}
\alias{grass7_v_out_ascii}
\title{QGIS algorithm v.out.ascii}
\usage{
grass7_v_out_ascii(
  input = qgisprocess::qgis_default_value(),
  type = qgisprocess::qgis_default_value(),
  columns = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  format = qgisprocess::qgis_default_value(),
  separator = qgisprocess::qgis_default_value(),
  precision = qgisprocess::qgis_default_value(),
  .o = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{source} - Name of input vector map. Path to a vector layer.}

\item{type}{\code{enum}  of \verb{("point", "line", "boundary", "centroid", "area", "face", "kernel")} - Input feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{columns}{\code{field} - Name of attribute column(s) to be exported. The name of an existing field. ; delimited list of existing field names.}

\item{where}{\code{string} - WHERE conditions of SQL statement without 'where' keyword. String value.}

\item{format}{\code{enum}  of \verb{("point", "standard", "wkt")} - Output format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{separator}{\code{enum}  of \verb{("pipe", "comma", "space", "tab", "newline")} - Field separator. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{precision}{\code{number} - Number of significant digits (floating point only). A numeric value.}

\item{.o}{\code{boolean} - Create old (version 4) ASCII file. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -o.}

\item{.c}{\code{boolean} - Include column names in output (points mode). 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{output}{\code{fileDestination} - Name for output ASCII file or ASCII vector name if '-o' is defined. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.out.ascii (grass7:v.out.ascii)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - Name for output ASCII file or ASCII vector name if
}
}
}
