% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_out_pov.R
\name{grass7_v_out_pov}
\alias{grass7_v_out_pov}
\title{QGIS algorithm v.out.pov}
\usage{
grass7_v_out_pov(
  input = qgisprocess:::qgis_default_value(),
  type = qgisprocess:::qgis_default_value(),
  size = qgisprocess:::qgis_default_value(),
  zmod = qgisprocess:::qgis_default_value(),
  objmod = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{source} - Name of input vector map. Path to a vector layer.}

\item{type}{\code{enum}  of \verb{("point", "line", "boundary", "centroid", "area", "face", "kernel")} - Input feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{size}{\code{number} - Radius of sphere for points and tube for lines. A numeric value.}

\item{zmod}{\code{string} - Modifier for z coordinates, this string is appended to each z coordinate. String value.}

\item{objmod}{\code{string} - Object modifier (OBJECT_MODIFIER in POV-Ray documentation). String value.}

\item{output}{\code{fileDestination} - POV vector. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.out.pov (grass7:v.out.pov)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - POV vector
}
}
}
