% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_out_vtk.R
\name{grass7_v_out_vtk}
\alias{grass7_v_out_vtk}
\title{QGIS algorithm v.out.vtk}
\usage{
grass7_v_out_vtk(
  input = qgisprocess::qgis_default_value(),
  type = qgisprocess::qgis_default_value(),
  precision = qgisprocess::qgis_default_value(),
  zscale = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{source} - Name of input vector map. Path to a vector layer.}

\item{type}{\code{enum}  of \verb{("point", "kernel", "centroid", "line", "boundary", "area", "face")} - Input feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{precision}{\code{number} - Number of significant digits (floating point only). A numeric value.}

\item{zscale}{\code{number} - Scale factor for elevation. A numeric value.}

\item{.c}{\code{boolean} - Correct the coordinates to fit the VTK-OpenGL precision. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{.n}{\code{boolean} - Export numeric attribute table fields as VTK scalar variables. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -n.}

\item{output}{\code{fileDestination} - VTK File. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.out.vtk (grass7:v.out.vtk)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - VTK File
}
}
}
