% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_overlay.R
\name{grass7_v_overlay}
\alias{grass7_v_overlay}
\title{QGIS algorithm v.overlay}
\usage{
grass7_v_overlay(
  ainput = qgisprocess::qgis_default_value(),
  atype = qgisprocess::qgis_default_value(),
  binput = qgisprocess::qgis_default_value(),
  btype = qgisprocess::qgis_default_value(),
  operator = qgisprocess::qgis_default_value(),
  snap = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ainput}{\code{source} - Input layer (A). Path to a vector layer.}

\item{atype}{\code{enum}  of \verb{("area", "line", "auto")} - Input layer (A) Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{binput}{\code{source} - Input layer (B). Path to a vector layer.}

\item{btype}{\code{enum}  of \code{("area")} - Input layer (B) Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{operator}{\code{enum}  of \verb{("and", "or", "not", "xor")} - Operator to use. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{snap}{\code{number} - Snapping threshold for boundaries. A numeric value.}

\item{.t}{\code{boolean} - Do not create attribute table. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{vectorDestination} - Overlay. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.overlay (grass7:v.overlay)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Overlay
}
}
}
