% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_proj.R
\name{grass7_v_proj}
\alias{grass7_v_proj}
\title{QGIS algorithm v.proj}
\usage{
grass7_v_proj(
  input = qgisprocess::qgis_default_value(),
  crs = qgisprocess::qgis_default_value(),
  smax = qgisprocess::qgis_default_value(),
  .z = qgisprocess::qgis_default_value(),
  .w = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{source} - Input vector to reproject. Path to a vector layer.}

\item{crs}{\code{crs} - New coordinate reference system. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{smax}{\code{number} - Maximum segment length in meters in output vector map. A numeric value.}

\item{.z}{\code{boolean} - Assume z coordinate is ellipsoidal height and transform if possible. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.w}{\code{boolean} - Disable wrapping to -180,180 for latlon output. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{vectorDestination} - Output vector map. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.proj (grass7:v.proj)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Output vector ma
}
}
}
