% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_random.R
\name{grass7_v_random}
\alias{grass7_v_random}
\title{QGIS algorithm v.random}
\usage{
grass7_v_random(
  npoints = qgisprocess::qgis_default_value(),
  restrict = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  zmin = qgisprocess::qgis_default_value(),
  zmax = qgisprocess::qgis_default_value(),
  seed = qgisprocess::qgis_default_value(),
  column = qgisprocess::qgis_default_value(),
  column_type = qgisprocess::qgis_default_value(),
  .z = qgisprocess::qgis_default_value(),
  .a = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{npoints}{\code{number} - Number of points to be created. A numeric value.}

\item{restrict}{\code{source} - Restrict points to areas in input vector. Path to a vector layer.}

\item{where}{\code{string} - WHERE conditions of SQL statement without 'where' keyword. String value.}

\item{zmin}{\code{number} - Minimum z height for 3D output. A numeric value.}

\item{zmax}{\code{number} - Maximum z height for 3D output. A numeric value.}

\item{seed}{\code{number} - Seed for random number generation. A numeric value.}

\item{column}{\code{string} - Column for Z values. String value.}

\item{column_type}{\code{enum}  of \verb{("integer", "double precision")} - Type of column for z values. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.z}{\code{boolean} - Create 3D output. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -z.}

\item{.a}{\code{boolean} - Generate n points for each individual area. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -a.}

\item{output}{\code{vectorDestination} - Random. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.random (grass7:v.random)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Random
}
}
}
