% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_rast_stats.R
\name{grass7_v_rast_stats}
\alias{grass7_v_rast_stats}
\title{QGIS algorithm v.rast.stats}
\usage{
grass7_v_rast_stats(
  map = qgisprocess:::qgis_default_value(),
  raster = qgisprocess:::qgis_default_value(),
  column_prefix = qgisprocess:::qgis_default_value(),
  method = qgisprocess:::qgis_default_value(),
  percentile = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{map}{\code{source} - Name of vector polygon map. Path to a vector layer.}

\item{raster}{\code{raster} - Name of raster map to calculate statistics from. Path to a raster layer.}

\item{column_prefix}{\code{string} - Column prefix for new attribute columns. String value.}

\item{method}{\code{enum}  of \verb{("number", "minimum", "maximum", "range", "average", "stddev", "variance", "coeff_var", "sum", "first_quartile", "median", "third_quartile", "percentile")} - The methods to use. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{percentile}{\code{number} - Percentile to calculate. A numeric value.}

\item{output}{\code{vectorDestination} - Rast stats. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.rast.stats (grass7:v.rast.stats)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Rast stats
}
}
}
