% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_select.R
\name{grass7_v_select}
\alias{grass7_v_select}
\title{QGIS algorithm v.select}
\usage{
grass7_v_select(
  ainput = qgisprocess::qgis_default_value(),
  atype = qgisprocess::qgis_default_value(),
  binput = qgisprocess::qgis_default_value(),
  btype = qgisprocess::qgis_default_value(),
  operator = qgisprocess::qgis_default_value(),
  relate = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  .r = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ainput}{\code{source} - Input layer (A). Path to a vector layer.}

\item{atype}{\code{enum}  of \verb{("point", "line", "boundary", "centroid", "area")} - Input layer (A) Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{binput}{\code{source} - Input layer (B). Path to a vector layer.}

\item{btype}{\code{enum}  of \verb{("point", "line", "boundary", "centroid", "area")} - Input layer (B) Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{operator}{\code{enum}  of \verb{("overlap", "equals", "disjoint", "intersect", "touches", "crosses", "within", "contains", "overlaps", "relate")} - Operator to use. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{relate}{\code{string} - Intersection Matrix Pattern used for 'relate' operator. String value.}

\item{.t}{\code{boolean} - Do not create attribute table. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{.c}{\code{boolean} - Do not skip features without category. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -c.}

\item{.r}{\code{boolean} - Reverse selection. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -r.}

\item{output}{\code{vectorDestination} - Selected. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.select (grass7:v.select)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Selected
}
}
}
