% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_surf_bspline.R
\name{grass7_v_surf_bspline}
\alias{grass7_v_surf_bspline}
\title{QGIS algorithm v.surf.bspline}
\usage{
grass7_v_surf_bspline(
  input = qgisprocess::qgis_default_value(),
  column = qgisprocess::qgis_default_value(),
  sparse_input = qgisprocess::qgis_default_value(),
  ew_step = qgisprocess::qgis_default_value(),
  ns_step = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  lambda_i = qgisprocess::qgis_default_value(),
  solver = qgisprocess::qgis_default_value(),
  maxit = qgisprocess::qgis_default_value(),
  error = qgisprocess::qgis_default_value(),
  memory = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  raster_output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{source} - Input points layer. Path to a vector layer.}

\item{column}{\code{field} - Attribute table column with values to interpolate. The name of an existing field. ; delimited list of existing field names.}

\item{sparse_input}{\code{source} - Sparse points layer. Path to a vector layer.}

\item{ew_step}{\code{number} - Length of each spline step in the east-west direction. A numeric value.}

\item{ns_step}{\code{number} - Length of each spline step in the north-south direction. A numeric value.}

\item{method}{\code{enum}  of \verb{("bilinear", "bicubic")} - Spline interpolation algorithm. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{lambda_i}{\code{number} - Tykhonov regularization parameter (affects smoothing). A numeric value.}

\item{solver}{\code{enum}  of \verb{("cholesky", "cg")} - Type of solver which should solve the symmetric linear equation system. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{maxit}{\code{number} - Maximum number of iteration used to solve the linear equation system. A numeric value.}

\item{error}{\code{number} - Error break criteria for iterative solver. A numeric value.}

\item{memory}{\code{number} - Maximum memory to be used (in MB). A numeric value.}

\item{output}{\code{vectorDestination} - Output vector. Path for new vector layer.}

\item{raster_output}{\code{rasterDestination} - Interpolated spline. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.surf.bspline (grass7:v.surf.bspline)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Output vector
\item raster_output - outputRaster - Interpolated spline
}
}
}
