% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_surf_rst.R
\name{grass7_v_surf_rst}
\alias{grass7_v_surf_rst}
\title{QGIS algorithm v.surf.rst}
\usage{
grass7_v_surf_rst(
  input = qgisprocess::qgis_default_value(),
  zcolumn = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  mask = qgisprocess::qgis_default_value(),
  tension = qgisprocess::qgis_default_value(),
  smooth = qgisprocess::qgis_default_value(),
  smooth_column = qgisprocess::qgis_default_value(),
  segmax = qgisprocess::qgis_default_value(),
  npmin = qgisprocess::qgis_default_value(),
  dmin = qgisprocess::qgis_default_value(),
  dmax = qgisprocess::qgis_default_value(),
  zscale = qgisprocess::qgis_default_value(),
  theta = qgisprocess::qgis_default_value(),
  scalex = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  .d = qgisprocess::qgis_default_value(),
  elevation = qgisprocess::qgis_default_value(),
  slope = qgisprocess::qgis_default_value(),
  aspect = qgisprocess::qgis_default_value(),
  pcurvature = qgisprocess::qgis_default_value(),
  tcurvature = qgisprocess::qgis_default_value(),
  mcurvature = qgisprocess::qgis_default_value(),
  deviations = qgisprocess::qgis_default_value(),
  treeseg = qgisprocess::qgis_default_value(),
  overwin = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{source} - Input points layer. Path to a vector layer.}

\item{zcolumn}{\code{field} - Name of the attribute column with values to be used for approximation. The name of an existing field. ; delimited list of existing field names.}

\item{where}{\code{string} - WHERE conditions of SQL statement without 'where' keyword. String value.}

\item{mask}{\code{raster} - Name of the raster map used as mask. Path to a raster layer.}

\item{tension}{\code{number} - Tension parameter. A numeric value.}

\item{smooth}{\code{number} - Smoothing parameter. A numeric value.}

\item{smooth_column}{\code{field} - Name of the attribute column with smoothing parameters. The name of an existing field. ; delimited list of existing field names.}

\item{segmax}{\code{number} - Maximum number of points in a segment. A numeric value.}

\item{npmin}{\code{number} - Minimum number of points for approximation in a segment (segmax). A numeric value.}

\item{dmin}{\code{number} - Minimum distance between points (to remove almost identical points). A numeric value.}

\item{dmax}{\code{number} - Maximum distance between points on isoline (to insert additional points). A numeric value.}

\item{zscale}{\code{number} - Conversion factor for values used for approximation. A numeric value.}

\item{theta}{\code{number} - Anisotropy angle (in degrees counterclockwise from East). A numeric value.}

\item{scalex}{\code{number} - Anisotropy scaling factor. A numeric value.}

\item{.t}{\code{boolean} - Use scale dependent tension. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -t.}

\item{.d}{\code{boolean} - Output partial derivatives instead of topographic parameters. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -d.}

\item{elevation}{\code{rasterDestination} - Interpolated RST. Path for new raster layer.}

\item{slope}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{aspect}{\code{rasterDestination} - Aspect. Path for new raster layer.}

\item{pcurvature}{\code{rasterDestination} - Profile curvature. Path for new raster layer.}

\item{tcurvature}{\code{rasterDestination} - Tangential curvature. Path for new raster layer.}

\item{mcurvature}{\code{rasterDestination} - Mean curvature. Path for new raster layer.}

\item{deviations}{\code{vectorDestination} - Deviations. Path for new vector layer.}

\item{treeseg}{\code{vectorDestination} - Quadtree Segmentation. Path for new vector layer.}

\item{overwin}{\code{vectorDestination} - Overlapping Windows. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.surf.rst (grass7:v.surf.rst)
}
\details{
\subsection{Outputs description}{
\itemize{
\item aspect - outputRaster - Aspect
\item deviations - outputVector - Deviations
\item elevation - outputRaster - Interpolated RST
\item mcurvature - outputRaster - Mean curvature
\item overwin - outputVector - Overlapping Windows
\item pcurvature - outputRaster - Profile curvature
\item slope - outputRaster - Slope
\item tcurvature - outputRaster - Tangential curvature
\item treeseg - outputVector - Quadtree Segmentation
}
}
}
