% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_to_rast.R
\name{grass7_v_to_rast}
\alias{grass7_v_to_rast}
\title{QGIS algorithm v.to.rast}
\usage{
grass7_v_to_rast(
  input = qgisprocess::qgis_default_value(),
  type = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  use = qgisprocess::qgis_default_value(),
  attribute_column = qgisprocess::qgis_default_value(),
  rgb_column = qgisprocess::qgis_default_value(),
  label_column = qgisprocess::qgis_default_value(),
  value = qgisprocess::qgis_default_value(),
  memory = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{source} - Input vector layer. Path to a vector layer.}

\item{type}{\code{enum}  of \verb{("point", "line", "boundary", "area")} - Input feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{where}{\code{string} - WHERE conditions of SQL statement without 'where' keyword. String value.}

\item{use}{\code{enum}  of \verb{("attr", "cat", "val", "z", "dir")} - Source of raster values. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{attribute_column}{\code{field} - Name of column for 'attr' parameter (data type must be numeric). The name of an existing field. ; delimited list of existing field names.}

\item{rgb_column}{\code{field} - Name of color definition column (with RRR:GGG:BBB entries). The name of an existing field. ; delimited list of existing field names.}

\item{label_column}{\code{field} - Name of column used as raster category labels. The name of an existing field. ; delimited list of existing field names.}

\item{value}{\code{number} - Raster value (for use=val). A numeric value.}

\item{memory}{\code{number} - Maximum memory to be used (in MB). A numeric value.}

\item{output}{\code{rasterDestination} - Rasterized. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.to.rast (grass7:v.to.rast)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Rasterized
}
}
}
