% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_g_extension_manage.R
\name{grass_g_extension_manage}
\alias{grass_g_extension_manage}
\title{QGIS algorithm - g.extension.manage}
\usage{
grass_g_extension_manage(
  extension = qgisprocess:::qgis_default_value(),
  operation = qgisprocess:::qgis_default_value(),
  .f = qgisprocess:::qgis_default_value(),
  .t = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{extension}{\code{string} - Name of Extension. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{operation}{\code{enum}  of \verb{("add", "remove")} - Operation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.f}{\code{boolean} - Force (required for removal). 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -f.}

\item{.t}{\code{boolean} - Operate on toolboxes instead of single modules (experimental). 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -t.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS g.extension.manage (grass:g.extension.manage). g.extension.manage - Install or uninstall GRASS addons.
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
