% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_i_aster_toar.R
\name{grass_i_aster_toar}
\alias{grass_i_aster_toar}
\title{QGIS algorithm - i.aster.toar}
\usage{
grass_i_aster_toar(
  input = qgisprocess:::qgis_default_value(),
  dayofyear = qgisprocess:::qgis_default_value(),
  sun_elevation = qgisprocess:::qgis_default_value(),
  .r = qgisprocess:::qgis_default_value(),
  .a = qgisprocess:::qgis_default_value(),
  .b = qgisprocess:::qgis_default_value(),
  .c = qgisprocess:::qgis_default_value(),
  .d = qgisprocess:::qgis_default_value(),
  .e = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{multilayer} - Names of ASTER DN layers (15 layers). .}

\item{dayofyear}{\code{number} - Day of Year of satellite overpass (0-366). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{sun_elevation}{\code{number} - Sun elevation angle (degrees,  90.0). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{.r}{\code{boolean} - Output is radiance (W\code{/}m2). 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -r.}

\item{.a}{\code{boolean} - VNIR is High Gain. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -a.}

\item{.b}{\code{boolean} - SWIR is High Gain. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -b.}

\item{.c}{\code{boolean} - VNIR is Low Gain 1. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -c.}

\item{.d}{\code{boolean} - SWIR is Low Gain 1. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -d.}

\item{.e}{\code{boolean} - SWIR is Low Gain 2. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -e.}

\item{output}{\code{folderDestination} - Output Directory. Path for an existing or new folder.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS i.aster.toar (grass:i.aster.toar). Calculates Top of Atmosphere Radiance/Reflectance/Brightness Temperature from ASTER DN.
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFolder - Output Directory
}
}
}
