% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_i_eb_netrad.R
\name{grass_i_eb_netrad}
\alias{grass_i_eb_netrad}
\title{QGIS algorithm - i.eb.netrad}
\usage{
grass_i_eb_netrad(
  albedo = qgisprocess:::qgis_default_value(),
  ndvi = qgisprocess:::qgis_default_value(),
  temperature = qgisprocess:::qgis_default_value(),
  localutctime = qgisprocess:::qgis_default_value(),
  temperaturedifference2m = qgisprocess:::qgis_default_value(),
  emissivity = qgisprocess:::qgis_default_value(),
  transmissivity_singleway = qgisprocess:::qgis_default_value(),
  dayofyear = qgisprocess:::qgis_default_value(),
  sunzenithangle = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{albedo}{\code{raster} - Name of albedo raster map (0.0;1.0). Path to a raster layer.}

\item{ndvi}{\code{raster} - Name of NDVI raster map (-1.0;+1.0). Path to a raster layer.}

\item{temperature}{\code{raster} - Name of surface temperature raster map (K). Path to a raster layer.}

\item{localutctime}{\code{raster} - Name of time of satellite overpass raster map (local time in UTC). Path to a raster layer.}

\item{temperaturedifference2m}{\code{raster} - Name of the difference map of temperature from surface skin to about 2 m height (K). Path to a raster layer.}

\item{emissivity}{\code{raster} - Name of the emissivity map (-). Path to a raster layer.}

\item{transmissivity_singleway}{\code{raster} - Name of the single-way atmospheric transmissivitymap (-). Path to a raster layer.}

\item{dayofyear}{\code{raster} - Name of the Day Of Year (DOY) map (-). Path to a raster layer.}

\item{sunzenithangle}{\code{raster} - Name of the sun zenith angle map (degrees). Path to a raster layer.}

\item{output}{\code{rasterDestination} - Net Radiation. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS i.eb.netrad (grass:i.eb.netrad). Net radiation approximation (Bastiaanssen, 1995).
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Net Radiation
}
}
}
