% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_i_eb_soilheatflux.R
\name{grass_i_eb_soilheatflux}
\alias{grass_i_eb_soilheatflux}
\title{QGIS algorithm - i.eb.soilheatflux}
\usage{
grass_i_eb_soilheatflux(
  albedo = qgisprocess:::qgis_default_value(),
  ndvi = qgisprocess:::qgis_default_value(),
  temperature = qgisprocess:::qgis_default_value(),
  netradiation = qgisprocess:::qgis_default_value(),
  localutctime = qgisprocess:::qgis_default_value(),
  .r = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{albedo}{\code{raster} - Name of albedo raster map (0.0;1.0). Path to a raster layer.}

\item{ndvi}{\code{raster} - Name of NDVI raster map (-1.0;+1.0). Path to a raster layer.}

\item{temperature}{\code{raster} - Name of Surface temperature raster map (K). Path to a raster layer.}

\item{netradiation}{\code{raster} - Name of Net Radiation raster map (W\code{/}m2). Path to a raster layer.}

\item{localutctime}{\code{raster} - Name of time of satellite overpass raster map (local time in UTC). Path to a raster layer.}

\item{.r}{\code{boolean} - HAPEX-Sahel empirical correction (Roerink, 1995). 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -r.}

\item{output}{\code{rasterDestination} - Soil Heat Flux. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS i.eb.soilheatflux (grass:i.eb.soilheatflux). Soil heat flux approximation (Bastiaanssen, 1995).
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Soil Heat Flux
}
}
}
