% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_i_evapo_time.R
\name{grass_i_evapo_time}
\alias{grass_i_evapo_time}
\title{QGIS algorithm - i.evapo.time}
\usage{
grass_i_evapo_time(
  eta = qgisprocess:::qgis_default_value(),
  eta_doy = qgisprocess:::qgis_default_value(),
  eto = qgisprocess:::qgis_default_value(),
  eto_doy_min = qgisprocess:::qgis_default_value(),
  start_period = qgisprocess:::qgis_default_value(),
  end_period = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{eta}{\code{multilayer} - Names of satellite ETa raster maps (mm\code{/}d or cm\code{/}d). .}

\item{eta_doy}{\code{multilayer} - Names of satellite ETa Day of Year (DOY) raster maps (0-400) (-). .}

\item{eto}{\code{multilayer} - Names of meteorological station ETo raster maps (0-400) (mm\code{/}d or cm\code{/}d). .}

\item{eto_doy_min}{\code{number} - Value of DOY for ETo first day. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{start_period}{\code{number} - Value of DOY for the first day of the period studied. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{end_period}{\code{number} - Value of DOY for the last day of the period studied. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{output}{\code{rasterDestination} - Temporal integration. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS i.evapo.time (grass:i.evapo.time). Computes temporal integration of satellite ET actual (ETa) following the daily ET reference (ETo) from meteorological station(s).
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Temporal integration
}
}
}
