% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_r_blend_rgb.R
\name{grass_r_blend_rgb}
\alias{grass_r_blend_rgb}
\title{QGIS algorithm - r.blend.rgb}
\usage{
grass_r_blend_rgb(
  first = qgisprocess:::qgis_default_value(),
  second = qgisprocess:::qgis_default_value(),
  percent = qgisprocess:::qgis_default_value(),
  output_red = qgisprocess:::qgis_default_value(),
  output_green = qgisprocess:::qgis_default_value(),
  output_blue = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{first}{\code{raster} - Name of first raster map for blending. Path to a raster layer.}

\item{second}{\code{raster} - Name of second raster map for blending. Path to a raster layer.}

\item{percent}{\code{number} - Percentage weight of first map for color blending. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{output_red}{\code{rasterDestination} - Blended Red. Path for new raster layer.}

\item{output_green}{\code{rasterDestination} - Blended Green. Path for new raster layer.}

\item{output_blue}{\code{rasterDestination} - Blended Blue. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.blend.rgb (grass:r.blend.rgb). r.blend.rgb - Blends color components of two raster maps by a given ratio and exports into three rasters.
}
\details{
\subsection{Outputs description}{
\itemize{
\item output_blue - outputRaster - Blended Blue
\item output_green - outputRaster - Blended Green
\item output_red - outputRaster - Blended Red
}
}
}
