% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_r_fill_dir.R
\name{grass_r_fill_dir}
\alias{grass_r_fill_dir}
\title{QGIS algorithm - r.fill.dir}
\usage{
grass_r_fill_dir(
  input = qgisprocess:::qgis_default_value(),
  format = qgisprocess:::qgis_default_value(),
  .f = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  direction = qgisprocess:::qgis_default_value(),
  areas = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{input}{\code{raster} - Elevation. Path to a raster layer.}

\item{format}{\code{enum}  of \verb{("grass", "agnps", "answers")} - Output aspect direction format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.f}{\code{boolean} - Find unresolved areas only. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -f.}

\item{output}{\code{rasterDestination} - Depressionless DEM. Path for new raster layer.}

\item{direction}{\code{rasterDestination} - Flow direction. Path for new raster layer.}

\item{areas}{\code{rasterDestination} - Problem areas. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.fill.dir (grass:r.fill.dir). Filters and generates a depressionless elevation layer and a flow direction layer from a given elevation raster layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item areas - outputRaster - Problem areas
\item direction - outputRaster - Flow direction
\item output - outputRaster - Depressionless DEM
}
}
}
