% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_r_gwflow.R
\name{grass_r_gwflow}
\alias{grass_r_gwflow}
\title{QGIS algorithm - r.gwflow}
\usage{
grass_r_gwflow(
  phead = qgisprocess:::qgis_default_value(),
  status = qgisprocess:::qgis_default_value(),
  hc_x = qgisprocess:::qgis_default_value(),
  hc_y = qgisprocess:::qgis_default_value(),
  q = qgisprocess:::qgis_default_value(),
  s = qgisprocess:::qgis_default_value(),
  recharge = qgisprocess:::qgis_default_value(),
  top = qgisprocess:::qgis_default_value(),
  bottom = qgisprocess:::qgis_default_value(),
  type = qgisprocess:::qgis_default_value(),
  river_bed = qgisprocess:::qgis_default_value(),
  river_head = qgisprocess:::qgis_default_value(),
  river_leak = qgisprocess:::qgis_default_value(),
  drain_bed = qgisprocess:::qgis_default_value(),
  drain_leak = qgisprocess:::qgis_default_value(),
  dtime = qgisprocess:::qgis_default_value(),
  maxit = qgisprocess:::qgis_default_value(),
  error = qgisprocess:::qgis_default_value(),
  solver = qgisprocess:::qgis_default_value(),
  relax = qgisprocess:::qgis_default_value(),
  .f = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  vx = qgisprocess:::qgis_default_value(),
  vy = qgisprocess:::qgis_default_value(),
  budget = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{phead}{\code{raster} - The initial piezometric head in (m). Path to a raster layer.}

\item{status}{\code{raster} - Boundary condition status, 0-inactive, 1-active, 2-dirichlet. Path to a raster layer.}

\item{hc_x}{\code{raster} - X-part of the hydraulic conductivity tensor in (m\code{/}s). Path to a raster layer.}

\item{hc_y}{\code{raster} - Y-part of the hydraulic conductivity tensor in (m\code{/}s). Path to a raster layer.}

\item{q}{\code{raster} - Water sources and sinks in (m^3\code{/}s). Path to a raster layer.}

\item{s}{\code{raster} - Specific yield in (1\code{/}m). Path to a raster layer.}

\item{recharge}{\code{raster} - Recharge map e.g: 6\emph{10^-9 per cell in (m^3\code{/}s}m^2). Path to a raster layer.}

\item{top}{\code{raster} - Top surface of the aquifer in (m). Path to a raster layer.}

\item{bottom}{\code{raster} - Bottom surface of the aquifer in (m). Path to a raster layer.}

\item{type}{\code{enum}  of \verb{("confined", "unconfined")} - The type of groundwater flow. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{river_bed}{\code{raster} - The height of the river bed in (m). Path to a raster layer.}

\item{river_head}{\code{raster} - Water level (head) of the river with leakage connection in (m). Path to a raster layer.}

\item{river_leak}{\code{raster} - The leakage coefficient of the river bed in (1\code{/}s). Path to a raster layer.}

\item{drain_bed}{\code{raster} - The height of the drainage bed in (m). Path to a raster layer.}

\item{drain_leak}{\code{raster} - The leakage coefficient of the drainage bed in (1\code{/}s). Path to a raster layer.}

\item{dtime}{\code{number} - The calculation time in seconds. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{maxit}{\code{number} - Maximum number of iteration used to solver the linear equation system. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{error}{\code{number} - Error break criteria for iterative solvers (jacobi, sor, cg or bicgstab). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{solver}{\code{enum}  of \verb{("cg", "pcg", "cholesky")} - The type of solver which should solve the symmetric linear equation system. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{relax}{\code{string} - The relaxation parameter used by the jacobi and sor solver for speedup or stabilizing. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{.f}{\code{boolean} - Allocate a full quadratic linear equation system, default is a sparse linear equation system. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -f.}

\item{output}{\code{rasterDestination} - Groundwater flow. Path for new raster layer.}

\item{vx}{\code{rasterDestination} - Groundwater filter velocity vector part in x direction (m\code{/}s). Path for new raster layer.}

\item{vy}{\code{rasterDestination} - Groundwater filter velocity vector part in y direction (m\code{/}s). Path for new raster layer.}

\item{budget}{\code{rasterDestination} - Groundwater budget for each cell (m^3\code{/}s). Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.gwflow (grass:r.gwflow). Numerical calculation program for transient, confined and unconfined groundwater flow in two dimensions.
}
\details{
\subsection{Outputs description}{
\itemize{
\item budget - outputRaster - Groundwater budget for each cell (m^3\code{/}s)
\item output - outputRaster - Groundwater flow
\item vx - outputRaster - Groundwater filter velocity vector part in x direction (m\code{/}s)
\item vy - outputRaster - Groundwater filter velocity vector part in y direction (m\code{/}s)
}
}
}
