% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_r_regression_multi.R
\name{grass_r_regression_multi}
\alias{grass_r_regression_multi}
\title{QGIS algorithm - r.regression.multi}
\usage{
grass_r_regression_multi(
  mapx = qgisprocess:::qgis_default_value(),
  mapy = qgisprocess:::qgis_default_value(),
  residuals = qgisprocess:::qgis_default_value(),
  estimates = qgisprocess:::qgis_default_value(),
  html = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{mapx}{\code{multilayer} - Map(s) for x coefficient. .}

\item{mapy}{\code{raster} - Map for y coefficient. Path to a raster layer.}

\item{residuals}{\code{rasterDestination} - Residual Map. Path for new raster layer.}

\item{estimates}{\code{rasterDestination} - Estimates Map. Path for new raster layer.}

\item{html}{\code{fileDestination} - Regression coefficients. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.regression.multi (grass:r.regression.multi). Calculates multiple linear regression from raster maps.
}
\details{
\subsection{Outputs description}{
\itemize{
\item estimates - outputRaster - Estimates Map
\item html - outputHtml - Regression coefficients
\item residuals - outputRaster - Residual Map
}
}
}
