% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_r_sim_sediment.R
\name{grass_r_sim_sediment}
\alias{grass_r_sim_sediment}
\title{QGIS algorithm - r.sim.sediment}
\usage{
grass_r_sim_sediment(
  elevation = qgisprocess:::qgis_default_value(),
  water_depth = qgisprocess:::qgis_default_value(),
  dx = qgisprocess:::qgis_default_value(),
  dy = qgisprocess:::qgis_default_value(),
  detachment_coeff = qgisprocess:::qgis_default_value(),
  transport_coeff = qgisprocess:::qgis_default_value(),
  shear_stress = qgisprocess:::qgis_default_value(),
  man = qgisprocess:::qgis_default_value(),
  man_value = qgisprocess:::qgis_default_value(),
  observation = qgisprocess:::qgis_default_value(),
  nwalkers = qgisprocess:::qgis_default_value(),
  niterations = qgisprocess:::qgis_default_value(),
  output_step = qgisprocess:::qgis_default_value(),
  diffusion_coeff = qgisprocess:::qgis_default_value(),
  transport_capacity = qgisprocess:::qgis_default_value(),
  tlimit_erosion_deposition = qgisprocess:::qgis_default_value(),
  sediment_concentration = qgisprocess:::qgis_default_value(),
  sediment_flux = qgisprocess:::qgis_default_value(),
  erosion_deposition = qgisprocess:::qgis_default_value(),
  walkers_output = qgisprocess:::qgis_default_value(),
  logfile = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{elevation}{\code{raster} - Name of the elevation raster map (m). Path to a raster layer.}

\item{water_depth}{\code{raster} - Name of the water depth raster map (m). Path to a raster layer.}

\item{dx}{\code{raster} - Name of the x-derivatives raster map (m\code{/}m). Path to a raster layer.}

\item{dy}{\code{raster} - Name of the y-derivatives raster map (m\code{/}m). Path to a raster layer.}

\item{detachment_coeff}{\code{raster} - Name of the detachment capacity coefficient raster map (s\code{/}m). Path to a raster layer.}

\item{transport_coeff}{\code{raster} - Name of the transport capacity coefficient raster map (s). Path to a raster layer.}

\item{shear_stress}{\code{raster} - Name of the critical shear stress raster map (Pa). Path to a raster layer.}

\item{man}{\code{raster} - Name of the Mannings n raster map. Path to a raster layer.}

\item{man_value}{\code{number} - Name of the Mannings n value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{observation}{\code{source} - Sampling locations vector points. Path to a vector layer.}

\item{nwalkers}{\code{number} - Number of walkers. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{niterations}{\code{number} - Time used for iterations (minutes). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{output_step}{\code{number} - Time interval for creating output maps (minutes). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{diffusion_coeff}{\code{number} - Water diffusion constant. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{transport_capacity}{\code{rasterDestination} - Transport capacity (kg\code{/}ms). Path for new raster layer.}

\item{tlimit_erosion_deposition}{\code{rasterDestination} - Transport limited erosion-deposition (kg\code{/}m2s). Path for new raster layer.}

\item{sediment_concentration}{\code{rasterDestination} - Sediment concentration (particle\code{/}m3). Path for new raster layer.}

\item{sediment_flux}{\code{rasterDestination} - Sediment flux (kg\code{/}ms). Path for new raster layer.}

\item{erosion_deposition}{\code{rasterDestination} - Erosion-deposition (kg\code{/}m2s). Path for new raster layer.}

\item{walkers_output}{\code{vectorDestination} - Name of the output walkers vector points layer. Path for new vector layer.}

\item{logfile}{\code{fileDestination} - Name for sampling points output text file.. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.sim.sediment (grass:r.sim.sediment). Sediment transport and erosion/deposition simulation using path sampling method (SIMWE).
}
\details{
\subsection{Outputs description}{
\itemize{
\item erosion_deposition - outputRaster - Erosion-deposition (kg\code{/}m2s)
\item logfile - outputFile - Name for sampling points output text file.
\item sediment_concentration - outputRaster - Sediment concentration (particle\code{/}m3)
\item sediment_flux - outputRaster - Sediment flux (kg\code{/}ms)
\item tlimit_erosion_deposition - outputRaster - Transport limited erosion-deposition (kg\code{/}m2s)
\item transport_capacity - outputRaster - Transport capacity (kg\code{/}ms)
\item walkers_output - outputVector - Name of the output walkers vector points layer
}
}
}
