% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_r_slope_aspect.R
\name{grass_r_slope_aspect}
\alias{grass_r_slope_aspect}
\title{QGIS algorithm - r.slope.aspect}
\usage{
grass_r_slope_aspect(
  elevation = qgisprocess:::qgis_default_value(),
  format = qgisprocess:::qgis_default_value(),
  precision = qgisprocess:::qgis_default_value(),
  .a = qgisprocess:::qgis_default_value(),
  .e = qgisprocess:::qgis_default_value(),
  .n = qgisprocess:::qgis_default_value(),
  zscale = qgisprocess:::qgis_default_value(),
  min_slope = qgisprocess:::qgis_default_value(),
  slope = qgisprocess:::qgis_default_value(),
  aspect = qgisprocess:::qgis_default_value(),
  pcurvature = qgisprocess:::qgis_default_value(),
  tcurvature = qgisprocess:::qgis_default_value(),
  dx = qgisprocess:::qgis_default_value(),
  dy = qgisprocess:::qgis_default_value(),
  dxx = qgisprocess:::qgis_default_value(),
  dyy = qgisprocess:::qgis_default_value(),
  dxy = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{elevation}{\code{raster} - Elevation. Path to a raster layer.}

\item{format}{\code{enum}  of \verb{("degrees", "percent")} - Format for reporting the slope. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{precision}{\code{enum}  of \verb{("FCELL", "CELL", "DCELL")} - Type of output aspect and slope layer. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.a}{\code{boolean} - Do not align the current region to the elevation layer. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -a.}

\item{.e}{\code{boolean} - Compute output at edges and near NULL values. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -e.}

\item{.n}{\code{boolean} - Create aspect as degrees clockwise from North (azimuth), with flat = -9999. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -n.}

\item{zscale}{\code{number} - Multiplicative factor to convert elevation units to meters. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{min_slope}{\code{number} - Minimum slope val. (in percent) for which aspect is computed. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{slope}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{aspect}{\code{rasterDestination} - Aspect. Path for new raster layer.}

\item{pcurvature}{\code{rasterDestination} - Profile curvature. Path for new raster layer.}

\item{tcurvature}{\code{rasterDestination} - Tangential curvature. Path for new raster layer.}

\item{dx}{\code{rasterDestination} - First order partial derivative dx (E-W slope). Path for new raster layer.}

\item{dy}{\code{rasterDestination} - First order partial derivative dy (N-S slope). Path for new raster layer.}

\item{dxx}{\code{rasterDestination} - Second order partial derivative dxx. Path for new raster layer.}

\item{dyy}{\code{rasterDestination} - Second order partial derivative dyy. Path for new raster layer.}

\item{dxy}{\code{rasterDestination} - Second order partial derivative dxy. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.slope.aspect (grass:r.slope.aspect). Generates raster layers of slope, aspect, curvatures and partial derivatives from a elevation raster layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item aspect - outputRaster - Aspect
\item dx - outputRaster - First order partial derivative dx (E-W slope)
\item dxx - outputRaster - Second order partial derivative dxx
\item dxy - outputRaster - Second order partial derivative dxy
\item dy - outputRaster - First order partial derivative dy (N-S slope)
\item dyy - outputRaster - Second order partial derivative dyy
\item pcurvature - outputRaster - Profile curvature
\item slope - outputRaster - Slope
\item tcurvature - outputRaster - Tangential curvature
}
}
}
