% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_r_spread.R
\name{grass_r_spread}
\alias{grass_r_spread}
\title{QGIS algorithm - r.spread}
\usage{
grass_r_spread(
  base_ros = qgisprocess:::qgis_default_value(),
  max_ros = qgisprocess:::qgis_default_value(),
  direction_ros = qgisprocess:::qgis_default_value(),
  start = qgisprocess:::qgis_default_value(),
  spotting_distance = qgisprocess:::qgis_default_value(),
  wind_speed = qgisprocess:::qgis_default_value(),
  fuel_moisture = qgisprocess:::qgis_default_value(),
  backdrop = qgisprocess:::qgis_default_value(),
  least_size = qgisprocess:::qgis_default_value(),
  comp_dens = qgisprocess:::qgis_default_value(),
  init_time = qgisprocess:::qgis_default_value(),
  lag = qgisprocess:::qgis_default_value(),
  .s = qgisprocess:::qgis_default_value(),
  .i = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  x_output = qgisprocess:::qgis_default_value(),
  y_output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess:::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{base_ros}{\code{raster} - Raster map containing base ROS (cm\code{/}min). Path to a raster layer.}

\item{max_ros}{\code{raster} - Raster map containing maximal ROS (cm\code{/}min). Path to a raster layer.}

\item{direction_ros}{\code{raster} - Raster map containing directions of maximal ROS (degree). Path to a raster layer.}

\item{start}{\code{raster} - Raster map containing starting sources. Path to a raster layer.}

\item{spotting_distance}{\code{raster} - Raster map containing maximal spotting distance (m, required with -s). Path to a raster layer.}

\item{wind_speed}{\code{raster} - Raster map containing midflame wind speed (ft\code{/}min, required with -s). Path to a raster layer.}

\item{fuel_moisture}{\code{raster} - Raster map containing fine fuel moisture of the cell receiving a spotting firebrand (\%, required with -s). Path to a raster layer.}

\item{backdrop}{\code{raster} - Name of raster map as a display backdrop. Path to a raster layer.}

\item{least_size}{\code{enum}  of \verb{("3", "5", "7", "9", "11", "13", "15")} - Basic sampling window size needed to meet certain accuracy (3). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{comp_dens}{\code{number} - Sampling density for additional computing (range: 0.0 - 1.0 (0.5)). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{init_time}{\code{number} - Initial time for current simulation (0) (min). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{lag}{\code{number} - Simulating time duration LAG (fill the region) (min). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{.s}{\code{boolean} - Consider spotting effect (for wildfires). 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -s.}

\item{.i}{\code{boolean} - Use start raster map values in output spread time raster map. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -i.}

\item{output}{\code{rasterDestination} - Spread Time. Path for new raster layer.}

\item{x_output}{\code{rasterDestination} - X Back Coordinates. Path for new raster layer.}

\item{y_output}{\code{rasterDestination} - Y Back Coordinates. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.spread (grass:r.spread). Simulates elliptically anisotropic spread.
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Spread Time
\item x_output - outputRaster - X Back Coordinates
\item y_output - outputRaster - Y Back Coordinates
}
}
}
