% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_r_univar.R
\name{grass_r_univar}
\alias{grass_r_univar}
\title{QGIS algorithm - r.univar}
\usage{
grass_r_univar(
  map = qgisprocess:::qgis_default_value(),
  zones = qgisprocess:::qgis_default_value(),
  percentile = qgisprocess:::qgis_default_value(),
  separator = qgisprocess:::qgis_default_value(),
  .e = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{map}{\code{multilayer} - Name of raster map(s). .}

\item{zones}{\code{raster} - Raster map used for zoning, must be of type CELL. Path to a raster layer.}

\item{percentile}{\code{string} - Percentile to calculate (comma separated list if multiple) (requires extended statistics flag). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{separator}{\code{string} - Field separator. Special characters: pipe, comma, space, tab, newline. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{.e}{\code{boolean} - Calculate extended statistics. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -e.}

\item{output}{\code{fileDestination} - Univariate results. Path for new file.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS region cellsize (leave 0 for default). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS r.univar (grass:r.univar). Calculates univariate statistics from the non-null cells of a raster map.
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputFile - Univariate results
}
}
}
