% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_v_distance.R
\name{grass_v_distance}
\alias{grass_v_distance}
\title{QGIS algorithm - v.distance}
\usage{
grass_v_distance(
  from = qgisprocess:::qgis_default_value(),
  from_type = qgisprocess:::qgis_default_value(),
  to = qgisprocess:::qgis_default_value(),
  to_type = qgisprocess:::qgis_default_value(),
  dmax = qgisprocess:::qgis_default_value(),
  dmin = qgisprocess:::qgis_default_value(),
  upload = qgisprocess:::qgis_default_value(),
  column = qgisprocess:::qgis_default_value(),
  to_column = qgisprocess:::qgis_default_value(),
  from_output = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{from}{\code{source} - 'from' vector map. Path to a vector layer.}

\item{from_type}{\code{enum}  of \verb{("point", "line", "boundary", "area", "centroid")} - 'from' feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{to}{\code{source} - 'to' vector map. Path to a vector layer.}

\item{to_type}{\code{enum}  of \verb{("point", "line", "boundary", "area", "centroid")} - 'to' feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{dmax}{\code{number} - Maximum distance or -1.0 for no limit. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{dmin}{\code{number} - Minimum distance or -1.0 for no limit. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{upload}{\code{enum}  of \verb{("cat", "dist", "to_x", "to_y", "to_along", "to_angle", "to_attr")} - 'upload': Values describing the relation between two nearest features. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{column}{\code{field} - Column name(s) where values specified by 'upload' option will be uploaded. The name of an existing field. ; delimited list of existing field names.}

\item{to_column}{\code{field} - Column name of nearest feature (used with upload=to_attr). The name of an existing field. ; delimited list of existing field names.}

\item{from_output}{\code{vectorDestination} - Nearest. Path for new vector layer.}

\item{output}{\code{vectorDestination} - Distance. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.distance (grass:v.distance). Finds the nearest element in vector map 'to' for elements in vector map 'from'.
}
\details{
\subsection{Outputs description}{
\itemize{
\item from_output - outputVector - Nearest
\item output - outputVector - Distance
}
}
}
