% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_v_mkgrid.R
\name{grass_v_mkgrid}
\alias{grass_v_mkgrid}
\title{QGIS algorithm - v.mkgrid}
\usage{
grass_v_mkgrid(
  grid = qgisprocess:::qgis_default_value(),
  position = qgisprocess:::qgis_default_value(),
  coordinates = qgisprocess:::qgis_default_value(),
  box = qgisprocess:::qgis_default_value(),
  angle = qgisprocess:::qgis_default_value(),
  breaks = qgisprocess:::qgis_default_value(),
  type = qgisprocess:::qgis_default_value(),
  .h = qgisprocess:::qgis_default_value(),
  .a = qgisprocess:::qgis_default_value(),
  map = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{grid}{\code{string} - Number of rows and columns in grid. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{position}{\code{enum}  of \code{("coor")} - Where to place the grid. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{coordinates}{\code{point} - Lower left easting and northing coordinates of map. Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{box}{\code{string} - Width and height of boxes in grid. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{angle}{\code{number} - Angle of rotation (in degrees counter-clockwise). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{breaks}{\code{number} - Number of vertex points per grid cell. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{type}{\code{enum}  of \verb{("point", "line", "area")} - Output feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{.h}{\code{boolean} - Create hexagons (default: rectangles). 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -h.}

\item{.a}{\code{boolean} - Allow asymmetric hexagons. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression. Original algorithm parameter name: -a.}

\item{map}{\code{vectorDestination} - Grid. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.mkgrid (grass:v.mkgrid). Creates a GRASS vector layer of a user-defined grid.
}
\details{
\subsection{Outputs description}{
\itemize{
\item map - outputVector - Grid
}
}
}
