% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_v_vect_stats.R
\name{grass_v_vect_stats}
\alias{grass_v_vect_stats}
\title{QGIS algorithm - v.vect.stats}
\usage{
grass_v_vect_stats(
  points = qgisprocess:::qgis_default_value(),
  areas = qgisprocess:::qgis_default_value(),
  type = qgisprocess:::qgis_default_value(),
  method = qgisprocess:::qgis_default_value(),
  points_column = qgisprocess:::qgis_default_value(),
  count_column = qgisprocess:::qgis_default_value(),
  stats_column = qgisprocess:::qgis_default_value(),
  output = qgisprocess:::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess:::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{points}{\code{source} - Name of existing vector map with points. Path to a vector layer.}

\item{areas}{\code{source} - Name of existing vector map with areas. Path to a vector layer.}

\item{type}{\code{enum}  of \verb{("point", "centroid")} - Input feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{method}{\code{enum}  of \verb{("sum", "average", "median", "mode", "minimum", "min_cat", "maximum", "max_cat", "range", "stddev", "variance", "diversity")} - Method for aggregate statistics. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{points_column}{\code{field} - Column name of points map to use for statistics. The name of an existing field. ; delimited list of existing field names.}

\item{count_column}{\code{string} - Column name to upload points count (integer, created if doesn't exists). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{stats_column}{\code{string} - Column name to upload statistics (double, created if doesn't exists). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{output}{\code{vectorDestination} - Updated. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GRASS v.vect.stats (grass:v.vect.stats). Count points in areas and calculate statistics.
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Updated
}
}
}
