% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_addautoincrementalfield.R
\name{qgis_addautoincrementalfield}
\alias{qgis_addautoincrementalfield}
\title{QGIS algorithm Add autoincremental field}
\usage{
qgis_addautoincrementalfield(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD_NAME = qgisprocess::qgis_default_value(),
  START = qgisprocess::qgis_default_value(),
  MODULUS = qgisprocess::qgis_default_value(),
  GROUP_FIELDS = qgisprocess::qgis_default_value(),
  SORT_EXPRESSION = qgisprocess::qgis_default_value(),
  SORT_ASCENDING = qgisprocess::qgis_default_value(),
  SORT_NULLS_FIRST = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD_NAME}{\code{string} - Field name. String value.}

\item{START}{\code{number} - Start values at. A numeric value.}

\item{MODULUS}{\code{number} - Modulus value. A numeric value.}

\item{GROUP_FIELDS}{\code{field} - Group values by. The name of an existing field. ; delimited list of existing field names.}

\item{SORT_EXPRESSION}{\code{expression} - Sort expression. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{SORT_ASCENDING}{\code{boolean} - Sort ascending. 1 for true/yes. 0 for false/no.}

\item{SORT_NULLS_FIRST}{\code{boolean} - Sort nulls first. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Incremented. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Add autoincremental field (native:addautoincrementalfield)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Incremented
}
}
}
