% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_addfieldtoattributestable.R
\name{qgis_addfieldtoattributestable}
\alias{qgis_addfieldtoattributestable}
\title{QGIS algorithm Add field to attributes table}
\usage{
qgis_addfieldtoattributestable(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD_NAME = qgisprocess::qgis_default_value(),
  FIELD_TYPE = qgisprocess::qgis_default_value(),
  FIELD_LENGTH = qgisprocess::qgis_default_value(),
  FIELD_PRECISION = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD_NAME}{\code{string} - Field name. String value.}

\item{FIELD_TYPE}{\code{enum}  of \verb{("Integer (32 bit)", "Decimal (double)", "Text (string)", "Boolean", "Date", "Time", "Date & Time", "Binary Object (BLOB)", "String List", "Integer List", "Decimal (double) List")} - Field type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FIELD_LENGTH}{\code{number} - Field length. A numeric value.}

\item{FIELD_PRECISION}{\code{number} - Field precision. A numeric value.}

\item{OUTPUT}{\code{sink} - Added. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Add field to attributes table (native:addfieldtoattributestable)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Added
}
}
}
