% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_advancedpythonfieldcalculator.R
\name{qgis_advancedpythonfieldcalculator}
\alias{qgis_advancedpythonfieldcalculator}
\title{QGIS algorithm Advanced Python field calculator}
\usage{
qgis_advancedpythonfieldcalculator(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD_NAME = qgisprocess::qgis_default_value(),
  FIELD_TYPE = qgisprocess::qgis_default_value(),
  FIELD_LENGTH = qgisprocess::qgis_default_value(),
  FIELD_PRECISION = qgisprocess::qgis_default_value(),
  GLOBAL = qgisprocess::qgis_default_value(),
  FORMULA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD_NAME}{\code{string} - Result field name. String value.}

\item{FIELD_TYPE}{\code{enum}  of \verb{("Integer (32 bit)", "Decimal (double)", "Text (string)", "Boolean", "Date", "Time", "Date & Time", "Binary Object (BLOB)", "String List", "Integer List", "Decimal (double) List")} - Field type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FIELD_LENGTH}{\code{number} - Field length. A numeric value.}

\item{FIELD_PRECISION}{\code{number} - Field precision. A numeric value.}

\item{GLOBAL}{\code{string} - Global expression. String value.}

\item{FORMULA}{\code{string} - Formula. String value.}

\item{OUTPUT}{\code{sink} - Calculated. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Advanced Python field calculator (qgis:advancedpythonfieldcalculator)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Calculated
}
}
}
