% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_affinetransform.R
\name{qgis_affinetransform}
\alias{qgis_affinetransform}
\title{QGIS algorithm Affine transform}
\usage{
qgis_affinetransform(
  INPUT = qgisprocess::qgis_default_value(),
  DELTA_X = qgisprocess::qgis_default_value(),
  DELTA_Y = qgisprocess::qgis_default_value(),
  DELTA_Z = qgisprocess::qgis_default_value(),
  DELTA_M = qgisprocess::qgis_default_value(),
  SCALE_X = qgisprocess::qgis_default_value(),
  SCALE_Y = qgisprocess::qgis_default_value(),
  SCALE_Z = qgisprocess::qgis_default_value(),
  SCALE_M = qgisprocess::qgis_default_value(),
  ROTATION_Z = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DELTA_X}{\code{distance} - Translation (x-axis). A numeric value.}

\item{DELTA_Y}{\code{distance} - Translation (y-axis). A numeric value.}

\item{DELTA_Z}{\code{number} - Translation (z-axis). A numeric value.}

\item{DELTA_M}{\code{number} - Translation (m values). A numeric value.}

\item{SCALE_X}{\code{number} - Scale factor (x-axis). A numeric value.}

\item{SCALE_Y}{\code{number} - Scale factor (y-axis). A numeric value.}

\item{SCALE_Z}{\code{number} - Scale factor (z-axis). A numeric value.}

\item{SCALE_M}{\code{number} - Scale factor (m values). A numeric value.}

\item{ROTATION_Z}{\code{number} - Rotation around z-axis (degrees counter-clockwise). A numeric value.}

\item{OUTPUT}{\code{sink} - Transformed. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Affine transform (native:affinetransform)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Transformed
}
}
}
