% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_aggregate.R
\name{qgis_aggregate}
\alias{qgis_aggregate}
\title{QGIS algorithm Aggregate}
\usage{
qgis_aggregate(
  INPUT = qgisprocess::qgis_default_value(),
  GROUP_BY = qgisprocess::qgis_default_value(),
  AGGREGATES = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{GROUP_BY}{\code{expression} - Group by expression (NULL to group all features). A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{AGGREGATES}{\code{aggregates} - Aggregates. .}

\item{OUTPUT}{\code{sink} - Aggregated. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Aggregate (native:aggregate)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Aggregated
}
}
}
