% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_alignsingleraster.R
\name{qgis_alignsingleraster}
\alias{qgis_alignsingleraster}
\title{QGIS algorithm - Align raster}
\usage{
qgis_alignsingleraster(
  INPUT = qgisprocess:::qgis_default_value(),
  RESAMPLING_METHOD = qgisprocess:::qgis_default_value(),
  RESCALE = qgisprocess:::qgis_default_value(),
  REFERENCE_LAYER = qgisprocess:::qgis_default_value(),
  CRS = qgisprocess:::qgis_default_value(),
  CELL_SIZE_X = qgisprocess:::qgis_default_value(),
  CELL_SIZE_Y = qgisprocess:::qgis_default_value(),
  GRID_OFFSET_X = qgisprocess:::qgis_default_value(),
  GRID_OFFSET_Y = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{RESAMPLING_METHOD}{\code{enum}  of \verb{("Nearest Neighbour", "Bilinear (2x2 Kernel)", "Cubic (4x4 Kernel)", "Cubic B-Spline (4x4 Kernel)", "Lanczos (6x6 Kernel)", "Average", "Mode", "Maximum", "Minimum", "Median", "First Quartile (Q1)", "Third Quartile (Q3)")} - Resampling method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESCALE}{\code{boolean} - Rescale values according to the cell size. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{REFERENCE_LAYER}{\code{raster} - Reference layer. Path to a raster layer.}

\item{CRS}{\code{crs} - Override reference CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{CELL_SIZE_X}{\code{number} - Override reference cell size X. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CELL_SIZE_Y}{\code{number} - Override reference cell size Y. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRID_OFFSET_X}{\code{number} - Override reference grid offset X. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRID_OFFSET_Y}{\code{number} - Override reference grid offset Y. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EXTENT}{\code{extent} - Clip to extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{OUTPUT}{\code{rasterDestination} - Aligned raster. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Align raster (native:alignsingleraster). Aligns raster by resampling it to the same cell size and reprojecting to the same CRS as a reference raster.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Aligned raster
}
}
}
