% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_angletonearest.R
\name{qgis_angletonearest}
\alias{qgis_angletonearest}
\title{QGIS algorithm Align points to features}
\usage{
qgis_angletonearest(
  INPUT = qgisprocess::qgis_default_value(),
  REFERENCE_LAYER = qgisprocess::qgis_default_value(),
  MAX_DISTANCE = qgisprocess::qgis_default_value(),
  FIELD_NAME = qgisprocess::qgis_default_value(),
  APPLY_SYMBOLOGY = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{REFERENCE_LAYER}{\code{source} - Reference layer. Path to a vector layer.}

\item{MAX_DISTANCE}{\code{distance} - Maximum distance to consider. A numeric value.}

\item{FIELD_NAME}{\code{string} - Angle field name. String value.}

\item{APPLY_SYMBOLOGY}{\code{boolean} - Automatically apply symbology. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Aligned layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Align points to features (native:angletonearest)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Aligned layer
}
}
}
