% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_assignprojection.R
\name{qgis_assignprojection}
\alias{qgis_assignprojection}
\title{QGIS algorithm Assign projection}
\usage{
qgis_assignprojection(
  INPUT = qgisprocess::qgis_default_value(),
  CRS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{CRS}{\code{crs} - Assigned CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT}{\code{sink} - Assigned CRS. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Assign projection (native:assignprojection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Assigned CR
}
}
}
