% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_b3dmtogltf.R
\name{qgis_b3dmtogltf}
\alias{qgis_b3dmtogltf}
\title{QGIS algorithm - Convert B3DM to GLTF}
\usage{
qgis_b3dmtogltf(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{file} - Input B3DM. Path to a file.}

\item{OUTPUT}{\code{fileDestination} - Output file. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convert B3DM to GLTF (native:b3dmtogltf). Converts files from the legacy B3DM format to GLTF. This algorithm converts files from the legacy B3DM format to GLTF.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - Output file
}
}
}
