% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_barplot.R
\name{qgis_barplot}
\alias{qgis_barplot}
\title{QGIS algorithm - Bar plot}
\usage{
qgis_barplot(
  INPUT = qgisprocess:::qgis_default_value(),
  NAME_FIELD = qgisprocess:::qgis_default_value(),
  VALUE_FIELD = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  TITLE = qgisprocess:::qgis_default_value(),
  XAXIS_TITLE = qgisprocess:::qgis_default_value(),
  YAXIS_TITLE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{NAME_FIELD}{\code{field} - Category name field. The name of an existing field. ; delimited list of existing field names.}

\item{VALUE_FIELD}{\code{field} - Value field. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{fileDestination} - Bar plot. Path for new file.}

\item{TITLE}{\code{string} - Title. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{XAXIS_TITLE}{\code{string} - X-axis title. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YAXIS_TITLE}{\code{string} - Y-axis title. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Bar plot (qgis:barplot). Creates a bar plot from a category and a layer field. This algorithm creates a bar plot from a category and a layer field.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputHtml - Bar plot
}
}
}
