% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_basicstatisticsforfields.R
\name{qgis_basicstatisticsforfields}
\alias{qgis_basicstatisticsforfields}
\title{QGIS algorithm - Basic statistics for fields}
\usage{
qgis_basicstatisticsforfields(
  INPUT_LAYER = qgisprocess:::qgis_default_value(),
  FIELD_NAME = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  OUTPUT_HTML_FILE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_LAYER}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD_NAME}{\code{field} - Field to calculate statistics on. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Statistics. Path for new vector layer.}

\item{OUTPUT_HTML_FILE}{\code{fileDestination} - Statistics report. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Basic statistics for fields (native:basicstatisticsforfields). Generates basic statistics from the values in a field of a vector layer. This algorithm generates basic statistics from the analysis of values in a field in the attribute table of a vector layer. Numeric, date, time and string fields are supported. The statistics returned will depend on the field type.
}
\details{
\subsection{Outputs description}{
\itemize{
\item COUNT - outputNumber - Count
\item CV - outputNumber - Coefficient of Variation
\item EMPTY - outputNumber - Number of empty (null) values
\item FILLED - outputNumber - Number of non-empty values
\item FIRSTQUARTILE - outputNumber - First quartile
\item IQR - outputNumber - Interquartile Range (IQR)
\item MAJORITY - outputNumber - Majority (most frequently occurring value)
\item MAX - outputNumber - Maximum value
\item MAX_LENGTH - outputNumber - Maximum length
\item MEAN - outputNumber - Mean value
\item MEAN_LENGTH - outputNumber - Mean length
\item MEDIAN - outputNumber - Median
\item MIN - outputNumber - Minimum value
\item MINORITY - outputNumber - Minority (rarest occurring value)
\item MIN_LENGTH - outputNumber - Minimum length
\item OUTPUT - outputVector - Statistics
\item OUTPUT_HTML_FILE - outputHtml - Statistics report
\item RANGE - outputNumber - Range
\item STD_DEV - outputNumber - Standard deviation
\item SUM - outputNumber - Sum
\item THIRDQUARTILE - outputNumber - Third quartile
\item UNIQUE - outputNumber - Number of unique values
}
}
}
