% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_batchnominatimgeocoder.R
\name{qgis_batchnominatimgeocoder}
\alias{qgis_batchnominatimgeocoder}
\title{QGIS algorithm Batch Nominatim geocoder}
\usage{
qgis_batchnominatimgeocoder(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Address field. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Geocoded. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Batch Nominatim geocoder (native:batchnominatimgeocoder)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Geocoded
}
}
}
