% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_bookmarkstolayer.R
\name{qgis_bookmarkstolayer}
\alias{qgis_bookmarkstolayer}
\title{QGIS algorithm Convert spatial bookmarks to layer}
\usage{
qgis_bookmarkstolayer(
  SOURCE = qgisprocess::qgis_default_value(),
  CRS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SOURCE}{\code{enum}  of \verb{("Project bookmarks", "User bookmarks")} - Bookmark source. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CRS}{\code{crs} - Output CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT}{\code{sink} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convert spatial bookmarks to layer (native:bookmarkstolayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output
}
}
}
