% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_buffer.R
\name{qgis_buffer}
\alias{qgis_buffer}
\title{QGIS algorithm Buffer}
\usage{
qgis_buffer(
  INPUT = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  SEGMENTS = qgisprocess::qgis_default_value(),
  END_CAP_STYLE = qgisprocess::qgis_default_value(),
  JOIN_STYLE = qgisprocess::qgis_default_value(),
  MITER_LIMIT = qgisprocess::qgis_default_value(),
  DISSOLVE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DISTANCE}{\code{distance} - Distance. A numeric value.}

\item{SEGMENTS}{\code{number} - Segments. A numeric value.}

\item{END_CAP_STYLE}{\code{enum}  of \verb{("Round", "Flat", "Square")} - End cap style. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{JOIN_STYLE}{\code{enum}  of \verb{("Round", "Miter", "Bevel")} - Join style. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MITER_LIMIT}{\code{number} - Miter limit. A numeric value.}

\item{DISSOLVE}{\code{boolean} - Dissolve result. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Buffered. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Buffer (native:buffer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Buffered
}
}
}
