% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_bufferbym.R
\name{qgis_bufferbym}
\alias{qgis_bufferbym}
\title{QGIS algorithm - Variable width buffer (by M value)}
\usage{
qgis_bufferbym(
  INPUT = qgisprocess:::qgis_default_value(),
  SEGMENTS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{SEGMENTS}{\code{number} - Segments. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Buffered. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Variable width buffer (by M value) (native:bufferbym). This algorithm creates variable width buffers along lines, using the M value of the line geometries as the diameter of the buffer at each vertex.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Buffered
}
}
}
