% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_calculateexpression.R
\name{qgis_calculateexpression}
\alias{qgis_calculateexpression}
\title{QGIS algorithm - Calculate expression}
\usage{
qgis_calculateexpression(
  INPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{string} - Input. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Calculate expression (native:calculateexpression). This algorithm calculates the result of a QGIS expression and makes it available for use in other parts of the model.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVariant - Value
}
}
}
