% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_cellstackpercentile.R
\name{qgis_cellstackpercentile}
\alias{qgis_cellstackpercentile}
\title{QGIS algorithm Cell stack percentile}
\usage{
qgis_cellstackpercentile(
  INPUT = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  PERCENTILE = qgisprocess::qgis_default_value(),
  IGNORE_NODATA = qgisprocess::qgis_default_value(),
  REFERENCE_LAYER = qgisprocess::qgis_default_value(),
  OUTPUT_NODATA_VALUE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input layers. .}

\item{METHOD}{\code{enum}  of \verb{("Nearest rank", "Inclusive linear interpolation (PERCENTILE.INC)", "Exclusive linear interpolation (PERCENTILE.EXC)")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PERCENTILE}{\code{number} - Percentile. A numeric value.}

\item{IGNORE_NODATA}{\code{boolean} - Ignore NoData values. 1 for true/yes. 0 for false/no.}

\item{REFERENCE_LAYER}{\code{raster} - Reference layer. Path to a raster layer.}

\item{OUTPUT_NODATA_VALUE}{\code{number} - Output NoData value. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Output layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Cell stack percentile (native:cellstackpercentile)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CRS_AUTHID - outputString - CRS authority identifier
\item EXTENT - outputString - Extent
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item OUTPUT - outputRaster - Output layer
\item TOTAL_PIXEL_COUNT - outputNumber - Total pixel count
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
}
}
}
