% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_cellstackpercentrankfromrasterlayer.R
\name{qgis_cellstackpercentrankfromrasterlayer}
\alias{qgis_cellstackpercentrankfromrasterlayer}
\title{QGIS algorithm Cell stack percentrank from raster layer}
\usage{
qgis_cellstackpercentrankfromrasterlayer(
  INPUT = qgisprocess::qgis_default_value(),
  INPUT_VALUE_RASTER = qgisprocess::qgis_default_value(),
  VALUE_RASTER_BAND = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  IGNORE_NODATA = qgisprocess::qgis_default_value(),
  REFERENCE_LAYER = qgisprocess::qgis_default_value(),
  OUTPUT_NODATA_VALUE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input layers. .}

\item{INPUT_VALUE_RASTER}{\code{raster} - Value raster layer. Path to a raster layer.}

\item{VALUE_RASTER_BAND}{\code{band} - Value raster band. Integer value representing an existing raster band number.}

\item{METHOD}{\code{enum}  of \verb{("Inclusive linear interpolation (PERCENTRANK.INC)", "Exclusive linear interpolation (PERCENTRANK.EXC)")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{IGNORE_NODATA}{\code{boolean} - Ignore NoData values. 1 for true/yes. 0 for false/no.}

\item{REFERENCE_LAYER}{\code{raster} - Reference layer. Path to a raster layer.}

\item{OUTPUT_NODATA_VALUE}{\code{number} - Output NoData value. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Output layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Cell stack percentrank from raster layer (native:cellstackpercentrankfromrasterlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output layer
\item EXTENT - outputString - Extent
\item CRS_AUTHID - outputString - CRS authority identifier
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item TOTAL_PIXEL_COUNT - outputNumber - Total pixel count
}
}
}
