% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_centroids.R
\name{qgis_centroids}
\alias{qgis_centroids}
\title{QGIS algorithm Centroids}
\usage{
qgis_centroids(
  INPUT = qgisprocess::qgis_default_value(),
  ALL_PARTS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{ALL_PARTS}{\code{boolean} - Create centroid for each part. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Centroids. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Centroids (native:centroids)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Centroids
}
}
}
