% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_checkgeometrycontained.R
\name{qgis_checkgeometrycontained}
\alias{qgis_checkgeometrycontained}
\title{QGIS algorithm - Features inside polygon}
\usage{
qgis_checkgeometrycontained(
  INPUT = qgisprocess:::qgis_default_value(),
  UNIQUE_ID = qgisprocess:::qgis_default_value(),
  POLYGONS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ERRORS = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{UNIQUE_ID}{\code{field} - Unique feature identifier. The name of an existing field. ; delimited list of existing field names.}

\item{POLYGONS}{\code{multilayer} - Polygon layers. .}

\item{OUTPUT}{\code{sink} - Errors from contained features. Path for new vector layer.}

\item{ERRORS}{\code{sink} - Contained features. Path for new vector layer.}

\item{TOLERANCE}{\code{number} - Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Features inside polygon (native:checkgeometrycontained). Detects features contained inside polygons from a list of polygon layers. This algorithm checks the input geometries contained in the polygons from the polygon layers list. A polygon layer can be checked against itself. Input features contained in the polygon layers features are errors.
}
\details{
\subsection{Outputs description}{
\itemize{
\item ERRORS - outputVector - Contained features
\item OUTPUT - outputVector - Errors from contained features
}
}
}
