% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_checkgeometrygap.R
\name{qgis_checkgeometrygap}
\alias{qgis_checkgeometrygap}
\title{QGIS algorithm - Small gaps}
\usage{
qgis_checkgeometrygap(
  INPUT = qgisprocess:::qgis_default_value(),
  UNIQUE_ID = qgisprocess:::qgis_default_value(),
  GAP_THRESHOLD = qgisprocess:::qgis_default_value(),
  ALLOWED_GAPS_LAYER = qgisprocess:::qgis_default_value(),
  ALLOWED_GAPS_BUFFER = qgisprocess:::qgis_default_value(),
  NEIGHBORS = qgisprocess:::qgis_default_value(),
  ERRORS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{UNIQUE_ID}{\code{field} - Unique feature identifier. The name of an existing field. ; delimited list of existing field names.}

\item{GAP_THRESHOLD}{\code{number} - Gap threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ALLOWED_GAPS_LAYER}{\code{vector} - Allowed gaps layer. Path to a vector layer.}

\item{ALLOWED_GAPS_BUFFER}{\code{distance} - Allowed gaps buffer. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NEIGHBORS}{\code{sink} - Neighbors layer. Path for new vector layer.}

\item{ERRORS}{\code{sink} - Gap errors. Path for new vector layer.}

\item{OUTPUT}{\code{sink} - Gap features. Path for new vector layer.}

\item{TOLERANCE}{\code{number} - Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Small gaps (native:checkgeometrygap). Detects gaps between polygons smaller than a given area. This algorithm checks the gaps between polygons. Gaps with an area smaller than the gap threshold are errors.  If an allowed gaps layer is given, the gaps contained in polygons from this layer will be ignored. An optional buffer can be applied to the allowed gaps.  The neighbors output layer is needed for the fix geometry (gaps) algorithm. It is a 1-N relational table for correspondance between a gap and the unique id of its neighbor features.
}
\details{
\subsection{Outputs description}{
\itemize{
\item ERRORS - outputVector - Gap errors
\item NEIGHBORS - outputVector - Neighbors layer
\item OUTPUT - outputVector - Gap features
}
}
}
