% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_checkgeometrymissingvertex.R
\name{qgis_checkgeometrymissingvertex}
\alias{qgis_checkgeometrymissingvertex}
\title{QGIS algorithm - Missing vertices along borders}
\usage{
qgis_checkgeometrymissingvertex(
  INPUT = qgisprocess:::qgis_default_value(),
  UNIQUE_ID = qgisprocess:::qgis_default_value(),
  ERRORS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{UNIQUE_ID}{\code{field} - Unique feature identifier. The name of an existing field. ; delimited list of existing field names.}

\item{ERRORS}{\code{sink} - Missing vertices errors. Path for new vector layer.}

\item{OUTPUT}{\code{sink} - Missing vertices features. Path for new vector layer.}

\item{TOLERANCE}{\code{number} - Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Missing vertices along borders (native:checkgeometrymissingvertex). Detects missing vertices along polygon borders for topology compliance. This algorithm checks for missing vertices along polygon borders. To be topologically correct, a vertex at the junction of two polygons must be present on both polygons. Missing vertices are errors.
}
\details{
\subsection{Outputs description}{
\itemize{
\item ERRORS - outputVector - Missing vertices errors
\item OUTPUT - outputVector - Missing vertices features
}
}
}
