% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_checkgeometrysliverpolygon.R
\name{qgis_checkgeometrysliverpolygon}
\alias{qgis_checkgeometrysliverpolygon}
\title{QGIS algorithm - Sliver polygons}
\usage{
qgis_checkgeometrysliverpolygon(
  INPUT = qgisprocess:::qgis_default_value(),
  UNIQUE_ID = qgisprocess:::qgis_default_value(),
  ERRORS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  MAX_THINNESS = qgisprocess:::qgis_default_value(),
  MAX_AREA = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{UNIQUE_ID}{\code{field} - Unique feature identifier. The name of an existing field. ; delimited list of existing field names.}

\item{ERRORS}{\code{sink} - Sliver polygon errors. Path for new vector layer.}

\item{OUTPUT}{\code{sink} - Sliver polygon features. Path for new vector layer.}

\item{MAX_THINNESS}{\code{number} - Maximum thinness. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAX_AREA}{\code{number} - Maximum area (map units squared). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TOLERANCE}{\code{number} - Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Sliver polygons (native:checkgeometrysliverpolygon). Detects sliver polygons that are too thin. This algorithm checks sliver polygons.  The thinness is the ratio between the area of the minimum square containing the polygon and the area of the polygon itself (a square has a thinness value of  1). The thinness value is between 1 and +infinity. If a polygon has an area higher than the maximum area, it is skipped (a maximum area value of 0 means no area check).  Polygons having a thinness higher than the maximum thinness are errors.  To fix sliver polygons, use the "Fix small polygons" algorithm.
}
\details{
\subsection{Outputs description}{
\itemize{
\item ERRORS - outputVector - Sliver polygon errors
\item OUTPUT - outputVector - Sliver polygon features
}
}
}
