% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_checkvalidity.R
\name{qgis_checkvalidity}
\alias{qgis_checkvalidity}
\title{QGIS algorithm - Check validity}
\usage{
qgis_checkvalidity(
  INPUT_LAYER = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  IGNORE_RING_SELF_INTERSECTION = qgisprocess:::qgis_default_value(),
  VALID_OUTPUT = qgisprocess:::qgis_default_value(),
  INVALID_OUTPUT = qgisprocess:::qgis_default_value(),
  ERROR_OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_LAYER}{\code{source} - Input layer. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("The one selected in digitizing settings", "QGIS", "GEOS")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{IGNORE_RING_SELF_INTERSECTION}{\code{boolean} - Ignore ring self intersections. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{VALID_OUTPUT}{\code{sink} - Valid output. Path for new vector layer.}

\item{INVALID_OUTPUT}{\code{sink} - Invalid output. Path for new vector layer.}

\item{ERROR_OUTPUT}{\code{sink} - Error output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Check validity (native:checkvalidity). Performs a validity check on the geometries of a vector layer and classifies them in three groups (valid, invalid and error). This algorithm performs a validity check on the geometries of a vector layer.  The geometries are classified in three groups (valid, invalid and error), and a vector layer is generated with the features in each of these categories.  By default the algorithm uses the strict OGC definition of polygon validity, where a polygon is marked as invalid if a self-intersecting ring causes an interior hole. If the 'Ignore ring self intersections' option is checked, then this rule will be ignored and a more lenient validity check will be performed.  The GEOS method is faster and performs better on larger geometries, but is limited to only returning the first error encountered in a geometry. The QGIS method will be slower but reports all errors encountered in the geometry, not just the first.
}
\details{
\subsection{Outputs description}{
\itemize{
\item ERROR_COUNT - outputNumber - Count of errors
\item ERROR_OUTPUT - outputVector - Error output
\item INVALID_COUNT - outputNumber - Count of invalid features
\item INVALID_OUTPUT - outputVector - Invalid output
\item VALID_COUNT - outputNumber - Count of valid features
\item VALID_OUTPUT - outputVector - Valid output
}
}
}
