% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_climbalongline.R
\name{qgis_climbalongline}
\alias{qgis_climbalongline}
\title{QGIS algorithm - Climb along line}
\usage{
qgis_climbalongline(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Line layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Climb layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Climb along line (native:climbalongline). Calculates the total climb and descent along line geometries with Z values. This algorithm calculates the total climb and descent along line geometries.  Input layer must have Z values present. If Z values are not available, the "Drape" (set Z value from raster) algorithm may be used to add Z values from a DEM layer.  The output layer is a copy of the input layer with additional fields that contain the total climb, total descent, the minimum elevation and the maximum elevation for each line geometry.If the input layer contains fields with the same names as these added fields, they will be renamed (field names will be altered to "name_2", "name_3", etc, finding the first non-duplicate name).
}
\details{
\subsection{Outputs description}{
\itemize{
\item MAXELEVATION - outputNumber - Maximum elevation
\item MINELEVATION - outputNumber - Minimum elevation
\item OUTPUT - outputVector - Climb layer
\item TOTALCLIMB - outputNumber - Total climb
\item TOTALDESCENT - outputNumber - Total descent
}
}
}
