% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_condition.R
\name{qgis_condition}
\alias{qgis_condition}
\title{QGIS algorithm - Conditional branch}
\usage{
qgis_condition(
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Conditional branch (native:condition). Adds a conditional branch into a model, allowing parts of the model to be selectively executed. This algorithm adds a conditional branch into a model, allowing parts of the model to be executed based on the result of an expression evaluation.
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
