% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_convertgeometrytype.R
\name{qgis_convertgeometrytype}
\alias{qgis_convertgeometrytype}
\title{QGIS algorithm Convert geometry type}
\usage{
qgis_convertgeometrytype(
  INPUT = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{TYPE}{\code{enum}  of \verb{("Centroids", "Nodes", "Linestrings", "Multilinestrings", "Polygons")} - New geometry type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Converted. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Convert geometry type (qgis:convertgeometrytype)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Converted
}
}
}
