% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_convertgpxfeaturetype.R
\name{qgis_convertgpxfeaturetype}
\alias{qgis_convertgpxfeaturetype}
\title{QGIS algorithm Convert GPX feature type}
\usage{
qgis_convertgpxfeaturetype(
  INPUT = qgisprocess::qgis_default_value(),
  CONVERSION = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{file} - Input file. Path to a file.}

\item{CONVERSION}{\code{enum}  of \verb{("Waypoints from a Route", "Waypoints from a Track", "Route from Waypoints", "Track from Waypoints")} - Conversion. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{fileDestination} - Output. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convert GPX feature type (native:convertgpxfeaturetype)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - Output
\item OUTPUT_LAYER - outputVector - Output layer
}
}
}
