% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_converttocurves.R
\name{qgis_converttocurves}
\alias{qgis_converttocurves}
\title{QGIS algorithm - Convert to curved geometries}
\usage{
qgis_converttocurves(
  INPUT = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  ANGLE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DISTANCE}{\code{number} - Maximum distance tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ANGLE}{\code{number} - Maximum angle tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Curves. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convert to curved geometries (native:converttocurves). This algorithm converts a geometry into its curved geometry equivalent.  Already curved geometries will be retained without change.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Curves
}
}
}
