% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_convexhull.R
\name{qgis_convexhull}
\alias{qgis_convexhull}
\title{QGIS algorithm Convex hull}
\usage{
qgis_convexhull(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Convex hulls. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convex hull (native:convexhull)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Convex hulls
}
}
}
