% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_copylayermetadata.R
\name{qgis_copylayermetadata}
\alias{qgis_copylayermetadata}
\title{QGIS algorithm - Copy layer metadata}
\usage{
qgis_copylayermetadata(
  SOURCE = qgisprocess:::qgis_default_value(),
  TARGET = qgisprocess:::qgis_default_value(),
  DEFAULT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SOURCE}{\code{layer} - Source layer. Path to a vector, raster or mesh layer.}

\item{TARGET}{\code{layer} - Target layer. Path to a vector, raster or mesh layer.}

\item{DEFAULT}{\code{boolean} - Save metadata as default. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Copy layer metadata (native:copylayermetadata). Copies the metadata from one layer to another. This algorithm copies metadata from a source layer to a target layer.  Any existing metadata in the target layer will be replaced.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputLayer - Updated layer
}
}
}
