% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_coverageunion.R
\name{qgis_coverageunion}
\alias{qgis_coverageunion}
\title{QGIS algorithm - Dissolve coverage}
\usage{
qgis_coverageunion(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Dissolved. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Dissolve coverage (native:coverageunion). Dissolves a coverage of polygon features This algorithm operates on a coverage (represented as a set of polygon features with exactly matching edge geometry) to dissolve (union) the geometries.  It provides a heavily optimized approach for unioning these features compared with the standard Dissolve tools.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Dissolved
}
}
}
