% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_coveragevalidate.R
\name{qgis_coveragevalidate}
\alias{qgis_coveragevalidate}
\title{QGIS algorithm - Validate coverage}
\usage{
qgis_coveragevalidate(
  INPUT = qgisprocess:::qgis_default_value(),
  GAP_WIDTH = qgisprocess:::qgis_default_value(),
  INVALID_EDGES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{GAP_WIDTH}{\code{distance} - Gap width. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INVALID_EDGES}{\code{sink} - Invalid edges. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Validate coverage (native:coveragevalidate). Analyzes a coverage of polygon features to find places where the assumption of exactly matching edges is not met. This algorithm analyzes a coverage (represented as a set of polygon features with exactly matching edge geometry) to find places where the assumption of exactly matching edges is not met.  Invalidity includes polygons that overlap or that have gaps smaller than the specified gap width.
}
\details{
\subsection{Outputs description}{
\itemize{
\item INVALID_EDGES - outputVector - Invalid edges
\item IS_VALID - outputBoolean - Coverage is valid
}
}
}
