% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_createattributeindex.R
\name{qgis_createattributeindex}
\alias{qgis_createattributeindex}
\title{QGIS algorithm - Create attribute index}
\usage{
qgis_createattributeindex(
  INPUT = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{vector} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute to index. The name of an existing field. ; delimited list of existing field names.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create attribute index (native:createattributeindex). Creates an index to speed up queries made against a field in a table. Support for index creation is dependent on the layer's data provider and the field type.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Indexed layer
}
}
}
