% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_createconstantrasterlayer.R
\name{qgis_createconstantrasterlayer}
\alias{qgis_createconstantrasterlayer}
\title{QGIS algorithm - Create constant raster layer}
\usage{
qgis_createconstantrasterlayer(
  EXTENT = qgisprocess:::qgis_default_value(),
  TARGET_CRS = qgisprocess:::qgis_default_value(),
  PIXEL_SIZE = qgisprocess:::qgis_default_value(),
  NUMBER = qgisprocess:::qgis_default_value(),
  OUTPUT_TYPE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{EXTENT}{\code{extent} - Desired extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{TARGET_CRS}{\code{crs} - Target CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{PIXEL_SIZE}{\code{number} - Pixel size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NUMBER}{\code{number} - Constant value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT_TYPE}{\code{enum}  of \verb{("Byte", "Integer16", "Unsigned Integer16", "Integer32", "Unsigned Integer32", "Float32", "Float64")} - Output raster data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{rasterDestination} - Constant. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create constant raster layer (native:createconstantrasterlayer). Generates raster layer for given extent and cell size filled with the specified value. Additionally an output data type can be specified. The algorithm will abort if a value has been entered that cannot be represented by the selected output raster data type.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Constant
}
}
}
