% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_createdirectory.R
\name{qgis_createdirectory}
\alias{qgis_createdirectory}
\title{QGIS algorithm Create directory}
\usage{
qgis_createdirectory(
  PATH = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{PATH}{\code{layer} - Directory path. Path to a vector, raster or mesh layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create directory (native:createdirectory)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFolder - Directory
}
}
}
