% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_createpointslayerfromtable.R
\name{qgis_createpointslayerfromtable}
\alias{qgis_createpointslayerfromtable}
\title{QGIS algorithm Create points layer from table}
\usage{
qgis_createpointslayerfromtable(
  INPUT = qgisprocess::qgis_default_value(),
  XFIELD = qgisprocess::qgis_default_value(),
  YFIELD = qgisprocess::qgis_default_value(),
  ZFIELD = qgisprocess::qgis_default_value(),
  MFIELD = qgisprocess::qgis_default_value(),
  TARGET_CRS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{XFIELD}{\code{field} - X field. The name of an existing field. ; delimited list of existing field names.}

\item{YFIELD}{\code{field} - Y field. The name of an existing field. ; delimited list of existing field names.}

\item{ZFIELD}{\code{field} - Z field. The name of an existing field. ; delimited list of existing field names.}

\item{MFIELD}{\code{field} - M field. The name of an existing field. ; delimited list of existing field names.}

\item{TARGET_CRS}{\code{crs} - Target CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT}{\code{sink} - Points from table. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create points layer from table (native:createpointslayerfromtable)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Points from table
}
}
}
