% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_createrandomnormalrasterlayer.R
\name{qgis_createrandomnormalrasterlayer}
\alias{qgis_createrandomnormalrasterlayer}
\title{QGIS algorithm Create random raster layer (normal distribution)}
\usage{
qgis_createrandomnormalrasterlayer(
  EXTENT = qgisprocess::qgis_default_value(),
  TARGET_CRS = qgisprocess::qgis_default_value(),
  PIXEL_SIZE = qgisprocess::qgis_default_value(),
  OUTPUT_TYPE = qgisprocess::qgis_default_value(),
  MEAN = qgisprocess::qgis_default_value(),
  STDDEV = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{EXTENT}{\code{extent} - Desired extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{TARGET_CRS}{\code{crs} - Target CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{PIXEL_SIZE}{\code{number} - Pixel size. A numeric value.}

\item{OUTPUT_TYPE}{\code{enum}  of \verb{("Float32", "Float64")} - Output raster data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MEAN}{\code{number} - Mean of normal distribution. A numeric value.}

\item{STDDEV}{\code{number} - Standard deviation of normal distribution. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Output raster. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create random raster layer (normal distribution) (native:createrandomnormalrasterlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output raste
}
}
}
