% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_createspatialindex.R
\name{qgis_createspatialindex}
\alias{qgis_createspatialindex}
\title{QGIS algorithm Create spatial index}
\usage{
qgis_createspatialindex(
  INPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{vector} - Input layer. Path to a vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create spatial index (native:createspatialindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Indexed layer
}
}
}
