% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_dbscanclustering.R
\name{qgis_dbscanclustering}
\alias{qgis_dbscanclustering}
\title{QGIS algorithm DBSCAN clustering}
\usage{
qgis_dbscanclustering(
  INPUT = qgisprocess::qgis_default_value(),
  MIN_SIZE = qgisprocess::qgis_default_value(),
  EPS = qgisprocess::qgis_default_value(),
  DBSCAN = qgisprocess::qgis_default_value(),
  FIELD_NAME = qgisprocess::qgis_default_value(),
  SIZE_FIELD_NAME = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{MIN_SIZE}{\code{number} - Minimum cluster size. A numeric value.}

\item{EPS}{\code{distance} - Maximum distance between clustered points. A numeric value.}

\item{DBSCAN}{\code{boolean} - Treat border points as noise (DBSCAN*). 1 for true/yes. 0 for false/no. Original algorithm parameter name: DBSCAN*.}

\item{FIELD_NAME}{\code{string} - Cluster field name. String value.}

\item{SIZE_FIELD_NAME}{\code{string} - Cluster size field name. String value.}

\item{OUTPUT}{\code{sink} - Clusters. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) DBSCAN clustering (native:dbscanclustering)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Clusters
\item NUM_CLUSTERS - outputNumber - Number of clusters
}
}
}
