% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_definecurrentprojection.R
\name{qgis_definecurrentprojection}
\alias{qgis_definecurrentprojection}
\title{QGIS algorithm - Define Shapefile projection}
\usage{
qgis_definecurrentprojection(
  INPUT = qgisprocess:::qgis_default_value(),
  CRS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{vector} - Input Shapefile. Path to a vector layer.}

\item{CRS}{\code{crs} - CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Define Shapefile projection (qgis:definecurrentprojection). Changes a Shapefile's projection to a new CRS without reprojecting features This algorithm sets an existing Shapefile's projection to the provided CRS. Contrary to the "Assign projection" algorithm, it will not output a new layer. The .prj and .qpj files associated with the Shapefile will be overwritten - or created if missing - to match the provided CRS.
}
\details{
\subsection{Outputs description}{
\itemize{
\item INPUT - outputVector - Layer with projection
}
}
}
