% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_deletecolumn.R
\name{qgis_deletecolumn}
\alias{qgis_deletecolumn}
\title{QGIS algorithm Drop field(s)}
\usage{
qgis_deletecolumn(
  INPUT = qgisprocess::qgis_default_value(),
  COLUMN = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{COLUMN}{\code{field} - Fields to drop. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Remaining fields. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Drop field(s) (native:deletecolumn)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Remaining fields
}
}
}
