% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_deleteduplicategeometries.R
\name{qgis_deleteduplicategeometries}
\alias{qgis_deleteduplicategeometries}
\title{QGIS algorithm Delete duplicate geometries}
\usage{
qgis_deleteduplicategeometries(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Cleaned. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Delete duplicate geometries (native:deleteduplicategeometries)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DUPLICATE_COUNT - outputNumber - Count of discarded duplicate records
\item OUTPUT - outputVector - Cleaned
\item RETAINED_COUNT - outputNumber - Count of retained records
}
}
}
