% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_deleteholes.R
\name{qgis_deleteholes}
\alias{qgis_deleteholes}
\title{QGIS algorithm Delete holes}
\usage{
qgis_deleteholes(
  INPUT = qgisprocess::qgis_default_value(),
  MIN_AREA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{MIN_AREA}{\code{number} - Remove holes with area less than. A numeric value.}

\item{OUTPUT}{\code{sink} - Cleaned. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Delete holes (native:deleteholes)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Cleaned
}
}
}
