% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_distancematrix.R
\name{qgis_distancematrix}
\alias{qgis_distancematrix}
\title{QGIS algorithm Distance matrix}
\usage{
qgis_distancematrix(
  INPUT = qgisprocess::qgis_default_value(),
  INPUT_FIELD = qgisprocess::qgis_default_value(),
  TARGET = qgisprocess::qgis_default_value(),
  TARGET_FIELD = qgisprocess::qgis_default_value(),
  MATRIX_TYPE = qgisprocess::qgis_default_value(),
  NEAREST_POINTS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input point layer. Path to a vector layer.}

\item{INPUT_FIELD}{\code{field} - Input unique ID field. The name of an existing field. ; delimited list of existing field names.}

\item{TARGET}{\code{source} - Target point layer. Path to a vector layer.}

\item{TARGET_FIELD}{\code{field} - Target unique ID field. The name of an existing field. ; delimited list of existing field names.}

\item{MATRIX_TYPE}{\code{enum}  of \verb{("Linear (N*k x 3) distance matrix", "Standard (N x T) distance matrix", "Summary distance matrix (mean, std. dev., min, max)")} - Output matrix type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NEAREST_POINTS}{\code{number} - Use only the nearest (k) target points. A numeric value.}

\item{OUTPUT}{\code{sink} - Distance matrix. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Distance matrix (qgis:distancematrix)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Distance matrix
}
}
}
