% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_distancetonearesthubpoints.R
\name{qgis_distancetonearesthubpoints}
\alias{qgis_distancetonearesthubpoints}
\title{QGIS algorithm Distance to nearest hub (points)}
\usage{
qgis_distancetonearesthubpoints(
  INPUT = qgisprocess:::qgis_default_value(),
  HUBS = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  UNIT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Source points layer. Path to a vector layer.}

\item{HUBS}{\code{source} - Destination hubs layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Hub layer name attribute. The name of an existing field. ; delimited list of existing field names.}

\item{UNIT}{\code{enum}  of \verb{("Meters", "Feet", "Miles", "Kilometers", "Layer units")} - Measurement unit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Hub distance. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Distance to nearest hub (points) (qgis:distancetonearesthubpoints)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Hub distance
}
}
}
