% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_downloadvectortiles.R
\name{qgis_downloadvectortiles}
\alias{qgis_downloadvectortiles}
\title{QGIS algorithm - Download vector tiles}
\usage{
qgis_downloadvectortiles(
  INPUT = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  MAX_ZOOM = qgisprocess:::qgis_default_value(),
  TILE_LIMIT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{layer} - Input layer. Path to a vector, raster or mesh layer.}

\item{EXTENT}{\code{extent} - Extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{MAX_ZOOM}{\code{number} - Maximum zoom level to download. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TILE_LIMIT}{\code{number} - Tile limit. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{vectorTileDestination} - Output. Path for new vector tile layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Download vector tiles (native:downloadvectortiles). Downloads vector tiles of the input vector tile layer and saves them in the local vector tile file.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVectorTile - Output
}
}
}
