% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_dropgeometries.R
\name{qgis_dropgeometries}
\alias{qgis_dropgeometries}
\title{QGIS algorithm - Drop geometries}
\usage{
qgis_dropgeometries(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Dropped geometries. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Drop geometries (native:dropgeometries). Removes any geometries from an input layer and returns a layer containing only the feature attributes. This algorithm removes any geometries from an input layer and returns a layer containing only the feature attributes.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Dropped geometries
}
}
}
