% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_dropmzvalues.R
\name{qgis_dropmzvalues}
\alias{qgis_dropmzvalues}
\title{QGIS algorithm Drop M/Z values}
\usage{
qgis_dropmzvalues(
  INPUT = qgisprocess::qgis_default_value(),
  DROP_M_VALUES = qgisprocess::qgis_default_value(),
  DROP_Z_VALUES = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DROP_M_VALUES}{\code{boolean} - Drop M Values. 1 for true/yes. 0 for false/no.}

\item{DROP_Z_VALUES}{\code{boolean} - Drop Z Values. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Z/M Dropped. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Drop M/Z values (native:dropmzvalues)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Z
}
}
}
