% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_dtmslopebasedfilter.R
\name{qgis_dtmslopebasedfilter}
\alias{qgis_dtmslopebasedfilter}
\title{QGIS algorithm - DTM filter (slope-based)}
\usage{
qgis_dtmslopebasedfilter(
  INPUT = qgisprocess:::qgis_default_value(),
  BAND = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  TERRAIN_SLOPE = qgisprocess:::qgis_default_value(),
  FILTER_MODIFICATION = qgisprocess:::qgis_default_value(),
  STANDARD_DEVIATION = qgisprocess:::qgis_default_value(),
  CREATION_OPTIONS = qgisprocess:::qgis_default_value(),
  OUTPUT_GROUND = qgisprocess:::qgis_default_value(),
  OUTPUT_NONGROUND = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{RADIUS}{\code{number} - Kernel radius (pixels). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TERRAIN_SLOPE}{\code{number} - Terrain slope (\%, pixel size\code{/}vertical units). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FILTER_MODIFICATION}{\code{enum}  of \verb{("None", "Relax filter", "Amplify")} - Filter modification. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{STANDARD_DEVIATION}{\code{number} - Standard deviation. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CREATION_OPTIONS}{\code{string} - Creation options. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT_GROUND}{\code{rasterDestination} - Output layer (ground). Path for new raster layer.}

\item{OUTPUT_NONGROUND}{\code{rasterDestination} - Output layer (non-ground objects). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) DTM filter (slope-based) (native:dtmslopebasedfilter). Filters a Digital Elevation Model in order to classify its cells into ground and object (non-ground) cells. This algorithm can be used to filter a Digital Elevation Model in order to classify its cells into ground and object (non-ground) cells.  The tool uses concepts as described by Vosselman (2000) and is based on the assumption that a large height difference between two nearby cells is unlikely to be caused by a steep slope in the terrain. The probability that the higher cell might be non-ground increases when the distance between the two cells decreases. Therefore the filter defines a maximum height difference (idz_max/i) between two cells as a function of the distance (id/i) between the cells (idz_max( d ) = d/i).  A cell is classified as terrain if there is no cell within the kernel radius to which the height difference is larger than the allowed maximum height difference at the distance between these two cells.  The approximate terrain slope (is/i) parameter is used to modify the filter function to match the overall slope in the study area (idz_max( d ) = d * s/i).  A 5 \% confidence interval (ici = 1.65 * sqrt( 2 * stddev )/i) may be used to modify the filter function even further by either relaxing (idz_max( d ) = d * s + ci/i) or amplifying (idz_max( d ) = d * s - ci/i) the filter criterium.  References: Vosselman, G. (2000): Slope based filtering of laser altimetry data. IAPRS, Vol. XXXIII, Part B3, Amsterdam, The Netherlands, 935-942  This algorithm is a port of the SAGA 'DTM Filter (slope-based)' tool.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_GROUND - outputRaster - Output layer (ground)
\item OUTPUT_NONGROUND - outputRaster - Output layer (non-ground objects)
}
}
}
