% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_dxfexport.R
\name{qgis_dxfexport}
\alias{qgis_dxfexport}
\title{QGIS algorithm Export layers to DXF}
\usage{
qgis_dxfexport(
  LAYERS = qgisprocess::qgis_default_value(),
  SYMBOLOGY_MODE = qgisprocess::qgis_default_value(),
  SYMBOLOGY_SCALE = qgisprocess::qgis_default_value(),
  ENCODING = qgisprocess::qgis_default_value(),
  CRS = qgisprocess::qgis_default_value(),
  USE_LAYER_TITLE = qgisprocess::qgis_default_value(),
  FORCE_2D = qgisprocess::qgis_default_value(),
  MTEXT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{LAYERS}{\code{dxflayers} - Input layers. .}

\item{SYMBOLOGY_MODE}{\code{enum}  of \verb{("No Symbology", "Feature Symbology", "Symbol Layer Symbology")} - Symbology mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SYMBOLOGY_SCALE}{\code{scale} - Symbology scale. A numeric value representing the scale denominator.}

\item{ENCODING}{\code{enum}  of \verb{("ISO-8859-1", "ISO-8859-2", "ISO-8859-3", "ISO-8859-4", "ISO-8859-5", "ISO-8859-6", "ISO-8859-7", "ISO-8859-8", "ISO-8859-9", "Shift_JIS", "Big5", "CP936", "GB2312")} - Encoding. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CRS}{\code{crs} - CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{USE_LAYER_TITLE}{\code{boolean} - Use layer title as name. 1 for true/yes. 0 for false/no.}

\item{FORCE_2D}{\code{boolean} - Force 2D output. 1 for true/yes. 0 for false/no.}

\item{MTEXT}{\code{boolean} - Export labels as MTEXT elements. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{fileDestination} - DXF. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Export layers to DXF (native:dxfexport)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - DX
}
}
}
