% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_eliminateselectedpolygons.R
\name{qgis_eliminateselectedpolygons}
\alias{qgis_eliminateselectedpolygons}
\title{QGIS algorithm Eliminate selected polygons}
\usage{
qgis_eliminateselectedpolygons(
  INPUT = qgisprocess::qgis_default_value(),
  MODE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{vector} - Input layer. Path to a vector layer.}

\item{MODE}{\code{enum}  of \verb{("Largest Area", "Smallest Area", "Largest Common Boundary")} - Merge selection with the neighbouring polygon with the. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Eliminated. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Eliminate selected polygons (qgis:eliminateselectedpolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Eliminated
}
}
}
