% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_exportlayersinformation.R
\name{qgis_exportlayersinformation}
\alias{qgis_exportlayersinformation}
\title{QGIS algorithm Export layer(s) information}
\usage{
qgis_exportlayersinformation(
  LAYERS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{LAYERS}{\code{multilayer} - Input layer(s). .}

\item{OUTPUT}{\code{sink} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Export layer(s) information (native:exportlayersinformation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output
}
}
}
