% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_exportmeshongrid.R
\name{qgis_exportmeshongrid}
\alias{qgis_exportmeshongrid}
\title{QGIS algorithm Export mesh on grid}
\usage{
qgis_exportmeshongrid(
  INPUT = qgisprocess:::qgis_default_value(),
  DATASET_GROUPS = qgisprocess:::qgis_default_value(),
  DATASET_TIME = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  GRID_SPACING = qgisprocess:::qgis_default_value(),
  CRS_OUTPUT = qgisprocess:::qgis_default_value(),
  VECTOR_OPTION = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{mesh} - Input mesh layer. Path to a mesh layer.}

\item{DATASET_GROUPS}{\code{meshdatasetgroups} - Dataset groups. .}

\item{DATASET_TIME}{\code{meshdatasettime} - Dataset time. .}

\item{EXTENT}{\code{extent} - Extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRID_SPACING}{\code{distance} - Grid spacing. A numeric value.}

\item{CRS_OUTPUT}{\code{crs} - Output coordinate system. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{VECTOR_OPTION}{\code{enum}  of \verb{("Cartesian (x,y)", "Polar (magnitude,degree)", "Cartesian and Polar")} - Export vector option. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Output vector layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Export mesh on grid (native:exportmeshongrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output vector layer
}
}
}
