% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_extractbyexpression.R
\name{qgis_extractbyexpression}
\alias{qgis_extractbyexpression}
\title{QGIS algorithm Extract by expression}
\usage{
qgis_extractbyexpression(
  INPUT = qgisprocess:::qgis_default_value(),
  EXPRESSION = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  FAIL_OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{EXPRESSION}{\code{expression} - Expression. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{OUTPUT}{\code{sink} - Matching features. Path for new vector layer.}

\item{FAIL_OUTPUT}{\code{sink} - Non-matching. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Extract by expression (native:extractbyexpression)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FAIL_OUTPUT - outputVector - Non-matching
\item OUTPUT - outputVector - Matching features
}
}
}
