% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_extractbylocation.R
\name{qgis_extractbylocation}
\alias{qgis_extractbylocation}
\title{QGIS algorithm - Extract by location}
\usage{
qgis_extractbylocation(
  INPUT = qgisprocess:::qgis_default_value(),
  PREDICATE = qgisprocess:::qgis_default_value(),
  INTERSECT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Extract features from. Path to a vector layer.}

\item{PREDICATE}{\code{enum}  of \verb{("intersect", "contain", "disjoint", "equal", "touch", "overlap", "are within", "cross")} - Where the features (geometric predicate). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{INTERSECT}{\code{source} - By comparing to the features from. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Extracted (location). Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Extract by location (native:extractbylocation). This algorithm creates a new vector layer that only contains matching features from an input layer. The criteria for adding features to the resulting layer is defined based on the spatial relationship between each feature and the features in an additional layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Extracted (location)
}
}
}
