% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_extractwithindistance.R
\name{qgis_extractwithindistance}
\alias{qgis_extractwithindistance}
\title{QGIS algorithm Extract within distance}
\usage{
qgis_extractwithindistance(
  INPUT = qgisprocess::qgis_default_value(),
  REFERENCE = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Extract features from. Path to a vector layer.}

\item{REFERENCE}{\code{source} - By comparing to the features from. Path to a vector layer.}

\item{DISTANCE}{\code{distance} - Where the features are within. A numeric value.}

\item{OUTPUT}{\code{sink} - Extracted (location). Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Extract within distance (native:extractwithindistance)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Extracted (location)
}
}
}
