% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_filedownloader.R
\name{qgis_filedownloader}
\alias{qgis_filedownloader}
\title{QGIS algorithm Download file}
\usage{
qgis_filedownloader(
  URL = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{URL}{\code{string} - URL. String value.}

\item{OUTPUT}{\code{fileDestination} - File destination. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Download file (native:filedownloader)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - File destination
}
}
}
