% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_fillsinkswangliu.R
\name{qgis_fillsinkswangliu}
\alias{qgis_fillsinkswangliu}
\title{QGIS algorithm - Fill sinks (Wang & Liu)}
\usage{
qgis_fillsinkswangliu(
  INPUT = qgisprocess:::qgis_default_value(),
  BAND = qgisprocess:::qgis_default_value(),
  MIN_SLOPE = qgisprocess:::qgis_default_value(),
  CREATION_OPTIONS = qgisprocess:::qgis_default_value(),
  OUTPUT_FILLED_DEM = qgisprocess:::qgis_default_value(),
  OUTPUT_FLOW_DIRECTIONS = qgisprocess:::qgis_default_value(),
  OUTPUT_WATERSHED_BASINS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{MIN_SLOPE}{\code{number} - Minimum slope (degrees). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CREATION_OPTIONS}{\code{string} - Creation options. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT_FILLED_DEM}{\code{rasterDestination} - Output layer (filled DEM). Path for new raster layer.}

\item{OUTPUT_FLOW_DIRECTIONS}{\code{rasterDestination} - Output layer (flow directions). Path for new raster layer.}

\item{OUTPUT_WATERSHED_BASINS}{\code{rasterDestination} - Output layer (watershed basins). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Fill sinks (Wang & Liu) (native:fillsinkswangliu). Identifies and fills surface depressions in digital elevation models using a method proposed by Wang & Liu. This algorithm uses a method proposed by Wang & Liu to identify and fill surface depressions in digital elevation models.  The method was enhanced to allow the creation of hydrologically sound elevation models, i.e. not only to fill the depression(s) but also to preserve a downward slope along the flow path. If desired, this is accomplished by preserving a minimum slope gradient (and thus elevation difference) between cells.  References: Wang, L. & H. Liu (2006): An efficient method for identifying and filling surface depressions in digital elevation models for hydrologic analysis and modelling. International Journal of Geographical Information Science, Vol. 20, No. 2: 193-213.  This algorithm is a port of the SAGA 'Fill Sinks (Wang & Liu)' tool.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_FILLED_DEM - outputRaster - Output layer (filled DEM)
\item OUTPUT_FLOW_DIRECTIONS - outputRaster - Output layer (flow directions)
\item OUTPUT_WATERSHED_BASINS - outputRaster - Output layer (watershed basins)
}
}
}
