% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_filter.R
\name{qgis_filter}
\alias{qgis_filter}
\title{QGIS algorithm - Feature filter}
\usage{
qgis_filter(
  INPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Feature filter (native:filter). Filters features from the input layer and redirects them to one or more outputs. This algorithm filters features from the input layer and redirects them to one or more outputs.
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
