% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_filterlayersbytype.R
\name{qgis_filterlayersbytype}
\alias{qgis_filterlayersbytype}
\title{QGIS algorithm Filter layers by type}
\usage{
qgis_filterlayersbytype(
  INPUT = qgisprocess::qgis_default_value(),
  VECTOR = qgisprocess::qgis_default_value(),
  RASTER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{layer} - Input layer. Path to a vector, raster or mesh layer.}

\item{VECTOR}{\code{vectorDestination} - Vector features. Path for new vector layer.}

\item{RASTER}{\code{rasterDestination} - Raster layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Filter layers by type (native:filterlayersbytype)
}
\details{
\subsection{Outputs description}{
\itemize{
\item VECTOR - outputVector - Vector features
\item RASTER - outputRaster - Raster layer
}
}
}
