% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_filterverticesbyz.R
\name{qgis_filterverticesbyz}
\alias{qgis_filterverticesbyz}
\title{QGIS algorithm Filter vertices by Z value}
\usage{
qgis_filterverticesbyz(
  INPUT = qgisprocess::qgis_default_value(),
  MIN = qgisprocess::qgis_default_value(),
  MAX = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{MIN}{\code{number} - Minimum. A numeric value.}

\item{MAX}{\code{number} - Maximum. A numeric value.}

\item{OUTPUT}{\code{sink} - Filtered. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Filter vertices by Z value (native:filterverticesbyz)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Filtered
}
}
}
