% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_findprojection.R
\name{qgis_findprojection}
\alias{qgis_findprojection}
\title{QGIS algorithm - Find projection}
\usage{
qgis_findprojection(
  INPUT = qgisprocess:::qgis_default_value(),
  TARGET_AREA = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{TARGET_AREA}{\code{extent} - Target area for layer. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{OUTPUT}{\code{sink} - CRS candidates. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Find projection (qgis:findprojection). This algorithm allows creation of a shortlist of possible candidate coordinate reference systems for a layer with an unknown projection. The expected area which the layer should reside in must be specified via the target area parameter. The algorithm operates by testing the layer's extent in every known reference system and listing any in which the bounds would fall near the target area if the layer was in this projection.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - CRS candidates
}
}
}
