% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_fixgeometries.R
\name{qgis_fixgeometries}
\alias{qgis_fixgeometries}
\title{QGIS algorithm - Fix geometries}
\usage{
qgis_fixgeometries(
  INPUT = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("Linework", "Structure")} - Repair method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Fixed geometries. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Fix geometries (native:fixgeometries). This algorithm attempts to create a valid representation of a given invalid geometry without losing any of the input vertices. Already-valid geometries are returned without further intervention. Always outputs multi-geometry layer.  NOTE: M values will be dropped from the output.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Fixed geometries
}
}
}
