% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_fixgeometryarea.R
\name{qgis_fixgeometryarea}
\alias{qgis_fixgeometryarea}
\title{QGIS algorithm - Fix small polygons}
\usage{
qgis_fixgeometryarea(
  INPUT = qgisprocess:::qgis_default_value(),
  ERRORS = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  MERGE_ATTRIBUTE = qgisprocess:::qgis_default_value(),
  UNIQUE_ID = qgisprocess:::qgis_default_value(),
  PART_IDX = qgisprocess:::qgis_default_value(),
  RING_IDX = qgisprocess:::qgis_default_value(),
  VERTEX_IDX = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  REPORT = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{ERRORS}{\code{source} - Error layer. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("Merge with neighboring polygon with longest shared edge", "Merge with neighboring polygon with largest area", "Merge with neighboring polygon with identical attribute value, if any, or leave as is")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MERGE_ATTRIBUTE}{\code{field} - Field to consider when merging polygons with the identical attribute method. The name of an existing field. ; delimited list of existing field names.}

\item{UNIQUE_ID}{\code{field} - Field of original feature unique identifier. The name of an existing field. ; delimited list of existing field names.}

\item{PART_IDX}{\code{field} - Field of part index. The name of an existing field. ; delimited list of existing field names.}

\item{RING_IDX}{\code{field} - Field of ring index. The name of an existing field. ; delimited list of existing field names.}

\item{VERTEX_IDX}{\code{field} - Field of vertex index. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Small polygons merged layer. Path for new vector layer.}

\item{REPORT}{\code{sink} - Report layer from merging small polygons. Path for new vector layer.}

\item{TOLERANCE}{\code{number} - Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Fix small polygons (native:fixgeometryarea). Merges small polygons detected with the "Small polygons" algorithm from the "Check geometry" section. This algorithm merges neighboring polygons according to the chosen method, based on an error layer from the "Small polygons" algorithm in the "Check geometry" section.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Small polygons merged layer
\item REPORT - outputVector - Report layer from merging small polygons
}
}
}
