% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_fixgeometrydeletefeatures.R
\name{qgis_fixgeometrydeletefeatures}
\alias{qgis_fixgeometrydeletefeatures}
\title{QGIS algorithm - Delete features}
\usage{
qgis_fixgeometrydeletefeatures(
  INPUT = qgisprocess:::qgis_default_value(),
  ERRORS = qgisprocess:::qgis_default_value(),
  UNIQUE_ID = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  REPORT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{ERRORS}{\code{source} - Error layer. Path to a vector layer.}

\item{UNIQUE_ID}{\code{field} - Field of original feature unique identifier. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Cleaned layer. Path for new vector layer.}

\item{REPORT}{\code{sink} - Report layer from deleting features. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Delete features (native:fixgeometrydeletefeatures). Deletes features detected with an algorithm from the "Check geometry" section. This algorithm deletes error features listed in the errors layer from an algorithm in the "Check geometry" section. The required inputs are the original layer used in the check algorithm, its unique id field, and its corresponding errors layer.  For instance, it can be used after the following check algorithms to delete error features:htmlulliFeature inside polygon/liliDegenerate polygons/liliSmall segments/liliDuplicated geometries/lilietc./li/ul/html
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Cleaned layer
\item REPORT - outputVector - Report layer from deleting features
}
}
}
