% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_fixgeometrygap.R
\name{qgis_fixgeometrygap}
\alias{qgis_fixgeometrygap}
\title{QGIS algorithm - Fill gaps}
\usage{
qgis_fixgeometrygap(
  INPUT = qgisprocess:::qgis_default_value(),
  NEIGHBORS = qgisprocess:::qgis_default_value(),
  GAPS = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  UNIQUE_ID = qgisprocess:::qgis_default_value(),
  ERROR_ID_IDX = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  REPORT = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{NEIGHBORS}{\code{source} - Neighbors layer. Path to a vector layer.}

\item{GAPS}{\code{source} - Gaps layer. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("Add to longest shared edge", "Create new feature", "Add to largest neighbouring area")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{UNIQUE_ID}{\code{field} - Field of original feature unique identifier. The name of an existing field. ; delimited list of existing field names.}

\item{ERROR_ID_IDX}{\code{field} - Field of error id. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Gaps-filled layer. Path for new vector layer.}

\item{REPORT}{\code{sink} - Report layer from fixing gaps. Path for new vector layer.}

\item{TOLERANCE}{\code{number} - Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Fill gaps (native:fixgeometrygap). Fills gaps detected with the "Small gaps" algorithm from the "Check geometry" section. This algorithm fills the gaps based on a gap and neighbors layer from the "Small gaps" algorithm in the "Check geometry" section.  3 different fixing methods are available, which will give different results.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Gaps-filled layer
\item REPORT - outputVector - Report layer from fixing gaps
}
}
}
