% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_fuzzifyrasterpowermembership.R
\name{qgis_fuzzifyrasterpowermembership}
\alias{qgis_fuzzifyrasterpowermembership}
\title{QGIS algorithm Fuzzify raster (power membership)}
\usage{
qgis_fuzzifyrasterpowermembership(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  FUZZYLOWBOUND = qgisprocess::qgis_default_value(),
  FUZZYHIGHBOUND = qgisprocess::qgis_default_value(),
  FUZZYEXPONENT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input Raster. Path to a raster layer.}

\item{BAND}{\code{band} - Band Number. Integer value representing an existing raster band number.}

\item{FUZZYLOWBOUND}{\code{number} - Low fuzzy membership bound. A numeric value.}

\item{FUZZYHIGHBOUND}{\code{number} - High fuzzy membership bound. A numeric value.}

\item{FUZZYEXPONENT}{\code{number} - Membership function exponent. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Fuzzified raster. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Fuzzify raster (power membership) (native:fuzzifyrasterpowermembership)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CRS_AUTHID - outputString - CRS authority identifier
\item EXTENT - outputString - Extent
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item OUTPUT - outputRaster - Fuzzified raster
\item TOTAL_PIXEL_COUNT - outputNumber - Total pixel count
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
}
}
}
