% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_generateelevationprofileimage.R
\name{qgis_generateelevationprofileimage}
\alias{qgis_generateelevationprofileimage}
\title{QGIS algorithm - Generate elevation profile image}
\usage{
qgis_generateelevationprofileimage(
  CURVE = qgisprocess:::qgis_default_value(),
  MAP_LAYERS = qgisprocess:::qgis_default_value(),
  WIDTH = qgisprocess:::qgis_default_value(),
  HEIGHT = qgisprocess:::qgis_default_value(),
  TERRAIN_LAYER = qgisprocess:::qgis_default_value(),
  MINIMUM_DISTANCE = qgisprocess:::qgis_default_value(),
  MAXIMUM_DISTANCE = qgisprocess:::qgis_default_value(),
  MINIMUM_ELEVATION = qgisprocess:::qgis_default_value(),
  MAXIMUM_ELEVATION = qgisprocess:::qgis_default_value(),
  TEXT_COLOR = qgisprocess:::qgis_default_value(),
  BACKGROUND_COLOR = qgisprocess:::qgis_default_value(),
  BORDER_COLOR = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  DPI = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CURVE}{\code{geometry} - Profile curve. Well-Known Text string (WKT).}

\item{MAP_LAYERS}{\code{multilayer} - Map layers. .}

\item{WIDTH}{\code{number} - Chart width (in pixels). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{HEIGHT}{\code{number} - Chart height (in pixels). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TERRAIN_LAYER}{\code{layer} - Terrain layer. Path to a vector, raster or mesh layer.}

\item{MINIMUM_DISTANCE}{\code{number} - Chart minimum distance (X axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAXIMUM_DISTANCE}{\code{number} - Chart maximum distance (X axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MINIMUM_ELEVATION}{\code{number} - Chart minimum elevation (Y axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAXIMUM_ELEVATION}{\code{number} - Chart maximum elevation (Y axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TEXT_COLOR}{\code{color} - Chart text color. String representation of color, e.g #ff0000 or rgba(200,100,50,0.8). field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{BACKGROUND_COLOR}{\code{color} - Chart background color. String representation of color, e.g #ff0000 or rgba(200,100,50,0.8). field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{BORDER_COLOR}{\code{color} - Chart border color. String representation of color, e.g #ff0000 or rgba(200,100,50,0.8). field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TOLERANCE}{\code{number} - Profile tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DPI}{\code{number} - Chart DPI. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{fileDestination} - Output image. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Generate elevation profile image (native:generateelevationprofileimage). Creates an elevation profile image from a list of map layer and an optional terrain. This algorithm creates an elevation profile image from a list of map layer and an optional terrain.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - Output image
}
}
}
