% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_generatepointspixelcentroidsalongline.R
\name{qgis_generatepointspixelcentroidsalongline}
\alias{qgis_generatepointspixelcentroidsalongline}
\title{QGIS algorithm - Generate points (pixel centroids) along line}
\usage{
qgis_generatepointspixelcentroidsalongline(
  INPUT_RASTER = qgisprocess:::qgis_default_value(),
  INPUT_VECTOR = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_RASTER}{\code{raster} - Raster layer. Path to a raster layer.}

\item{INPUT_VECTOR}{\code{source} - Vector layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Points along lines. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Generate points (pixel centroids) along line (qgis:generatepointspixelcentroidsalongline). This algorithm generates a point vector layer from an input raster and line layer. The points correspond to the pixel centroids that intersect the line layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Points along lines
}
}
}
