% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/qgis_generatepointspixelcentroidsinsidepolygons.R
\name{qgis_generatepointspixelcentroidsinsidepolygons}
\alias{qgis_generatepointspixelcentroidsinsidepolygons}
\title{QGIS algorithm Generate points (pixel centroids) inside polygons}
\usage{
qgis_generatepointspixelcentroidsinsidepolygons(
  INPUT_RASTER = qgisprocess::qgis_default_value(),
  INPUT_VECTOR = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT_RASTER}{\code{raster} - Raster layer. Path to a raster layer.}

\item{INPUT_VECTOR}{\code{source} - Vector layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Pixel centroids. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Generate points (pixel centroids) inside polygons (native:generatepointspixelcentroidsinsidepolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Pixel centroids
}
}
}
