% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_geometrybyexpression.R
\name{qgis_geometrybyexpression}
\alias{qgis_geometrybyexpression}
\title{QGIS algorithm Geometry by expression}
\usage{
qgis_geometrybyexpression(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT_GEOMETRY = qgisprocess::qgis_default_value(),
  WITH_Z = qgisprocess::qgis_default_value(),
  WITH_M = qgisprocess::qgis_default_value(),
  EXPRESSION = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{OUTPUT_GEOMETRY}{\code{enum}  of \verb{("Polygon", "Line", "Point")} - Output geometry type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{WITH_Z}{\code{boolean} - Output geometry has z dimension. 1 for true/yes. 0 for false/no.}

\item{WITH_M}{\code{boolean} - Output geometry has m values. 1 for true/yes. 0 for false/no.}

\item{EXPRESSION}{\code{expression} - Geometry expression. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{OUTPUT}{\code{sink} - Modified geometry. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Geometry by expression (native:geometrybyexpression)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Modified geometry
}
}
}
