% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_gltftovector.R
\name{qgis_gltftovector}
\alias{qgis_gltftovector}
\title{QGIS algorithm - Convert GLTF to vector features}
\usage{
qgis_gltftovector(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT_POLYGONS = qgisprocess:::qgis_default_value(),
  OUTPUT_LINES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{file} - Input GLTF. Path to a file.}

\item{OUTPUT_POLYGONS}{\code{sink} - Output polygons. Path for new vector layer.}

\item{OUTPUT_LINES}{\code{sink} - Output lines. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convert GLTF to vector features (native:gltftovector). Converts GLTF content to standard vector layer formats. This algorithm converts GLTF content to standard vector layer formats.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_LINES - outputVector - Output lines
\item OUTPUT_POLYGONS - outputVector - Output polygons
}
}
}
