% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_greaterthanfrequency.R
\name{qgis_greaterthanfrequency}
\alias{qgis_greaterthanfrequency}
\title{QGIS algorithm Greater than frequency}
\usage{
qgis_greaterthanfrequency(
  INPUT_VALUE_RASTER = qgisprocess::qgis_default_value(),
  INPUT_VALUE_RASTER_BAND = qgisprocess::qgis_default_value(),
  INPUT_RASTERS = qgisprocess::qgis_default_value(),
  IGNORE_NODATA = qgisprocess::qgis_default_value(),
  OUTPUT_NODATA_VALUE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT_VALUE_RASTER}{\code{raster} - Input value raster. Path to a raster layer.}

\item{INPUT_VALUE_RASTER_BAND}{\code{band} - Value raster band. Integer value representing an existing raster band number.}

\item{INPUT_RASTERS}{\code{multilayer} - Input raster layers. .}

\item{IGNORE_NODATA}{\code{boolean} - Ignore NoData values. 1 for true/yes. 0 for false/no.}

\item{OUTPUT_NODATA_VALUE}{\code{number} - Output NoData value. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Output layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Greater than frequency (native:greaterthanfrequency)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output layer
\item OCCURRENCE_COUNT - outputNumber - Count of value occurrences
\item FOUND_LOCATIONS_COUNT - outputNumber - Count of cells with equal value occurrences
\item MEAN_FREQUENCY_PER_LOCATION - outputNumber - Mean frequency at valid cell locations
\item EXTENT - outputString - Extent
\item CRS_AUTHID - outputString - CRS authority identifier
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item TOTAL_PIXEL_COUNT - outputNumber - Total pixel count
}
}
}
