% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_heatmapkerneldensityestimation.R
\name{qgis_heatmapkerneldensityestimation}
\alias{qgis_heatmapkerneldensityestimation}
\title{QGIS algorithm Heatmap (Kernel Density Estimation)}
\usage{
qgis_heatmapkerneldensityestimation(
  INPUT = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  RADIUS_FIELD = qgisprocess::qgis_default_value(),
  PIXEL_SIZE = qgisprocess::qgis_default_value(),
  WEIGHT_FIELD = qgisprocess::qgis_default_value(),
  KERNEL = qgisprocess::qgis_default_value(),
  DECAY = qgisprocess::qgis_default_value(),
  OUTPUT_VALUE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Point layer. Path to a vector layer.}

\item{RADIUS}{\code{distance} - Radius. A numeric value.}

\item{RADIUS_FIELD}{\code{field} - Radius from field. The name of an existing field. ; delimited list of existing field names.}

\item{PIXEL_SIZE}{\code{number} - Output raster size. A numeric value.}

\item{WEIGHT_FIELD}{\code{field} - Weight from field. The name of an existing field. ; delimited list of existing field names.}

\item{KERNEL}{\code{enum}  of \verb{("Quartic", "Triangular", "Uniform", "Triweight", "Epanechnikov")} - Kernel shape. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DECAY}{\code{number} - Decay ratio (Triangular kernels only). A numeric value.}

\item{OUTPUT_VALUE}{\code{enum}  of \verb{("Raw", "Scaled")} - Output value scaling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{rasterDestination} - Heatmap. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Heatmap (Kernel Density Estimation) (qgis:heatmapkerneldensityestimation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Heatmap
}
}
}
