% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_hillshade.R
\name{qgis_hillshade}
\alias{qgis_hillshade}
\title{QGIS algorithm Hillshade}
\usage{
qgis_hillshade(
  INPUT = qgisprocess::qgis_default_value(),
  Z_FACTOR = qgisprocess::qgis_default_value(),
  AZIMUTH = qgisprocess::qgis_default_value(),
  V_ANGLE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Elevation layer. Path to a raster layer.}

\item{Z_FACTOR}{\code{number} - Z factor. A numeric value.}

\item{AZIMUTH}{\code{number} - Azimuth (horizontal angle). A numeric value.}

\item{V_ANGLE}{\code{number} - Vertical angle. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Hillshade. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Hillshade (native:hillshade)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Hillshade
}
}
}
