% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_hypsometriccurves.R
\name{qgis_hypsometriccurves}
\alias{qgis_hypsometriccurves}
\title{QGIS algorithm Hypsometric curves}
\usage{
qgis_hypsometriccurves(
  INPUT_DEM = qgisprocess:::qgis_default_value(),
  BOUNDARY_LAYER = qgisprocess:::qgis_default_value(),
  STEP = qgisprocess:::qgis_default_value(),
  USE_PERCENTAGE = qgisprocess:::qgis_default_value(),
  OUTPUT_DIRECTORY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_DEM}{\code{raster} - DEM to analyze. Path to a raster layer.}

\item{BOUNDARY_LAYER}{\code{source} - Boundary layer. Path to a vector layer.}

\item{STEP}{\code{number} - Step. A numeric value.}

\item{USE_PERCENTAGE}{\code{boolean} - Use \% of area instead of absolute value. 1 for true/yes. 0 for false/no.}

\item{OUTPUT_DIRECTORY}{\code{folderDestination} - Hypsometric curves. Path for an existing or new folder.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Hypsometric curves (qgis:hypsometriccurves)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_DIRECTORY - outputFolder - Hypsometric curves
}
}
}
