% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_idwinterpolation.R
\name{qgis_idwinterpolation}
\alias{qgis_idwinterpolation}
\title{QGIS algorithm IDW interpolation}
\usage{
qgis_idwinterpolation(
  INTERPOLATION_DATA = qgisprocess::qgis_default_value(),
  DISTANCE_COEFFICIENT = qgisprocess::qgis_default_value(),
  EXTENT = qgisprocess::qgis_default_value(),
  PIXEL_SIZE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INTERPOLATION_DATA}{\code{idw_interpolation_data} - Input layer(s). .}

\item{DISTANCE_COEFFICIENT}{\code{number} - Distance coefficient P. A numeric value.}

\item{EXTENT}{\code{extent} - Extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{PIXEL_SIZE}{\code{number} - Output raster size. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Interpolated. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS IDW interpolation (qgis:idwinterpolation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Interpolated
}
}
}
