% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_importintospatialite.R
\name{qgis_importintospatialite}
\alias{qgis_importintospatialite}
\title{QGIS algorithm - Export to SpatiaLite}
\usage{
qgis_importintospatialite(
  INPUT = qgisprocess:::qgis_default_value(),
  DATABASE = qgisprocess:::qgis_default_value(),
  TABLENAME = qgisprocess:::qgis_default_value(),
  PRIMARY_KEY = qgisprocess:::qgis_default_value(),
  GEOMETRY_COLUMN = qgisprocess:::qgis_default_value(),
  ENCODING = qgisprocess:::qgis_default_value(),
  OVERWRITE = qgisprocess:::qgis_default_value(),
  CREATEINDEX = qgisprocess:::qgis_default_value(),
  LOWERCASE_NAMES = qgisprocess:::qgis_default_value(),
  DROP_STRING_LENGTH = qgisprocess:::qgis_default_value(),
  FORCE_SINGLEPART = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Layer to import. Path to a vector layer.}

\item{DATABASE}{\code{vector} - File database. Path to a vector layer.}

\item{TABLENAME}{\code{string} - Table to import to (leave blank to use layer name). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PRIMARY_KEY}{\code{field} - Primary key field. The name of an existing field. ; delimited list of existing field names.}

\item{GEOMETRY_COLUMN}{\code{string} - Geometry column. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ENCODING}{\code{string} - Encoding. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OVERWRITE}{\code{boolean} - Overwrite. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CREATEINDEX}{\code{boolean} - Create spatial index. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LOWERCASE_NAMES}{\code{boolean} - Convert field names to lowercase. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DROP_STRING_LENGTH}{\code{boolean} - Drop length constraints on character fields. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FORCE_SINGLEPART}{\code{boolean} - Create single-part geometries instead of multi-part. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Export to SpatiaLite (qgis:importintospatialite). Exports a vector layer to a SpatiaLite database. This algorithm imports a vector layer into a SpatiaLite database, creating a new table.
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
