% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_joinattributesbylocation.R
\name{qgis_joinattributesbylocation}
\alias{qgis_joinattributesbylocation}
\title{QGIS algorithm Join attributes by location}
\usage{
qgis_joinattributesbylocation(
  INPUT = qgisprocess::qgis_default_value(),
  PREDICATE = qgisprocess::qgis_default_value(),
  JOIN = qgisprocess::qgis_default_value(),
  JOIN_FIELDS = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  DISCARD_NONMATCHING = qgisprocess::qgis_default_value(),
  PREFIX = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  NON_MATCHING = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Join to features in. Path to a vector layer.}

\item{PREDICATE}{\code{enum}  of \verb{("intersect", "contain", "equal", "touch", "overlap", "are within", "cross")} - Features they (geometric predicate). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{JOIN}{\code{source} - By comparing to. Path to a vector layer.}

\item{JOIN_FIELDS}{\code{field} - Fields to add (leave empty to use all fields). The name of an existing field. ; delimited list of existing field names.}

\item{METHOD}{\code{enum}  of \verb{("Create separate feature for each matching feature (one-to-many)", "Take attributes of the first matching feature only (one-to-one)", "Take attributes of the feature with largest overlap only (one-to-one)")} - Join type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DISCARD_NONMATCHING}{\code{boolean} - Discard records which could not be joined. 1 for true/yes. 0 for false/no.}

\item{PREFIX}{\code{string} - Joined field prefix. String value.}

\item{OUTPUT}{\code{sink} - Joined layer. Path for new vector layer.}

\item{NON_MATCHING}{\code{sink} - Unjoinable features from first layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Join attributes by location (native:joinattributesbylocation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item JOINED_COUNT - outputNumber - Number of joined features from input table
\item NON_MATCHING - outputVector - Unjoinable features from first layer
\item OUTPUT - outputVector - Joined layer
}
}
}
