% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_kmeansclustering.R
\name{qgis_kmeansclustering}
\alias{qgis_kmeansclustering}
\title{QGIS algorithm K-means clustering}
\usage{
qgis_kmeansclustering(
  INPUT = qgisprocess:::qgis_default_value(),
  CLUSTERS = qgisprocess:::qgis_default_value(),
  FIELD_NAME = qgisprocess:::qgis_default_value(),
  SIZE_FIELD_NAME = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{CLUSTERS}{\code{number} - Number of clusters. A numeric value.}

\item{FIELD_NAME}{\code{string} - Cluster field name. String value.}

\item{SIZE_FIELD_NAME}{\code{string} - Cluster size field name. String value.}

\item{OUTPUT}{\code{sink} - Clusters. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) K-means clustering (native:kmeansclustering)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Clusters
}
}
}
