% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_layertobookmarks.R
\name{qgis_layertobookmarks}
\alias{qgis_layertobookmarks}
\title{QGIS algorithm - Convert layer to spatial bookmarks}
\usage{
qgis_layertobookmarks(
  INPUT = qgisprocess:::qgis_default_value(),
  DESTINATION = qgisprocess:::qgis_default_value(),
  NAME_EXPRESSION = qgisprocess:::qgis_default_value(),
  GROUP_EXPRESSION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DESTINATION}{\code{enum}  of \verb{("Project bookmarks", "User bookmarks")} - Bookmark destination. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NAME_EXPRESSION}{\code{expression} - Name field. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{GROUP_EXPRESSION}{\code{expression} - Group field. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convert layer to spatial bookmarks (native:layertobookmarks). Converts feature extents to stored spatial bookmarks. This algorithm creates spatial bookmarks corresponding to the extent of features contained in a layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item COUNT - outputNumber - Count of bookmarks added
}
}
}
