% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_lineintersections.R
\name{qgis_lineintersections}
\alias{qgis_lineintersections}
\title{QGIS algorithm Line intersections}
\usage{
qgis_lineintersections(
  INPUT = qgisprocess:::qgis_default_value(),
  INTERSECT = qgisprocess:::qgis_default_value(),
  INPUT_FIELDS = qgisprocess:::qgis_default_value(),
  INTERSECT_FIELDS = qgisprocess:::qgis_default_value(),
  INTERSECT_FIELDS_PREFIX = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{INTERSECT}{\code{source} - Intersect layer. Path to a vector layer.}

\item{INPUT_FIELDS}{\code{field} - Input fields to keep (leave empty to keep all fields). The name of an existing field. ; delimited list of existing field names.}

\item{INTERSECT_FIELDS}{\code{field} - Intersect fields to keep (leave empty to keep all fields). The name of an existing field. ; delimited list of existing field names.}

\item{INTERSECT_FIELDS_PREFIX}{\code{string} - Intersect fields prefix. String value.}

\item{OUTPUT}{\code{sink} - Intersections. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Line intersections (native:lineintersections)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Intersections
}
}
}
