% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_loadlayer.R
\name{qgis_loadlayer}
\alias{qgis_loadlayer}
\title{QGIS algorithm Load layer into project}
\usage{
qgis_loadlayer(
  INPUT = qgisprocess::qgis_default_value(),
  NAME = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{layer} - Layer. Path to a vector, raster or mesh layer.}

\item{NAME}{\code{string} - Loaded layer name. String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Load layer into project (native:loadlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputLayer - Layer
}
}
}
