% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_lowestpositioninrasterstack.R
\name{qgis_lowestpositioninrasterstack}
\alias{qgis_lowestpositioninrasterstack}
\title{QGIS algorithm Lowest position in raster stack}
\usage{
qgis_lowestpositioninrasterstack(
  INPUT_RASTERS = qgisprocess::qgis_default_value(),
  REFERENCE_LAYER = qgisprocess::qgis_default_value(),
  IGNORE_NODATA = qgisprocess::qgis_default_value(),
  OUTPUT_NODATA_VALUE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT_RASTERS}{\code{multilayer} - Input raster layers. .}

\item{REFERENCE_LAYER}{\code{raster} - Reference layer. Path to a raster layer.}

\item{IGNORE_NODATA}{\code{boolean} - Ignore NoData values. 1 for true/yes. 0 for false/no.}

\item{OUTPUT_NODATA_VALUE}{\code{number} - Output NoData value. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Output layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Lowest position in raster stack (native:lowestpositioninrasterstack)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CRS_AUTHID - outputString - CRS authority identifier
\item EXTENT - outputString - Extent
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item OUTPUT - outputRaster - Output layer
\item TOTAL_PIXEL_COUNT - outputNumber - Total pixel count
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
}
}
}
