% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_meancoordinates.R
\name{qgis_meancoordinates}
\alias{qgis_meancoordinates}
\title{QGIS algorithm Mean coordinate(s)}
\usage{
qgis_meancoordinates(
  INPUT = qgisprocess::qgis_default_value(),
  WEIGHT = qgisprocess::qgis_default_value(),
  UID = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{WEIGHT}{\code{field} - Weight field. The name of an existing field. ; delimited list of existing field names.}

\item{UID}{\code{field} - Unique ID field. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Mean coordinates. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Mean coordinate(s) (native:meancoordinates)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Mean coordinates
}
}
}
