% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_meshcontours.R
\name{qgis_meshcontours}
\alias{qgis_meshcontours}
\title{QGIS algorithm Export contours}
\usage{
qgis_meshcontours(
  INPUT = qgisprocess::qgis_default_value(),
  DATASET_GROUPS = qgisprocess::qgis_default_value(),
  DATASET_TIME = qgisprocess::qgis_default_value(),
  INCREMENT = qgisprocess::qgis_default_value(),
  MINIMUM = qgisprocess::qgis_default_value(),
  MAXIMUM = qgisprocess::qgis_default_value(),
  CONTOUR_LEVEL_LIST = qgisprocess::qgis_default_value(),
  CRS_OUTPUT = qgisprocess::qgis_default_value(),
  OUTPUT_LINES = qgisprocess::qgis_default_value(),
  OUTPUT_POLYGONS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{mesh} - Input mesh layer. Path to a mesh layer.}

\item{DATASET_GROUPS}{\code{meshdatasetgroups} - Dataset groups. .}

\item{DATASET_TIME}{\code{meshdatasettime} - Dataset time. .}

\item{INCREMENT}{\code{number} - Increment between contour levels. A numeric value.}

\item{MINIMUM}{\code{number} - Minimum contour level. A numeric value.}

\item{MAXIMUM}{\code{number} - Maximum contour level. A numeric value.}

\item{CONTOUR_LEVEL_LIST}{\code{string} - List of contours level. String value.}

\item{CRS_OUTPUT}{\code{crs} - Output coordinate system. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT_LINES}{\code{sink} - Exported contour lines. Path for new vector layer.}

\item{OUTPUT_POLYGONS}{\code{sink} - Exported contour polygons. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Export contours (native:meshcontours)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_LINES - outputVector - Exported contour lines
\item OUTPUT_POLYGONS - outputVector - Exported contour polygons
}
}
}
