% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_meshexportcrosssection.R
\name{qgis_meshexportcrosssection}
\alias{qgis_meshexportcrosssection}
\title{QGIS algorithm - Export cross section dataset values on lines from mesh}
\usage{
qgis_meshexportcrosssection(
  INPUT = qgisprocess:::qgis_default_value(),
  DATASET_GROUPS = qgisprocess:::qgis_default_value(),
  DATASET_TIME = qgisprocess:::qgis_default_value(),
  INPUT_LINES = qgisprocess:::qgis_default_value(),
  RESOLUTION = qgisprocess:::qgis_default_value(),
  COORDINATES_DIGITS = qgisprocess:::qgis_default_value(),
  DATASET_DIGITS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{mesh} - Input mesh layer. Path to a mesh layer.}

\item{DATASET_GROUPS}{\code{meshdatasetgroups} - Dataset groups. .}

\item{DATASET_TIME}{\code{meshdatasettime} - Dataset time. .}

\item{INPUT_LINES}{\code{source} - Lines for data export. Path to a vector layer.}

\item{RESOLUTION}{\code{distance} - Line segmentation resolution. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{COORDINATES_DIGITS}{\code{number} - Digits count for coordinates. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DATASET_DIGITS}{\code{number} - Digits count for dataset value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{fileDestination} - Exported data CSV file. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Export cross section dataset values on lines from mesh (native:meshexportcrosssection). Extracts a mesh dataset's values from lines contained in a vector layer. This algorithm extracts mesh's dataset values from line contained in a vector layer. Each line is discretized with a resolution distance parameter for extraction of values on its vertices.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - Exported data CSV file
}
}
}
