% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_minimumboundinggeometry.R
\name{qgis_minimumboundinggeometry}
\alias{qgis_minimumboundinggeometry}
\title{QGIS algorithm Minimum bounding geometry}
\usage{
qgis_minimumboundinggeometry(
  INPUT = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Field (optional, set if features should be grouped by class). The name of an existing field. ; delimited list of existing field names.}

\item{TYPE}{\code{enum}  of \verb{("Envelope (Bounding Box)", "Minimum Oriented Rectangle", "Minimum Enclosing Circle", "Convex Hull")} - Geometry type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Bounding geometry. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Minimum bounding geometry (qgis:minimumboundinggeometry)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Bounding geometry
}
}
}
