% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_modelerrastercalc.R
\name{qgis_modelerrastercalc}
\alias{qgis_modelerrastercalc}
\title{QGIS algorithm - Raster calculator}
\usage{
qgis_modelerrastercalc(
  LAYERS = qgisprocess:::qgis_default_value(),
  EXPRESSION = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  CELL_SIZE = qgisprocess:::qgis_default_value(),
  CRS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LAYERS}{\code{multilayer} - Input layers. .}

\item{EXPRESSION}{\code{expression} - Expression. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{EXTENT}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{CELL_SIZE}{\code{number} - Output cell size (leave empty to set automatically). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CRS}{\code{crs} - Output CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT}{\code{rasterDestination} - Calculated. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raster calculator (native:modelerrastercalc). Performing algebraic operations using raster layers.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Calculated
}
}
}
