% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_multiringconstantbuffer.R
\name{qgis_multiringconstantbuffer}
\alias{qgis_multiringconstantbuffer}
\title{QGIS algorithm Multi-ring buffer (constant distance)}
\usage{
qgis_multiringconstantbuffer(
  INPUT = qgisprocess::qgis_default_value(),
  RINGS = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{RINGS}{\code{number} - Number of rings. A numeric value.}

\item{DISTANCE}{\code{distance} - Distance between rings. A numeric value.}

\item{OUTPUT}{\code{sink} - Multi-ring buffer (constant distance). Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Multi-ring buffer (constant distance) (native:multiringconstantbuffer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Multi-ring buffer (constant distance)
}
}
}
