% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_openurl.R
\name{qgis_openurl}
\alias{qgis_openurl}
\title{QGIS algorithm - Open file or URL}
\usage{
qgis_openurl(
  URL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{URL}{\code{string} - URL or file path. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Open file or URL (native:openurl). Opens files in their default associated application, or URLs in the user's default web browser. This algorithm opens files in their default associated application, or URLs in the user's default web browser.
}
\details{
\subsection{Outputs description}{
\itemize{
\item SUCCESS - outputBoolean - Successfully performed opening file or URL
}
}
}
