% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_pixelstopolygons.R
\name{qgis_pixelstopolygons}
\alias{qgis_pixelstopolygons}
\title{QGIS algorithm Raster pixels to polygons}
\usage{
qgis_pixelstopolygons(
  INPUT_RASTER = qgisprocess:::qgis_default_value(),
  RASTER_BAND = qgisprocess:::qgis_default_value(),
  FIELD_NAME = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_RASTER}{\code{raster} - Raster layer. Path to a raster layer.}

\item{RASTER_BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{FIELD_NAME}{\code{string} - Field name. String value.}

\item{OUTPUT}{\code{sink} - Vector polygons. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raster pixels to polygons (native:pixelstopolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Vector polygons
}
}
}
