% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_pointsdisplacement.R
\name{qgis_pointsdisplacement}
\alias{qgis_pointsdisplacement}
\title{QGIS algorithm Points displacement}
\usage{
qgis_pointsdisplacement(
  INPUT = qgisprocess::qgis_default_value(),
  PROXIMITY = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  HORIZONTAL = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{PROXIMITY}{\code{distance} - Minimum distance to other points. A numeric value.}

\item{DISTANCE}{\code{distance} - Displacement distance. A numeric value.}

\item{HORIZONTAL}{\code{boolean} - Horizontal distribution for two point case. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Displaced. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Points displacement (qgis:pointsdisplacement)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Displaced
}
}
}
