% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_polygonfromlayerextent.R
\name{qgis_polygonfromlayerextent}
\alias{qgis_polygonfromlayerextent}
\title{QGIS algorithm Extract layer extent}
\usage{
qgis_polygonfromlayerextent(
  INPUT = qgisprocess:::qgis_default_value(),
  ROUND_TO = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{layer} - Input layer. Path to a vector, raster or mesh layer.}

\item{ROUND_TO}{\code{distance} - Round values to. A numeric value.}

\item{OUTPUT}{\code{sink} - Extent. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Extract layer extent (native:polygonfromlayerextent)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Extent
}
}
}
