% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_postgisexecuteandloadsql.R
\name{qgis_postgisexecuteandloadsql}
\alias{qgis_postgisexecuteandloadsql}
\title{QGIS algorithm PostgreSQL execute and load SQL}
\usage{
qgis_postgisexecuteandloadsql(
  DATABASE = qgisprocess::qgis_default_value(),
  SQL = qgisprocess::qgis_default_value(),
  ID_FIELD = qgisprocess::qgis_default_value(),
  GEOMETRY_FIELD = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DATABASE}{\code{providerconnection} - Database (connection name). Name of registered database connection.}

\item{SQL}{\code{string} - SQL query. String value.}

\item{ID_FIELD}{\code{string} - Unique ID field name. String value.}

\item{GEOMETRY_FIELD}{\code{string} - Geometry field name. String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS PostgreSQL execute and load SQL (qgis:postgisexecuteandloadsql)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output layer
}
}
}
