% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_printlayouttoimage.R
\name{qgis_printlayouttoimage}
\alias{qgis_printlayouttoimage}
\title{QGIS algorithm Export print layout as image}
\usage{
qgis_printlayouttoimage(
  LAYOUT = qgisprocess::qgis_default_value(),
  LAYERS = qgisprocess::qgis_default_value(),
  DPI = qgisprocess::qgis_default_value(),
  GEOREFERENCE = qgisprocess::qgis_default_value(),
  INCLUDE_METADATA = qgisprocess::qgis_default_value(),
  ANTIALIAS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{LAYOUT}{\code{layout} - Print layout. Name of print layout in current project.}

\item{LAYERS}{\code{multilayer} - Map layers to assign to unlocked map item(s). .}

\item{DPI}{\code{number} - DPI (leave blank for default layout DPI). A numeric value.}

\item{GEOREFERENCE}{\code{boolean} - Generate world file. 1 for true/yes. 0 for false/no.}

\item{INCLUDE_METADATA}{\code{boolean} - Export RDF metadata (title, author, etc.). 1 for true/yes. 0 for false/no.}

\item{ANTIALIAS}{\code{boolean} - Enable antialiasing. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{fileDestination} - Image file. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Export print layout as image (native:printlayouttoimage)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - Image file
}
}
}
