% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_printlayouttopdf.R
\name{qgis_printlayouttopdf}
\alias{qgis_printlayouttopdf}
\title{QGIS algorithm Export print layout as PDF}
\usage{
qgis_printlayouttopdf(
  LAYOUT = qgisprocess::qgis_default_value(),
  LAYERS = qgisprocess::qgis_default_value(),
  DPI = qgisprocess::qgis_default_value(),
  FORCE_VECTOR = qgisprocess::qgis_default_value(),
  GEOREFERENCE = qgisprocess::qgis_default_value(),
  INCLUDE_METADATA = qgisprocess::qgis_default_value(),
  DISABLE_TILED = qgisprocess::qgis_default_value(),
  SIMPLIFY = qgisprocess::qgis_default_value(),
  TEXT_FORMAT = qgisprocess::qgis_default_value(),
  SEPARATE_LAYERS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LAYOUT}{\code{layout} - Print layout. Name of print layout in current project.}

\item{LAYERS}{\code{multilayer} - Map layers to assign to unlocked map item(s). .}

\item{DPI}{\code{number} - DPI (leave blank for default layout DPI). A numeric value.}

\item{FORCE_VECTOR}{\code{boolean} - Always export as vectors. 1 for true/yes. 0 for false/no.}

\item{GEOREFERENCE}{\code{boolean} - Append georeference information. 1 for true/yes. 0 for false/no.}

\item{INCLUDE_METADATA}{\code{boolean} - Export RDF metadata (title, author, etc.). 1 for true/yes. 0 for false/no.}

\item{DISABLE_TILED}{\code{boolean} - Disable tiled raster layer exports. 1 for true/yes. 0 for false/no.}

\item{SIMPLIFY}{\code{boolean} - Simplify geometries to reduce output file size. 1 for true/yes. 0 for false/no.}

\item{TEXT_FORMAT}{\code{enum}  of \verb{("Always Export Text as Paths (Recommended)", "Always Export Text as Text Objects")} - Text export. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEPARATE_LAYERS}{\code{boolean} - Export layers as separate PDF files. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{fileDestination} - PDF file. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Export print layout as PDF (native:printlayouttopdf)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - PDF file
}
}
}
