% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_raisemessage.R
\name{qgis_raisemessage}
\alias{qgis_raisemessage}
\title{QGIS algorithm - Raise message}
\usage{
qgis_raisemessage(
  MESSAGE = qgisprocess:::qgis_default_value(),
  CONDITION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{MESSAGE}{\code{string} - Information message. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CONDITION}{\code{expression} - Condition. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raise message (native:raisemessage). Raises an information message. This algorithm raises an information message in the log.  The message can be customized, and optionally an expression based condition can be specified. If an expression condition is used, then the message will only be logged if the expression result is true. A false result indicates that no message will be logged.
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
