% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_randompointsinlayerbounds.R
\name{qgis_randompointsinlayerbounds}
\alias{qgis_randompointsinlayerbounds}
\title{QGIS algorithm Random points in layer bounds}
\usage{
qgis_randompointsinlayerbounds(
  INPUT = qgisprocess::qgis_default_value(),
  POINTS_NUMBER = qgisprocess::qgis_default_value(),
  MIN_DISTANCE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{POINTS_NUMBER}{\code{number} - Number of points. A numeric value.}

\item{MIN_DISTANCE}{\code{distance} - Minimum distance between points. A numeric value.}

\item{OUTPUT}{\code{sink} - Random points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Random points in layer bounds (qgis:randompointsinlayerbounds)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Random points
}
}
}
