% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_randompointsinsidepolygons.R
\name{qgis_randompointsinsidepolygons}
\alias{qgis_randompointsinsidepolygons}
\title{QGIS algorithm Random points inside polygons}
\usage{
qgis_randompointsinsidepolygons(
  INPUT = qgisprocess::qgis_default_value(),
  STRATEGY = qgisprocess::qgis_default_value(),
  VALUE = qgisprocess::qgis_default_value(),
  MIN_DISTANCE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{STRATEGY}{\code{enum}  of \verb{("Points count", "Points density")} - Sampling strategy. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{VALUE}{\code{number} - Point count or density. A numeric value.}

\item{MIN_DISTANCE}{\code{distance} - Minimum distance between points. A numeric value.}

\item{OUTPUT}{\code{sink} - Random points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Random points inside polygons (qgis:randompointsinsidepolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Random points
}
}
}
