% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_randomselection.R
\name{qgis_randomselection}
\alias{qgis_randomselection}
\title{QGIS algorithm Random selection}
\usage{
qgis_randomselection(
  INPUT = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  NUMBER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{vector} - Input layer. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("Number of selected features", "Percentage of selected features")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NUMBER}{\code{number} - Number/percentage of selected features. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Random selection (qgis:randomselection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Selected
}
}
}
