% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rasterize.R
\name{qgis_rasterize}
\alias{qgis_rasterize}
\title{QGIS algorithm Convert map to raster}
\usage{
qgis_rasterize(
  EXTENT = qgisprocess::qgis_default_value(),
  EXTENT_BUFFER = qgisprocess::qgis_default_value(),
  TILE_SIZE = qgisprocess::qgis_default_value(),
  MAP_UNITS_PER_PIXEL = qgisprocess::qgis_default_value(),
  MAKE_BACKGROUND_TRANSPARENT = qgisprocess::qgis_default_value(),
  MAP_THEME = qgisprocess::qgis_default_value(),
  LAYERS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{EXTENT}{\code{extent} - Minimum extent to render. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{EXTENT_BUFFER}{\code{number} - Buffer around tiles in map units. A numeric value.}

\item{TILE_SIZE}{\code{number} - Tile size. A numeric value.}

\item{MAP_UNITS_PER_PIXEL}{\code{number} - Map units per pixel. A numeric value.}

\item{MAKE_BACKGROUND_TRANSPARENT}{\code{boolean} - Make background transparent. 1 for true/yes. 0 for false/no.}

\item{MAP_THEME}{\code{maptheme} - Map theme to render. Name of an existing map theme.}

\item{LAYERS}{\code{multilayer} - Layers to render. .}

\item{OUTPUT}{\code{rasterDestination} - Output layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convert map to raster (native:rasterize)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output layer
}
}
}
