% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rasterlayerproperties.R
\name{qgis_rasterlayerproperties}
\alias{qgis_rasterlayerproperties}
\title{QGIS algorithm Raster layer properties}
\usage{
qgis_rasterlayerproperties(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raster layer properties (native:rasterlayerproperties)
}
\details{
\subsection{Outputs description}{
\itemize{
\item X_MIN - outputNumber - Minimum x
\item X_MAX - outputNumber - Maximum x
\item Y_MIN - outputNumber - Minimum y
\item Y_MAX - outputNumber - Maximum y
\item EXTENT - outputString - Extent
\item PIXEL_WIDTH - outputNumber - Pixel size
\item PIXEL_HEIGHT - outputNumber - Pixel size
\item CRS_AUTHID - outputString - CRS authority identifier
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item HAS_NODATA_VALUE - outputBoolean - Band has a nodata value set
\item NODATA_VALUE - outputNumber - Band nodata value
\item BAND_COUNT - outputNumber - Number of bands in raster
}
}
}
