% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rasterlayerstatistics.R
\name{qgis_rasterlayerstatistics}
\alias{qgis_rasterlayerstatistics}
\title{QGIS algorithm Raster layer statistics}
\usage{
qgis_rasterlayerstatistics(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  OUTPUT_HTML_FILE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{OUTPUT_HTML_FILE}{\code{fileDestination} - Statistics. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raster layer statistics (native:rasterlayerstatistics)
}
\details{
\subsection{Outputs description}{
\itemize{
\item MAX - outputNumber - Maximum value
\item MEAN - outputNumber - Mean value
\item MIN - outputNumber - Minimum value
\item OUTPUT_HTML_FILE - outputHtml - Statistics
\item RANGE - outputNumber - Range
\item STD_DEV - outputNumber - Standard deviation
\item SUM - outputNumber - Sum
\item SUM_OF_SQUARES - outputNumber - Sum of the squares
}
}
}
