% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rasterlogicalor.R
\name{qgis_rasterlogicalor}
\alias{qgis_rasterlogicalor}
\title{QGIS algorithm Raster boolean OR}
\usage{
qgis_rasterlogicalor(
  INPUT = qgisprocess::qgis_default_value(),
  REF_LAYER = qgisprocess::qgis_default_value(),
  NODATA_AS_FALSE = qgisprocess::qgis_default_value(),
  NO_DATA = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input layers. .}

\item{REF_LAYER}{\code{raster} - Reference layer. Path to a raster layer.}

\item{NODATA_AS_FALSE}{\code{boolean} - Treat nodata values as false. 1 for true/yes. 0 for false/no.}

\item{NO_DATA}{\code{number} - Output no data value. A numeric value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Byte", "Int16", "UInt16", "Int32", "UInt32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{rasterDestination} - Output layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raster boolean OR (native:rasterlogicalor)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CRS_AUTHID - outputString - CRS authority identifier
\item EXTENT - outputString - Extent
\item FALSE_PIXEL_COUNT - outputNumber - False pixel count
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item NODATA_PIXEL_COUNT - outputNumber - NODATA pixel count
\item OUTPUT - outputRaster - Output layer
\item TOTAL_PIXEL_COUNT - outputNumber - Total pixel count
\item TRUE_PIXEL_COUNT - outputNumber - True pixel count
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
}
}
}
