% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rasterminmax.R
\name{qgis_rasterminmax}
\alias{qgis_rasterminmax}
\title{QGIS algorithm - Raster minimum/maximum}
\usage{
qgis_rasterminmax(
  INPUT = qgisprocess:::qgis_default_value(),
  BAND = qgisprocess:::qgis_default_value(),
  EXTRACT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{EXTRACT}{\code{enum}  of \verb{("Minimum and Maximum", "Minimum", "Maximum")} - Extract extrema. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raster minimum/maximum (native:rasterminmax). Calculates the minimum and maximum pixel in a raster layer. This algorithm extracts extrema (minimum and maximum) values from a given band of the raster layer.  The output is a vector layer containing point features for the selected extrema, at the center of the associated pixel.  If multiple pixels in the raster share the minimum or maximum value, then only one of these pixels will be included in the output.
}
\details{
\subsection{Outputs description}{
\itemize{
\item MAXIMUM - outputNumber - Maximum
\item MINIMUM - outputNumber - Minimum
\item OUTPUT - outputVector - Output
}
}
}
