% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rasterrank.R
\name{qgis_rasterrank}
\alias{qgis_rasterrank}
\title{QGIS algorithm - Raster rank}
\usage{
qgis_rasterrank(
  INPUT_RASTERS = qgisprocess:::qgis_default_value(),
  RANKS = qgisprocess:::qgis_default_value(),
  NODATA_HANDLING = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  CELL_SIZE = qgisprocess:::qgis_default_value(),
  CRS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_RASTERS}{\code{multilayer} - Input raster layers. .}

\item{RANKS}{\code{string} - Rank(s) (separate multiple ranks using commas). String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NODATA_HANDLING}{\code{enum}  of \verb{("Exclude NoData from values lists", "Presence of NoData in a values list results in NoData output cell")} - NoData value handling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXTENT}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{CELL_SIZE}{\code{number} - Output cell size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CRS}{\code{crs} - Output CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT}{\code{rasterDestination} - Ranked. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raster rank (native:rasterrank). Performs a cell-by-cell analysis in which output values match the rank of a sorted list of overlapping cell values from input layers. This algorithm performs a cell-by-cell analysis in which output values match the rank of a sorted list of overlapping cell values from input layers. The output raster will be multi-band if multiple ranks are provided. If multiband rasters are used in the data raster stack, the algorithm will always perform the analysis on the first band of the rasters.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Ranked
}
}
}
