% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rastersurfacevolume.R
\name{qgis_rastersurfacevolume}
\alias{qgis_rastersurfacevolume}
\title{QGIS algorithm Raster surface volume}
\usage{
qgis_rastersurfacevolume(
  INPUT = qgisprocess:::qgis_default_value(),
  BAND = qgisprocess:::qgis_default_value(),
  LEVEL = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  OUTPUT_HTML_FILE = qgisprocess:::qgis_default_value(),
  OUTPUT_TABLE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{LEVEL}{\code{number} - Base level. A numeric value.}

\item{METHOD}{\code{enum}  of \verb{("Count Only Above Base Level", "Count Only Below Base Level", "Subtract Volumes Below Base Level", "Add Volumes Below Base Level")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT_HTML_FILE}{\code{fileDestination} - Surface volume report. Path for new file.}

\item{OUTPUT_TABLE}{\code{sink} - Surface volume table. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raster surface volume (native:rastersurfacevolume)
}
\details{
\subsection{Outputs description}{
\itemize{
\item AREA - outputNumber - Area
\item OUTPUT_HTML_FILE - outputHtml - Surface volume report
\item OUTPUT_TABLE - outputVector - Surface volume table
\item PIXEL_COUNT - outputNumber - Pixel count
\item VOLUME - outputNumber - Volume
}
}
}
