% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_reclassifybytable.R
\name{qgis_reclassifybytable}
\alias{qgis_reclassifybytable}
\title{QGIS algorithm - Reclassify by table}
\usage{
qgis_reclassifybytable(
  INPUT_RASTER = qgisprocess:::qgis_default_value(),
  RASTER_BAND = qgisprocess:::qgis_default_value(),
  TABLE = qgisprocess:::qgis_default_value(),
  NO_DATA = qgisprocess:::qgis_default_value(),
  RANGE_BOUNDARIES = qgisprocess:::qgis_default_value(),
  NODATA_FOR_MISSING = qgisprocess:::qgis_default_value(),
  DATA_TYPE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_RASTER}{\code{raster} - Raster layer. Path to a raster layer.}

\item{RASTER_BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{TABLE}{\code{matrix} - Reclassification table. A comma delimited list of values.}

\item{NO_DATA}{\code{number} - Output no data value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RANGE_BOUNDARIES}{\code{enum}  of \verb{("min  value = max", "min = value  max", "min = value = max", "min  value  max")} - Range boundaries. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NODATA_FOR_MISSING}{\code{boolean} - Use no data when no range matches value. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Byte", "Int16", "UInt16", "Int32", "UInt32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64", "Int8")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{rasterDestination} - Reclassified raster. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Reclassify by table (native:reclassifybytable). This algorithm reclassifies a raster band by assigning new class values based on the ranges specified in a fixed table.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Reclassified raster
}
}
}
