% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rectanglesovalsdiamondsvariable.R
\name{qgis_rectanglesovalsdiamondsvariable}
\alias{qgis_rectanglesovalsdiamondsvariable}
\title{QGIS algorithm Rectangles, ovals, diamonds (variable)}
\usage{
qgis_rectanglesovalsdiamondsvariable(
  INPUT = qgisprocess:::qgis_default_value(),
  SHAPE = qgisprocess:::qgis_default_value(),
  WIDTH = qgisprocess:::qgis_default_value(),
  HEIGHT = qgisprocess:::qgis_default_value(),
  ROTATION = qgisprocess:::qgis_default_value(),
  SEGMENTS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{SHAPE}{\code{enum}  of \verb{("Rectangles", "Diamonds", "Ovals")} - Buffer shape. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{WIDTH}{\code{field} - Width field. The name of an existing field. ; delimited list of existing field names.}

\item{HEIGHT}{\code{field} - Height field. The name of an existing field. ; delimited list of existing field names.}

\item{ROTATION}{\code{field} - Rotation field. The name of an existing field. ; delimited list of existing field names.}

\item{SEGMENTS}{\code{number} - Number of segments. A numeric value.}

\item{OUTPUT}{\code{sink} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Rectangles, ovals, diamonds (variable) (qgis:rectanglesovalsdiamondsvariable)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output
}
}
}
