% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_refactorfields.R
\name{qgis_refactorfields}
\alias{qgis_refactorfields}
\title{QGIS algorithm - Refactor fields}
\usage{
qgis_refactorfields(
  INPUT = qgisprocess:::qgis_default_value(),
  FIELDS_MAPPING = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELDS_MAPPING}{\code{fields_mapping} - Fields mapping. .}

\item{OUTPUT}{\code{sink} - Refactored. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Refactor fields (native:refactorfields). Allows editing the structure of the attributes table of a vector layer, permitting field renaming, creation and deletion. This algorithm allows editing the structure of the attributes table of a vector layer. Fields can be modified in their type and name, using a fields mapping.  The original layer is not modified. A new layer is generated, which contains a modified attribute table, according to the provided fields mapping.  Rows in orange have constraints in the template layer from which these fields were loaded. Treat this information as a hint during configuration. No constraints will be added on an output layer nor will they be checked or enforced by the algorithm.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Refactored
}
}
}
