% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_regularpoints.R
\name{qgis_regularpoints}
\alias{qgis_regularpoints}
\title{QGIS algorithm - Regular points}
\usage{
qgis_regularpoints(
  EXTENT = qgisprocess:::qgis_default_value(),
  SPACING = qgisprocess:::qgis_default_value(),
  INSET = qgisprocess:::qgis_default_value(),
  RANDOMIZE = qgisprocess:::qgis_default_value(),
  IS_SPACING = qgisprocess:::qgis_default_value(),
  CRS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{EXTENT}{\code{extent} - Input extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{SPACING}{\code{distance} - Point spacing\code{/}count. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INSET}{\code{distance} - Initial inset from corner (LH side). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RANDOMIZE}{\code{boolean} - Apply random offset to point spacing. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{IS_SPACING}{\code{boolean} - Use point spacing. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CRS}{\code{crs} - Output layer CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT}{\code{sink} - Regular points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Regular points (qgis:regularpoints). This algorithm creates a point layer with a given number of regular points, all of them within a given extent.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Regular points
}
}
}
