% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_relief.R
\name{qgis_relief}
\alias{qgis_relief}
\title{QGIS algorithm Relief}
\usage{
qgis_relief(
  INPUT = qgisprocess::qgis_default_value(),
  Z_FACTOR = qgisprocess::qgis_default_value(),
  AUTO_COLORS = qgisprocess::qgis_default_value(),
  COLORS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  FREQUENCY_DISTRIBUTION = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Elevation layer. Path to a raster layer.}

\item{Z_FACTOR}{\code{number} - Z factor. A numeric value.}

\item{AUTO_COLORS}{\code{boolean} - Generate relief classes automatically. 1 for true/yes. 0 for false/no.}

\item{COLORS}{\code{relief_colors} - Relief colors. .}

\item{OUTPUT}{\code{rasterDestination} - Relief. Path for new raster layer.}

\item{FREQUENCY_DISTRIBUTION}{\code{fileDestination} - Frequency distribution. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Relief (qgis:relief)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FREQUENCY_DISTRIBUTION - outputFile - Frequency distribution
\item OUTPUT - outputRaster - Relief
}
}
}
