% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_renametablefield.R
\name{qgis_renametablefield}
\alias{qgis_renametablefield}
\title{QGIS algorithm Rename field}
\usage{
qgis_renametablefield(
  INPUT = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  NEW_NAME = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Field to rename. The name of an existing field. ; delimited list of existing field names.}

\item{NEW_NAME}{\code{string} - New field name. String value.}

\item{OUTPUT}{\code{sink} - Renamed. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Rename field (native:renametablefield)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Renamed
}
}
}
