% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_repairshapefile.R
\name{qgis_repairshapefile}
\alias{qgis_repairshapefile}
\title{QGIS algorithm Repair Shapefile}
\usage{
qgis_repairshapefile(
  INPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{file} - Input Shapefile. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Repair Shapefile (native:repairshapefile)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Repaired layer
}
}
}
