% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_reprojectlayer.R
\name{qgis_reprojectlayer}
\alias{qgis_reprojectlayer}
\title{QGIS algorithm Reproject layer}
\usage{
qgis_reprojectlayer(
  INPUT = qgisprocess::qgis_default_value(),
  TARGET_CRS = qgisprocess::qgis_default_value(),
  OPERATION = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{TARGET_CRS}{\code{crs} - Target CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OPERATION}{\code{coordinateoperation} - Coordinate operation. String representation of Proj coordinate operation.}

\item{OUTPUT}{\code{sink} - Reprojected. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Reproject layer (native:reprojectlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Reprojecte
}
}
}
