% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rescaleraster.R
\name{qgis_rescaleraster}
\alias{qgis_rescaleraster}
\title{QGIS algorithm Rescale raster}
\usage{
qgis_rescaleraster(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  MINIMUM = qgisprocess::qgis_default_value(),
  MAXIMUM = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input raster. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{MINIMUM}{\code{number} - New minimum value. A numeric value.}

\item{MAXIMUM}{\code{number} - New maximum value. A numeric value.}

\item{NODATA}{\code{number} - New NODATA value. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Rescaled. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Rescale raster (native:rescaleraster)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Rescaled
}
}
}
