% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rotatefeatures.R
\name{qgis_rotatefeatures}
\alias{qgis_rotatefeatures}
\title{QGIS algorithm - Rotate}
\usage{
qgis_rotatefeatures(
  INPUT = qgisprocess:::qgis_default_value(),
  ANGLE = qgisprocess:::qgis_default_value(),
  ANCHOR = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{ANGLE}{\code{number} - Rotation (degrees clockwise). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ANCHOR}{\code{point} - Rotation anchor point. Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{OUTPUT}{\code{sink} - Rotated. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Rotate (native:rotatefeatures). This algorithm rotates feature geometries, by the specified angle clockwise  Optionally, the rotation can occur around a preset point. If not set the rotation occurs around each feature's centroid.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Rotated
}
}
}
