% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_roundrastervalues.R
\name{qgis_roundrastervalues}
\alias{qgis_roundrastervalues}
\title{QGIS algorithm Round raster}
\usage{
qgis_roundrastervalues(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  ROUNDING_DIRECTION = qgisprocess::qgis_default_value(),
  DECIMAL_PLACES = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  BASE_N = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input raster. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{ROUNDING_DIRECTION}{\code{enum}  of \verb{("Round up", "Round to nearest", "Round down")} - Rounding direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DECIMAL_PLACES}{\code{number} - Number of decimals places. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Output raster. Path for new raster layer.}

\item{BASE_N}{\code{number} - Base n for rounding to multiples of n. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Round raster (native:roundrastervalues)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output raster
}
}
}
