% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_segmentizebymaxdistance.R
\name{qgis_segmentizebymaxdistance}
\alias{qgis_segmentizebymaxdistance}
\title{QGIS algorithm Segmentize by maximum distance}
\usage{
qgis_segmentizebymaxdistance(
  INPUT = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DISTANCE}{\code{distance} - Maximum offset distance. A numeric value.}

\item{OUTPUT}{\code{sink} - Segmentized. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Segmentize by maximum distance (native:segmentizebymaxdistance)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Segmentized
}
}
}
