% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_selectbyattribute.R
\name{qgis_selectbyattribute}
\alias{qgis_selectbyattribute}
\title{QGIS algorithm Select by attribute}
\usage{
qgis_selectbyattribute(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  OPERATOR = qgisprocess::qgis_default_value(),
  VALUE = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{vector} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Selection attribute. The name of an existing field. ; delimited list of existing field names.}

\item{OPERATOR}{\code{enum}  of \verb{("=", "≠", ">", "≥", "<", "≤", "begins with", "contains", "is null", "is not null", "does not contain")} - Operator. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{VALUE}{\code{string} - Value. String value.}

\item{METHOD}{\code{enum}  of \verb{("creating new selection", "adding to current selection", "removing from current selection", "selecting within current selection")} - Modify current selection by. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Select by attribute (qgis:selectbyattribute)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Selected
}
}
}
