% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_selectbyexpression.R
\name{qgis_selectbyexpression}
\alias{qgis_selectbyexpression}
\title{QGIS algorithm Select by expression}
\usage{
qgis_selectbyexpression(
  INPUT = qgisprocess::qgis_default_value(),
  EXPRESSION = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{vector} - Input layer. Path to a vector layer.}

\item{EXPRESSION}{\code{expression} - Expression. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{METHOD}{\code{enum}  of \verb{("creating new selection", "adding to current selection", "removing from current selection", "selecting within current selection")} - Modify current selection by. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Select by expression (qgis:selectbyexpression)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Selected
}
}
}
