% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_serviceareafrompoint.R
\name{qgis_serviceareafrompoint}
\alias{qgis_serviceareafrompoint}
\title{QGIS algorithm Service area (from point)}
\usage{
qgis_serviceareafrompoint(
  INPUT = qgisprocess:::qgis_default_value(),
  STRATEGY = qgisprocess:::qgis_default_value(),
  DIRECTION_FIELD = qgisprocess:::qgis_default_value(),
  VALUE_FORWARD = qgisprocess:::qgis_default_value(),
  VALUE_BACKWARD = qgisprocess:::qgis_default_value(),
  VALUE_BOTH = qgisprocess:::qgis_default_value(),
  DEFAULT_DIRECTION = qgisprocess:::qgis_default_value(),
  SPEED_FIELD = qgisprocess:::qgis_default_value(),
  DEFAULT_SPEED = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  START_POINT = qgisprocess:::qgis_default_value(),
  TRAVEL_COST2 = qgisprocess:::qgis_default_value(),
  INCLUDE_BOUNDS = qgisprocess:::qgis_default_value(),
  OUTPUT_LINES = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Vector layer representing network. Path to a vector layer.}

\item{STRATEGY}{\code{enum}  of \verb{("Shortest", "Fastest")} - Path type to calculate. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DIRECTION_FIELD}{\code{field} - Direction field. The name of an existing field. ; delimited list of existing field names.}

\item{VALUE_FORWARD}{\code{string} - Value for forward direction. String value.}

\item{VALUE_BACKWARD}{\code{string} - Value for backward direction. String value.}

\item{VALUE_BOTH}{\code{string} - Value for both directions. String value.}

\item{DEFAULT_DIRECTION}{\code{enum}  of \verb{("Forward direction", "Backward direction", "Both directions")} - Default direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SPEED_FIELD}{\code{field} - Speed field. The name of an existing field. ; delimited list of existing field names.}

\item{DEFAULT_SPEED}{\code{number} - Default speed (km/h). A numeric value.}

\item{TOLERANCE}{\code{distance} - Topology tolerance. A numeric value.}

\item{START_POINT}{\code{point} - Start point. Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{TRAVEL_COST2}{\code{number} - Travel cost (distance for 'Shortest', time for 'Fastest'). A numeric value.}

\item{INCLUDE_BOUNDS}{\code{boolean} - Include upper/lower bound points. 1 for true/yes. 0 for false/no.}

\item{OUTPUT_LINES}{\code{sink} - Service area (lines). Path for new vector layer.}

\item{OUTPUT}{\code{sink} - Service area (boundary nodes). Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Service area (from point) (native:serviceareafrompoint)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Service area (boundary nodes)
\item OUTPUT_LINES - outputVector - Service area (lines)
}
}
}
