% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_setlayerencoding.R
\name{qgis_setlayerencoding}
\alias{qgis_setlayerencoding}
\title{QGIS algorithm - Set layer encoding}
\usage{
qgis_setlayerencoding(
  INPUT = qgisprocess:::qgis_default_value(),
  ENCODING = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{vector} - Input layer. Path to a vector layer.}

\item{ENCODING}{\code{string} - Encoding. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Set layer encoding (native:setlayerencoding). Sets the encoding used for reading a layer's attributes This algorithm sets the encoding used for reading a layer's attributes. No permanent changes are made to the layer, rather it affects only how the layer is read during the current session.  Changing the encoding is only supported for some vector layer data sources.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output layer
}
}
}
