% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_setlayerstyle.R
\name{qgis_setlayerstyle}
\alias{qgis_setlayerstyle}
\title{QGIS algorithm - Set layer style}
\usage{
qgis_setlayerstyle(
  INPUT = qgisprocess:::qgis_default_value(),
  STYLE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{layer} - Layer. Path to a vector, raster or mesh layer.}

\item{STYLE}{\code{file} - Style file. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Set layer style (native:setlayerstyle). Applies the style from a QML file to a layer. This algorithm applies the style to a layer. The style must be defined as QML file.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputLayer - Styled
}
}
}
