% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_setmvalue.R
\name{qgis_setmvalue}
\alias{qgis_setmvalue}
\title{QGIS algorithm Set M value}
\usage{
qgis_setmvalue(
  INPUT = qgisprocess::qgis_default_value(),
  M_VALUE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{M_VALUE}{\code{number} - M Value. A numeric value.}

\item{OUTPUT}{\code{sink} - M Added. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Set M value (native:setmvalue)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - M Added
}
}
}
