% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_setstyleforrasterlayer.R
\name{qgis_setstyleforrasterlayer}
\alias{qgis_setstyleforrasterlayer}
\title{QGIS algorithm Set style for raster layer}
\usage{
qgis_setstyleforrasterlayer(
  INPUT = qgisprocess::qgis_default_value(),
  STYLE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Raster layer. Path to a raster layer.}

\item{STYLE}{\code{file} - Style file. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Set style for raster layer (qgis:setstyleforrasterlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INPUT - outputRaster - Styled
}
}
}
