% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_setzvalue.R
\name{qgis_setzvalue}
\alias{qgis_setzvalue}
\title{QGIS algorithm Set Z value}
\usage{
qgis_setzvalue(
  INPUT = qgisprocess::qgis_default_value(),
  Z_VALUE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{Z_VALUE}{\code{number} - Z Value. A numeric value.}

\item{OUTPUT}{\code{sink} - Z Added. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Set Z value (native:setzvalue)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Z Added
}
}
}
