% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_shortestline.R
\name{qgis_shortestline}
\alias{qgis_shortestline}
\title{QGIS algorithm - Shortest line between features}
\usage{
qgis_shortestline(
  SOURCE = qgisprocess:::qgis_default_value(),
  DESTINATION = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  NEIGHBORS = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SOURCE}{\code{source} - Source layer. Path to a vector layer.}

\item{DESTINATION}{\code{source} - Destination layer. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("Distance to Nearest Point on feature", "Distance to Feature Centroid")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NEIGHBORS}{\code{number} - Maximum number of neighbors. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DISTANCE}{\code{distance} - Maximum distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Shortest lines. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Shortest line between features (native:shortestline). Calculates the shortest lines between features in source and destination layers. This algorithm creates a line layer as the shortest line between the source and the destination layer. By default only the first nearest feature of the destination layer is taken into account. The n-nearest neighboring features number can be specified.  If a maximum distance is specified, then only features which are closer than this distance will be considered.  The output features will contain all the source layer attributes, all the attributes from the n-nearest feature and the additional field of the distance.  This algorithm uses purely Cartesian calculations for distance, and does not consider geodetic or ellipsoid properties when determining feature proximity. The measurement and output coordinate system is based on the coordinate system of the source layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Shortest lines
}
}
}
