% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_shortestpathlayertopoint.R
\name{qgis_shortestpathlayertopoint}
\alias{qgis_shortestpathlayertopoint}
\title{QGIS algorithm Shortest path (layer to point)}
\usage{
qgis_shortestpathlayertopoint(
  INPUT = qgisprocess::qgis_default_value(),
  STRATEGY = qgisprocess::qgis_default_value(),
  DIRECTION_FIELD = qgisprocess::qgis_default_value(),
  VALUE_FORWARD = qgisprocess::qgis_default_value(),
  VALUE_BACKWARD = qgisprocess::qgis_default_value(),
  VALUE_BOTH = qgisprocess::qgis_default_value(),
  DEFAULT_DIRECTION = qgisprocess::qgis_default_value(),
  SPEED_FIELD = qgisprocess::qgis_default_value(),
  DEFAULT_SPEED = qgisprocess::qgis_default_value(),
  TOLERANCE = qgisprocess::qgis_default_value(),
  START_POINTS = qgisprocess::qgis_default_value(),
  END_POINT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Vector layer representing network. Path to a vector layer.}

\item{STRATEGY}{\code{enum}  of \verb{("Shortest", "Fastest")} - Path type to calculate. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DIRECTION_FIELD}{\code{field} - Direction field. The name of an existing field. ; delimited list of existing field names.}

\item{VALUE_FORWARD}{\code{string} - Value for forward direction. String value.}

\item{VALUE_BACKWARD}{\code{string} - Value for backward direction. String value.}

\item{VALUE_BOTH}{\code{string} - Value for both directions. String value.}

\item{DEFAULT_DIRECTION}{\code{enum}  of \verb{("Forward direction", "Backward direction", "Both directions")} - Default direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SPEED_FIELD}{\code{field} - Speed field. The name of an existing field. ; delimited list of existing field names.}

\item{DEFAULT_SPEED}{\code{number} - Default speed (km/h). A numeric value.}

\item{TOLERANCE}{\code{distance} - Topology tolerance. A numeric value.}

\item{START_POINTS}{\code{source} - Vector layer with start points. Path to a vector layer.}

\item{END_POINT}{\code{point} - End point. Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{OUTPUT}{\code{sink} - Shortest path. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Shortest path (layer to point) (native:shortestpathlayertopoint)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Shortest path
}
}
}
