% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_shpencodinginfo.R
\name{qgis_shpencodinginfo}
\alias{qgis_shpencodinginfo}
\title{QGIS algorithm Extract Shapefile encoding}
\usage{
qgis_shpencodinginfo(
  INPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{file} - Input layer. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Extract Shapefile encoding (native:shpencodinginfo)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ENCODING - outputString - Shapefile Encoding
\item CPG_ENCODING - outputString - CPG Encoding
\item LDID_ENCODING - outputString - LDID Encoding
}
}
}
