% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_simplifygeometries.R
\name{qgis_simplifygeometries}
\alias{qgis_simplifygeometries}
\title{QGIS algorithm Simplify}
\usage{
qgis_simplifygeometries(
  INPUT = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  TOLERANCE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("Distance (Douglas-Peucker)", "Snap to grid", "Area (Visvalingam)")} - Simplification method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TOLERANCE}{\code{distance} - Tolerance. A numeric value.}

\item{OUTPUT}{\code{sink} - Simplified. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Simplify (native:simplifygeometries)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Simplified
}
}
}
