% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_snappointstogrid.R
\name{qgis_snappointstogrid}
\alias{qgis_snappointstogrid}
\title{QGIS algorithm Snap points to grid}
\usage{
qgis_snappointstogrid(
  INPUT = qgisprocess::qgis_default_value(),
  HSPACING = qgisprocess::qgis_default_value(),
  VSPACING = qgisprocess::qgis_default_value(),
  ZSPACING = qgisprocess::qgis_default_value(),
  MSPACING = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{HSPACING}{\code{distance} - X Grid Spacing. A numeric value.}

\item{VSPACING}{\code{distance} - Y Grid Spacing. A numeric value.}

\item{ZSPACING}{\code{number} - Z Grid Spacing. A numeric value.}

\item{MSPACING}{\code{number} - M Grid Spacing. A numeric value.}

\item{OUTPUT}{\code{sink} - Snapped. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Snap points to grid (native:snappointstogrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Snapped
}
}
}
