% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_splitlinesbylength.R
\name{qgis_splitlinesbylength}
\alias{qgis_splitlinesbylength}
\title{QGIS algorithm Split lines by maximum length}
\usage{
qgis_splitlinesbylength(
  INPUT = qgisprocess::qgis_default_value(),
  LENGTH = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{LENGTH}{\code{distance} - Maximum line length. A numeric value.}

\item{OUTPUT}{\code{sink} - Split. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Split lines by maximum length (native:splitlinesbylength)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Split
}
}
}
