% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_stringconcatenation.R
\name{qgis_stringconcatenation}
\alias{qgis_stringconcatenation}
\title{QGIS algorithm String concatenation}
\usage{
qgis_stringconcatenation(
  INPUT_1 = qgisprocess:::qgis_default_value(),
  INPUT_2 = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_1}{\code{string} - Input 1. String value.}

\item{INPUT_2}{\code{string} - Input 2. String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) String concatenation (native:stringconcatenation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONCATENATION - outputString - Concatenation
}
}
}
