% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_stylefromproject.R
\name{qgis_stylefromproject}
\alias{qgis_stylefromproject}
\title{QGIS algorithm - Create style database from project}
\usage{
qgis_stylefromproject(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  OBJECTS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{file} - Input project (leave blank to use current). Path to a file.}

\item{OUTPUT}{\code{fileDestination} - Output style database. Path for new file.}

\item{OBJECTS}{\code{enum}  of \verb{("Symbols", "Color ramps", "Text formats", "Label settings")} - Objects to extract. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create style database from project (native:stylefromproject). Creates a style database by extracting all symbols, color ramps, text formats and label settings from a QGIS project. This algorithm extracts all style objects (including symbols, color ramps, text formats and label settings) from a QGIS project.  The extracted symbols are saved to a QGIS style database (XML format), which can be managed and imported via the Style Manager dialog.
}
\details{
\subsection{Outputs description}{
\itemize{
\item COLORRAMPS - outputNumber - Color ramp count
\item LABELSETTINGS - outputNumber - Label settings count
\item OUTPUT - outputFile - Output style database
\item SYMBOLS - outputNumber - Symbol count
\item TEXTFORMATS - outputNumber - Text format count
}
}
}
