% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_subdivide.R
\name{qgis_subdivide}
\alias{qgis_subdivide}
\title{QGIS algorithm Subdivide}
\usage{
qgis_subdivide(
  INPUT = qgisprocess::qgis_default_value(),
  MAX_NODES = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{MAX_NODES}{\code{number} - Maximum nodes in parts. A numeric value.}

\item{OUTPUT}{\code{sink} - Subdivided. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Subdivide (native:subdivide)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Subdivided
}
}
}
