% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_sumlinelengths.R
\name{qgis_sumlinelengths}
\alias{qgis_sumlinelengths}
\title{QGIS algorithm Sum line lengths}
\usage{
qgis_sumlinelengths(
  POLYGONS = qgisprocess::qgis_default_value(),
  LINES = qgisprocess::qgis_default_value(),
  LEN_FIELD = qgisprocess::qgis_default_value(),
  COUNT_FIELD = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{LEN_FIELD}{\code{string} - Lines length field name. String value.}

\item{COUNT_FIELD}{\code{string} - Lines count field name. String value.}

\item{OUTPUT}{\code{sink} - Line length. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Sum line lengths (native:sumlinelengths)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Line length
}
}
}
