% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_surfacetopolygon.R
\name{qgis_surfacetopolygon}
\alias{qgis_surfacetopolygon}
\title{QGIS algorithm - Surface to polygon}
\usage{
qgis_surfacetopolygon(
  INPUT = qgisprocess:::qgis_default_value(),
  CRS_OUTPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{mesh} - Input mesh layer. Path to a mesh layer.}

\item{CRS_OUTPUT}{\code{crs} - Output coordinate system. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT}{\code{sink} - Output vector layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Surface to polygon (native:surfacetopolygon). Exports a polygon layer containing a mesh layer's boundary. This algorithm exports a polygon layer containing a mesh layer's boundary. It may contain holes and it may be a multi-part polygon.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output vector layer
}
}
}
