% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_taperedbuffer.R
\name{qgis_taperedbuffer}
\alias{qgis_taperedbuffer}
\title{QGIS algorithm Tapered buffers}
\usage{
qgis_taperedbuffer(
  INPUT = qgisprocess::qgis_default_value(),
  START_WIDTH = qgisprocess::qgis_default_value(),
  END_WIDTH = qgisprocess::qgis_default_value(),
  SEGMENTS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{START_WIDTH}{\code{number} - Start width. A numeric value.}

\item{END_WIDTH}{\code{number} - End width. A numeric value.}

\item{SEGMENTS}{\code{number} - Segments. A numeric value.}

\item{OUTPUT}{\code{sink} - Buffered. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Tapered buffers (native:taperedbuffer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Buffered
}
}
}
