% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_texttofloat.R
\name{qgis_texttofloat}
\alias{qgis_texttofloat}
\title{QGIS algorithm Text to float}
\usage{
qgis_texttofloat(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Text attribute to convert to float. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Float from text. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Text to float (qgis:texttofloat)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Float from text
}
}
}
