% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_tilesxyzdirectory.R
\name{qgis_tilesxyzdirectory}
\alias{qgis_tilesxyzdirectory}
\title{QGIS algorithm Generate XYZ tiles (Directory)}
\usage{
qgis_tilesxyzdirectory(
  EXTENT = qgisprocess::qgis_default_value(),
  ZOOM_MIN = qgisprocess::qgis_default_value(),
  ZOOM_MAX = qgisprocess::qgis_default_value(),
  DPI = qgisprocess::qgis_default_value(),
  BACKGROUND_COLOR = qgisprocess::qgis_default_value(),
  TILE_FORMAT = qgisprocess::qgis_default_value(),
  QUALITY = qgisprocess::qgis_default_value(),
  METATILESIZE = qgisprocess::qgis_default_value(),
  TILE_WIDTH = qgisprocess::qgis_default_value(),
  TILE_HEIGHT = qgisprocess::qgis_default_value(),
  TMS_CONVENTION = qgisprocess::qgis_default_value(),
  OUTPUT_DIRECTORY = qgisprocess::qgis_default_value(),
  OUTPUT_HTML = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{EXTENT}{\code{extent} - Extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{ZOOM_MIN}{\code{number} - Minimum zoom. A numeric value.}

\item{ZOOM_MAX}{\code{number} - Maximum zoom. A numeric value.}

\item{DPI}{\code{number} - DPI. A numeric value.}

\item{BACKGROUND_COLOR}{\code{color} - Background color. String representation of color, e.g #ff0000 or rgba(200,100,50,0.8).}

\item{TILE_FORMAT}{\code{enum}  of \verb{("PNG", "JPG")} - Tile format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{QUALITY}{\code{number} - Quality (JPG only). A numeric value.}

\item{METATILESIZE}{\code{number} - Metatile size. A numeric value.}

\item{TILE_WIDTH}{\code{number} - Tile width. A numeric value.}

\item{TILE_HEIGHT}{\code{number} - Tile height. A numeric value.}

\item{TMS_CONVENTION}{\code{boolean} - Use inverted tile Y axis (TMS convention). 1 for true/yes. 0 for false/no.}

\item{OUTPUT_DIRECTORY}{\code{folderDestination} - Output directory. Path for an existing or new folder.}

\item{OUTPUT_HTML}{\code{fileDestination} - Output html (Leaflet). Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Generate XYZ tiles (Directory) (qgis:tilesxyzdirectory)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_DIRECTORY - outputFolder - Output directory
\item OUTPUT_HTML - outputHtml - Output html
}
}
}
