% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_tilesxyzmbtiles.R
\name{qgis_tilesxyzmbtiles}
\alias{qgis_tilesxyzmbtiles}
\title{QGIS algorithm Generate XYZ tiles (MBTiles)}
\usage{
qgis_tilesxyzmbtiles(
  EXTENT = qgisprocess::qgis_default_value(),
  ZOOM_MIN = qgisprocess::qgis_default_value(),
  ZOOM_MAX = qgisprocess::qgis_default_value(),
  DPI = qgisprocess::qgis_default_value(),
  BACKGROUND_COLOR = qgisprocess::qgis_default_value(),
  TILE_FORMAT = qgisprocess::qgis_default_value(),
  QUALITY = qgisprocess::qgis_default_value(),
  METATILESIZE = qgisprocess::qgis_default_value(),
  OUTPUT_FILE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{EXTENT}{\code{extent} - Extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{ZOOM_MIN}{\code{number} - Minimum zoom. A numeric value.}

\item{ZOOM_MAX}{\code{number} - Maximum zoom. A numeric value.}

\item{DPI}{\code{number} - DPI. A numeric value.}

\item{BACKGROUND_COLOR}{\code{color} - Background color. String representation of color, e.g #ff0000 or rgba(200,100,50,0.8).}

\item{TILE_FORMAT}{\code{enum}  of \verb{("PNG", "JPG")} - Tile format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{QUALITY}{\code{number} - Quality (JPG only). A numeric value.}

\item{METATILESIZE}{\code{number} - Metatile size. A numeric value.}

\item{OUTPUT_FILE}{\code{fileDestination} - Output file (for MBTiles). Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Generate XYZ tiles (MBTiles) (qgis:tilesxyzmbtiles)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_FILE - outputFile - Output file (for MBTiles)
}
}
}
