% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_tininterpolation.R
\name{qgis_tininterpolation}
\alias{qgis_tininterpolation}
\title{QGIS algorithm - TIN interpolation}
\usage{
qgis_tininterpolation(
  INTERPOLATION_DATA = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  PIXEL_SIZE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  TRIANGULATION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INTERPOLATION_DATA}{\code{idw_interpolation_data} - Input layer(s). .}

\item{METHOD}{\code{enum}  of \verb{("Linear", "Clough-Toucher (cubic)")} - Interpolation method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXTENT}{\code{extent} - Extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{PIXEL_SIZE}{\code{number} - Output raster size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{rasterDestination} - Interpolated. Path for new raster layer.}

\item{TRIANGULATION}{\code{sink} - Triangulation. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS TIN interpolation (qgis:tininterpolation). Generates a Triangulated Irregular Network (TIN) interpolation of a point vector layer. Generates a Triangulated Irregular Network (TIN) interpolation of a point vector layer. With the TIN method you can create a surface formed by triangles of nearest neighbor points. To do this, circumcircles around selected sample points are created and their intersections are connected to a network of non overlapping and as compact as possible triangles. The resulting surfaces are not smooth. The algorithm creates both the raster layer of the interpolated values and the vector line layer with the triangulation boundaries.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Interpolated
\item TRIANGULATION - outputVector - Triangulation
}
}
}
