% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_tinmeshcreation.R
\name{qgis_tinmeshcreation}
\alias{qgis_tinmeshcreation}
\title{QGIS algorithm - TIN Mesh Creation}
\usage{
qgis_tinmeshcreation(
  SOURCE_DATA = qgisprocess:::qgis_default_value(),
  MESH_FORMAT = qgisprocess:::qgis_default_value(),
  CRS_OUTPUT = qgisprocess:::qgis_default_value(),
  OUTPUT_MESH = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SOURCE_DATA}{\code{tininputlayers} - Input layers. .}

\item{MESH_FORMAT}{\code{enum}  of \verb{("2DM", "SELAFIN", "PLY", "Ugrid", "Mike21")} - Output format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CRS_OUTPUT}{\code{crs} - Output coordinate system. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT_MESH}{\code{fileDestination} - Output file. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) TIN Mesh Creation (native:tinmeshcreation). Creates a TIN mesh layer from vector layers. This algorithm creates a TIN mesh layer from vector layers.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_MESH - outputFile - Output file
}
}
}
