% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_transect.R
\name{qgis_transect}
\alias{qgis_transect}
\title{QGIS algorithm Transect}
\usage{
qgis_transect(
  INPUT = qgisprocess::qgis_default_value(),
  LENGTH = qgisprocess::qgis_default_value(),
  ANGLE = qgisprocess::qgis_default_value(),
  SIDE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{LENGTH}{\code{distance} - Length of the transect. A numeric value.}

\item{ANGLE}{\code{number} - Angle in degrees from the original line at the vertices. A numeric value.}

\item{SIDE}{\code{enum}  of \verb{("Left", "Right", "Both")} - Side to create the transects. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Transect. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Transect (native:transect)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Transect
}
}
}
