% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_transferannotationsfrommain.R
\name{qgis_transferannotationsfrommain}
\alias{qgis_transferannotationsfrommain}
\title{QGIS algorithm Transfer annotations from main layer}
\usage{
qgis_transferannotationsfrommain(
  LAYER_NAME = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{LAYER_NAME}{\code{string} - New layer name. String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Transfer annotations from main layer (native:transferannotationsfrommain)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputLayer - New annotation layer
}
}
}
