% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_truncatetable.R
\name{qgis_truncatetable}
\alias{qgis_truncatetable}
\title{QGIS algorithm Truncate table}
\usage{
qgis_truncatetable(
  INPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{vector} - Input layer. Path to a vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Truncate table (native:truncatetable)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Truncated layer
}
}
}
