% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_updatelayermetadata.R
\name{qgis_updatelayermetadata}
\alias{qgis_updatelayermetadata}
\title{QGIS algorithm - Update layer metadata}
\usage{
qgis_updatelayermetadata(
  SOURCE = qgisprocess:::qgis_default_value(),
  TARGET = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SOURCE}{\code{layer} - Source layer. Path to a vector, raster or mesh layer.}

\item{TARGET}{\code{layer} - Target layer. Path to a vector, raster or mesh layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Update layer metadata (native:updatelayermetadata). Copies all non-empty metadata fields from one layer to another. This algorithm copies all non-empty metadata fields from a source layer to a target layer.  Leaves empty input fields unchanged in the target.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputLayer - Updated layer
}
}
}
