% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_voronoipolygons.R
\name{qgis_voronoipolygons}
\alias{qgis_voronoipolygons}
\title{QGIS algorithm Voronoi polygons}
\usage{
qgis_voronoipolygons(
  INPUT = qgisprocess::qgis_default_value(),
  BUFFER = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{BUFFER}{\code{number} - Buffer region (\% of extent). A numeric value.}

\item{OUTPUT}{\code{sink} - Voronoi polygons. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Voronoi polygons (qgis:voronoipolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Voronoi polygons
}
}
}
