% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_writevectortiles_mbtiles.R
\name{qgis_writevectortiles_mbtiles}
\alias{qgis_writevectortiles_mbtiles}
\title{QGIS algorithm - Write Vector Tiles (MBTiles)}
\usage{
qgis_writevectortiles_mbtiles(
  OUTPUT = qgisprocess:::qgis_default_value(),
  LAYERS = qgisprocess:::qgis_default_value(),
  MIN_ZOOM = qgisprocess:::qgis_default_value(),
  MAX_ZOOM = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  META_NAME = qgisprocess:::qgis_default_value(),
  META_DESCRIPTION = qgisprocess:::qgis_default_value(),
  META_ATTRIBUTION = qgisprocess:::qgis_default_value(),
  META_VERSION = qgisprocess:::qgis_default_value(),
  META_TYPE = qgisprocess:::qgis_default_value(),
  META_CENTER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{OUTPUT}{\code{vectorTileDestination} - Destination MBTiles. Path for new vector tile layer.}

\item{LAYERS}{\code{vectortilewriterlayers} - Input layers. .}

\item{MIN_ZOOM}{\code{number} - Minimum zoom level. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAX_ZOOM}{\code{number} - Maximum zoom level. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EXTENT}{\code{extent} - Extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{META_NAME}{\code{string} - Metadata: Name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{META_DESCRIPTION}{\code{string} - Metadata: Description. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{META_ATTRIBUTION}{\code{string} - Metadata: Attribution. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{META_VERSION}{\code{string} - Metadata: Version. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{META_TYPE}{\code{string} - Metadata: Type. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{META_CENTER}{\code{string} - Metadata: Center. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Write Vector Tiles (MBTiles) (native:writevectortiles_mbtiles). This algorithm exports one or more vector layers to vector tiles - a data format optimized for fast map rendering and small data size.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVectorTile - Destination MBTiles
}
}
}
