% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_writevectortiles_xyz.R
\name{qgis_writevectortiles_xyz}
\alias{qgis_writevectortiles_xyz}
\title{QGIS algorithm - Write Vector Tiles (XYZ)}
\usage{
qgis_writevectortiles_xyz(
  OUTPUT_DIRECTORY = qgisprocess:::qgis_default_value(),
  XYZ_TEMPLATE = qgisprocess:::qgis_default_value(),
  LAYERS = qgisprocess:::qgis_default_value(),
  MIN_ZOOM = qgisprocess:::qgis_default_value(),
  MAX_ZOOM = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{OUTPUT_DIRECTORY}{\code{folderDestination} - Output directory. Path for an existing or new folder.}

\item{XYZ_TEMPLATE}{\code{string} - File template. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LAYERS}{\code{vectortilewriterlayers} - Input layers. .}

\item{MIN_ZOOM}{\code{number} - Minimum zoom level. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAX_ZOOM}{\code{number} - Maximum zoom level. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EXTENT}{\code{extent} - Extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Write Vector Tiles (XYZ) (native:writevectortiles_xyz). Exports one or more vector layers to vector tiles. This algorithm exports one or more vector layers to vector tiles - a data format optimized for fast map rendering and small data size.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_DIRECTORY - outputFolder - Output directory
}
}
}
