% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_zonalminmaxpoint.R
\name{qgis_zonalminmaxpoint}
\alias{qgis_zonalminmaxpoint}
\title{QGIS algorithm - Zonal minimum/maximum point}
\usage{
qgis_zonalminmaxpoint(
  INPUT = qgisprocess:::qgis_default_value(),
  INPUT_RASTER = qgisprocess:::qgis_default_value(),
  RASTER_BAND = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{INPUT_RASTER}{\code{raster} - Raster layer. Path to a raster layer.}

\item{RASTER_BAND}{\code{band} - Raster band. Integer value representing an existing raster band number.}

\item{OUTPUT}{\code{sink} - Zonal Extrema. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Zonal minimum/maximum point (native:zonalminmaxpoint). Extracts point locations for minimum and maximum raster values within polygon zones. This algorithm extracts point features corresponding to the minimum and maximum pixel values contained within polygon zones.  The output will contain one point feature for the minimum and one for the maximum raster value for every individual zonal feature from a polygon layer.  The created point layer will be in the same spatial reference system as the selected raster layer.  ---------------- Notes ----------------   - This algorithm drops existing primary keys or FID values and regenerates them in output layers.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Zonal Extrema
}
}
}
