% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_zonalstatistics.R
\name{qgis_zonalstatistics}
\alias{qgis_zonalstatistics}
\title{QGIS algorithm Zonal statistics (in place)}
\usage{
qgis_zonalstatistics(
  INPUT_RASTER = qgisprocess::qgis_default_value(),
  RASTER_BAND = qgisprocess::qgis_default_value(),
  INPUT_VECTOR = qgisprocess::qgis_default_value(),
  COLUMN_PREFIX = qgisprocess::qgis_default_value(),
  STATISTICS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_RASTER}{\code{raster} - Raster layer. Path to a raster layer.}

\item{RASTER_BAND}{\code{band} - Raster band. Integer value representing an existing raster band number.}

\item{INPUT_VECTOR}{\code{vector} - Vector layer containing zones. Path to a vector layer.}

\item{COLUMN_PREFIX}{\code{string} - Output column prefix. String value.}

\item{STATISTICS}{\code{enum}  of \verb{("Count", "Sum", "Mean", "Median", "St dev", "Minimum", "Maximum", "Range", "Minority", "Majority", "Variety", "Variance")} - Statistics to calculate. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Zonal statistics (in place) (native:zonalstatistics)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INPUT_VECTOR - outputVector - Zonal statistics
}
}
}
