% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_zonalstatisticsfb.R
\name{qgis_zonalstatisticsfb}
\alias{qgis_zonalstatisticsfb}
\title{QGIS algorithm - Zonal statistics}
\usage{
qgis_zonalstatisticsfb(
  INPUT = qgisprocess:::qgis_default_value(),
  INPUT_RASTER = qgisprocess:::qgis_default_value(),
  RASTER_BAND = qgisprocess:::qgis_default_value(),
  COLUMN_PREFIX = qgisprocess:::qgis_default_value(),
  STATISTICS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{INPUT_RASTER}{\code{raster} - Raster layer. Path to a raster layer.}

\item{RASTER_BAND}{\code{band} - Raster band. Integer value representing an existing raster band number.}

\item{COLUMN_PREFIX}{\code{string} - Output column prefix. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{STATISTICS}{\code{enum}  of \verb{("Count", "Sum", "Mean", "Median", "St dev", "Minimum", "Maximum", "Range", "Minority", "Majority", "Variety", "Variance")} - Statistics to calculate. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Zonal Statistics. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Zonal statistics (native:zonalstatisticsfb). Calculates statistics of a raster layer for each feature of an overlapping polygon vector layer. This algorithm calculates statistics of a raster layer for each feature of an overlapping polygon vector layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Zonal Statistics
}
}
}
