% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_accumulatedcostanisotropic.R
\name{saga_accumulatedcostanisotropic}
\alias{saga_accumulatedcostanisotropic}
\title{QGIS algorithm Accumulated cost (anisotropic)}
\usage{
saga_accumulatedcostanisotropic(
  COST = qgisprocess::qgis_default_value(),
  DIRECTION = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  K = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  ACCCOST = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{COST}{\code{raster} - Cost Grid. Path to a raster layer.}

\item{DIRECTION}{\code{raster} - Direction of max cost. Path to a raster layer.}

\item{POINTS}{\code{raster} - Destination Points. Path to a raster layer.}

\item{K}{\code{number} - k factor. A numeric value.}

\item{THRESHOLD}{\code{number} - Threshold for different route. A numeric value.}

\item{ACCCOST}{\code{rasterDestination} - Accumulated Cost. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Accumulated cost (anisotropic) (saga:accumulatedcostanisotropic)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ACCCOST - outputRaster - Accumulated Cost
}
}
}
