% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_accumulatedcostisotropic.R
\name{saga_accumulatedcostisotropic}
\alias{saga_accumulatedcostisotropic}
\title{QGIS algorithm Accumulated cost (isotropic)}
\usage{
saga_accumulatedcostisotropic(
  COST = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  ACCCOST = qgisprocess::qgis_default_value(),
  CLOSESTPT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{COST}{\code{raster} - Cost Grid. Path to a raster layer.}

\item{POINTS}{\code{raster} - Destination Points. Path to a raster layer.}

\item{THRESHOLD}{\code{number} - Threshold for different route. A numeric value.}

\item{ACCCOST}{\code{rasterDestination} - Accumulated Cost. Path for new raster layer.}

\item{CLOSESTPT}{\code{rasterDestination} - Closest Point. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Accumulated cost (isotropic) (saga:accumulatedcostisotropic)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ACCCOST - outputRaster - Accumulated Cost
\item CLOSESTPT - outputRaster - Closest Point
}
}
}
