% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_accumulationfunctions.R
\name{saga_accumulationfunctions}
\alias{saga_accumulationfunctions}
\title{QGIS algorithm Accumulation functions}
\usage{
saga_accumulationfunctions(
  SURFACE = qgisprocess::qgis_default_value(),
  INPUT = qgisprocess::qgis_default_value(),
  STATE_IN = qgisprocess::qgis_default_value(),
  CONTROL = qgisprocess::qgis_default_value(),
  CTRL_LINEAR = qgisprocess::qgis_default_value(),
  FLUX = qgisprocess::qgis_default_value(),
  STATE_OUT = qgisprocess::qgis_default_value(),
  OPERATION = qgisprocess::qgis_default_value(),
  LINEAR = qgisprocess::qgis_default_value(),
  THRES_LINEAR = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SURFACE}{\code{raster} - Surface. Path to a raster layer.}

\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{STATE_IN}{\code{raster} - State t. Path to a raster layer.}

\item{CONTROL}{\code{raster} - Operation Control. Path to a raster layer.}

\item{CTRL_LINEAR}{\code{raster} - Linear Flow Control Grid. Path to a raster layer.}

\item{FLUX}{\code{rasterDestination} - Flux. Path for new raster layer.}

\item{STATE_OUT}{\code{rasterDestination} - State t + 1. Path for new raster layer.}

\item{OPERATION}{\code{enum}  of \verb{("[0] accuflux", "[1] accucapacityflux / state", "[2] accufractionflux / state", "[3] accuthresholdflux / state", "[4] accutriggerflux / state")} - Operation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LINEAR}{\code{boolean} - Switch to Linear Flow. 1 for true/yes. 0 for false/no.}

\item{THRES_LINEAR}{\code{number} - Threshold Linear Flow. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Accumulation functions (saga:accumulationfunctions)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLUX - outputRaster - Flux
\item STATE_OUT - outputRaster - State t + 1
}
}
}
