% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_addcoordinatestopoints.R
\name{saga_addcoordinatestopoints}
\alias{saga_addcoordinatestopoints}
\title{QGIS algorithm Add coordinates to points}
\usage{
saga_addcoordinatestopoints(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Points. Path to a vector layer.}

\item{OUTPUT}{\code{vectorDestination} - Points with coordinates. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Add coordinates to points (saga:addcoordinatestopoints)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Points with coordinates
}
}
}
