% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_addindicatorfieldsforcategories.R
\name{saga_addindicatorfieldsforcategories}
\alias{saga_addindicatorfieldsforcategories}
\title{QGIS algorithm Add indicator fields for categories}
\usage{
saga_addindicatorfieldsforcategories(
  TABLE = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  OUT_TABLE = qgisprocess::qgis_default_value(),
  OUT_SHAPES = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TABLE}{\code{source} - Table. Path to a vector layer.}

\item{FIELD}{\code{field} - Categories. The name of an existing field. ; delimited list of existing field names.}

\item{OUT_TABLE}{\code{source} - Output table with field(s) deleted. Path to a vector layer.}

\item{OUT_SHAPES}{\code{vectorDestination} - Output shapes with field(s) deleted. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Add indicator fields for categories (saga:addindicatorfieldsforcategories)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUT_SHAPES - outputVector - Output shapes with field(s) deleted
}
}
}
