% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_addpointattributestopolygons.R
\name{saga_addpointattributestopolygons}
\alias{saga_addpointattributestopolygons}
\title{QGIS algorithm Add point attributes to polygons}
\usage{
saga_addpointattributestopolygons(
  INPUT = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  FIELDS = qgisprocess::qgis_default_value(),
  ADD_LOCATION_INFO = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Polygons. Path to a vector layer.}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{FIELDS}{\code{field} - Attributes. The name of an existing field. ; delimited list of existing field names.}

\item{ADD_LOCATION_INFO}{\code{boolean} - Add location info. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Add point attributes to polygons (saga:addpointattributestopolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Result
}
}
}
