% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_addpolygonattributestopoints.R
\name{saga_addpolygonattributestopoints}
\alias{saga_addpolygonattributestopoints}
\title{QGIS algorithm Add polygon attributes to points}
\usage{
saga_addpolygonattributestopoints(
  INPUT = qgisprocess::qgis_default_value(),
  POLYGONS = qgisprocess::qgis_default_value(),
  FIELDS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Points. Path to a vector layer.}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{FIELDS}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Add polygon attributes to points (saga:addpolygonattributestopoints)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Result
}
}
}
