% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_aggregatepointobservations.R
\name{saga_aggregatepointobservations}
\alias{saga_aggregatepointobservations}
\title{QGIS algorithm Aggregate point observations}
\usage{
saga_aggregatepointobservations(
  REFERENCE = qgisprocess::qgis_default_value(),
  REFERENCE_ID = qgisprocess::qgis_default_value(),
  OBSERVATIONS = qgisprocess::qgis_default_value(),
  X = qgisprocess::qgis_default_value(),
  Y = qgisprocess::qgis_default_value(),
  TRACK = qgisprocess::qgis_default_value(),
  DATE = qgisprocess::qgis_default_value(),
  TIME = qgisprocess::qgis_default_value(),
  PARAMETER = qgisprocess::qgis_default_value(),
  EPS_TIME = qgisprocess::qgis_default_value(),
  EPS_SPACE = qgisprocess::qgis_default_value(),
  AGGREGATED = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{REFERENCE}{\code{source} - Reference Points. Path to a vector layer.}

\item{REFERENCE_ID}{\code{field} - ID. The name of an existing field. ; delimited list of existing field names.}

\item{OBSERVATIONS}{\code{source} - Observations. Path to a vector layer.}

\item{X}{\code{field} - X. The name of an existing field. ; delimited list of existing field names.}

\item{Y}{\code{field} - Y. The name of an existing field. ; delimited list of existing field names.}

\item{TRACK}{\code{field} - Track. The name of an existing field. ; delimited list of existing field names.}

\item{DATE}{\code{field} - Date. The name of an existing field. ; delimited list of existing field names.}

\item{TIME}{\code{field} - Time. The name of an existing field. ; delimited list of existing field names.}

\item{PARAMETER}{\code{field} - Parameter. The name of an existing field. ; delimited list of existing field names.}

\item{EPS_TIME}{\code{number} - Maximum Time Span (Seconds). A numeric value.}

\item{EPS_SPACE}{\code{number} - Maximum Distance. A numeric value.}

\item{AGGREGATED}{\code{vectorDestination} - Aggregated. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Aggregate point observations (saga:aggregatepointobservations)
}
\details{
\subsection{Outputs description}{
\itemize{
\item AGGREGATED - outputVector - Aggregated
}
}
}
