% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_analyticalhierarchyprocess.R
\name{saga_analyticalhierarchyprocess}
\alias{saga_analyticalhierarchyprocess}
\title{QGIS algorithm Analytical hierarchy process}
\usage{
saga_analyticalhierarchyprocess(
  GRIDS = qgisprocess::qgis_default_value(),
  TABLE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Input Grids. .}

\item{TABLE}{\code{source} - Pairwise Comparisons Table. Path to a vector layer.}

\item{OUTPUT}{\code{rasterDestination} - Output Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Analytical hierarchy process (saga:analyticalhierarchyprocess)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output Grid
}
}
}
