% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_analyticalhillshading.R
\name{saga_analyticalhillshading}
\alias{saga_analyticalhillshading}
\title{QGIS algorithm Analytical hillshading}
\usage{
saga_analyticalhillshading(
  ELEVATION = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  AZIMUTH = qgisprocess::qgis_default_value(),
  DECLINATION = qgisprocess::qgis_default_value(),
  EXAGGERATION = qgisprocess::qgis_default_value(),
  SHADE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Standard", "[1] Standard (max. 90Degree)", "[2] Combined Shading", "[3] Ray Tracing")} - Shading Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{AZIMUTH}{\code{number} - Azimuth (Degree). A numeric value.}

\item{DECLINATION}{\code{number} - Declination (Degree). A numeric value.}

\item{EXAGGERATION}{\code{number} - Exaggeration. A numeric value.}

\item{SHADE}{\code{rasterDestination} - Analytical Hillshading. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Analytical hillshading (saga:analyticalhillshading)
}
\details{
\subsection{Outputs description}{
\itemize{
\item SHADE - outputRaster - Analytical Hillshading
}
}
}
