% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_angmap.R
\name{saga_angmap}
\alias{saga_angmap}
\title{QGIS algorithm Angmap}
\usage{
saga_angmap(
  DEM = qgisprocess::qgis_default_value(),
  C = qgisprocess::qgis_default_value(),
  D = qgisprocess::qgis_default_value(),
  fB = qgisprocess::qgis_default_value(),
  fC = qgisprocess::qgis_default_value(),
  E = qgisprocess::qgis_default_value(),
  F = qgisprocess::qgis_default_value(),
  G = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{C}{\code{raster} - Dip grid (degrees). Path to a raster layer.}

\item{D}{\code{raster} - Dip direction grid (degrees). Path to a raster layer.}

\item{fB}{\code{number} - Global structure dip (degrees). A numeric value.}

\item{fC}{\code{number} - Global structure dip direction (degrees). A numeric value.}

\item{E}{\code{rasterDestination} - Angle. Path for new raster layer.}

\item{F}{\code{rasterDestination} - CL dipdir. Path for new raster layer.}

\item{G}{\code{rasterDestination} - CL dip. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Angmap (saga:angmap)
}
\details{
\subsection{Outputs description}{
\itemize{
\item E - outputRaster - Angle
\item F - outputRaster - CL dipdir
\item G - outputRaster - CL dip
}
}
}
