% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/saga_artificialneuralnetworkclassificationopencv.R
\name{saga_artificialneuralnetworkclassificationopencv}
\alias{saga_artificialneuralnetworkclassificationopencv}
\title{QGIS algorithm Artificial neural network classification (opencv)}
\usage{
saga_artificialneuralnetworkclassificationopencv(
  FEATURES = qgisprocess::qgis_default_value(),
  NORMALIZE = qgisprocess::qgis_default_value(),
  TRAIN_AREAS = qgisprocess::qgis_default_value(),
  TRAIN_CLASS = qgisprocess::qgis_default_value(),
  CLASSES = qgisprocess::qgis_default_value(),
  ANN_LAYERS = qgisprocess::qgis_default_value(),
  ANN_NEURONS = qgisprocess::qgis_default_value(),
  ANN_MAXITER = qgisprocess::qgis_default_value(),
  ANN_EPSILON = qgisprocess::qgis_default_value(),
  ANN_ACTIVATION = qgisprocess::qgis_default_value(),
  ANN_ACT_ALPHA = qgisprocess::qgis_default_value(),
  ANN_ACT_BETA = qgisprocess::qgis_default_value(),
  ANN_PROPAGATION = qgisprocess::qgis_default_value(),
  ANN_RP_DW0 = qgisprocess::qgis_default_value(),
  ANN_RP_DW_PLUS = qgisprocess::qgis_default_value(),
  ANN_RP_DW_MINUS = qgisprocess::qgis_default_value(),
  ANN_RP_DW_MIN = qgisprocess::qgis_default_value(),
  ANN_RP_DW_MAX = qgisprocess::qgis_default_value(),
  ANN_BP_DW = qgisprocess::qgis_default_value(),
  ANN_BP_MOMENT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{FEATURES}{\code{multilayer} - Features. .}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no.}

\item{TRAIN_AREAS}{\code{source} - Training Areas. Path to a vector layer.}

\item{TRAIN_CLASS}{\code{source} - Class Identifier. Path to a vector layer.}

\item{CLASSES}{\code{rasterDestination} - Classification. Path for new raster layer.}

\item{ANN_LAYERS}{\code{number} - Number of Layers. A numeric value.}

\item{ANN_NEURONS}{\code{number} - Number of Neurons. A numeric value.}

\item{ANN_MAXITER}{\code{number} - Maximum Number of Iterations. A numeric value.}

\item{ANN_EPSILON}{\code{number} - Error Change (Epsilon). A numeric value.}

\item{ANN_ACTIVATION}{\code{enum}  of \verb{("[0] Identity", "[1] Sigmoid", "[2] Gaussian")} - Activation Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ANN_ACT_ALPHA}{\code{number} - Function's Alpha. A numeric value.}

\item{ANN_ACT_BETA}{\code{number} - Function's Beta. A numeric value.}

\item{ANN_PROPAGATION}{\code{enum}  of \verb{("[0] resilient propagation", "[1] back propagation")} - Training Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ANN_RP_DW0}{\code{number} - Initial Update Value. A numeric value.}

\item{ANN_RP_DW_PLUS}{\code{number} - Increase Factor. A numeric value.}

\item{ANN_RP_DW_MINUS}{\code{number} - Decrease Factor. A numeric value.}

\item{ANN_RP_DW_MIN}{\code{number} - Lower Value Update Limit. A numeric value.}

\item{ANN_RP_DW_MAX}{\code{number} - Upper Value Update Limit. A numeric value.}

\item{ANN_BP_DW}{\code{number} - Weight Gradient Term. A numeric value.}

\item{ANN_BP_MOMENT}{\code{number} - Moment Term. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Artificial neural network classification (opencv) (saga:artificialneuralnetworkclassificationopencv)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Classification
}
}
}
