% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_aspectslopegrid.R
\name{saga_aspectslopegrid}
\alias{saga_aspectslopegrid}
\title{QGIS algorithm Aspect-slope grid}
\usage{
saga_aspectslopegrid(
  ASPECT = qgisprocess:::qgis_default_value(),
  SLOPE = qgisprocess:::qgis_default_value(),
  ASPECT_SLOPE = qgisprocess:::qgis_default_value(),
  LUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ASPECT}{\code{raster} - Aspect. Path to a raster layer.}

\item{SLOPE}{\code{raster} - Slope. Path to a raster layer.}

\item{ASPECT_SLOPE}{\code{rasterDestination} - Aspect-Slope. Path for new raster layer.}

\item{LUT}{\code{source} - Lookup Table. Path to a vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Aspect-slope grid (saga:aspectslopegrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ASPECT_SLOPE - outputRaster - Aspect-Slope
}
}
}
