% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_automatedcloudcoverassessment.R
\name{saga_automatedcloudcoverassessment}
\alias{saga_automatedcloudcoverassessment}
\title{QGIS algorithm Automated cloud cover assessment}
\usage{
saga_automatedcloudcoverassessment(
  BAND2 = qgisprocess::qgis_default_value(),
  BAND3 = qgisprocess::qgis_default_value(),
  BAND4 = qgisprocess::qgis_default_value(),
  BAND5 = qgisprocess::qgis_default_value(),
  BAND6 = qgisprocess::qgis_default_value(),
  CLOUD = qgisprocess::qgis_default_value(),
  FILTER = qgisprocess::qgis_default_value(),
  B56C = qgisprocess::qgis_default_value(),
  B45R = qgisprocess::qgis_default_value(),
  CSIG = qgisprocess::qgis_default_value(),
  PASS2 = qgisprocess::qgis_default_value(),
  SHADOW = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{BAND2}{\code{raster} - Landsat Band 2. Path to a raster layer.}

\item{BAND3}{\code{raster} - Landsat Band 3. Path to a raster layer.}

\item{BAND4}{\code{raster} - Landsat Band 4. Path to a raster layer.}

\item{BAND5}{\code{raster} - Landsat Band 5. Path to a raster layer.}

\item{BAND6}{\code{raster} - Landsat Band 6. Path to a raster layer.}

\item{CLOUD}{\code{rasterDestination} - Cloud Cover. Path for new raster layer.}

\item{FILTER}{\code{boolean} - Apply post-processing filter to remove small holes. 1 for true/yes. 0 for false/no.}

\item{B56C}{\code{number} - B56 Composite (step 6). A numeric value.}

\item{B45R}{\code{number} - B45 Ratio: Desert detection (step 10). A numeric value.}

\item{CSIG}{\code{boolean} - Always use cloud signature (step 14). 1 for true/yes. 0 for false/no.}

\item{PASS2}{\code{boolean} - Bypass second-pass processing, and merge warm (not ambiguous) and cold clouds. 1 for true/yes. 0 for false/no.}

\item{SHADOW}{\code{boolean} - Include a category for cloud shadows. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Automated cloud cover assessment (saga:automatedcloudcoverassessment)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLOUD - outputRaster - Cloud Cover
}
}
}
