% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_averagewithmask1.R
\name{saga_averagewithmask1}
\alias{saga_averagewithmask1}
\title{QGIS algorithm Average with mask 1}
\usage{
saga_averagewithmask1(
  INPUT = qgisprocess::qgis_default_value(),
  MASK = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  V = qgisprocess::qgis_default_value(),
  RX = qgisprocess::qgis_default_value(),
  RY = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{MASK}{\code{raster} - Mask Grid. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - AWM1 Grid. Path for new raster layer.}

\item{V}{\code{number} - Mask value. A numeric value.}

\item{RX}{\code{number} - Radius X. A numeric value.}

\item{RY}{\code{number} - Radius Y. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Average with mask 1 (saga:averagewithmask1)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - AWM1 Grid
}
}
}
