% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_averagewiththereshold1.R
\name{saga_averagewiththereshold1}
\alias{saga_averagewiththereshold1}
\title{QGIS algorithm Average with thereshold 1}
\usage{
saga_averagewiththereshold1(
  INPUT = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  RX = qgisprocess::qgis_default_value(),
  RY = qgisprocess::qgis_default_value(),
  THRESH = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - AWT Grid. Path for new raster layer.}

\item{RX}{\code{number} - Radius X. A numeric value.}

\item{RY}{\code{number} - Radius Y. A numeric value.}

\item{THRESH}{\code{number} - Threshold. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Average with thereshold 1 (saga:averagewiththereshold1)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - AWT Grid
}
}
}
