% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_averagewiththereshold2.R
\name{saga_averagewiththereshold2}
\alias{saga_averagewiththereshold2}
\title{QGIS algorithm Average with thereshold 2}
\usage{
saga_averagewiththereshold2(
  INPUT = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  RX = qgisprocess::qgis_default_value(),
  RY = qgisprocess::qgis_default_value(),
  THRESH = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - AWT Grid. Path for new raster layer.}

\item{RX}{\code{number} - Radius X. A numeric value.}

\item{RY}{\code{number} - Radius Y. A numeric value.}

\item{THRESH}{\code{number} - Threshold. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Average with thereshold 2 (saga:averagewiththereshold2)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - AWT Grid
}
}
}
