% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_basicterrainanalysis.R
\name{saga_basicterrainanalysis}
\alias{saga_basicterrainanalysis}
\title{QGIS algorithm Basic terrain analysis}
\usage{
saga_basicterrainanalysis(
  ELEVATION = qgisprocess::qgis_default_value(),
  SHADE = qgisprocess::qgis_default_value(),
  SLOPE = qgisprocess::qgis_default_value(),
  ASPECT = qgisprocess::qgis_default_value(),
  HCURV = qgisprocess::qgis_default_value(),
  VCURV = qgisprocess::qgis_default_value(),
  CONVERGENCE = qgisprocess::qgis_default_value(),
  SINKS = qgisprocess::qgis_default_value(),
  FLOW = qgisprocess::qgis_default_value(),
  WETNESS = qgisprocess::qgis_default_value(),
  LSFACTOR = qgisprocess::qgis_default_value(),
  CHANNELS = qgisprocess::qgis_default_value(),
  BASINS = qgisprocess::qgis_default_value(),
  CHNL_BASE = qgisprocess::qgis_default_value(),
  CHNL_DIST = qgisprocess::qgis_default_value(),
  VALL_DEPTH = qgisprocess::qgis_default_value(),
  RSP = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{SHADE}{\code{rasterDestination} - Analytical Hillshading. Path for new raster layer.}

\item{SLOPE}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{ASPECT}{\code{rasterDestination} - Aspect. Path for new raster layer.}

\item{HCURV}{\code{rasterDestination} - Plan Curvature. Path for new raster layer.}

\item{VCURV}{\code{rasterDestination} - Profile Curvature. Path for new raster layer.}

\item{CONVERGENCE}{\code{rasterDestination} - Convergence Index. Path for new raster layer.}

\item{SINKS}{\code{rasterDestination} - Closed Depressions. Path for new raster layer.}

\item{FLOW}{\code{rasterDestination} - Total Catchment Area. Path for new raster layer.}

\item{WETNESS}{\code{rasterDestination} - Topographic Wetness Index. Path for new raster layer.}

\item{LSFACTOR}{\code{rasterDestination} - LS-Factor. Path for new raster layer.}

\item{CHANNELS}{\code{vectorDestination} - Channel Network. Path for new vector layer.}

\item{BASINS}{\code{vectorDestination} - Drainage Basins. Path for new vector layer.}

\item{CHNL_BASE}{\code{rasterDestination} - Channel Network Base Level. Path for new raster layer.}

\item{CHNL_DIST}{\code{rasterDestination} - Channel Network Distance. Path for new raster layer.}

\item{VALL_DEPTH}{\code{rasterDestination} - Valley Depth. Path for new raster layer.}

\item{RSP}{\code{rasterDestination} - Relative Slope Position. Path for new raster layer.}

\item{THRESHOLD}{\code{number} - Channel Density. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Basic terrain analysis (saga:basicterrainanalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ASPECT - outputRaster - Aspect
\item BASINS - outputVector - Drainage Basins
\item CHANNELS - outputVector - Channel Network
\item CHNL_BASE - outputRaster - Channel Network Base Level
\item CHNL_DIST - outputRaster - Channel Network Distance
\item CONVERGENCE - outputRaster - Convergence Index
\item FLOW - outputRaster - Total Catchment Area
\item HCURV - outputRaster - Plan Curvature
\item LSFACTOR - outputRaster - LS-Factor
\item RSP - outputRaster - Relative Slope Position
\item SHADE - outputRaster - Analytical Hillshading
\item SINKS - outputRaster - Closed Depressions
\item SLOPE - outputRaster - Slope
\item VALL_DEPTH - outputRaster - Valley Depth
\item VCURV - outputRaster - Profile Curvature
\item WETNESS - outputRaster - Topographic Wetness Index
}
}
}
