% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_binaryerosionreconstruction.R
\name{saga_binaryerosionreconstruction}
\alias{saga_binaryerosionreconstruction}
\title{QGIS algorithm Binary erosion-reconstruction}
\usage{
saga_binaryerosionreconstruction(
  INPUT_GRID = qgisprocess::qgis_default_value(),
  OUTPUT_GRID = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT_GRID}{\code{raster} - Input Grid. Path to a raster layer.}

\item{OUTPUT_GRID}{\code{rasterDestination} - Output Grid. Path for new raster layer.}

\item{RADIUS}{\code{number} - Filter Size (Radius). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Binary erosion-reconstruction (saga:binaryerosionreconstruction)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_GRID - outputRaster - Output Grid
}
}
}
