% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_boostingclassificationopencv.R
\name{saga_boostingclassificationopencv}
\alias{saga_boostingclassificationopencv}
\title{QGIS algorithm Boosting classification (opencv)}
\usage{
saga_boostingclassificationopencv(
  FEATURES = qgisprocess::qgis_default_value(),
  NORMALIZE = qgisprocess::qgis_default_value(),
  TRAIN_AREAS = qgisprocess::qgis_default_value(),
  TRAIN_CLASS = qgisprocess::qgis_default_value(),
  CLASSES = qgisprocess::qgis_default_value(),
  MAX_DEPTH = qgisprocess::qgis_default_value(),
  MIN_SAMPLES = qgisprocess::qgis_default_value(),
  MAX_CATEGRS = qgisprocess::qgis_default_value(),
  arg_1SE_RULE = qgisprocess::qgis_default_value(),
  TRUNC_PRUNED = qgisprocess::qgis_default_value(),
  REG_ACCURACY = qgisprocess::qgis_default_value(),
  WEAK_COUNT = qgisprocess::qgis_default_value(),
  WGT_TRIM_RATE = qgisprocess::qgis_default_value(),
  BOOST_TYPE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{FEATURES}{\code{multilayer} - Features. .}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no.}

\item{TRAIN_AREAS}{\code{source} - Training Areas. Path to a vector layer.}

\item{TRAIN_CLASS}{\code{source} - Class Identifier. Path to a vector layer.}

\item{CLASSES}{\code{rasterDestination} - Classification. Path for new raster layer.}

\item{MAX_DEPTH}{\code{number} - Maximum Tree Depth. A numeric value.}

\item{MIN_SAMPLES}{\code{number} - Minimum Sample Count. A numeric value.}

\item{MAX_CATEGRS}{\code{number} - Maximum Categories. A numeric value.}

\item{arg_1SE_RULE}{\code{boolean} - Use 1SE Rule. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{TRUNC_PRUNED}{\code{boolean} - Truncate Pruned Trees. 1 for true/yes. 0 for false/no.}

\item{REG_ACCURACY}{\code{number} - Regression Accuracy. A numeric value.}

\item{WEAK_COUNT}{\code{number} - Weak Count. A numeric value.}

\item{WGT_TRIM_RATE}{\code{number} - Weight Trim Rate. A numeric value.}

\item{BOOST_TYPE}{\code{enum}  of \verb{("[0] Discrete AdaBoost", "[1] Real AdaBoost", "[2] LogitBoost", "[3] Gentle AdaBoost")} - Boost Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Boosting classification (opencv) (saga:boostingclassificationopencv)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Classification
}
}
}
