% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_burnstreamnetworkintodem.R
\name{saga_burnstreamnetworkintodem}
\alias{saga_burnstreamnetworkintodem}
\title{QGIS algorithm Burn stream network into dem}
\usage{
saga_burnstreamnetworkintodem(
  DEM = qgisprocess::qgis_default_value(),
  STREAM = qgisprocess::qgis_default_value(),
  FLOWDIR = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  EPSILON = qgisprocess::qgis_default_value(),
  BURN = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{STREAM}{\code{raster} - Streams. Path to a raster layer.}

\item{FLOWDIR}{\code{raster} - Flow direction. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] simply decrease cell's value by epsilon", "[1] lower cell's value to neighbours minimum value minus epsilon", "[2] trace stream network downstream")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EPSILON}{\code{number} - Epsilon. A numeric value.}

\item{BURN}{\code{rasterDestination} - Processed DEM. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Burn stream network into dem (saga:burnstreamnetworkintodem)
}
\details{
\subsection{Outputs description}{
\itemize{
\item BURN - outputRaster - Processed DEM
}
}
}
