% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_catchmentarea.R
\name{saga_catchmentarea}
\alias{saga_catchmentarea}
\title{QGIS algorithm Catchment area}
\usage{
saga_catchmentarea(
  ELEVATION = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  FLOW = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Deterministic 8", "[1] Rho 8", "[2] Braunschweiger Reliefmodell", "[3] Deterministic Infinity", "[4] Multiple Flow Direction", "[5] Multiple Triangular Flow Direction")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FLOW}{\code{rasterDestination} - Catchment Area. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Catchment area (saga:catchmentarea)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLOW - outputRaster - Catchment Area
}
}
}
