% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_catchmentareaflowtracing.R
\name{saga_catchmentareaflowtracing}
\alias{saga_catchmentareaflowtracing}
\title{QGIS algorithm Catchment area (flow tracing)}
\usage{
saga_catchmentareaflowtracing(
  ELEVATION = qgisprocess:::qgis_default_value(),
  FLOW_UNIT = qgisprocess:::qgis_default_value(),
  SINKROUTE = qgisprocess:::qgis_default_value(),
  WEIGHTS = qgisprocess:::qgis_default_value(),
  ACCU_MATERIAL = qgisprocess:::qgis_default_value(),
  VAL_INPUT = qgisprocess:::qgis_default_value(),
  ACCU_TARGET = qgisprocess:::qgis_default_value(),
  STEP = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  MINDQV = qgisprocess:::qgis_default_value(),
  CORRECT = qgisprocess:::qgis_default_value(),
  FLOW = qgisprocess:::qgis_default_value(),
  VAL_MEAN = qgisprocess:::qgis_default_value(),
  ACCU_TOTAL = qgisprocess:::qgis_default_value(),
  ACCU_LEFT = qgisprocess:::qgis_default_value(),
  ACCU_RIGHT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{FLOW_UNIT}{\code{enum}  of \verb{("[0] Number of Cells", "[1] Cell Area")} - Flow Accumulation Units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SINKROUTE}{\code{raster} - Sink Routes. Path to a raster layer.}

\item{WEIGHTS}{\code{raster} - Weights. Path to a raster layer.}

\item{ACCU_MATERIAL}{\code{raster} - Material for Accumulation. Path to a raster layer.}

\item{VAL_INPUT}{\code{raster} - Input for Mean over Catchment. Path to a raster layer.}

\item{ACCU_TARGET}{\code{raster} - Accumulation Target. Path to a raster layer.}

\item{STEP}{\code{number} - Step. A numeric value.}

\item{METHOD}{\code{enum}  of \verb{("[0] Rho 8", "[1] Kinematic Routing Algorithm", "[2] DEMON")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MINDQV}{\code{number} - DEMON - Min. DQV. A numeric value.}

\item{CORRECT}{\code{boolean} - Flow Correction. 1 for true/yes. 0 for false/no.}

\item{FLOW}{\code{rasterDestination} - Flow Accumulation. Path for new raster layer.}

\item{VAL_MEAN}{\code{rasterDestination} - Mean over Catchment. Path for new raster layer.}

\item{ACCU_TOTAL}{\code{rasterDestination} - Accumulated Material. Path for new raster layer.}

\item{ACCU_LEFT}{\code{rasterDestination} - Accumulated Material from left side. Path for new raster layer.}

\item{ACCU_RIGHT}{\code{rasterDestination} - Accumulated Material from right side. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Catchment area (flow tracing) (saga:catchmentareaflowtracing)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ACCU_LEFT - outputRaster - Accumulated Material from left side
\item ACCU_RIGHT - outputRaster - Accumulated Material from right side
\item ACCU_TOTAL - outputRaster - Accumulated Material
\item FLOW - outputRaster - Flow Accumulation
\item VAL_MEAN - outputRaster - Mean over Catchment
}
}
}
