% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_catchmentarearecursive.R
\name{saga_catchmentarearecursive}
\alias{saga_catchmentarearecursive}
\title{QGIS algorithm Catchment area (recursive)}
\usage{
saga_catchmentarearecursive(
  ELEVATION = qgisprocess::qgis_default_value(),
  FLOW_UNIT = qgisprocess::qgis_default_value(),
  SINKROUTE = qgisprocess::qgis_default_value(),
  WEIGHTS = qgisprocess::qgis_default_value(),
  ACCU_MATERIAL = qgisprocess::qgis_default_value(),
  VAL_INPUT = qgisprocess::qgis_default_value(),
  ACCU_TARGET = qgisprocess::qgis_default_value(),
  STEP = qgisprocess::qgis_default_value(),
  TARGETS = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  CONVERGENCE = qgisprocess::qgis_default_value(),
  NO_NEGATIVES = qgisprocess::qgis_default_value(),
  FLOW = qgisprocess::qgis_default_value(),
  VAL_MEAN = qgisprocess::qgis_default_value(),
  ACCU_TOTAL = qgisprocess::qgis_default_value(),
  ACCU_LEFT = qgisprocess::qgis_default_value(),
  ACCU_RIGHT = qgisprocess::qgis_default_value(),
  FLOW_LENGTH = qgisprocess::qgis_default_value(),
  WEIGHT_LOSS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{FLOW_UNIT}{\code{enum}  of \verb{("[0] Number of Cells", "[1] Cell Area")} - Flow Accumulation Units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SINKROUTE}{\code{raster} - Sink Routes. Path to a raster layer.}

\item{WEIGHTS}{\code{raster} - Weights. Path to a raster layer.}

\item{ACCU_MATERIAL}{\code{raster} - Material for Accumulation. Path to a raster layer.}

\item{VAL_INPUT}{\code{raster} - Input for Mean over Catchment. Path to a raster layer.}

\item{ACCU_TARGET}{\code{raster} - Accumulation Target. Path to a raster layer.}

\item{STEP}{\code{number} - Step. A numeric value.}

\item{TARGETS}{\code{raster} - Target Areas. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Deterministic 8", "[1] Rho 8", "[2] Deterministic Infinity", "[3] Multiple Flow Direction")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONVERGENCE}{\code{number} - Convergence. A numeric value.}

\item{NO_NEGATIVES}{\code{boolean} - Prevent Negative Flow Accumulation. 1 for true/yes. 0 for false/no.}

\item{FLOW}{\code{rasterDestination} - Catchment Area. Path for new raster layer.}

\item{VAL_MEAN}{\code{rasterDestination} - Mean over Catchment. Path for new raster layer.}

\item{ACCU_TOTAL}{\code{rasterDestination} - Accumulated Material. Path for new raster layer.}

\item{ACCU_LEFT}{\code{rasterDestination} - Accumulated Material from left side. Path for new raster layer.}

\item{ACCU_RIGHT}{\code{rasterDestination} - Accumulated Material from right side. Path for new raster layer.}

\item{FLOW_LENGTH}{\code{rasterDestination} - Flow Path Length. Path for new raster layer.}

\item{WEIGHT_LOSS}{\code{rasterDestination} - Loss through Negative Weights. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Catchment area (recursive) (saga:catchmentarearecursive)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ACCU_LEFT - outputRaster - Accumulated Material from left side
\item ACCU_RIGHT - outputRaster - Accumulated Material from right side
\item ACCU_TOTAL - outputRaster - Accumulated Material
\item FLOW - outputRaster - Catchment Area
\item FLOW_LENGTH - outputRaster - Flow Path Length
\item VAL_MEAN - outputRaster - Mean over Catchment
\item WEIGHT_LOSS - outputRaster - Loss through Negative Weights
}
}
}
