% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_cellbalance.R
\name{saga_cellbalance}
\alias{saga_cellbalance}
\title{QGIS algorithm Cell balance}
\usage{
saga_cellbalance(
  DEM = qgisprocess::qgis_default_value(),
  WEIGHTS = qgisprocess::qgis_default_value(),
  WEIGHTS_DEFAULT = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  BALANCE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{WEIGHTS}{\code{raster} - Parameter. Path to a raster layer.}

\item{WEIGHTS_DEFAULT}{\code{number} - Default Weight. A numeric value.}

\item{METHOD}{\code{enum}  of \verb{("[0] Deterministic 8", "[1] Multiple Flow Direction")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{BALANCE}{\code{rasterDestination} - Cell Balance. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Cell balance (saga:cellbalance)
}
\details{
\subsection{Outputs description}{
\itemize{
\item BALANCE - outputRaster - Cell Balance
}
}
}
