% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_changedatastorage.R
\name{saga_changedatastorage}
\alias{saga_changedatastorage}
\title{QGIS algorithm Change data storage}
\usage{
saga_changedatastorage(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  OFFSET = qgisprocess::qgis_default_value(),
  SCALE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Converted Grid. Path for new raster layer.}

\item{TYPE}{\code{enum}  of \verb{("[0] bit", "[1] unsigned 1 byte integer", "[2] signed 1 byte integer", "[3] unsigned 2 byte integer", "[4] signed 2 byte integer", "[5] unsigned 4 byte integer", "[6] signed 4 byte integer", "[7] 4 byte floating point number", "[8] 8 byte floating point number")} - Data storage type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OFFSET}{\code{number} - Offset. A numeric value.}

\item{SCALE}{\code{number} - Scale. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Change data storage (saga:changedatastorage)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Converted Grid
}
}
}
