% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_changedetection.R
\name{saga_changedetection}
\alias{saga_changedetection}
\title{QGIS algorithm Change detection}
\usage{
saga_changedetection(
  INITIAL = qgisprocess::qgis_default_value(),
  INI_LUT = qgisprocess::qgis_default_value(),
  INI_LUT_MIN = qgisprocess::qgis_default_value(),
  INI_LUT_MAX = qgisprocess::qgis_default_value(),
  INI_LUT_NAM = qgisprocess::qgis_default_value(),
  FINAL = qgisprocess::qgis_default_value(),
  FIN_LUT = qgisprocess::qgis_default_value(),
  FIN_LUT_MIN = qgisprocess::qgis_default_value(),
  FIN_LUT_MAX = qgisprocess::qgis_default_value(),
  FIN_LUT_NAM = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  CHANGE = qgisprocess::qgis_default_value(),
  CHANGES = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INITIAL}{\code{raster} - Initial State. Path to a raster layer.}

\item{INI_LUT}{\code{source} - Look-up Table. Path to a vector layer.}

\item{INI_LUT_MIN}{\code{field} - Value. The name of an existing field. ; delimited list of existing field names.}

\item{INI_LUT_MAX}{\code{field} - Value (Maximum). The name of an existing field. ; delimited list of existing field names.}

\item{INI_LUT_NAM}{\code{field} - Name. The name of an existing field. ; delimited list of existing field names.}

\item{FINAL}{\code{raster} - Final State. Path to a raster layer.}

\item{FIN_LUT}{\code{source} - Look-up Table. Path to a vector layer.}

\item{FIN_LUT_MIN}{\code{field} - Value. The name of an existing field. ; delimited list of existing field names.}

\item{FIN_LUT_MAX}{\code{field} - Value (Maximum). The name of an existing field. ; delimited list of existing field names.}

\item{FIN_LUT_NAM}{\code{field} - Name. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{enum}  of \verb{("[0] cells", "[1] percent", "[2] area")} - Output as.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CHANGE}{\code{rasterDestination} - Changes. Path for new raster layer.}

\item{CHANGES}{\code{vectorDestination} - Changes. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Change detection (saga:changedetection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CHANGE - outputRaster - Changes
\item CHANGES - outputVector - Changes
}
}
}
