% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_channelnetwork.R
\name{saga_channelnetwork}
\alias{saga_channelnetwork}
\title{QGIS algorithm Channel network}
\usage{
saga_channelnetwork(
  ELEVATION = qgisprocess:::qgis_default_value(),
  SINKROUTE = qgisprocess:::qgis_default_value(),
  INIT_GRID = qgisprocess:::qgis_default_value(),
  INIT_METHOD = qgisprocess:::qgis_default_value(),
  INIT_VALUE = qgisprocess:::qgis_default_value(),
  DIV_GRID = qgisprocess:::qgis_default_value(),
  DIV_CELLS = qgisprocess:::qgis_default_value(),
  TRACE_WEIGHT = qgisprocess:::qgis_default_value(),
  MINLEN = qgisprocess:::qgis_default_value(),
  CHNLNTWRK = qgisprocess:::qgis_default_value(),
  CHNLROUTE = qgisprocess:::qgis_default_value(),
  SHAPES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{SINKROUTE}{\code{raster} - Flow Direction. Path to a raster layer.}

\item{INIT_GRID}{\code{raster} - Initiation Grid. Path to a raster layer.}

\item{INIT_METHOD}{\code{enum}  of \verb{("[0] Less than", "[1] Equals", "[2] Greater than")} - Initiation Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{INIT_VALUE}{\code{number} - Initiation Threshold. A numeric value.}

\item{DIV_GRID}{\code{raster} - Divergence. Path to a raster layer.}

\item{DIV_CELLS}{\code{number} - Tracing: Max. Divergence. A numeric value.}

\item{TRACE_WEIGHT}{\code{raster} - Tracing: Weight. Path to a raster layer.}

\item{MINLEN}{\code{number} - Min. Segment Length. A numeric value.}

\item{CHNLNTWRK}{\code{rasterDestination} - Channel Network. Path for new raster layer.}

\item{CHNLROUTE}{\code{rasterDestination} - Channel Direction. Path for new raster layer.}

\item{SHAPES}{\code{vectorDestination} - Channel Network. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Channel network (saga:channelnetwork)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CHNLNTWRK - outputRaster - Channel Network
\item CHNLROUTE - outputRaster - Channel Direction
\item SHAPES - outputVector - Channel Network
}
}
}
