% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_clippointswithpolygons.R
\name{saga_clippointswithpolygons}
\alias{saga_clippointswithpolygons}
\title{QGIS algorithm Clip points with polygons}
\usage{
saga_clippointswithpolygons(
  POINTS = qgisprocess:::qgis_default_value(),
  POLYGONS = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  CLIPS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{FIELD}{\code{field} - Add Attribute to Clipped Points. The name of an existing field. ; delimited list of existing field names.}

\item{METHOD}{\code{enum}  of \verb{("[0] one layer for all points", "[1] separate layer for each polygon")} - Clipping Options. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CLIPS}{\code{vectorDestination} - Clipped Points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Clip points with polygons (saga:clippointswithpolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLIPS - outputVector - Clipped Points
}
}
}
