% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_cliprasterwithpolygon.R
\name{saga_cliprasterwithpolygon}
\alias{saga_cliprasterwithpolygon}
\title{QGIS algorithm Clip raster with polygon}
\usage{
saga_cliprasterwithpolygon(
  INPUT = qgisprocess::qgis_default_value(),
  POLYGONS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{OUTPUT}{\code{rasterDestination} - Clipped. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Clip raster with polygon (saga:cliprasterwithpolygon)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Clipped
}
}
}
