% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_closegaps.R
\name{saga_closegaps}
\alias{saga_closegaps}
\title{QGIS algorithm Close gaps}
\usage{
saga_closegaps(
  INPUT = qgisprocess::qgis_default_value(),
  MASK = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{MASK}{\code{raster} - Mask. Path to a raster layer.}

\item{THRESHOLD}{\code{number} - Tension Threshold. A numeric value.}

\item{RESULT}{\code{rasterDestination} - Changed Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Close gaps (saga:closegaps)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Changed Grid
}
}
}
