% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_closegapswithspline.R
\name{saga_closegapswithspline}
\alias{saga_closegapswithspline}
\title{QGIS algorithm Close gaps with spline}
\usage{
saga_closegapswithspline(
  GRID = qgisprocess::qgis_default_value(),
  MASK = qgisprocess::qgis_default_value(),
  MAXGAPCELLS = qgisprocess::qgis_default_value(),
  MAXPOINTS = qgisprocess::qgis_default_value(),
  LOCALPOINTS = qgisprocess::qgis_default_value(),
  NEIGHBOURS = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  RELAXATION = qgisprocess::qgis_default_value(),
  CLOSED = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{MASK}{\code{raster} - Mask. Path to a raster layer.}

\item{MAXGAPCELLS}{\code{number} - Only Process Gaps with Less Cells. A numeric value.}

\item{MAXPOINTS}{\code{number} - Maximum Points. A numeric value.}

\item{LOCALPOINTS}{\code{number} - Number of Points for Local Interpolation. A numeric value.}

\item{NEIGHBOURS}{\code{enum}  of \verb{("[0] Neumann", "[1] Moore")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radius (Cells). A numeric value.}

\item{RELAXATION}{\code{number} - Relaxation. A numeric value.}

\item{CLOSED}{\code{rasterDestination} - Closed Gaps Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Close gaps with spline (saga:closegapswithspline)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLOSED - outputRaster - Closed Gaps Grid
}
}
}
