% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_closeonecellgaps.R
\name{saga_closeonecellgaps}
\alias{saga_closeonecellgaps}
\title{QGIS algorithm Close one cell gaps}
\usage{
saga_closeonecellgaps(
  INPUT = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - Changed Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Close one cell gaps (saga:closeonecellgaps)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Changed Grid
}
}
}
