% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_clusteranalysis.R
\name{saga_clusteranalysis}
\alias{saga_clusteranalysis}
\title{QGIS algorithm Cluster analysis}
\usage{
saga_clusteranalysis(
  INPUT = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  STATISTICS = qgisprocess:::qgis_default_value(),
  FIELDS = qgisprocess:::qgis_default_value(),
  CLUSTER = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  NCLUSTER = qgisprocess:::qgis_default_value(),
  NORMALISE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Table. Path to a vector layer.}

\item{RESULT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{STATISTICS}{\code{vectorDestination} - Statistics. Path for new vector layer.}

\item{FIELDS}{\code{field} - Attributes. The name of an existing field. ; delimited list of existing field names.}

\item{CLUSTER}{\code{field} - Cluster. The name of an existing field. ; delimited list of existing field names.}

\item{METHOD}{\code{enum}  of \verb{("[0] Iterative Minimum Distance (Forgy 1965)", "[1] Hill-Climbing (Rubin 1967)", "[2] Combined Minimum Distance / Hillclimbing")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NCLUSTER}{\code{number} - Clusters. A numeric value.}

\item{NORMALISE}{\code{boolean} - Normalise. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Cluster analysis (saga:clusteranalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Result
\item STATISTICS - outputVector - Statistics
}
}
}
