% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_confusionmatrixpolygonsgrid.R
\name{saga_confusionmatrixpolygonsgrid}
\alias{saga_confusionmatrixpolygonsgrid}
\title{QGIS algorithm Confusion matrix (polygons / grid)}
\usage{
saga_confusionmatrixpolygonsgrid(
  GRID = qgisprocess::qgis_default_value(),
  GRID_LUT = qgisprocess::qgis_default_value(),
  GRID_LUT_MIN = qgisprocess::qgis_default_value(),
  GRID_LUT_MAX = qgisprocess::qgis_default_value(),
  GRID_LUT_NAM = qgisprocess::qgis_default_value(),
  POLYGONS = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  CONFUSION = qgisprocess::qgis_default_value(),
  CLASSES = qgisprocess::qgis_default_value(),
  SUMMARY = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Classification. Path to a raster layer.}

\item{GRID_LUT}{\code{source} - Look-up Table. Path to a vector layer.}

\item{GRID_LUT_MIN}{\code{field} - Value. The name of an existing field. ; delimited list of existing field names.}

\item{GRID_LUT_MAX}{\code{field} - Value (Maximum). The name of an existing field. ; delimited list of existing field names.}

\item{GRID_LUT_NAM}{\code{field} - Name. The name of an existing field. ; delimited list of existing field names.}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{FIELD}{\code{field} - Class Values. The name of an existing field. ; delimited list of existing field names.}

\item{CONFUSION}{\code{vectorDestination} - Confusion Matrix. Path for new vector layer.}

\item{CLASSES}{\code{vectorDestination} - Class Values. Path for new vector layer.}

\item{SUMMARY}{\code{vectorDestination} - Summary. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Confusion matrix (polygons / grid) (saga:confusionmatrixpolygonsgrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputVector - Class Values
\item CONFUSION - outputVector - Confusion Matrix
\item SUMMARY - outputVector - Summary
}
}
}
