% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_confusionmatrixtwogrids.R
\name{saga_confusionmatrixtwogrids}
\alias{saga_confusionmatrixtwogrids}
\title{QGIS algorithm Confusion matrix (two grids)}
\usage{
saga_confusionmatrixtwogrids(
  ONE = qgisprocess::qgis_default_value(),
  ONE_LUT = qgisprocess::qgis_default_value(),
  ONE_LUT_MIN = qgisprocess::qgis_default_value(),
  ONE_LUT_MAX = qgisprocess::qgis_default_value(),
  ONE_LUT_NAM = qgisprocess::qgis_default_value(),
  TWO = qgisprocess::qgis_default_value(),
  TWO_LUT = qgisprocess::qgis_default_value(),
  TWO_LUT_MIN = qgisprocess::qgis_default_value(),
  TWO_LUT_MAX = qgisprocess::qgis_default_value(),
  TWO_LUT_NAM = qgisprocess::qgis_default_value(),
  COMBINED = qgisprocess::qgis_default_value(),
  NOCHANGE = qgisprocess::qgis_default_value(),
  CONFUSION = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  CLASSES = qgisprocess::qgis_default_value(),
  SUMMARY = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ONE}{\code{raster} - Classification 1. Path to a raster layer.}

\item{ONE_LUT}{\code{source} - Look-up Table. Path to a vector layer.}

\item{ONE_LUT_MIN}{\code{source} - Value. Path to a vector layer.}

\item{ONE_LUT_MAX}{\code{source} - Value (Maximum). Path to a vector layer.}

\item{ONE_LUT_NAM}{\code{source} - Name. Path to a vector layer.}

\item{TWO}{\code{raster} - Classification 2. Path to a raster layer.}

\item{TWO_LUT}{\code{source} - Look-up Table. Path to a vector layer.}

\item{TWO_LUT_MIN}{\code{source} - Value. Path to a vector layer.}

\item{TWO_LUT_MAX}{\code{source} - Value (Maximum). Path to a vector layer.}

\item{TWO_LUT_NAM}{\code{source} - Name. Path to a vector layer.}

\item{COMBINED}{\code{rasterDestination} - Combined Classes. Path for new raster layer.}

\item{NOCHANGE}{\code{boolean} - Report Unchanged Classes. 1 for true/yes. 0 for false/no.}

\item{CONFUSION}{\code{vectorDestination} - Confusion Matrix. Path for new vector layer.}

\item{OUTPUT}{\code{enum}  of \verb{("[0] cells", "[1] percent", "[2] area")} - Output as.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CLASSES}{\code{vectorDestination} - Class Values. Path for new vector layer.}

\item{SUMMARY}{\code{vectorDestination} - Summary. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Confusion matrix (two grids) (saga:confusionmatrixtwogrids)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputVector - Class Values
\item COMBINED - outputRaster - Combined Classes
\item CONFUSION - outputVector - Confusion Matrix
\item SUMMARY - outputVector - Summary
}
}
}
