% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_connectivityanalysis.R
\name{saga_connectivityanalysis}
\alias{saga_connectivityanalysis}
\title{QGIS algorithm Connectivity analysis}
\usage{
saga_connectivityanalysis(
  INPUT_GRID = qgisprocess::qgis_default_value(),
  FILTERED_MASK = qgisprocess::qgis_default_value(),
  FILTER = qgisprocess::qgis_default_value(),
  SIZE = qgisprocess::qgis_default_value(),
  SYMBOLIC_IMAGE = qgisprocess::qgis_default_value(),
  OUTLINES = qgisprocess::qgis_default_value(),
  BORDER_PIXEL_CENTERS = qgisprocess::qgis_default_value(),
  REMOVE_MARGINAL_REGIONS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT_GRID}{\code{raster} - Input Binary Grid. Path to a raster layer.}

\item{FILTERED_MASK}{\code{rasterDestination} - Filtered Image. Path for new raster layer.}

\item{FILTER}{\code{boolean} - Apply Filter?. 1 for true/yes. 0 for false/no.}

\item{SIZE}{\code{number} - Filter Size (Radius). A numeric value.}

\item{SYMBOLIC_IMAGE}{\code{rasterDestination} - Symbolic Image. Path for new raster layer.}

\item{OUTLINES}{\code{vectorDestination} - Outlines. Path for new vector layer.}

\item{BORDER_PIXEL_CENTERS}{\code{boolean} - Pixel Centers?. 1 for true/yes. 0 for false/no.}

\item{REMOVE_MARGINAL_REGIONS}{\code{boolean} - Remove Border Regions?. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Connectivity analysis (saga:connectivityanalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FILTERED_MASK - outputRaster - Filtered Image
\item OUTLINES - outputVector - Outlines
\item SYMBOLIC_IMAGE - outputRaster - Symbolic Image
}
}
}
