% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_constantgrid.R
\name{saga_constantgrid}
\alias{saga_constantgrid}
\title{QGIS algorithm Constant grid}
\usage{
saga_constantgrid(
  NAME = qgisprocess:::qgis_default_value(),
  CONST = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  DEFINITION = qgisprocess:::qgis_default_value(),
  USER_SIZE = qgisprocess:::qgis_default_value(),
  USER_XMIN = qgisprocess:::qgis_default_value(),
  USER_XMAX = qgisprocess:::qgis_default_value(),
  USER_YMIN = qgisprocess:::qgis_default_value(),
  USER_YMAX = qgisprocess:::qgis_default_value(),
  USER_FITS = qgisprocess:::qgis_default_value(),
  TEMPLATE = qgisprocess:::qgis_default_value(),
  OUT_GRID = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{NAME}{\code{string} - Name. String value.}

\item{CONST}{\code{number} - Constant Value. A numeric value.}

\item{TYPE}{\code{enum}  of \verb{("[0] bit", "[1] unsigned 1 byte integer", "[2] signed 1 byte integer", "[3] unsigned 2 byte integer", "[4] signed 2 byte integer", "[5] unsigned 8 byte integer", "[6] signed 8 byte integer", "[7] 4 byte floating point number", "[8] 8 byte floating point number")} - Data Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DEFINITION}{\code{enum}  of \verb{("[0] user defined", "[1] grid or grid system")} - Target Grid System. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{USER_XMIN}{\code{number} - Left. A numeric value.}

\item{USER_XMAX}{\code{number} - Right. A numeric value.}

\item{USER_YMIN}{\code{number} - Bottom. A numeric value.}

\item{USER_YMAX}{\code{number} - Top. A numeric value.}

\item{USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TEMPLATE}{\code{raster} - Target System. Path to a raster layer.}

\item{OUT_GRID}{\code{rasterDestination} - Target Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Constant grid (saga:constantgrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUT_GRID - outputRaster - Target Grid
}
}
}
