% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_contourlines.R
\name{saga_contourlines}
\alias{saga_contourlines}
\title{QGIS algorithm Contour lines}
\usage{
saga_contourlines(
  GRID = qgisprocess::qgis_default_value(),
  VERTEX = qgisprocess::qgis_default_value(),
  ZMIN = qgisprocess::qgis_default_value(),
  ZMAX = qgisprocess::qgis_default_value(),
  ZSTEP = qgisprocess::qgis_default_value(),
  CONTOUR = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{VERTEX}{\code{enum}  of \verb{("[0] x,y", "[1] x,y,z")} - Vertex type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ZMIN}{\code{number} - Minimum Contour Value. A numeric value.}

\item{ZMAX}{\code{number} - Maximum Contour Value. A numeric value.}

\item{ZSTEP}{\code{number} - Equidistance. A numeric value.}

\item{CONTOUR}{\code{vectorDestination} - Contour Lines. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Contour lines (saga:contourlines)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONTOUR - outputVector - Contour Lines
}
}
}
