% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_convergenceindex.R
\name{saga_convergenceindex}
\alias{saga_convergenceindex}
\title{QGIS algorithm Convergence index}
\usage{
saga_convergenceindex(
  ELEVATION = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  NEIGHBOURS = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Aspect", "[1] Gradient")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NEIGHBOURS}{\code{enum}  of \verb{("[0] 2 x 2", "[1] 3 x 3")} - Gradient Calculation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESULT}{\code{rasterDestination} - Convergence Index. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Convergence index (saga:convergenceindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Convergence Index
}
}
}
