% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_convergenceindexsearchradius.R
\name{saga_convergenceindexsearchradius}
\alias{saga_convergenceindexsearchradius}
\title{QGIS algorithm Convergence index (search radius)}
\usage{
saga_convergenceindexsearchradius(
  ELEVATION = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_WEIGHTING = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_IDW_POWER = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_IDW_OFFSET = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_BANDWIDTH = qgisprocess::qgis_default_value(),
  SLOPE = qgisprocess::qgis_default_value(),
  DIFFERENCE = qgisprocess::qgis_default_value(),
  CONVERGENCE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{RADIUS}{\code{number} - Radius (Cells). A numeric value.}

\item{DISTANCE_WEIGHTING_DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DISTANCE_WEIGHTING_DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DISTANCE_WEIGHTING_DW_IDW_OFFSET}{\code{boolean} - Inverse Distance Offset. 1 for true/yes. 0 for false/no.}

\item{DISTANCE_WEIGHTING_DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{SLOPE}{\code{boolean} - Gradient. 1 for true/yes. 0 for false/no.}

\item{DIFFERENCE}{\code{enum}  of \verb{("[0] direction to the center cell", "[1] center cell's aspect direction")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONVERGENCE}{\code{rasterDestination} - Convergence Index. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Convergence index (search radius) (saga:convergenceindexsearchradius)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONVERGENCE - outputRaster - Convergence Index
}
}
}
