% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_convertlinestopolygons.R
\name{saga_convertlinestopolygons}
\alias{saga_convertlinestopolygons}
\title{QGIS algorithm Convert lines to polygons}
\usage{
saga_convertlinestopolygons(
  LINES = qgisprocess::qgis_default_value(),
  POLYGONS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{POLYGONS}{\code{vectorDestination} - Polygons. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Convert lines to polygons (saga:convertlinestopolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item POLYGONS - outputVector - Polygons
}
}
}
