% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_convertpointstolines.R
\name{saga_convertpointstolines}
\alias{saga_convertpointstolines}
\title{QGIS algorithm Convert points to line(s)}
\usage{
saga_convertpointstolines(
  POINTS = qgisprocess::qgis_default_value(),
  ORDER = qgisprocess::qgis_default_value(),
  SEPARATE = qgisprocess::qgis_default_value(),
  LINES = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{ORDER}{\code{field} - Order by.... The name of an existing field. ; delimited list of existing field names.}

\item{SEPARATE}{\code{field} - Separate by.... The name of an existing field. ; delimited list of existing field names.}

\item{LINES}{\code{vectorDestination} - Lines. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Convert points to line(s) (saga:convertpointstolines)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LINES - outputVector - Lines
}
}
}
