% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_convertpolygonstolines.R
\name{saga_convertpolygonstolines}
\alias{saga_convertpolygonstolines}
\title{QGIS algorithm Convert polygons to lines}
\usage{
saga_convertpolygonstolines(
  POLYGONS = qgisprocess::qgis_default_value(),
  LINES = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{LINES}{\code{vectorDestination} - Lines. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Convert polygons to lines (saga:convertpolygonstolines)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LINES - outputVector - Lines
}
}
}
