% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_covereddistance.R
\name{saga_covereddistance}
\alias{saga_covereddistance}
\title{QGIS algorithm Covered distance}
\usage{
saga_covereddistance(
  INPUT = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Grids. .}

\item{RESULT}{\code{rasterDestination} - Covered Distance. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Covered distance (saga:covereddistance)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Covered Distance
}
}
}
