% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_croptodata.R
\name{saga_croptodata}
\alias{saga_croptodata}
\title{QGIS algorithm Crop to data}
\usage{
saga_croptodata(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Cropped. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Crop to data (saga:croptodata)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Cropped
}
}
}
