% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_crossclassificationandtabulation.R
\name{saga_crossclassificationandtabulation}
\alias{saga_crossclassificationandtabulation}
\title{QGIS algorithm Cross-classification and tabulation}
\usage{
saga_crossclassificationandtabulation(
  INPUT = qgisprocess::qgis_default_value(),
  INPUT2 = qgisprocess::qgis_default_value(),
  MAXNUMCLASS = qgisprocess::qgis_default_value(),
  RESULTGRID = qgisprocess::qgis_default_value(),
  RESULTTABLE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input Grid 1. Path to a raster layer.}

\item{INPUT2}{\code{raster} - Input Grid 2. Path to a raster layer.}

\item{MAXNUMCLASS}{\code{number} - Max. Number of Classes. A numeric value.}

\item{RESULTGRID}{\code{rasterDestination} - Cross-Classification Grid. Path for new raster layer.}

\item{RESULTTABLE}{\code{vectorDestination} - Cross-Tabulation Table. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Cross-classification and tabulation (saga:crossclassificationandtabulation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULTGRID - outputRaster - Cross-Classification Grid
\item RESULTTABLE - outputVector - Cross-Tabulation Table
}
}
}
