% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_curvatureclassification.R
\name{saga_curvatureclassification}
\alias{saga_curvatureclassification}
\title{QGIS algorithm Curvature classification}
\usage{
saga_curvatureclassification(
  DEM = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  CLASS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{THRESHOLD}{\code{number} - Threshold for plane. A numeric value.}

\item{CLASS}{\code{rasterDestination} - Curvature Classification. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Curvature classification (saga:curvatureclassification)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASS - outputRaster - Curvature Classification
}
}
}
