% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_cutvectorlayer.R
\name{saga_cutvectorlayer}
\alias{saga_cutvectorlayer}
\title{QGIS algorithm Cut vector layer}
\usage{
saga_cutvectorlayer(
  SHAPES = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  POLYGONS_POLYGONS = qgisprocess:::qgis_default_value(),
  CUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Vector layer to cut. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] completely contained", "[1] intersects", "[2] center")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{POLYGONS_POLYGONS}{\code{source} - Cutting polygons. Path to a vector layer.}

\item{CUT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Cut vector layer (saga:cutvectorlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CUT - outputVector - Result
}
}
}
