% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_destripingwithmask.R
\name{saga_destripingwithmask}
\alias{saga_destripingwithmask}
\title{QGIS algorithm Destriping with mask}
\usage{
saga_destripingwithmask(
  INPUT = qgisprocess:::qgis_default_value(),
  MASK = qgisprocess:::qgis_default_value(),
  RESULT3 = qgisprocess:::qgis_default_value(),
  RESULT1 = qgisprocess:::qgis_default_value(),
  RESULT2 = qgisprocess:::qgis_default_value(),
  ANG = qgisprocess:::qgis_default_value(),
  R = qgisprocess:::qgis_default_value(),
  D = qgisprocess:::qgis_default_value(),
  MIN = qgisprocess:::qgis_default_value(),
  MAX = qgisprocess:::qgis_default_value(),
  MMIN = qgisprocess:::qgis_default_value(),
  MMAX = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{MASK}{\code{raster} - Mask Grid. Path to a raster layer.}

\item{RESULT3}{\code{rasterDestination} - Destriped Grid. Path for new raster layer.}

\item{RESULT1}{\code{rasterDestination} - Low-pass 1. Path for new raster layer.}

\item{RESULT2}{\code{rasterDestination} - Low-pass 2. Path for new raster layer.}

\item{ANG}{\code{number} - Angle (in degrees). A numeric value.}

\item{R}{\code{number} - Radius. A numeric value.}

\item{D}{\code{number} - Stripes distance. A numeric value.}

\item{MIN}{\code{number} - Min. A numeric value.}

\item{MAX}{\code{number} - Max. A numeric value.}

\item{MMIN}{\code{number} - Mask Min. A numeric value.}

\item{MMAX}{\code{number} - Mask Max. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Destriping with mask (saga:destripingwithmask)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT1 - outputRaster - Low-pass 1
\item RESULT2 - outputRaster - Low-pass 2
\item RESULT3 - outputRaster - Destriped Grid
}
}
}
