% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_diffusivehillslopeevolutionadi.R
\name{saga_diffusivehillslopeevolutionadi}
\alias{saga_diffusivehillslopeevolutionadi}
\title{QGIS algorithm Diffusive hillslope evolution (adi)}
\usage{
saga_diffusivehillslopeevolutionadi(
  DEM = qgisprocess::qgis_default_value(),
  CHANNELS = qgisprocess::qgis_default_value(),
  UPDATE = qgisprocess::qgis_default_value(),
  KAPPA = qgisprocess::qgis_default_value(),
  DURATION = qgisprocess::qgis_default_value(),
  TIMESTEP = qgisprocess::qgis_default_value(),
  DTIME = qgisprocess::qgis_default_value(),
  MODEL = qgisprocess::qgis_default_value(),
  DIFF = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{CHANNELS}{\code{raster} - Channel Mask. Path to a raster layer.}

\item{UPDATE}{\code{boolean} - Update. 1 for true/yes. 0 for false/no.}

\item{KAPPA}{\code{number} - Diffusivity (m2/kyr). A numeric value.}

\item{DURATION}{\code{number} - Simulation Time (kyr). A numeric value.}

\item{TIMESTEP}{\code{enum}  of \verb{("[0] user defined", "[1] automatically")} - Time Step. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DTIME}{\code{number} - Time Step (kyr). A numeric value.}

\item{MODEL}{\code{rasterDestination} - Modelled Elevation. Path for new raster layer.}

\item{DIFF}{\code{rasterDestination} - Elevation Difference. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Diffusive hillslope evolution (adi) (saga:diffusivehillslopeevolutionadi)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIFF - outputRaster - Elevation Difference
\item MODEL - outputRaster - Modelled Elevation
}
}
}
