% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_diffusivehillslopeevolutionftcs.R
\name{saga_diffusivehillslopeevolutionftcs}
\alias{saga_diffusivehillslopeevolutionftcs}
\title{QGIS algorithm Diffusive hillslope evolution (ftcs)}
\usage{
saga_diffusivehillslopeevolutionftcs(
  DEM = qgisprocess::qgis_default_value(),
  UPDATE = qgisprocess::qgis_default_value(),
  KAPPA = qgisprocess::qgis_default_value(),
  DURATION = qgisprocess::qgis_default_value(),
  TIMESTEP = qgisprocess::qgis_default_value(),
  DTIME = qgisprocess::qgis_default_value(),
  NEIGHBOURS = qgisprocess::qgis_default_value(),
  MODEL = qgisprocess::qgis_default_value(),
  DIFF = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{UPDATE}{\code{boolean} - Update. 1 for true/yes. 0 for false/no.}

\item{KAPPA}{\code{number} - Diffusivity (m2/kyr). A numeric value.}

\item{DURATION}{\code{number} - Simulation Time (kyr). A numeric value.}

\item{TIMESTEP}{\code{enum}  of \verb{("[0] user defined", "[1] automatically")} - Time Step. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DTIME}{\code{number} - Time Step (kyr). A numeric value.}

\item{NEIGHBOURS}{\code{enum}  of \verb{("[0] Neumann", "[1] Moore")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MODEL}{\code{rasterDestination} - Modelled Elevation. Path for new raster layer.}

\item{DIFF}{\code{rasterDestination} - Elevation Difference. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Diffusive hillslope evolution (ftcs) (saga:diffusivehillslopeevolutionftcs)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIFF - outputRaster - Elevation Difference
\item MODEL - outputRaster - Modelled Elevation
}
}
}
