% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_directionalaverage.R
\name{saga_directionalaverage}
\alias{saga_directionalaverage}
\title{QGIS algorithm Directional average}
\usage{
saga_directionalaverage(
  INPUT = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ANG = qgisprocess::qgis_default_value(),
  R1 = qgisprocess::qgis_default_value(),
  R2 = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - Output Grid. Path for new raster layer.}

\item{ANG}{\code{number} - Angle (in degrees). A numeric value.}

\item{R1}{\code{number} - Main Radius. A numeric value.}

\item{R2}{\code{number} - Transversal radius. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Directional average (saga:directionalaverage)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Output Grid
}
}
}
