% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_directionalstatisticsforrasterlayer.R
\name{saga_directionalstatisticsforrasterlayer}
\alias{saga_directionalstatisticsforrasterlayer}
\title{QGIS algorithm Directional statistics for raster layer}
\usage{
saga_directionalstatisticsforrasterlayer(
  GRID = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  DIRECTION = qgisprocess::qgis_default_value(),
  TOLERANCE = qgisprocess::qgis_default_value(),
  MAXDISTANCE = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_WEIGHTING = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_IDW_POWER = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_IDW_OFFSET = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_BANDWIDTH = qgisprocess::qgis_default_value(),
  MEAN = qgisprocess::qgis_default_value(),
  DIFMEAN = qgisprocess::qgis_default_value(),
  MIN = qgisprocess::qgis_default_value(),
  MAX = qgisprocess::qgis_default_value(),
  RANGE = qgisprocess::qgis_default_value(),
  VAR = qgisprocess::qgis_default_value(),
  STDDEV = qgisprocess::qgis_default_value(),
  STDDEVLO = qgisprocess::qgis_default_value(),
  STDDEVHI = qgisprocess::qgis_default_value(),
  DEVMEAN = qgisprocess::qgis_default_value(),
  PERCENT = qgisprocess::qgis_default_value(),
  POINTS_OUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{DIRECTION}{\code{number} - Direction (Degree). A numeric value.}

\item{TOLERANCE}{\code{number} - Tolerance (Degree). A numeric value.}

\item{MAXDISTANCE}{\code{number} - Maximum Distance (Cells). A numeric value.}

\item{DISTANCE_WEIGHTING_DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Distance Weighting. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DISTANCE_WEIGHTING_DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DISTANCE_WEIGHTING_DW_IDW_OFFSET}{\code{boolean} - Inverse Distance Offset. 1 for true/yes. 0 for false/no.}

\item{DISTANCE_WEIGHTING_DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{MEAN}{\code{rasterDestination} - Arithmetic Mean. Path for new raster layer.}

\item{DIFMEAN}{\code{rasterDestination} - Difference from Arithmetic Mean. Path for new raster layer.}

\item{MIN}{\code{rasterDestination} - Minimum. Path for new raster layer.}

\item{MAX}{\code{rasterDestination} - Maximum. Path for new raster layer.}

\item{RANGE}{\code{rasterDestination} - Range. Path for new raster layer.}

\item{VAR}{\code{rasterDestination} - Variance. Path for new raster layer.}

\item{STDDEV}{\code{rasterDestination} - Standard Deviation. Path for new raster layer.}

\item{STDDEVLO}{\code{rasterDestination} - Mean less Standard Deviation. Path for new raster layer.}

\item{STDDEVHI}{\code{rasterDestination} - Mean plus Standard Deviation. Path for new raster layer.}

\item{DEVMEAN}{\code{rasterDestination} - Deviation from Arithmetic Mean. Path for new raster layer.}

\item{PERCENT}{\code{rasterDestination} - Percentile. Path for new raster layer.}

\item{POINTS_OUT}{\code{vectorDestination} - Directional Statistics for Points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Directional statistics for raster layer (saga:directionalstatisticsforrasterlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DEVMEAN - outputRaster - Deviation from Arithmetic Mean
\item DIFMEAN - outputRaster - Difference from Arithmetic Mean
\item MAX - outputRaster - Maximum
\item MEAN - outputRaster - Arithmetic Mean
\item MIN - outputRaster - Minimum
\item PERCENT - outputRaster - Percentile
\item POINTS_OUT - outputVector - Directional Statistics for Points
\item RANGE - outputRaster - Range
\item STDDEV - outputRaster - Standard Deviation
\item STDDEVHI - outputRaster - Mean plus Standard Deviation
\item STDDEVLO - outputRaster - Mean less Standard Deviation
\item VAR - outputRaster - Variance
}
}
}
