% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_distancevigra.R
\name{saga_distancevigra}
\alias{saga_distancevigra}
\title{QGIS algorithm Distance (vigra)}
\usage{
saga_distancevigra(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  NORM = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Features. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Distance. Path for new raster layer.}

\item{NORM}{\code{enum}  of \verb{("[0] Chessboard", "[1] Manhattan", "[2] Euclidean")} - Type of distance calculation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Distance (vigra) (saga:distancevigra)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Distance
}
}
}
