% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_diurnalanisotropicheating.R
\name{saga_diurnalanisotropicheating}
\alias{saga_diurnalanisotropicheating}
\title{QGIS algorithm Diurnal anisotropic heating}
\usage{
saga_diurnalanisotropicheating(
  DEM = qgisprocess::qgis_default_value(),
  ALPHA_MAX = qgisprocess::qgis_default_value(),
  DAH = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{ALPHA_MAX}{\code{number} - Alpha Max (Degree). A numeric value.}

\item{DAH}{\code{rasterDestination} - Diurnal Anisotropic Heating. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Diurnal anisotropic heating (saga:diurnalanisotropicheating)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DAH - outputRaster - Diurnal Anisotropic Heating
}
}
}
