% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_diversityofcategories.R
\name{saga_diversityofcategories}
\alias{saga_diversityofcategories}
\title{QGIS algorithm Diversity of categories}
\usage{
saga_diversityofcategories(
  CATEGORIES = qgisprocess::qgis_default_value(),
  SEARCH_MODE = qgisprocess::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess::qgis_default_value(),
  NB_CASE = qgisprocess::qgis_default_value(),
  DW_WEIGHTING = qgisprocess::qgis_default_value(),
  DW_IDW_POWER = qgisprocess::qgis_default_value(),
  DW_IDW_OFFSET = qgisprocess::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess::qgis_default_value(),
  COUNT = qgisprocess::qgis_default_value(),
  DIVERSITY = qgisprocess::qgis_default_value(),
  SIZE_MEAN = qgisprocess::qgis_default_value(),
  SIZE_SKEW = qgisprocess::qgis_default_value(),
  CONNECTIVITY = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{CATEGORIES}{\code{raster} - Categories. Path to a raster layer.}

\item{SEARCH_MODE}{\code{enum}  of \verb{("[0] square", "[1] circle")} - Search Mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_RADIUS}{\code{number} - Search Radius Distance. A numeric value.}

\item{NB_CASE}{\code{enum}  of \verb{("[0] Rook's case", "[1] Queen's case")} - Connectivity Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DW_IDW_OFFSET}{\code{boolean} - Inverse Distance Offset. 1 for true/yes. 0 for false/no.}

\item{DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{COUNT}{\code{rasterDestination} - Number of Categories. Path for new raster layer.}

\item{DIVERSITY}{\code{rasterDestination} - Diversity. Path for new raster layer.}

\item{SIZE_MEAN}{\code{rasterDestination} - Average Size. Path for new raster layer.}

\item{SIZE_SKEW}{\code{rasterDestination} - Skewness. Path for new raster layer.}

\item{CONNECTIVITY}{\code{rasterDestination} - Connectivity. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Diversity of categories (saga:diversityofcategories)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONNECTIVITY - outputRaster - Connectivity
\item COUNT - outputRaster - Number of Categories
\item DIVERSITY - outputRaster - Diversity
\item SIZE_MEAN - outputRaster - Average Size
\item SIZE_SKEW - outputRaster - Skewness
}
}
}
