% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_downslopedistancegradient.R
\name{saga_downslopedistancegradient}
\alias{saga_downslopedistancegradient}
\title{QGIS algorithm Downslope distance gradient}
\usage{
saga_downslopedistancegradient(
  DEM = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  GRADIENT = qgisprocess::qgis_default_value(),
  DIFFERENCE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{DISTANCE}{\code{number} - Vertical Distance. A numeric value.}

\item{OUTPUT}{\code{enum}  of \verb{("[0] distance", "[1] gradient (tangens)", "[2] gradient (degree)")} - Output. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRADIENT}{\code{rasterDestination} - Gradient. Path for new raster layer.}

\item{DIFFERENCE}{\code{rasterDestination} - Gradient Difference. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Downslope distance gradient (saga:downslopedistancegradient)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIFFERENCE - outputRaster - Gradient Difference
\item GRADIENT - outputRaster - Gradient
}
}
}
