% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_edgecontamination.R
\name{saga_edgecontamination}
\alias{saga_edgecontamination}
\title{QGIS algorithm Edge contamination}
\usage{
saga_edgecontamination(
  DEM = qgisprocess::qgis_default_value(),
  CONTAMINATION = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{CONTAMINATION}{\code{rasterDestination} - Edge Contamination. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Edge contamination (saga:edgecontamination)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONTAMINATION - outputRaster - Edge Contamination
}
}
}
