% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_edgedetectionvigra.R
\name{saga_edgedetectionvigra}
\alias{saga_edgedetectionvigra}
\title{QGIS algorithm Edge detection (vigra)}
\usage{
saga_edgedetectionvigra(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  SCALE = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Edges. Path for new raster layer.}

\item{TYPE}{\code{enum}  of \verb{("[0] Canny", "[1] Shen-Castan")} - Detector type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SCALE}{\code{number} - Operator scale. A numeric value.}

\item{THRESHOLD}{\code{number} - Gradient threshold. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Edge detection (vigra) (saga:edgedetectionvigra)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Edges
}
}
}
