% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_enhancedvegetationindex.R
\name{saga_enhancedvegetationindex}
\alias{saga_enhancedvegetationindex}
\title{QGIS algorithm Enhanced vegetation index}
\usage{
saga_enhancedvegetationindex(
  BLUE = qgisprocess::qgis_default_value(),
  RED = qgisprocess::qgis_default_value(),
  NIR = qgisprocess::qgis_default_value(),
  EVI = qgisprocess::qgis_default_value(),
  GAIN = qgisprocess::qgis_default_value(),
  L = qgisprocess::qgis_default_value(),
  CBLUE = qgisprocess::qgis_default_value(),
  CRED = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{BLUE}{\code{raster} - Blue Reflectance. Path to a raster layer.}

\item{RED}{\code{raster} - Red Reflectance. Path to a raster layer.}

\item{NIR}{\code{raster} - Near Infrared Reflectance. Path to a raster layer.}

\item{EVI}{\code{rasterDestination} - Enhanced Vegetation Index. Path for new raster layer.}

\item{GAIN}{\code{number} - Gain. A numeric value.}

\item{L}{\code{number} - Canopy Background Adjustment. A numeric value.}

\item{CBLUE}{\code{number} - Aerosol Resistance Coefficient (Blue). A numeric value.}

\item{CRED}{\code{number} - Aerosol Resistance Coefficient (Red). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Enhanced vegetation index (saga:enhancedvegetationindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item EVI - outputRaster - Enhanced Vegetation Index
}
}
}
