% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_featureextents.R
\name{saga_featureextents}
\alias{saga_featureextents}
\title{QGIS algorithm Feature extents}
\usage{
saga_featureextents(
  SHAPES = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  EXTENTS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{OUTPUT}{\code{enum}  of \verb{("[0] all shapes", "[1] each shape", "[2] each shape's part")} - Get Extent for. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXTENTS}{\code{vectorDestination} - Extents. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Feature extents (saga:featureextents)
}
\details{
\subsection{Outputs description}{
\itemize{
\item EXTENTS - outputVector - Extents
}
}
}
