% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fillgapsinrecords.R
\name{saga_fillgapsinrecords}
\alias{saga_fillgapsinrecords}
\title{QGIS algorithm Fill gaps in records}
\usage{
saga_fillgapsinrecords(
  TABLE = qgisprocess::qgis_default_value(),
  ORDER = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  NOGAPS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{TABLE}{\code{source} - Table. Path to a vector layer.}

\item{ORDER}{\code{field} - Order. The name of an existing field. ; delimited list of existing field names.}

\item{METHOD}{\code{enum}  of \verb{("[0] Nearest Neighbour", "[1] Linear", "[2] Spline")} - Interpolation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NOGAPS}{\code{vectorDestination} - Table without Gaps. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Fill gaps in records (saga:fillgapsinrecords)
}
\details{
\subsection{Outputs description}{
\itemize{
\item NOGAPS - outputVector - Table without Gaps
}
}
}
