% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fillsinksqmofesp.R
\name{saga_fillsinksqmofesp}
\alias{saga_fillsinksqmofesp}
\title{QGIS algorithm Fill sinks (qm of esp)}
\usage{
saga_fillsinksqmofesp(
  DEM = qgisprocess::qgis_default_value(),
  DZFILL = qgisprocess::qgis_default_value(),
  FILLED = qgisprocess::qgis_default_value(),
  SINKS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{DZFILL}{\code{number} - Fill Increment. A numeric value.}

\item{FILLED}{\code{rasterDestination} - DEM Without Sinks. Path for new raster layer.}

\item{SINKS}{\code{rasterDestination} - Sinks. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Fill sinks (qm of esp) (saga:fillsinksqmofesp)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FILLED - outputRaster - DEM Without Sinks
\item SINKS - outputRaster - Sinks
}
}
}
