% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fillsinkswangliu.R
\name{saga_fillsinkswangliu}
\alias{saga_fillsinkswangliu}
\title{QGIS algorithm Fill sinks (wang & liu)}
\usage{
saga_fillsinkswangliu(
  ELEV = qgisprocess::qgis_default_value(),
  MINSLOPE = qgisprocess::qgis_default_value(),
  FILLED = qgisprocess::qgis_default_value(),
  FDIR = qgisprocess::qgis_default_value(),
  WSHED = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEV}{\code{raster} - DEM. Path to a raster layer.}

\item{MINSLOPE}{\code{number} - Minimum Slope (Degree). A numeric value.}

\item{FILLED}{\code{rasterDestination} - Filled DEM. Path for new raster layer.}

\item{FDIR}{\code{rasterDestination} - Flow Directions. Path for new raster layer.}

\item{WSHED}{\code{rasterDestination} - Watershed Basins. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Fill sinks (wang & liu) (saga:fillsinkswangliu)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FDIR - outputRaster - Flow Directions
\item FILLED - outputRaster - Filled DEM
\item WSHED - outputRaster - Watershed Basins
}
}
}
