% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_findfieldofextremevalue.R
\name{saga_findfieldofextremevalue}
\alias{saga_findfieldofextremevalue}
\title{QGIS algorithm Find field of extreme value}
\usage{
saga_findfieldofextremevalue(
  INPUT = qgisprocess::qgis_default_value(),
  FIELDS = qgisprocess::qgis_default_value(),
  EXTREME_ID = qgisprocess::qgis_default_value(),
  EXTREME_VALUE = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  IDENTIFY = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input. Path to a vector layer.}

\item{FIELDS}{\code{field} - Attributes. The name of an existing field. ; delimited list of existing field names.}

\item{EXTREME_ID}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{EXTREME_VALUE}{\code{field} - Value. The name of an existing field. ; delimited list of existing field names.}

\item{TYPE}{\code{enum}  of \verb{("[0] minimum", "[1] maximum")} - TYPE. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{IDENTIFY}{\code{enum}  of \verb{("[0] name", "[1] index")} - Attribute Identification. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{vectorDestination} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Find field of extreme value (saga:findfieldofextremevalue)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output
}
}
}
