% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fireriskanalysis.R
\name{saga_fireriskanalysis}
\alias{saga_fireriskanalysis}
\title{QGIS algorithm Fire risk analysis}
\usage{
saga_fireriskanalysis(
  DEM = qgisprocess:::qgis_default_value(),
  FUEL = qgisprocess:::qgis_default_value(),
  WINDSPD = qgisprocess:::qgis_default_value(),
  WINDDIR = qgisprocess:::qgis_default_value(),
  M1H = qgisprocess:::qgis_default_value(),
  M10H = qgisprocess:::qgis_default_value(),
  M100H = qgisprocess:::qgis_default_value(),
  MHERB = qgisprocess:::qgis_default_value(),
  MWOOD = qgisprocess:::qgis_default_value(),
  VALUE = qgisprocess:::qgis_default_value(),
  BASEPROB = qgisprocess:::qgis_default_value(),
  MONTECARLO = qgisprocess:::qgis_default_value(),
  INTERVAL = qgisprocess:::qgis_default_value(),
  DANGER = qgisprocess:::qgis_default_value(),
  COMPPROB = qgisprocess:::qgis_default_value(),
  PRIORITY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{FUEL}{\code{raster} - Fuel Model. Path to a raster layer.}

\item{WINDSPD}{\code{raster} - Wind Speed. Path to a raster layer.}

\item{WINDDIR}{\code{raster} - Wind Direction. Path to a raster layer.}

\item{M1H}{\code{raster} - Dead Fuel Moisture 1H. Path to a raster layer.}

\item{M10H}{\code{raster} - Dead Fuel Moisture 10H. Path to a raster layer.}

\item{M100H}{\code{raster} - Dead Fuel Moisture 100H. Path to a raster layer.}

\item{MHERB}{\code{raster} - Herbaceous Fuel Moisture. Path to a raster layer.}

\item{MWOOD}{\code{raster} - Wood Fuel Moisture. Path to a raster layer.}

\item{VALUE}{\code{raster} - Value. Path to a raster layer.}

\item{BASEPROB}{\code{raster} - Base Probability. Path to a raster layer.}

\item{MONTECARLO}{\code{number} - Number of Events. A numeric value.}

\item{INTERVAL}{\code{number} - Fire Length. A numeric value.}

\item{DANGER}{\code{rasterDestination} - Danger. Path for new raster layer.}

\item{COMPPROB}{\code{rasterDestination} - Compound Probability. Path for new raster layer.}

\item{PRIORITY}{\code{rasterDestination} - Priority Index. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Fire risk analysis (saga:fireriskanalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item COMPPROB - outputRaster - Compound Probability
\item DANGER - outputRaster - Danger
\item PRIORITY - outputRaster - Priority Index
}
}
}
