% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_flatdetection.R
\name{saga_flatdetection}
\alias{saga_flatdetection}
\title{QGIS algorithm Flat detection}
\usage{
saga_flatdetection(
  DEM = qgisprocess::qgis_default_value(),
  FLAT_OUTPUT = qgisprocess::qgis_default_value(),
  NOFLATS = qgisprocess::qgis_default_value(),
  FLATS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{FLAT_OUTPUT}{\code{enum}  of \verb{("[0] elevation", "[1] enumeration")} - Flat Area Values. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NOFLATS}{\code{rasterDestination} - No Flats. Path for new raster layer.}

\item{FLATS}{\code{rasterDestination} - Flat Areas. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Flat detection (saga:flatdetection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLATS - outputRaster - Flat Areas
\item NOFLATS - outputRaster - No Flats
}
}
}
