% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_flattenpolygonlayer.R
\name{saga_flattenpolygonlayer}
\alias{saga_flattenpolygonlayer}
\title{QGIS algorithm Flatten polygon layer}
\usage{
saga_flattenpolygonlayer(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input. Path to a vector layer.}

\item{OUTPUT}{\code{vectorDestination} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Flatten polygon layer (saga:flattenpolygonlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output
}
}
}
