% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_flowaccumulationqmofesp.R
\name{saga_flowaccumulationqmofesp}
\alias{saga_flowaccumulationqmofesp}
\title{QGIS algorithm Flow accumulation (qm of esp)}
\usage{
saga_flowaccumulationqmofesp(
  DEM = qgisprocess::qgis_default_value(),
  PREPROC = qgisprocess::qgis_default_value(),
  DZFILL = qgisprocess::qgis_default_value(),
  FLOW = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{PREPROC}{\code{enum}  of \verb{("[0] none", "[1] fill sinks temporarily", "[2] fill sinks permanently")} - Preprocessing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DZFILL}{\code{number} - Fill Increment. A numeric value.}

\item{FLOW}{\code{rasterDestination} - Contributing Area. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Flow accumulation (qm of esp) (saga:flowaccumulationqmofesp)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLOW - outputRaster - Contributing Area
}
}
}
