% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_flowwidthandspecificcatchmentarea.R
\name{saga_flowwidthandspecificcatchmentarea}
\alias{saga_flowwidthandspecificcatchmentarea}
\title{QGIS algorithm Flow width and specific catchment area}
\usage{
saga_flowwidthandspecificcatchmentarea(
  DEM = qgisprocess::qgis_default_value(),
  TCA = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  WIDTH = qgisprocess::qgis_default_value(),
  SCA = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{TCA}{\code{raster} - Total Catchment Area (TCA). Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Deterministic 8", "[1] Multiple Flow Direction (Quinn et al. 1991)", "[2] Aspect")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{WIDTH}{\code{rasterDestination} - Flow Width. Path for new raster layer.}

\item{SCA}{\code{rasterDestination} - Specific Catchment Area (SCA). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Flow width and specific catchment area (saga:flowwidthandspecificcatchmentarea)
}
\details{
\subsection{Outputs description}{
\itemize{
\item SCA - outputRaster - Specific Catchment Area (SCA)
\item WIDTH - outputRaster - Flow Width
}
}
}
