% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fourierfiltervigra.R
\name{saga_fourierfiltervigra}
\alias{saga_fourierfiltervigra}
\title{QGIS algorithm Fourier filter (vigra)}
\usage{
saga_fourierfiltervigra(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  SCALE = qgisprocess:::qgis_default_value(),
  POWER = qgisprocess:::qgis_default_value(),
  RANGE_MIN = qgisprocess:::qgis_default_value(),
  RANGE_MAX = qgisprocess:::qgis_default_value(),
  FILTER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{SCALE}{\code{number} - Size of smoothing filter. A numeric value.}

\item{POWER}{\code{number} - Power. A numeric value.}

\item{RANGE_MIN}{\code{number} - Range (Min). A numeric value.}

\item{RANGE_MAX}{\code{number} - Range (Max). A numeric value.}

\item{FILTER}{\code{enum}  of \verb{("[0] gaussian", "[1] power of distance", "[2] include range", "[3] exclude range")} - Filter. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Fourier filter (vigra) (saga:fourierfiltervigra)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
