% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fouriertransformationopencv.R
\name{saga_fouriertransformationopencv}
\alias{saga_fouriertransformationopencv}
\title{QGIS algorithm Fourier transformation (opencv)}
\usage{
saga_fouriertransformationopencv(
  INPUT = qgisprocess::qgis_default_value(),
  REAL = qgisprocess::qgis_default_value(),
  IMAG = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{REAL}{\code{rasterDestination} - Fourier Transformation (Real). Path for new raster layer.}

\item{IMAG}{\code{rasterDestination} - Fourier Transformation (Imaginary). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Fourier transformation (opencv) (saga:fouriertransformationopencv)
}
\details{
\subsection{Outputs description}{
\itemize{
\item IMAG - outputRaster - Fourier Transformation (Imaginary)
\item REAL - outputRaster - Fourier Transformation (Real)
}
}
}
