% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fouriertransforminversevigra.R
\name{saga_fouriertransforminversevigra}
\alias{saga_fouriertransforminversevigra}
\title{QGIS algorithm Fourier transform inverse (vigra)}
\usage{
saga_fouriertransforminversevigra(
  REAL = qgisprocess::qgis_default_value(),
  IMAG = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  CENTER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{REAL}{\code{raster} - Real. Path to a raster layer.}

\item{IMAG}{\code{raster} - Imaginary. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{CENTER}{\code{boolean} - Centered. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Fourier transform inverse (vigra) (saga:fouriertransforminversevigra)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
