% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fouriertransformvigra.R
\name{saga_fouriertransformvigra}
\alias{saga_fouriertransformvigra}
\title{QGIS algorithm Fourier transform (vigra)}
\usage{
saga_fouriertransformvigra(
  INPUT = qgisprocess::qgis_default_value(),
  REAL = qgisprocess::qgis_default_value(),
  IMAG = qgisprocess::qgis_default_value(),
  CENTER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{REAL}{\code{rasterDestination} - Real. Path for new raster layer.}

\item{IMAG}{\code{rasterDestination} - Imaginary. Path for new raster layer.}

\item{CENTER}{\code{boolean} - Centered. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Fourier transform (vigra) (saga:fouriertransformvigra)
}
\details{
\subsection{Outputs description}{
\itemize{
\item IMAG - outputRaster - Imaginary
\item REAL - outputRaster - Real
}
}
}
