% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fragmentationalternative.R
\name{saga_fragmentationalternative}
\alias{saga_fragmentationalternative}
\title{QGIS algorithm Fragmentation (alternative)}
\usage{
saga_fragmentationalternative(
  CLASSES = qgisprocess:::qgis_default_value(),
  CLASS = qgisprocess:::qgis_default_value(),
  NEIGHBORHOOD_MIN = qgisprocess:::qgis_default_value(),
  NEIGHBORHOOD_MAX = qgisprocess:::qgis_default_value(),
  AGGREGATION = qgisprocess:::qgis_default_value(),
  WEIGHT = qgisprocess:::qgis_default_value(),
  DENSITY_MIN = qgisprocess:::qgis_default_value(),
  DENSITY_INT = qgisprocess:::qgis_default_value(),
  LEVEL_GROW = qgisprocess:::qgis_default_value(),
  DENSITY = qgisprocess:::qgis_default_value(),
  CONNECTIVITY = qgisprocess:::qgis_default_value(),
  FRAGMENTATION = qgisprocess:::qgis_default_value(),
  FRAGSTATS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CLASSES}{\code{raster} - Classification. Path to a raster layer.}

\item{CLASS}{\code{number} - Class Identifier. A numeric value.}

\item{NEIGHBORHOOD_MIN}{\code{number} - Neighborhood Min. A numeric value.}

\item{NEIGHBORHOOD_MAX}{\code{number} - Neighborhood Max. A numeric value.}

\item{AGGREGATION}{\code{enum}  of \verb{("[0] average", "[1] multiplicative")} - Level Aggregation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{WEIGHT}{\code{number} - Connectivity Weighting. A numeric value.}

\item{DENSITY_MIN}{\code{number} - Minimum Density (Percent). A numeric value.}

\item{DENSITY_INT}{\code{number} - Minimum Density for Interior Forest (Percent). A numeric value.}

\item{LEVEL_GROW}{\code{number} - Search Distance Increment. A numeric value.}

\item{DENSITY}{\code{rasterDestination} - Density (Percent). Path for new raster layer.}

\item{CONNECTIVITY}{\code{rasterDestination} - Connectivity (Percent). Path for new raster layer.}

\item{FRAGMENTATION}{\code{rasterDestination} - Fragmentation. Path for new raster layer.}

\item{FRAGSTATS}{\code{vectorDestination} - Summary. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Fragmentation (alternative) (saga:fragmentationalternative)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONNECTIVITY - outputRaster - Connectivity (Percent)
\item DENSITY - outputRaster - Density (Percent)
\item FRAGMENTATION - outputRaster - Fragmentation
\item FRAGSTATS - outputVector - Summary
}
}
}
