% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/saga_fragmentationclassesfromdensityandconnectivity.R
\name{saga_fragmentationclassesfromdensityandconnectivity}
\alias{saga_fragmentationclassesfromdensityandconnectivity}
\title{QGIS algorithm Fragmentation classes from density and connectivity}
\usage{
saga_fragmentationclassesfromdensityandconnectivity(
  DENSITY = qgisprocess::qgis_default_value(),
  CONNECTIVITY = qgisprocess::qgis_default_value(),
  BORDER = qgisprocess::qgis_default_value(),
  WEIGHT = qgisprocess::qgis_default_value(),
  DENSITY_MIN = qgisprocess::qgis_default_value(),
  DENSITY_INT = qgisprocess::qgis_default_value(),
  FRAGMENTATION = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DENSITY}{\code{raster} - Density (Percent). Path to a raster layer.}

\item{CONNECTIVITY}{\code{raster} - Connectivity (Percent). Path to a raster layer.}

\item{BORDER}{\code{boolean} - Add Border. 1 for true/yes. 0 for false/no.}

\item{WEIGHT}{\code{number} - Connectivity Weighting. A numeric value.}

\item{DENSITY_MIN}{\code{number} - Minimum Density (Percent). A numeric value.}

\item{DENSITY_INT}{\code{number} - Minimum Density for Interior Forest (Percent). A numeric value.}

\item{FRAGMENTATION}{\code{rasterDestination} - Fragmentation. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Fragmentation classes from density and connectivity (saga:fragmentationclassesfromdensityandconnectivity)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FRAGMENTATION - outputRaster - Fragmentation
}
}
}
