% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fragmentationstandard.R
\name{saga_fragmentationstandard}
\alias{saga_fragmentationstandard}
\title{QGIS algorithm Fragmentation (standard)}
\usage{
saga_fragmentationstandard(
  CLASSES = qgisprocess::qgis_default_value(),
  CLASS = qgisprocess::qgis_default_value(),
  NEIGHBORHOOD_MIN = qgisprocess::qgis_default_value(),
  NEIGHBORHOOD_MAX = qgisprocess::qgis_default_value(),
  AGGREGATION = qgisprocess::qgis_default_value(),
  BORDER = qgisprocess::qgis_default_value(),
  WEIGHT = qgisprocess::qgis_default_value(),
  DENSITY_MIN = qgisprocess::qgis_default_value(),
  DENSITY_INT = qgisprocess::qgis_default_value(),
  CIRCULAR = qgisprocess::qgis_default_value(),
  DIAGONAL = qgisprocess::qgis_default_value(),
  DENSITY = qgisprocess::qgis_default_value(),
  CONNECTIVITY = qgisprocess::qgis_default_value(),
  FRAGMENTATION = qgisprocess::qgis_default_value(),
  FRAGSTATS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{CLASSES}{\code{raster} - Classification. Path to a raster layer.}

\item{CLASS}{\code{number} - Class Identifier. A numeric value.}

\item{NEIGHBORHOOD_MIN}{\code{number} - Neighborhood Min. A numeric value.}

\item{NEIGHBORHOOD_MAX}{\code{number} - Neighborhood Max. A numeric value.}

\item{AGGREGATION}{\code{enum}  of \verb{("[0] average", "[1] multiplicative")} - Level Aggregation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{BORDER}{\code{boolean} - Add Border. 1 for true/yes. 0 for false/no.}

\item{WEIGHT}{\code{number} - Connectivity Weighting. A numeric value.}

\item{DENSITY_MIN}{\code{number} - Minimum Density (Percent). A numeric value.}

\item{DENSITY_INT}{\code{number} - Minimum Density for Interior Forest (Percent). A numeric value.}

\item{CIRCULAR}{\code{enum}  of \verb{("[0] square", "[1] circle")} - Neighborhood Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DIAGONAL}{\code{boolean} - Include diagonal neighbour relations. 1 for true/yes. 0 for false/no.}

\item{DENSITY}{\code{rasterDestination} - Density (Percent). Path for new raster layer.}

\item{CONNECTIVITY}{\code{rasterDestination} - Connectivity (Percent). Path for new raster layer.}

\item{FRAGMENTATION}{\code{rasterDestination} - Fragmentation. Path for new raster layer.}

\item{FRAGSTATS}{\code{vectorDestination} - Summary. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Fragmentation (standard) (saga:fragmentationstandard)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DENSITY - outputRaster - Density
\item CONNECTIVITY - outputRaster - Connectivity
\item FRAGMENTATION - outputRaster - Fragmentation
\item FRAGSTATS - outputVector - Summary
}
}
}
