% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_function.R
\name{saga_function}
\alias{saga_function}
\title{QGIS algorithm Function}
\usage{
saga_function(
  XMIN = qgisprocess::qgis_default_value(),
  XMAX = qgisprocess::qgis_default_value(),
  YMIN = qgisprocess::qgis_default_value(),
  YMAX = qgisprocess::qgis_default_value(),
  FORMUL = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{XMIN}{\code{number} - xmin. A numeric value.}

\item{XMAX}{\code{number} - xmax. A numeric value.}

\item{YMIN}{\code{number} - ymin. A numeric value.}

\item{YMAX}{\code{number} - ymax. A numeric value.}

\item{FORMUL}{\code{string} - Formula. String value.}

\item{RESULT}{\code{rasterDestination} - Function. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Function (saga:function)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Function
}
}
}
