% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fuzzyintersectionand.R
\name{saga_fuzzyintersectionand}
\alias{saga_fuzzyintersectionand}
\title{QGIS algorithm Fuzzy intersection (and)}
\usage{
saga_fuzzyintersectionand(
  GRIDS = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  AND = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{TYPE}{\code{enum}  of \verb{("[0] min(a, b) (non-interactive)", "[1] a * b", "[2] max(0, a + b - 1)")} - Operator Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{AND}{\code{rasterDestination} - Intersection. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Fuzzy intersection (and) (saga:fuzzyintersectionand)
}
\details{
\subsection{Outputs description}{
\itemize{
\item AND - outputRaster - Intersection
}
}
}
