% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fuzzylandformelementclassification.R
\name{saga_fuzzylandformelementclassification}
\alias{saga_fuzzylandformelementclassification}
\title{QGIS algorithm Fuzzy landform element classification}
\usage{
saga_fuzzylandformelementclassification(
  SLOPE = qgisprocess:::qgis_default_value(),
  MINCURV = qgisprocess:::qgis_default_value(),
  MAXCURV = qgisprocess:::qgis_default_value(),
  PCURV = qgisprocess:::qgis_default_value(),
  TCURV = qgisprocess:::qgis_default_value(),
  PLAIN = qgisprocess:::qgis_default_value(),
  PIT = qgisprocess:::qgis_default_value(),
  PEAK = qgisprocess:::qgis_default_value(),
  RIDGE = qgisprocess:::qgis_default_value(),
  CHANNEL = qgisprocess:::qgis_default_value(),
  SADDLE = qgisprocess:::qgis_default_value(),
  BSLOPE = qgisprocess:::qgis_default_value(),
  FSLOPE = qgisprocess:::qgis_default_value(),
  SSLOPE = qgisprocess:::qgis_default_value(),
  HOLLOW = qgisprocess:::qgis_default_value(),
  FHOLLOW = qgisprocess:::qgis_default_value(),
  SHOLLOW = qgisprocess:::qgis_default_value(),
  SPUR = qgisprocess:::qgis_default_value(),
  FSPUR = qgisprocess:::qgis_default_value(),
  SSPUR = qgisprocess:::qgis_default_value(),
  FORM = qgisprocess:::qgis_default_value(),
  MEM = qgisprocess:::qgis_default_value(),
  ENTROPY = qgisprocess:::qgis_default_value(),
  CI = qgisprocess:::qgis_default_value(),
  SLOPETODEG = qgisprocess:::qgis_default_value(),
  T_SLOPE_MIN = qgisprocess:::qgis_default_value(),
  T_SLOPE_MAX = qgisprocess:::qgis_default_value(),
  T_CURVE_MIN = qgisprocess:::qgis_default_value(),
  T_CURVE_MAX = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SLOPE}{\code{raster} - Slope. Path to a raster layer.}

\item{MINCURV}{\code{raster} - Minimum Curvature. Path to a raster layer.}

\item{MAXCURV}{\code{raster} - Maximum Curvature. Path to a raster layer.}

\item{PCURV}{\code{raster} - Profile Curvature. Path to a raster layer.}

\item{TCURV}{\code{raster} - Tangential Curvature. Path to a raster layer.}

\item{PLAIN}{\code{rasterDestination} - Plain. Path for new raster layer.}

\item{PIT}{\code{rasterDestination} - Pit. Path for new raster layer.}

\item{PEAK}{\code{rasterDestination} - Peak. Path for new raster layer.}

\item{RIDGE}{\code{rasterDestination} - Ridge. Path for new raster layer.}

\item{CHANNEL}{\code{rasterDestination} - Channel. Path for new raster layer.}

\item{SADDLE}{\code{rasterDestination} - Saddle. Path for new raster layer.}

\item{BSLOPE}{\code{rasterDestination} - Back Slope. Path for new raster layer.}

\item{FSLOPE}{\code{rasterDestination} - Foot Slope. Path for new raster layer.}

\item{SSLOPE}{\code{rasterDestination} - Shoulder Slope. Path for new raster layer.}

\item{HOLLOW}{\code{rasterDestination} - Hollow. Path for new raster layer.}

\item{FHOLLOW}{\code{rasterDestination} - Foot Hollow. Path for new raster layer.}

\item{SHOLLOW}{\code{rasterDestination} - Shoulder Hollow. Path for new raster layer.}

\item{SPUR}{\code{rasterDestination} - Spur. Path for new raster layer.}

\item{FSPUR}{\code{rasterDestination} - Foot Spur. Path for new raster layer.}

\item{SSPUR}{\code{rasterDestination} - Shoulder Spur. Path for new raster layer.}

\item{FORM}{\code{rasterDestination} - Landform. Path for new raster layer.}

\item{MEM}{\code{rasterDestination} - Maximum Membership. Path for new raster layer.}

\item{ENTROPY}{\code{rasterDestination} - Entropy. Path for new raster layer.}

\item{CI}{\code{rasterDestination} - Confusion Index. Path for new raster layer.}

\item{SLOPETODEG}{\code{enum}  of \verb{("[0] degree", "[1] radians")} - Slope Grid Units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{T_SLOPE_MIN}{\code{number} - Slope Thresholds (Degree) (Min). A numeric value.}

\item{T_SLOPE_MAX}{\code{number} - Slope Thresholds (Degree) (Max). A numeric value.}

\item{T_CURVE_MIN}{\code{number} - Curvature Thresholds (1 / m) (Min). A numeric value.}

\item{T_CURVE_MAX}{\code{number} - Curvature Thresholds (1 / m) (Max). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Fuzzy landform element classification (saga:fuzzylandformelementclassification)
}
\details{
\subsection{Outputs description}{
\itemize{
\item BSLOPE - outputRaster - Back Slope
\item CHANNEL - outputRaster - Channel
\item CI - outputRaster - Confusion Index
\item ENTROPY - outputRaster - Entropy
\item FHOLLOW - outputRaster - Foot Hollow
\item FORM - outputRaster - Landform
\item FSLOPE - outputRaster - Foot Slope
\item FSPUR - outputRaster - Foot Spur
\item HOLLOW - outputRaster - Hollow
\item MEM - outputRaster - Maximum Membership
\item PEAK - outputRaster - Peak
\item PIT - outputRaster - Pit
\item PLAIN - outputRaster - Plain
\item RIDGE - outputRaster - Ridge
\item SADDLE - outputRaster - Saddle
\item SHOLLOW - outputRaster - Shoulder Hollow
\item SPUR - outputRaster - Spur
\item SSLOPE - outputRaster - Shoulder Slope
\item SSPUR - outputRaster - Shoulder Spur
}
}
}
