% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fuzzyunionor.R
\name{saga_fuzzyunionor}
\alias{saga_fuzzyunionor}
\title{QGIS algorithm Fuzzy union (or)}
\usage{
saga_fuzzyunionor(
  GRIDS = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  OR = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{TYPE}{\code{enum}  of \verb{("[0] max(a, b) (non-interactive)", "[1] a + b - a * b", "[2] min(1, a + b)")} - Operator Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OR}{\code{rasterDestination} - Union. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Fuzzy union (or) (saga:fuzzyunionor)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OR - outputRaster - Union
}
}
}
