% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_gaussianfilter.R
\name{saga_gaussianfilter}
\alias{saga_gaussianfilter}
\title{QGIS algorithm Gaussian filter}
\usage{
saga_gaussianfilter(
  INPUT = qgisprocess::qgis_default_value(),
  SIGMA = qgisprocess::qgis_default_value(),
  MODE = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{SIGMA}{\code{number} - Standard Deviation. A numeric value.}

\item{MODE}{\code{enum}  of \verb{("[0] Square", "[1] Circle")} - Search Mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Search Radius. A numeric value.}

\item{RESULT}{\code{rasterDestination} - Filtered Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Gaussian filter (saga:gaussianfilter)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Filtered Grid
}
}
}
