% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_generateshapes.R
\name{saga_generateshapes}
\alias{saga_generateshapes}
\title{QGIS algorithm Generate shapes}
\usage{
saga_generateshapes(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD_ID = qgisprocess::qgis_default_value(),
  FIELD_X = qgisprocess::qgis_default_value(),
  FIELD_Y = qgisprocess::qgis_default_value(),
  SHAPE_TYPE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input. Path to a vector layer.}

\item{FIELD_ID}{\code{field} - ID. The name of an existing field. ; delimited list of existing field names.}

\item{FIELD_X}{\code{field} - X. The name of an existing field. ; delimited list of existing field names.}

\item{FIELD_Y}{\code{field} - Y. The name of an existing field. ; delimited list of existing field names.}

\item{SHAPE_TYPE}{\code{enum}  of \verb{("[0] Point(s)", "[1] Line(s)", "[2] Polygon(s)")} - Shape Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{vectorDestination} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Generate shapes (saga:generateshapes)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output
}
}
}
