% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_geodesicmorphologicalreconstruction.R
\name{saga_geodesicmorphologicalreconstruction}
\alias{saga_geodesicmorphologicalreconstruction}
\title{QGIS algorithm Geodesic morphological reconstruction}
\usage{
saga_geodesicmorphologicalreconstruction(
  INPUT_GRID = qgisprocess::qgis_default_value(),
  OBJECT_GRID = qgisprocess::qgis_default_value(),
  DIFFERENCE_GRID = qgisprocess::qgis_default_value(),
  SHIFT_VALUE = qgisprocess::qgis_default_value(),
  BORDER_YES_NO = qgisprocess::qgis_default_value(),
  BIN_YES_NO = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT_GRID}{\code{raster} - Input Grid. Path to a raster layer.}

\item{OBJECT_GRID}{\code{rasterDestination} - Object Grid. Path for new raster layer.}

\item{DIFFERENCE_GRID}{\code{rasterDestination} - Difference Input - Reconstruction. Path for new raster layer.}

\item{SHIFT_VALUE}{\code{number} - Shift value. A numeric value.}

\item{BORDER_YES_NO}{\code{boolean} - Preserve 1px border Yes/No. 1 for true/yes. 0 for false/no.}

\item{BIN_YES_NO}{\code{boolean} - Create a binary mask Yes/No. 1 for true/yes. 0 for false/no.}

\item{THRESHOLD}{\code{number} - Threshold. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Geodesic morphological reconstruction (saga:geodesicmorphologicalreconstruction)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OBJECT_GRID - outputRaster - Object Grid
\item DIFFERENCE_GRID - outputRaster - Difference Input
}
}
}
