% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_geometricfigures.R
\name{saga_geometricfigures}
\alias{saga_geometricfigures}
\title{QGIS algorithm Geometric figures}
\usage{
saga_geometricfigures(
  CELL_COUNT = qgisprocess::qgis_default_value(),
  CELL_SIZE = qgisprocess::qgis_default_value(),
  FIGURE = qgisprocess::qgis_default_value(),
  PLANE = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CELL_COUNT}{\code{number} - Cell Count. A numeric value.}

\item{CELL_SIZE}{\code{number} - Cell Size. A numeric value.}

\item{FIGURE}{\code{enum}  of \verb{("[0] Cone (up)", "[1] Cone (down)", "[2] Plane")} - Figure. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PLANE}{\code{number} - Direction of Plane (Degree). A numeric value.}

\item{RESULT}{\code{rasterDestination} - Result. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Geometric figures (saga:geometricfigures)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Result
}
}
}
