% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/saga_gradientvectorfrompolartocartesiancoordinates.R
\name{saga_gradientvectorfrompolartocartesiancoordinates}
\alias{saga_gradientvectorfrompolartocartesiancoordinates}
\title{QGIS algorithm Gradient vector from polar to cartesian coordinates}
\usage{
saga_gradientvectorfrompolartocartesiancoordinates(
  DIR = qgisprocess::qgis_default_value(),
  LEN = qgisprocess::qgis_default_value(),
  UNITS = qgisprocess::qgis_default_value(),
  SYSTEM = qgisprocess::qgis_default_value(),
  SYSTEM_ZERO = qgisprocess::qgis_default_value(),
  SYSTEM_ORIENT = qgisprocess::qgis_default_value(),
  DX = qgisprocess::qgis_default_value(),
  DY = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DIR}{\code{raster} - Direction. Path to a raster layer.}

\item{LEN}{\code{raster} - Length. Path to a raster layer.}

\item{UNITS}{\code{enum}  of \verb{("[0] radians", "[1] degree")} - Polar Angle Units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SYSTEM}{\code{enum}  of \verb{("[0] mathematical", "[1] geographical", "[2] user defined")} - Polar Coordinate System. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SYSTEM_ZERO}{\code{number} - User defined Zero Direction. A numeric value.}

\item{SYSTEM_ORIENT}{\code{enum}  of \verb{("[0] clockwise", "[1] counterclockwise")} - User defined Orientation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DX}{\code{rasterDestination} - X Component. Path for new raster layer.}

\item{DY}{\code{rasterDestination} - Y Component. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Gradient vector from polar to cartesian coordinates (saga:gradientvectorfrompolartocartesiancoordinates)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DX - outputRaster - X Component
\item DY - outputRaster - Y Component
}
}
}
