% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_gradientvectorsfromsurface.R
\name{saga_gradientvectorsfromsurface}
\alias{saga_gradientvectorsfromsurface}
\title{QGIS algorithm Gradient vectors from surface}
\usage{
saga_gradientvectorsfromsurface(
  SURFACE = qgisprocess::qgis_default_value(),
  STEP = qgisprocess::qgis_default_value(),
  SIZE_MIN = qgisprocess::qgis_default_value(),
  SIZE_MAX = qgisprocess::qgis_default_value(),
  AGGR = qgisprocess::qgis_default_value(),
  STYLE = qgisprocess::qgis_default_value(),
  VECTORS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SURFACE}{\code{raster} - Surface. Path to a raster layer.}

\item{STEP}{\code{number} - Step. A numeric value.}

\item{SIZE_MIN}{\code{number} - Size Range Min. A numeric value.}

\item{SIZE_MAX}{\code{number} - Size Range Max. A numeric value.}

\item{AGGR}{\code{enum}  of \verb{("[0] nearest neighbour", "[1] mean value")} - Aggregation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{STYLE}{\code{enum}  of \verb{("[0] simple line", "[1] arrow", "[2] arrow (centered to cell)")} - Style. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{VECTORS}{\code{vectorDestination} - Gradient Vectors. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Gradient vectors from surface (saga:gradientvectorsfromsurface)
}
\details{
\subsection{Outputs description}{
\itemize{
\item VECTORS - outputVector - Gradient Vectors
}
}
}
