% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_gridstatisticsforpoints.R
\name{saga_gridstatisticsforpoints}
\alias{saga_gridstatisticsforpoints}
\title{QGIS algorithm Grid statistics for points}
\usage{
saga_gridstatisticsforpoints(
  GRIDS = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  KERNEL_TYPE = qgisprocess:::qgis_default_value(),
  KERNEL_SIZE = qgisprocess:::qgis_default_value(),
  NAMING = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  COUNT = qgisprocess:::qgis_default_value(),
  MIN = qgisprocess:::qgis_default_value(),
  MAX = qgisprocess:::qgis_default_value(),
  RANGE = qgisprocess:::qgis_default_value(),
  SUM = qgisprocess:::qgis_default_value(),
  MEAN = qgisprocess:::qgis_default_value(),
  VAR = qgisprocess:::qgis_default_value(),
  STDDEV = qgisprocess:::qgis_default_value(),
  QUANTILE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{KERNEL_TYPE}{\code{enum}  of \verb{("[0] square", "[1] circle")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_SIZE}{\code{number} - Kernel Size. A numeric value.}

\item{NAMING}{\code{enum}  of \verb{("[0] grid number", "[1] grid name")} - Field Naming. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESULT}{\code{vectorDestination} - Statistics. Path for new vector layer.}

\item{COUNT}{\code{boolean} - Number of Cells. 1 for true/yes. 0 for false/no.}

\item{MIN}{\code{boolean} - Minimum. 1 for true/yes. 0 for false/no.}

\item{MAX}{\code{boolean} - Maximum. 1 for true/yes. 0 for false/no.}

\item{RANGE}{\code{boolean} - Range. 1 for true/yes. 0 for false/no.}

\item{SUM}{\code{boolean} - Sum. 1 for true/yes. 0 for false/no.}

\item{MEAN}{\code{boolean} - Mean. 1 for true/yes. 0 for false/no.}

\item{VAR}{\code{boolean} - Variance. 1 for true/yes. 0 for false/no.}

\item{STDDEV}{\code{boolean} - Standard Deviation. 1 for true/yes. 0 for false/no.}

\item{QUANTILE}{\code{number} - Quantile. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Grid statistics for points (saga:gridstatisticsforpoints)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Statistics
}
}
}
