% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_gwrformultiplepredictorlayers.R
\name{saga_gwrformultiplepredictorlayers}
\alias{saga_gwrformultiplepredictorlayers}
\title{QGIS algorithm Gwr for multiple predictor layers}
\usage{
saga_gwrformultiplepredictorlayers(
  PREDICTORS = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  DEPENDENT = qgisprocess::qgis_default_value(),
  RESOLUTION = qgisprocess::qgis_default_value(),
  RESOLUTION_VAL = qgisprocess::qgis_default_value(),
  DW_WEIGHTING = qgisprocess::qgis_default_value(),
  DW_IDW_POWER = qgisprocess::qgis_default_value(),
  DW_IDW_OFFSET = qgisprocess::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess::qgis_default_value(),
  SEARCH_RANGE = qgisprocess::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_ALL = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MIN = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MAX = qgisprocess::qgis_default_value(),
  SEARCH_DIRECTION = qgisprocess::qgis_default_value(),
  REGRESSION = qgisprocess::qgis_default_value(),
  QUALITY = qgisprocess::qgis_default_value(),
  RESIDUALS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{PREDICTORS}{\code{multilayer} - Predictors. .}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{DEPENDENT}{\code{field} - Dependent Variable. The name of an existing field. ; delimited list of existing field names.}

\item{RESOLUTION}{\code{enum}  of \verb{("[0] same as predictors", "[1] user defined")} - Model Resolution. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESOLUTION_VAL}{\code{number} - Resolution. A numeric value.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DW_IDW_OFFSET}{\code{boolean} - Inverse Distance Offset. 1 for true/yes. 0 for false/no.}

\item{DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{SEARCH_RANGE}{\code{enum}  of \verb{("[0] search radius (local)", "[1] no search radius (global)")} - Search Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_RADIUS}{\code{number} - Search Radius. A numeric value.}

\item{SEARCH_POINTS_ALL}{\code{enum}  of \verb{("[0] maximum number of nearest points", "[1] all points within search distance")} - Number of Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_POINTS_MIN}{\code{number} - Minimum. A numeric value.}

\item{SEARCH_POINTS_MAX}{\code{number} - Maximum. A numeric value.}

\item{SEARCH_DIRECTION}{\code{enum}  of \verb{("[0] all directions", "[1] quadrants")} - Search Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{REGRESSION}{\code{rasterDestination} - Regression. Path for new raster layer.}

\item{QUALITY}{\code{rasterDestination} - Coefficient of Determination. Path for new raster layer.}

\item{RESIDUALS}{\code{vectorDestination} - Residuals. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Gwr for multiple predictor layers (saga:gwrformultiplepredictorlayers)
}
\details{
\subsection{Outputs description}{
\itemize{
\item QUALITY - outputRaster - Coefficient of Determination
\item REGRESSION - outputRaster - Regression
\item RESIDUALS - outputVector - Residuals
}
}
}
