% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_gwrforsinglepredictorgriddedmodeloutput.R
\name{saga_gwrforsinglepredictorgriddedmodeloutput}
\alias{saga_gwrforsinglepredictorgriddedmodeloutput}
\title{QGIS algorithm Gwr for single predictor (gridded model output)}
\usage{
saga_gwrforsinglepredictorgriddedmodeloutput(
  POINTS = qgisprocess::qgis_default_value(),
  DEPENDENT = qgisprocess::qgis_default_value(),
  PREDICTOR = qgisprocess::qgis_default_value(),
  TARGET_DEFINITION = qgisprocess::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess::qgis_default_value(),
  TARGET_USER_XMIN = qgisprocess::qgis_default_value(),
  TARGET_USER_XMAX = qgisprocess::qgis_default_value(),
  TARGET_USER_YMIN = qgisprocess::qgis_default_value(),
  TARGET_USER_YMAX = qgisprocess::qgis_default_value(),
  TARGET_USER_FITS = qgisprocess::qgis_default_value(),
  TARGET_TEMPLATE = qgisprocess::qgis_default_value(),
  TARGET_OUT_GRID = qgisprocess::qgis_default_value(),
  INTERCEPT = qgisprocess::qgis_default_value(),
  SLOPE = qgisprocess::qgis_default_value(),
  QUALITY = qgisprocess::qgis_default_value(),
  DW_WEIGHTING = qgisprocess::qgis_default_value(),
  DW_IDW_POWER = qgisprocess::qgis_default_value(),
  DW_IDW_OFFSET = qgisprocess::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess::qgis_default_value(),
  SEARCH_RANGE = qgisprocess::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_ALL = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MIN = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MAX = qgisprocess::qgis_default_value(),
  SEARCH_DIRECTION = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{DEPENDENT}{\code{field} - Dependent Variable. The name of an existing field. ; delimited list of existing field names.}

\item{PREDICTOR}{\code{field} - Predictor. The name of an existing field. ; delimited list of existing field names.}

\item{TARGET_DEFINITION}{\code{enum}  of \verb{("[0] user defined", "[1] grid or grid system")} - Target Grid System. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{TARGET_USER_XMIN}{\code{number} - Left. A numeric value.}

\item{TARGET_USER_XMAX}{\code{number} - Right. A numeric value.}

\item{TARGET_USER_YMIN}{\code{number} - Bottom. A numeric value.}

\item{TARGET_USER_YMAX}{\code{number} - Top. A numeric value.}

\item{TARGET_USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_TEMPLATE}{\code{raster} - Target System. Path to a raster layer.}

\item{TARGET_OUT_GRID}{\code{rasterDestination} - Target Grid. Path for new raster layer.}

\item{INTERCEPT}{\code{rasterDestination} - Intercept. Path for new raster layer.}

\item{SLOPE}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{QUALITY}{\code{rasterDestination} - Quality. Path for new raster layer.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DW_IDW_OFFSET}{\code{boolean} - Inverse Distance Offset. 1 for true/yes. 0 for false/no.}

\item{DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{SEARCH_RANGE}{\code{enum}  of \verb{("[0] local", "[1] global")} - Search Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_RADIUS}{\code{number} - Maximum Search Distance. A numeric value.}

\item{SEARCH_POINTS_ALL}{\code{enum}  of \verb{("[0] maximum number of nearest points", "[1] all points within search distance")} - Number of Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_POINTS_MIN}{\code{number} - Minimum. A numeric value.}

\item{SEARCH_POINTS_MAX}{\code{number} - Maximum. A numeric value.}

\item{SEARCH_DIRECTION}{\code{enum}  of \verb{("[0] all directions", "[1] quadrants")} - Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Gwr for single predictor (gridded model output) (saga:gwrforsinglepredictorgriddedmodeloutput)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TARGET_OUT_GRID - outputRaster - Target Grid
\item INTERCEPT - outputRaster - Intercept
\item SLOPE - outputRaster - Slope
\item QUALITY - outputRaster - Quality
}
}
}
