% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_gwrforsinglepredictorlayer.R
\name{saga_gwrforsinglepredictorlayer}
\alias{saga_gwrforsinglepredictorlayer}
\title{QGIS algorithm Gwr for single predictor layer}
\usage{
saga_gwrforsinglepredictorlayer(
  POINTS = qgisprocess::qgis_default_value(),
  PREDICTOR = qgisprocess::qgis_default_value(),
  DEPENDENT = qgisprocess::qgis_default_value(),
  DW_WEIGHTING = qgisprocess::qgis_default_value(),
  DW_IDW_POWER = qgisprocess::qgis_default_value(),
  DW_IDW_OFFSET = qgisprocess::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess::qgis_default_value(),
  SEARCH_RANGE = qgisprocess::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_ALL = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MIN = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MAX = qgisprocess::qgis_default_value(),
  SEARCH_DIRECTION = qgisprocess::qgis_default_value(),
  RESIDUALS = qgisprocess::qgis_default_value(),
  REGRESSION = qgisprocess::qgis_default_value(),
  QUALITY = qgisprocess::qgis_default_value(),
  INTERCEPT = qgisprocess::qgis_default_value(),
  SLOPE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{PREDICTOR}{\code{raster} - Predictor. Path to a raster layer.}

\item{DEPENDENT}{\code{field} - Dependent Variable. The name of an existing field. ; delimited list of existing field names.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DW_IDW_OFFSET}{\code{boolean} - Inverse Distance Offset. 1 for true/yes. 0 for false/no.}

\item{DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{SEARCH_RANGE}{\code{enum}  of \verb{("[0] local", "[1] global")} - Search Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_RADIUS}{\code{number} - Maximum Search Distance. A numeric value.}

\item{SEARCH_POINTS_ALL}{\code{enum}  of \verb{("[0] maximum number of nearest points", "[1] all points within search distance")} - Number of Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_POINTS_MIN}{\code{number} - Minimum. A numeric value.}

\item{SEARCH_POINTS_MAX}{\code{number} - Maximum. A numeric value.}

\item{SEARCH_DIRECTION}{\code{enum}  of \verb{("[0] all directions", "[1] quadrants")} - Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESIDUALS}{\code{vectorDestination} - Residuals. Path for new vector layer.}

\item{REGRESSION}{\code{rasterDestination} - Regression. Path for new raster layer.}

\item{QUALITY}{\code{rasterDestination} - Coefficient of Determination. Path for new raster layer.}

\item{INTERCEPT}{\code{rasterDestination} - Intercept. Path for new raster layer.}

\item{SLOPE}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Gwr for single predictor layer (saga:gwrforsinglepredictorlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INTERCEPT - outputRaster - Intercept
\item QUALITY - outputRaster - Coefficient of Determination
\item REGRESSION - outputRaster - Regression
\item RESIDUALS - outputVector - Residuals
\item SLOPE - outputRaster - Slope
}
}
}
