% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_histogramsurface.R
\name{saga_histogramsurface}
\alias{saga_histogramsurface}
\title{QGIS algorithm Histogram surface}
\usage{
saga_histogramsurface(
  GRID = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  HIST = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] rows", "[1] columns", "[2] circle")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{HIST}{\code{rasterDestination} - Histogram. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Histogram surface (saga:histogramsurface)
}
\details{
\subsection{Outputs description}{
\itemize{
\item HIST - outputRaster - Histogram
}
}
}
