% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_hypsometry.R
\name{saga_hypsometry}
\alias{saga_hypsometry}
\title{QGIS algorithm Hypsometry}
\usage{
saga_hypsometry(
  ELEVATION = qgisprocess::qgis_default_value(),
  COUNT = qgisprocess::qgis_default_value(),
  SORTING = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  BZRANGE = qgisprocess::qgis_default_value(),
  ZRANGE_MIN = qgisprocess::qgis_default_value(),
  ZRANGE_MAX = qgisprocess::qgis_default_value(),
  TABLE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{COUNT}{\code{number} - Number of Classes. A numeric value.}

\item{SORTING}{\code{enum}  of \verb{("[0] up", "[1] down")} - Sort. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("[0] height", "[1] area")} - Classification Constant. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{BZRANGE}{\code{boolean} - Use Z-Range. 1 for true/yes. 0 for false/no.}

\item{ZRANGE_MIN}{\code{number} - Z-Range Min. A numeric value.}

\item{ZRANGE_MAX}{\code{number} - Z-Range Max. A numeric value.}

\item{TABLE}{\code{vectorDestination} - Hypsometry. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Hypsometry (saga:hypsometry)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TABLE - outputVector - Hypsometry
}
}
}
