% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_imcorrfeaturetracking.R
\name{saga_imcorrfeaturetracking}
\alias{saga_imcorrfeaturetracking}
\title{QGIS algorithm Imcorr - feature tracking}
\usage{
saga_imcorrfeaturetracking(
  GRID_1 = qgisprocess::qgis_default_value(),
  GRID_2 = qgisprocess::qgis_default_value(),
  DTM_1 = qgisprocess::qgis_default_value(),
  DTM_2 = qgisprocess::qgis_default_value(),
  CORRPOINTS = qgisprocess::qgis_default_value(),
  CORRLINES = qgisprocess::qgis_default_value(),
  SEARCH_CHIPSIZE = qgisprocess::qgis_default_value(),
  REF_CHIPSIZE = qgisprocess::qgis_default_value(),
  GRID_SPACING = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID_1}{\code{raster} - Grid 1. Path to a raster layer.}

\item{GRID_2}{\code{raster} - Grid 2. Path to a raster layer.}

\item{DTM_1}{\code{raster} - DTM 1. Path to a raster layer.}

\item{DTM_2}{\code{raster} - DTM 2. Path to a raster layer.}

\item{CORRPOINTS}{\code{vectorDestination} - Correlated Points. Path for new vector layer.}

\item{CORRLINES}{\code{vectorDestination} - Displacement Vector. Path for new vector layer.}

\item{SEARCH_CHIPSIZE}{\code{enum}  of \verb{("[0] 16x16", "[1] 32x32", "[2] 64x64", "[3] 128x128", "[4] 256x256")} - Search Chip Size (Cells). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{REF_CHIPSIZE}{\code{enum}  of \verb{("[0] 16x16", "[1] 32x32", "[2] 64x64", "[3] 128x128")} - Reference Chip Size (Cells). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRID_SPACING}{\code{number} - Grid Spacing (Map Units). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Imcorr - feature tracking (saga:imcorrfeaturetracking)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CORRPOINTS - outputVector - Correlated Points
\item CORRLINES - outputVector - Displacement Vector
}
}
}
