% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_invertdatanodata.R
\name{saga_invertdatanodata}
\alias{saga_invertdatanodata}
\title{QGIS algorithm Invert data/no-data}
\usage{
saga_invertdatanodata(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Result. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Invert data/no-data (saga:invertdatanodata)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Result
}
}
}
