% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_invertgrid.R
\name{saga_invertgrid}
\alias{saga_invertgrid}
\title{QGIS algorithm Invert grid}
\usage{
saga_invertgrid(
  GRID = qgisprocess:::qgis_default_value(),
  INVERSE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{INVERSE}{\code{rasterDestination} - Inverse Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Invert grid (saga:invertgrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INVERSE - outputRaster - Inverse Grid
}
}
}
