% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_kmeansclusteringforgrids.R
\name{saga_kmeansclusteringforgrids}
\alias{saga_kmeansclusteringforgrids}
\title{QGIS algorithm K-means clustering for grids}
\usage{
saga_kmeansclusteringforgrids(
  GRIDS = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  NCLUSTER = qgisprocess::qgis_default_value(),
  MAXITER = qgisprocess::qgis_default_value(),
  NORMALISE = qgisprocess::qgis_default_value(),
  OLDVERSION = qgisprocess::qgis_default_value(),
  UPDATEVIEW = qgisprocess::qgis_default_value(),
  CLUSTER = qgisprocess::qgis_default_value(),
  STATISTICS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{METHOD}{\code{enum}  of \verb{("[0] Iterative Minimum Distance (Forgy 1965)", "[1] Hill-Climbing (Rubin 1967)", "[2] Combined Minimum Distance / Hillclimbing")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NCLUSTER}{\code{number} - Clusters. A numeric value.}

\item{MAXITER}{\code{number} - Maximum Iterations. A numeric value.}

\item{NORMALISE}{\code{boolean} - Normalise. 1 for true/yes. 0 for false/no.}

\item{OLDVERSION}{\code{boolean} - Old Version. 1 for true/yes. 0 for false/no.}

\item{UPDATEVIEW}{\code{boolean} - Update View. 1 for true/yes. 0 for false/no.}

\item{CLUSTER}{\code{rasterDestination} - Clusters. Path for new raster layer.}

\item{STATISTICS}{\code{vectorDestination} - Statistics. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA K-means clustering for grids (saga:kmeansclusteringforgrids)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLUSTER - outputRaster - Clusters
\item STATISTICS - outputVector - Statistics
}
}
}
