% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_knearestneighboursclassificationopencv.R
\name{saga_knearestneighboursclassificationopencv}
\alias{saga_knearestneighboursclassificationopencv}
\title{QGIS algorithm K-nearest neighbours classification (opencv)}
\usage{
saga_knearestneighboursclassificationopencv(
  FEATURES = qgisprocess::qgis_default_value(),
  NORMALIZE = qgisprocess::qgis_default_value(),
  TRAIN_AREAS = qgisprocess::qgis_default_value(),
  TRAIN_CLASS = qgisprocess::qgis_default_value(),
  CLASSES = qgisprocess::qgis_default_value(),
  NEIGHBOURS = qgisprocess::qgis_default_value(),
  TRAINING = qgisprocess::qgis_default_value(),
  ALGORITHM = qgisprocess::qgis_default_value(),
  EMAX = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FEATURES}{\code{multilayer} - Features. .}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no.}

\item{TRAIN_AREAS}{\code{source} - Training Areas. Path to a vector layer.}

\item{TRAIN_CLASS}{\code{source} - Class Identifier. Path to a vector layer.}

\item{CLASSES}{\code{rasterDestination} - Classification. Path for new raster layer.}

\item{NEIGHBOURS}{\code{number} - Default Number of Neighbours. A numeric value.}

\item{TRAINING}{\code{enum}  of \verb{("[0] classification", "[1] regression model")} - Training Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ALGORITHM}{\code{enum}  of \verb{("[0] brute force", "[1] KD Tree")} - Algorithm Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EMAX}{\code{number} - Parameter for KD Tree implementation. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA K-nearest neighbours classification (opencv) (saga:knearestneighboursclassificationopencv)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Classification
}
}
}
