% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_lakeflood.R
\name{saga_lakeflood}
\alias{saga_lakeflood}
\title{QGIS algorithm Lake flood}
\usage{
saga_lakeflood(
  ELEV = qgisprocess:::qgis_default_value(),
  SEEDS = qgisprocess:::qgis_default_value(),
  OUTDEPTH = qgisprocess:::qgis_default_value(),
  OUTLEVEL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEV}{\code{raster} - DEM. Path to a raster layer.}

\item{SEEDS}{\code{raster} - Seeds. Path to a raster layer.}

\item{OUTDEPTH}{\code{rasterDestination} - Lake. Path for new raster layer.}

\item{OUTLEVEL}{\code{rasterDestination} - Surface. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Lake flood (saga:lakeflood)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTDEPTH - outputRaster - Lake
\item OUTLEVEL - outputRaster - Surface
}
}
}
