% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_landsurfacetemperature.R
\name{saga_landsurfacetemperature}
\alias{saga_landsurfacetemperature}
\title{QGIS algorithm Land surface temperature}
\usage{
saga_landsurfacetemperature(
  DEM = qgisprocess:::qgis_default_value(),
  SWR = qgisprocess:::qgis_default_value(),
  LAI = qgisprocess:::qgis_default_value(),
  Z_REFERENCE = qgisprocess:::qgis_default_value(),
  T_REFERENCE = qgisprocess:::qgis_default_value(),
  T_GRADIENT = qgisprocess:::qgis_default_value(),
  C_FACTOR = qgisprocess:::qgis_default_value(),
  LST = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation (m). Path to a raster layer.}

\item{SWR}{\code{raster} - Short Wave Radiation (kW/m2). Path to a raster layer.}

\item{LAI}{\code{raster} - Leaf Area Index. Path to a raster layer.}

\item{Z_REFERENCE}{\code{number} - Elevation at Reference Station (m). A numeric value.}

\item{T_REFERENCE}{\code{number} - Temperature at Reference Station (Deg.Celsius). A numeric value.}

\item{T_GRADIENT}{\code{number} - Temperature Gradient (Deg.Celsius/km). A numeric value.}

\item{C_FACTOR}{\code{number} - C Factor. A numeric value.}

\item{LST}{\code{rasterDestination} - Land Surface Temperature (Deg.Celsius). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Land surface temperature (saga:landsurfacetemperature)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LST - outputRaster - Land Surface Temperature (Deg.Celsius)
}
}
}
