% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_laplacianfilter.R
\name{saga_laplacianfilter}
\alias{saga_laplacianfilter}
\title{QGIS algorithm Laplacian filter}
\usage{
saga_laplacianfilter(
  INPUT = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  SIGMA = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  MODE = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] standard kernel 1", "[1] standard kernel 2", "[2] Standard kernel 3", "[3] user defined kernel")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SIGMA}{\code{number} - Standard Deviation (Percent of Radius). A numeric value.}

\item{RADIUS}{\code{number} - Radius. A numeric value.}

\item{MODE}{\code{enum}  of \verb{("[0] square", "[1] circle")} - Search Mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESULT}{\code{rasterDestination} - Filtered Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Laplacian filter (saga:laplacianfilter)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Filtered Grid
}
}
}
