% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_leastcostpaths.R
\name{saga_leastcostpaths}
\alias{saga_leastcostpaths}
\title{QGIS algorithm Least cost paths}
\usage{
saga_leastcostpaths(
  SOURCE = qgisprocess::qgis_default_value(),
  DEM = qgisprocess::qgis_default_value(),
  VALUES = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  LINE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SOURCE}{\code{source} - Source Point(s). Path to a vector layer.}

\item{DEM}{\code{raster} - Accumulated cost. Path to a raster layer.}

\item{VALUES}{\code{multilayer} - Values. .}

\item{POINTS}{\code{vectorDestination} - Profile (points). Path for new vector layer.}

\item{LINE}{\code{vectorDestination} - Profile (lines). Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Least cost paths (saga:leastcostpaths)
}
\details{
\subsection{Outputs description}{
\itemize{
\item POINTS - outputVector - Profile
\item LINE - outputVector - Profile
}
}
}
