% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_linepolygonintersection.R
\name{saga_linepolygonintersection}
\alias{saga_linepolygonintersection}
\title{QGIS algorithm Line-polygon intersection}
\usage{
saga_linepolygonintersection(
  LINES = qgisprocess:::qgis_default_value(),
  POLYGONS = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  INTERSECT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] one multi-line per polygon", "[1] keep original line attributes")} - Output. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{INTERSECT}{\code{vectorDestination} - Intersection. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Line-polygon intersection (saga:linepolygonintersection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INTERSECT - outputVector - Intersection
}
}
}
