% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_linesmoothing.R
\name{saga_linesmoothing}
\alias{saga_linesmoothing}
\title{QGIS algorithm Line smoothing}
\usage{
saga_linesmoothing(
  LINES_IN = qgisprocess::qgis_default_value(),
  LINES_OUT = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  SENSITIVITY = qgisprocess::qgis_default_value(),
  ITERATIONS = qgisprocess::qgis_default_value(),
  PRESERVATION = qgisprocess::qgis_default_value(),
  SIGMA = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LINES_IN}{\code{source} - Lines. Path to a vector layer.}

\item{LINES_OUT}{\code{vectorDestination} - Smoothed Lines. Path for new vector layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] basic SIA model", "[1] improved SIA model", "[2] Gaussian Filtering")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SENSITIVITY}{\code{number} - Sensitivity. A numeric value.}

\item{ITERATIONS}{\code{number} - Iterations. A numeric value.}

\item{PRESERVATION}{\code{number} - Preservation. A numeric value.}

\item{SIGMA}{\code{number} - Sigma. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Line smoothing (saga:linesmoothing)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LINES_OUT - outputVector - Smoothed Lines
}
}
}
