% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_lsfactor.R
\name{saga_lsfactor}
\alias{saga_lsfactor}
\title{QGIS algorithm Ls factor}
\usage{
saga_lsfactor(
  SLOPE = qgisprocess::qgis_default_value(),
  AREA = qgisprocess::qgis_default_value(),
  CONV = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  EROSIVITY = qgisprocess::qgis_default_value(),
  STABILITY = qgisprocess::qgis_default_value(),
  LS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SLOPE}{\code{raster} - Slope. Path to a raster layer.}

\item{AREA}{\code{raster} - Catchment Area. Path to a raster layer.}

\item{CONV}{\code{enum}  of \verb{("[0] no conversion (areas already given as specific catchment area)", "[1] 1 / cell size (specific catchment area)", "[2] square root (catchment length)")} - Area to Length Conversion. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("[0] Moore et al. 1991", "[1] Desmet & Govers 1996", "[2] Boehner & Selige 2006")} - Method (LS). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EROSIVITY}{\code{number} - Rill/Interrill Erosivity. A numeric value.}

\item{STABILITY}{\code{enum}  of \verb{("[0] stable", "[1] instable (thawing)")} - Stability. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LS}{\code{rasterDestination} - LS Factor. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Ls factor (saga:lsfactor)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LS - outputRaster - LS Factor
}
}
}
