% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_lsfactorfieldbased.R
\name{saga_lsfactorfieldbased}
\alias{saga_lsfactorfieldbased}
\title{QGIS algorithm Ls-factor, field based}
\usage{
saga_lsfactorfieldbased(
  DEM = qgisprocess::qgis_default_value(),
  FIELDS = qgisprocess::qgis_default_value(),
  STATISTICS = qgisprocess::qgis_default_value(),
  UPSLOPE_AREA = qgisprocess::qgis_default_value(),
  UPSLOPE_LENGTH = qgisprocess::qgis_default_value(),
  UPSLOPE_SLOPE = qgisprocess::qgis_default_value(),
  LS_FACTOR = qgisprocess::qgis_default_value(),
  BALANCE = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  METHOD_SLOPE = qgisprocess::qgis_default_value(),
  METHOD_AREA = qgisprocess::qgis_default_value(),
  STOP_AT_EDGE = qgisprocess::qgis_default_value(),
  EROSIVITY = qgisprocess::qgis_default_value(),
  STABILITY = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{FIELDS}{\code{source} - Fields. Path to a vector layer.}

\item{STATISTICS}{\code{vectorDestination} - Field Statistics. Path for new vector layer.}

\item{UPSLOPE_AREA}{\code{rasterDestination} - Upslope Length Factor. Path for new raster layer.}

\item{UPSLOPE_LENGTH}{\code{rasterDestination} - Effective Flow Length. Path for new raster layer.}

\item{UPSLOPE_SLOPE}{\code{rasterDestination} - Upslope Slope. Path for new raster layer.}

\item{LS_FACTOR}{\code{rasterDestination} - LS Factor. Path for new raster layer.}

\item{BALANCE}{\code{rasterDestination} - Sediment Balance. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Moore & Nieber 1989", "[1] Desmet & Govers 1996", "[2] Wischmeier & Smith 1978")} - LS Calculation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD_SLOPE}{\code{enum}  of \verb{("[0] local slope", "[1] distance weighted average catchment slope")} - Type of Slope. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD_AREA}{\code{enum}  of \verb{("[0] specific catchment area (contour length simply as cell size)", "[1] specific catchment area (contour length dependent on aspect)", "[2] catchment length (square root of catchment area)", "[3] effective flow length")} - Specific Catchment Area. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{STOP_AT_EDGE}{\code{boolean} - Stop at Edge. 1 for true/yes. 0 for false/no.}

\item{EROSIVITY}{\code{number} - Rill/Interrill Erosivity. A numeric value.}

\item{STABILITY}{\code{enum}  of \verb{("[0] stable", "[1] instable (thawing)")} - Stability. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Ls-factor, field based (saga:lsfactorfieldbased)
}
\details{
\subsection{Outputs description}{
\itemize{
\item BALANCE - outputRaster - Sediment Balance
\item LS_FACTOR - outputRaster - LS Factor
\item STATISTICS - outputVector - Field Statistics
\item UPSLOPE_AREA - outputRaster - Upslope Length Factor
\item UPSLOPE_LENGTH - outputRaster - Effective Flow Length
\item UPSLOPE_SLOPE - outputRaster - Upslope Slope
}
}
}
