% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_massbalanceindex.R
\name{saga_massbalanceindex}
\alias{saga_massbalanceindex}
\title{QGIS algorithm Mass balance index}
\usage{
saga_massbalanceindex(
  DEM = qgisprocess::qgis_default_value(),
  HREL = qgisprocess::qgis_default_value(),
  TSLOPE = qgisprocess::qgis_default_value(),
  TCURVE = qgisprocess::qgis_default_value(),
  THREL = qgisprocess::qgis_default_value(),
  MBI = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{HREL}{\code{raster} - Vertical Distance to Channel Network. Path to a raster layer.}

\item{TSLOPE}{\code{number} - T Slope. A numeric value.}

\item{TCURVE}{\code{number} - T Curvature. A numeric value.}

\item{THREL}{\code{number} - T Vertical Distance to Channel Network. A numeric value.}

\item{MBI}{\code{rasterDestination} - Mass Balance Index. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Mass balance index (saga:massbalanceindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item MBI - outputRaster - Mass Balance Index
}
}
}
