% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_maximumentropypresenceprediction.R
\name{saga_maximumentropypresenceprediction}
\alias{saga_maximumentropypresenceprediction}
\title{QGIS algorithm Maximum entropy presence prediction}
\usage{
saga_maximumentropypresenceprediction(
  PRESENCE = qgisprocess::qgis_default_value(),
  FEATURES_NUM = qgisprocess::qgis_default_value(),
  FEATURES_CAT = qgisprocess::qgis_default_value(),
  PREDICTION = qgisprocess::qgis_default_value(),
  PROBABILITY = qgisprocess::qgis_default_value(),
  BACKGROUND = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  YT_FILE_LOAD = qgisprocess::qgis_default_value(),
  YT_FILE_SAVE = qgisprocess::qgis_default_value(),
  YT_REGUL = qgisprocess::qgis_default_value(),
  YT_REGUL_VAL = qgisprocess::qgis_default_value(),
  YT_NUMASREAL = qgisprocess::qgis_default_value(),
  DL_ALPHA = qgisprocess::qgis_default_value(),
  DL_THRESHOLD = qgisprocess::qgis_default_value(),
  DL_ITERATIONS = qgisprocess::qgis_default_value(),
  NUM_CLASSES = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{PRESENCE}{\code{source} - Presence Data. Path to a vector layer.}

\item{FEATURES_NUM}{\code{multilayer} - Numerical Features. .}

\item{FEATURES_CAT}{\code{multilayer} - Categorical Features. .}

\item{PREDICTION}{\code{rasterDestination} - Presence Prediction. Path for new raster layer.}

\item{PROBABILITY}{\code{rasterDestination} - Presence Probability. Path for new raster layer.}

\item{BACKGROUND}{\code{number} - Background Sample Density (Percent). A numeric value.}

\item{METHOD}{\code{enum}  of \verb{("[0] Yoshimasa Tsuruoka", "[1] Dekang Lin")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{YT_FILE_LOAD}{\code{file} - Load from File.... Path to a file.}

\item{YT_FILE_SAVE}{\code{file} - Save to File.... Path to a file.}

\item{YT_REGUL}{\code{enum}  of \verb{("[0] none", "[1] L1", "[2] L2")} - Regularization. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{YT_REGUL_VAL}{\code{number} - Regularization Factor. A numeric value.}

\item{YT_NUMASREAL}{\code{boolean} - Real-valued Numerical Features. 1 for true/yes. 0 for false/no.}

\item{DL_ALPHA}{\code{number} - Alpha. A numeric value.}

\item{DL_THRESHOLD}{\code{number} - Threshold. A numeric value.}

\item{DL_ITERATIONS}{\code{number} - Maximum Iterations. A numeric value.}

\item{NUM_CLASSES}{\code{number} - Number of Numeric Value Classes. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Maximum entropy presence prediction (saga:maximumentropypresenceprediction)
}
\details{
\subsection{Outputs description}{
\itemize{
\item PREDICTION - outputRaster - Presence Prediction
\item PROBABILITY - outputRaster - Presence Probability
}
}
}
