% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_mergevectorlayers.R
\name{saga_mergevectorlayers}
\alias{saga_mergevectorlayers}
\title{QGIS algorithm Merge vector layers}
\usage{
saga_mergevectorlayers(
  INPUT = qgisprocess::qgis_default_value(),
  SRCINFO = qgisprocess::qgis_default_value(),
  MATCH = qgisprocess::qgis_default_value(),
  MERGED = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input Layers. .}

\item{SRCINFO}{\code{boolean} - Add source information. 1 for true/yes. 0 for false/no.}

\item{MATCH}{\code{boolean} - Match Fields by Name. 1 for true/yes. 0 for false/no.}

\item{MERGED}{\code{vectorDestination} - Merged Layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Merge vector layers (saga:mergevectorlayers)
}
\details{
\subsection{Outputs description}{
\itemize{
\item MERGED - outputVector - Merged Layer
}
}
}
