% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_meshdenoise.R
\name{saga_meshdenoise}
\alias{saga_meshdenoise}
\title{QGIS algorithm Mesh denoise}
\usage{
saga_meshdenoise(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  SIGMA = qgisprocess::qgis_default_value(),
  ITER = qgisprocess::qgis_default_value(),
  VITER = qgisprocess::qgis_default_value(),
  NB_CV = qgisprocess::qgis_default_value(),
  ZONLY = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Denoised Grid. Path for new raster layer.}

\item{SIGMA}{\code{number} - Threshold. A numeric value.}

\item{ITER}{\code{number} - Number of Iterations for Normal Updating. A numeric value.}

\item{VITER}{\code{number} - Number of Iterations for Vertex Updating. A numeric value.}

\item{NB_CV}{\code{enum}  of \verb{("[0] Common Vertex", "[1] Common Edge")} - Common Edge Type of Face Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ZONLY}{\code{boolean} - Only Z-Direction Position is Updated. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Mesh denoise (saga:meshdenoise)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Denoised Grid
}
}
}
