% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_metricconversions.R
\name{saga_metricconversions}
\alias{saga_metricconversions}
\title{QGIS algorithm Metric conversions}
\usage{
saga_metricconversions(
  GRID = qgisprocess:::qgis_default_value(),
  CONVERSION = qgisprocess:::qgis_default_value(),
  CONV = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{CONVERSION}{\code{enum}  of \verb{("[0] radians to degree", "[1] degree to radians", "[2] Celsius to Fahrenheit", "[3] Fahrenheit to Celsius")} - Conversion. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONV}{\code{rasterDestination} - Converted Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Metric conversions (saga:metricconversions)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONV - outputRaster - Converted Grid
}
}
}
