% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_mirrorgrid.R
\name{saga_mirrorgrid}
\alias{saga_mirrorgrid}
\title{QGIS algorithm Mirror grid}
\usage{
saga_mirrorgrid(
  GRID = qgisprocess::qgis_default_value(),
  MIRROR = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{MIRROR}{\code{rasterDestination} - Mirror Grid. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] horizontally", "[1] vertically", "[2] both")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Mirror grid (saga:mirrorgrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item MIRROR - outputRaster - Mirror Grid
}
}
}
