% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_morphologicalfilteropencv.R
\name{saga_morphologicalfilteropencv}
\alias{saga_morphologicalfilteropencv}
\title{QGIS algorithm Morphological filter (opencv)}
\usage{
saga_morphologicalfilteropencv(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  SHAPE = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  ITERATIONS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{TYPE}{\code{enum}  of \verb{("[0] dilation", "[1] erosion", "[2] opening", "[3] closing", "[4] morpological gradient", "[5] top hat", "[6] black hat")} - Operation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SHAPE}{\code{enum}  of \verb{("[0] ellipse", "[1] rectangle", "[2] cross")} - Element Shape. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radius (cells). A numeric value.}

\item{ITERATIONS}{\code{number} - Iterations. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Morphological filter (opencv) (saga:morphologicalfilteropencv)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
