% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_morphologicalfiltervigra.R
\name{saga_morphologicalfiltervigra}
\alias{saga_morphologicalfiltervigra}
\title{QGIS algorithm Morphological filter (vigra)}
\usage{
saga_morphologicalfiltervigra(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  RANK = qgisprocess:::qgis_default_value(),
  RESCALE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{TYPE}{\code{enum}  of \verb{("[0] Dilation", "[1] Erosion", "[2] Median", "[3] User defined rank")} - Operation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radius (cells). A numeric value.}

\item{RANK}{\code{number} - User defined rank. A numeric value.}

\item{RESCALE}{\code{boolean} - Rescale Values (0-255). 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Morphological filter (vigra) (saga:morphologicalfiltervigra)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
