% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_morphometricfeatures.R
\name{saga_morphometricfeatures}
\alias{saga_morphometricfeatures}
\title{QGIS algorithm Morphometric features}
\usage{
saga_morphometricfeatures(
  DEM = qgisprocess::qgis_default_value(),
  FEATURES = qgisprocess::qgis_default_value(),
  ELEVATION = qgisprocess::qgis_default_value(),
  SLOPE = qgisprocess::qgis_default_value(),
  ASPECT = qgisprocess::qgis_default_value(),
  PROFC = qgisprocess::qgis_default_value(),
  PLANC = qgisprocess::qgis_default_value(),
  LONGC = qgisprocess::qgis_default_value(),
  CROSC = qgisprocess::qgis_default_value(),
  MAXIC = qgisprocess::qgis_default_value(),
  MINIC = qgisprocess::qgis_default_value(),
  SIZE = qgisprocess::qgis_default_value(),
  TOL_SLOPE = qgisprocess::qgis_default_value(),
  TOL_CURVE = qgisprocess::qgis_default_value(),
  EXPONENT = qgisprocess::qgis_default_value(),
  ZSCALE = qgisprocess::qgis_default_value(),
  CONSTRAIN = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{FEATURES}{\code{rasterDestination} - Morphometric Features. Path for new raster layer.}

\item{ELEVATION}{\code{rasterDestination} - Generalized Surface. Path for new raster layer.}

\item{SLOPE}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{ASPECT}{\code{rasterDestination} - Aspect. Path for new raster layer.}

\item{PROFC}{\code{rasterDestination} - Profile Curvature. Path for new raster layer.}

\item{PLANC}{\code{rasterDestination} - Plan Curvature. Path for new raster layer.}

\item{LONGC}{\code{rasterDestination} - Longitudinal Curvature. Path for new raster layer.}

\item{CROSC}{\code{rasterDestination} - Cross-Sectional Curvature. Path for new raster layer.}

\item{MAXIC}{\code{rasterDestination} - Maximum Curvature. Path for new raster layer.}

\item{MINIC}{\code{rasterDestination} - Minimum Curvature. Path for new raster layer.}

\item{SIZE}{\code{number} - Scale Radius (Cells). A numeric value.}

\item{TOL_SLOPE}{\code{number} - Slope Tolerance. A numeric value.}

\item{TOL_CURVE}{\code{number} - Curvature Tolerance. A numeric value.}

\item{EXPONENT}{\code{number} - Distance Weighting Exponent. A numeric value.}

\item{ZSCALE}{\code{number} - Vertical Scaling. A numeric value.}

\item{CONSTRAIN}{\code{boolean} - Constrain. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Morphometric features (saga:morphometricfeatures)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ASPECT - outputRaster - Aspect
\item CROSC - outputRaster - Cross-Sectional Curvature
\item ELEVATION - outputRaster - Generalized Surface
\item FEATURES - outputRaster - Morphometric Features
\item LONGC - outputRaster - Longitudinal Curvature
\item MAXIC - outputRaster - Maximum Curvature
\item MINIC - outputRaster - Minimum Curvature
\item PLANC - outputRaster - Plan Curvature
\item PROFC - outputRaster - Profile Curvature
\item SLOPE - outputRaster - Slope
}
}
}
