% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_mosaicrasterlayers.R
\name{saga_mosaicrasterlayers}
\alias{saga_mosaicrasterlayers}
\title{QGIS algorithm Mosaic raster layers}
\usage{
saga_mosaicrasterlayers(
  GRIDS = qgisprocess::qgis_default_value(),
  NAME = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  RESAMPLING = qgisprocess::qgis_default_value(),
  OVERLAP = qgisprocess::qgis_default_value(),
  BLEND_DIST = qgisprocess::qgis_default_value(),
  MATCH = qgisprocess::qgis_default_value(),
 
    TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX = qgisprocess::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess::qgis_default_value(),
  TARGET_USER_FITS = qgisprocess::qgis_default_value(),
  TARGET_OUT_GRID = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Input Grids. .}

\item{NAME}{\code{string} - Name. String value.}

\item{TYPE}{\code{enum}  of \verb{("[0] 1 bit", "[1] 1 byte unsigned integer", "[2] 1 byte signed integer", "[3] 2 byte unsigned integer", "[4] 2 byte signed integer", "[5] 4 byte unsigned integer", "[6] 4 byte signed integer", "[7] 4 byte floating point", "[8] 8 byte floating point")} - Preferred data storage type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESAMPLING}{\code{enum}  of \verb{("[0] Nearest Neighbor", "[1] Bilinear Interpolation", "[2] Inverse Distance Interpolation", "[3] Bicubic Spline Interpolation", "[4] B-Spline Interpolation")} - Interpolation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OVERLAP}{\code{enum}  of \verb{("[0] first", "[1] last", "[2] minimum", "[3] maximum", "[4] mean", "[5] blend boundary", "[6] feathering")} - Overlapping Areas. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{BLEND_DIST}{\code{number} - Blending Distance. A numeric value.}

\item{MATCH}{\code{enum}  of \verb{("[0] none", "[1] regression")} - Match. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: {name}.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{TARGET_USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_OUT_GRID}{\code{rasterDestination} - Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Mosaic raster layers (saga:mosaicrasterlayers)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TARGET_OUT_GRID - outputRaster - Grid
}
}
}
