% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_multibandvariation.R
\name{saga_multibandvariation}
\alias{saga_multibandvariation}
\title{QGIS algorithm Multi-band variation}
\usage{
saga_multibandvariation(
  BANDS = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  DISTANCE_WEIGHTING_WEIGHTING = qgisprocess:::qgis_default_value(),
  DISTANCE_WEIGHTING_IDW_POWER = qgisprocess:::qgis_default_value(),
  DISTANCE_WEIGHTING_BANDWIDTH = qgisprocess:::qgis_default_value(),
  MEAN = qgisprocess:::qgis_default_value(),
  STDDEV = qgisprocess:::qgis_default_value(),
  DIFF = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{BANDS}{\code{multilayer} - Grids. .}

\item{RADIUS}{\code{number} - Radius (Cells). A numeric value.}

\item{DISTANCE_WEIGHTING_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Distance Weighting. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DISTANCE_WEIGHTING_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DISTANCE_WEIGHTING_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{MEAN}{\code{rasterDestination} - Mean Distance. Path for new raster layer.}

\item{STDDEV}{\code{rasterDestination} - Standard Deviation. Path for new raster layer.}

\item{DIFF}{\code{rasterDestination} - Distance. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Multi-band variation (saga:multibandvariation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIFF - outputRaster - Distance
\item MEAN - outputRaster - Mean Distance
\item STDDEV - outputRaster - Standard Deviation
}
}
}
