% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_multidirectionleefilter.R
\name{saga_multidirectionleefilter}
\alias{saga_multidirectionleefilter}
\title{QGIS algorithm Multi direction lee filter}
\usage{
saga_multidirectionleefilter(
  INPUT = qgisprocess::qgis_default_value(),
  NOISE_ABS = qgisprocess::qgis_default_value(),
  NOISE_REL = qgisprocess::qgis_default_value(),
  WEIGHTED = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  STDDEV = qgisprocess::qgis_default_value(),
  DIR = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{NOISE_ABS}{\code{number} - Estimated Noise (absolute). A numeric value.}

\item{NOISE_REL}{\code{number} - Estimated Noise (relative). A numeric value.}

\item{WEIGHTED}{\code{boolean} - Weighted. 1 for true/yes. 0 for false/no.}

\item{METHOD}{\code{enum}  of \verb{("[0] noise variance given as absolute value", "[1] noise variance given relative to mean standard deviation", "[2] original calculation (Ringeler)")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESULT}{\code{rasterDestination} - Filtered Grid. Path for new raster layer.}

\item{STDDEV}{\code{rasterDestination} - Minimum Standard Deviation. Path for new raster layer.}

\item{DIR}{\code{rasterDestination} - Direction of Minimum Standard Deviation. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Multi direction lee filter (saga:multidirectionleefilter)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Filtered Grid
\item STDDEV - outputRaster - Minimum Standard Deviation
\item DIR - outputRaster - Direction of Minimum Standard Deviation
}
}
}
