% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/saga_multipleregressionanalysisgridandpredictorgrids.R
\name{saga_multipleregressionanalysisgridandpredictorgrids}
\alias{saga_multipleregressionanalysisgridandpredictorgrids}
\title{QGIS algorithm Multiple regression analysis (grid and predictor grids)}
\usage{
saga_multipleregressionanalysisgridandpredictorgrids(
  DEPENDENT = qgisprocess::qgis_default_value(),
  PREDICTORS = qgisprocess::qgis_default_value(),
  REGRESSION = qgisprocess::qgis_default_value(),
  RESIDUALS = qgisprocess::qgis_default_value(),
  INFO_COEFF = qgisprocess::qgis_default_value(),
  INFO_MODEL = qgisprocess::qgis_default_value(),
  INFO_STEPS = qgisprocess::qgis_default_value(),
  RESAMPLING = qgisprocess::qgis_default_value(),
  COORD_X = qgisprocess::qgis_default_value(),
  COORD_Y = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  P_VALUE = qgisprocess::qgis_default_value(),
  CROSSVAL = qgisprocess::qgis_default_value(),
  CROSSVAL_K = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEPENDENT}{\code{raster} - Dependent Variable. Path to a raster layer.}

\item{PREDICTORS}{\code{multilayer} - Predictors. .}

\item{REGRESSION}{\code{rasterDestination} - Regression. Path for new raster layer.}

\item{RESIDUALS}{\code{rasterDestination} - Residuals. Path for new raster layer.}

\item{INFO_COEFF}{\code{source} - Details: Coefficients. Path to a vector layer.}

\item{INFO_MODEL}{\code{source} - Details: Model. Path to a vector layer.}

\item{INFO_STEPS}{\code{source} - Details: Steps. Path to a vector layer.}

\item{RESAMPLING}{\code{enum}  of \verb{("[0] Nearest Neighbour", "[1] Bilinear Interpolation", "[2] Bicubic Spline Interpolation", "[3] B-Spline Interpolation")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{COORD_X}{\code{boolean} - Include X Coordinate. 1 for true/yes. 0 for false/no.}

\item{COORD_Y}{\code{boolean} - Include Y Coordinate. 1 for true/yes. 0 for false/no.}

\item{METHOD}{\code{enum}  of \verb{("[0] include all", "[1] forward", "[2] backward", "[3] stepwise")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{P_VALUE}{\code{number} - Significance Level. A numeric value.}

\item{CROSSVAL}{\code{enum}  of \verb{("[0] none", "[1] leave one out", "[2] 2-fold", "[3] k-fold")} - Cross Validation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CROSSVAL_K}{\code{number} - Cross Validation Subsamples. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Multiple regression analysis (grid and predictor grids) (saga:multipleregressionanalysisgridandpredictorgrids)
}
\details{
\subsection{Outputs description}{
\itemize{
\item REGRESSION - outputRaster - Regression
\item RESIDUALS - outputRaster - Residuals
}
}
}
