% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/saga_multipleregressionanalysispointsandpredictorgrids.R
\name{saga_multipleregressionanalysispointsandpredictorgrids}
\alias{saga_multipleregressionanalysispointsandpredictorgrids}
\title{QGIS algorithm Multiple regression analysis (points and predictor grids)}
\usage{
saga_multipleregressionanalysispointsandpredictorgrids(
  PREDICTORS = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  ATTRIBUTE = qgisprocess::qgis_default_value(),
  INFO_COEFF = qgisprocess::qgis_default_value(),
  INFO_MODEL = qgisprocess::qgis_default_value(),
  INFO_STEPS = qgisprocess::qgis_default_value(),
  RESIDUALS = qgisprocess::qgis_default_value(),
  REGRESSION = qgisprocess::qgis_default_value(),
  REGRESCORR = qgisprocess::qgis_default_value(),
  RESAMPLING = qgisprocess::qgis_default_value(),
  COORD_X = qgisprocess::qgis_default_value(),
  COORD_Y = qgisprocess::qgis_default_value(),
  INTERCEPT = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  P_VALUE = qgisprocess::qgis_default_value(),
  CROSSVAL = qgisprocess::qgis_default_value(),
  CROSSVAL_K = qgisprocess::qgis_default_value(),
  RESIDUAL_COR = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{PREDICTORS}{\code{multilayer} - Predictors. .}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{ATTRIBUTE}{\code{field} - Dependent Variable. The name of an existing field. ; delimited list of existing field names.}

\item{INFO_COEFF}{\code{vectorDestination} - Details: Coefficients. Path for new vector layer.}

\item{INFO_MODEL}{\code{vectorDestination} - Details: Model. Path for new vector layer.}

\item{INFO_STEPS}{\code{vectorDestination} - Details: Steps. Path for new vector layer.}

\item{RESIDUALS}{\code{vectorDestination} - Residuals. Path for new vector layer.}

\item{REGRESSION}{\code{rasterDestination} - Regression. Path for new raster layer.}

\item{REGRESCORR}{\code{rasterDestination} - Regression with Residual Correction. Path for new raster layer.}

\item{RESAMPLING}{\code{enum}  of \verb{("[0] Nearest Neighbour", "[1] Bilinear Interpolation", "[2] Bicubic Spline Interpolation", "[3] B-Spline Interpolation")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{COORD_X}{\code{boolean} - Include X Coordinate. 1 for true/yes. 0 for false/no.}

\item{COORD_Y}{\code{boolean} - Include Y Coordinate. 1 for true/yes. 0 for false/no.}

\item{INTERCEPT}{\code{boolean} - Intercept. 1 for true/yes. 0 for false/no.}

\item{METHOD}{\code{enum}  of \verb{("[0] include all", "[1] forward", "[2] backward", "[3] stepwise")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{P_VALUE}{\code{number} - Significance Level. A numeric value.}

\item{CROSSVAL}{\code{enum}  of \verb{("[0] none", "[1] leave one out", "[2] 2-fold", "[3] k-fold")} - Cross Validation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CROSSVAL_K}{\code{number} - Cross Validation Subsamples. A numeric value.}

\item{RESIDUAL_COR}{\code{enum}  of \verb{("[0] Multleve B-Spline Interpolation", "[1] Inverse Distance Weighted")} - Residual Interpolation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Multiple regression analysis (points and predictor grids) (saga:multipleregressionanalysispointsandpredictorgrids)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INFO_COEFF - outputVector - Details
\item INFO_MODEL - outputVector - Details
\item INFO_STEPS - outputVector - Details
\item RESIDUALS - outputVector - Residuals
\item REGRESSION - outputRaster - Regression
\item REGRESCORR - outputRaster - Regression with Residual Correction
}
}
}
