% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_multipleregressionanalysisrasterraster.R
\name{saga_multipleregressionanalysisrasterraster}
\alias{saga_multipleregressionanalysisrasterraster}
\title{QGIS algorithm Multiple regression analysis (raster/raster)}
\usage{
saga_multipleregressionanalysisrasterraster(
  DEPENDENT = qgisprocess::qgis_default_value(),
  GRIDS = qgisprocess::qgis_default_value(),
  INTERPOL = qgisprocess::qgis_default_value(),
  COORD_X = qgisprocess::qgis_default_value(),
  COORD_Y = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  P_IN = qgisprocess::qgis_default_value(),
  P_OUT = qgisprocess::qgis_default_value(),
  REGRESSION = qgisprocess::qgis_default_value(),
  RESIDUALS = qgisprocess::qgis_default_value(),
  INFO_COEFF = qgisprocess::qgis_default_value(),
  INFO_MODEL = qgisprocess::qgis_default_value(),
  INFO_STEPS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEPENDENT}{\code{raster} - Dependent. Path to a raster layer.}

\item{GRIDS}{\code{multilayer} - Grids. .}

\item{INTERPOL}{\code{enum}  of \verb{("[0] Nearest Neighbor", "[1] Bilinear Interpolation", "[2] Inverse Distance Interpolation", "[3] Bicubic Spline Interpolation", "[4] B-Spline Interpolation")} - Grid Interpolation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{COORD_X}{\code{boolean} - Include X Coordinate. 1 for true/yes. 0 for false/no.}

\item{COORD_Y}{\code{boolean} - Include Y Coordinate. 1 for true/yes. 0 for false/no.}

\item{METHOD}{\code{enum}  of \verb{("[0] include all", "[1] forward", "[2] backward", "[3] stepwise")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{P_IN}{\code{number} - P in. A numeric value.}

\item{P_OUT}{\code{number} - P out. A numeric value.}

\item{REGRESSION}{\code{rasterDestination} - Regression. Path for new raster layer.}

\item{RESIDUALS}{\code{rasterDestination} - Residuals. Path for new raster layer.}

\item{INFO_COEFF}{\code{vectorDestination} - Details: Coefficients. Path for new vector layer.}

\item{INFO_MODEL}{\code{vectorDestination} - Details: Model. Path for new vector layer.}

\item{INFO_STEPS}{\code{vectorDestination} - Details: Steps. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Multiple regression analysis (raster/raster) (saga:multipleregressionanalysisrasterraster)
}
\details{
\subsection{Outputs description}{
\itemize{
\item REGRESSION - outputRaster - Regression
\item RESIDUALS - outputRaster - Residuals
\item INFO_COEFF - outputVector - Details
\item INFO_MODEL - outputVector - Details
\item INFO_STEPS - outputVector - Details
}
}
}
