% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_naturalneighbour.R
\name{saga_naturalneighbour}
\alias{saga_naturalneighbour}
\title{QGIS algorithm Natural neighbour}
\usage{
saga_naturalneighbour(
  SHAPES = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
 
    TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX = qgisprocess::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess::qgis_default_value(),
  TARGET_USER_FITS = qgisprocess::qgis_default_value(),
  TARGET_OUT_GRID = qgisprocess::qgis_default_value(),
  WEIGHT = qgisprocess::qgis_default_value(),
  TARGET_TEMPLATE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SHAPES}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{METHOD}{\code{enum}  of \verb{("[0] Linear", "[1] Sibson", "[2] Non-Sibson")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: {name}.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{TARGET_USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_OUT_GRID}{\code{rasterDestination} - Grid. Path for new raster layer.}

\item{WEIGHT}{\code{number} - Minimum Weight. A numeric value.}

\item{TARGET_TEMPLATE}{\code{raster} - Target System. Path to a raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Natural neighbour (saga:naturalneighbour)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TARGET_OUT_GRID - outputRaster - Grid
}
}
}
