% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_ordinarykriging.R
\name{saga_ordinarykriging}
\alias{saga_ordinarykriging}
\title{QGIS algorithm Ordinary kriging}
\usage{
saga_ordinarykriging(
  POINTS = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  TQUALITY = qgisprocess::qgis_default_value(),
  LOG = qgisprocess::qgis_default_value(),
  BLOCK = qgisprocess::qgis_default_value(),
  DBLOCK = qgisprocess::qgis_default_value(),
  VAR_MAXDIST = qgisprocess::qgis_default_value(),
  VAR_NCLASSES = qgisprocess::qgis_default_value(),
  VAR_NSKIP = qgisprocess::qgis_default_value(),
  VAR_MODEL = qgisprocess::qgis_default_value(),
 
    TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX = qgisprocess::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess::qgis_default_value(),
  TARGET_USER_FITS = qgisprocess::qgis_default_value(),
  PREDICTION = qgisprocess::qgis_default_value(),
  VARIANCE = qgisprocess::qgis_default_value(),
  SEARCH_RANGE = qgisprocess::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_ALL = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MIN = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MAX = qgisprocess::qgis_default_value(),
  SEARCH_DIRECTION = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{TQUALITY}{\code{enum}  of \verb{("[0] standard deviation", "[1] variance")} - Type of Quality Measure. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LOG}{\code{boolean} - Logarithmic Transformation. 1 for true/yes. 0 for false/no.}

\item{BLOCK}{\code{boolean} - Block Kriging. 1 for true/yes. 0 for false/no.}

\item{DBLOCK}{\code{number} - Block Size. A numeric value.}

\item{VAR_MAXDIST}{\code{number} - Maximum Distance. A numeric value.}

\item{VAR_NCLASSES}{\code{number} - Lag Distance Classes. A numeric value.}

\item{VAR_NSKIP}{\code{number} - Skip. A numeric value.}

\item{VAR_MODEL}{\code{string} - Variogram Model. String value.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{TARGET_USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PREDICTION}{\code{rasterDestination} - Prediction. Path for new raster layer.}

\item{VARIANCE}{\code{rasterDestination} - Quality Measure. Path for new raster layer.}

\item{SEARCH_RANGE}{\code{enum}  of \verb{("[0] local", "[1] global")} - Search Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_RADIUS}{\code{number} - Maximum Search Distance. A numeric value.}

\item{SEARCH_POINTS_ALL}{\code{enum}  of \verb{("[0] maximum number of nearest points", "[1] all points within search distance")} - Number of Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_POINTS_MIN}{\code{number} - Minimum. A numeric value.}

\item{SEARCH_POINTS_MAX}{\code{number} - Maximum. A numeric value.}

\item{SEARCH_DIRECTION}{\code{enum}  of \verb{("[0] all directions", "[1] quadrants")} - Search Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Ordinary kriging (saga:ordinarykriging)
}
\details{
\subsection{Outputs description}{
\itemize{
\item PREDICTION - outputRaster - Prediction
\item VARIANCE - outputRaster - Quality Measure
}
}
}
