% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_overlandflowkinematicwaved8.R
\name{saga_overlandflowkinematicwaved8}
\alias{saga_overlandflowkinematicwaved8}
\title{QGIS algorithm Overland flow - kinematic wave d8}
\usage{
saga_overlandflowkinematicwaved8(
  DEM = qgisprocess:::qgis_default_value(),
  GAUGES = qgisprocess:::qgis_default_value(),
  TIME_SPAN = qgisprocess:::qgis_default_value(),
  TIME_STEP = qgisprocess:::qgis_default_value(),
  ROUGHNESS = qgisprocess:::qgis_default_value(),
  NEWTON_MAXITER = qgisprocess:::qgis_default_value(),
  NEWTON_EPSILON = qgisprocess:::qgis_default_value(),
  PRECIP = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  FLOW = qgisprocess:::qgis_default_value(),
  GAUGES_FLOW = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{GAUGES}{\code{source} - Gauges. Path to a vector layer.}

\item{TIME_SPAN}{\code{number} - Simulation Time (h). A numeric value.}

\item{TIME_STEP}{\code{number} - Simulation Time Step (h). A numeric value.}

\item{ROUGHNESS}{\code{number} - Manning's Roughness. A numeric value.}

\item{NEWTON_MAXITER}{\code{number} - Max. Iterations. A numeric value.}

\item{NEWTON_EPSILON}{\code{number} - Epsilon. A numeric value.}

\item{PRECIP}{\code{enum}  of \verb{("[0] Homogeneous", "[1] Above Elevation", "[2] Left Half")} - Precipitation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{THRESHOLD}{\code{number} - Threshold Elevation. A numeric value.}

\item{FLOW}{\code{rasterDestination} - Runoff. Path for new raster layer.}

\item{GAUGES_FLOW}{\code{vectorDestination} - Flow at Gauges. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Overland flow - kinematic wave d8 (saga:overlandflowkinematicwaved8)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLOW - outputRaster - Runoff
\item GAUGES_FLOW - outputVector - Flow at Gauges
}
}
}
