% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_patching.R
\name{saga_patching}
\alias{saga_patching}
\title{QGIS algorithm Patching}
\usage{
saga_patching(
  ORIGINAL = qgisprocess::qgis_default_value(),
  ADDITIONAL = qgisprocess::qgis_default_value(),
  RESAMPLING = qgisprocess::qgis_default_value(),
  COMPLETED = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ORIGINAL}{\code{raster} - Grid. Path to a raster layer.}

\item{ADDITIONAL}{\code{raster} - Patch Grid. Path to a raster layer.}

\item{RESAMPLING}{\code{enum}  of \verb{("[0] Nearest Neighbor", "[1] Bilinear Interpolation", "[2] Inverse Distance Interpolation", "[3] Bicubic Spline Interpolation", "[4] B-Spline Interpolation")} - Interpolation Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{COMPLETED}{\code{rasterDestination} - Completed Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Patching (saga:patching)
}
\details{
\subsection{Outputs description}{
\itemize{
\item COMPLETED - outputRaster - Completed Grid
}
}
}
