% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_patternanalysis.R
\name{saga_patternanalysis}
\alias{saga_patternanalysis}
\title{QGIS algorithm Pattern analysis}
\usage{
saga_patternanalysis(
  INPUT = qgisprocess::qgis_default_value(),
  WINSIZE = qgisprocess::qgis_default_value(),
  MAXNUMCLASS = qgisprocess::qgis_default_value(),
  RELATIVE = qgisprocess::qgis_default_value(),
  DIVERSITY = qgisprocess::qgis_default_value(),
  DOMINANCE = qgisprocess::qgis_default_value(),
  FRAGMENTATION = qgisprocess::qgis_default_value(),
  NDC = qgisprocess::qgis_default_value(),
  CVN = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input Grid. Path to a raster layer.}

\item{WINSIZE}{\code{enum}  of \verb{("[0] 3 X 3", "[1] 5 X 5", "[2] 7 X 7")} - Size of Analysis Window. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MAXNUMCLASS}{\code{number} - Max. Number of Classes. A numeric value.}

\item{RELATIVE}{\code{rasterDestination} - Relative Richness. Path for new raster layer.}

\item{DIVERSITY}{\code{rasterDestination} - Diversity. Path for new raster layer.}

\item{DOMINANCE}{\code{rasterDestination} - Dominance. Path for new raster layer.}

\item{FRAGMENTATION}{\code{rasterDestination} - Fragmentation. Path for new raster layer.}

\item{NDC}{\code{rasterDestination} - Number of Different Classes. Path for new raster layer.}

\item{CVN}{\code{rasterDestination} - Center Versus Neighbours. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Pattern analysis (saga:patternanalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CVN - outputRaster - Center Versus Neighbours
\item DIVERSITY - outputRaster - Diversity
\item DOMINANCE - outputRaster - Dominance
\item FRAGMENTATION - outputRaster - Fragmentation
\item NDC - outputRaster - Number of Different Classes
\item RELATIVE - outputRaster - Relative Richness
}
}
}
