% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_petafterhargreavesgrid.R
\name{saga_petafterhargreavesgrid}
\alias{saga_petafterhargreavesgrid}
\title{QGIS algorithm Pet (after hargreaves, grid)}
\usage{
saga_petafterhargreavesgrid(
  T = qgisprocess::qgis_default_value(),
  T_MIN = qgisprocess::qgis_default_value(),
  T_MAX = qgisprocess::qgis_default_value(),
  PET = qgisprocess::qgis_default_value(),
  LAT = qgisprocess::qgis_default_value(),
  TIME = qgisprocess::qgis_default_value(),
  MONTH = qgisprocess::qgis_default_value(),
  DAY = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{T}{\code{raster} - Mean Temperature. Path to a raster layer.}

\item{T_MIN}{\code{raster} - Minimum Temperature. Path to a raster layer.}

\item{T_MAX}{\code{raster} - Maximum Temperature. Path to a raster layer.}

\item{PET}{\code{rasterDestination} - Potential Evapotranspiration. Path for new raster layer.}

\item{LAT}{\code{number} - Latitude (Degree). A numeric value.}

\item{TIME}{\code{enum}  of \verb{("[0] day", "[1] month")} - Time. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MONTH}{\code{enum}  of \verb{("[0] January", "[1] February", "[2] March", "[3] April", "[4] May", "[5] June", "[6] July", "[7] August", "[8] September", "[9] October", "[10] November", "[11] December")} - Month. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DAY}{\code{number} - Day of Month. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Pet (after hargreaves, grid) (saga:petafterhargreavesgrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item PET - outputRaster - Potential Evapotranspiration
}
}
}
