% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_pointdistances.R
\name{saga_pointdistances}
\alias{saga_pointdistances}
\title{QGIS algorithm Point distances}
\usage{
saga_pointdistances(
  POINTS = qgisprocess:::qgis_default_value(),
  ID_POINTS = qgisprocess:::qgis_default_value(),
  NEAR = qgisprocess:::qgis_default_value(),
  ID_NEAR = qgisprocess:::qgis_default_value(),
  FORMAT = qgisprocess:::qgis_default_value(),
  MAX_DIST = qgisprocess:::qgis_default_value(),
  DISTANCES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{ID_POINTS}{\code{field} - Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{NEAR}{\code{source} - Near Points. Path to a vector layer.}

\item{ID_NEAR}{\code{field} - Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{FORMAT}{\code{enum}  of \verb{("[0] complete input times near points matrix", "[1] each pair with a single record")} - Output Format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MAX_DIST}{\code{number} - Maximum Distance. A numeric value.}

\item{DISTANCES}{\code{vectorDestination} - Distances. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Point distances (saga:pointdistances)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DISTANCES - outputVector - Distances
}
}
}
