% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_pointsthinning.R
\name{saga_pointsthinning}
\alias{saga_pointsthinning}
\title{QGIS algorithm Points thinning}
\usage{
saga_pointsthinning(
  POINTS = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  RESOLUTION = qgisprocess::qgis_default_value(),
  THINNED = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{RESOLUTION}{\code{number} - Resolution. A numeric value.}

\item{THINNED}{\code{vectorDestination} - Thinned Points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Points thinning (saga:pointsthinning)
}
\details{
\subsection{Outputs description}{
\itemize{
\item THINNED - outputVector - Thinned Points
}
}
}
