% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_polartocartesiancoordinates.R
\name{saga_polartocartesiancoordinates}
\alias{saga_polartocartesiancoordinates}
\title{QGIS algorithm Polar to cartesian coordinates}
\usage{
saga_polartocartesiancoordinates(
  POLAR = qgisprocess::qgis_default_value(),
  F_EXAGG = qgisprocess::qgis_default_value(),
  D_EXAGG = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  DEGREE = qgisprocess::qgis_default_value(),
  CARTES = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POLAR}{\code{source} - Polar Coordinates. Path to a vector layer.}

\item{F_EXAGG}{\code{field} - Exaggeration. The name of an existing field. ; delimited list of existing field names.}

\item{D_EXAGG}{\code{number} - Exaggeration Factor. A numeric value.}

\item{RADIUS}{\code{number} - Radius. A numeric value.}

\item{DEGREE}{\code{boolean} - Degree. 1 for true/yes. 0 for false/no.}

\item{CARTES}{\code{vectorDestination} - Cartesian Coordinates. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Polar to cartesian coordinates (saga:polartocartesiancoordinates)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CARTES - outputVector - Cartesian Coordinates
}
}
}
