% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_polygoncentroids.R
\name{saga_polygoncentroids}
\alias{saga_polygoncentroids}
\title{QGIS algorithm Polygon centroids}
\usage{
saga_polygoncentroids(
  POLYGONS = qgisprocess::qgis_default_value(),
  CENTROIDS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{CENTROIDS}{\code{vectorDestination} - Centroids. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Polygon centroids (saga:polygoncentroids)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CENTROIDS - outputVector - Centroids
}
}
}
