% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_polygonclipping.R
\name{saga_polygonclipping}
\alias{saga_polygonclipping}
\title{QGIS algorithm Polygon clipping}
\usage{
saga_polygonclipping(
  CLIP = qgisprocess::qgis_default_value(),
  S_INPUT = qgisprocess::qgis_default_value(),
  S_OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{CLIP}{\code{source} - Clip features. Path to a vector layer.}

\item{S_INPUT}{\code{source} - Input features. Path to a vector layer.}

\item{S_OUTPUT}{\code{vectorDestination} - Output features. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Polygon clipping (saga:polygonclipping)
}
\details{
\subsection{Outputs description}{
\itemize{
\item S_OUTPUT - outputVector - Output features
}
}
}
