% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_polygonpartstoseparatepolygons.R
\name{saga_polygonpartstoseparatepolygons}
\alias{saga_polygonpartstoseparatepolygons}
\title{QGIS algorithm Polygon parts to separate polygons}
\usage{
saga_polygonpartstoseparatepolygons(
  POLYGONS = qgisprocess::qgis_default_value(),
  LAKES = qgisprocess::qgis_default_value(),
  PARTS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{LAKES}{\code{boolean} - Ignore Lakes. 1 for true/yes. 0 for false/no.}

\item{PARTS}{\code{vectorDestination} - Polygon Parts. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Polygon parts to separate polygons (saga:polygonpartstoseparatepolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item PARTS - outputVector - Polygon Parts
}
}
}
