% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_polygonproperties.R
\name{saga_polygonproperties}
\alias{saga_polygonproperties}
\title{QGIS algorithm Polygon properties}
\usage{
saga_polygonproperties(
  POLYGONS = qgisprocess::qgis_default_value(),
  BPARTS = qgisprocess::qgis_default_value(),
  BPOINTS = qgisprocess::qgis_default_value(),
  BLENGTH = qgisprocess::qgis_default_value(),
  BAREA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{BPARTS}{\code{boolean} - Number of Parts. 1 for true/yes. 0 for false/no.}

\item{BPOINTS}{\code{boolean} - Number of Vertices. 1 for true/yes. 0 for false/no.}

\item{BLENGTH}{\code{boolean} - Perimeter. 1 for true/yes. 0 for false/no.}

\item{BAREA}{\code{boolean} - Area. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{vectorDestination} - Polygons with Property Attributes. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Polygon properties (saga:polygonproperties)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Polygons with Property Attributes
}
}
}
