% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_polygonselfintersection.R
\name{saga_polygonselfintersection}
\alias{saga_polygonselfintersection}
\title{QGIS algorithm Polygon self-intersection}
\usage{
saga_polygonselfintersection(
  POLYGONS = qgisprocess::qgis_default_value(),
  ID = qgisprocess::qgis_default_value(),
  INTERSECT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{ID}{\code{field} - Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{INTERSECT}{\code{vectorDestination} - Intersection. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Polygon self-intersection (saga:polygonselfintersection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INTERSECT - outputVector - Intersection
}
}
}
