% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_polygonshapeindices.R
\name{saga_polygonshapeindices}
\alias{saga_polygonshapeindices}
\title{QGIS algorithm Polygon shape indices}
\usage{
saga_polygonshapeindices(
  SHAPES = qgisprocess::qgis_default_value(),
  INDEX = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{INDEX}{\code{vectorDestination} - Shape Index. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Polygon shape indices (saga:polygonshapeindices)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INDEX - outputVector - Shape Index
}
}
}
