% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_polynomialregression.R
\name{saga_polynomialregression}
\alias{saga_polynomialregression}
\title{QGIS algorithm Polynomial regression}
\usage{
saga_polynomialregression(
  POINTS = qgisprocess::qgis_default_value(),
  ATTRIBUTE = qgisprocess::qgis_default_value(),
  POLYNOM = qgisprocess::qgis_default_value(),
  XORDER = qgisprocess::qgis_default_value(),
  YORDER = qgisprocess::qgis_default_value(),
  TORDER = qgisprocess::qgis_default_value(),
 
    TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX = qgisprocess::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess::qgis_default_value(),
  TARGET_USER_FITS = qgisprocess::qgis_default_value(),
  TARGET_OUT_GRID = qgisprocess::qgis_default_value(),
  RESIDUALS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{ATTRIBUTE}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{POLYNOM}{\code{enum}  of \verb{("[0] simple planar surface", "[1] bi-linear saddle", "[2] quadratic surface", "[3] cubic surface", "[4] user defined")} - Polynom. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{XORDER}{\code{number} - Maximum X Order. A numeric value.}

\item{YORDER}{\code{number} - Maximum Y Order. A numeric value.}

\item{TORDER}{\code{number} - Maximum Total Order. A numeric value.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: {name}.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{TARGET_USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_OUT_GRID}{\code{rasterDestination} - Grid. Path for new raster layer.}

\item{RESIDUALS}{\code{vectorDestination} - Residuals. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Polynomial regression (saga:polynomialregression)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TARGET_OUT_GRID - outputRaster - Grid
\item RESIDUALS - outputVector - Residuals
}
}
}
