% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_principlecomponentsanalysis.R
\name{saga_principlecomponentsanalysis}
\alias{saga_principlecomponentsanalysis}
\title{QGIS algorithm Principle components analysis}
\usage{
saga_principlecomponentsanalysis(
  TABLE = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  NFIRST = qgisprocess::qgis_default_value(),
  PCA = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{TABLE}{\code{source} - Table. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] correlation matrix", "[1] variance-covariance matrix", "[2] sums-of-squares-and-cross-products matrix")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NFIRST}{\code{number} - Number of Components. A numeric value.}

\item{PCA}{\code{vectorDestination} - Principle Components. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Principle components analysis (saga:principlecomponentsanalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item PCA - outputVector - Principle Components
}
}
}
