% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_profilefrompointstable.R
\name{saga_profilefrompointstable}
\alias{saga_profilefrompointstable}
\title{QGIS algorithm Profile from points table}
\usage{
saga_profilefrompointstable(
  GRID = qgisprocess::qgis_default_value(),
  TABLE = qgisprocess::qgis_default_value(),
  X = qgisprocess::qgis_default_value(),
  Y = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{TABLE}{\code{source} - Input. Path to a vector layer.}

\item{X}{\code{field} - X. The name of an existing field. ; delimited list of existing field names.}

\item{Y}{\code{field} - Y. The name of an existing field. ; delimited list of existing field names.}

\item{RESULT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Profile from points table (saga:profilefrompointstable)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Result
}
}
}
