% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_proximityraster.R
\name{saga_proximityraster}
\alias{saga_proximityraster}
\title{QGIS algorithm Proximity raster}
\usage{
saga_proximityraster(
  FEATURES = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  DIRECTION = qgisprocess::qgis_default_value(),
  ALLOCATION = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{FEATURES}{\code{raster} - Features. Path to a raster layer.}

\item{DISTANCE}{\code{rasterDestination} - Distance. Path for new raster layer.}

\item{DIRECTION}{\code{rasterDestination} - Direction. Path for new raster layer.}

\item{ALLOCATION}{\code{rasterDestination} - Allocation. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Proximity raster (saga:proximityraster)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DISTANCE - outputRaster - Distance
\item DIRECTION - outputRaster - Direction
\item ALLOCATION - outputRaster - Allocation
}
}
}
