% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_quadtreestructuretopolygons.R
\name{saga_quadtreestructuretopolygons}
\alias{saga_quadtreestructuretopolygons}
\title{QGIS algorithm Quadtree structure to polygons}
\usage{
saga_quadtreestructuretopolygons(
  SHAPES = qgisprocess::qgis_default_value(),
  ATTRIBUTE = qgisprocess::qgis_default_value(),
  POLYGONS = qgisprocess::qgis_default_value(),
  LINES = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{ATTRIBUTE}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{POLYGONS}{\code{vectorDestination} - Polygons. Path for new vector layer.}

\item{LINES}{\code{vectorDestination} - Lines. Path for new vector layer.}

\item{POINTS}{\code{vectorDestination} - Duplicated Points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Quadtree structure to polygons (saga:quadtreestructuretopolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LINES - outputVector - Lines
\item POINTS - outputVector - Duplicated Points
\item POLYGONS - outputVector - Polygons
}
}
}
