% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_radiusofvarianceraster.R
\name{saga_radiusofvarianceraster}
\alias{saga_radiusofvarianceraster}
\title{QGIS algorithm Radius of variance (raster)}
\usage{
saga_radiusofvarianceraster(
  INPUT = qgisprocess::qgis_default_value(),
  VARIANCE = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{VARIANCE}{\code{number} - Standard Deviation. A numeric value.}

\item{RADIUS}{\code{number} - Maximum Search Radius (cells). A numeric value.}

\item{OUTPUT}{\code{enum}  of \verb{("[0] Cells", "[1] Map Units")} - Type of Output. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESULT}{\code{rasterDestination} - Variance Radius. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Radius of variance (raster) (saga:radiusofvarianceraster)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Variance Radius
}
}
}
