% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_randomfield.R
\name{saga_randomfield}
\alias{saga_randomfield}
\title{QGIS algorithm Random field}
\usage{
saga_randomfield(
  METHOD = qgisprocess::qgis_default_value(),
  RANGE_MIN = qgisprocess::qgis_default_value(),
  RANGE_MAX = qgisprocess::qgis_default_value(),
  MEAN = qgisprocess::qgis_default_value(),
  STDDEV = qgisprocess::qgis_default_value(),
  USER_XMIN_USER_XMAX_USER_YMIN_USER_YMAX = qgisprocess::qgis_default_value(),
  USER_SIZE = qgisprocess::qgis_default_value(),
  USER_FITS = qgisprocess::qgis_default_value(),
  OUT_GRID = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{METHOD}{\code{enum}  of \verb{("[0] Uniform", "[1] Gaussian")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RANGE_MIN}{\code{number} - Range Min. A numeric value.}

\item{RANGE_MAX}{\code{number} - Range Max. A numeric value.}

\item{MEAN}{\code{number} - Arithmetic Mean. A numeric value.}

\item{STDDEV}{\code{number} - Standard Deviation. A numeric value.}

\item{USER_XMIN_USER_XMAX_USER_YMIN_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: USER_XMIN USER_XMAX USER_YMIN USER_YMAX.}

\item{USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUT_GRID}{\code{rasterDestination} - Random Field. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Random field (saga:randomfield)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUT_GRID - outputRaster - Random Field
}
}
}
