% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_randomforestclassificationopencv.R
\name{saga_randomforestclassificationopencv}
\alias{saga_randomforestclassificationopencv}
\title{QGIS algorithm Random forest classification (opencv)}
\usage{
saga_randomforestclassificationopencv(
  FEATURES = qgisprocess::qgis_default_value(),
  NORMALIZE = qgisprocess::qgis_default_value(),
  TRAIN_AREAS = qgisprocess::qgis_default_value(),
  TRAIN_CLASS = qgisprocess::qgis_default_value(),
  CLASSES = qgisprocess::qgis_default_value(),
  MAX_DEPTH = qgisprocess::qgis_default_value(),
  MIN_SAMPLES = qgisprocess::qgis_default_value(),
  MAX_CATEGRS = qgisprocess::qgis_default_value(),
  arg_1SE_RULE = qgisprocess::qgis_default_value(),
  TRUNC_PRUNED = qgisprocess::qgis_default_value(),
  REG_ACCURACY = qgisprocess::qgis_default_value(),
  ACTIVE_VARS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{FEATURES}{\code{multilayer} - Features. .}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no.}

\item{TRAIN_AREAS}{\code{source} - Training Areas. Path to a vector layer.}

\item{TRAIN_CLASS}{\code{source} - Class Identifier. Path to a vector layer.}

\item{CLASSES}{\code{rasterDestination} - Classification. Path for new raster layer.}

\item{MAX_DEPTH}{\code{number} - Maximum Tree Depth. A numeric value.}

\item{MIN_SAMPLES}{\code{number} - Minimum Sample Count. A numeric value.}

\item{MAX_CATEGRS}{\code{number} - Maximum Categories. A numeric value.}

\item{arg_1SE_RULE}{\code{boolean} - Use 1SE Rule. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{TRUNC_PRUNED}{\code{boolean} - Truncate Pruned Trees. 1 for true/yes. 0 for false/no.}

\item{REG_ACCURACY}{\code{number} - Regression Accuracy. A numeric value.}

\item{ACTIVE_VARS}{\code{number} - Active Variable Count. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Random forest classification (opencv) (saga:randomforestclassificationopencv)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Classification
}
}
}
