% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_randomterraingeneration.R
\name{saga_randomterraingeneration}
\alias{saga_randomterraingeneration}
\title{QGIS algorithm Random terrain generation}
\usage{
saga_randomterraingeneration(
  RADIUS = qgisprocess::qgis_default_value(),
  ITERATIONS = qgisprocess::qgis_default_value(),
  TARGET_TYPE = qgisprocess::qgis_default_value(),
  USER_CELL_SIZE = qgisprocess::qgis_default_value(),
  USER_COLS = qgisprocess::qgis_default_value(),
  USER_ROWS = qgisprocess::qgis_default_value(),
  TARGET_GRID = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{RADIUS}{\code{number} - Radius (cells). A numeric value.}

\item{ITERATIONS}{\code{number} - Iterations. A numeric value.}

\item{TARGET_TYPE}{\code{enum}  of \code{("[0] User defined")} - Target Dimensions. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{USER_CELL_SIZE}{\code{number} - Grid Size. A numeric value.}

\item{USER_COLS}{\code{number} - Cols. A numeric value.}

\item{USER_ROWS}{\code{number} - Rows. A numeric value.}

\item{TARGET_GRID}{\code{rasterDestination} - Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Random terrain generation (saga:randomterraingeneration)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TARGET_GRID - outputRaster - Grid
}
}
}
