% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rasterbuffer.R
\name{saga_rasterbuffer}
\alias{saga_rasterbuffer}
\title{QGIS algorithm Raster buffer}
\usage{
saga_rasterbuffer(
  FEATURES = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  BUFFER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FEATURES}{\code{raster} - Features Grid. Path to a raster layer.}

\item{DISTANCE}{\code{number} - Distance. A numeric value.}

\item{TYPE}{\code{enum}  of \verb{("[0] fixed", "[1] cell's value")} - Buffer Distance. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{BUFFER}{\code{rasterDestination} - Buffer Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster buffer (saga:rasterbuffer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item BUFFER - outputRaster - Buffer Grid
}
}
}
