% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rastercalculator.R
\name{saga_rastercalculator}
\alias{saga_rastercalculator}
\title{QGIS algorithm Raster calculator}
\usage{
saga_rastercalculator(
  GRIDS = qgisprocess:::qgis_default_value(),
  XGRIDS = qgisprocess:::qgis_default_value(),
  FORMULA = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  USE_NODATA = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{raster} - Main input layer. Path to a raster layer.}

\item{XGRIDS}{\code{multilayer} - Additional layers. .}

\item{FORMULA}{\code{string} - Formula. String value.}

\item{RESAMPLING}{\code{enum}  of \verb{("[0] Nearest Neighbour", "[1] Bilinear Interpolation", "[2] Bicubic Spline Interpolation", "[3] B-Spline Interpolation")} - Resampling Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{USE_NODATA}{\code{boolean} - Use NoData. 1 for true/yes. 0 for false/no.}

\item{TYPE}{\code{enum}  of \verb{("[0] bit", "[1] unsigned 1 byte integer", "[2] signed 1 byte integer", "[3] unsigned 2 byte integer", "[4] signed 2 byte integer", "[5] unsigned 4 byte integer", "[6] signed 4 byte integer", "[7] 4 byte floating point number", "[8] 8 byte floating point number")} - Output Data Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESULT}{\code{rasterDestination} - Calculated. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster calculator (saga:rastercalculator)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Calculated
}
}
}
