% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rastercellindex.R
\name{saga_rastercellindex}
\alias{saga_rastercellindex}
\title{QGIS algorithm Raster cell index}
\usage{
saga_rastercellindex(
  GRID = qgisprocess::qgis_default_value(),
  ORDER = qgisprocess::qgis_default_value(),
  INDEX = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Input Grid. Path to a raster layer.}

\item{ORDER}{\code{enum}  of \verb{("[0] ascending", "[1] descending")} - Index. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{INDEX}{\code{rasterDestination} - Sorted Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster cell index (saga:rastercellindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INDEX - outputRaster - Sorted Grid
}
}
}
