% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rasterdifference.R
\name{saga_rasterdifference}
\alias{saga_rasterdifference}
\title{QGIS algorithm Raster difference}
\usage{
saga_rasterdifference(
  A = qgisprocess::qgis_default_value(),
  B = qgisprocess::qgis_default_value(),
  C = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{A}{\code{raster} - A. Path to a raster layer.}

\item{B}{\code{raster} - B. Path to a raster layer.}

\item{C}{\code{rasterDestination} - Difference (A - B). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster difference (saga:rasterdifference)
}
\details{
\subsection{Outputs description}{
\itemize{
\item C - outputRaster - Difference (A - B)
}
}
}
