% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rasterize.R
\name{saga_rasterize}
\alias{saga_rasterize}
\title{QGIS algorithm Rasterize}
\usage{
saga_rasterize(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  MULTIPLE = qgisprocess::qgis_default_value(),
  LINE_TYPE = qgisprocess::qgis_default_value(),
  POLY_TYPE = qgisprocess::qgis_default_value(),
  GRID_TYPE = qgisprocess::qgis_default_value(),
 
    TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX = qgisprocess::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess::qgis_default_value(),
  TARGET_USER_FITS = qgisprocess::qgis_default_value(),
  GRID = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Shapes. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{enum}  of \verb{("[0] data / no-data", "[1] index number", "[2] attribute")} - Output Values. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MULTIPLE}{\code{enum}  of \verb{("[0] first", "[1] last", "[2] minimum", "[3] maximum", "[4] mean")} - Method for Multiple Values. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LINE_TYPE}{\code{enum}  of \verb{("[0] thin", "[1] thick")} - Method for Lines. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{POLY_TYPE}{\code{enum}  of \verb{("[0] node", "[1] cell")} - Method for Lines. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRID_TYPE}{\code{enum}  of \verb{("[0] Integer (1 byte)", "[1] Integer (2 byte)", "[2] Integer (4 byte)", "[3] Floating Point (4 byte)", "[4] Floating Point (8 byte)")} - Preferred Target Grid Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: {name}.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{TARGET_USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRID}{\code{rasterDestination} - Rasterized. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Rasterize (saga:rasterize)
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRID - outputRaster - Rasterized
}
}
}
