% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rasternormalisation.R
\name{saga_rasternormalisation}
\alias{saga_rasternormalisation}
\title{QGIS algorithm Raster normalisation}
\usage{
saga_rasternormalisation(
  INPUT = qgisprocess:::qgis_default_value(),
  RANGE_MIN = qgisprocess:::qgis_default_value(),
  RANGE_MAX = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{RANGE_MIN}{\code{number} - Target Range (min). A numeric value.}

\item{RANGE_MAX}{\code{number} - Target Range (max). A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Normalised Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster normalisation (saga:rasternormalisation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Normalised Grid
}
}
}
