% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rasterssum.R
\name{saga_rasterssum}
\alias{saga_rasterssum}
\title{QGIS algorithm Rasters sum}
\usage{
saga_rasterssum(
  GRIDS = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{RESULT}{\code{rasterDestination} - Sum. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Rasters sum (saga:rasterssum)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Sum
}
}
}
