% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rasterstatisticsforpolygons.R
\name{saga_rasterstatisticsforpolygons}
\alias{saga_rasterstatisticsforpolygons}
\title{QGIS algorithm Raster statistics for polygons}
\usage{
saga_rasterstatisticsforpolygons(
  GRIDS = qgisprocess::qgis_default_value(),
  POLYGONS = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  NAMING = qgisprocess::qgis_default_value(),
  COUNT = qgisprocess::qgis_default_value(),
  MIN = qgisprocess::qgis_default_value(),
  MAX = qgisprocess::qgis_default_value(),
  RANGE = qgisprocess::qgis_default_value(),
  SUM = qgisprocess::qgis_default_value(),
  MEAN = qgisprocess::qgis_default_value(),
  VAR = qgisprocess::qgis_default_value(),
  STDDEV = qgisprocess::qgis_default_value(),
  QUANTILE = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Standard", "[1] Shape wise, supports overlapping polygons")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NAMING}{\code{enum}  of \verb{("[0] Grid number", "[1] Grid name")} - Grid Naming. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{COUNT}{\code{boolean} - Number of Cells. 1 for true/yes. 0 for false/no.}

\item{MIN}{\code{boolean} - Minimum. 1 for true/yes. 0 for false/no.}

\item{MAX}{\code{boolean} - Maximum. 1 for true/yes. 0 for false/no.}

\item{RANGE}{\code{boolean} - Range. 1 for true/yes. 0 for false/no.}

\item{SUM}{\code{boolean} - Sum. 1 for true/yes. 0 for false/no.}

\item{MEAN}{\code{boolean} - Mean. 1 for true/yes. 0 for false/no.}

\item{VAR}{\code{boolean} - Variance. 1 for true/yes. 0 for false/no.}

\item{STDDEV}{\code{boolean} - Standard Deviation. 1 for true/yes. 0 for false/no.}

\item{QUANTILE}{\code{number} - Quantiles. A numeric value.}

\item{RESULT}{\code{vectorDestination} - Statistics. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster statistics for polygons (saga:rasterstatisticsforpolygons)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Statistics
}
}
}
