% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rastervaluestopointsrandomly.R
\name{saga_rastervaluestopointsrandomly}
\alias{saga_rastervaluestopointsrandomly}
\title{QGIS algorithm Raster values to points (randomly)}
\usage{
saga_rastervaluestopointsrandomly(
  GRID = qgisprocess::qgis_default_value(),
  FREQ = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{FREQ}{\code{number} - Frequency. A numeric value.}

\item{POINTS}{\code{vectorDestination} - Points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster values to points (randomly) (saga:rastervaluestopointsrandomly)
}
\details{
\subsection{Outputs description}{
\itemize{
\item POINTS - outputVector - Points
}
}
}
