% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_realsurfacearea.R
\name{saga_realsurfacearea}
\alias{saga_realsurfacearea}
\title{QGIS algorithm Real surface area}
\usage{
saga_realsurfacearea(
  DEM = qgisprocess::qgis_default_value(),
  AREA = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{AREA}{\code{rasterDestination} - Surface Area. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Real surface area (saga:realsurfacearea)
}
\details{
\subsection{Outputs description}{
\itemize{
\item AREA - outputRaster - Surface Area
}
}
}
