% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_reclassifyvaluessimple.R
\name{saga_reclassifyvaluessimple}
\alias{saga_reclassifyvaluessimple}
\title{QGIS algorithm Reclassify values (simple)}
\usage{
saga_reclassifyvaluessimple(
  GRID_IN = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  LOOKUP = qgisprocess::qgis_default_value(),
  GRID_OUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID_IN}{\code{raster} - Grid. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Grid value equals low value", "[1] Low value < grid value < high value", "[2] Low value <= grid value < high value")} - Replace Condition. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LOOKUP}{\code{matrix} - Lookup Table. A comma delimited list of values.}

\item{GRID_OUT}{\code{rasterDestination} - Changed Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Reclassify values (simple) (saga:reclassifyvaluessimple)
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRID_OUT - outputRaster - Changed Grid
}
}
}
