% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_reclassifyvaluessingle.R
\name{saga_reclassifyvaluessingle}
\alias{saga_reclassifyvaluessingle}
\title{QGIS algorithm Reclassify values (single)}
\usage{
saga_reclassifyvaluessingle(
  INPUT = qgisprocess::qgis_default_value(),
  OLD = qgisprocess::qgis_default_value(),
  NEW = qgisprocess::qgis_default_value(),
  SOPERATOR = qgisprocess::qgis_default_value(),
  NODATAOPT = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  OTHEROPT = qgisprocess::qgis_default_value(),
  OTHERS = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{OLD}{\code{number} - old value (for single value change). A numeric value.}

\item{NEW}{\code{number} - new value (for single value change). A numeric value.}

\item{SOPERATOR}{\code{enum}  of \verb{("[0] =", "[1] <", "[2] <=", "[3] >=", "[4] >")} - operator (for single value change). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NODATAOPT}{\code{boolean} - replace no data values. 1 for true/yes. 0 for false/no.}

\item{NODATA}{\code{number} - new value for no data values. A numeric value.}

\item{OTHEROPT}{\code{boolean} - replace other values. 1 for true/yes. 0 for false/no.}

\item{OTHERS}{\code{number} - new value for other values. A numeric value.}

\item{RESULT}{\code{rasterDestination} - Reclassified Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Reclassify values (single) (saga:reclassifyvaluessingle)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Reclassified Grid
}
}
}
