% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_regressionanalysis.R
\name{saga_regressionanalysis}
\alias{saga_regressionanalysis}
\title{QGIS algorithm Regression analysis}
\usage{
saga_regressionanalysis(
  GRID = qgisprocess::qgis_default_value(),
  SHAPES = qgisprocess::qgis_default_value(),
  ATTRIBUTE = qgisprocess::qgis_default_value(),
  INTERPOL = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  REGRESSION = qgisprocess::qgis_default_value(),
  RESIDUAL = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{ATTRIBUTE}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{INTERPOL}{\code{enum}  of \verb{("[0] Nearest Neighbor", "[1] Bilinear Interpolation", "[2] Inverse Distance Interpolation", "[3] Bicubic Spline Interpolation", "[4] B-Spline Interpolation")} - Grid Interpolation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("[0] Y = a + b * X (linear)", "[1] Y = a + b / X", "[2] Y = a / (b - X)", "[3] Y = a * X^b (power)", "[4] Y = a e^(b * X) (exponential)", "[5] Y = a + b * ln(X) (logarithmic)")} - Regression Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{REGRESSION}{\code{rasterDestination} - Regression. Path for new raster layer.}

\item{RESIDUAL}{\code{vectorDestination} - Residuals. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Regression analysis (saga:regressionanalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item REGRESSION - outputRaster - Regression
\item RESIDUAL - outputVector - Residuals
}
}
}
