% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_regressionkriging.R
\name{saga_regressionkriging}
\alias{saga_regressionkriging}
\title{QGIS algorithm Regression kriging}
\usage{
saga_regressionkriging(
  POINTS = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  PREDICTORS = qgisprocess::qgis_default_value(),
  REGRESSION = qgisprocess::qgis_default_value(),
  PREDICTION = qgisprocess::qgis_default_value(),
  RESIDUALS = qgisprocess::qgis_default_value(),
  VARIANCE = qgisprocess::qgis_default_value(),
  INFO_COEFF = qgisprocess::qgis_default_value(),
  INFO_MODEL = qgisprocess::qgis_default_value(),
  INFO_STEPS = qgisprocess::qgis_default_value(),
  TQUALITY = qgisprocess::qgis_default_value(),
  LOG = qgisprocess::qgis_default_value(),
  BLOCK = qgisprocess::qgis_default_value(),
  DBLOCK = qgisprocess::qgis_default_value(),
  VAR_MAXDIST = qgisprocess::qgis_default_value(),
  VAR_NCLASSES = qgisprocess::qgis_default_value(),
  VAR_NSKIP = qgisprocess::qgis_default_value(),
  VAR_MODEL = qgisprocess::qgis_default_value(),
  COORD_X = qgisprocess::qgis_default_value(),
  COORD_Y = qgisprocess::qgis_default_value(),
  INTERCEPT = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  P_VALUE = qgisprocess::qgis_default_value(),
  RESAMPLING = qgisprocess::qgis_default_value(),
  SEARCH_RANGE = qgisprocess::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_ALL = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MIN = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MAX = qgisprocess::qgis_default_value(),
  SEARCH_DIRECTION = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{PREDICTORS}{\code{multilayer} - Predictors. .}

\item{REGRESSION}{\code{rasterDestination} - Regression. Path for new raster layer.}

\item{PREDICTION}{\code{rasterDestination} - Prediction. Path for new raster layer.}

\item{RESIDUALS}{\code{rasterDestination} - Residuals. Path for new raster layer.}

\item{VARIANCE}{\code{rasterDestination} - Quality Measure. Path for new raster layer.}

\item{INFO_COEFF}{\code{vectorDestination} - Regression: Coefficients. Path for new vector layer.}

\item{INFO_MODEL}{\code{vectorDestination} - Regression: Model. Path for new vector layer.}

\item{INFO_STEPS}{\code{vectorDestination} - Regression: Steps. Path for new vector layer.}

\item{TQUALITY}{\code{enum}  of \verb{("[0] standard deviation", "[1] variance")} - Type of Quality Measure. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LOG}{\code{boolean} - Logarithmic Transformation. 1 for true/yes. 0 for false/no.}

\item{BLOCK}{\code{boolean} - Block Kriging. 1 for true/yes. 0 for false/no.}

\item{DBLOCK}{\code{number} - Block Size. A numeric value.}

\item{VAR_MAXDIST}{\code{number} - Maximum Distance. A numeric value.}

\item{VAR_NCLASSES}{\code{number} - Lag Distance Classes. A numeric value.}

\item{VAR_NSKIP}{\code{number} - Skip. A numeric value.}

\item{VAR_MODEL}{\code{string} - Variogram Model. String value.}

\item{COORD_X}{\code{boolean} - Include X Coordinate. 1 for true/yes. 0 for false/no.}

\item{COORD_Y}{\code{boolean} - Include Y Coordinate. 1 for true/yes. 0 for false/no.}

\item{INTERCEPT}{\code{boolean} - Intercept. 1 for true/yes. 0 for false/no.}

\item{METHOD}{\code{enum}  of \verb{("[0] include all", "[1] forward", "[2] backward", "[3] stepwise")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{P_VALUE}{\code{number} - Significance Level. A numeric value.}

\item{RESAMPLING}{\code{enum}  of \verb{("[0] Nearest Neighbour", "[1] Bilinear Interpolation", "[2] Bicubic Spline Interpolation", "[3] B-Spline Interpolation")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_RANGE}{\code{enum}  of \verb{("[0] local", "[1] global")} - Search Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_RADIUS}{\code{number} - Maximum Search Distance. A numeric value.}

\item{SEARCH_POINTS_ALL}{\code{enum}  of \verb{("[0] maximum number of nearest points", "[1] all points within search distance")} - Number of Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_POINTS_MIN}{\code{number} - Minimum. A numeric value.}

\item{SEARCH_POINTS_MAX}{\code{number} - Maximum. A numeric value.}

\item{SEARCH_DIRECTION}{\code{enum}  of \verb{("[0] all directions", "[1] quadrants")} - Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Regression kriging (saga:regressionkriging)
}
\details{
\subsection{Outputs description}{
\itemize{
\item REGRESSION - outputRaster - Regression
\item PREDICTION - outputRaster - Prediction
\item RESIDUALS - outputRaster - Residuals
\item VARIANCE - outputRaster - Quality Measure
\item INFO_COEFF - outputVector - Regression
\item INFO_MODEL - outputVector - Regression
\item INFO_STEPS - outputVector - Regression
}
}
}
