% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_relativeheightsandslopepositions.R
\name{saga_relativeheightsandslopepositions}
\alias{saga_relativeheightsandslopepositions}
\title{QGIS algorithm Relative heights and slope positions}
\usage{
saga_relativeheightsandslopepositions(
  DEM = qgisprocess::qgis_default_value(),
  W = qgisprocess::qgis_default_value(),
  T = qgisprocess::qgis_default_value(),
  E = qgisprocess::qgis_default_value(),
  HO = qgisprocess::qgis_default_value(),
  HU = qgisprocess::qgis_default_value(),
  NH = qgisprocess::qgis_default_value(),
  SH = qgisprocess::qgis_default_value(),
  MS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{W}{\code{number} - w. A numeric value.}

\item{T}{\code{number} - t. A numeric value.}

\item{E}{\code{number} - e. A numeric value.}

\item{HO}{\code{rasterDestination} - Slope Height. Path for new raster layer.}

\item{HU}{\code{rasterDestination} - Valley Depth. Path for new raster layer.}

\item{NH}{\code{rasterDestination} - Normalized Height. Path for new raster layer.}

\item{SH}{\code{rasterDestination} - Standardized Height. Path for new raster layer.}

\item{MS}{\code{rasterDestination} - Mid-Slope Position. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Relative heights and slope positions (saga:relativeheightsandslopepositions)
}
\details{
\subsection{Outputs description}{
\itemize{
\item HO - outputRaster - Slope Height
\item HU - outputRaster - Valley Depth
\item NH - outputRaster - Normalized Height
\item SH - outputRaster - Standardized Height
\item MS - outputRaster - Mid
}
}
}
