% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_resampling.R
\name{saga_resampling}
\alias{saga_resampling}
\title{QGIS algorithm Resampling}
\usage{
saga_resampling(
  INPUT = qgisprocess::qgis_default_value(),
  KEEP_TYPE = qgisprocess::qgis_default_value(),
  SCALE_UP = qgisprocess::qgis_default_value(),
  SCALE_DOWN = qgisprocess::qgis_default_value(),
 
    TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX = qgisprocess::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess::qgis_default_value(),
  TARGET_USER_FITS = qgisprocess::qgis_default_value(),
  TARGET_TEMPLATE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{KEEP_TYPE}{\code{boolean} - Preserve Data Type. 1 for true/yes. 0 for false/no.}

\item{SCALE_UP}{\code{enum}  of \verb{("[0] Nearest Neighbor", "[1] Bilinear Interpolation", "[2] Bicubic Spline Interpolation", "[3] B-Spline Interpolation", "[4] Mean Value", "[5] Mean Value (cell area weighted)", "[6] Minimum Value", "[7] Maximum Value", "[8] Majority")} - Upscaling Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SCALE_DOWN}{\code{enum}  of \verb{("[0] Nearest Neighbor", "[1] Bilinear Interpolation", "[2] Bicubic Spline Interpolation", "[3] B-Spline Interpolation")} - Downscaling Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{TARGET_USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_TEMPLATE}{\code{raster} - Target system. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Resampling (saga:resampling)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Grid
}
}
}
