% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_resamplingfilter.R
\name{saga_resamplingfilter}
\alias{saga_resamplingfilter}
\title{QGIS algorithm Resampling filter}
\usage{
saga_resamplingfilter(
  GRID = qgisprocess::qgis_default_value(),
  LOPASS = qgisprocess::qgis_default_value(),
  HIPASS = qgisprocess::qgis_default_value(),
  SCALE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{LOPASS}{\code{rasterDestination} - Low Pass Filter. Path for new raster layer.}

\item{HIPASS}{\code{rasterDestination} - High Pass Filter. Path for new raster layer.}

\item{SCALE}{\code{number} - Scale Factor. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Resampling filter (saga:resamplingfilter)
}
\details{
\subsection{Outputs description}{
\itemize{
\item HIPASS - outputRaster - High Pass Filter
\item LOPASS - outputRaster - Low Pass Filter
}
}
}
