% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_residualanalysis.R
\name{saga_residualanalysis}
\alias{saga_residualanalysis}
\title{QGIS algorithm Residual analysis}
\usage{
saga_residualanalysis(
  GRID = qgisprocess::qgis_default_value(),
  MODE = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  BCENTER = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_WEIGHTING = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_IDW_POWER = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_IDW_OFFSET = qgisprocess::qgis_default_value(),
  DISTANCE_WEIGHTING_DW_BANDWIDTH = qgisprocess::qgis_default_value(),
  MEAN = qgisprocess::qgis_default_value(),
  DIFF = qgisprocess::qgis_default_value(),
  STDDEV = qgisprocess::qgis_default_value(),
  RANGE = qgisprocess::qgis_default_value(),
  MIN = qgisprocess::qgis_default_value(),
  MAX = qgisprocess::qgis_default_value(),
  DEVMEAN = qgisprocess::qgis_default_value(),
  PERCENT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{MODE}{\code{enum}  of \verb{("[0] square", "[1] circle")} - Search Mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radius (Cells). A numeric value.}

\item{BCENTER}{\code{boolean} - Include Center Cell. 1 for true/yes. 0 for false/no.}

\item{DISTANCE_WEIGHTING_DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Distance Weighting. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DISTANCE_WEIGHTING_DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DISTANCE_WEIGHTING_DW_IDW_OFFSET}{\code{boolean} - Inverse Distance Offset. 1 for true/yes. 0 for false/no.}

\item{DISTANCE_WEIGHTING_DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{MEAN}{\code{rasterDestination} - Mean Value. Path for new raster layer.}

\item{DIFF}{\code{rasterDestination} - Difference from Mean Value. Path for new raster layer.}

\item{STDDEV}{\code{rasterDestination} - Standard Deviation. Path for new raster layer.}

\item{RANGE}{\code{rasterDestination} - Value Range. Path for new raster layer.}

\item{MIN}{\code{rasterDestination} - Minimum Value. Path for new raster layer.}

\item{MAX}{\code{rasterDestination} - Maximum Value. Path for new raster layer.}

\item{DEVMEAN}{\code{rasterDestination} - Deviation from Mean Value. Path for new raster layer.}

\item{PERCENT}{\code{rasterDestination} - Percentile. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Residual analysis (saga:residualanalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DEVMEAN - outputRaster - Deviation from Mean Value
\item DIFF - outputRaster - Difference from Mean Value
\item MAX - outputRaster - Maximum Value
\item MEAN - outputRaster - Mean Value
\item MIN - outputRaster - Minimum Value
\item PERCENT - outputRaster - Percentile
\item RANGE - outputRaster - Value Range
\item STDDEV - outputRaster - Standard Deviation
}
}
}
