% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rgbcomposite.R
\name{saga_rgbcomposite}
\alias{saga_rgbcomposite}
\title{QGIS algorithm Rgb composite}
\usage{
saga_rgbcomposite(
  R_GRID = qgisprocess::qgis_default_value(),
  G_GRID = qgisprocess::qgis_default_value(),
  B_GRID = qgisprocess::qgis_default_value(),
  R_METHOD = qgisprocess::qgis_default_value(),
  G_METHOD = qgisprocess::qgis_default_value(),
  B_METHOD = qgisprocess::qgis_default_value(),
  R_RANGE_MIN = qgisprocess::qgis_default_value(),
  R_RANGE_MAX = qgisprocess::qgis_default_value(),
  R_PERCTL_MIN = qgisprocess::qgis_default_value(),
  R_PERCTL_MAX = qgisprocess::qgis_default_value(),
  R_STDDEV = qgisprocess::qgis_default_value(),
  G_RANGE_MIN = qgisprocess::qgis_default_value(),
  G_RANGE_MAX = qgisprocess::qgis_default_value(),
  G_PERCTL_MIN = qgisprocess::qgis_default_value(),
  G_PERCTL_MAX = qgisprocess::qgis_default_value(),
  G_STDDEV = qgisprocess::qgis_default_value(),
  B_RANGE_MIN = qgisprocess::qgis_default_value(),
  B_RANGE_MAX = qgisprocess::qgis_default_value(),
  B_PERCTL_MIN = qgisprocess::qgis_default_value(),
  B_PERCTL_MAX = qgisprocess::qgis_default_value(),
  B_STDDEV = qgisprocess::qgis_default_value(),
  RGB = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{R_GRID}{\code{raster} - R. Path to a raster layer.}

\item{G_GRID}{\code{raster} - G. Path to a raster layer.}

\item{B_GRID}{\code{raster} - B. Path to a raster layer.}

\item{R_METHOD}{\code{enum}  of \verb{("0 - 255", "Rescale to 0 - 255", "User defined rescale", "Percentiles", "Percentage of standard deviation")} - Method for R value. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{G_METHOD}{\code{enum}  of \verb{("0 - 255", "Rescale to 0 - 255", "User defined rescale", "Percentiles", "Percentage of standard deviation")} - Method for G value. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{B_METHOD}{\code{enum}  of \verb{("0 - 255", "Rescale to 0 - 255", "User defined rescale", "Percentiles", "Percentage of standard deviation")} - Method for B value. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{R_RANGE_MIN}{\code{number} - Rescale Range for RED min. A numeric value.}

\item{R_RANGE_MAX}{\code{number} - Rescale Range for RED max. A numeric value.}

\item{R_PERCTL_MIN}{\code{number} - Percentiles Range for RED max. A numeric value.}

\item{R_PERCTL_MAX}{\code{number} - Percentiles Range for RED max. A numeric value.}

\item{R_STDDEV}{\code{number} - Standard deviation for RED. A numeric value.}

\item{G_RANGE_MIN}{\code{number} - Rescale Range for GREEN min. A numeric value.}

\item{G_RANGE_MAX}{\code{number} - Rescale Range for GREEN max. A numeric value.}

\item{G_PERCTL_MIN}{\code{number} - Percentiles Range for GREEN max. A numeric value.}

\item{G_PERCTL_MAX}{\code{number} - Percentiles Range for GREEN max. A numeric value.}

\item{G_STDDEV}{\code{number} - Standard deviation for GREEN. A numeric value.}

\item{B_RANGE_MIN}{\code{number} - Rescale Range for BLUE min. A numeric value.}

\item{B_RANGE_MAX}{\code{number} - Rescale Range for BLUE max. A numeric value.}

\item{B_PERCTL_MIN}{\code{number} - Percentiles Range for BLUE max. A numeric value.}

\item{B_PERCTL_MAX}{\code{number} - Percentiles Range for BLUE max. A numeric value.}

\item{B_STDDEV}{\code{number} - Standard deviation for BLUE. A numeric value.}

\item{RGB}{\code{rasterDestination} - Output RGB. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Rgb composite (saga:rgbcomposite)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RGB - outputRaster - Output RGB
}
}
}
