% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_riverbasin.R
\name{saga_riverbasin}
\alias{saga_riverbasin}
\title{QGIS algorithm Riverbasin}
\usage{
saga_riverbasin(
  INPUT = qgisprocess:::qgis_default_value(),
  INPUT2 = qgisprocess:::qgis_default_value(),
  INPUT3 = qgisprocess:::qgis_default_value(),
  WCons = qgisprocess:::qgis_default_value(),
  WCons2 = qgisprocess:::qgis_default_value(),
  OUTPUT2 = qgisprocess:::qgis_default_value(),
  OUTPUT3 = qgisprocess:::qgis_default_value(),
  OUTPUT4 = qgisprocess:::qgis_default_value(),
  OUTPUT5 = qgisprocess:::qgis_default_value(),
  OUTPUT6 = qgisprocess:::qgis_default_value(),
  OUTPUT7 = qgisprocess:::qgis_default_value(),
  OUTPUT8 = qgisprocess:::qgis_default_value(),
  OUTPUT9 = qgisprocess:::qgis_default_value(),
  pCr = qgisprocess:::qgis_default_value(),
  nCr = qgisprocess:::qgis_default_value(),
  EnfVmax = qgisprocess:::qgis_default_value(),
  VTresh = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - DTM. Path to a raster layer.}

\item{INPUT2}{\code{raster} - HGGrid. Path to a raster layer.}

\item{INPUT3}{\code{raster} - statisches Entnahmeraster. Path to a raster layer.}

\item{WCons}{\code{boolean} - Anteilige Flaechenwasserentnahme. 1 for true/yes. 0 for false/no.}

\item{WCons2}{\code{enum}  of \verb{("[0] ...anteilig aus den Flussrasterzellen", "[1] ...anteilig aus Rasterzellen der Teileinzugegebiete")} - Dynamische Flaechenwassernutzung.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT2}{\code{rasterDestination} - Grad. Path for new raster layer.}

\item{OUTPUT3}{\code{rasterDestination} - Direc. Path for new raster layer.}

\item{OUTPUT4}{\code{rasterDestination} - HGGrad. Path for new raster layer.}

\item{OUTPUT5}{\code{rasterDestination} - RivSpeed. Path for new raster layer.}

\item{OUTPUT6}{\code{rasterDestination} - Coordinates. Path for new raster layer.}

\item{OUTPUT7}{\code{rasterDestination} - BasinShare. Path for new raster layer.}

\item{OUTPUT8}{\code{rasterDestination} - statWUse. Path for new raster layer.}

\item{OUTPUT9}{\code{rasterDestination} - NumInFlowCells. Path for new raster layer.}

\item{pCr}{\code{number} - Hauptgerinne-Parameter pHG. A numeric value.}

\item{nCr}{\code{number} - Hauptgerinne-Speicherkaskade nHG. A numeric value.}

\item{EnfVmax}{\code{boolean} - Maximal Geschwindigkeit des Hauptgerinnes beruecksichtigen. 1 for true/yes. 0 for false/no.}

\item{VTresh}{\code{number} - Maximalgeschwindigkeit im Hauptgerinne in km/h. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Riverbasin (saga:riverbasin)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT2 - outputRaster - Grad
\item OUTPUT3 - outputRaster - Direc
\item OUTPUT4 - outputRaster - HGGrad
\item OUTPUT5 - outputRaster - RivSpeed
\item OUTPUT6 - outputRaster - Coordinates
\item OUTPUT7 - outputRaster - BasinShare
\item OUTPUT8 - outputRaster - statWUse
\item OUTPUT9 - outputRaster - NumInFlowCells
}
}
}
