% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rivergridgeneration.R
\name{saga_rivergridgeneration}
\alias{saga_rivergridgeneration}
\title{QGIS algorithm Rivergridgeneration}
\usage{
saga_rivergridgeneration(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  SX = qgisprocess::qgis_default_value(),
  SY = qgisprocess::qgis_default_value(),
  MX = qgisprocess::qgis_default_value(),
  MY = qgisprocess::qgis_default_value(),
  Owrite = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Geländemodell (DTM). Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - HG Raster. Path for new raster layer.}

\item{SX}{\code{number} - Abflusspfad-Quelle, x-Wert. A numeric value.}

\item{SY}{\code{number} - Abflusspfad-Quelle, y-Wert. A numeric value.}

\item{MX}{\code{number} - Abflusspfad-Mündung, x-Wert. A numeric value.}

\item{MY}{\code{number} - Abflusspfad-Mündung, y-Wert. A numeric value.}

\item{Owrite}{\code{boolean} - Overwrite RiverGridCells. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Rivergridgeneration (saga:rivergridgeneration)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - HG Raster
}
}
}
