% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_sagawetnessindex.R
\name{saga_sagawetnessindex}
\alias{saga_sagawetnessindex}
\title{QGIS algorithm Saga wetness index}
\usage{
saga_sagawetnessindex(
  DEM = qgisprocess::qgis_default_value(),
  SUCTION = qgisprocess::qgis_default_value(),
  AREA_TYPE = qgisprocess::qgis_default_value(),
  SLOPE_TYPE = qgisprocess::qgis_default_value(),
  SLOPE_MIN = qgisprocess::qgis_default_value(),
  SLOPE_OFF = qgisprocess::qgis_default_value(),
  SLOPE_WEIGHT = qgisprocess::qgis_default_value(),
  AREA = qgisprocess::qgis_default_value(),
  SLOPE = qgisprocess::qgis_default_value(),
  AREA_MOD = qgisprocess::qgis_default_value(),
  TWI = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{SUCTION}{\code{number} - Suction. A numeric value.}

\item{AREA_TYPE}{\code{enum}  of \verb{("[0] absolute catchment area", "[1] square root of catchment area", "[2] specific catchment area")} - Type of Area. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SLOPE_TYPE}{\code{enum}  of \verb{("[0] local slope", "[1] catchment slope")} - Type of Slope. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SLOPE_MIN}{\code{number} - Suction. A numeric value.}

\item{SLOPE_OFF}{\code{number} - Suction. A numeric value.}

\item{SLOPE_WEIGHT}{\code{number} - Suction. A numeric value.}

\item{AREA}{\code{rasterDestination} - Catchment area. Path for new raster layer.}

\item{SLOPE}{\code{rasterDestination} - Catchment slope. Path for new raster layer.}

\item{AREA_MOD}{\code{rasterDestination} - Modified catchment area. Path for new raster layer.}

\item{TWI}{\code{rasterDestination} - Topographic Wetness Index. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Saga wetness index (saga:sagawetnessindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item AREA - outputRaster - Catchment area
\item AREA_MOD - outputRaster - Modified catchment area
\item SLOPE - outputRaster - Catchment slope
\item TWI - outputRaster - Topographic Wetness Index
}
}
}
