% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_seededregiongrowing.R
\name{saga_seededregiongrowing}
\alias{saga_seededregiongrowing}
\title{QGIS algorithm Seeded region growing}
\usage{
saga_seededregiongrowing(
  SEEDS = qgisprocess:::qgis_default_value(),
  FEATURES = qgisprocess:::qgis_default_value(),
  SEGMENTS = qgisprocess:::qgis_default_value(),
  SIMILARITY = qgisprocess:::qgis_default_value(),
  TABLE = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  NEIGHBOUR = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  SIG_1 = qgisprocess:::qgis_default_value(),
  SIG_2 = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  REFRESH = qgisprocess:::qgis_default_value(),
  LEAFSIZE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SEEDS}{\code{raster} - Seeds. Path to a raster layer.}

\item{FEATURES}{\code{multilayer} - Features. .}

\item{SEGMENTS}{\code{rasterDestination} - Segments. Path for new raster layer.}

\item{SIMILARITY}{\code{rasterDestination} - Similarity. Path for new raster layer.}

\item{TABLE}{\code{source} - Seeds. Path to a vector layer.}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no.}

\item{NEIGHBOUR}{\code{enum}  of \verb{("[0] 4 (von Neumann)", "[1] 8 (Moore)")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("[0] feature space and position", "[1] feature space")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SIG_1}{\code{number} - Variance in Feature Space. A numeric value.}

\item{SIG_2}{\code{number} - Variance in Position Space. A numeric value.}

\item{THRESHOLD}{\code{number} - Similarity Threshold. A numeric value.}

\item{REFRESH}{\code{boolean} - Refresh. 1 for true/yes. 0 for false/no.}

\item{LEAFSIZE}{\code{number} - Leaf Size (for Speed Optimisation). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Seeded region growing (saga:seededregiongrowing)
}
\details{
\subsection{Outputs description}{
\itemize{
\item SEGMENTS - outputRaster - Segments
\item SIMILARITY - outputRaster - Similarity
}
}
}
