% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_seedgeneration.R
\name{saga_seedgeneration}
\alias{saga_seedgeneration}
\title{QGIS algorithm Seed generation}
\usage{
saga_seedgeneration(
  FEATURES = qgisprocess:::qgis_default_value(),
  SEED_TYPE = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  BAND_WIDTH = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  DW_WEIGHTING = qgisprocess:::qgis_default_value(),
  DW_IDW_POWER = qgisprocess:::qgis_default_value(),
  DW_IDW_OFFSET = qgisprocess:::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess:::qgis_default_value(),
  SEED_GRID = qgisprocess:::qgis_default_value(),
  VARIANCE = qgisprocess:::qgis_default_value(),
  SEED_POINTS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FEATURES}{\code{multilayer} - Features. .}

\item{SEED_TYPE}{\code{enum}  of \verb{("[0] minima", "[1] maxima")} - Seed Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("[0] band width smoothing", "[1] band width search")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{BAND_WIDTH}{\code{number} - Bandwidth (Cells). A numeric value.}

\item{NORMALIZE}{\code{boolean} - Normalize Features. 1 for true/yes. 0 for false/no.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DW_IDW_OFFSET}{\code{boolean} - Inverse Distance Offset. 1 for true/yes. 0 for false/no.}

\item{DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{SEED_GRID}{\code{rasterDestination} - Seeds Grid. Path for new raster layer.}

\item{VARIANCE}{\code{rasterDestination} - Variance. Path for new raster layer.}

\item{SEED_POINTS}{\code{vectorDestination} - Seeds Points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Seed generation (saga:seedgeneration)
}
\details{
\subsection{Outputs description}{
\itemize{
\item SEED_GRID - outputRaster - Seeds Grid
\item SEED_POINTS - outputVector - Seeds Points
\item VARIANCE - outputRaster - Variance
}
}
}
