% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_sharedpolygonedges.R
\name{saga_sharedpolygonedges}
\alias{saga_sharedpolygonedges}
\title{QGIS algorithm Shared polygon edges}
\usage{
saga_sharedpolygonedges(
  POLYGONS = qgisprocess::qgis_default_value(),
  ATTRIBUTE = qgisprocess::qgis_default_value(),
  EPSILON = qgisprocess::qgis_default_value(),
  VERTICES = qgisprocess::qgis_default_value(),
  DOUBLE = qgisprocess::qgis_default_value(),
  EDGES = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{ATTRIBUTE}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{EPSILON}{\code{number} - Tolerance. A numeric value.}

\item{VERTICES}{\code{boolean} - Check vertices. 1 for true/yes. 0 for false/no.}

\item{DOUBLE}{\code{boolean} - Double edges. 1 for true/yes. 0 for false/no.}

\item{EDGES}{\code{vectorDestination} - Edges. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Shared polygon edges (saga:sharedpolygonedges)
}
\details{
\subsection{Outputs description}{
\itemize{
\item EDGES - outputVector - Edges
}
}
}
