% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_shrinkandexpand.R
\name{saga_shrinkandexpand}
\alias{saga_shrinkandexpand}
\title{QGIS algorithm Shrink and expand}
\usage{
saga_shrinkandexpand(
  INPUT = qgisprocess::qgis_default_value(),
  OPERATION = qgisprocess::qgis_default_value(),
  CIRCLE = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  EXPAND = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{OPERATION}{\code{enum}  of \verb{("[0] Shrink", "[1] Expand", "[2] shrink and expand", "[3] expand and shrink")} - Operation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CIRCLE}{\code{enum}  of \verb{("[0] Square", "[1] Circle")} - Search Mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radius. A numeric value.}

\item{EXPAND}{\code{enum}  of \verb{("[0] min", "[1] max", "[2] mean", "[3] majority")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESULT}{\code{rasterDestination} - Result Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Shrink and expand (saga:shrinkandexpand)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Result Grid
}
}
}
