% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_sievingclasses.R
\name{saga_sievingclasses}
\alias{saga_sievingclasses}
\title{QGIS algorithm Sieving classes}
\usage{
saga_sievingclasses(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  MODE = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  ALL = qgisprocess::qgis_default_value(),
  CLASS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Classes. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Sieved Classes. Path for new raster layer.}

\item{MODE}{\code{enum}  of \verb{("[0] Neumann", "[1] Moore")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{THRESHOLD}{\code{number} - Minimum Threshold. A numeric value.}

\item{ALL}{\code{enum}  of \verb{("[0] single class", "[1] all classes")} - Class Selection. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CLASS}{\code{number} - Class Identifier. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Sieving classes (saga:sievingclasses)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Sieved Classes
}
}
}
