% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_simplefilterwithinshapes.R
\name{saga_simplefilterwithinshapes}
\alias{saga_simplefilterwithinshapes}
\title{QGIS algorithm Simple filter within shapes}
\usage{
saga_simplefilterwithinshapes(
  INPUT = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  SHAPES = qgisprocess::qgis_default_value(),
  MODE = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - Filtered Grid. Path for new raster layer.}

\item{SHAPES}{\code{source} - Boundaries. Path to a vector layer.}

\item{MODE}{\code{enum}  of \verb{("[0] Square", "[1] Circle")} - Search Mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("[0] Smooth", "[1] Sharpen", "[2] Edge")} - Filter. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radius. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Simple filter within shapes (saga:simplefilterwithinshapes)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Filtered Grid
}
}
}
