% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_singlevaluedecompositionopencv.R
\name{saga_singlevaluedecompositionopencv}
\alias{saga_singlevaluedecompositionopencv}
\title{QGIS algorithm Single value decomposition (opencv)}
\usage{
saga_singlevaluedecompositionopencv(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  RANGE_MIN = qgisprocess::qgis_default_value(),
  RANGE_MAX = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{RANGE_MIN}{\code{number} - Range (Min). A numeric value.}

\item{RANGE_MAX}{\code{number} - Range (Max). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Single value decomposition (opencv) (saga:singlevaluedecompositionopencv)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
