% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_sinkdrainageroutedetection.R
\name{saga_sinkdrainageroutedetection}
\alias{saga_sinkdrainageroutedetection}
\title{QGIS algorithm Sink drainage route detection}
\usage{
saga_sinkdrainageroutedetection(
  ELEVATION = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  THRSHEIGHT = qgisprocess::qgis_default_value(),
  SINKROUTE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{THRESHOLD}{\code{boolean} - Threshold. 1 for true/yes. 0 for false/no.}

\item{THRSHEIGHT}{\code{number} - Threshold Height. A numeric value.}

\item{SINKROUTE}{\code{rasterDestination} - Sink Route. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Sink drainage route detection (saga:sinkdrainageroutedetection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item SINKROUTE - outputRaster - Sink Route
}
}
}
