% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_skyviewfactor.R
\name{saga_skyviewfactor}
\alias{saga_skyviewfactor}
\title{QGIS algorithm Sky view factor}
\usage{
saga_skyviewfactor(
  DEM = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  DLEVEL = qgisprocess:::qgis_default_value(),
  NDIRS = qgisprocess:::qgis_default_value(),
  VISIBLE = qgisprocess:::qgis_default_value(),
  SVF = qgisprocess:::qgis_default_value(),
  SIMPLE = qgisprocess:::qgis_default_value(),
  TERRAIN = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{RADIUS}{\code{number} - Maximum Search Radius. A numeric value.}

\item{METHOD}{\code{enum}  of \verb{("[0] multi scale", "[1] sectors")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DLEVEL}{\code{number} - Multi Scale Factor. A numeric value.}

\item{NDIRS}{\code{number} - Number of Sectors. A numeric value.}

\item{VISIBLE}{\code{rasterDestination} - Visible Sky. Path for new raster layer.}

\item{SVF}{\code{rasterDestination} - Sky View Factor. Path for new raster layer.}

\item{SIMPLE}{\code{rasterDestination} - Sky View Factor (Simplified). Path for new raster layer.}

\item{TERRAIN}{\code{rasterDestination} - Terrain View Factor. Path for new raster layer.}

\item{DISTANCE}{\code{rasterDestination} - Terrain View Factor. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Sky view factor (saga:skyviewfactor)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DISTANCE - outputRaster - Terrain View Factor
\item SIMPLE - outputRaster - Sky View Factor (Simplified)
\item SVF - outputRaster - Sky View Factor
\item TERRAIN - outputRaster - Terrain View Factor
\item VISIBLE - outputRaster - Visible Sky
}
}
}
