% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_slopeaspectcurvature.R
\name{saga_slopeaspectcurvature}
\alias{saga_slopeaspectcurvature}
\title{QGIS algorithm Slope, aspect, curvature}
\usage{
saga_slopeaspectcurvature(
  ELEVATION = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  UNIT_SLOPE = qgisprocess::qgis_default_value(),
  UNIT_ASPECT = qgisprocess::qgis_default_value(),
  SLOPE = qgisprocess::qgis_default_value(),
  ASPECT = qgisprocess::qgis_default_value(),
  C_GENE = qgisprocess::qgis_default_value(),
  C_PLAN = qgisprocess::qgis_default_value(),
  C_PROF = qgisprocess::qgis_default_value(),
  C_TANG = qgisprocess::qgis_default_value(),
  C_LONG = qgisprocess::qgis_default_value(),
  C_CROS = qgisprocess::qgis_default_value(),
  C_MINI = qgisprocess::qgis_default_value(),
  C_MAXI = qgisprocess::qgis_default_value(),
  C_TOTA = qgisprocess::qgis_default_value(),
  C_ROTO = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Maximum Slope (Travis et al. 1975)", "[1] Maximum Triangle Slope (Tarboton 1997)", "[2] Least Squares Fitted Plane (Horn 1981, Costa-Cabral & Burgess 1996)", "[3] 6 parameter 2nd order polynom (Evans 1979)", "[4] 6 parameter 2nd order polynom (Heerdegen & Beran 1982)", "[5] 6 parameter 2nd order polynom (Bauer, Rohdenburg, Bork 1985)", "[6] 9 parameter 2nd order polynom (Zevenbergen & Thorne 1987)", "[7]10 parameter 3rd order polynom (Haralick 1983)")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{UNIT_SLOPE}{\code{enum}  of \verb{("[0] radians", "[1] degree", "[2] percent")} - Slope Units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{UNIT_ASPECT}{\code{enum}  of \verb{("[0] radians", "[1] degree")} - Aspect Units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SLOPE}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{ASPECT}{\code{rasterDestination} - Aspect. Path for new raster layer.}

\item{C_GENE}{\code{rasterDestination} - General Curvature. Path for new raster layer.}

\item{C_PLAN}{\code{rasterDestination} - Plan Curvature. Path for new raster layer.}

\item{C_PROF}{\code{rasterDestination} - Profile Curvature. Path for new raster layer.}

\item{C_TANG}{\code{rasterDestination} - Tangential Curvature. Path for new raster layer.}

\item{C_LONG}{\code{rasterDestination} - Longitudinal Curvature. Path for new raster layer.}

\item{C_CROS}{\code{rasterDestination} - Cross-Sectional Curvature. Path for new raster layer.}

\item{C_MINI}{\code{rasterDestination} - Minimal Curvature. Path for new raster layer.}

\item{C_MAXI}{\code{rasterDestination} - Maximal Curvature. Path for new raster layer.}

\item{C_TOTA}{\code{rasterDestination} - Total Curvature. Path for new raster layer.}

\item{C_ROTO}{\code{rasterDestination} - Flow-Line Curvature. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Slope, aspect, curvature (saga:slopeaspectcurvature)
}
\details{
\subsection{Outputs description}{
\itemize{
\item SLOPE - outputRaster - Slope
\item ASPECT - outputRaster - Aspect
\item C_GENE - outputRaster - General Curvature
\item C_PLAN - outputRaster - Plan Curvature
\item C_PROF - outputRaster - Profile Curvature
\item C_TANG - outputRaster - Tangential Curvature
\item C_LONG - outputRaster - Longitudinal Curvature
\item C_CROS - outputRaster - Cross
\item C_MINI - outputRaster - Minimal Curvature
\item C_MAXI - outputRaster - Maximal Curvature
\item C_TOTA - outputRaster - Total Curvature
\item C_ROTO - outputRaster - Flow
}
}
}
