% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_slopelength.R
\name{saga_slopelength}
\alias{saga_slopelength}
\title{QGIS algorithm Slope length}
\usage{
saga_slopelength(
  DEM = qgisprocess::qgis_default_value(),
  LENGTH = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{LENGTH}{\code{rasterDestination} - Slope Length. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Slope length (saga:slopelength)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LENGTH - outputRaster - Slope Length
}
}
}
