% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_slopelimitedflowaccumulation.R
\name{saga_slopelimitedflowaccumulation}
\alias{saga_slopelimitedflowaccumulation}
\title{QGIS algorithm Slope limited flow accumulation}
\usage{
saga_slopelimitedflowaccumulation(
  DEM = qgisprocess::qgis_default_value(),
  WEIGHT = qgisprocess::qgis_default_value(),
  FLOW = qgisprocess::qgis_default_value(),
  SLOPE_MIN = qgisprocess::qgis_default_value(),
  SLOPE_MAX = qgisprocess::qgis_default_value(),
  B_FLOW = qgisprocess::qgis_default_value(),
  T_FLOW_MIN = qgisprocess::qgis_default_value(),
  T_FLOW_MAX = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{WEIGHT}{\code{raster} - Weight. Path to a raster layer.}

\item{FLOW}{\code{rasterDestination} - Flow Accumulation. Path for new raster layer.}

\item{SLOPE_MIN}{\code{number} - Slope Minimum. A numeric value.}

\item{SLOPE_MAX}{\code{number} - Slope Threshold. A numeric value.}

\item{B_FLOW}{\code{boolean} - Use Flow Threshold. 1 for true/yes. 0 for false/no.}

\item{T_FLOW_MIN}{\code{number} - Flow Threshold (Min). A numeric value.}

\item{T_FLOW_MAX}{\code{number} - Flow Threshold (Max). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Slope limited flow accumulation (saga:slopelimitedflowaccumulation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLOW - outputRaster - Flow Accumulation
}
}
}
