% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_smoothingvigra.R
\name{saga_smoothingvigra}
\alias{saga_smoothingvigra}
\title{QGIS algorithm Smoothing (vigra)}
\usage{
saga_smoothingvigra(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  SCALE = qgisprocess::qgis_default_value(),
  EDGE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{TYPE}{\code{enum}  of \verb{("[0] exponential", "[1] nonlinear", "[2] gaussian")} - Type of smoothing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SCALE}{\code{number} - Size of smoothing filter. A numeric value.}

\item{EDGE}{\code{number} - Edge threshold for nonlinear smoothing. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Smoothing (vigra) (saga:smoothingvigra)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
