% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_snappointstogrid.R
\name{saga_snappointstogrid}
\alias{saga_snappointstogrid}
\title{QGIS algorithm Snap points to grid}
\usage{
saga_snappointstogrid(
  INPUT = qgisprocess:::qgis_default_value(),
  GRID = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  MOVES = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  SHAPE = qgisprocess:::qgis_default_value(),
  EXTREME = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Points. Path to a vector layer.}

\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{OUTPUT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{MOVES}{\code{vectorDestination} - Moves. Path for new vector layer.}

\item{DISTANCE}{\code{number} - Search Distance (Map Units). A numeric value.}

\item{SHAPE}{\code{enum}  of \verb{("[0] circle", "[1] square")} - Search Shape. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXTREME}{\code{enum}  of \verb{("[0] minimum", "[1] maximum")} - Extreme. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Snap points to grid (saga:snappointstogrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item MOVES - outputVector - Moves
\item OUTPUT - outputVector - Result
}
}
}
