% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_snappointstolines.R
\name{saga_snappointstolines}
\alias{saga_snappointstolines}
\title{QGIS algorithm Snap points to lines}
\usage{
saga_snappointstolines(
  INPUT = qgisprocess::qgis_default_value(),
  SNAP = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  MOVES = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Points. Path to a vector layer.}

\item{SNAP}{\code{source} - Snap Features. Path to a vector layer.}

\item{OUTPUT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{MOVES}{\code{vectorDestination} - Moves. Path for new vector layer.}

\item{DISTANCE}{\code{number} - Search Distance. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Snap points to lines (saga:snappointstolines)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Result
\item MOVES - outputVector - Moves
}
}
}
