% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_spatialpointpatternanalysis.R
\name{saga_spatialpointpatternanalysis}
\alias{saga_spatialpointpatternanalysis}
\title{QGIS algorithm Spatial point pattern analysis}
\usage{
saga_spatialpointpatternanalysis(
  POINTS = qgisprocess::qgis_default_value(),
  STEP = qgisprocess::qgis_default_value(),
  CENTRE = qgisprocess::qgis_default_value(),
  STDDIST = qgisprocess::qgis_default_value(),
  BBOX = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{STEP}{\code{number} - Vertex Distance (Degree). A numeric value.}

\item{CENTRE}{\code{vectorDestination} - Mean Centre. Path for new vector layer.}

\item{STDDIST}{\code{vectorDestination} - Standard Distance. Path for new vector layer.}

\item{BBOX}{\code{vectorDestination} - Bounding Box. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Spatial point pattern analysis (saga:spatialpointpatternanalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CENTRE - outputVector - Mean Centre
\item STDDIST - outputVector - Standard Distance
\item BBOX - outputVector - Bounding Box
}
}
}
