% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_splitlinesatpoints.R
\name{saga_splitlinesatpoints}
\alias{saga_splitlinesatpoints}
\title{QGIS algorithm Split lines at points}
\usage{
saga_splitlinesatpoints(
  LINES = qgisprocess::qgis_default_value(),
  SPLIT = qgisprocess::qgis_default_value(),
  INTERSECT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  EPSILON = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{SPLIT}{\code{source} - Split Features. Path to a vector layer.}

\item{INTERSECT}{\code{vectorDestination} - Intersection. Path for new vector layer.}

\item{OUTPUT}{\code{enum}  of \verb{("[0] polylines", "[1] separate lines")} - Output. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EPSILON}{\code{number} - Epsilon. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Split lines at points (saga:splitlinesatpoints)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INTERSECT - outputVector - Intersection
}
}
}
