% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_splitlineswithlines.R
\name{saga_splitlineswithlines}
\alias{saga_splitlineswithlines}
\title{QGIS algorithm Split lines with lines}
\usage{
saga_splitlineswithlines(
  LINES = qgisprocess::qgis_default_value(),
  SPLIT = qgisprocess::qgis_default_value(),
  INTERSECT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{SPLIT}{\code{source} - Split Features. Path to a vector layer.}

\item{INTERSECT}{\code{vectorDestination} - Intersection. Path for new vector layer.}

\item{OUTPUT}{\code{enum}  of \verb{("[0] polylines", "[1] separate lines")} - Output. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Split lines with lines (saga:splitlineswithlines)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INTERSECT - outputVector - Intersection
}
}
}
