% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_splitrgbbands.R
\name{saga_splitrgbbands}
\alias{saga_splitrgbbands}
\title{QGIS algorithm Split RGB bands}
\usage{
saga_splitrgbbands(
  INPUT = qgisprocess::qgis_default_value(),
  R = qgisprocess::qgis_default_value(),
  G = qgisprocess::qgis_default_value(),
  B = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{R}{\code{rasterDestination} - Output R band layer. Path for new raster layer.}

\item{G}{\code{rasterDestination} - Output G band layer. Path for new raster layer.}

\item{B}{\code{rasterDestination} - Output B band layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Split RGB bands (saga:splitrgbbands)
}
\details{
\subsection{Outputs description}{
\itemize{
\item B - outputRaster - Output B band layer
\item G - outputRaster - Output G band layer
\item R - outputRaster - Output R band layer
}
}
}
