% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_statisticsforrasters.R
\name{saga_statisticsforrasters}
\alias{saga_statisticsforrasters}
\title{QGIS algorithm Statistics for rasters}
\usage{
saga_statisticsforrasters(
  GRIDS = qgisprocess:::qgis_default_value(),
  PCTL_VAL = qgisprocess:::qgis_default_value(),
  MEAN = qgisprocess:::qgis_default_value(),
  MIN = qgisprocess:::qgis_default_value(),
  MAX = qgisprocess:::qgis_default_value(),
  VAR = qgisprocess:::qgis_default_value(),
  SUM = qgisprocess:::qgis_default_value(),
  RANGE = qgisprocess:::qgis_default_value(),
  PCTL = qgisprocess:::qgis_default_value(),
  STDDEV = qgisprocess:::qgis_default_value(),
  STDDEVLO = qgisprocess:::qgis_default_value(),
  STDDEVHI = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{PCTL_VAL}{\code{number} - Percentile. A numeric value.}

\item{MEAN}{\code{rasterDestination} - Arithmetic Mean. Path for new raster layer.}

\item{MIN}{\code{rasterDestination} - Minimum. Path for new raster layer.}

\item{MAX}{\code{rasterDestination} - Maximum. Path for new raster layer.}

\item{VAR}{\code{rasterDestination} - Variance. Path for new raster layer.}

\item{SUM}{\code{rasterDestination} - Sum. Path for new raster layer.}

\item{RANGE}{\code{rasterDestination} - Range. Path for new raster layer.}

\item{PCTL}{\code{rasterDestination} - Percentile. Path for new raster layer.}

\item{STDDEV}{\code{rasterDestination} - Standard Deviation. Path for new raster layer.}

\item{STDDEVLO}{\code{rasterDestination} - Mean less Standard Deviation. Path for new raster layer.}

\item{STDDEVHI}{\code{rasterDestination} - Mean plus Standard Deviation. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Statistics for rasters (saga:statisticsforrasters)
}
\details{
\subsection{Outputs description}{
\itemize{
\item MAX - outputRaster - Maximum
\item MEAN - outputRaster - Arithmetic Mean
\item MIN - outputRaster - Minimum
\item PCTL - outputRaster - Percentile
\item RANGE - outputRaster - Range
\item STDDEV - outputRaster - Standard Deviation
\item STDDEVHI - outputRaster - Mean plus Standard Deviation
\item STDDEVLO - outputRaster - Mean less Standard Deviation
\item SUM - outputRaster - Sum
\item VAR - outputRaster - Variance
}
}
}
