% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_streampowerindex.R
\name{saga_streampowerindex}
\alias{saga_streampowerindex}
\title{QGIS algorithm Stream power index}
\usage{
saga_streampowerindex(
  SLOPE = qgisprocess::qgis_default_value(),
  AREA = qgisprocess::qgis_default_value(),
  CONV = qgisprocess::qgis_default_value(),
  SPI = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SLOPE}{\code{raster} - Slope. Path to a raster layer.}

\item{AREA}{\code{raster} - Catchment Area. Path to a raster layer.}

\item{CONV}{\code{enum}  of \verb{("[0] no conversion (areas already given as specific catchment area)", "[1] 1 / cell size (pseudo specific catchment area)")} - Area Conversion. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SPI}{\code{rasterDestination} - Stream Power Index. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Stream power index (saga:streampowerindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item SPI - outputRaster - Stream Power Index
}
}
}
