% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_supervisedclassification.R
\name{saga_supervisedclassification}
\alias{saga_supervisedclassification}
\title{QGIS algorithm Supervised classification}
\usage{
saga_supervisedclassification(
  GRIDS = qgisprocess:::qgis_default_value(),
  ROI = qgisprocess:::qgis_default_value(),
  ROI_ID = qgisprocess:::qgis_default_value(),
  STATS = qgisprocess:::qgis_default_value(),
  STATS_SRC = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  NORMALISE = qgisprocess:::qgis_default_value(),
  THRESHOLD_DIST = qgisprocess:::qgis_default_value(),
  THRESHOLD_PROB = qgisprocess:::qgis_default_value(),
  RELATIVE_PROB = qgisprocess:::qgis_default_value(),
  THRESHOLD_ANGLE = qgisprocess:::qgis_default_value(),
  WTA_0 = qgisprocess:::qgis_default_value(),
  WTA_1 = qgisprocess:::qgis_default_value(),
  WTA_2 = qgisprocess:::qgis_default_value(),
  WTA_3 = qgisprocess:::qgis_default_value(),
  WTA_4 = qgisprocess:::qgis_default_value(),
  WTA_5 = qgisprocess:::qgis_default_value(),
  CLASS_INFO = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  QUALITY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{ROI}{\code{source} - Training Areas. Path to a vector layer.}

\item{ROI_ID}{\code{field} - Class Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{STATS}{\code{source} - Class Statistics. Path to a vector layer.}

\item{STATS_SRC}{\code{enum}  of \verb{("[0] training areas", "[1] table")} - Get Class Statistics from.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("[0] Binary Encoding", "[1] Parallelepiped", "[2] Minimum Distance", "[3] Mahalanobis Distance", "[4] Maximum Likelihood", "[5] Spectral Angle Mapping", " [6] Winner Takes All")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NORMALISE}{\code{boolean} - Normalise. 1 for true/yes. 0 for false/no.}

\item{THRESHOLD_DIST}{\code{number} - Distance Threshold. A numeric value.}

\item{THRESHOLD_PROB}{\code{number} - Probability Threshold (Percent). A numeric value.}

\item{RELATIVE_PROB}{\code{enum}  of \verb{("[0] absolute", "[1] relative")} - Probability Reference. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{THRESHOLD_ANGLE}{\code{number} - Spectral Angle Threshold (Degree). A numeric value.}

\item{WTA_0}{\code{boolean} - Binary Encoding. 1 for true/yes. 0 for false/no.}

\item{WTA_1}{\code{boolean} - Parallelepiped. 1 for true/yes. 0 for false/no.}

\item{WTA_2}{\code{boolean} - Minimum Distance. 1 for true/yes. 0 for false/no.}

\item{WTA_3}{\code{boolean} - Mahalanobis Distance. 1 for true/yes. 0 for false/no.}

\item{WTA_4}{\code{boolean} - Maximum Likelihood. 1 for true/yes. 0 for false/no.}

\item{WTA_5}{\code{boolean} - Spectral Angle Mapping. 1 for true/yes. 0 for false/no.}

\item{CLASS_INFO}{\code{vectorDestination} - Class Information. Path for new vector layer.}

\item{CLASSES}{\code{rasterDestination} - Classification. Path for new raster layer.}

\item{QUALITY}{\code{rasterDestination} - Quality. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Supervised classification (saga:supervisedclassification)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Classification
\item CLASS_INFO - outputVector - Class Information
\item QUALITY - outputRaster - Quality
}
}
}
