% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_supportvectormachineclassificationopencv.R
\name{saga_supportvectormachineclassificationopencv}
\alias{saga_supportvectormachineclassificationopencv}
\title{QGIS algorithm Support vector machine classification (opencv)}
\usage{
saga_supportvectormachineclassificationopencv(
  FEATURES = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  TRAIN_AREAS = qgisprocess:::qgis_default_value(),
  TRAIN_CLASS = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  SVM_TYPE = qgisprocess:::qgis_default_value(),
  C = qgisprocess:::qgis_default_value(),
  NU = qgisprocess:::qgis_default_value(),
  P = qgisprocess:::qgis_default_value(),
  KERNEL = qgisprocess:::qgis_default_value(),
  COEF0 = qgisprocess:::qgis_default_value(),
  DEGREE = qgisprocess:::qgis_default_value(),
  GAMMA = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FEATURES}{\code{multilayer} - Features. .}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no.}

\item{TRAIN_AREAS}{\code{source} - Training Areas. Path to a vector layer.}

\item{TRAIN_CLASS}{\code{field} - Class Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{CLASSES}{\code{rasterDestination} - Classification. Path for new raster layer.}

\item{SVM_TYPE}{\code{enum}  of \verb{("[0] c-support vector classification", "[1] nu support vector classification", "[2] distribution estimation (one class)", "[3] epsilon support vector regression", "[4] nu support vector regression")} - SVM Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{C}{\code{number} - C. A numeric value.}

\item{NU}{\code{number} - Nu. A numeric value.}

\item{P}{\code{number} - P. A numeric value.}

\item{KERNEL}{\code{enum}  of \verb{("[0] linear", "[1] polynomial", "[2] radial basis function", "[3] sigmoid", "[4] exponential chi2", "[5] histogram intersection")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{COEF0}{\code{number} - Coefficient 0. A numeric value.}

\item{DEGREE}{\code{number} - Degree. A numeric value.}

\item{GAMMA}{\code{number} - Gamma. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Support vector machine classification (opencv) (saga:supportvectormachineclassificationopencv)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Classification
}
}
}
