% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_surfacegradientandconcentration.R
\name{saga_surfacegradientandconcentration}
\alias{saga_surfacegradientandconcentration}
\title{QGIS algorithm Surface, gradient and concentration}
\usage{
saga_surfacegradientandconcentration(
  MASK = qgisprocess::qgis_default_value(),
  SURF = qgisprocess::qgis_default_value(),
  GRAD = qgisprocess::qgis_default_value(),
  CONC = qgisprocess::qgis_default_value(),
  SURF_E = qgisprocess::qgis_default_value(),
  CONC_IN = qgisprocess::qgis_default_value(),
  CONC_OUT = qgisprocess::qgis_default_value(),
  CONC_E = qgisprocess::qgis_default_value(),
  GRAD_MIN = qgisprocess::qgis_default_value(),
  NEIGHBOURS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{MASK}{\code{raster} - Mask. Path to a raster layer.}

\item{SURF}{\code{rasterDestination} - Surface. Path for new raster layer.}

\item{GRAD}{\code{rasterDestination} - Gradient. Path for new raster layer.}

\item{CONC}{\code{rasterDestination} - Concentration. Path for new raster layer.}

\item{SURF_E}{\code{number} - Surface Approximation Threshold. A numeric value.}

\item{CONC_IN}{\code{number} - Inlet Concentration. A numeric value.}

\item{CONC_OUT}{\code{number} - Outlet Concentration. A numeric value.}

\item{CONC_E}{\code{number} - Concentration Approximation Threshold. A numeric value.}

\item{GRAD_MIN}{\code{number} - Minimum Gradient. A numeric value.}

\item{NEIGHBOURS}{\code{enum}  of \verb{("[0] Moore (8)", "[1] Neumann (4)", "[2] Optimised")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Surface, gradient and concentration (saga:surfacegradientandconcentration)
}
\details{
\subsection{Outputs description}{
\itemize{
\item SURF - outputRaster - Surface
\item GRAD - outputRaster - Gradient
\item CONC - outputRaster - Concentration
}
}
}
