% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_surfacespecificpoints.R
\name{saga_surfacespecificpoints}
\alias{saga_surfacespecificpoints}
\title{QGIS algorithm Surface specific points}
\usage{
saga_surfacespecificpoints(
  ELEVATION = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Mark Highest Neighbour", "[1] Opposite Neighbours", "[2] Flow Direction", "[3] Flow Direction (up and down)", "[4] Peucker & Douglas")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{THRESHOLD}{\code{number} - Threshold. A numeric value.}

\item{RESULT}{\code{rasterDestination} - Result. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Surface specific points (saga:surfacespecificpoints)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Result
}
}
}
