% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_svmclassification.R
\name{saga_svmclassification}
\alias{saga_svmclassification}
\title{QGIS algorithm Svm classification}
\usage{
saga_svmclassification(
  GRIDS = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  SCALING = qgisprocess:::qgis_default_value(),
  MESSAGE = qgisprocess:::qgis_default_value(),
  MODEL_SRC = qgisprocess:::qgis_default_value(),
  MODEL_LOAD = qgisprocess:::qgis_default_value(),
  ROI = qgisprocess:::qgis_default_value(),
  ROI_ID = qgisprocess:::qgis_default_value(),
  MODEL_SAVE = qgisprocess:::qgis_default_value(),
  SVM_TYPE = qgisprocess:::qgis_default_value(),
  KERNEL_TYPE = qgisprocess:::qgis_default_value(),
  DEGREE = qgisprocess:::qgis_default_value(),
  GAMMA = qgisprocess:::qgis_default_value(),
  COEF0 = qgisprocess:::qgis_default_value(),
  COST = qgisprocess:::qgis_default_value(),
  NU = qgisprocess:::qgis_default_value(),
  EPS_SVR = qgisprocess:::qgis_default_value(),
  CACHE_SIZE = qgisprocess:::qgis_default_value(),
  EPS = qgisprocess:::qgis_default_value(),
  SHRINKING = qgisprocess:::qgis_default_value(),
  PROBABILITY = qgisprocess:::qgis_default_value(),
  CROSSVAL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{CLASSES}{\code{rasterDestination} - Classification. Path for new raster layer.}

\item{SCALING}{\code{enum}  of \verb{("[0] none", "[1] normalize (0-1)", "[2] standardize")} - Scaling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MESSAGE}{\code{boolean} - Verbose Messages. 1 for true/yes. 0 for false/no.}

\item{MODEL_SRC}{\code{enum}  of \verb{("[0] create from training areas", "[1] restore from file")} - Model Source. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MODEL_LOAD}{\code{file} - Restore Model from File. Path to a file.}

\item{ROI}{\code{source} - Training Areas. Path to a vector layer.}

\item{ROI_ID}{\code{field} - Class Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{MODEL_SAVE}{\code{file} - Store Model to File. Path to a file.}

\item{SVM_TYPE}{\code{enum}  of \verb{("[0] C-SVC", "[1] nu-SVC", "[2] one-class SVM", "[3] epsilon-SVR", "[4] nu-SVR")} - SVM Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_TYPE}{\code{enum}  of \verb{("[0] linear", "[1] polynomial", "[2] radial basis function", "[3] sigmoid")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DEGREE}{\code{number} - Degree. A numeric value.}

\item{GAMMA}{\code{number} - Gamma. A numeric value.}

\item{COEF0}{\code{number} - coef0. A numeric value.}

\item{COST}{\code{number} - C. A numeric value.}

\item{NU}{\code{number} - nu-SVR. A numeric value.}

\item{EPS_SVR}{\code{number} - SVR Epsilon. A numeric value.}

\item{CACHE_SIZE}{\code{number} - Cache Size. A numeric value.}

\item{EPS}{\code{number} - Epsilon. A numeric value.}

\item{SHRINKING}{\code{boolean} - Shrinking. 1 for true/yes. 0 for false/no.}

\item{PROBABILITY}{\code{boolean} - Probability Estimates. 1 for true/yes. 0 for false/no.}

\item{CROSSVAL}{\code{number} - Cross Validation. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Svm classification (saga:svmclassification)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Classification
}
}
}
