% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_tasseledcaptransformation.R
\name{saga_tasseledcaptransformation}
\alias{saga_tasseledcaptransformation}
\title{QGIS algorithm Tasseled cap transformation}
\usage{
saga_tasseledcaptransformation(
  BLUE = qgisprocess::qgis_default_value(),
  RED = qgisprocess::qgis_default_value(),
  GREEN = qgisprocess::qgis_default_value(),
  NIR = qgisprocess::qgis_default_value(),
  MIR1 = qgisprocess::qgis_default_value(),
  MIR2 = qgisprocess::qgis_default_value(),
  BRIGHTNESS = qgisprocess::qgis_default_value(),
  GREENNESS = qgisprocess::qgis_default_value(),
  WETNESS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{BLUE}{\code{raster} - Blue (TM 1). Path to a raster layer.}

\item{RED}{\code{raster} - Red (TM 2). Path to a raster layer.}

\item{GREEN}{\code{raster} - Green (TM 3). Path to a raster layer.}

\item{NIR}{\code{raster} - Near Infrared (TM 4). Path to a raster layer.}

\item{MIR1}{\code{raster} - Mid Infrared (TM 5). Path to a raster layer.}

\item{MIR2}{\code{raster} - Mid Infrared (TM 7). Path to a raster layer.}

\item{BRIGHTNESS}{\code{rasterDestination} - Brightness. Path for new raster layer.}

\item{GREENNESS}{\code{rasterDestination} - Greenness. Path for new raster layer.}

\item{WETNESS}{\code{rasterDestination} - Wetness. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Tasseled cap transformation (saga:tasseledcaptransformation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item BRIGHTNESS - outputRaster - Brightness
\item GREENNESS - outputRaster - Greenness
\item WETNESS - outputRaster - Wetness
}
}
}
