% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_tcilow.R
\name{saga_tcilow}
\alias{saga_tcilow}
\title{QGIS algorithm Tci low}
\usage{
saga_tcilow(
  DISTANCE = qgisprocess:::qgis_default_value(),
  TWI = qgisprocess:::qgis_default_value(),
  TCILOW = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DISTANCE}{\code{raster} - Vertical Distance to Channel Network. Path to a raster layer.}

\item{TWI}{\code{raster} - Topographic Wetness Index. Path to a raster layer.}

\item{TCILOW}{\code{rasterDestination} - TCI Low. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Tci low (saga:tcilow)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TCILOW - outputRaster - TCI Low
}
}
}
