% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_terrainmapview.R
\name{saga_terrainmapview}
\alias{saga_terrainmapview}
\title{QGIS algorithm Terrain map view}
\usage{
saga_terrainmapview(
  DEM = qgisprocess:::qgis_default_value(),
  SHADE = qgisprocess:::qgis_default_value(),
  OPENNESS = qgisprocess:::qgis_default_value(),
  SLOPE = qgisprocess:::qgis_default_value(),
  CONTOURS = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  CONTOUR_LINES = qgisprocess:::qgis_default_value(),
  EQUIDISTANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{SHADE}{\code{rasterDestination} - Shade. Path for new raster layer.}

\item{OPENNESS}{\code{rasterDestination} - Openness. Path for new raster layer.}

\item{SLOPE}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{CONTOURS}{\code{vectorDestination} - Contours. Path for new vector layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Topography", "[1] Morphology")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radial Limit. A numeric value.}

\item{CONTOUR_LINES}{\code{boolean} - Contour Lines. 1 for true/yes. 0 for false/no.}

\item{EQUIDISTANCE}{\code{number} - Equidistance. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Terrain map view (saga:terrainmapview)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONTOURS - outputVector - Contours
\item OPENNESS - outputRaster - Openness
\item SHADE - outputRaster - Shade
\item SLOPE - outputRaster - Slope
}
}
}
