% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/saga_terrainsurfaceclassificationiwahashiandpike.R
\name{saga_terrainsurfaceclassificationiwahashiandpike}
\alias{saga_terrainsurfaceclassificationiwahashiandpike}
\title{QGIS algorithm Terrain surface classification (iwahashi and pike)}
\usage{
saga_terrainsurfaceclassificationiwahashiandpike(
  DEM = qgisprocess::qgis_default_value(),
  SLOPE = qgisprocess::qgis_default_value(),
  CONVEXITY = qgisprocess::qgis_default_value(),
  CONV_RECALC = qgisprocess::qgis_default_value(),
  TEXTURE = qgisprocess::qgis_default_value(),
  TEXT_RECALC = qgisprocess::qgis_default_value(),
  LANDFORMS = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  CONV_SCALE = qgisprocess::qgis_default_value(),
  CONV_KERNEL = qgisprocess::qgis_default_value(),
  CONV_TYPE = qgisprocess::qgis_default_value(),
  CONV_EPSILON = qgisprocess::qgis_default_value(),
  TEXT_SCALE = qgisprocess::qgis_default_value(),
  TEXT_EPSILON = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{SLOPE}{\code{raster} - Slope. Path to a raster layer.}

\item{CONVEXITY}{\code{raster} - Convexity. Path to a raster layer.}

\item{CONV_RECALC}{\code{boolean} - Recalculate. 1 for true/yes. 0 for false/no.}

\item{TEXTURE}{\code{raster} - Texture. Path to a raster layer.}

\item{TEXT_RECALC}{\code{boolean} - Recalculate. 1 for true/yes. 0 for false/no.}

\item{LANDFORMS}{\code{rasterDestination} - Landforms. Path for new raster layer.}

\item{TYPE}{\code{enum}  of \verb{("[0] 8", "[1] 12", "[2] 16")} - Number of Classes. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONV_SCALE}{\code{number} - Scale (Cells). A numeric value.}

\item{CONV_KERNEL}{\code{enum}  of \verb{("[0] four-neighbourhood", "[1] eight-neihbourhood", "[2] eight-neihbourhood (distance based weighting)")} - Laplacian Filter Kernel. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONV_TYPE}{\code{enum}  of \verb{("[0] convexity", "[1] concavity")} - Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONV_EPSILON}{\code{number} - Flat Area Threshold. A numeric value.}

\item{TEXT_SCALE}{\code{number} - Scale (Cells). A numeric value.}

\item{TEXT_EPSILON}{\code{number} - Flat Area Threshold. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Terrain surface classification (iwahashi and pike) (saga:terrainsurfaceclassificationiwahashiandpike)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LANDFORMS - outputRaster - Landforms
}
}
}
