% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_terrainsurfaceconvexity.R
\name{saga_terrainsurfaceconvexity}
\alias{saga_terrainsurfaceconvexity}
\title{QGIS algorithm Terrain surface convexity}
\usage{
saga_terrainsurfaceconvexity(
  DEM = qgisprocess::qgis_default_value(),
  CONVEXITY = qgisprocess::qgis_default_value(),
  KERNEL = qgisprocess::qgis_default_value(),
  TYPE = qgisprocess::qgis_default_value(),
  EPSILON = qgisprocess::qgis_default_value(),
  SCALE = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  DW_WEIGHTING = qgisprocess::qgis_default_value(),
  DW_IDW_POWER = qgisprocess::qgis_default_value(),
  DW_IDW_OFFSET = qgisprocess::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{CONVEXITY}{\code{rasterDestination} - Convexity. Path for new raster layer.}

\item{KERNEL}{\code{enum}  of \verb{("[0] conventional four-neighbourhood", "[1] conventional eight-neihbourhood", "[2] eight-neihbourhood (distance based weighting)")} - Laplacian Filter Kernel. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TYPE}{\code{enum}  of \verb{("[0] convexity", "[1] concavity")} - Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EPSILON}{\code{number} - Flat Area Threshold. A numeric value.}

\item{SCALE}{\code{number} - Scale (Cells). A numeric value.}

\item{METHOD}{\code{enum}  of \verb{("[0] counting cells", "[1] resampling")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DW_IDW_OFFSET}{\code{boolean} - Inverse Distance Offset. 1 for true/yes. 0 for false/no.}

\item{DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Terrain surface convexity (saga:terrainsurfaceconvexity)
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONVEXITY - outputRaster - Convexity
}
}
}
