% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_thinplatespline.R
\name{saga_thinplatespline}
\alias{saga_thinplatespline}
\title{QGIS algorithm Thin plate spline}
\usage{
saga_thinplatespline(
  SHAPES = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  REGULARISATION = qgisprocess::qgis_default_value(),
  SEARCH_RANGE = qgisprocess::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_ALL = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MIN = qgisprocess::qgis_default_value(),
  SEARCH_POINTS_MAX = qgisprocess::qgis_default_value(),
  SEARCH_DIRECTION = qgisprocess::qgis_default_value(),
 
    TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX = qgisprocess::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess::qgis_default_value(),
  TARGET_USER_FITS = qgisprocess::qgis_default_value(),
  TARGET_TEMPLATE = qgisprocess::qgis_default_value(),
  TARGET_OUT_GRID = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SHAPES}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{REGULARISATION}{\code{number} - Regularisation. A numeric value.}

\item{SEARCH_RANGE}{\code{enum}  of \verb{("[0] local", "[1] global")} - Search Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_RADIUS}{\code{number} - Search Radius. A numeric value.}

\item{SEARCH_POINTS_ALL}{\code{enum}  of \verb{("[0] maximum number of nearest points", "[1] all points within search distance")} - Number of Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_POINTS_MIN}{\code{number} - Maximum Number of Points. A numeric value.}

\item{SEARCH_POINTS_MAX}{\code{number} - Maximum Number of Points. A numeric value.}

\item{SEARCH_DIRECTION}{\code{enum}  of \verb{("[0] all directions", "[1] quadrants")} - Search Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{TARGET_USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_TEMPLATE}{\code{raster} - Target system. Path to a raster layer.}

\item{TARGET_OUT_GRID}{\code{rasterDestination} - Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Thin plate spline (saga:thinplatespline)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TARGET_OUT_GRID - outputRaster - Grid
}
}
}
