% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_thinplatesplinetin.R
\name{saga_thinplatesplinetin}
\alias{saga_thinplatesplinetin}
\title{QGIS algorithm Thin plate spline (tin)}
\usage{
saga_thinplatesplinetin(
  SHAPES = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  REGULARISATION = qgisprocess::qgis_default_value(),
  LEVEL = qgisprocess::qgis_default_value(),
  FRAME = qgisprocess::qgis_default_value(),
  TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX =
    qgisprocess::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess::qgis_default_value(),
  TARGET_USER_FITS = qgisprocess::qgis_default_value(),
  TARGET_OUT_GRID = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{REGULARISATION}{\code{number} - Regularisation. A numeric value.}

\item{LEVEL}{\code{enum}  of \verb{("[0] immediate", "[1] level 1", "[2] level 2")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FRAME}{\code{boolean} - Add Frame. 1 for true/yes. 0 for false/no.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value.}

\item{TARGET_USER_FITS}{\code{enum}  of \verb{("[0] nodes", "[1] cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_OUT_GRID}{\code{rasterDestination} - Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Thin plate spline (tin) (saga:thinplatesplinetin)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TARGET_OUT_GRID - outputRaster - Grid
}
}
}
