% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_topofatmospherereflectance.R
\name{saga_topofatmospherereflectance}
\alias{saga_topofatmospherereflectance}
\title{QGIS algorithm Top of atmosphere reflectance}
\usage{
saga_topofatmospherereflectance(
  DN_MSS01 = qgisprocess::qgis_default_value(),
  DN_MSS02 = qgisprocess::qgis_default_value(),
  DN_MSS03 = qgisprocess::qgis_default_value(),
  DN_MSS04 = qgisprocess::qgis_default_value(),
  RF_MSS01 = qgisprocess::qgis_default_value(),
  RF_MSS02 = qgisprocess::qgis_default_value(),
  RF_MSS03 = qgisprocess::qgis_default_value(),
  RF_MSS04 = qgisprocess::qgis_default_value(),
  DN_ETM01 = qgisprocess::qgis_default_value(),
  DN_ETM02 = qgisprocess::qgis_default_value(),
  DN_ETM03 = qgisprocess::qgis_default_value(),
  DN_ETM04 = qgisprocess::qgis_default_value(),
  DN_ETM05 = qgisprocess::qgis_default_value(),
  DN_ETM07 = qgisprocess::qgis_default_value(),
  RF_ETM01 = qgisprocess::qgis_default_value(),
  RF_ETM02 = qgisprocess::qgis_default_value(),
  RF_ETM03 = qgisprocess::qgis_default_value(),
  RF_ETM04 = qgisprocess::qgis_default_value(),
  RF_ETM05 = qgisprocess::qgis_default_value(),
  RF_ETM07 = qgisprocess::qgis_default_value(),
  DN_OLI01 = qgisprocess::qgis_default_value(),
  DN_OLI02 = qgisprocess::qgis_default_value(),
  DN_OLI03 = qgisprocess::qgis_default_value(),
  DN_OLI04 = qgisprocess::qgis_default_value(),
  DN_OLI05 = qgisprocess::qgis_default_value(),
  DN_OLI06 = qgisprocess::qgis_default_value(),
  DN_OLI07 = qgisprocess::qgis_default_value(),
  DN_OLI09 = qgisprocess::qgis_default_value(),
  RF_OLI01 = qgisprocess::qgis_default_value(),
  RF_OLI02 = qgisprocess::qgis_default_value(),
  RF_OLI03 = qgisprocess::qgis_default_value(),
  RF_OLI04 = qgisprocess::qgis_default_value(),
  RF_OLI05 = qgisprocess::qgis_default_value(),
  RF_OLI06 = qgisprocess::qgis_default_value(),
  RF_OLI07 = qgisprocess::qgis_default_value(),
  RF_OLI09 = qgisprocess::qgis_default_value(),
  DN__TM06 = qgisprocess::qgis_default_value(),
  RF__TM06 = qgisprocess::qgis_default_value(),
  DN_ETM61 = qgisprocess::qgis_default_value(),
  DN_ETM62 = qgisprocess::qgis_default_value(),
  RF_ETM61 = qgisprocess::qgis_default_value(),
  RF_ETM62 = qgisprocess::qgis_default_value(),
  DN_OLI10 = qgisprocess::qgis_default_value(),
  DN_OLI11 = qgisprocess::qgis_default_value(),
  RF_OLI10 = qgisprocess::qgis_default_value(),
  RF_OLI11 = qgisprocess::qgis_default_value(),
  DN_PAN08 = qgisprocess::qgis_default_value(),
  RF_PAN08 = qgisprocess::qgis_default_value(),
  METAFILE = qgisprocess::qgis_default_value(),
  SENSOR = qgisprocess::qgis_default_value(),
  DATE_ACQU = qgisprocess::qgis_default_value(),
  DATE_PROD = qgisprocess::qgis_default_value(),
  SUN_HGT = qgisprocess::qgis_default_value(),
  AS_RAD = qgisprocess::qgis_default_value(),
  AC_METHOD = qgisprocess::qgis_default_value(),
  AC_DO_CELLS = qgisprocess::qgis_default_value(),
  AC_RAYLEIGH = qgisprocess::qgis_default_value(),
  AC_SUN_RAD = qgisprocess::qgis_default_value(),
  ETM_GAIN_10 = qgisprocess::qgis_default_value(),
  ETM_GAIN_20 = qgisprocess::qgis_default_value(),
  ETM_GAIN_30 = qgisprocess::qgis_default_value(),
  ETM_GAIN_40 = qgisprocess::qgis_default_value(),
  ETM_GAIN_50 = qgisprocess::qgis_default_value(),
  ETM_GAIN_61 = qgisprocess::qgis_default_value(),
  ETM_GAIN_62 = qgisprocess::qgis_default_value(),
  ETM_GAIN_70 = qgisprocess::qgis_default_value(),
  ETM_GAIN_80 = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DN_MSS01}{\code{raster} - DN Band 1. Path to a raster layer.}

\item{DN_MSS02}{\code{raster} - DN Band 2. Path to a raster layer.}

\item{DN_MSS03}{\code{raster} - DN Band 3. Path to a raster layer.}

\item{DN_MSS04}{\code{raster} - DN Band 4. Path to a raster layer.}

\item{RF_MSS01}{\code{rasterDestination} - Reflectance Band 1. Path for new raster layer.}

\item{RF_MSS02}{\code{rasterDestination} - Reflectance Band 2. Path for new raster layer.}

\item{RF_MSS03}{\code{rasterDestination} - Reflectance Band 3. Path for new raster layer.}

\item{RF_MSS04}{\code{rasterDestination} - Reflectance Band 4. Path for new raster layer.}

\item{DN_ETM01}{\code{raster} - DN Band 1. Path to a raster layer.}

\item{DN_ETM02}{\code{raster} - DN Band 2. Path to a raster layer.}

\item{DN_ETM03}{\code{raster} - DN Band 3. Path to a raster layer.}

\item{DN_ETM04}{\code{raster} - DN Band 4. Path to a raster layer.}

\item{DN_ETM05}{\code{raster} - DN Band 5. Path to a raster layer.}

\item{DN_ETM07}{\code{raster} - DN Band 7. Path to a raster layer.}

\item{RF_ETM01}{\code{rasterDestination} - Reflectance Band 1. Path for new raster layer.}

\item{RF_ETM02}{\code{rasterDestination} - Reflectance Band 2. Path for new raster layer.}

\item{RF_ETM03}{\code{rasterDestination} - Reflectance Band 3. Path for new raster layer.}

\item{RF_ETM04}{\code{rasterDestination} - Reflectance Band 4. Path for new raster layer.}

\item{RF_ETM05}{\code{rasterDestination} - Reflectance Band 5. Path for new raster layer.}

\item{RF_ETM07}{\code{rasterDestination} - Reflectance Band 7. Path for new raster layer.}

\item{DN_OLI01}{\code{raster} - DN Band 1. Path to a raster layer.}

\item{DN_OLI02}{\code{raster} - DN Band 2. Path to a raster layer.}

\item{DN_OLI03}{\code{raster} - DN Band 3. Path to a raster layer.}

\item{DN_OLI04}{\code{raster} - DN Band 4. Path to a raster layer.}

\item{DN_OLI05}{\code{raster} - DN Band 5. Path to a raster layer.}

\item{DN_OLI06}{\code{raster} - DN Band 6. Path to a raster layer.}

\item{DN_OLI07}{\code{raster} - DN Band 7. Path to a raster layer.}

\item{DN_OLI09}{\code{raster} - DN Band 9. Path to a raster layer.}

\item{RF_OLI01}{\code{rasterDestination} - Reflectance Band 1. Path for new raster layer.}

\item{RF_OLI02}{\code{rasterDestination} - Reflectance Band 2. Path for new raster layer.}

\item{RF_OLI03}{\code{rasterDestination} - Reflectance Band 3. Path for new raster layer.}

\item{RF_OLI04}{\code{rasterDestination} - Reflectance Band 4. Path for new raster layer.}

\item{RF_OLI05}{\code{rasterDestination} - Reflectance Band 5. Path for new raster layer.}

\item{RF_OLI06}{\code{rasterDestination} - Reflectance Band 6. Path for new raster layer.}

\item{RF_OLI07}{\code{rasterDestination} - Reflectance Band 7. Path for new raster layer.}

\item{RF_OLI09}{\code{rasterDestination} - Reflectance Band 9. Path for new raster layer.}

\item{DN__TM06}{\code{raster} - DN Band 6. Path to a raster layer.}

\item{RF__TM06}{\code{rasterDestination} - Reflectance Band 6. Path for new raster layer.}

\item{DN_ETM61}{\code{raster} - DN Band 61. Path to a raster layer.}

\item{DN_ETM62}{\code{raster} - DN Band 62. Path to a raster layer.}

\item{RF_ETM61}{\code{rasterDestination} - Reflectance Band 61. Path for new raster layer.}

\item{RF_ETM62}{\code{rasterDestination} - Reflectance Band 62. Path for new raster layer.}

\item{DN_OLI10}{\code{raster} - DN Band 10. Path to a raster layer.}

\item{DN_OLI11}{\code{raster} - DN Band 11. Path to a raster layer.}

\item{RF_OLI10}{\code{rasterDestination} - Reflectance Band 10. Path for new raster layer.}

\item{RF_OLI11}{\code{rasterDestination} - Reflectance Band 11. Path for new raster layer.}

\item{DN_PAN08}{\code{raster} - DN Band 8. Path to a raster layer.}

\item{RF_PAN08}{\code{rasterDestination} - Reflectance Band 8. Path for new raster layer.}

\item{METAFILE}{\code{file} - Metadata File. Path to a file.}

\item{SENSOR}{\code{enum}  of \verb{("[0] Landsat-1 MSS", "[1] Landsat-2 MSS", "[2] Landsat-3 MSS", "[3] Landsat-4 MSS", "[4] Landsat-5 MSS", "[5] Landsat-4 TM", "[6] Landsat-5 TM", "[7] Landsat-7 ETM+", "[8] Landsat-8 OLI/TIRS")} - Spacecraft Sensor. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DATE_ACQU}{\code{string} - Image Acquisition Date. String value.}

\item{DATE_PROD}{\code{string} - Image Creation Date. String value.}

\item{SUN_HGT}{\code{number} - Suns's Height. A numeric value.}

\item{AS_RAD}{\code{boolean} - At-Sensor Radiance. 1 for true/yes. 0 for false/no.}

\item{AC_METHOD}{\code{enum}  of \verb{("[0] uncorrected", "[1] corrected", "[2] dark object subtraction 1", "[3] dark object subtraction 2", "[4] dark object subtraction 2b", "[5] dark object subtraction 3", "[6] dark object subtraction 4")} - Atmospheric Correction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{AC_DO_CELLS}{\code{number} - Minimum Number of Dark Object Cells. A numeric value.}

\item{AC_RAYLEIGH}{\code{number} - Rayleigh Scattering. A numeric value.}

\item{AC_SUN_RAD}{\code{number} - Solar Radiance. A numeric value.}

\item{ETM_GAIN_10}{\code{enum}  of \verb{("[0] low", "[1] high")} - Band 1. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ETM_GAIN_20}{\code{enum}  of \verb{("[0] low", "[1] high")} - Band 2. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ETM_GAIN_30}{\code{enum}  of \verb{("[0] low", "[1] high")} - Band 3. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ETM_GAIN_40}{\code{enum}  of \verb{("[0] low", "[1] high")} - Band 4. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ETM_GAIN_50}{\code{enum}  of \verb{("[0] low", "[1] high")} - Band 5. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ETM_GAIN_61}{\code{enum}  of \verb{("[0] low", "[1] high")} - Band 61. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ETM_GAIN_62}{\code{enum}  of \verb{("[0] low", "[1] high")} - Band 62. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ETM_GAIN_70}{\code{enum}  of \verb{("[0] low", "[1] high")} - Band 7. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ETM_GAIN_80}{\code{enum}  of \verb{("[0] low", "[1] high")} - Band 8. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Top of atmosphere reflectance (saga:topofatmospherereflectance)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RF_MSS01 - outputRaster - Reflectance Band 1
\item RF_MSS02 - outputRaster - Reflectance Band 2
\item RF_MSS03 - outputRaster - Reflectance Band 3
\item RF_MSS04 - outputRaster - Reflectance Band 4
\item RF_ETM01 - outputRaster - Reflectance Band 1
\item RF_ETM02 - outputRaster - Reflectance Band 2
\item RF_ETM03 - outputRaster - Reflectance Band 3
\item RF_ETM04 - outputRaster - Reflectance Band 4
\item RF_ETM05 - outputRaster - Reflectance Band 5
\item RF_ETM07 - outputRaster - Reflectance Band 7
\item RF_OLI01 - outputRaster - Reflectance Band 1
\item RF_OLI02 - outputRaster - Reflectance Band 2
\item RF_OLI03 - outputRaster - Reflectance Band 3
\item RF_OLI04 - outputRaster - Reflectance Band 4
\item RF_OLI05 - outputRaster - Reflectance Band 5
\item RF_OLI06 - outputRaster - Reflectance Band 6
\item RF_OLI07 - outputRaster - Reflectance Band 7
\item RF_OLI09 - outputRaster - Reflectance Band 9
\item RF__TM06 - outputRaster - Reflectance Band 6
\item RF_ETM61 - outputRaster - Reflectance Band 61
\item RF_ETM62 - outputRaster - Reflectance Band 62
\item RF_OLI10 - outputRaster - Reflectance Band 10
\item RF_OLI11 - outputRaster - Reflectance Band 11
\item RF_PAN08 - outputRaster - Reflectance Band 8
}
}
}
