% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_topographicopenness.R
\name{saga_topographicopenness}
\alias{saga_topographicopenness}
\title{QGIS algorithm Topographic openness}
\usage{
saga_topographicopenness(
  DEM = qgisprocess::qgis_default_value(),
  POS = qgisprocess::qgis_default_value(),
  NEG = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  DLEVEL = qgisprocess::qgis_default_value(),
  NDIRS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{POS}{\code{rasterDestination} - Positive Openness. Path for new raster layer.}

\item{NEG}{\code{rasterDestination} - Negative Openness. Path for new raster layer.}

\item{RADIUS}{\code{number} - Radial Limit. A numeric value.}

\item{METHOD}{\code{enum}  of \verb{("[0] multi scale", "[1] sectors")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DLEVEL}{\code{number} - Multi Scale Factor. A numeric value.}

\item{NDIRS}{\code{number} - Number of Sectors. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Topographic openness (saga:topographicopenness)
}
\details{
\subsection{Outputs description}{
\itemize{
\item NEG - outputRaster - Negative Openness
\item POS - outputRaster - Positive Openness
}
}
}
