% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_tpibasedlandformclassification.R
\name{saga_tpibasedlandformclassification}
\alias{saga_tpibasedlandformclassification}
\title{QGIS algorithm Tpi based landform classification}
\usage{
saga_tpibasedlandformclassification(
  DEM = qgisprocess::qgis_default_value(),
  RADIUS_A_MIN = qgisprocess::qgis_default_value(),
  RADIUS_A_MAX = qgisprocess::qgis_default_value(),
  RADIUS_B_MIN = qgisprocess::qgis_default_value(),
  RADIUS_B_MAX = qgisprocess::qgis_default_value(),
  DW_WEIGHTING = qgisprocess::qgis_default_value(),
  DW_IDW_POWER = qgisprocess::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess::qgis_default_value(),
  LANDFORMS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{RADIUS_A_MIN}{\code{number} - Min Radius A. A numeric value.}

\item{RADIUS_A_MAX}{\code{number} - Max Radius A. A numeric value.}

\item{RADIUS_B_MIN}{\code{number} - Min Radius B. A numeric value.}

\item{RADIUS_B_MAX}{\code{number} - Max Radius B. A numeric value.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("[0] no distance weighting", "[1] inverse distance to a power", "[2] exponential", "[3] gaussian weighting")} - Distance Weighting. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Inverse Distance Weighting Power. A numeric value.}

\item{DW_BANDWIDTH}{\code{number} - Gaussian and Exponential Weighting Bandwidth. A numeric value.}

\item{LANDFORMS}{\code{rasterDestination} - Landforms. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Tpi based landform classification (saga:tpibasedlandformclassification)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LANDFORMS - outputRaster - Landforms
}
}
}
