% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_transectthroughpolygonshapefile.R
\name{saga_transectthroughpolygonshapefile}
\alias{saga_transectthroughpolygonshapefile}
\title{QGIS algorithm Transect through polygon shapefile}
\usage{
saga_transectthroughpolygonshapefile(
  TRANSECT = qgisprocess::qgis_default_value(),
  THEME = qgisprocess::qgis_default_value(),
  THEME_FIELD = qgisprocess::qgis_default_value(),
  TRANSECT_RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{TRANSECT}{\code{source} - Line Transect(s). Path to a vector layer.}

\item{THEME}{\code{source} - Theme. Path to a vector layer.}

\item{THEME_FIELD}{\code{field} - Theme Field. The name of an existing field. ; delimited list of existing field names.}

\item{TRANSECT_RESULT}{\code{vectorDestination} - Result table. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Transect through polygon shapefile (saga:transectthroughpolygonshapefile)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TRANSECT_RESULT - outputVector - Result table
}
}
}
