% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_transformvectorlayer.R
\name{saga_transformvectorlayer}
\alias{saga_transformvectorlayer}
\title{QGIS algorithm Transform vector layer}
\usage{
saga_transformvectorlayer(
  IN = qgisprocess:::qgis_default_value(),
  DX = qgisprocess:::qgis_default_value(),
  DY = qgisprocess:::qgis_default_value(),
  ANGLE = qgisprocess:::qgis_default_value(),
  SCALEX = qgisprocess:::qgis_default_value(),
  SCALEY = qgisprocess:::qgis_default_value(),
  ANCHORX = qgisprocess:::qgis_default_value(),
  ANCHORY = qgisprocess:::qgis_default_value(),
  OUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{IN}{\code{source} - Shapes. Path to a vector layer.}

\item{DX}{\code{number} - dX. A numeric value.}

\item{DY}{\code{number} - dY. A numeric value.}

\item{ANGLE}{\code{number} - Angle. A numeric value.}

\item{SCALEX}{\code{number} - Scale Factor X. A numeric value.}

\item{SCALEY}{\code{number} - Scale Factor Y. A numeric value.}

\item{ANCHORX}{\code{number} - X. A numeric value.}

\item{ANCHORY}{\code{number} - Y. A numeric value.}

\item{OUT}{\code{vectorDestination} - Transformed. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Transform vector layer (saga:transformvectorlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUT - outputVector - Transformed
}
}
}
