% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_transposerasterlayers.R
\name{saga_transposerasterlayers}
\alias{saga_transposerasterlayers}
\title{QGIS algorithm Transpose raster layers}
\usage{
saga_transposerasterlayers(
  GRIDS = qgisprocess::qgis_default_value(),
  MIRROR_X = qgisprocess::qgis_default_value(),
  MIRROR_Y = qgisprocess::qgis_default_value(),
  TRANSPOSED = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{raster} - Input Grid. Path to a raster layer.}

\item{MIRROR_X}{\code{boolean} - Mirror Horizontally. 1 for true/yes. 0 for false/no.}

\item{MIRROR_Y}{\code{boolean} - Mirror Vertically. 1 for true/yes. 0 for false/no.}

\item{TRANSPOSED}{\code{rasterDestination} - Transposed Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Transpose raster layers (saga:transposerasterlayers)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TRANSPOSED - outputRaster - Transposed Grid
}
}
}
