% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_upslopeanddownslopecurvature.R
\name{saga_upslopeanddownslopecurvature}
\alias{saga_upslopeanddownslopecurvature}
\title{QGIS algorithm Upslope and downslope curvature}
\usage{
saga_upslopeanddownslopecurvature(
  DEM = qgisprocess::qgis_default_value(),
  C_LOCAL = qgisprocess::qgis_default_value(),
  C_UP = qgisprocess::qgis_default_value(),
  C_UP_LOCAL = qgisprocess::qgis_default_value(),
  C_DOWN = qgisprocess::qgis_default_value(),
  C_DOWN_LOCAL = qgisprocess::qgis_default_value(),
  WEIGHTING = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{C_LOCAL}{\code{rasterDestination} - Local Curvature. Path for new raster layer.}

\item{C_UP}{\code{rasterDestination} - Upslope Curvature. Path for new raster layer.}

\item{C_UP_LOCAL}{\code{rasterDestination} - Local Upslope Curvature. Path for new raster layer.}

\item{C_DOWN}{\code{rasterDestination} - Downslope Curvature. Path for new raster layer.}

\item{C_DOWN_LOCAL}{\code{rasterDestination} - Local Downslope Curvature. Path for new raster layer.}

\item{WEIGHTING}{\code{number} - Upslope Weighting. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Upslope and downslope curvature (saga:upslopeanddownslopecurvature)
}
\details{
\subsection{Outputs description}{
\itemize{
\item C_DOWN - outputRaster - Downslope Curvature
\item C_DOWN_LOCAL - outputRaster - Local Downslope Curvature
\item C_LOCAL - outputRaster - Local Curvature
\item C_UP - outputRaster - Upslope Curvature
\item C_UP_LOCAL - outputRaster - Local Upslope Curvature
}
}
}
