% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_upslopearea.R
\name{saga_upslopearea}
\alias{saga_upslopearea}
\title{QGIS algorithm Upslope area}
\usage{
saga_upslopearea(
  TARGET = qgisprocess::qgis_default_value(),
  TARGET_PT_X = qgisprocess::qgis_default_value(),
  TARGET_PT_Y = qgisprocess::qgis_default_value(),
  ELEVATION = qgisprocess::qgis_default_value(),
  SINKROUTE = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  CONVERGE = qgisprocess::qgis_default_value(),
  AREA = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TARGET}{\code{raster} - Target Area. Path to a raster layer.}

\item{TARGET_PT_X}{\code{number} - Target X coordinate. A numeric value.}

\item{TARGET_PT_Y}{\code{number} - Target Y coordinate. A numeric value.}

\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{SINKROUTE}{\code{raster} - Sink Routes. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Deterministic 8", "[1] Deterministic Infinity", "[2] Multiple Flow Direction")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONVERGE}{\code{number} - Convergence. A numeric value.}

\item{AREA}{\code{rasterDestination} - Upslope Area. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Upslope area (saga:upslopearea)
}
\details{
\subsection{Outputs description}{
\itemize{
\item AREA - outputRaster - Upslope Area
}
}
}
