% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_userdefinedfilter.R
\name{saga_userdefinedfilter}
\alias{saga_userdefinedfilter}
\title{QGIS algorithm User defined filter}
\usage{
saga_userdefinedfilter(
  INPUT = qgisprocess::qgis_default_value(),
  FILTER = qgisprocess::qgis_default_value(),
  FILTER_3X3 = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{FILTER}{\code{source} - Filter Matrix. Path to a vector layer.}

\item{FILTER_3X3}{\code{matrix} - Default Filter Matrix (3x3). A comma delimited list of values.}

\item{RESULT}{\code{rasterDestination} - Filtered Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA User defined filter (saga:userdefinedfilter)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Filtered Grid
}
}
}
