% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_valleyandridgedetectiontophatapproach.R
\name{saga_valleyandridgedetectiontophatapproach}
\alias{saga_valleyandridgedetectiontophatapproach}
\title{QGIS algorithm Valley and ridge detection (top hat approach)}
\usage{
saga_valleyandridgedetectiontophatapproach(
  DEM = qgisprocess::qgis_default_value(),
  VALLEY = qgisprocess::qgis_default_value(),
  HILL = qgisprocess::qgis_default_value(),
  VALLEY_IDX = qgisprocess::qgis_default_value(),
  HILL_IDX = qgisprocess::qgis_default_value(),
  SLOPE_IDX = qgisprocess::qgis_default_value(),
  RADIUS_VALLEY = qgisprocess::qgis_default_value(),
  RADIUS_HILL = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{VALLEY}{\code{rasterDestination} - Valley Depth. Path for new raster layer.}

\item{HILL}{\code{rasterDestination} - Hill Height. Path for new raster layer.}

\item{VALLEY_IDX}{\code{rasterDestination} - Valley Index. Path for new raster layer.}

\item{HILL_IDX}{\code{rasterDestination} - Hill Index. Path for new raster layer.}

\item{SLOPE_IDX}{\code{rasterDestination} - Hillslope Index. Path for new raster layer.}

\item{RADIUS_VALLEY}{\code{number} - Valley Radius. A numeric value.}

\item{RADIUS_HILL}{\code{number} - Hill Radius. A numeric value.}

\item{THRESHOLD}{\code{number} - Elevation Threshold. A numeric value.}

\item{METHOD}{\code{enum}  of \verb{("[0] default", "[1] alternative")} - Slope Index. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Valley and ridge detection (top hat approach) (saga:valleyandridgedetectiontophatapproach)
}
\details{
\subsection{Outputs description}{
\itemize{
\item HILL - outputRaster - Hill Height
\item HILL_IDX - outputRaster - Hill Index
\item SLOPE_IDX - outputRaster - Hillslope Index
\item VALLEY - outputRaster - Valley Depth
\item VALLEY_IDX - outputRaster - Valley Index
}
}
}
