% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_valleydepth.R
\name{saga_valleydepth}
\alias{saga_valleydepth}
\title{QGIS algorithm Valley depth}
\usage{
saga_valleydepth(
  ELEVATION = qgisprocess::qgis_default_value(),
  VALLEY_DEPTH = qgisprocess::qgis_default_value(),
  RIDGE_LEVEL = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  NOUNDERGROUND = qgisprocess::qgis_default_value(),
  ORDER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{VALLEY_DEPTH}{\code{rasterDestination} - Valley Depth. Path for new raster layer.}

\item{RIDGE_LEVEL}{\code{rasterDestination} - Ridge Level. Path for new raster layer.}

\item{THRESHOLD}{\code{number} - Tension Threshold (Percentage of Cell Size). A numeric value.}

\item{NOUNDERGROUND}{\code{boolean} - Keep Ridge Level above Surface. 1 for true/yes. 0 for false/no.}

\item{ORDER}{\code{number} - Ridge Detection Threshold. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Valley depth (saga:valleydepth)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RIDGE_LEVEL - outputRaster - Ridge Level
\item VALLEY_DEPTH - outputRaster - Valley Depth
}
}
}
