% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_variogramcloud.R
\name{saga_variogramcloud}
\alias{saga_variogramcloud}
\title{QGIS algorithm Variogram cloud}
\usage{
saga_variogramcloud(
  POINTS = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  DISTMAX = qgisprocess:::qgis_default_value(),
  NSKIP = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{DISTMAX}{\code{number} - Maximum Distance. A numeric value.}

\item{NSKIP}{\code{number} - Skip Number. A numeric value.}

\item{RESULT}{\code{vectorDestination} - Variogram Cloud. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Variogram cloud (saga:variogramcloud)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Variogram Cloud
}
}
}
