% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_vectorisinggridclasses.R
\name{saga_vectorisinggridclasses}
\alias{saga_vectorisinggridclasses}
\title{QGIS algorithm Vectorising grid classes}
\usage{
saga_vectorisinggridclasses(
  GRID = qgisprocess::qgis_default_value(),
  CLASS_ALL = qgisprocess::qgis_default_value(),
  CLASS_ID = qgisprocess::qgis_default_value(),
  SPLIT = qgisprocess::qgis_default_value(),
  POLYGONS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{CLASS_ALL}{\code{enum}  of \verb{("[0] one single class specified by class identifier", "[1] all classes")} - Class Selection. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CLASS_ID}{\code{number} - Class Identifier. A numeric value.}

\item{SPLIT}{\code{enum}  of \verb{("[0] one single (multi-)polygon object", "[1] each island as separated polygon")} - Vectorised class as.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{POLYGONS}{\code{vectorDestination} - Vectorized. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Vectorising grid classes (saga:vectorisinggridclasses)
}
\details{
\subsection{Outputs description}{
\itemize{
\item POLYGONS - outputVector - Vectorized
}
}
}
