% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_vegetationindexdistancebased.R
\name{saga_vegetationindexdistancebased}
\alias{saga_vegetationindexdistancebased}
\title{QGIS algorithm Vegetation index (distance based)}
\usage{
saga_vegetationindexdistancebased(
  RED = qgisprocess::qgis_default_value(),
  NIR = qgisprocess::qgis_default_value(),
  PVI0 = qgisprocess::qgis_default_value(),
  PVI1 = qgisprocess::qgis_default_value(),
  PVI2 = qgisprocess::qgis_default_value(),
  PVI3 = qgisprocess::qgis_default_value(),
  TSAVI = qgisprocess::qgis_default_value(),
  ATSAVI = qgisprocess::qgis_default_value(),
  INTERCEPT = qgisprocess::qgis_default_value(),
  SLOPE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{RED}{\code{raster} - Red Reflectance. Path to a raster layer.}

\item{NIR}{\code{raster} - Near Infrared Reflectance. Path to a raster layer.}

\item{PVI0}{\code{rasterDestination} - Perpendicular Vegetation Index (Richardson and Wiegand, 1977). Path for new raster layer.}

\item{PVI1}{\code{rasterDestination} - Perpendicular Vegetation Index (Perry and Lautenschlager, 1984). Path for new raster layer.}

\item{PVI2}{\code{rasterDestination} - Perpendicular Vegetation Index (Walther and Shabaani). Path for new raster layer.}

\item{PVI3}{\code{rasterDestination} - Perpendicular Vegetation Index (Qi, et al., 1994). Path for new raster layer.}

\item{TSAVI}{\code{rasterDestination} - Transformed Soil Adjusted Vegetation Index (Baret et al. 1989). Path for new raster layer.}

\item{ATSAVI}{\code{rasterDestination} - Transformed Soil Adjusted Vegetation Index (Baret and Guyot, 1991). Path for new raster layer.}

\item{INTERCEPT}{\code{number} - Intercept of Soil Line. A numeric value.}

\item{SLOPE}{\code{number} - Slope of Soil Line. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Vegetation index (distance based) (saga:vegetationindexdistancebased)
}
\details{
\subsection{Outputs description}{
\itemize{
\item PVI0 - outputRaster - Perpendicular Vegetation Index
\item PVI1 - outputRaster - Perpendicular Vegetation Index
\item PVI2 - outputRaster - Perpendicular Vegetation Index
\item PVI3 - outputRaster - Perpendicular Vegetation Index
\item TSAVI - outputRaster - Transformed Soil Adjusted Vegetation Index
\item ATSAVI - outputRaster - Transformed Soil Adjusted Vegetation Index
}
}
}
