% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_vegetationindexslopebased.R
\name{saga_vegetationindexslopebased}
\alias{saga_vegetationindexslopebased}
\title{QGIS algorithm Vegetation index (slope based)}
\usage{
saga_vegetationindexslopebased(
  NIR = qgisprocess::qgis_default_value(),
  RED = qgisprocess::qgis_default_value(),
  SOIL = qgisprocess::qgis_default_value(),
  DVI = qgisprocess::qgis_default_value(),
  NDVI = qgisprocess::qgis_default_value(),
  RVI = qgisprocess::qgis_default_value(),
  NRVI = qgisprocess::qgis_default_value(),
  TVI = qgisprocess::qgis_default_value(),
  CTVI = qgisprocess::qgis_default_value(),
  TTVI = qgisprocess::qgis_default_value(),
  SAVI = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{NIR}{\code{raster} - Near Infrared Reflectance. Path to a raster layer.}

\item{RED}{\code{raster} - Red Reflectance. Path to a raster layer.}

\item{SOIL}{\code{number} - Soil Adjustment Factor. A numeric value.}

\item{DVI}{\code{rasterDestination} - Difference Vegetation Index. Path for new raster layer.}

\item{NDVI}{\code{rasterDestination} - Normalized Difference Vegetation Index. Path for new raster layer.}

\item{RVI}{\code{rasterDestination} - Ratio Vegetation Index. Path for new raster layer.}

\item{NRVI}{\code{rasterDestination} - Normalized Ratio Vegetation Index. Path for new raster layer.}

\item{TVI}{\code{rasterDestination} - Transformed Vegetation Index. Path for new raster layer.}

\item{CTVI}{\code{rasterDestination} - Corrected Transformed Vegetation Index. Path for new raster layer.}

\item{TTVI}{\code{rasterDestination} - Thiam's Transformed Vegetation Index. Path for new raster layer.}

\item{SAVI}{\code{rasterDestination} - Soil Adjusted Vegetation Index. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Vegetation index (slope based) (saga:vegetationindexslopebased)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DVI - outputRaster - Difference Vegetation Index
\item NDVI - outputRaster - Normalized Difference Vegetation Index
\item RVI - outputRaster - Ratio Vegetation Index
\item NRVI - outputRaster - Normalized Ratio Vegetation Index
\item TVI - outputRaster - Transformed Vegetation Index
\item CTVI - outputRaster - Corrected Transformed Vegetation Index
\item TTVI - outputRaster - Thiam
\item SAVI - outputRaster - Soil Adjusted Vegetation Index
}
}
}
