% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_warpingshapes.R
\name{saga_warpingshapes}
\alias{saga_warpingshapes}
\title{QGIS algorithm Warping shapes}
\usage{
saga_warpingshapes(
  REF_SOURCE = qgisprocess:::qgis_default_value(),
  REF_TARGET = qgisprocess:::qgis_default_value(),
  XFIELD = qgisprocess:::qgis_default_value(),
  YFIELD = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ORDER = qgisprocess:::qgis_default_value(),
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{REF_SOURCE}{\code{source} - Reference Points (Origin). Path to a vector layer.}

\item{REF_TARGET}{\code{source} - Reference Points (Projection). Path to a vector layer.}

\item{XFIELD}{\code{source} - x Position. Path to a vector layer.}

\item{YFIELD}{\code{source} - y Position. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Automatic", "[1] Triangulation", "[2] Spline", "[3] Affine", "[4] 1st Order Polynomial", "[5] 2nd Order Polynomial", "[6] 3rd Order Polynomial", "[7] Polynomial, Order")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ORDER}{\code{number} - Polynomial Order. A numeric value.}

\item{INPUT}{\code{source} - Input. Path to a vector layer.}

\item{OUTPUT}{\code{vectorDestination} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Warping shapes (saga:warpingshapes)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output
}
}
}
