% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_waterretentioncapacity.R
\name{saga_waterretentioncapacity}
\alias{saga_waterretentioncapacity}
\title{QGIS algorithm Water retention capacity}
\usage{
saga_waterretentioncapacity(
  SHAPES = qgisprocess::qgis_default_value(),
  DEM = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  RETENTION = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Plot Holes. Path to a vector layer.}

\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{OUTPUT}{\code{vectorDestination} - Final Parameters. Path for new vector layer.}

\item{RETENTION}{\code{rasterDestination} - Water Retention Capacity. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Water retention capacity (saga:waterretentioncapacity)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Final Parameters
\item RETENTION - outputRaster - Water Retention Capacity
}
}
}
