% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_watershedsegmentation.R
\name{saga_watershedsegmentation}
\alias{saga_watershedsegmentation}
\title{QGIS algorithm Watershed segmentation}
\usage{
saga_watershedsegmentation(
  GRID = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  DOWN = qgisprocess::qgis_default_value(),
  JOIN = qgisprocess::qgis_default_value(),
  THRESHOLD = qgisprocess::qgis_default_value(),
  EDGE = qgisprocess::qgis_default_value(),
  BBORDERS = qgisprocess::qgis_default_value(),
  SEGMENTS = qgisprocess::qgis_default_value(),
  SEEDS = qgisprocess::qgis_default_value(),
  BORDERS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{OUTPUT}{\code{enum}  of \verb{("[0] Seed Value", "[1] Segment ID")} - Output. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DOWN}{\code{enum}  of \verb{("[0] Minima", "[1] Maxima")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{JOIN}{\code{enum}  of \verb{("[0] do not join", "[1] seed to saddle difference", "[2] seeds difference")} - Join Segments based on Threshold Value. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{THRESHOLD}{\code{number} - Threshold. A numeric value.}

\item{EDGE}{\code{boolean} - Allow Edge Pixels to be Seeds. 1 for true/yes. 0 for false/no.}

\item{BBORDERS}{\code{boolean} - Borders. 1 for true/yes. 0 for false/no.}

\item{SEGMENTS}{\code{rasterDestination} - Segments. Path for new raster layer.}

\item{SEEDS}{\code{vectorDestination} - Seed Points. Path for new vector layer.}

\item{BORDERS}{\code{rasterDestination} - Borders. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Watershed segmentation (saga:watershedsegmentation)
}
\details{
\subsection{Outputs description}{
\itemize{
\item BORDERS - outputRaster - Borders
\item SEEDS - outputVector - Seed Points
\item SEGMENTS - outputRaster - Segments
}
}
}
