% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_windexpositionindex.R
\name{saga_windexpositionindex}
\alias{saga_windexpositionindex}
\title{QGIS algorithm Wind exposition index}
\usage{
saga_windexpositionindex(
  DEM = qgisprocess:::qgis_default_value(),
  EXPOSITION = qgisprocess:::qgis_default_value(),
  MAXDIST = qgisprocess:::qgis_default_value(),
  STEP = qgisprocess:::qgis_default_value(),
  OLDVER = qgisprocess:::qgis_default_value(),
  ACCEL = qgisprocess:::qgis_default_value(),
  PYRAMIDS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{EXPOSITION}{\code{rasterDestination} - Wind Exposition. Path for new raster layer.}

\item{MAXDIST}{\code{number} - Search Distance (km). A numeric value.}

\item{STEP}{\code{number} - Angular Step Size (Degree). A numeric value.}

\item{OLDVER}{\code{boolean} - Old Version. 1 for true/yes. 0 for false/no.}

\item{ACCEL}{\code{number} - Acceleration. A numeric value.}

\item{PYRAMIDS}{\code{boolean} - Elevation Averaging. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Wind exposition index (saga:windexpositionindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item EXPOSITION - outputRaster - Wind Exposition
}
}
}
