% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sagang_02_carboncyclesimulationforterrestrialbiomass.R
\name{sagang_02_carboncyclesimulationforterrestrialbiomass}
\alias{sagang_02_carboncyclesimulationforterrestrialbiomass}
\title{QGIS algorithm - 02: carbon cycle simulation for terrestrial biomass}
\usage{
sagang_02_carboncyclesimulationforterrestrialbiomass(
  TABLE = qgisprocess:::qgis_default_value(),
  TIME_SPAN = qgisprocess:::qgis_default_value(),
  TIME_STEP = qgisprocess:::qgis_default_value(),
  PRIMPROD = qgisprocess:::qgis_default_value(),
  CHUMIFY = qgisprocess:::qgis_default_value(),
  CCARBON = qgisprocess:::qgis_default_value(),
  P_LEAV = qgisprocess:::qgis_default_value(),
  P_BRAN = qgisprocess:::qgis_default_value(),
  P_STEM = qgisprocess:::qgis_default_value(),
  P_ROOT = qgisprocess:::qgis_default_value(),
  K_LEAV_LITT = qgisprocess:::qgis_default_value(),
  K_BRAN_LITT = qgisprocess:::qgis_default_value(),
  K_STEM_LITT = qgisprocess:::qgis_default_value(),
  K_ROOT_HUMU = qgisprocess:::qgis_default_value(),
  K_LITT_HUMU = qgisprocess:::qgis_default_value(),
  K_HUMU_COAL = qgisprocess:::qgis_default_value(),
  K_COAL_ENVI = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TABLE}{\code{vectorDestination} - Results. Path for new vector layer.}

\item{TIME_SPAN}{\code{number} - Time Span. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME_STEP}{\code{number} - Time Interval (a). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PRIMPROD}{\code{number} - Net Primary Production (Gt Carbon \code{/} a). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CHUMIFY}{\code{number} - Humification Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CCARBON}{\code{number} - Carbonization Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_LEAV}{\code{number} - Leaves. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_BRAN}{\code{number} - Branches. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_STEM}{\code{number} - Stems. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_ROOT}{\code{number} - Roots. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{K_LEAV_LITT}{\code{number} - Leaves to Litter. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{K_BRAN_LITT}{\code{number} - Branches to Litter. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{K_STEM_LITT}{\code{number} - Stems to Litter. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{K_ROOT_HUMU}{\code{number} - Roots to Humus. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{K_LITT_HUMU}{\code{number} - Litter to Humus. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{K_HUMU_COAL}{\code{number} - Humus to Charcoal. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{K_COAL_ENVI}{\code{number} - Charcoal to Environment. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen 02: carbon cycle simulation for terrestrial biomass (sagang:02:carboncyclesimulationforterrestrialbiomass). ---------------- Arguments ----------------  TABLE: Results 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer TIME_SPAN: Time Span 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME_STEP: Time Interval (a) 	Default value:	0.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PRIMPROD: Net Primary Production (Gt Carbon / a) 	Default value:	27.8 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CHUMIFY: Humification Factor 	Default value:	0.4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CCARBON: Carbonization Factor 	Default value:	0.05 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_LEAV: Leaves 	Default value:	0.3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_BRAN: Branches 	Default value:	0.2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_STEM: Stems 	Default value:	0.3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_ROOT: Roots 	Default value:	0.2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression K_LEAV_LITT: Leaves to Litter 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression K_BRAN_LITT: Branches to Litter 	Default value:	0.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression K_STEM_LITT: Stems to Litter 	Default value:	0.033 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression K_ROOT_HUMU: Roots to Humus 	Default value:	0.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression K_LITT_HUMU: Litter to Humus 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression K_HUMU_COAL: Humus to Charcoal 	Default value:	0.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression K_COAL_ENVI: Charcoal to Environment 	Default value:	0.002 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item TABLE - outputVector - Results
}
}
}
