% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_3dpointsselection.R
\name{sagang_3dpointsselection}
\alias{sagang_3dpointsselection}
\title{QGIS algorithm - 3d points selection}
\usage{
sagang_3dpointsselection(
  LOWER = qgisprocess:::qgis_default_value(),
  UPPER = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  Z_FIELD = qgisprocess:::qgis_default_value(),
  COPY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LOWER}{\code{raster} - Lower Surface. Path to a raster layer.}

\item{UPPER}{\code{raster} - Upper Surface. Path to a raster layer.}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{Z_FIELD}{\code{field} - Z. The name of an existing field. ; delimited list of existing field names.}

\item{COPY}{\code{vectorDestination} - Copy Selection. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen 3d points selection (sagang:3dpointsselection). ---------------- Arguments ----------------  LOWER: Lower Surface 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer UPPER: Upper Surface 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer Z_FIELD: Z 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names COPY: Copy Selection 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item COPY - outputVector - Copy Selection
}
}
}
