% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_accumulatedcost.R
\name{sagang_accumulatedcost}
\alias{sagang_accumulatedcost}
\title{QGIS algorithm - Accumulated cost}
\usage{
sagang_accumulatedcost(
  DEST_TYPE = qgisprocess:::qgis_default_value(),
  DEST_POINTS = qgisprocess:::qgis_default_value(),
  DEST_GRID = qgisprocess:::qgis_default_value(),
  COST = qgisprocess:::qgis_default_value(),
  DIR_MAXCOST = qgisprocess:::qgis_default_value(),
  DIR_UNIT = qgisprocess:::qgis_default_value(),
  DIR_K = qgisprocess:::qgis_default_value(),
  ACCUMULATED = qgisprocess:::qgis_default_value(),
  ALLOCATION = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEST_TYPE}{\code{enum}  of \verb{("(0) Point", "(1) Grid")} - Input Type of Destinations. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DEST_POINTS}{\code{source} - Destinations. Path to a vector layer.}

\item{DEST_GRID}{\code{raster} - Destinations. Path to a raster layer.}

\item{COST}{\code{raster} - Local Cost. Path to a raster layer.}

\item{DIR_MAXCOST}{\code{raster} - Direction of Maximum Cost. Path to a raster layer.}

\item{DIR_UNIT}{\code{enum}  of \verb{("(0) radians", "(1) degree")} - Units of Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DIR_K}{\code{number} - K Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ACCUMULATED}{\code{rasterDestination} - Accumulated Cost. Path for new raster layer.}

\item{ALLOCATION}{\code{rasterDestination} - Allocation. Path for new raster layer.}

\item{THRESHOLD}{\code{number} - Threshold for different route. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Accumulated cost (sagang:accumulatedcost). ---------------- Arguments ----------------  DEST_TYPE: Input Type of Destinations 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Point 		- 1: (1) Grid 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DEST_POINTS: Destinations 	Argument type:	source 	Acceptable values: 		- Path to a vector layer DEST_GRID: Destinations 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer COST: Local Cost 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DIR_MAXCOST: Direction of Maximum Cost (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DIR_UNIT: Units of Direction 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) radians 		- 1: (1) degree 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DIR_K: K Factor 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ACCUMULATED: Accumulated Cost 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ALLOCATION: Allocation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer THRESHOLD: Threshold for different route 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item ACCUMULATED - outputRaster - Accumulated Cost
\item ALLOCATION - outputRaster - Allocation
}
}
}
