% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_accumulationfunctions.R
\name{sagang_accumulationfunctions}
\alias{sagang_accumulationfunctions}
\title{QGIS algorithm - Accumulation functions}
\usage{
sagang_accumulationfunctions(
  SURFACE = qgisprocess:::qgis_default_value(),
  INPUT = qgisprocess:::qgis_default_value(),
  STATE_IN = qgisprocess:::qgis_default_value(),
  CONTROL = qgisprocess:::qgis_default_value(),
  CTRL_LINEAR = qgisprocess:::qgis_default_value(),
  FLUX = qgisprocess:::qgis_default_value(),
  STATE_OUT = qgisprocess:::qgis_default_value(),
  OPERATION = qgisprocess:::qgis_default_value(),
  LINEAR = qgisprocess:::qgis_default_value(),
  THRES_LINEAR = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SURFACE}{\code{raster} - Surface. Path to a raster layer.}

\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{STATE_IN}{\code{raster} - State t. Path to a raster layer.}

\item{CONTROL}{\code{raster} - Operation Control. Path to a raster layer.}

\item{CTRL_LINEAR}{\code{raster} - Linear Flow Control Grid. Path to a raster layer.}

\item{FLUX}{\code{rasterDestination} - Flux. Path for new raster layer.}

\item{STATE_OUT}{\code{rasterDestination} - State t + 1. Path for new raster layer.}

\item{OPERATION}{\code{enum}  of \verb{("(0) accuflux", "(1) accucapacityflux / state", "(2) accufractionflux / state", "(3) accuthresholdflux / state", "(4) accutriggerflux / state")} - Operation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LINEAR}{\code{boolean} - Switch to Linear Flow. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRES_LINEAR}{\code{number} - Threshold Linear Flow. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Accumulation functions (sagang:accumulationfunctions). ---------------- Arguments ----------------  SURFACE: Surface 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer INPUT: Input 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer STATE_IN: State t (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CONTROL: Operation Control (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CTRL_LINEAR: Linear Flow Control Grid (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FLUX: Flux 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STATE_OUT: State t + 1 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer OPERATION: Operation 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) accuflux 		- 1: (1) accucapacityflux / state 		- 2: (2) accufractionflux / state 		- 3: (3) accuthresholdflux / state 		- 4: (4) accutriggerflux / state 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' LINEAR: Switch to Linear Flow 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRES_LINEAR: Threshold Linear Flow 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLUX - outputRaster - Flux
\item STATE_OUT - outputRaster - State t + 1
}
}
}
