% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_aggregatepointobservations.R
\name{sagang_aggregatepointobservations}
\alias{sagang_aggregatepointobservations}
\title{QGIS algorithm - Aggregate point observations}
\usage{
sagang_aggregatepointobservations(
  REFERENCE = qgisprocess:::qgis_default_value(),
  REFERENCE_ID = qgisprocess:::qgis_default_value(),
  OBSERVATIONS = qgisprocess:::qgis_default_value(),
  X = qgisprocess:::qgis_default_value(),
  Y = qgisprocess:::qgis_default_value(),
  TRACK = qgisprocess:::qgis_default_value(),
  DATE = qgisprocess:::qgis_default_value(),
  TIME = qgisprocess:::qgis_default_value(),
  PARAMETER = qgisprocess:::qgis_default_value(),
  AGGREGATED = qgisprocess:::qgis_default_value(),
  TIME_SPAN = qgisprocess:::qgis_default_value(),
  FIX_TIME = qgisprocess:::qgis_default_value(),
  OFF_TIME = qgisprocess:::qgis_default_value(),
  EPS_TIME = qgisprocess:::qgis_default_value(),
  EPS_SPACE = qgisprocess:::qgis_default_value(),
  VERBOSE = qgisprocess:::qgis_default_value(),
  POLAR = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{REFERENCE}{\code{source} - Reference Points. Path to a vector layer.}

\item{REFERENCE_ID}{\code{field} - ID. The name of an existing field. ; delimited list of existing field names.}

\item{OBSERVATIONS}{\code{source} - Observations. Path to a vector layer.}

\item{X}{\code{field} - X. The name of an existing field. ; delimited list of existing field names.}

\item{Y}{\code{field} - Y. The name of an existing field. ; delimited list of existing field names.}

\item{TRACK}{\code{field} - Track. The name of an existing field. ; delimited list of existing field names.}

\item{DATE}{\code{field} - Date. The name of an existing field. ; delimited list of existing field names.}

\item{TIME}{\code{field} - Time. The name of an existing field. ; delimited list of existing field names.}

\item{PARAMETER}{\code{field} - Parameter. The name of an existing field. ; delimited list of existing field names.}

\item{AGGREGATED}{\code{vectorDestination} - Aggregated. Path for new vector layer.}

\item{TIME_SPAN}{\code{enum}  of \verb{("(0) ignore", "(1) floating", "(2) fixed")} - Time Span Aggregation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FIX_TIME}{\code{number} - Fixed Time Span (minutes). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OFF_TIME}{\code{number} - Fixed Time Span Offset (minutes). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EPS_TIME}{\code{number} - Maximum Time Span (Seconds). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EPS_SPACE}{\code{number} - Maximum Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{VERBOSE}{\code{boolean} - Verbose. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{POLAR}{\code{boolean} - Polar Coordinates. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Aggregate point observations (sagang:aggregatepointobservations). ---------------- Arguments ----------------  REFERENCE: Reference Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer REFERENCE_ID: ID 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names OBSERVATIONS: Observations 	Argument type:	source 	Acceptable values: 		- Path to a vector layer X: X 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names Y: Y 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TRACK: Track 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names DATE: Date 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TIME: Time 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names PARAMETER: Parameter 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names AGGREGATED: Aggregated 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer TIME_SPAN: Time Span Aggregation 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) ignore 		- 1: (1) floating 		- 2: (2) fixed 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FIX_TIME: Fixed Time Span (minutes) 	Default value:	20 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OFF_TIME: Fixed Time Span Offset (minutes) 	Default value:	-10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EPS_TIME: Maximum Time Span (Seconds) 	Default value:	60 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EPS_SPACE: Maximum Distance 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression VERBOSE: Verbose 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression POLAR: Polar Coordinates 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item AGGREGATED - outputVector - Aggregated
}
}
}
