% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_airpressureadjustment.R
\name{sagang_airpressureadjustment}
\alias{sagang_airpressureadjustment}
\title{QGIS algorithm - Air pressure adjustment}
\usage{
sagang_airpressureadjustment(
  P = qgisprocess:::qgis_default_value(),
  P_DEFAULT = qgisprocess:::qgis_default_value(),
  Z = qgisprocess:::qgis_default_value(),
  Z_DEFAULT = qgisprocess:::qgis_default_value(),
  T = qgisprocess:::qgis_default_value(),
  T_DEFAULT = qgisprocess:::qgis_default_value(),
  L = qgisprocess:::qgis_default_value(),
  L_DEFAULT = qgisprocess:::qgis_default_value(),
  DEM = qgisprocess:::qgis_default_value(),
  P_ADJ = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{P}{\code{raster} - Air Pressure. Path to a raster layer.}

\item{P_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{Z}{\code{raster} - Air Pressure Elevation. Path to a raster layer.}

\item{Z_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T}{\code{raster} - Temperature. Path to a raster layer.}

\item{T_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{L}{\code{raster} - Temperature Lapse Rate. Path to a raster layer.}

\item{L_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{P_ADJ}{\code{rasterDestination} - Adjusted Air Pressure. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Air pressure adjustment (sagang:airpressureadjustment). ---------------- Arguments ----------------  P: Air Pressure (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer P_DEFAULT: Default 	Default value:	1013.25 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Z: Air Pressure Elevation (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Z_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T: Temperature (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer T_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression L: Temperature Lapse Rate (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer L_DEFAULT: Default 	Default value:	0.0065 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer P_ADJ: Adjusted Air Pressure 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item P_ADJ - outputRaster - Adjusted Air Pressure
}
}
}
