% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_analyticalhillshading.R
\name{sagang_analyticalhillshading}
\alias{sagang_analyticalhillshading}
\title{QGIS algorithm - Analytical hillshading}
\usage{
sagang_analyticalhillshading(
  ELEVATION = qgisprocess:::qgis_default_value(),
  SHADE = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  POSITION = qgisprocess:::qgis_default_value(),
  AZIMUTH = qgisprocess:::qgis_default_value(),
  DECLINATION = qgisprocess:::qgis_default_value(),
  DATE = qgisprocess:::qgis_default_value(),
  TIME = qgisprocess:::qgis_default_value(),
  EXAGGERATION = qgisprocess:::qgis_default_value(),
  UNIT = qgisprocess:::qgis_default_value(),
  SHADOW = qgisprocess:::qgis_default_value(),
  NDIRS = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{SHADE}{\code{rasterDestination} - Analytical Hillshading. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) Standard", "(1) Limited Maximum", "(2) With Shadows", "(3) Shadows Only", "(4) Ambient Occlusion", "(5) Combined Shading")} - Shading Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{POSITION}{\code{enum}  of \verb{("(0) azimuth and height", "(1) date and time")} - Sun's Position. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{AZIMUTH}{\code{number} - Azimuth. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DECLINATION}{\code{number} - Height. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DATE}{\code{string} - Day. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME}{\code{number} - Hour. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EXAGGERATION}{\code{number} - Exaggeration. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{UNIT}{\code{enum}  of \verb{("(0) radians", "(1) degree")} - Unit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SHADOW}{\code{enum}  of \verb{("(0) slim", "(1) fat")} - Shadow. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NDIRS}{\code{number} - Number of Directions. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RADIUS}{\code{number} - Search Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Analytical hillshading (sagang:analyticalhillshading). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SHADE: Analytical Hillshading 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Shading Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Standard 		- 1: (1) Limited Maximum 		- 2: (2) With Shadows 		- 3: (3) Shadows Only 		- 4: (4) Ambient Occlusion 		- 5: (5) Combined Shading 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' POSITION: Sun's Position 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) azimuth and height 		- 1: (1) date and time 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' AZIMUTH: Azimuth 	Default value:	315 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DECLINATION: Height 	Default value:	45 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DATE: Day 	Default value:	2021-03-24 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME: Hour 	Default value:	12 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EXAGGERATION: Exaggeration 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression UNIT: Unit 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) radians 		- 1: (1) degree 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SHADOW: Shadow 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) slim 		- 1: (1) fat 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NDIRS: Number of Directions 	Default value:	8 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RADIUS: Search Radius 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item SHADE - outputRaster - Analytical Hillshading
}
}
}
