% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_angmap.R
\name{sagang_angmap}
\alias{sagang_angmap}
\title{QGIS algorithm - Angmap}
\usage{
sagang_angmap(
  DEM = qgisprocess:::qgis_default_value(),
  C = qgisprocess:::qgis_default_value(),
  D = qgisprocess:::qgis_default_value(),
  fB = qgisprocess:::qgis_default_value(),
  fC = qgisprocess:::qgis_default_value(),
  E = qgisprocess:::qgis_default_value(),
  F = qgisprocess:::qgis_default_value(),
  G = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{C}{\code{raster} - Dip grid (degrees) . Path to a raster layer.}

\item{D}{\code{raster} - Dip direction grid (degrees) . Path to a raster layer.}

\item{fB}{\code{number} - Global structure dip (degrees). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fC}{\code{number} - Global structure dip direction (degrees). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{E}{\code{rasterDestination} - Angle. Path for new raster layer.}

\item{F}{\code{rasterDestination} - CL dipdir. Path for new raster layer.}

\item{G}{\code{rasterDestination} - CL dip. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Angmap (sagang:angmap). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer C: Dip grid (degrees)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer D: Dip direction grid (degrees)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer fB: Global structure dip (degrees) 	Default value:	45 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fC: Global structure dip direction (degrees) 	Default value:	90 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression E: Angle 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer F: CL dipdir 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer G: CL dip 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item E - outputRaster - Angle
\item F - outputRaster - CL dipdir
\item G - outputRaster - CL dip
}
}
}
