% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_automatedcloudcoverassessment.R
\name{sagang_automatedcloudcoverassessment}
\alias{sagang_automatedcloudcoverassessment}
\title{QGIS algorithm - Automated cloud cover assessment}
\usage{
sagang_automatedcloudcoverassessment(
  BAND2 = qgisprocess:::qgis_default_value(),
  BAND3 = qgisprocess:::qgis_default_value(),
  BAND4 = qgisprocess:::qgis_default_value(),
  BAND5 = qgisprocess:::qgis_default_value(),
  BAND6 = qgisprocess:::qgis_default_value(),
  CLOUD = qgisprocess:::qgis_default_value(),
  FILTER = qgisprocess:::qgis_default_value(),
  B56C = qgisprocess:::qgis_default_value(),
  B45R = qgisprocess:::qgis_default_value(),
  CSIG = qgisprocess:::qgis_default_value(),
  PASS2 = qgisprocess:::qgis_default_value(),
  SHADOW = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{BAND2}{\code{raster} - Green. Path to a raster layer.}

\item{BAND3}{\code{raster} - Red. Path to a raster layer.}

\item{BAND4}{\code{raster} - NIR. Path to a raster layer.}

\item{BAND5}{\code{raster} - SWIR. Path to a raster layer.}

\item{BAND6}{\code{raster} - Thermal. Path to a raster layer.}

\item{CLOUD}{\code{rasterDestination} - Cloud Cover. Path for new raster layer.}

\item{FILTER}{\code{boolean} - Apply post-processing filter to remove small holes. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{B56C}{\code{number} - B56 Composite (step 6). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{B45R}{\code{number} - B45 Ratio: Desert detection (step 10). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CSIG}{\code{boolean} - Always use cloud signature (step 14). 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PASS2}{\code{boolean} - Bypass second-pass processing, and merge warm (not ambiguous) and cold clouds. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SHADOW}{\code{boolean} - Include a category for cloud shadows. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Automated cloud cover assessment (sagang:automatedcloudcoverassessment). ---------------- Arguments ----------------  BAND2: Green 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer BAND3: Red 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer BAND4: NIR 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer BAND5: SWIR 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer BAND6: Thermal 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CLOUD: Cloud Cover 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FILTER: Apply post-processing filter to remove small holes 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression B56C: B56 Composite (step 6) 	Default value:	225 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression B45R: B45 Ratio: Desert detection (step 10) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CSIG: Always use cloud signature (step 14) 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PASS2: Bypass second-pass processing, and merge warm (not ambiguous) and cold clouds 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SHADOW: Include a category for cloud shadows 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLOUD - outputRaster - Cloud Cover
}
}
}
