% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_bioclimaticvariables.R
\name{sagang_bioclimaticvariables}
\alias{sagang_bioclimaticvariables}
\title{QGIS algorithm - Bioclimatic variables}
\usage{
sagang_bioclimaticvariables(
  TMEAN = qgisprocess:::qgis_default_value(),
  TMIN = qgisprocess:::qgis_default_value(),
  TMAX = qgisprocess:::qgis_default_value(),
  P = qgisprocess:::qgis_default_value(),
  BIO_01 = qgisprocess:::qgis_default_value(),
  BIO_02 = qgisprocess:::qgis_default_value(),
  BIO_03 = qgisprocess:::qgis_default_value(),
  BIO_04 = qgisprocess:::qgis_default_value(),
  BIO_05 = qgisprocess:::qgis_default_value(),
  BIO_06 = qgisprocess:::qgis_default_value(),
  BIO_07 = qgisprocess:::qgis_default_value(),
  BIO_08 = qgisprocess:::qgis_default_value(),
  BIO_09 = qgisprocess:::qgis_default_value(),
  BIO_10 = qgisprocess:::qgis_default_value(),
  BIO_11 = qgisprocess:::qgis_default_value(),
  BIO_12 = qgisprocess:::qgis_default_value(),
  BIO_13 = qgisprocess:::qgis_default_value(),
  BIO_14 = qgisprocess:::qgis_default_value(),
  BIO_15 = qgisprocess:::qgis_default_value(),
  BIO_16 = qgisprocess:::qgis_default_value(),
  BIO_17 = qgisprocess:::qgis_default_value(),
  BIO_18 = qgisprocess:::qgis_default_value(),
  BIO_19 = qgisprocess:::qgis_default_value(),
  SEASONALITY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TMEAN}{\code{multilayer} - Mean Temperature. .}

\item{TMIN}{\code{multilayer} - Minimum Temperature. .}

\item{TMAX}{\code{multilayer} - Maximum Temperature. .}

\item{P}{\code{multilayer} - Precipitation. .}

\item{BIO_01}{\code{rasterDestination} - Annual Mean Temperature. Path for new raster layer.}

\item{BIO_02}{\code{rasterDestination} - Mean Diurnal Range. Path for new raster layer.}

\item{BIO_03}{\code{rasterDestination} - Isothermality. Path for new raster layer.}

\item{BIO_04}{\code{rasterDestination} - Temperature Seasonality. Path for new raster layer.}

\item{BIO_05}{\code{rasterDestination} - Maximum Temperature of Warmest Month. Path for new raster layer.}

\item{BIO_06}{\code{rasterDestination} - Minimum Temperature of Coldest Month. Path for new raster layer.}

\item{BIO_07}{\code{rasterDestination} - Temperature Annual Range. Path for new raster layer.}

\item{BIO_08}{\code{rasterDestination} - Mean Temperature of Wettest Quarter. Path for new raster layer.}

\item{BIO_09}{\code{rasterDestination} - Mean Temperature of Driest Quarter. Path for new raster layer.}

\item{BIO_10}{\code{rasterDestination} - Mean Temperature of Warmest Quarter. Path for new raster layer.}

\item{BIO_11}{\code{rasterDestination} - Mean Temperature of Coldest Quarter. Path for new raster layer.}

\item{BIO_12}{\code{rasterDestination} - Annual Precipitation. Path for new raster layer.}

\item{BIO_13}{\code{rasterDestination} - Precipitation of Wettest Month. Path for new raster layer.}

\item{BIO_14}{\code{rasterDestination} - Precipitation of Driest Month. Path for new raster layer.}

\item{BIO_15}{\code{rasterDestination} - Precipitation Seasonality. Path for new raster layer.}

\item{BIO_16}{\code{rasterDestination} - Precipitation of Wettest Quarter. Path for new raster layer.}

\item{BIO_17}{\code{rasterDestination} - Precipitation of Driest Quarter. Path for new raster layer.}

\item{BIO_18}{\code{rasterDestination} - Precipitation of Warmest Quarter. Path for new raster layer.}

\item{BIO_19}{\code{rasterDestination} - Precipitation of Coldest Quarter. Path for new raster layer.}

\item{SEASONALITY}{\code{enum}  of \verb{("(0) Coefficient of Variation", "(1) Standard Deviation")} - Temperature Seasonality. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Bioclimatic variables (sagang:bioclimaticvariables). ---------------- Arguments ----------------  TMEAN: Mean Temperature 	Argument type:	multilayer TMIN: Minimum Temperature 	Argument type:	multilayer TMAX: Maximum Temperature 	Argument type:	multilayer P: Precipitation 	Argument type:	multilayer BIO_01: Annual Mean Temperature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_02: Mean Diurnal Range 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_03: Isothermality 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_04: Temperature Seasonality 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_05: Maximum Temperature of Warmest Month 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_06: Minimum Temperature of Coldest Month 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_07: Temperature Annual Range 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_08: Mean Temperature of Wettest Quarter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_09: Mean Temperature of Driest Quarter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_10: Mean Temperature of Warmest Quarter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_11: Mean Temperature of Coldest Quarter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_12: Annual Precipitation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_13: Precipitation of Wettest Month 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_14: Precipitation of Driest Month 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_15: Precipitation Seasonality 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_16: Precipitation of Wettest Quarter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_17: Precipitation of Driest Quarter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_18: Precipitation of Warmest Quarter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BIO_19: Precipitation of Coldest Quarter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SEASONALITY: Temperature Seasonality 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Coefficient of Variation 		- 1: (1) Standard Deviation 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item BIO_01 - outputRaster - Annual Mean Temperature
\item BIO_02 - outputRaster - Mean Diurnal Range
\item BIO_03 - outputRaster - Isothermality
\item BIO_04 - outputRaster - Temperature Seasonality
\item BIO_05 - outputRaster - Maximum Temperature of Warmest Month
\item BIO_06 - outputRaster - Minimum Temperature of Coldest Month
\item BIO_07 - outputRaster - Temperature Annual Range
\item BIO_08 - outputRaster - Mean Temperature of Wettest Quarter
\item BIO_09 - outputRaster - Mean Temperature of Driest Quarter
\item BIO_10 - outputRaster - Mean Temperature of Warmest Quarter
\item BIO_11 - outputRaster - Mean Temperature of Coldest Quarter
\item BIO_12 - outputRaster - Annual Precipitation
\item BIO_13 - outputRaster - Precipitation of Wettest Month
\item BIO_14 - outputRaster - Precipitation of Driest Month
\item BIO_15 - outputRaster - Precipitation Seasonality
\item BIO_16 - outputRaster - Precipitation of Wettest Quarter
\item BIO_17 - outputRaster - Precipitation of Driest Quarter
\item BIO_18 - outputRaster - Precipitation of Warmest Quarter
\item BIO_19 - outputRaster - Precipitation of Coldest Quarter
}
}
}
