% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_changeagridsnodatavalue.R
\name{sagang_changeagridsnodatavalue}
\alias{sagang_changeagridsnodatavalue}
\title{QGIS algorithm - Change a grid's no-data value}
\usage{
sagang_changeagridsnodatavalue(
  GRID = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  VALUE = qgisprocess:::qgis_default_value(),
  CHANGE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Changed Grid. Path for new raster layer.}

\item{TYPE}{\code{enum}  of \verb{("(0) single value", "(1) value range")} - Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{VALUE}{\code{number} - No-Data Value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CHANGE}{\code{boolean} - Change Values. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Change a grid's no-data value (sagang:changeagridsnodatavalue). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OUTPUT: Changed Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TYPE: Type 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) single value 		- 1: (1) value range 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' VALUE: No-Data Value 	Default value:	-99999 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CHANGE: Change Values 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Changed Grid
}
}
}
