% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_changelongitudinalrangeforgrids.R
\name{sagang_changelongitudinalrangeforgrids}
\alias{sagang_changelongitudinalrangeforgrids}
\title{QGIS algorithm - Change longitudinal range for grids}
\usage{
sagang_changelongitudinalrangeforgrids(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  DIRECTION = qgisprocess:::qgis_default_value(),
  PATCH = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input. .}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{DIRECTION}{\code{enum}  of \verb{("(0) 0 - 360  -180 - 180", "(1) -180 - 180  0 - 360")} - Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PATCH}{\code{boolean} - Patch Last Column. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Change longitudinal range for grids (sagang:changelongitudinalrangeforgrids). ---------------- Arguments ----------------  INPUT: Input 	Argument type:	multilayer OUTPUT: Output 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIRECTION: Direction 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) 0 - 360  -180 - 180 		- 1: (1) -180 - 180  0 - 360 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' PATCH: Patch Last Column 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
