% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_channelnetworkanddrainagebasins.R
\name{sagang_channelnetworkanddrainagebasins}
\alias{sagang_channelnetworkanddrainagebasins}
\title{QGIS algorithm - Channel network and drainage basins}
\usage{
sagang_channelnetworkanddrainagebasins(
  DEM = qgisprocess:::qgis_default_value(),
  DIRECTION = qgisprocess:::qgis_default_value(),
  CONNECTION = qgisprocess:::qgis_default_value(),
  ORDER = qgisprocess:::qgis_default_value(),
  BASIN = qgisprocess:::qgis_default_value(),
  SEGMENTS = qgisprocess:::qgis_default_value(),
  BASINS = qgisprocess:::qgis_default_value(),
  NODES = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{DIRECTION}{\code{rasterDestination} - Flow Direction. Path for new raster layer.}

\item{CONNECTION}{\code{rasterDestination} - Flow Connectivity. Path for new raster layer.}

\item{ORDER}{\code{rasterDestination} - Strahler Order. Path for new raster layer.}

\item{BASIN}{\code{rasterDestination} - Drainage Basins. Path for new raster layer.}

\item{SEGMENTS}{\code{vectorDestination} - Channels. Path for new vector layer.}

\item{BASINS}{\code{vectorDestination} - Drainage Basins. Path for new vector layer.}

\item{NODES}{\code{vectorDestination} - Junctions. Path for new vector layer.}

\item{THRESHOLD}{\code{number} - Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Channel network and drainage basins (sagang:channelnetworkanddrainagebasins). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DIRECTION: Flow Direction 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONNECTION: Flow Connectivity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ORDER: Strahler Order 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BASIN: Drainage Basins 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SEGMENTS: Channels 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer BASINS: Drainage Basins 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer NODES: Junctions 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer THRESHOLD: Threshold 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item BASIN - outputRaster - Drainage Basins
\item BASINS - outputVector - Drainage Basins
\item CONNECTION - outputRaster - Flow Connectivity
\item DIRECTION - outputRaster - Flow Direction
\item NODES - outputVector - Junctions
\item ORDER - outputRaster - Strahler Order
\item SEGMENTS - outputVector - Channels
}
}
}
