% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_cliffmetrics.R
\name{sagang_cliffmetrics}
\alias{sagang_cliffmetrics}
\title{QGIS algorithm - Cliffmetrics}
\usage{
sagang_cliffmetrics(
  DEM = qgisprocess:::qgis_default_value(),
  SEDIMENT_TOP = qgisprocess:::qgis_default_value(),
  RASTER_COAST = qgisprocess:::qgis_default_value(),
  RASTER_NORMAL = qgisprocess:::qgis_default_value(),
  COAST_INITIAL = qgisprocess:::qgis_default_value(),
  CoastSeaHandiness = qgisprocess:::qgis_default_value(),
  StartEdgeUserCoastLine = qgisprocess:::qgis_default_value(),
  EndEdgeUserCoastLine = qgisprocess:::qgis_default_value(),
  COAST = qgisprocess:::qgis_default_value(),
  NORMALS = qgisprocess:::qgis_default_value(),
  CLIFF_TOP = qgisprocess:::qgis_default_value(),
  CLIFF_TOE = qgisprocess:::qgis_default_value(),
  COAST_POINT = qgisprocess:::qgis_default_value(),
  INVALID_NORMALS = qgisprocess:::qgis_default_value(),
  COAST_CURVATURE = qgisprocess:::qgis_default_value(),
  PROFILES = qgisprocess:::qgis_default_value(),
  StillWaterLevel = qgisprocess:::qgis_default_value(),
  CoastSmooth = qgisprocess:::qgis_default_value(),
  CoastSmoothWindow = qgisprocess:::qgis_default_value(),
  SavGolCoastPoly = qgisprocess:::qgis_default_value(),
  ScaleRasterOutput = qgisprocess:::qgis_default_value(),
  RandomCoastEdgeSearch = qgisprocess:::qgis_default_value(),
  CoastNormalLength = qgisprocess:::qgis_default_value(),
  EleTolerance = qgisprocess:::qgis_default_value(),
  OutPath = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{SEDIMENT_TOP}{\code{rasterDestination} - Sediment Top Elevation. Path for new raster layer.}

\item{RASTER_COAST}{\code{rasterDestination} - Coastline. Path for new raster layer.}

\item{RASTER_NORMAL}{\code{rasterDestination} - Normals. Path for new raster layer.}

\item{COAST_INITIAL}{\code{source} - User Defined Coastline Points. Path to a vector layer.}

\item{CoastSeaHandiness}{\code{enum}  of \verb{("(0) right", "(1) left")} - Sea handiness. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{StartEdgeUserCoastLine}{\code{enum}  of \verb{("(0) North", "(1) East", "(2) South", "(3) West")} - Start edge coastline. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EndEdgeUserCoastLine}{\code{enum}  of \verb{("(0) North", "(1) East", "(2) South", "(3) West")} - End edge coastline. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{COAST}{\code{vectorDestination} - Coastline. Path for new vector layer.}

\item{NORMALS}{\code{vectorDestination} - Coastline-Normal Profiles. Path for new vector layer.}

\item{CLIFF_TOP}{\code{vectorDestination} - Cliff Top Points. Path for new vector layer.}

\item{CLIFF_TOE}{\code{vectorDestination} - Cliff Toe Points. Path for new vector layer.}

\item{COAST_POINT}{\code{vectorDestination} - Coast Points. Path for new vector layer.}

\item{INVALID_NORMALS}{\code{vectorDestination} - Invalid Coastline-Normal Profiles. Path for new vector layer.}

\item{COAST_CURVATURE}{\code{vectorDestination} - Coastline Curvature. Path for new vector layer.}

\item{PROFILES}{\code{vectorDestination} - Profile Data. Path for new vector layer.}

\item{StillWaterLevel}{\code{number} - Still Water Level. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CoastSmooth}{\code{enum}  of \verb{("(0) none", "(1) running mean", "(2) Savitsky-Golay")} - Coastline Smoothing Algorithm. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CoastSmoothWindow}{\code{number} - Coastline Smoothing Window Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SavGolCoastPoly}{\code{number} - Polynomial Order for Savitsky-Golay. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ScaleRasterOutput}{\code{boolean} - Scale Raster Output Values. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RandomCoastEdgeSearch}{\code{boolean} - Random Edge for Coastline Search. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CoastNormalLength}{\code{number} - Length of Coastline Normals. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EleTolerance}{\code{number} - Vertical Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OutPath}{\code{file} - Main Output File Directory. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Cliffmetrics (sagang:cliffmetrics). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SEDIMENT_TOP: Sediment Top Elevation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RASTER_COAST: Coastline 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RASTER_NORMAL: Normals 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer COAST_INITIAL: User Defined Coastline Points (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer CoastSeaHandiness: Sea handiness 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) right 		- 1: (1) left 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' StartEdgeUserCoastLine: Start edge coastline 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) North 		- 1: (1) East 		- 2: (2) South 		- 3: (3) West 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' EndEdgeUserCoastLine: End edge coastline 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) North 		- 1: (1) East 		- 2: (2) South 		- 3: (3) West 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' COAST: Coastline 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer NORMALS: Coastline-Normal Profiles 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer CLIFF_TOP: Cliff Top Points 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer CLIFF_TOE: Cliff Toe Points 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer COAST_POINT: Coast Points 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer INVALID_NORMALS: Invalid Coastline-Normal Profiles 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer COAST_CURVATURE: Coastline Curvature 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer PROFILES: Profile Data 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer StillWaterLevel: Still Water Level 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CoastSmooth: Coastline Smoothing Algorithm 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) none 		- 1: (1) running mean 		- 2: (2) Savitsky-Golay 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CoastSmoothWindow: Coastline Smoothing Window Size 	Default value:	30 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SavGolCoastPoly: Polynomial Order for Savitsky-Golay 	Default value:	4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ScaleRasterOutput: Scale Raster Output Values 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RandomCoastEdgeSearch: Random Edge for Coastline Search 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CoastNormalLength: Length of Coastline Normals 	Default value:	500 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EleTolerance: Vertical Tolerance 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OutPath: Main Output File Directory 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLIFF_TOE - outputVector - Cliff Toe Points
\item CLIFF_TOP - outputVector - Cliff Top Points
\item COAST - outputVector - Coastline
\item COAST_CURVATURE - outputVector - Coastline Curvature
\item COAST_POINT - outputVector - Coast Points
\item INVALID_NORMALS - outputVector - Invalid Coastline-Normal Profiles
\item NORMALS - outputVector - Coastline-Normal Profiles
\item PROFILES - outputVector - Profile Data
\item RASTER_COAST - outputRaster - Coastline
\item RASTER_NORMAL - outputRaster - Normals
\item SEDIMENT_TOP - outputRaster - Sediment Top Elevation
}
}
}
