% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_climateclassification.R
\name{sagang_climateclassification}
\alias{sagang_climateclassification}
\title{QGIS algorithm - Climate classification}
\usage{
sagang_climateclassification(
  T = qgisprocess:::qgis_default_value(),
  P = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{T}{\code{multilayer} - Temperature. .}

\item{P}{\code{multilayer} - Precipitation. .}

\item{CLASSES}{\code{rasterDestination} - Climate Classification. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) Koeppen-Geiger", "(1) Koeppen-Geiger without As/Aw differentiation", "(2) Koeppen-Geiger after Peel et al. (2007)", "(3) Wissmann (1939)", "(4) Thornthwaite (1931)", "(5) Troll-Paffen")} - Classification. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Climate classification (sagang:climateclassification). ---------------- Arguments ----------------  T: Temperature 	Argument type:	multilayer P: Precipitation 	Argument type:	multilayer CLASSES: Climate Classification 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Classification 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Koeppen-Geiger 		- 1: (1) Koeppen-Geiger without As/Aw differentiation 		- 2: (2) Koeppen-Geiger after Peel et al. (2007) 		- 3: (3) Wissmann (1939) 		- 4: (4) Thornthwaite (1931) 		- 5: (5) Troll-Paffen 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Climate Classification
}
}
}
