% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_cliprasterwithpolygon.R
\name{sagang_cliprasterwithpolygon}
\alias{sagang_cliprasterwithpolygon}
\title{QGIS algorithm - Clip raster with polygon}
\usage{
sagang_cliprasterwithpolygon(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  POLYGONS = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input. .}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{EXTENT}{\code{enum}  of \verb{("(0) original", "(1) polygons", "(2) crop to data")} - Target Extent. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Clip raster with polygon (sagang:cliprasterwithpolygon). ---------------- Arguments ----------------  INPUT: Input 	Argument type:	multilayer OUTPUT: Output 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer POLYGONS: Polygons 	Argument type:	source 	Acceptable values: 		- Path to a vector layer EXTENT: Target Extent 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) original 		- 1: (1) polygons 		- 2: (2) crop to data 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
