% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_clouddetection.R
\name{sagang_clouddetection}
\alias{sagang_clouddetection}
\title{QGIS algorithm - Cloud detection}
\usage{
sagang_clouddetection(
  BLUE = qgisprocess:::qgis_default_value(),
  GREEN = qgisprocess:::qgis_default_value(),
  RED = qgisprocess:::qgis_default_value(),
  NIR = qgisprocess:::qgis_default_value(),
  CLOUDS = qgisprocess:::qgis_default_value(),
  SWIR1 = qgisprocess:::qgis_default_value(),
  SWIR2 = qgisprocess:::qgis_default_value(),
  THERMAL = qgisprocess:::qgis_default_value(),
  CIRRUS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{BLUE}{\code{raster} - Blue. Path to a raster layer.}

\item{GREEN}{\code{raster} - Green. Path to a raster layer.}

\item{RED}{\code{raster} - Red. Path to a raster layer.}

\item{NIR}{\code{raster} - Near Infrared. Path to a raster layer.}

\item{CLOUDS}{\code{rasterDestination} - Clouds. Path for new raster layer.}

\item{SWIR1}{\code{raster} - Shortwave Infrared 1. Path to a raster layer.}

\item{SWIR2}{\code{raster} - Shortwave Infrared 2. Path to a raster layer.}

\item{THERMAL}{\code{raster} - Temperature (Kelvin). Path to a raster layer.}

\item{CIRRUS}{\code{raster} - Cirrus. Path to a raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Cloud detection (sagang:clouddetection). ---------------- Arguments ----------------  BLUE: Blue 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GREEN: Green 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RED: Red 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer NIR: Near Infrared 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CLOUDS: Clouds 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SWIR1: Shortwave Infrared 1 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SWIR2: Shortwave Infrared 2 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer THERMAL: Temperature (Kelvin) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CIRRUS: Cirrus (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLOUDS - outputRaster - Clouds
}
}
}
