% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_cloudoverlap.R
\name{sagang_cloudoverlap}
\alias{sagang_cloudoverlap}
\title{QGIS algorithm - Cloud overlap}
\usage{
sagang_cloudoverlap(
  COVERS = qgisprocess:::qgis_default_value(),
  HEIGHTS = qgisprocess:::qgis_default_value(),
  GROUND = qgisprocess:::qgis_default_value(),
  WIND = qgisprocess:::qgis_default_value(),
  CBASE = qgisprocess:::qgis_default_value(),
  COVER = qgisprocess:::qgis_default_value(),
  BLOCKS = qgisprocess:::qgis_default_value(),
  INTERVAL = qgisprocess:::qgis_default_value(),
  MINCOVER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{COVERS}{\code{multilayer} - Cloud Fractions. .}

\item{HEIGHTS}{\code{multilayer} - Heights. .}

\item{GROUND}{\code{raster} - Surface Elevation. Path to a raster layer.}

\item{WIND}{\code{raster} - Wind effect. Path to a raster layer.}

\item{CBASE}{\code{raster} - Cloud Base. Path to a raster layer.}

\item{COVER}{\code{rasterDestination} - Total Cloud Cover. Path for new raster layer.}

\item{BLOCKS}{\code{rasterDestination} - Number of Cloud Blocks. Path for new raster layer.}

\item{INTERVAL}{\code{number} - Interval. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MINCOVER}{\code{number} - Minimum Cloud Cover Fraction. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Cloud overlap (sagang:cloudoverlap). ---------------- Arguments ----------------  COVERS: Cloud Fractions 	Argument type:	multilayer HEIGHTS: Heights 	Argument type:	multilayer GROUND: Surface Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer WIND: Wind effect 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CBASE: Cloud Base 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer COVER: Total Cloud Cover 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BLOCKS: Number of Cloud Blocks 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer INTERVAL: Interval 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MINCOVER: Minimum Cloud Cover Fraction 	Default value:	0.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item BLOCKS - outputRaster - Number of Cloud Blocks
\item COVER - outputRaster - Total Cloud Cover
}
}
}
