% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_coldairflow.R
\name{sagang_coldairflow}
\alias{sagang_coldairflow}
\title{QGIS algorithm - Cold air flow}
\usage{
sagang_coldairflow(
  DEM = qgisprocess:::qgis_default_value(),
  PRODUCTION = qgisprocess:::qgis_default_value(),
  PRODUCTION_DEFAULT = qgisprocess:::qgis_default_value(),
  FRICTION = qgisprocess:::qgis_default_value(),
  FRICTION_DEFAULT = qgisprocess:::qgis_default_value(),
  AIR = qgisprocess:::qgis_default_value(),
  VELOCITY = qgisprocess:::qgis_default_value(),
  RESET = qgisprocess:::qgis_default_value(),
  TIME_STOP = qgisprocess:::qgis_default_value(),
  TIME_UPDATE = qgisprocess:::qgis_default_value(),
  EDGE = qgisprocess:::qgis_default_value(),
  DELAY = qgisprocess:::qgis_default_value(),
  T_AIR = qgisprocess:::qgis_default_value(),
  T_AIR_COLD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{PRODUCTION}{\code{raster} - Production. Path to a raster layer.}

\item{PRODUCTION_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FRICTION}{\code{raster} - Surface Friction Coefficient. Path to a raster layer.}

\item{FRICTION_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{AIR}{\code{rasterDestination} - Cold Air Height. Path for new raster layer.}

\item{VELOCITY}{\code{rasterDestination} - Velocity. Path for new raster layer.}

\item{RESET}{\code{boolean} - Reset. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME_STOP}{\code{number} - Simulation Time (h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME_UPDATE}{\code{number} - Map Update Frequency (min). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EDGE}{\code{enum}  of \verb{("(0) closed", "(1) open")} - Edge. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DELAY}{\code{number} - Time Step Adjustment. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T_AIR}{\code{number} - Surrounding Air Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T_AIR_COLD}{\code{number} - Cold Air Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Cold air flow (sagang:coldairflow). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer PRODUCTION: Production (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer PRODUCTION_DEFAULT: Default 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FRICTION: Surface Friction Coefficient (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FRICTION_DEFAULT: Default 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression AIR: Cold Air Height 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VELOCITY: Velocity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RESET: Reset 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME_STOP: Simulation Time (h) 	Default value:	6 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME_UPDATE: Map Update Frequency (min) 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EDGE: Edge 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) closed 		- 1: (1) open 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DELAY: Time Step Adjustment 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T_AIR: Surrounding Air Temperature 	Default value:	15 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T_AIR_COLD: Cold Air Temperature 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item AIR - outputRaster - Cold Air Height
\item VELOCITY - outputRaster - Velocity
}
}
}
