% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_confusionmatrixpolygonsgrid.R
\name{sagang_confusionmatrixpolygonsgrid}
\alias{sagang_confusionmatrixpolygonsgrid}
\title{QGIS algorithm - Confusion matrix (polygons / grid)}
\usage{
sagang_confusionmatrixpolygonsgrid(
  GRID = qgisprocess:::qgis_default_value(),
  GRID_VALUES = qgisprocess:::qgis_default_value(),
  GRID_LUT = qgisprocess:::qgis_default_value(),
  GRID_LUT_MIN = qgisprocess:::qgis_default_value(),
  GRID_LUT_MAX = qgisprocess:::qgis_default_value(),
  GRID_LUT_NAM = qgisprocess:::qgis_default_value(),
  POLYGONS = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  CONFUSION = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  SUMMARY = qgisprocess:::qgis_default_value(),
  NO_CLASS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Classification. Path to a raster layer.}

\item{GRID_VALUES}{\code{enum}  of \verb{("(0) values are class identifiers", "(1) use look-up table")} - Value Interpretation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRID_LUT}{\code{source} - Look-up Table. Path to a vector layer.}

\item{GRID_LUT_MIN}{\code{field} - Value. The name of an existing field. ; delimited list of existing field names.}

\item{GRID_LUT_MAX}{\code{field} - Value (Maximum). The name of an existing field. ; delimited list of existing field names.}

\item{GRID_LUT_NAM}{\code{field} - Name. The name of an existing field. ; delimited list of existing field names.}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{FIELD}{\code{field} - Classes. The name of an existing field. ; delimited list of existing field names.}

\item{CONFUSION}{\code{vectorDestination} - Confusion Matrix. Path for new vector layer.}

\item{CLASSES}{\code{vectorDestination} - Class Values. Path for new vector layer.}

\item{SUMMARY}{\code{vectorDestination} - Summary. Path for new vector layer.}

\item{NO_CLASS}{\code{boolean} - Unclassified. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Confusion matrix (polygons / grid) (sagang:confusionmatrixpolygonsgrid). ---------------- Arguments ----------------  GRID: Classification 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GRID_VALUES: Value Interpretation 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) values are class identifiers 		- 1: (1) use look-up table 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' GRID_LUT: Look-up Table (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer GRID_LUT_MIN: Value 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names GRID_LUT_MAX: Value (Maximum) 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names GRID_LUT_NAM: Name 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names POLYGONS: Polygons 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Classes 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names CONFUSION: Confusion Matrix 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer CLASSES: Class Values 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer SUMMARY: Summary 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer NO_CLASS: Unclassified 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputVector - Class Values
\item CONFUSION - outputVector - Confusion Matrix
\item SUMMARY - outputVector - Summary
}
}
}
