% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_confusionmatrixtwogrids.R
\name{sagang_confusionmatrixtwogrids}
\alias{sagang_confusionmatrixtwogrids}
\title{QGIS algorithm - Confusion matrix (two grids)}
\usage{
sagang_confusionmatrixtwogrids(
  ONE = qgisprocess:::qgis_default_value(),
  ONE_LUT = qgisprocess:::qgis_default_value(),
  ONE_LUT_MIN = qgisprocess:::qgis_default_value(),
  ONE_LUT_MAX = qgisprocess:::qgis_default_value(),
  ONE_LUT_NAM = qgisprocess:::qgis_default_value(),
  TWO = qgisprocess:::qgis_default_value(),
  TWO_LUT = qgisprocess:::qgis_default_value(),
  TWO_LUT_MIN = qgisprocess:::qgis_default_value(),
  TWO_LUT_MAX = qgisprocess:::qgis_default_value(),
  TWO_LUT_NAM = qgisprocess:::qgis_default_value(),
  COMBINED = qgisprocess:::qgis_default_value(),
  NOCHANGE = qgisprocess:::qgis_default_value(),
  NODATA = qgisprocess:::qgis_default_value(),
  CONFUSION = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  SUMMARY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ONE}{\code{raster} - Classification 1. Path to a raster layer.}

\item{ONE_LUT}{\code{source} - Look-up Table. Path to a vector layer.}

\item{ONE_LUT_MIN}{\code{field} - Value. The name of an existing field. ; delimited list of existing field names.}

\item{ONE_LUT_MAX}{\code{field} - Value (Maximum). The name of an existing field. ; delimited list of existing field names.}

\item{ONE_LUT_NAM}{\code{field} - Name. The name of an existing field. ; delimited list of existing field names.}

\item{TWO}{\code{raster} - Classification 2. Path to a raster layer.}

\item{TWO_LUT}{\code{source} - Look-up Table. Path to a vector layer.}

\item{TWO_LUT_MIN}{\code{field} - Value. The name of an existing field. ; delimited list of existing field names.}

\item{TWO_LUT_MAX}{\code{field} - Value (Maximum). The name of an existing field. ; delimited list of existing field names.}

\item{TWO_LUT_NAM}{\code{field} - Name. The name of an existing field. ; delimited list of existing field names.}

\item{COMBINED}{\code{rasterDestination} - Combined Classes. Path for new raster layer.}

\item{NOCHANGE}{\code{boolean} - Report Unchanged Classes. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NODATA}{\code{boolean} - Include Unclassified Cells. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CONFUSION}{\code{vectorDestination} - Confusion Matrix. Path for new vector layer.}

\item{OUTPUT}{\code{enum}  of \verb{("(0) cells", "(1) percent", "(2) area")} - Output as.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CLASSES}{\code{vectorDestination} - Class Values. Path for new vector layer.}

\item{SUMMARY}{\code{vectorDestination} - Summary. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Confusion matrix (two grids) (sagang:confusionmatrixtwogrids). ---------------- Arguments ----------------  ONE: Classification 1 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer ONE_LUT: Look-up Table (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ONE_LUT_MIN: Value 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names ONE_LUT_MAX: Value (Maximum) 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names ONE_LUT_NAM: Name 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TWO: Classification 2 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TWO_LUT: Look-up Table (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer TWO_LUT_MIN: Value 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TWO_LUT_MAX: Value (Maximum) 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TWO_LUT_NAM: Name 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names COMBINED: Combined Classes 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer NOCHANGE: Report Unchanged Classes 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NODATA: Include Unclassified Cells 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CONFUSION: Confusion Matrix 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer OUTPUT: Output as... 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) cells 		- 1: (1) percent 		- 2: (2) area 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CLASSES: Class Values 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer SUMMARY: Summary 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputVector - Class Values
\item COMBINED - outputRaster - Combined Classes
\item CONFUSION - outputVector - Confusion Matrix
\item SUMMARY - outputVector - Summary
}
}
}
