% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_connectivityanalysis.R
\name{sagang_connectivityanalysis}
\alias{sagang_connectivityanalysis}
\title{QGIS algorithm - Connectivity analysis}
\usage{
sagang_connectivityanalysis(
  INPUT_GRID = qgisprocess:::qgis_default_value(),
  FILTERED_MASK = qgisprocess:::qgis_default_value(),
  FILTER = qgisprocess:::qgis_default_value(),
  SIZE = qgisprocess:::qgis_default_value(),
  SYMBOLIC_IMAGE = qgisprocess:::qgis_default_value(),
  OUTLINES = qgisprocess:::qgis_default_value(),
  BORDER_PIXEL_CENTERS = qgisprocess:::qgis_default_value(),
  REMOVE_MARGINAL_REGIONS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_GRID}{\code{raster} - Input Binary Grid. Path to a raster layer.}

\item{FILTERED_MASK}{\code{rasterDestination} - Filtered Image. Path for new raster layer.}

\item{FILTER}{\code{boolean} - Apply Filter?. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SIZE}{\code{number} - Filter Size (Radius). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SYMBOLIC_IMAGE}{\code{rasterDestination} - Symbolic Image. Path for new raster layer.}

\item{OUTLINES}{\code{vectorDestination} - Outlines. Path for new vector layer.}

\item{BORDER_PIXEL_CENTERS}{\code{boolean} - Pixel Centers?. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{REMOVE_MARGINAL_REGIONS}{\code{boolean} - Remove Border Regions?. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Connectivity analysis (sagang:connectivityanalysis). ---------------- Arguments ----------------  INPUT_GRID: Input Binary Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FILTERED_MASK: Filtered Image 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FILTER: Apply Filter? 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SIZE: Filter Size (Radius) 	Default value:	3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SYMBOLIC_IMAGE: Symbolic Image 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer OUTLINES: Outlines 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer BORDER_PIXEL_CENTERS: Pixel Centers? 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression REMOVE_MARGINAL_REGIONS: Remove Border Regions? 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FILTERED_MASK - outputRaster - Filtered Image
\item OUTLINES - outputVector - Outlines
\item SYMBOLIC_IMAGE - outputRaster - Symbolic Image
}
}
}
