% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_constantgrid.R
\name{sagang_constantgrid}
\alias{sagang_constantgrid}
\title{QGIS algorithm - Constant grid}
\usage{
sagang_constantgrid(
  NAME = qgisprocess:::qgis_default_value(),
  CONST = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  DEFINITION = qgisprocess:::qgis_default_value(),
  USER_SIZE = qgisprocess:::qgis_default_value(),
  USER_XMIN = qgisprocess:::qgis_default_value(),
  USER_XMAX = qgisprocess:::qgis_default_value(),
  USER_YMIN = qgisprocess:::qgis_default_value(),
  USER_YMAX = qgisprocess:::qgis_default_value(),
  USER_COLS = qgisprocess:::qgis_default_value(),
  USER_ROWS = qgisprocess:::qgis_default_value(),
  USER_FITS = qgisprocess:::qgis_default_value(),
  TEMPLATE = qgisprocess:::qgis_default_value(),
  OUT_GRID = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{NAME}{\code{string} - Name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CONST}{\code{number} - Constant Value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TYPE}{\code{enum}  of \verb{("(0) bit", "(1) unsigned 1 byte integer", "(2) signed 1 byte integer", "(3) unsigned 2 byte integer", "(4) signed 2 byte integer", "(5) unsigned 8 byte integer", "(6) signed 8 byte integer", "(7) 4 byte floating point number", "(8) 8 byte floating point number")} - Data Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DEFINITION}{\code{enum}  of \verb{("(0) user defined", "(1) grid or grid system")} - Target Grid System. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{USER_SIZE}{\code{number} - Cellsize. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{USER_XMIN}{\code{number} - West. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{USER_XMAX}{\code{number} - East. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{USER_YMIN}{\code{number} - South. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{USER_YMAX}{\code{number} - North. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{USER_COLS}{\code{number} - Columns. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{USER_ROWS}{\code{number} - Rows. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{USER_FITS}{\code{enum}  of \verb{("(0) nodes", "(1) cells")} - Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TEMPLATE}{\code{raster} - Target System. Path to a raster layer.}

\item{OUT_GRID}{\code{rasterDestination} - Target Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Constant grid (sagang:constantgrid). ---------------- Arguments ----------------  NAME: Name 	Default value:	Constant Grid 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CONST: Constant Value 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TYPE: Data Type 	Default value:	7 	Argument type:	enum 	Available values: 		- 0: (0) bit 		- 1: (1) unsigned 1 byte integer 		- 2: (2) signed 1 byte integer 		- 3: (3) unsigned 2 byte integer 		- 4: (4) signed 2 byte integer 		- 5: (5) unsigned 8 byte integer 		- 6: (6) signed 8 byte integer 		- 7: (7) 4 byte floating point number 		- 8: (8) 8 byte floating point number 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DEFINITION: Target Grid System 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) user defined 		- 1: (1) grid or grid system 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' USER_SIZE: Cellsize 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression USER_XMIN: West 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression USER_XMAX: East 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression USER_YMIN: South 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression USER_YMAX: North 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression USER_COLS: Columns 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression USER_ROWS: Rows 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression USER_FITS: Fit 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) nodes 		- 1: (1) cells 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' TEMPLATE: Target System (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OUT_GRID: Target Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUT_GRID - outputRaster - Target Grid
}
}
}
