% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_convertlinestopoints.R
\name{sagang_convertlinestopoints}
\alias{sagang_convertlinestopoints}
\title{QGIS algorithm - Convert lines to points}
\usage{
sagang_convertlinestopoints(
  LINES = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  ADD = qgisprocess:::qgis_default_value(),
  METHOD_INSERT = qgisprocess:::qgis_default_value(),
  DIST = qgisprocess:::qgis_default_value(),
  ADD_POINT_ORDER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{POINTS}{\code{vectorDestination} - Points. Path for new vector layer.}

\item{ADD}{\code{boolean} - Insert Additional Points. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD_INSERT}{\code{enum}  of \verb{("(0) per line segment", "(1) per line", "(2) from line center")} - Insertion. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DIST}{\code{number} - Insertion Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ADD_POINT_ORDER}{\code{boolean} - Add Point Order. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Convert lines to points (sagang:convertlinestopoints). ---------------- Arguments ----------------  LINES: Lines 	Argument type:	source 	Acceptable values: 		- Path to a vector layer POINTS: Points 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer ADD: Insert Additional Points 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD_INSERT: Insertion 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) per line segment 		- 1: (1) per line 		- 2: (2) from line center 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DIST: Insertion Distance 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ADD_POINT_ORDER: Add Point Order 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item POINTS - outputVector - Points
}
}
}
