% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_convertpointstolines.R
\name{sagang_convertpointstolines}
\alias{sagang_convertpointstolines}
\title{QGIS algorithm - Convert points to line(s)}
\usage{
sagang_convertpointstolines(
  LINES = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  ORDER = qgisprocess:::qgis_default_value(),
  SEPARATE = qgisprocess:::qgis_default_value(),
  ELEVATION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LINES}{\code{vectorDestination} - Lines. Path for new vector layer.}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{ORDER}{\code{field} - Order by.... The name of an existing field. ; delimited list of existing field names.}

\item{SEPARATE}{\code{field} - Separate by.... The name of an existing field. ; delimited list of existing field names.}

\item{ELEVATION}{\code{field} - Elevation. The name of an existing field. ; delimited list of existing field names.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Convert points to line(s) (sagang:convertpointstolines). ---------------- Arguments ----------------  LINES: Lines 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ORDER: Order by... 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names SEPARATE: Separate by... 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names ELEVATION: Elevation 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names
}
\details{
\subsection{Outputs description}{
\itemize{
\item LINES - outputVector - Lines
}
}
}
