% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_converttabletopoints.R
\name{sagang_converttabletopoints}
\alias{sagang_converttabletopoints}
\title{QGIS algorithm - Convert table to points}
\usage{
sagang_converttabletopoints(
  POINTS = qgisprocess:::qgis_default_value(),
  TABLE = qgisprocess:::qgis_default_value(),
  X = qgisprocess:::qgis_default_value(),
  Y = qgisprocess:::qgis_default_value(),
  Z = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{vectorDestination} - Points. Path for new vector layer.}

\item{TABLE}{\code{source} - Table. Path to a vector layer.}

\item{X}{\code{field} - X. The name of an existing field. ; delimited list of existing field names.}

\item{Y}{\code{field} - Y. The name of an existing field. ; delimited list of existing field names.}

\item{Z}{\code{field} - Z. The name of an existing field. ; delimited list of existing field names.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Convert table to points (sagang:converttabletopoints). ---------------- Arguments ----------------  POINTS: Points 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer TABLE: Table 	Argument type:	source 	Acceptable values: 		- Path to a vector layer X: X 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names Y: Y 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names Z: Z 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names
}
\details{
\subsection{Outputs description}{
\itemize{
\item POINTS - outputVector - Points
}
}
}
