% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_convertvertextype2d3d.R
\name{sagang_convertvertextype2d3d}
\alias{sagang_convertvertextype2d3d}
\title{QGIS algorithm - Convert vertex type (2d/3d)}
\usage{
sagang_convertvertextype2d3d(
  INPUT = qgisprocess:::qgis_default_value(),
  FIELD_Z = qgisprocess:::qgis_default_value(),
  FIELD_M = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input. Path to a vector layer.}

\item{FIELD_Z}{\code{field} - Z. The name of an existing field. ; delimited list of existing field names.}

\item{FIELD_M}{\code{field} - M. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{vectorDestination} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Convert vertex type (2d/3d) (sagang:convertvertextype2d3d). ---------------- Arguments ----------------  INPUT: Input 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD_Z: Z 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FIELD_M: M 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names OUTPUT: Output 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output
}
}
}
