% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_convexhull.R
\name{sagang_convexhull}
\alias{sagang_convexhull}
\title{QGIS algorithm - Convex hull}
\usage{
sagang_convexhull(
  SHAPES = qgisprocess:::qgis_default_value(),
  HULLS = qgisprocess:::qgis_default_value(),
  BOXES = qgisprocess:::qgis_default_value(),
  POLYPOINTS = qgisprocess:::qgis_default_value(),
  POLYGONCVX = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Points. Path to a vector layer.}

\item{HULLS}{\code{vectorDestination} - Convex Hull. Path for new vector layer.}

\item{BOXES}{\code{vectorDestination} - Minimum Bounding Box. Path for new vector layer.}

\item{POLYPOINTS}{\code{enum}  of \verb{("(0) one hull for all shapes", "(1) one hull per shape", "(2) one hull per shape part")} - Hull Construction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{POLYGONCVX}{\code{boolean} - Polygon Convexity. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Convex hull (sagang:convexhull). ---------------- Arguments ----------------  SHAPES: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer HULLS: Convex Hull 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer BOXES: Minimum Bounding Box 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer POLYPOINTS: Hull Construction 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) one hull for all shapes 		- 1: (1) one hull per shape 		- 2: (2) one hull per shape part 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' POLYGONCVX: Polygon Convexity 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item BOXES - outputVector - Minimum Bounding Box
\item HULLS - outputVector - Convex Hull
}
}
}
