% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_coordinateconversiongrids.R
\name{sagang_coordinateconversiongrids}
\alias{sagang_coordinateconversiongrids}
\title{QGIS algorithm - Coordinate conversion (grids)}
\usage{
sagang_coordinateconversiongrids(
  SOURCE_X = qgisprocess:::qgis_default_value(),
  SOURCE_Y = qgisprocess:::qgis_default_value(),
  TARGET_X = qgisprocess:::qgis_default_value(),
  TARGET_Y = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SOURCE_X}{\code{raster} - X Coordinate Source. Path to a raster layer.}

\item{SOURCE_Y}{\code{raster} - Y Coordinate Source. Path to a raster layer.}

\item{TARGET_X}{\code{rasterDestination} - Projected X Coordinates. Path for new raster layer.}

\item{TARGET_Y}{\code{rasterDestination} - Projected Y Coordinates. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Coordinate conversion (grids) (sagang:coordinateconversiongrids). ---------------- Arguments ----------------  SOURCE_X: X Coordinate Source 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SOURCE_Y: Y Coordinate Source 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TARGET_X: Projected X Coordinates 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TARGET_Y: Projected Y Coordinates 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item TARGET_X - outputRaster - Projected X Coordinates
\item TARGET_Y - outputRaster - Projected Y Coordinates
}
}
}
