% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_coordinateconversiontable.R
\name{sagang_coordinateconversiontable}
\alias{sagang_coordinateconversiontable}
\title{QGIS algorithm - Coordinate conversion (table)}
\usage{
sagang_coordinateconversiontable(
  TABLE = qgisprocess:::qgis_default_value(),
  SOURCE_X = qgisprocess:::qgis_default_value(),
  SOURCE_Y = qgisprocess:::qgis_default_value(),
  TARGET_X = qgisprocess:::qgis_default_value(),
  TARGET_Y = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TABLE}{\code{source} - Table. Path to a vector layer.}

\item{SOURCE_X}{\code{field} - X Coordinate Source. The name of an existing field. ; delimited list of existing field names.}

\item{SOURCE_Y}{\code{field} - Y Coordinate Source. The name of an existing field. ; delimited list of existing field names.}

\item{TARGET_X}{\code{field} - Projected X Coordinates. The name of an existing field. ; delimited list of existing field names.}

\item{TARGET_Y}{\code{field} - Projected Y Coordinates. The name of an existing field. ; delimited list of existing field names.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Coordinate conversion (table) (sagang:coordinateconversiontable). ---------------- Arguments ----------------  TABLE: Table 	Argument type:	source 	Acceptable values: 		- Path to a vector layer SOURCE_X: X Coordinate Source 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names SOURCE_Y: Y Coordinate Source 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TARGET_X: Projected X Coordinates 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TARGET_Y: Projected Y Coordinates 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
