% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_covereddistance.R
\name{sagang_covereddistance}
\alias{sagang_covereddistance}
\title{QGIS algorithm - Covered distance}
\usage{
sagang_covereddistance(
  INPUT = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Grids. .}

\item{RESULT}{\code{rasterDestination} - Covered Distance. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Covered distance (sagang:covereddistance). ---------------- Arguments ----------------  INPUT: Grids 	Argument type:	multilayer RESULT: Covered Distance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Covered Distance
}
}
}
