% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_createrastercataloguefromfiles.R
\name{sagang_createrastercataloguefromfiles}
\alias{sagang_createrastercataloguefromfiles}
\title{QGIS algorithm - Create raster catalogue from files}
\usage{
sagang_createrastercataloguefromfiles(
  CATALOGUE = qgisprocess:::qgis_default_value(),
  FILES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CATALOGUE}{\code{vectorDestination} - Raster Catalogue. Path for new vector layer.}

\item{FILES}{\code{file} - Files. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Create raster catalogue from files (sagang:createrastercataloguefromfiles). ---------------- Arguments ----------------  CATALOGUE: Raster Catalogue 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer FILES: Files 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file
}
\details{
\subsection{Outputs description}{
\itemize{
\item CATALOGUE - outputVector - Raster Catalogue
}
}
}
