% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_createtileshapefromvirtualpointcloud.R
\name{sagang_createtileshapefromvirtualpointcloud}
\alias{sagang_createtileshapefromvirtualpointcloud}
\title{QGIS algorithm - Create tileshape from virtual point cloud}
\usage{
sagang_createtileshapefromvirtualpointcloud(
  FILENAME = qgisprocess:::qgis_default_value(),
  TILE_SHP = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FILENAME}{\code{file} - Filename. Path to a file.}

\item{TILE_SHP}{\code{vectorDestination} - Tileshape. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Create tileshape from virtual point cloud (sagang:createtileshapefromvirtualpointcloud). ---------------- Arguments ----------------  FILENAME: Filename 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file TILE_SHP: Tileshape 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item TILE_SHP - outputVector - Tileshape
}
}
}
