% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_createvirtualpointclouddataset.R
\name{sagang_createvirtualpointclouddataset}
\alias{sagang_createvirtualpointclouddataset}
\title{QGIS algorithm - Create virtual point cloud dataset}
\usage{
sagang_createvirtualpointclouddataset(
  FILES = qgisprocess:::qgis_default_value(),
  INPUT_FILE_LIST = qgisprocess:::qgis_default_value(),
  FILENAME = qgisprocess:::qgis_default_value(),
  METHOD_PATHS = qgisprocess:::qgis_default_value(),
  USE_HEADER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FILES}{\code{file} - Input Files. Path to a file.}

\item{INPUT_FILE_LIST}{\code{file} - Input File List. Path to a file.}

\item{FILENAME}{\code{file} - Filename. Path to a file.}

\item{METHOD_PATHS}{\code{enum}  of \verb{("(0) absolute", "(1) relative")} - File Paths. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{USE_HEADER}{\code{boolean} - Use Header File. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Create virtual point cloud dataset (sagang:createvirtualpointclouddataset). ---------------- Arguments ----------------  FILES: Input Files 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file INPUT_FILE_LIST: Input File List 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file FILENAME: Filename 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file METHOD_PATHS: File Paths 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) absolute 		- 1: (1) relative 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' USE_HEADER: Use Header File 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
