% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_createvirtualrastervrt.R
\name{sagang_createvirtualrastervrt}
\alias{sagang_createvirtualrastervrt}
\title{QGIS algorithm - Create virtual raster (vrt)}
\usage{
sagang_createvirtualrastervrt(
  FILES = qgisprocess:::qgis_default_value(),
  FILE_LIST = qgisprocess:::qgis_default_value(),
  VRT_NAME = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  RESOLUTION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FILES}{\code{file} - Files. Path to a file.}

\item{FILE_LIST}{\code{file} - Input File List. Path to a file.}

\item{VRT_NAME}{\code{file} - VRT Filename. Path to a file.}

\item{RESAMPLING}{\code{enum}  of \verb{("(0) nearest", "(1) bilinear", "(2) cubic", "(3) cubic spline", "(4) lanczos", "(5) average", "(6) mode")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESOLUTION}{\code{enum}  of \verb{("(0) highest", "(1) lowest", "(2) average")} - Resolution. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Create virtual raster (vrt) (sagang:createvirtualrastervrt). ---------------- Arguments ----------------  FILES: Files 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file FILE_LIST: Input File List 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file VRT_NAME: VRT Filename 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file RESAMPLING: Resampling 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) nearest 		- 1: (1) bilinear 		- 2: (2) cubic 		- 3: (3) cubic spline 		- 4: (4) lanczos 		- 5: (5) average 		- 6: (6) mode 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RESOLUTION: Resolution 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) highest 		- 1: (1) lowest 		- 2: (2) average 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
