% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_croptodata.R
\name{sagang_croptodata}
\alias{sagang_croptodata}
\title{QGIS algorithm - Crop to data}
\usage{
sagang_croptodata(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Grids. .}

\item{OUTPUT}{\code{rasterDestination} - Cropped Grids. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Crop to data (sagang:croptodata). ---------------- Arguments ----------------  INPUT: Grids 	Argument type:	multilayer OUTPUT: Cropped Grids 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Cropped Grids
}
}
}
