% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_crossclassificationandtabulation.R
\name{sagang_crossclassificationandtabulation}
\alias{sagang_crossclassificationandtabulation}
\title{QGIS algorithm - Cross-classification and tabulation}
\usage{
sagang_crossclassificationandtabulation(
  INPUT = qgisprocess:::qgis_default_value(),
  INPUT2 = qgisprocess:::qgis_default_value(),
  RESULTGRID = qgisprocess:::qgis_default_value(),
  RESULTTABLE = qgisprocess:::qgis_default_value(),
  MAXNUMCLASS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input Grid 1. Path to a raster layer.}

\item{INPUT2}{\code{raster} - Input Grid 2. Path to a raster layer.}

\item{RESULTGRID}{\code{rasterDestination} - Cross-Classification Grid. Path for new raster layer.}

\item{RESULTTABLE}{\code{vectorDestination} - Cross-Tabulation Table. Path for new vector layer.}

\item{MAXNUMCLASS}{\code{number} - Max. Number of Classes. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Cross-classification and tabulation (sagang:crossclassificationandtabulation). ---------------- Arguments ----------------  INPUT: Input Grid 1 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer INPUT2: Input Grid 2 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RESULTGRID: Cross-Classification Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RESULTTABLE: Cross-Tabulation Table 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer MAXNUMCLASS: Max. Number of Classes 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULTGRID - outputRaster - Cross-Classification Grid
\item RESULTTABLE - outputVector - Cross-Tabulation Table
}
}
}
