% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_crossprofiles.R
\name{sagang_crossprofiles}
\alias{sagang_crossprofiles}
\title{QGIS algorithm - Cross profiles}
\usage{
sagang_crossprofiles(
  DEM = qgisprocess:::qgis_default_value(),
  LINES = qgisprocess:::qgis_default_value(),
  PROFILES = qgisprocess:::qgis_default_value(),
  DIST_LINE = qgisprocess:::qgis_default_value(),
  DIST_PROFILE = qgisprocess:::qgis_default_value(),
  NUM_PROFILE = qgisprocess:::qgis_default_value(),
  INTERPOLATION = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{PROFILES}{\code{vectorDestination} - Cross Profiles. Path for new vector layer.}

\item{DIST_LINE}{\code{number} - Profile Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DIST_PROFILE}{\code{number} - Profile Length. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NUM_PROFILE}{\code{number} - Profile Samples. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INTERPOLATION}{\code{enum}  of \verb{("(0) Nearest Neighbour", "(1) Bilinear Interpolation", "(2) Bicubic Spline Interpolation", "(3) B-Spline Interpolation")} - Interpolation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{enum}  of \verb{("(0) vertices", "(1) attributes", "(2) vertices and attributes")} - Output. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Cross profiles (sagang:crossprofiles). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LINES: Lines 	Argument type:	source 	Acceptable values: 		- Path to a vector layer PROFILES: Cross Profiles 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DIST_LINE: Profile Distance 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DIST_PROFILE: Profile Length 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NUM_PROFILE: Profile Samples 	Default value:	11 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression INTERPOLATION: Interpolation 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) Nearest Neighbour 		- 1: (1) Bilinear Interpolation 		- 2: (2) Bicubic Spline Interpolation 		- 3: (3) B-Spline Interpolation 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' OUTPUT: Output 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) vertices 		- 1: (1) attributes 		- 2: (2) vertices and attributes 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item PROFILES - outputVector - Cross Profiles
}
}
}
