% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_curvatureclassification.R
\name{sagang_curvatureclassification}
\alias{sagang_curvatureclassification}
\title{QGIS algorithm - Curvature classification}
\usage{
sagang_curvatureclassification(
  DEM = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  STRAIGHT = qgisprocess:::qgis_default_value(),
  VERTICAL = qgisprocess:::qgis_default_value(),
  HORIZONTAL = qgisprocess:::qgis_default_value(),
  SMOOTH = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{CLASSES}{\code{rasterDestination} - Curvature Classification. Path for new raster layer.}

\item{STRAIGHT}{\code{number} - Threshold Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{VERTICAL}{\code{enum}  of \verb{("(0) longitudinal curvature", "(1) profile curvature")} - Vertical Curvature. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{HORIZONTAL}{\code{enum}  of \verb{("(0) cross-sectional curvature", "(1) tangential curvature")} - Horizontal Curvature. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SMOOTH}{\code{number} - Smoothing. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Curvature classification (sagang:curvatureclassification). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CLASSES: Curvature Classification 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STRAIGHT: Threshold Radius 	Default value:	2000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression VERTICAL: Vertical Curvature 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) longitudinal curvature 		- 1: (1) profile curvature 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' HORIZONTAL: Horizontal Curvature 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) cross-sectional curvature 		- 1: (1) tangential curvature 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SMOOTH: Smoothing 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Curvature Classification
}
}
}
