% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_definegeoreferenceforgrids.R
\name{sagang_definegeoreferenceforgrids}
\alias{sagang_definegeoreferenceforgrids}
\title{QGIS algorithm - Define georeference for grids}
\usage{
sagang_definegeoreferenceforgrids(
  GRIDS = qgisprocess:::qgis_default_value(),
  REFERENCED = qgisprocess:::qgis_default_value(),
  DEFINITION = qgisprocess:::qgis_default_value(),
  SIZE = qgisprocess:::qgis_default_value(),
  XMIN = qgisprocess:::qgis_default_value(),
  XMAX = qgisprocess:::qgis_default_value(),
  YMIN = qgisprocess:::qgis_default_value(),
  YMAX = qgisprocess:::qgis_default_value(),
  CELL_REF = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{REFERENCED}{\code{rasterDestination} - Referenced Grids. Path for new raster layer.}

\item{DEFINITION}{\code{enum}  of \verb{("(0) cellsize and lower left cell coordinates", "(1) cellsize and upper left cell coordinates", "(2) lower left cell coordinates and left to right range", "(3) lower left cell coordinates and lower to upper range")} - Definition. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SIZE}{\code{number} - Cellsize. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{XMIN}{\code{number} - Left. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{XMAX}{\code{number} - Right. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YMIN}{\code{number} - Lower. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YMAX}{\code{number} - Upper. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CELL_REF}{\code{enum}  of \verb{("(0) center", "(1) corner")} - Cell Reference. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Define georeference for grids (sagang:definegeoreferenceforgrids). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	multilayer REFERENCED: Referenced Grids 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DEFINITION: Definition 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) cellsize and lower left cell coordinates 		- 1: (1) cellsize and upper left cell coordinates 		- 2: (2) lower left cell coordinates and left to right range 		- 3: (3) lower left cell coordinates and lower to upper range 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SIZE: Cellsize 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression XMIN: Left 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression XMAX: Right 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YMIN: Lower 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YMAX: Upper 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CELL_REF: Cell Reference 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) center 		- 1: (1) corner 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item REFERENCED - outputRaster - Referenced Grids
}
}
}
