% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_destriping.R
\name{sagang_destriping}
\alias{sagang_destriping}
\title{QGIS algorithm - Destriping}
\usage{
sagang_destriping(
  INPUT = qgisprocess:::qgis_default_value(),
  MASK = qgisprocess:::qgis_default_value(),
  RESULT3 = qgisprocess:::qgis_default_value(),
  RESULT1 = qgisprocess:::qgis_default_value(),
  RESULT2 = qgisprocess:::qgis_default_value(),
  STRIPES = qgisprocess:::qgis_default_value(),
  ANG = qgisprocess:::qgis_default_value(),
  R = qgisprocess:::qgis_default_value(),
  D = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{MASK}{\code{raster} - Mask. Path to a raster layer.}

\item{RESULT3}{\code{rasterDestination} - Destriped Grid. Path for new raster layer.}

\item{RESULT1}{\code{rasterDestination} - Low-pass 1. Path for new raster layer.}

\item{RESULT2}{\code{rasterDestination} - Low-pass 2. Path for new raster layer.}

\item{STRIPES}{\code{rasterDestination} - Stripes. Path for new raster layer.}

\item{ANG}{\code{number} - Angle. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{R}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{D}{\code{number} - Stripes Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Destriping (sagang:destriping). ---------------- Arguments ----------------  INPUT: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MASK: Mask (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RESULT3: Destriped Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RESULT1: Low-pass 1 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RESULT2: Low-pass 2 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STRIPES: Stripes 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ANG: Angle 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression R: Radius 	Default value:	20 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression D: Stripes Distance 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT1 - outputRaster - Low-pass 1
\item RESULT2 - outputRaster - Low-pass 2
\item RESULT3 - outputRaster - Destriped Grid
\item STRIPES - outputRaster - Stripes
}
}
}
