% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_diffusepollutionrisk.R
\name{sagang_diffusepollutionrisk}
\alias{sagang_diffusepollutionrisk}
\title{QGIS algorithm - Diffuse pollution risk}
\usage{
sagang_diffusepollutionrisk(
  DEM = qgisprocess:::qgis_default_value(),
  CHANNEL = qgisprocess:::qgis_default_value(),
  WEIGHT = qgisprocess:::qgis_default_value(),
  WEIGHT_DEFAULT = qgisprocess:::qgis_default_value(),
  RAIN = qgisprocess:::qgis_default_value(),
  RAIN_DEFAULT = qgisprocess:::qgis_default_value(),
  DELIVERY = qgisprocess:::qgis_default_value(),
  RISK_POINT = qgisprocess:::qgis_default_value(),
  RISK_DIFFUSE = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  CHANNEL_START = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{CHANNEL}{\code{raster} - Channel Network. Path to a raster layer.}

\item{WEIGHT}{\code{raster} - Land Cover Weights. Path to a raster layer.}

\item{WEIGHT_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RAIN}{\code{raster} - Rainfall. Path to a raster layer.}

\item{RAIN_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELIVERY}{\code{rasterDestination} - Delivery Index. Path for new raster layer.}

\item{RISK_POINT}{\code{rasterDestination} - Locational Risk. Path for new raster layer.}

\item{RISK_DIFFUSE}{\code{rasterDestination} - Diffuse Pollution Risk. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) D8", "(1) MFD")} - Flow Routing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CHANNEL_START}{\code{number} - Channel Initiation Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Diffuse pollution risk (sagang:diffusepollutionrisk). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CHANNEL: Channel Network (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer WEIGHT: Land Cover Weights (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer WEIGHT_DEFAULT: Default 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RAIN: Rainfall (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RAIN_DEFAULT: Default 	Default value:	500 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DELIVERY: Delivery Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RISK_POINT: Locational Risk 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RISK_DIFFUSE: Diffuse Pollution Risk 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Flow Routing 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) D8 		- 1: (1) MFD 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CHANNEL_START: Channel Initiation Threshold 	Default value:	150 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item DELIVERY - outputRaster - Delivery Index
\item RISK_DIFFUSE - outputRaster - Diffuse Pollution Risk
\item RISK_POINT - outputRaster - Locational Risk
}
}
}
