% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_diffusivehillslopeevolutionftcs.R
\name{sagang_diffusivehillslopeevolutionftcs}
\alias{sagang_diffusivehillslopeevolutionftcs}
\title{QGIS algorithm - Diffusive hillslope evolution (ftcs)}
\usage{
sagang_diffusivehillslopeevolutionftcs(
  DEM = qgisprocess:::qgis_default_value(),
  MODEL = qgisprocess:::qgis_default_value(),
  DIFF = qgisprocess:::qgis_default_value(),
  UPDATE = qgisprocess:::qgis_default_value(),
  KAPPA = qgisprocess:::qgis_default_value(),
  DURATION = qgisprocess:::qgis_default_value(),
  TIMESTEP = qgisprocess:::qgis_default_value(),
  DTIME = qgisprocess:::qgis_default_value(),
  NEIGHBOURS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{MODEL}{\code{rasterDestination} - Modelled Elevation. Path for new raster layer.}

\item{DIFF}{\code{rasterDestination} - Elevation Difference. Path for new raster layer.}

\item{UPDATE}{\code{boolean} - Update. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KAPPA}{\code{number} - Diffusivity (m2 \code{/} kyr). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DURATION}{\code{number} - Simulation Time (kyr). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIMESTEP}{\code{enum}  of \verb{("(0) user defined", "(1) automatically")} - Time Step. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DTIME}{\code{number} - Time Step (kyr). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NEIGHBOURS}{\code{enum}  of \verb{("(0) Neumann", "(1) Moore")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Diffusive hillslope evolution (ftcs) (sagang:diffusivehillslopeevolutionftcs). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MODEL: Modelled Elevation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIFF: Elevation Difference 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer UPDATE: Update 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KAPPA: Diffusivity (m2 / kyr) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DURATION: Simulation Time (kyr) 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIMESTEP: Time Step 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) user defined 		- 1: (1) automatically 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DTIME: Time Step (kyr) 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NEIGHBOURS: Neighbourhood 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Neumann 		- 1: (1) Moore 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIFF - outputRaster - Elevation Difference
\item MODEL - outputRaster - Modelled Elevation
}
}
}
