% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sagang_directgeoreferencingofairbornephotographs.R
\name{sagang_directgeoreferencingofairbornephotographs}
\alias{sagang_directgeoreferencingofairbornephotographs}
\title{QGIS algorithm - Direct georeferencing of airborne photographs}
\usage{
sagang_directgeoreferencingofairbornephotographs(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  DEM = qgisprocess:::qgis_default_value(),
  DEM_DEFAULT = qgisprocess:::qgis_default_value(),
  CFL = qgisprocess:::qgis_default_value(),
  PXSIZE = qgisprocess:::qgis_default_value(),
  X = qgisprocess:::qgis_default_value(),
  Y = qgisprocess:::qgis_default_value(),
  Z = qgisprocess:::qgis_default_value(),
  OMEGA = qgisprocess:::qgis_default_value(),
  PHI = qgisprocess:::qgis_default_value(),
  KAPPA = qgisprocess:::qgis_default_value(),
  KAPPA_OFF = qgisprocess:::qgis_default_value(),
  ORIENTATION = qgisprocess:::qgis_default_value(),
  ROW_ORDER = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  DATA_TYPE = qgisprocess:::qgis_default_value(),
  TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX =
    qgisprocess:::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Unreferenced Grids. .}

\item{OUTPUT}{\code{rasterDestination} - Referenced Grids. Path for new raster layer.}

\item{EXTENT}{\code{vectorDestination} - Extent. Path for new vector layer.}

\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{DEM_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CFL}{\code{number} - Focal Length (mm). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PXSIZE}{\code{number} - CCD Physical Pixel Size (micron). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{X}{\code{number} - X. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{Y}{\code{number} - Y. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{Z}{\code{number} - Z. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OMEGA}{\code{number} - Omega. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PHI}{\code{number} - Phi. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KAPPA}{\code{number} - Kappa. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KAPPA_OFF}{\code{number} - Offset. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ORIENTATION}{\code{enum}  of \verb{("(0) BLUH", "(1) PATB")} - Orientation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ROW_ORDER}{\code{enum}  of \verb{("(0) top down", "(1) bottom up")} - Row Order. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESAMPLING}{\code{enum}  of \verb{("(0) Nearest Neighbour", "(1) Bilinear Interpolation", "(2) Bicubic Spline Interpolation", "(3) B-Spline Interpolation")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DATA_TYPE}{\code{enum}  of \verb{("(0) 1 byte unsigned integer", "(1) 1 byte signed integer", "(2) 2 byte unsigned integer", "(3) 2 byte signed integer", "(4) 4 byte unsigned integer", "(5) 4 byte signed integer", "(6) 4 byte floating point", "(7) 8 byte floating point", "(8) same as original")} - Data Storage Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Direct georeferencing of airborne photographs (sagang:directgeoreferencingofairbornephotographs). ---------------- Arguments ----------------  INPUT: Unreferenced Grids 	Argument type:	multilayer OUTPUT: Referenced Grids 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer EXTENT: Extent 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DEM: Elevation (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DEM_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CFL: Focal Length (mm) 	Default value:	80 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PXSIZE: CCD Physical Pixel Size (micron) 	Default value:	5.2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression X: X 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Y: Y 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Z: Z 	Default value:	1000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OMEGA: Omega 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PHI: Phi 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KAPPA: Kappa 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KAPPA_OFF: Offset 	Default value:	90 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ORIENTATION: Orientation 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) BLUH 		- 1: (1) PATB 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' ROW_ORDER: Row Order 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) top down 		- 1: (1) bottom up 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RESAMPLING: Resampling 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) Nearest Neighbour 		- 1: (1) Bilinear Interpolation 		- 2: (2) Bicubic Spline Interpolation 		- 3: (3) B-Spline Interpolation 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DATA_TYPE: Data Storage Type 	Default value:	8 	Argument type:	enum 	Available values: 		- 0: (0) 1 byte unsigned integer 		- 1: (1) 1 byte signed integer 		- 2: (2) 2 byte unsigned integer 		- 3: (3) 2 byte signed integer 		- 4: (4) 4 byte unsigned integer 		- 5: (5) 4 byte signed integer 		- 6: (6) 4 byte floating point 		- 7: (7) 8 byte floating point 		- 8: (8) same as original 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX: Output extent (optional) 	Argument type:	extent 	Acceptable values: 		- A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105' 		- Path to a layer. The extent of the layer is used. TARGET_USER_SIZE: Cellsize 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item EXTENT - outputVector - Extent
\item OUTPUT - outputRaster - Referenced Grids
}
}
}
