% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_directionalstatisticsforrasterlayer.R
\name{sagang_directionalstatisticsforrasterlayer}
\alias{sagang_directionalstatisticsforrasterlayer}
\title{QGIS algorithm - Directional statistics for raster layer}
\usage{
sagang_directionalstatisticsforrasterlayer(
  GRID = qgisprocess:::qgis_default_value(),
  MEAN = qgisprocess:::qgis_default_value(),
  DIFMEAN = qgisprocess:::qgis_default_value(),
  MIN = qgisprocess:::qgis_default_value(),
  MAX = qgisprocess:::qgis_default_value(),
  RANGE = qgisprocess:::qgis_default_value(),
  VAR = qgisprocess:::qgis_default_value(),
  STDDEV = qgisprocess:::qgis_default_value(),
  STDDEVLO = qgisprocess:::qgis_default_value(),
  STDDEVHI = qgisprocess:::qgis_default_value(),
  DEVMEAN = qgisprocess:::qgis_default_value(),
  PERCENT = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  POINTS_OUT = qgisprocess:::qgis_default_value(),
  DIRECTION = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  MAXDISTANCE = qgisprocess:::qgis_default_value(),
  DW_WEIGHTING = qgisprocess:::qgis_default_value(),
  DW_IDW_POWER = qgisprocess:::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{MEAN}{\code{rasterDestination} - Arithmetic Mean. Path for new raster layer.}

\item{DIFMEAN}{\code{rasterDestination} - Difference from Arithmetic Mean. Path for new raster layer.}

\item{MIN}{\code{rasterDestination} - Minimum. Path for new raster layer.}

\item{MAX}{\code{rasterDestination} - Maximum. Path for new raster layer.}

\item{RANGE}{\code{rasterDestination} - Range. Path for new raster layer.}

\item{VAR}{\code{rasterDestination} - Variance. Path for new raster layer.}

\item{STDDEV}{\code{rasterDestination} - Standard Deviation. Path for new raster layer.}

\item{STDDEVLO}{\code{rasterDestination} - Mean less Standard Deviation. Path for new raster layer.}

\item{STDDEVHI}{\code{rasterDestination} - Mean plus Standard Deviation. Path for new raster layer.}

\item{DEVMEAN}{\code{rasterDestination} - Deviation from Arithmetic Mean. Path for new raster layer.}

\item{PERCENT}{\code{rasterDestination} - Percentile. Path for new raster layer.}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{POINTS_OUT}{\code{vectorDestination} - Directional Statistics for Points. Path for new vector layer.}

\item{DIRECTION}{\code{number} - Direction (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TOLERANCE}{\code{number} - Tolerance (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAXDISTANCE}{\code{number} - Maximum Distance (Cells). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("(0) no distance weighting", "(1) inverse distance to a power", "(2) exponential", "(3) gaussian")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Power. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DW_BANDWIDTH}{\code{number} - Bandwidth. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Directional statistics for raster layer (sagang:directionalstatisticsforrasterlayer). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MEAN: Arithmetic Mean 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIFMEAN: Difference from Arithmetic Mean 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MIN: Minimum 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MAX: Maximum 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RANGE: Range 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VAR: Variance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STDDEV: Standard Deviation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STDDEVLO: Mean less Standard Deviation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STDDEVHI: Mean plus Standard Deviation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DEVMEAN: Deviation from Arithmetic Mean 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PERCENT: Percentile 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer POINTS: Points (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer POINTS_OUT: Directional Statistics for Points 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DIRECTION: Direction (Degree) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TOLERANCE: Tolerance (Degree) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAXDISTANCE: Maximum Distance (Cells) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DW_WEIGHTING: Weighting Function 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) no distance weighting 		- 1: (1) inverse distance to a power 		- 2: (2) exponential 		- 3: (3) gaussian 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DW_IDW_POWER: Power 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DW_BANDWIDTH: Bandwidth 	Default value:	25 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item DEVMEAN - outputRaster - Deviation from Arithmetic Mean
\item DIFMEAN - outputRaster - Difference from Arithmetic Mean
\item MAX - outputRaster - Maximum
\item MEAN - outputRaster - Arithmetic Mean
\item MIN - outputRaster - Minimum
\item PERCENT - outputRaster - Percentile
\item POINTS_OUT - outputVector - Directional Statistics for Points
\item RANGE - outputRaster - Range
\item STDDEV - outputRaster - Standard Deviation
\item STDDEVHI - outputRaster - Mean plus Standard Deviation
\item STDDEVLO - outputRaster - Mean less Standard Deviation
\item VAR - outputRaster - Variance
}
}
}
