% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_diversityofcategories.R
\name{sagang_diversityofcategories}
\alias{sagang_diversityofcategories}
\title{QGIS algorithm - Diversity of categories}
\usage{
sagang_diversityofcategories(
  CATEGORIES = qgisprocess:::qgis_default_value(),
  COUNT = qgisprocess:::qgis_default_value(),
  DIVERSITY = qgisprocess:::qgis_default_value(),
  CONNECTIVITY = qgisprocess:::qgis_default_value(),
  CONNECTEDAVG = qgisprocess:::qgis_default_value(),
  NB_CASE = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  KERNEL_TYPE = qgisprocess:::qgis_default_value(),
  KERNEL_RADIUS = qgisprocess:::qgis_default_value(),
  DW_WEIGHTING = qgisprocess:::qgis_default_value(),
  DW_IDW_POWER = qgisprocess:::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CATEGORIES}{\code{raster} - Categories. Path to a raster layer.}

\item{COUNT}{\code{rasterDestination} - Number of Categories. Path for new raster layer.}

\item{DIVERSITY}{\code{rasterDestination} - Diversity. Path for new raster layer.}

\item{CONNECTIVITY}{\code{rasterDestination} - Connectivity. Path for new raster layer.}

\item{CONNECTEDAVG}{\code{rasterDestination} - Averaged Connectivity. Path for new raster layer.}

\item{NB_CASE}{\code{enum}  of \verb{("(0) Rook's case", "(1) Queen's case")} - Connectivity Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NORMALIZE}{\code{enum}  of \verb{("(0) no", "(1) by number of cells", "(2) by area size")} - Normalize. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_TYPE}{\code{enum}  of \verb{("(0) Square", "(1) Circle")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("(0) no distance weighting", "(1) inverse distance to a power", "(2) exponential", "(3) gaussian")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Power. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DW_BANDWIDTH}{\code{number} - Bandwidth. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Diversity of categories (sagang:diversityofcategories). ---------------- Arguments ----------------  CATEGORIES: Categories 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer COUNT: Number of Categories 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIVERSITY: Diversity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONNECTIVITY: Connectivity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONNECTEDAVG: Averaged Connectivity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer NB_CASE: Connectivity Neighbourhood 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Rook's case 		- 1: (1) Queen's case 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NORMALIZE: Normalize 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) no 		- 1: (1) by number of cells 		- 2: (2) by area size 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' KERNEL_TYPE: Kernel Type 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Square 		- 1: (1) Circle 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' KERNEL_RADIUS: Radius 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DW_WEIGHTING: Weighting Function 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) no distance weighting 		- 1: (1) inverse distance to a power 		- 2: (2) exponential 		- 3: (3) gaussian 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DW_IDW_POWER: Power 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DW_BANDWIDTH: Bandwidth 	Default value:	0.7 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONNECTEDAVG - outputRaster - Averaged Connectivity
\item CONNECTIVITY - outputRaster - Connectivity
\item COUNT - outputRaster - Number of Categories
\item DIVERSITY - outputRaster - Diversity
}
}
}
