% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_downslopedistancegradient.R
\name{sagang_downslopedistancegradient}
\alias{sagang_downslopedistancegradient}
\title{QGIS algorithm - Downslope distance gradient}
\usage{
sagang_downslopedistancegradient(
  DEM = qgisprocess:::qgis_default_value(),
  GRADIENT = qgisprocess:::qgis_default_value(),
  DIFFERENCE = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{GRADIENT}{\code{rasterDestination} - Gradient. Path for new raster layer.}

\item{DIFFERENCE}{\code{rasterDestination} - Gradient Difference. Path for new raster layer.}

\item{DISTANCE}{\code{number} - Vertical Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{enum}  of \verb{("(0) distance", "(1) gradient (tangens)", "(2) gradient (degree)")} - Output. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Downslope distance gradient (sagang:downslopedistancegradient). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GRADIENT: Gradient 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIFFERENCE: Gradient Difference 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DISTANCE: Vertical Distance 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OUTPUT: Output 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) distance 		- 1: (1) gradient (tangens) 		- 2: (2) gradient (degree) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIFFERENCE - outputRaster - Gradient Difference
\item GRADIENT - outputRaster - Gradient
}
}
}
