% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_dtmfilterslopebased.R
\name{sagang_dtmfilterslopebased}
\alias{sagang_dtmfilterslopebased}
\title{QGIS algorithm - Dtm filter (slope-based)}
\usage{
sagang_dtmfilterslopebased(
  INPUT = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  TERRAINSLOPE = qgisprocess:::qgis_default_value(),
  STDDEV = qgisprocess:::qgis_default_value(),
  GROUND = qgisprocess:::qgis_default_value(),
  NONGROUND = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid to filter. Path to a raster layer.}

\item{RADIUS}{\code{number} - Search Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TERRAINSLOPE}{\code{number} - Approx. Terrain Slope. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{STDDEV}{\code{boolean} - Use Confidence Interval. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GROUND}{\code{rasterDestination} - Bare Earth. Path for new raster layer.}

\item{NONGROUND}{\code{rasterDestination} - Removed Objects. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Dtm filter (slope-based) (sagang:dtmfilterslopebased). ---------------- Arguments ----------------  INPUT: Grid to filter 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RADIUS: Search Radius 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TERRAINSLOPE: Approx. Terrain Slope 	Default value:	30 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression STDDEV: Use Confidence Interval 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GROUND: Bare Earth 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer NONGROUND: Removed Objects 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item GROUND - outputRaster - Bare Earth
\item NONGROUND - outputRaster - Removed Objects
}
}
}
