% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_edgecontamination.R
\name{sagang_edgecontamination}
\alias{sagang_edgecontamination}
\title{QGIS algorithm - Edge contamination}
\usage{
sagang_edgecontamination(
  ELEVATION = qgisprocess:::qgis_default_value(),
  CONTAMINATION = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{CONTAMINATION}{\code{rasterDestination} - Edge Contamination. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) single flow direction", "(1) multiple flow direction")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Edge contamination (sagang:edgecontamination). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CONTAMINATION: Edge Contamination 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) single flow direction 		- 1: (1) multiple flow direction 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONTAMINATION - outputRaster - Edge Contamination
}
}
}
