% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_evapotranspirationtable.R
\name{sagang_evapotranspirationtable}
\alias{sagang_evapotranspirationtable}
\title{QGIS algorithm - Evapotranspiration (table)}
\usage{
sagang_evapotranspirationtable(
  TABLE = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  T = qgisprocess:::qgis_default_value(),
  T_MIN = qgisprocess:::qgis_default_value(),
  T_MAX = qgisprocess:::qgis_default_value(),
  RH = qgisprocess:::qgis_default_value(),
  SR = qgisprocess:::qgis_default_value(),
  WS = qgisprocess:::qgis_default_value(),
  P = qgisprocess:::qgis_default_value(),
  DATE = qgisprocess:::qgis_default_value(),
  ET = qgisprocess:::qgis_default_value(),
  LAT = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TABLE}{\code{source} - Data. Path to a vector layer.}

\item{RESULT}{\code{vectorDestination} - Evapotranspiration. Path for new vector layer.}

\item{T}{\code{field} - Mean Temperature. The name of an existing field. ; delimited list of existing field names.}

\item{T_MIN}{\code{field} - Minimum Temperature. The name of an existing field. ; delimited list of existing field names.}

\item{T_MAX}{\code{field} - Maximum Temperature. The name of an existing field. ; delimited list of existing field names.}

\item{RH}{\code{field} - Relative Humidity. The name of an existing field. ; delimited list of existing field names.}

\item{SR}{\code{field} - Solar Radiation. The name of an existing field. ; delimited list of existing field names.}

\item{WS}{\code{field} - Wind Speed. The name of an existing field. ; delimited list of existing field names.}

\item{P}{\code{field} - Air Pressure. The name of an existing field. ; delimited list of existing field names.}

\item{DATE}{\code{field} - Date. The name of an existing field. ; delimited list of existing field names.}

\item{ET}{\code{field} - Evapotranspiration. The name of an existing field. ; delimited list of existing field names.}

\item{LAT}{\code{number} - Latitude. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) Turc", "(1) Hargreave", "(2) Penman (simplified)", "(3) Penman-Monteith")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Evapotranspiration (table) (sagang:evapotranspirationtable). ---------------- Arguments ----------------  TABLE: Data 	Argument type:	source 	Acceptable values: 		- Path to a vector layer RESULT: Evapotranspiration 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer T: Mean Temperature 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names T_MIN: Minimum Temperature 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names T_MAX: Maximum Temperature 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names RH: Relative Humidity 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names SR: Solar Radiation 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names WS: Wind Speed 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names P: Air Pressure 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names DATE: Date 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names ET: Evapotranspiration 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names LAT: Latitude 	Default value:	53 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Turc 		- 1: (1) Hargreave 		- 2: (2) Penman (simplified) 		- 3: (3) Penman-Monteith 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Evapotranspiration
}
}
}
