% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_exportesriarcinfogrid.R
\name{sagang_exportesriarcinfogrid}
\alias{sagang_exportesriarcinfogrid}
\title{QGIS algorithm - Export esri arc/info grid}
\usage{
sagang_exportesriarcinfogrid(
  GRID = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  FORMAT = qgisprocess:::qgis_default_value(),
  GEOREF = qgisprocess:::qgis_default_value(),
  PREC = qgisprocess:::qgis_default_value(),
  DECSEP = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{FILE}{\code{file} - File. Path to a file.}

\item{FORMAT}{\code{enum}  of \verb{("(0) binary", "(1) ASCII")} - Format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GEOREF}{\code{enum}  of \verb{("(0) corner", "(1) center")} - Geo-Reference. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PREC}{\code{number} - ASCII Precision. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DECSEP}{\code{enum}  of \verb{("(0) point (.)", "(1) comma (,)")} - ASCII Decimal Separator. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Export esri arc/info grid (sagang:exportesriarcinfogrid). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FILE: File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file FORMAT: Format 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) binary 		- 1: (1) ASCII 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' GEOREF: Geo-Reference 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) corner 		- 1: (1) center 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' PREC: ASCII Precision 	Default value:	4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DECSEP: ASCII Decimal Separator 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) point (.) 		- 1: (1) comma (,) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
