% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_exportpolygonstohtmlimagemap.R
\name{sagang_exportpolygonstohtmlimagemap}
\alias{sagang_exportpolygonstohtmlimagemap}
\title{QGIS algorithm - Export polygons to html image map}
\usage{
sagang_exportpolygonstohtmlimagemap(
  POLYGONS = qgisprocess:::qgis_default_value(),
  LINK = qgisprocess:::qgis_default_value(),
  TITLE = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  LINK_PREFIX = qgisprocess:::qgis_default_value(),
  LINK_SUFFIX = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{LINK}{\code{field} - Link. The name of an existing field. ; delimited list of existing field names.}

\item{TITLE}{\code{field} - Title. The name of an existing field. ; delimited list of existing field names.}

\item{FILE}{\code{file} - File. Path to a file.}

\item{LINK_PREFIX}{\code{string} - Link Prefix. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LINK_SUFFIX}{\code{string} - Link Suffix. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Export polygons to html image map (sagang:exportpolygonstohtmlimagemap). ---------------- Arguments ----------------  POLYGONS: Polygons 	Argument type:	source 	Acceptable values: 		- Path to a vector layer LINK: Link 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TITLE: Title 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FILE: File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file LINK_PREFIX: Link Prefix 	Default value:	http://www.saga-gis.org/ 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LINK_SUFFIX: Link Suffix 	Default value:	index.html 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
