% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_exportshapes.R
\name{sagang_exportshapes}
\alias{sagang_exportshapes}
\title{QGIS algorithm - Export shapes}
\usage{
sagang_exportshapes(
  SHAPES = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  FORMAT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{FILE}{\code{file} - File. Path to a file.}

\item{FORMAT}{\code{enum}  of \verb{("(0) PCIDSK Database File", "(1) Network Common Data Format", "(2) NASA Planetary Data System 4", "(3) MIPL VICAR file", "(4) Geospatial PDF", "(5) MBTiles", "(6) IBM DB2 Spatial Database", "(7) ESRI Shapefile", "(8) MapInfo File", "(9) IHO S-57 (ENC)", "(10) Microstation DGN", "(11) Memory", "(12) Atlas BNA", "(13) Comma Separated Value (.csv)", "(14) Geography Markup Language (GML)", "(15) GPX", "(16) Keyhole Markup Language (LIBKML)", "(17) Keyhole Markup Language (KML)", "(18) GeoJSON", "(19) GeoJSON Sequence", "(20) Interlis 1", "(21) Interlis 2", "(22) GMT ASCII Vectors (.gmt)", "(23) GeoPackage", "(24) SQLite / Spatialite", "(25) ODBC", "(26) WAsP .map format", "(27) Microsoft SQL Server Spatial Database (BCP)", "(28) PostgreSQL/PostGIS", "(29) MySQL", "(30) AutoCAD DXF", "(31) FlatGeobuf", "(32) Geoconcept", "(33) GeoRSS", "(34) GPSTrackMaker", "(35) PostgreSQL SQL dump", "(36) GPSBabel", "(37) CouchDB / GeoCouch", "(38) Cloudant / CouchDB", "(39) Open Document/ LibreOffice / OpenOffice Spreadsheet ", "(40) MS Office Open XML spreadsheet", "(41) Elastic Search", "(42) Carto", "(43) AmigoCloud", "(44) Selafin", "(45) OpenJUMP JML", "(46) VDV-451/VDV-452/INTREST Data Format", "(47) Mapbox Vector Tiles", "(48) NextGIS Web", "(49) MapML", "(50) U.S. Census TIGER/Line")} - Format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Export shapes (sagang:exportshapes). ---------------- Arguments ----------------  SHAPES: Shapes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FILE: File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file FORMAT: Format 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) PCIDSK Database File 		- 1: (1) Network Common Data Format 		- 2: (2) NASA Planetary Data System 4 		- 3: (3) MIPL VICAR file 		- 4: (4) Geospatial PDF 		- 5: (5) MBTiles 		- 6: (6) IBM DB2 Spatial Database 		- 7: (7) ESRI Shapefile 		- 8: (8) MapInfo File 		- 9: (9) IHO S-57 (ENC) 		- 10: (10) Microstation DGN 		- 11: (11) Memory 		- 12: (12) Atlas BNA 		- 13: (13) Comma Separated Value (.csv) 		- 14: (14) Geography Markup Language (GML) 		- 15: (15) GPX 		- 16: (16) Keyhole Markup Language (LIBKML) 		- 17: (17) Keyhole Markup Language (KML) 		- 18: (18) GeoJSON 		- 19: (19) GeoJSON Sequence 		- 20: (20) Interlis 1 		- 21: (21) Interlis 2 		- 22: (22) GMT ASCII Vectors (.gmt) 		- 23: (23) GeoPackage 		- 24: (24) SQLite / Spatialite 		- 25: (25) ODBC 		- 26: (26) WAsP .map format 		- 27: (27) Microsoft SQL Server Spatial Database (BCP) 		- 28: (28) PostgreSQL/PostGIS 		- 29: (29) MySQL 		- 30: (30) AutoCAD DXF 		- 31: (31) FlatGeobuf 		- 32: (32) Geoconcept 		- 33: (33) GeoRSS 		- 34: (34) GPSTrackMaker 		- 35: (35) PostgreSQL SQL dump 		- 36: (36) GPSBabel 		- 37: (37) CouchDB / GeoCouch 		- 38: (38) Cloudant / CouchDB 		- 39: (39) Open Document/ LibreOffice / OpenOffice Spreadsheet  		- 40: (40) MS Office Open XML spreadsheet 		- 41: (41) Elastic Search 		- 42: (42) Carto 		- 43: (43) AmigoCloud 		- 44: (44) Selafin 		- 45: (45) OpenJUMP JML 		- 46: (46) VDV-451/VDV-452/INTREST Data Format 		- 47: (47) Mapbox Vector Tiles 		- 48: (48) NextGIS Web 		- 49: (49) MapML 		- 50: (50) U.S. Census TIGER/Line 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
