% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_exportshapestoxyz.R
\name{sagang_exportshapestoxyz}
\alias{sagang_exportshapestoxyz}
\title{QGIS algorithm - Export shapes to xyz}
\usage{
sagang_exportshapestoxyz(
  POINTS = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  HEADER = qgisprocess:::qgis_default_value(),
  SEPARATE = qgisprocess:::qgis_default_value(),
  FILENAME = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Shapes. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{HEADER}{\code{boolean} - Save Table Header. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SEPARATE}{\code{enum}  of \verb{("(0) none", "(1) *", "(2) number of points")} - Separate Line\code{/}Polygon Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FILENAME}{\code{file} - File. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Export shapes to xyz (sagang:exportshapestoxyz). ---------------- Arguments ----------------  POINTS: Shapes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Attribute 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names HEADER: Save Table Header 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEPARATE: Separate Line/Polygon Points 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) none 		- 1: (1) * 		- 2: (2) number of points 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FILENAME: File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
