% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_exportsurfergrid.R
\name{sagang_exportsurfergrid}
\alias{sagang_exportsurfergrid}
\title{QGIS algorithm - Export surfer grid}
\usage{
sagang_exportsurfergrid(
  GRID = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  FORMAT = qgisprocess:::qgis_default_value(),
  NODATA = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{FILE}{\code{file} - File. Path to a file.}

\item{FORMAT}{\code{enum}  of \verb{("(0) binary", "(1) ASCII")} - Format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NODATA}{\code{boolean} - Use Surfer's No-Data Value. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Export surfer grid (sagang:exportsurfergrid). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FILE: File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file FORMAT: Format 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) binary 		- 1: (1) ASCII 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NODATA: Use Surfer's No-Data Value 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
