% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_exportwaspterrainmapfile.R
\name{sagang_exportwaspterrainmapfile}
\alias{sagang_exportwaspterrainmapfile}
\title{QGIS algorithm - Export wasp terrain map file}
\usage{
sagang_exportwaspterrainmapfile(
  SHAPES = qgisprocess:::qgis_default_value(),
  ELEVATION = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Contour Lines. Path to a vector layer.}

\item{ELEVATION}{\code{field} - Map File. The name of an existing field. ; delimited list of existing field names.}

\item{FILE}{\code{file} - File Name. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Export wasp terrain map file (sagang:exportwaspterrainmapfile). ---------------- Arguments ----------------  SHAPES: Contour Lines 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ELEVATION: Map File 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FILE: File Name 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
