% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_exportwrfgeogridbinaryformat.R
\name{sagang_exportwrfgeogridbinaryformat}
\alias{sagang_exportwrfgeogridbinaryformat}
\title{QGIS algorithm - Export wrf geogrid binary format}
\usage{
sagang_exportwrfgeogridbinaryformat(
  GRIDS = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  DATATYPE = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  NAME_DIGITS = qgisprocess:::qgis_default_value(),
  MISSING = qgisprocess:::qgis_default_value(),
  SCALE = qgisprocess:::qgis_default_value(),
  UNITS = qgisprocess:::qgis_default_value(),
  DESCRIPTION = qgisprocess:::qgis_default_value(),
  MMINLU = qgisprocess:::qgis_default_value(),
  TILE_BDR = qgisprocess:::qgis_default_value(),
  PROJECTION = qgisprocess:::qgis_default_value(),
  SDTLON = qgisprocess:::qgis_default_value(),
  TRUELAT1 = qgisprocess:::qgis_default_value(),
  TRUELAT2 = qgisprocess:::qgis_default_value(),
  ISWATER = qgisprocess:::qgis_default_value(),
  ISLAKE = qgisprocess:::qgis_default_value(),
  ISICE = qgisprocess:::qgis_default_value(),
  ISURBAN = qgisprocess:::qgis_default_value(),
  ISOILWATER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{FILE}{\code{file} - Directory. Path to a file.}

\item{DATATYPE}{\code{enum}  of \verb{("(0) 1 byte unsigned", "(1) 1 byte signed", "(2) 2 byte unsigned", "(3) 2 byte signed", "(4) 4 byte unsigned", "(5) 4 byte signed")} - Data Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TYPE}{\code{enum}  of \verb{("(0) categorical", "(1) continuous")} - Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NAME_DIGITS}{\code{number} - Filename Digits. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MISSING}{\code{number} - Missing Value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SCALE}{\code{number} - Scale Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{UNITS}{\code{string} - Units. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DESCRIPTION}{\code{string} - Description. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MMINLU}{\code{string} - Look Up Section. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TILE_BDR}{\code{number} - Halo Width. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PROJECTION}{\code{enum}  of \verb{("(0) lambert", "(1) polar", "(2) mercator", "(3) regular_ll", "(4) albers_nad83", "(5) polar_wgs84")} - Projection. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SDTLON}{\code{number} - Standard Longitude. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TRUELAT1}{\code{number} - True Latitude 1. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TRUELAT2}{\code{number} - True Latitude 2. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ISWATER}{\code{number} - Water. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ISLAKE}{\code{number} - Lake. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ISICE}{\code{number} - Ice. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ISURBAN}{\code{number} - Urban. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ISOILWATER}{\code{number} - Soil Water. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Export wrf geogrid binary format (sagang:exportwrfgeogridbinaryformat). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	multilayer FILE: Directory 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file DATATYPE: Data Type 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) 1 byte unsigned 		- 1: (1) 1 byte signed 		- 2: (2) 2 byte unsigned 		- 3: (3) 2 byte signed 		- 4: (4) 4 byte unsigned 		- 5: (5) 4 byte signed 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' TYPE: Type 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) categorical 		- 1: (1) continuous 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NAME_DIGITS: Filename Digits 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MISSING: Missing Value 	Default value:	-99999 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SCALE: Scale Factor 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression UNITS: Units 	Default value:	 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DESCRIPTION: Description 	Default value:	 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MMINLU: Look Up Section 	Default value:	USGS 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TILE_BDR: Halo Width 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PROJECTION: Projection 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) lambert 		- 1: (1) polar 		- 2: (2) mercator 		- 3: (3) regular_ll 		- 4: (4) albers_nad83 		- 5: (5) polar_wgs84 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SDTLON: Standard Longitude 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TRUELAT1: True Latitude 1 	Default value:	45 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TRUELAT2: True Latitude 2 	Default value:	35 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ISWATER: Water 	Default value:	16 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ISLAKE: Lake 	Default value:	-1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ISICE: Ice 	Default value:	24 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ISURBAN: Urban 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ISOILWATER: Soil Water 	Default value:	14 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
