% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_fastrepresentativeness.R
\name{sagang_fastrepresentativeness}
\alias{sagang_fastrepresentativeness}
\title{QGIS algorithm - Fast representativeness}
\usage{
sagang_fastrepresentativeness(
  INPUT = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  RESULT_LOD = qgisprocess:::qgis_default_value(),
  SEEDS = qgisprocess:::qgis_default_value(),
  LOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{RESULT_LOD}{\code{rasterDestination} - Output Lod. Path for new raster layer.}

\item{SEEDS}{\code{rasterDestination} - Output Seeds. Path for new raster layer.}

\item{LOD}{\code{number} - Level of Generalisation. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Fast representativeness (sagang:fastrepresentativeness). ---------------- Arguments ----------------  INPUT: Input 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RESULT: Output 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RESULT_LOD: Output Lod 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SEEDS: Output Seeds 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LOD: Level of Generalisation 	Default value:	16 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Output
\item RESULT_LOD - outputRaster - Output Lod
\item SEEDS - outputRaster - Output Seeds
}
}
}
