% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_fillsinksqmofesp.R
\name{sagang_fillsinksqmofesp}
\alias{sagang_fillsinksqmofesp}
\title{QGIS algorithm - Fill sinks (qm of esp)}
\usage{
sagang_fillsinksqmofesp(
  DEM = qgisprocess:::qgis_default_value(),
  FILLED = qgisprocess:::qgis_default_value(),
  SINKS = qgisprocess:::qgis_default_value(),
  DZFILL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{FILLED}{\code{rasterDestination} - DEM without Sinks. Path for new raster layer.}

\item{SINKS}{\code{rasterDestination} - Sinks. Path for new raster layer.}

\item{DZFILL}{\code{number} - Fill Increment. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Fill sinks (qm of esp) (sagang:fillsinksqmofesp). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FILLED: DEM without Sinks 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SINKS: Sinks 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DZFILL: Fill Increment 	Default value:	0.01 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FILLED - outputRaster - DEM without Sinks
\item SINKS - outputRaster - Sinks
}
}
}
