% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_fillsinksxxlwangliu.R
\name{sagang_fillsinksxxlwangliu}
\alias{sagang_fillsinksxxlwangliu}
\title{QGIS algorithm - Fill sinks xxl (wang & liu)}
\usage{
sagang_fillsinksxxlwangliu(
  ELEV = qgisprocess:::qgis_default_value(),
  FILLED = qgisprocess:::qgis_default_value(),
  MINSLOPE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEV}{\code{raster} - DEM. Path to a raster layer.}

\item{FILLED}{\code{rasterDestination} - Filled DEM. Path for new raster layer.}

\item{MINSLOPE}{\code{number} - Minimum Slope (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Fill sinks xxl (wang & liu) (sagang:fillsinksxxlwangliu). ---------------- Arguments ----------------  ELEV: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FILLED: Filled DEM 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MINSLOPE: Minimum Slope (Degree) 	Default value:	0.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FILLED - outputRaster - Filled DEM
}
}
}
