% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_flatdetection.R
\name{sagang_flatdetection}
\alias{sagang_flatdetection}
\title{QGIS algorithm - Flat detection}
\usage{
sagang_flatdetection(
  DEM = qgisprocess:::qgis_default_value(),
  NOFLATS = qgisprocess:::qgis_default_value(),
  FLATS = qgisprocess:::qgis_default_value(),
  FLAT_OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{NOFLATS}{\code{rasterDestination} - No Flats. Path for new raster layer.}

\item{FLATS}{\code{rasterDestination} - Flat Areas. Path for new raster layer.}

\item{FLAT_OUTPUT}{\code{enum}  of \verb{("(0) elevation", "(1) enumeration")} - Flat Area Values. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Flat detection (sagang:flatdetection). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer NOFLATS: No Flats 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FLATS: Flat Areas 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FLAT_OUTPUT: Flat Area Values 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) elevation 		- 1: (1) enumeration 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLATS - outputRaster - Flat Areas
\item NOFLATS - outputRaster - No Flats
}
}
}
