% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_flowaccumulationonestep.R
\name{sagang_flowaccumulationonestep}
\alias{sagang_flowaccumulationonestep}
\title{QGIS algorithm - Flow accumulation (one step)}
\usage{
sagang_flowaccumulationonestep(
  DEM = qgisprocess:::qgis_default_value(),
  TCA = qgisprocess:::qgis_default_value(),
  SCA = qgisprocess:::qgis_default_value(),
  PREPROCESSING = qgisprocess:::qgis_default_value(),
  FLOW_ROUTING = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{TCA}{\code{rasterDestination} - Flow Accumulation. Path for new raster layer.}

\item{SCA}{\code{rasterDestination} - Specific Catchment Area. Path for new raster layer.}

\item{PREPROCESSING}{\code{enum}  of \verb{("(0) Fill Sinks (Wang & Liu)", "(1) Sink Removal")} - Preprocessing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FLOW_ROUTING}{\code{enum}  of \verb{("(0) Deterministic 8", "(1) Rho 8", "(2) Braunschweiger Reliefmodell", "(3) Deterministic Infinity", "(4) Multiple Flow Direction", "(5) Multiple Triangular Flow Direction", "(6) Multiple Maximum Downslope Gradient Based Flow Direction")} - Flow Routing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Flow accumulation (one step) (sagang:flowaccumulationonestep). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TCA: Flow Accumulation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SCA: Specific Catchment Area 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PREPROCESSING: Preprocessing 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Fill Sinks (Wang & Liu) 		- 1: (1) Sink Removal 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FLOW_ROUTING: Flow Routing 	Default value:	4 	Argument type:	enum 	Available values: 		- 0: (0) Deterministic 8 		- 1: (1) Rho 8 		- 2: (2) Braunschweiger Reliefmodell 		- 3: (3) Deterministic Infinity 		- 4: (4) Multiple Flow Direction 		- 5: (5) Multiple Triangular Flow Direction 		- 6: (6) Multiple Maximum Downslope Gradient Based Flow Direction 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item SCA - outputRaster - Specific Catchment Area
\item TCA - outputRaster - Flow Accumulation
}
}
}
