% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_flowaccumulationqmofesp.R
\name{sagang_flowaccumulationqmofesp}
\alias{sagang_flowaccumulationqmofesp}
\title{QGIS algorithm - Flow accumulation (qm of esp)}
\usage{
sagang_flowaccumulationqmofesp(
  DEM = qgisprocess:::qgis_default_value(),
  FLOW = qgisprocess:::qgis_default_value(),
  PREPROC = qgisprocess:::qgis_default_value(),
  DZFILL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{FLOW}{\code{rasterDestination} - Contributing Area. Path for new raster layer.}

\item{PREPROC}{\code{enum}  of \verb{("(0) none", "(1) fill sinks temporarily", "(2) fill sinks permanently")} - Preprocessing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DZFILL}{\code{number} - Fill Increment. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Flow accumulation (qm of esp) (sagang:flowaccumulationqmofesp). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FLOW: Contributing Area 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PREPROC: Preprocessing 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) none 		- 1: (1) fill sinks temporarily 		- 2: (2) fill sinks permanently 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DZFILL: Fill Increment 	Default value:	0.01 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLOW - outputRaster - Contributing Area
}
}
}
