% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_flowbetweenfields.R
\name{sagang_flowbetweenfields}
\alias{sagang_flowbetweenfields}
\title{QGIS algorithm - Flow between fields}
\usage{
sagang_flowbetweenfields(
  ELEVATION = qgisprocess:::qgis_default_value(),
  FIELDS = qgisprocess:::qgis_default_value(),
  FLOW = qgisprocess:::qgis_default_value(),
  UPAREA = qgisprocess:::qgis_default_value(),
  STOP = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{FIELDS}{\code{raster} - FIELDS. Path to a raster layer.}

\item{FLOW}{\code{vectorDestination} - Flow table. Path for new vector layer.}

\item{UPAREA}{\code{rasterDestination} - Uparea. Path for new raster layer.}

\item{STOP}{\code{boolean} - Stop at edge. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Flow between fields (sagang:flowbetweenfields). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FIELDS: FIELDS 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FLOW: Flow table 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer UPAREA: Uparea 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STOP: Stop at edge 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLOW - outputVector - Flow table
\item UPAREA - outputRaster - Uparea
}
}
}
