% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_flowpathlength.R
\name{sagang_flowpathlength}
\alias{sagang_flowpathlength}
\title{QGIS algorithm - Flow path length}
\usage{
sagang_flowpathlength(
  ELEVATION = qgisprocess:::qgis_default_value(),
  SEED = qgisprocess:::qgis_default_value(),
  LENGTH = qgisprocess:::qgis_default_value(),
  SEEDS_ONLY = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  CONVERGENCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{SEED}{\code{raster} - Seeds. Path to a raster layer.}

\item{LENGTH}{\code{rasterDestination} - Flow Path Length. Path for new raster layer.}

\item{SEEDS_ONLY}{\code{boolean} - Seeds Only. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) Deterministic 8 (D8)", "(1) Multiple Flow Direction (FD8)")} - Flow Routing Algorithm. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONVERGENCE}{\code{number} - Convergence (FD8). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Flow path length (sagang:flowpathlength). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SEED: Seeds (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LENGTH: Flow Path Length 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SEEDS_ONLY: Seeds Only 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Flow Routing Algorithm 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Deterministic 8 (D8) 		- 1: (1) Multiple Flow Direction (FD8) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CONVERGENCE: Convergence (FD8) 	Default value:	1.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item LENGTH - outputRaster - Flow Path Length
}
}
}
