% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_focalmechanismbeachballplots.R
\name{sagang_focalmechanismbeachballplots}
\alias{sagang_focalmechanismbeachballplots}
\title{QGIS algorithm - Focal mechanism (beachball plots)}
\usage{
sagang_focalmechanismbeachballplots(
  POINTS = qgisprocess:::qgis_default_value(),
  STRIKE = qgisprocess:::qgis_default_value(),
  DIP = qgisprocess:::qgis_default_value(),
  RAKE = qgisprocess:::qgis_default_value(),
  SIZE = qgisprocess:::qgis_default_value(),
  SIZE_DEF = qgisprocess:::qgis_default_value(),
  PLOTS = qgisprocess:::qgis_default_value(),
  DARC = qgisprocess:::qgis_default_value(),
  STYLE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{STRIKE}{\code{field} - Strike. The name of an existing field. ; delimited list of existing field names.}

\item{DIP}{\code{field} - Dip. The name of an existing field. ; delimited list of existing field names.}

\item{RAKE}{\code{field} - Rake. The name of an existing field. ; delimited list of existing field names.}

\item{SIZE}{\code{field} - Size. The name of an existing field. ; delimited list of existing field names.}

\item{SIZE_DEF}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PLOTS}{\code{vectorDestination} - Focal Mechanism Beachballs. Path for new vector layer.}

\item{DARC}{\code{number} - Arc Vertex Distance (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{STYLE}{\code{enum}  of \verb{("(0) one", "(1) two")} - Style. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Focal mechanism (beachball plots) (sagang:focalmechanismbeachballplots). ---------------- Arguments ----------------  POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer STRIKE: Strike 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names DIP: Dip 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names RAKE: Rake 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names SIZE: Size 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names SIZE_DEF: Default 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PLOTS: Focal Mechanism Beachballs 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DARC: Arc Vertex Distance (Degree) 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression STYLE: Style 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) one 		- 1: (1) two 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item PLOTS - outputVector - Focal Mechanism Beachballs
}
}
}
