% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_focalstatistics.R
\name{sagang_focalstatistics}
\alias{sagang_focalstatistics}
\title{QGIS algorithm - Focal statistics}
\usage{
sagang_focalstatistics(
  GRID = qgisprocess:::qgis_default_value(),
  MEAN = qgisprocess:::qgis_default_value(),
  MIN = qgisprocess:::qgis_default_value(),
  MAX = qgisprocess:::qgis_default_value(),
  RANGE = qgisprocess:::qgis_default_value(),
  STDDEV = qgisprocess:::qgis_default_value(),
  VARIANCE = qgisprocess:::qgis_default_value(),
  SUM = qgisprocess:::qgis_default_value(),
  DIFF = qgisprocess:::qgis_default_value(),
  DEVMEAN = qgisprocess:::qgis_default_value(),
  PERCENT = qgisprocess:::qgis_default_value(),
  MEDIAN = qgisprocess:::qgis_default_value(),
  MINORITY = qgisprocess:::qgis_default_value(),
  MAJORITY = qgisprocess:::qgis_default_value(),
  BCENTER = qgisprocess:::qgis_default_value(),
  KERNEL_TYPE = qgisprocess:::qgis_default_value(),
  KERNEL_INNER = qgisprocess:::qgis_default_value(),
  KERNEL_RADIUS = qgisprocess:::qgis_default_value(),
  KERNEL_DIRECTION = qgisprocess:::qgis_default_value(),
  KERNEL_TOLERANCE = qgisprocess:::qgis_default_value(),
  DW_WEIGHTING = qgisprocess:::qgis_default_value(),
  DW_IDW_POWER = qgisprocess:::qgis_default_value(),
  DW_BANDWIDTH = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{MEAN}{\code{rasterDestination} - Mean Value. Path for new raster layer.}

\item{MIN}{\code{rasterDestination} - Minimum Value. Path for new raster layer.}

\item{MAX}{\code{rasterDestination} - Maximum Value. Path for new raster layer.}

\item{RANGE}{\code{rasterDestination} - Value Range. Path for new raster layer.}

\item{STDDEV}{\code{rasterDestination} - Standard Deviation. Path for new raster layer.}

\item{VARIANCE}{\code{rasterDestination} - Variance. Path for new raster layer.}

\item{SUM}{\code{rasterDestination} - Sum. Path for new raster layer.}

\item{DIFF}{\code{rasterDestination} - Difference from Mean Value. Path for new raster layer.}

\item{DEVMEAN}{\code{rasterDestination} - Deviation from Mean Value. Path for new raster layer.}

\item{PERCENT}{\code{rasterDestination} - Percentile. Path for new raster layer.}

\item{MEDIAN}{\code{rasterDestination} - Median. Path for new raster layer.}

\item{MINORITY}{\code{rasterDestination} - Minority. Path for new raster layer.}

\item{MAJORITY}{\code{rasterDestination} - Majority. Path for new raster layer.}

\item{BCENTER}{\code{boolean} - Include Center Cell. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KERNEL_TYPE}{\code{enum}  of \verb{("(0) Square", "(1) Circle", "(2) Annulus", "(3) Sector")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_INNER}{\code{number} - Inner Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KERNEL_RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KERNEL_DIRECTION}{\code{number} - Direction. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KERNEL_TOLERANCE}{\code{number} - Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DW_WEIGHTING}{\code{enum}  of \verb{("(0) no distance weighting", "(1) inverse distance to a power", "(2) exponential", "(3) gaussian")} - Weighting Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DW_IDW_POWER}{\code{number} - Power. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DW_BANDWIDTH}{\code{number} - Bandwidth. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Focal statistics (sagang:focalstatistics). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MEAN: Mean Value 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MIN: Minimum Value 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MAX: Maximum Value 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RANGE: Value Range 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STDDEV: Standard Deviation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VARIANCE: Variance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SUM: Sum 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIFF: Difference from Mean Value 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DEVMEAN: Deviation from Mean Value 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PERCENT: Percentile 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MEDIAN: Median 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MINORITY: Minority 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MAJORITY: Majority 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BCENTER: Include Center Cell 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KERNEL_TYPE: Kernel Type 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Square 		- 1: (1) Circle 		- 2: (2) Annulus 		- 3: (3) Sector 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' KERNEL_INNER: Inner Radius 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KERNEL_RADIUS: Radius 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KERNEL_DIRECTION: Direction 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KERNEL_TOLERANCE: Tolerance 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DW_WEIGHTING: Weighting Function 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) no distance weighting 		- 1: (1) inverse distance to a power 		- 2: (2) exponential 		- 3: (3) gaussian 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DW_IDW_POWER: Power 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DW_BANDWIDTH: Bandwidth 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item DEVMEAN - outputRaster - Deviation from Mean Value
\item DIFF - outputRaster - Difference from Mean Value
\item MAJORITY - outputRaster - Majority
\item MAX - outputRaster - Maximum Value
\item MEAN - outputRaster - Mean Value
\item MEDIAN - outputRaster - Median
\item MIN - outputRaster - Minimum Value
\item MINORITY - outputRaster - Minority
\item PERCENT - outputRaster - Percentile
\item RANGE - outputRaster - Value Range
\item STDDEV - outputRaster - Standard Deviation
\item SUM - outputRaster - Sum
\item VARIANCE - outputRaster - Variance
}
}
}
