% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_fragmentationalternative.R
\name{sagang_fragmentationalternative}
\alias{sagang_fragmentationalternative}
\title{QGIS algorithm - Fragmentation (alternative)}
\usage{
sagang_fragmentationalternative(
  CLASSES = qgisprocess:::qgis_default_value(),
  DENSITY = qgisprocess:::qgis_default_value(),
  CONNECTIVITY = qgisprocess:::qgis_default_value(),
  FRAGMENTATION = qgisprocess:::qgis_default_value(),
  FRAGSTATS = qgisprocess:::qgis_default_value(),
  CLASS = qgisprocess:::qgis_default_value(),
  AGGREGATION = qgisprocess:::qgis_default_value(),
  BORDER = qgisprocess:::qgis_default_value(),
  WEIGHT = qgisprocess:::qgis_default_value(),
  DENSITY_MIN = qgisprocess:::qgis_default_value(),
  DENSITY_INT = qgisprocess:::qgis_default_value(),
  LEVEL_GROW = qgisprocess:::qgis_default_value(),
  DENSITY_MEAN = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CLASSES}{\code{raster} - Classification. Path to a raster layer.}

\item{DENSITY}{\code{rasterDestination} - Density (Percent). Path for new raster layer.}

\item{CONNECTIVITY}{\code{rasterDestination} - Connectivity (Percent). Path for new raster layer.}

\item{FRAGMENTATION}{\code{rasterDestination} - Fragmentation. Path for new raster layer.}

\item{FRAGSTATS}{\code{vectorDestination} - Summary. Path for new vector layer.}

\item{CLASS}{\code{number} - Class Identifier. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{AGGREGATION}{\code{enum}  of \verb{("(0) average", "(1) multiplicative")} - Level Aggregation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{BORDER}{\code{boolean} - Add Border. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WEIGHT}{\code{number} - Connectivity Weighting. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DENSITY_MIN}{\code{number} - Minimum Density (Percent). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DENSITY_INT}{\code{number} - Minimum Density for Interior Forest (Percent). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LEVEL_GROW}{\code{number} - Search Distance Increment. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DENSITY_MEAN}{\code{boolean} - Density from Neighbourhood. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Fragmentation (alternative) (sagang:fragmentationalternative). ---------------- Arguments ----------------  CLASSES: Classification 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DENSITY: Density (Percent) 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONNECTIVITY: Connectivity (Percent) 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FRAGMENTATION: Fragmentation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FRAGSTATS: Summary 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer CLASS: Class Identifier 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression AGGREGATION: Level Aggregation 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) average 		- 1: (1) multiplicative 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' BORDER: Add Border 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WEIGHT: Connectivity Weighting 	Default value:	1.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DENSITY_MIN: Minimum Density (Percent) 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DENSITY_INT: Minimum Density for Interior Forest (Percent) 	Default value:	99 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LEVEL_GROW: Search Distance Increment 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DENSITY_MEAN: Density from Neighbourhood 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONNECTIVITY - outputRaster - Connectivity (Percent)
\item DENSITY - outputRaster - Density (Percent)
\item FRAGMENTATION - outputRaster - Fragmentation
\item FRAGSTATS - outputVector - Summary
}
}
}
