% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sagang_fragmentationclassesfromdensityandconnectivity.R
\name{sagang_fragmentationclassesfromdensityandconnectivity}
\alias{sagang_fragmentationclassesfromdensityandconnectivity}
\title{QGIS algorithm - Fragmentation classes from density and connectivity}
\usage{
sagang_fragmentationclassesfromdensityandconnectivity(
  DENSITY = qgisprocess:::qgis_default_value(),
  CONNECTIVITY = qgisprocess:::qgis_default_value(),
  FRAGMENTATION = qgisprocess:::qgis_default_value(),
  BORDER = qgisprocess:::qgis_default_value(),
  WEIGHT = qgisprocess:::qgis_default_value(),
  DENSITY_MIN = qgisprocess:::qgis_default_value(),
  DENSITY_INT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DENSITY}{\code{raster} - Density (Percent). Path to a raster layer.}

\item{CONNECTIVITY}{\code{raster} - Connectivity (Percent). Path to a raster layer.}

\item{FRAGMENTATION}{\code{rasterDestination} - Fragmentation. Path for new raster layer.}

\item{BORDER}{\code{boolean} - Add Border. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WEIGHT}{\code{number} - Connectivity Weighting. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DENSITY_MIN}{\code{number} - Minimum Density (Percent). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DENSITY_INT}{\code{number} - Minimum Density for Interior Forest (Percent). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Fragmentation classes from density and connectivity (sagang:fragmentationclassesfromdensityandconnectivity). ---------------- Arguments ----------------  DENSITY: Density (Percent) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CONNECTIVITY: Connectivity (Percent) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FRAGMENTATION: Fragmentation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BORDER: Add Border 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WEIGHT: Connectivity Weighting 	Default value:	1.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DENSITY_MIN: Minimum Density (Percent) 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DENSITY_INT: Minimum Density for Interior Forest (Percent) 	Default value:	99 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FRAGMENTATION - outputRaster - Fragmentation
}
}
}
