% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_frostchangefrequency.R
\name{sagang_frostchangefrequency}
\alias{sagang_frostchangefrequency}
\title{QGIS algorithm - Frost change frequency}
\usage{
sagang_frostchangefrequency(
  TMIN = qgisprocess:::qgis_default_value(),
  TMAX = qgisprocess:::qgis_default_value(),
  FREQUENCY = qgisprocess:::qgis_default_value(),
  DT_MEAN = qgisprocess:::qgis_default_value(),
  DT_MAX = qgisprocess:::qgis_default_value(),
  DT_STDV = qgisprocess:::qgis_default_value(),
  TMIN_MEAN = qgisprocess:::qgis_default_value(),
  TMIN_MIN = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TMIN}{\code{multilayer} - Minimum Temperature. .}

\item{TMAX}{\code{multilayer} - Maximum Temperature. .}

\item{FREQUENCY}{\code{rasterDestination} - Frost Change Frequency. Path for new raster layer.}

\item{DT_MEAN}{\code{rasterDestination} - Mean Temperature Span. Path for new raster layer.}

\item{DT_MAX}{\code{rasterDestination} - Maximum Temperature Span. Path for new raster layer.}

\item{DT_STDV}{\code{rasterDestination} - Standard Deviation of Temperature Span. Path for new raster layer.}

\item{TMIN_MEAN}{\code{rasterDestination} - Mean Minimum Temperature. Path for new raster layer.}

\item{TMIN_MIN}{\code{rasterDestination} - Minimum Temperature. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Frost change frequency (sagang:frostchangefrequency). ---------------- Arguments ----------------  TMIN: Minimum Temperature 	Argument type:	multilayer TMAX: Maximum Temperature 	Argument type:	multilayer FREQUENCY: Frost Change Frequency 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DT_MEAN: Mean Temperature Span 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DT_MAX: Maximum Temperature Span 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DT_STDV: Standard Deviation of Temperature Span 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TMIN_MEAN: Mean Minimum Temperature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TMIN_MIN: Minimum Temperature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item DT_MAX - outputRaster - Maximum Temperature Span
\item DT_MEAN - outputRaster - Mean Temperature Span
\item DT_STDV - outputRaster - Standard Deviation of Temperature Span
\item FREQUENCY - outputRaster - Frost Change Frequency
\item TMIN_MEAN - outputRaster - Mean Minimum Temperature
\item TMIN_MIN - outputRaster - Minimum Temperature
}
}
}
