% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_fuzzify.R
\name{sagang_fuzzify}
\alias{sagang_fuzzify}
\title{QGIS algorithm - Fuzzify}
\usage{
sagang_fuzzify(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  INC_MIN = qgisprocess:::qgis_default_value(),
  INC_MAX = qgisprocess:::qgis_default_value(),
  DEC_MIN = qgisprocess:::qgis_default_value(),
  DEC_MAX = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  TRANSITION = qgisprocess:::qgis_default_value(),
  INVERT = qgisprocess:::qgis_default_value(),
  AUTOFIT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Fuzzified Grid. Path for new raster layer.}

\item{INC_MIN}{\code{number} - From. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INC_MAX}{\code{number} - To. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEC_MIN}{\code{number} - From. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEC_MAX}{\code{number} - To. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) Increase", "(1) Decrease", "(2) Increase and Decrease")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TRANSITION}{\code{enum}  of \verb{("(0) linear", "(1) sigmoidal", "(2) j-shaped")} - Transition. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{INVERT}{\code{boolean} - Invert. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{AUTOFIT}{\code{boolean} - Adjust. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Fuzzify (sagang:fuzzify). ---------------- Arguments ----------------  INPUT: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OUTPUT: Fuzzified Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer INC_MIN: From 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression INC_MAX: To 	Default value:	0.3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DEC_MIN: From 	Default value:	0.7 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DEC_MAX: To 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Increase 		- 1: (1) Decrease 		- 2: (2) Increase and Decrease 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' TRANSITION: Transition 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) linear 		- 1: (1) sigmoidal 		- 2: (2) j-shaped 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' INVERT: Invert 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression AUTOFIT: Adjust 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Fuzzified Grid
}
}
}
