% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_fuzzyunionor.R
\name{sagang_fuzzyunionor}
\alias{sagang_fuzzyunionor}
\title{QGIS algorithm - Fuzzy union (or)}
\usage{
sagang_fuzzyunionor(
  GRIDS = qgisprocess:::qgis_default_value(),
  OR = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{OR}{\code{rasterDestination} - Union. Path for new raster layer.}

\item{TYPE}{\code{enum}  of \verb{("(0) max(a, b) (non-interactive)", "(1) a + b - a * b", "(2) min(1, a + b)")} - Operator Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Fuzzy union (or) (sagang:fuzzyunionor). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	multilayer OR: Union 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TYPE: Operator Type 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) max(a, b) (non-interactive) 		- 1: (1) a + b - a * b 		- 2: (2) min(1, a + b) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item OR - outputRaster - Union
}
}
}
