% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_gdalformats.R
\name{sagang_gdalformats}
\alias{sagang_gdalformats}
\title{QGIS algorithm - Gdal formats}
\usage{
sagang_gdalformats(
  FORMATS = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  ACCESS = qgisprocess:::qgis_default_value(),
  RECOGNIZED = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FORMATS}{\code{vectorDestination} - GDAL Formats. Path for new vector layer.}

\item{TYPE}{\code{enum}  of \verb{("(0) raster", "(1) vector", "(2) all")} - Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ACCESS}{\code{enum}  of \verb{("(0) read", "(1) write", "(2) read or write")} - Access. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RECOGNIZED}{\code{boolean} - All Recognized Files. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Gdal formats (sagang:gdalformats). ---------------- Arguments ----------------  FORMATS: GDAL Formats 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer TYPE: Type 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) raster 		- 1: (1) vector 		- 2: (2) all 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' ACCESS: Access 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) read 		- 1: (1) write 		- 2: (2) read or write 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RECOGNIZED: All Recognized Files 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FORMATS - outputVector - GDAL Formats
}
}
}
