% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_generateshapes.R
\name{sagang_generateshapes}
\alias{sagang_generateshapes}
\title{QGIS algorithm - Generate shapes}
\usage{
sagang_generateshapes(
  INPUT = qgisprocess:::qgis_default_value(),
  FIELD_ID = qgisprocess:::qgis_default_value(),
  FIELD_X = qgisprocess:::qgis_default_value(),
  FIELD_Y = qgisprocess:::qgis_default_value(),
  FIELD_Z = qgisprocess:::qgis_default_value(),
  SHAPE_TYPE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input. Path to a vector layer.}

\item{FIELD_ID}{\code{field} - ID. The name of an existing field. ; delimited list of existing field names.}

\item{FIELD_X}{\code{field} - X. The name of an existing field. ; delimited list of existing field names.}

\item{FIELD_Y}{\code{field} - Y. The name of an existing field. ; delimited list of existing field names.}

\item{FIELD_Z}{\code{field} - Z. The name of an existing field. ; delimited list of existing field names.}

\item{SHAPE_TYPE}{\code{enum}  of \verb{("(0) Point(s)", "(1) Line(s)", "(2) Polygon(s)")} - Shape Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{vectorDestination} - Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Generate shapes (sagang:generateshapes). ---------------- Arguments ----------------  INPUT: Input 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD_ID: ID 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FIELD_X: X 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FIELD_Y: Y 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FIELD_Z: Z 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names SHAPE_TYPE: Shape Type 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Point(s) 		- 1: (1) Line(s) 		- 2: (2) Polygon(s) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' OUTPUT: Output 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output
}
}
}
