% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_geocoding.R
\name{sagang_geocoding}
\alias{sagang_geocoding}
\title{QGIS algorithm - Geocoding}
\usage{
sagang_geocoding(
  LOCATIONS = qgisprocess:::qgis_default_value(),
  ADDRESSES = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  ADDRESS = qgisprocess:::qgis_default_value(),
  PROVIDER = qgisprocess:::qgis_default_value(),
  API_KEY = qgisprocess:::qgis_default_value(),
  METADATA = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LOCATIONS}{\code{vectorDestination} - Locations. Path for new vector layer.}

\item{ADDRESSES}{\code{source} - Address List. Path to a vector layer.}

\item{FIELD}{\code{field} - Address Field. The name of an existing field. ; delimited list of existing field names.}

\item{ADDRESS}{\code{string} - Single Address. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PROVIDER}{\code{enum}  of \verb{("(0) Nominatim (OpenStreetMap)", "(1) The Data Science Toolkit", "(2) Google", "(3) Bing", "(4) MapQuest")} - Service Provider. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{API_KEY}{\code{string} - API Key. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METADATA}{\code{boolean} - Store Metadata. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Geocoding (sagang:geocoding). ---------------- Arguments ----------------  LOCATIONS: Locations 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer ADDRESSES: Address List (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Address Field 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names ADDRESS: Single Address 	Default value:	Bundesstrasse 55, Hamburg, Germany 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PROVIDER: Service Provider 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Nominatim (OpenStreetMap) 		- 1: (1) The Data Science Toolkit 		- 2: (2) Google 		- 3: (3) Bing 		- 4: (4) MapQuest 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' API_KEY: API Key 	Default value:	 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METADATA: Store Metadata 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item LOCATIONS - outputVector - Locations
}
}
}
