% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_geodesicmorphologicalreconstruction.R
\name{sagang_geodesicmorphologicalreconstruction}
\alias{sagang_geodesicmorphologicalreconstruction}
\title{QGIS algorithm - Geodesic morphological reconstruction}
\usage{
sagang_geodesicmorphologicalreconstruction(
  INPUT_GRID = qgisprocess:::qgis_default_value(),
  OBJECT_GRID = qgisprocess:::qgis_default_value(),
  DIFFERENCE_GRID = qgisprocess:::qgis_default_value(),
  SHIFT_VALUE = qgisprocess:::qgis_default_value(),
  BORDER_YES_NO = qgisprocess:::qgis_default_value(),
  BIN_YES_NO = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_GRID}{\code{raster} - Input Grid. Path to a raster layer.}

\item{OBJECT_GRID}{\code{rasterDestination} - Object Grid. Path for new raster layer.}

\item{DIFFERENCE_GRID}{\code{rasterDestination} - Difference Input - Reconstruction. Path for new raster layer.}

\item{SHIFT_VALUE}{\code{number} - Shift value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{BORDER_YES_NO}{\code{boolean} - Preserve 1px border Yes\code{/}No. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{BIN_YES_NO}{\code{boolean} - Create a binary mask Yes\code{/}No. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRESHOLD}{\code{number} - Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Geodesic morphological reconstruction (sagang:geodesicmorphologicalreconstruction). ---------------- Arguments ----------------  INPUT_GRID: Input Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OBJECT_GRID: Object Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIFFERENCE_GRID: Difference Input - Reconstruction 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SHIFT_VALUE: Shift value 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression BORDER_YES_NO: Preserve 1px border Yes/No 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression BIN_YES_NO: Create a binary mask Yes/No 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRESHOLD: Threshold 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIFFERENCE_GRID - outputRaster - Difference Input - Reconstruction
\item OBJECT_GRID - outputRaster - Object Grid
}
}
}
