% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_geographicdistances.R
\name{sagang_geographicdistances}
\alias{sagang_geographicdistances}
\title{QGIS algorithm - Geographic distances}
\usage{
sagang_geographicdistances(
  PLANAR = qgisprocess:::qgis_default_value(),
  ORTHODROME = qgisprocess:::qgis_default_value(),
  LOXODROME = qgisprocess:::qgis_default_value(),
  EPSILON = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{PLANAR}{\code{source} - Segments. Path to a vector layer.}

\item{ORTHODROME}{\code{vectorDestination} - Great Elliptic. Path for new vector layer.}

\item{LOXODROME}{\code{vectorDestination} - Loxodrome. Path for new vector layer.}

\item{EPSILON}{\code{number} - Epsilon. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Geographic distances (sagang:geographicdistances). ---------------- Arguments ----------------  PLANAR: Segments 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ORTHODROME: Great Elliptic 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer LOXODROME: Loxodrome 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer EPSILON: Epsilon 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item LOXODROME - outputVector - Loxodrome
\item ORTHODROME - outputVector - Great Elliptic
}
}
}
