% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_geometricfigures.R
\name{sagang_geometricfigures}
\alias{sagang_geometricfigures}
\title{QGIS algorithm - Geometric figures}
\usage{
sagang_geometricfigures(
  RESULT = qgisprocess:::qgis_default_value(),
  CELL_COUNT = qgisprocess:::qgis_default_value(),
  CELL_SIZE = qgisprocess:::qgis_default_value(),
  FIGURE = qgisprocess:::qgis_default_value(),
  PLANE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{RESULT}{\code{rasterDestination} - Result. Path for new raster layer.}

\item{CELL_COUNT}{\code{number} - Cell Count. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CELL_SIZE}{\code{number} - Cell Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FIGURE}{\code{enum}  of \verb{("(0) Cone (up)", "(1) Cone (down)", "(2) Plane")} - Figure. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PLANE}{\code{number} - Direction of Plane (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Geometric figures (sagang:geometricfigures). ---------------- Arguments ----------------  RESULT: Result 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CELL_COUNT: Cell Count 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CELL_SIZE: Cell Size 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FIGURE: Figure 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Cone (up) 		- 1: (1) Cone (down) 		- 2: (2) Plane 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' PLANE: Direction of Plane (Degree) 	Default value:	22.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Result
}
}
}
