% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_getgridfromvirtualpointcloud.R
\name{sagang_getgridfromvirtualpointcloud}
\alias{sagang_getgridfromvirtualpointcloud}
\title{QGIS algorithm - Get grid from virtual point cloud}
\usage{
sagang_getgridfromvirtualpointcloud(
  FILENAME = qgisprocess:::qgis_default_value(),
  GRID_OUT = qgisprocess:::qgis_default_value(),
  FILEPATH = qgisprocess:::qgis_default_value(),
  ATTR_FIELD_GRID = qgisprocess:::qgis_default_value(),
  CELL_SIZE = qgisprocess:::qgis_default_value(),
  GRID_SYSTEM_FIT = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  CONSTRAIN_QUERY = qgisprocess:::qgis_default_value(),
  ATTR_FIELD = qgisprocess:::qgis_default_value(),
  AOI_SHP = qgisprocess:::qgis_default_value(),
  FIELD_TILENAME = qgisprocess:::qgis_default_value(),
  AOI_GRID = qgisprocess:::qgis_default_value(),
  AOI_ADD_OVERLAP = qgisprocess:::qgis_default_value(),
  OVERLAP = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FILENAME}{\code{file} - Filename. Path to a file.}

\item{GRID_OUT}{\code{rasterDestination} - Grid. Path for new raster layer.}

\item{FILEPATH}{\code{file} - Optional Output Filepath. Path to a file.}

\item{ATTR_FIELD_GRID}{\code{number} - Attribute Field to Grid. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CELL_SIZE}{\code{number} - Cellsize. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRID_SYSTEM_FIT}{\code{enum}  of \verb{("(0) nodes", "(1) cells")} - Grid System Fit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) lowest", "(1) highest")} - Aggregation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONSTRAIN_QUERY}{\code{boolean} - Constrain Query. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ATTR_FIELD}{\code{number} - Attribute Field. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{AOI_SHP}{\code{source} - Shape. Path to a vector layer.}

\item{FIELD_TILENAME}{\code{field} - Tilename. The name of an existing field. ; delimited list of existing field names.}

\item{AOI_GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{AOI_ADD_OVERLAP}{\code{boolean} - Add Overlap. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OVERLAP}{\code{number} - Overlap. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Get grid from virtual point cloud (sagang:getgridfromvirtualpointcloud). ---------------- Arguments ----------------  FILENAME: Filename 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file GRID_OUT: Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FILEPATH: Optional Output Filepath 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file ATTR_FIELD_GRID: Attribute Field to Grid 	Default value:	3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CELL_SIZE: Cellsize 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GRID_SYSTEM_FIT: Grid System Fit 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) nodes 		- 1: (1) cells 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Aggregation 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) lowest 		- 1: (1) highest 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CONSTRAIN_QUERY: Constrain Query 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ATTR_FIELD: Attribute Field 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression AOI_SHP: Shape (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD_TILENAME: Tilename 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names AOI_GRID: Grid (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer AOI_ADD_OVERLAP: Add Overlap 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OVERLAP: Overlap 	Default value:	50 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRID_OUT - outputRaster - Grid
}
}
}
