% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_globalmoransiforrasterlayer.R
\name{sagang_globalmoransiforrasterlayer}
\alias{sagang_globalmoransiforrasterlayer}
\title{QGIS algorithm - Global moran's i for raster layer}
\usage{
sagang_globalmoransiforrasterlayer(
  GRID = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  CONTIGUITY = qgisprocess:::qgis_default_value(),
  DIALOG = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{RESULT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{CONTIGUITY}{\code{enum}  of \verb{("(0) Rook", "(1) Queen")} - Case of contiguity. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DIALOG}{\code{boolean} - Show Result in Dialog. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Global moran's i for raster layer (sagang:globalmoransiforrasterlayer). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RESULT: Result 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer CONTIGUITY: Case of contiguity 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Rook 		- 1: (1) Queen 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DIALOG: Show Result in Dialog 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Result
}
}
}
