% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_gpxtoshapefile.R
\name{sagang_gpxtoshapefile}
\alias{sagang_gpxtoshapefile}
\title{QGIS algorithm - Gpx to shapefile}
\usage{
sagang_gpxtoshapefile(
  BASEPATH = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  TRACKPOINTS = qgisprocess:::qgis_default_value(),
  WAYPOINTS = qgisprocess:::qgis_default_value(),
  ROUTES = qgisprocess:::qgis_default_value(),
  ADD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{BASEPATH}{\code{file} - Gpx2shp path. Path to a file.}

\item{FILE}{\code{file} - GPX file. Path to a file.}

\item{TRACKPOINTS}{\code{boolean} - Convert track points. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WAYPOINTS}{\code{boolean} - Convert way points. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ROUTES}{\code{boolean} - Convert routes. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ADD}{\code{boolean} - Load shapefile. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Gpx to shapefile (sagang:gpxtoshapefile). ---------------- Arguments ----------------  BASEPATH: Gpx2shp path 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file FILE: GPX file 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file TRACKPOINTS: Convert track points 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WAYPOINTS: Convert way points 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ROUTES: Convert routes 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ADD: Load shapefile 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
