% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sagang_gradientvectorfromcartesiantopolarcoordinates.R
\name{sagang_gradientvectorfromcartesiantopolarcoordinates}
\alias{sagang_gradientvectorfromcartesiantopolarcoordinates}
\title{QGIS algorithm - Gradient vector from cartesian to polar coordinates}
\usage{
sagang_gradientvectorfromcartesiantopolarcoordinates(
  DX = qgisprocess:::qgis_default_value(),
  DY = qgisprocess:::qgis_default_value(),
  DIR = qgisprocess:::qgis_default_value(),
  LEN = qgisprocess:::qgis_default_value(),
  UNITS = qgisprocess:::qgis_default_value(),
  SYSTEM = qgisprocess:::qgis_default_value(),
  SYSTEM_ZERO = qgisprocess:::qgis_default_value(),
  SYSTEM_ORIENT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DX}{\code{raster} - X Component. Path to a raster layer.}

\item{DY}{\code{raster} - Y Component. Path to a raster layer.}

\item{DIR}{\code{rasterDestination} - Direction. Path for new raster layer.}

\item{LEN}{\code{rasterDestination} - Length. Path for new raster layer.}

\item{UNITS}{\code{enum}  of \verb{("(0) radians", "(1) degree")} - Polar Angle Units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SYSTEM}{\code{enum}  of \verb{("(0) mathematical", "(1) geographical", "(2) user defined")} - Polar Coordinate System. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SYSTEM_ZERO}{\code{number} - User defined Zero Direction. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SYSTEM_ORIENT}{\code{enum}  of \verb{("(0) clockwise", "(1) counterclockwise")} - User defined Orientation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Gradient vector from cartesian to polar coordinates (sagang:gradientvectorfromcartesiantopolarcoordinates). ---------------- Arguments ----------------  DX: X Component 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DY: Y Component 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DIR: Direction 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LEN: Length 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer UNITS: Polar Angle Units 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) radians 		- 1: (1) degree 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SYSTEM: Polar Coordinate System 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) mathematical 		- 1: (1) geographical 		- 2: (2) user defined 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SYSTEM_ZERO: User defined Zero Direction 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SYSTEM_ORIENT: User defined Orientation 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) clockwise 		- 1: (1) counterclockwise 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIR - outputRaster - Direction
\item LEN - outputRaster - Length
}
}
}
