% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sagang_gradientvectorsfromdirectionalcomponents.R
\name{sagang_gradientvectorsfromdirectionalcomponents}
\alias{sagang_gradientvectorsfromdirectionalcomponents}
\title{QGIS algorithm - Gradient vectors from directional components}
\usage{
sagang_gradientvectorsfromdirectionalcomponents(
  X = qgisprocess:::qgis_default_value(),
  Y = qgisprocess:::qgis_default_value(),
  VECTORS = qgisprocess:::qgis_default_value(),
  STEP = qgisprocess:::qgis_default_value(),
  AGGR = qgisprocess:::qgis_default_value(),
  STYLE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{X}{\code{raster} - X Component. Path to a raster layer.}

\item{Y}{\code{raster} - Y Component. Path to a raster layer.}

\item{VECTORS}{\code{vectorDestination} - Gradient Vectors. Path for new vector layer.}

\item{STEP}{\code{number} - Step. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{AGGR}{\code{enum}  of \verb{("(0) nearest neighbour", "(1) mean value")} - Aggregation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{STYLE}{\code{enum}  of \verb{("(0) simple line", "(1) arrow", "(2) arrow (centered to cell)")} - Style. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Gradient vectors from directional components (sagang:gradientvectorsfromdirectionalcomponents). ---------------- Arguments ----------------  X: X Component 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Y: Y Component 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer VECTORS: Gradient Vectors 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer STEP: Step 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression AGGR: Aggregation 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) nearest neighbour 		- 1: (1) mean value 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' STYLE: Style 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) simple line 		- 1: (1) arrow 		- 2: (2) arrow (centered to cell) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item VECTORS - outputVector - Gradient Vectors
}
}
}
