% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_gravitationalprocesspathmodel.R
\name{sagang_gravitationalprocesspathmodel}
\alias{sagang_gravitationalprocesspathmodel}
\title{QGIS algorithm - Gravitational process path model}
\usage{
sagang_gravitationalprocesspathmodel(
  DEM = qgisprocess:::qgis_default_value(),
  RELEASE_AREAS = qgisprocess:::qgis_default_value(),
  MATERIAL = qgisprocess:::qgis_default_value(),
  FRICTION_ANGLE_GRID = qgisprocess:::qgis_default_value(),
  SLOPE_IMPACT_GRID = qgisprocess:::qgis_default_value(),
  FRICTION_MU_GRID = qgisprocess:::qgis_default_value(),
  FRICTION_MASS_TO_DRAG_GRID = qgisprocess:::qgis_default_value(),
  PROCESS_AREA = qgisprocess:::qgis_default_value(),
  DEPOSITION = qgisprocess:::qgis_default_value(),
  MAX_VELOCITY = qgisprocess:::qgis_default_value(),
  STOP_POSITIONS = qgisprocess:::qgis_default_value(),
  PROCESS_PATH_MODEL = qgisprocess:::qgis_default_value(),
  RW_SLOPE_THRES = qgisprocess:::qgis_default_value(),
  RW_EXPONENT = qgisprocess:::qgis_default_value(),
  RW_PERSISTENCE = qgisprocess:::qgis_default_value(),
  GPP_ITERATIONS = qgisprocess:::qgis_default_value(),
  GPP_PROCESSING_ORDER = qgisprocess:::qgis_default_value(),
  GPP_SEED = qgisprocess:::qgis_default_value(),
  FRICTION_MODEL = qgisprocess:::qgis_default_value(),
  FRICTION_THRES_FREE_FALL = qgisprocess:::qgis_default_value(),
  FRICTION_METHOD_IMPACT = qgisprocess:::qgis_default_value(),
  FRICTION_IMPACT_REDUCTION = qgisprocess:::qgis_default_value(),
  FRICTION_ANGLE = qgisprocess:::qgis_default_value(),
  FRICTION_MU = qgisprocess:::qgis_default_value(),
  FRICTION_MODE_OF_MOTION = qgisprocess:::qgis_default_value(),
  FRICTION_MASS_TO_DRAG = qgisprocess:::qgis_default_value(),
  FRICTION_INIT_VELOCITY = qgisprocess:::qgis_default_value(),
  DEPOSITION_MODEL = qgisprocess:::qgis_default_value(),
  DEPOSITION_INITIAL = qgisprocess:::qgis_default_value(),
  DEPOSITION_SLOPE_THRES = qgisprocess:::qgis_default_value(),
  DEPOSITION_VELOCITY_THRES = qgisprocess:::qgis_default_value(),
  DEPOSITION_MAX = qgisprocess:::qgis_default_value(),
  DEPOSITION_MIN_PATH = qgisprocess:::qgis_default_value(),
  SINK_MIN_SLOPE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{RELEASE_AREAS}{\code{raster} - Release Areas. Path to a raster layer.}

\item{MATERIAL}{\code{raster} - Material. Path to a raster layer.}

\item{FRICTION_ANGLE_GRID}{\code{raster} - Friction Angle. Path to a raster layer.}

\item{SLOPE_IMPACT_GRID}{\code{raster} - Slope Impact Areas. Path to a raster layer.}

\item{FRICTION_MU_GRID}{\code{raster} - Friction Parameter Mu. Path to a raster layer.}

\item{FRICTION_MASS_TO_DRAG_GRID}{\code{raster} - Mass to Drag Ratio. Path to a raster layer.}

\item{PROCESS_AREA}{\code{rasterDestination} - Process Area. Path for new raster layer.}

\item{DEPOSITION}{\code{rasterDestination} - Deposition. Path for new raster layer.}

\item{MAX_VELOCITY}{\code{rasterDestination} - Maximum Velocity. Path for new raster layer.}

\item{STOP_POSITIONS}{\code{rasterDestination} - Stopping Positions. Path for new raster layer.}

\item{PROCESS_PATH_MODEL}{\code{enum}  of \verb{("(0) Maximum Slope", "(1) Random Walk")} - Model. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RW_SLOPE_THRES}{\code{number} - Slope Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RW_EXPONENT}{\code{number} - Exponent. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RW_PERSISTENCE}{\code{number} - Persistence Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GPP_ITERATIONS}{\code{number} - Iterations. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GPP_PROCESSING_ORDER}{\code{enum}  of \verb{("(0) RAs in Sequence", "(1) RAs in Sequence per Iteration", "(2) RAs in Parallel per Iteration")} - Processing Order. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GPP_SEED}{\code{number} - Seed Value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FRICTION_MODEL}{\code{enum}  of \verb{("(0) None", "(1) Geometric Gradient (Heim 1932)", "(2) Fahrboeschung Principle (Heim 1932)", "(3) Shadow Angle (Evans & Hungr 1988)", "(4) 1-parameter friction model (Scheidegger 1975)", "(5) PCM Model (Perla et al. 1980)")} - Model. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FRICTION_THRES_FREE_FALL}{\code{number} - Threshold Angle Free Fall. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FRICTION_METHOD_IMPACT}{\code{enum}  of \verb{("(0) Energy Reduction (Scheidegger 1975)", "(1) Preserved Component of Velocity (Kirkby & Statham 1975)")} - Method Impact. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FRICTION_IMPACT_REDUCTION}{\code{number} - Reduction. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FRICTION_ANGLE}{\code{number} - Friction Angle. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FRICTION_MU}{\code{number} - Mu. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FRICTION_MODE_OF_MOTION}{\code{enum}  of \verb{("(0) Sliding", "(1) Rolling")} - Mode of Motion. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FRICTION_MASS_TO_DRAG}{\code{number} - Mass to Drag Ratio. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FRICTION_INIT_VELOCITY}{\code{number} - Initial Velocity. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEPOSITION_MODEL}{\code{enum}  of \verb{("(0) None", "(1) On Stop", "(2) Slope & On Stop", "(3) Velocity & On Stop", "(4) min(Slope,Velocity) & On Stop")} - Model. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DEPOSITION_INITIAL}{\code{number} - Initial Deposition on Stop. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEPOSITION_SLOPE_THRES}{\code{number} - Slope Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEPOSITION_VELOCITY_THRES}{\code{number} - Velocity Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEPOSITION_MAX}{\code{number} - Maximum Deposition along Path. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEPOSITION_MIN_PATH}{\code{number} - Minimum Path Length. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SINK_MIN_SLOPE}{\code{number} - Minimum Slope. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Gravitational process path model (sagang:gravitationalprocesspathmodel). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RELEASE_AREAS: Release Areas 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MATERIAL: Material (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FRICTION_ANGLE_GRID: Friction Angle (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SLOPE_IMPACT_GRID: Slope Impact Areas (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FRICTION_MU_GRID: Friction Parameter Mu (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FRICTION_MASS_TO_DRAG_GRID: Mass to Drag Ratio (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer PROCESS_AREA: Process Area 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DEPOSITION: Deposition 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MAX_VELOCITY: Maximum Velocity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STOP_POSITIONS: Stopping Positions 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PROCESS_PATH_MODEL: Model 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Maximum Slope 		- 1: (1) Random Walk 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RW_SLOPE_THRES: Slope Threshold 	Default value:	40 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RW_EXPONENT: Exponent 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RW_PERSISTENCE: Persistence Factor 	Default value:	1.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GPP_ITERATIONS: Iterations 	Default value:	1000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GPP_PROCESSING_ORDER: Processing Order 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) RAs in Sequence 		- 1: (1) RAs in Sequence per Iteration 		- 2: (2) RAs in Parallel per Iteration 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' GPP_SEED: Seed Value 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FRICTION_MODEL: Model 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) None 		- 1: (1) Geometric Gradient (Heim 1932) 		- 2: (2) Fahrboeschung Principle (Heim 1932) 		- 3: (3) Shadow Angle (Evans & Hungr 1988) 		- 4: (4) 1-parameter friction model (Scheidegger 1975) 		- 5: (5) PCM Model (Perla et al. 1980) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FRICTION_THRES_FREE_FALL: Threshold Angle Free Fall 	Default value:	60 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FRICTION_METHOD_IMPACT: Method Impact 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Energy Reduction (Scheidegger 1975) 		- 1: (1) Preserved Component of Velocity (Kirkby & Statham 1975) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FRICTION_IMPACT_REDUCTION: Reduction 	Default value:	75 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FRICTION_ANGLE: Friction Angle 	Default value:	30 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FRICTION_MU: Mu 	Default value:	0.25 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FRICTION_MODE_OF_MOTION: Mode of Motion 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Sliding 		- 1: (1) Rolling 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FRICTION_MASS_TO_DRAG: Mass to Drag Ratio 	Default value:	200 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FRICTION_INIT_VELOCITY: Initial Velocity 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DEPOSITION_MODEL: Model 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) None 		- 1: (1) On Stop 		- 2: (2) Slope & On Stop 		- 3: (3) Velocity & On Stop 		- 4: (4) min(Slope,Velocity) & On Stop 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DEPOSITION_INITIAL: Initial Deposition on Stop 	Default value:	20 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DEPOSITION_SLOPE_THRES: Slope Threshold 	Default value:	20 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DEPOSITION_VELOCITY_THRES: Velocity Threshold 	Default value:	15 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DEPOSITION_MAX: Maximum Deposition along Path 	Default value:	20 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DEPOSITION_MIN_PATH: Minimum Path Length 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SINK_MIN_SLOPE: Minimum Slope 	Default value:	2.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item DEPOSITION - outputRaster - Deposition
\item MAX_VELOCITY - outputRaster - Maximum Velocity
\item PROCESS_AREA - outputRaster - Process Area
\item STOP_POSITIONS - outputRaster - Stopping Positions
}
}
}
