% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_gridclassesareaforpolygons.R
\name{sagang_gridclassesareaforpolygons}
\alias{sagang_gridclassesareaforpolygons}
\title{QGIS algorithm - Grid classes area for polygons}
\usage{
sagang_gridclassesareaforpolygons(
  POLYGONS = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  GRID = qgisprocess:::qgis_default_value(),
  PROCESS = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  GRID_VALUES = qgisprocess:::qgis_default_value(),
  GRID_LUT = qgisprocess:::qgis_default_value(),
  GRID_LUT_MIN = qgisprocess:::qgis_default_value(),
  GRID_LUT_MAX = qgisprocess:::qgis_default_value(),
  GRID_LUT_NAM = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{RESULT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{PROCESS}{\code{enum}  of \verb{("(0) cell by cell", "(1) polygon by polygon")} - Processing Order. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) cell center", "(1) cell area")} - Cell Area Intersection. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{enum}  of \verb{("(0) total area", "(1) percentage")} - Output Measurment. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRID_VALUES}{\code{enum}  of \verb{("(0) values are class identifiers", "(1) use look-up table")} - Class Definition. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRID_LUT}{\code{source} - Look-up Table. Path to a vector layer.}

\item{GRID_LUT_MIN}{\code{field} - Value. The name of an existing field. ; delimited list of existing field names.}

\item{GRID_LUT_MAX}{\code{field} - Value (Maximum). The name of an existing field. ; delimited list of existing field names.}

\item{GRID_LUT_NAM}{\code{field} - Name. The name of an existing field. ; delimited list of existing field names.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Grid classes area for polygons (sagang:gridclassesareaforpolygons). ---------------- Arguments ----------------  POLYGONS: Polygons 	Argument type:	source 	Acceptable values: 		- Path to a vector layer RESULT: Result 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer PROCESS: Processing Order 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) cell by cell 		- 1: (1) polygon by polygon 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Cell Area Intersection 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) cell center 		- 1: (1) cell area 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' OUTPUT: Output Measurment 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) total area 		- 1: (1) percentage 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' GRID_VALUES: Class Definition 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) values are class identifiers 		- 1: (1) use look-up table 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' GRID_LUT: Look-up Table (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer GRID_LUT_MIN: Value 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names GRID_LUT_MAX: Value (Maximum) 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names GRID_LUT_NAM: Name 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Result
}
}
}
