% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_gridcombination.R
\name{sagang_gridcombination}
\alias{sagang_gridcombination}
\title{QGIS algorithm - Gridcombination}
\usage{
sagang_gridcombination(
  INPUT = qgisprocess:::qgis_default_value(),
  Folder1 = qgisprocess:::qgis_default_value(),
  sY = qgisprocess:::qgis_default_value(),
  eY = qgisprocess:::qgis_default_value(),
  DomW = qgisprocess:::qgis_default_value(),
  ElecW = qgisprocess:::qgis_default_value(),
  LiveW = qgisprocess:::qgis_default_value(),
  ManW = qgisprocess:::qgis_default_value(),
  IrrW = qgisprocess:::qgis_default_value(),
  FvA = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Gelaendemodell (DTM). Path to a raster layer.}

\item{Folder1}{\code{file} - Pfad WaterGap Raster. Path to a file.}

\item{sY}{\code{number} - Start-Jahr. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{eY}{\code{number} - End-Jahr. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DomW}{\code{boolean} - Domestic Water. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ElecW}{\code{boolean} - Electricity Water. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LiveW}{\code{boolean} - Livestock Water. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ManW}{\code{boolean} - Manufacturing Water. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{IrrW}{\code{boolean} - Irrigation Water. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FvA}{\code{enum}  of \verb{("(0) Resultierendes Raster ueber WasserENTNAHME erstellen", "(1) Resultierendes Raster ueber WasserNUTZUNG erstellen")} - Flaechenverbrauch-Auswahl (FvA). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Gridcombination (sagang:gridcombination). ---------------- Arguments ----------------  INPUT: Gelaendemodell (DTM) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Folder1: Pfad WaterGap Raster 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file sY: Start-Jahr 	Default value:	1990 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression eY: End-Jahr 	Default value:	1990 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DomW: Domestic Water 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ElecW: Electricity Water 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LiveW: Livestock Water 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ManW: Manufacturing Water 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression IrrW: Irrigation Water 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FvA: Flaechenverbrauch-Auswahl (FvA) 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Resultierendes Raster ueber WasserENTNAHME erstellen 		- 1: (1) Resultierendes Raster ueber WasserNUTZUNG erstellen 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
