% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_gridsfromclassifiedgridandtable.R
\name{sagang_gridsfromclassifiedgridandtable}
\alias{sagang_gridsfromclassifiedgridandtable}
\title{QGIS algorithm - Grids from classified grid and table}
\usage{
sagang_gridsfromclassifiedgridandtable(
  TABLE = qgisprocess:::qgis_default_value(),
  ID_FIELD = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  GRIDS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TABLE}{\code{source} - Table. Path to a vector layer.}

\item{ID_FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{CLASSES}{\code{raster} - Classes. Path to a raster layer.}

\item{GRIDS}{\code{rasterDestination} - Grids. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Grids from classified grid and table (sagang:gridsfromclassifiedgridandtable). ---------------- Arguments ----------------  TABLE: Table 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ID_FIELD: Attribute 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names CLASSES: Classes 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GRIDS: Grids 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRIDS - outputRaster - Grids
}
}
}
