% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_gridsystemextent.R
\name{sagang_gridsystemextent}
\alias{sagang_gridsystemextent}
\title{QGIS algorithm - Grid system extent}
\usage{
sagang_gridsystemextent(
  EXTENT = qgisprocess:::qgis_default_value(),
  BORDER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{EXTENT}{\code{vectorDestination} - Extent. Path for new vector layer.}

\item{BORDER}{\code{enum}  of \verb{("(0) grid cells", "(1) grid nodes")} - Border. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Grid system extent (sagang:gridsystemextent). ---------------- Arguments ----------------  EXTENT: Extent 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer BORDER: Border 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) grid cells 		- 1: (1) grid nodes 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item EXTENT - outputVector - Extent
}
}
}
