% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_growingdegreedays.R
\name{sagang_growingdegreedays}
\alias{sagang_growingdegreedays}
\title{QGIS algorithm - Growing degree days}
\usage{
sagang_growingdegreedays(
  TMEAN = qgisprocess:::qgis_default_value(),
  TBASE = qgisprocess:::qgis_default_value(),
  NGDD = qgisprocess:::qgis_default_value(),
  TSUM = qgisprocess:::qgis_default_value(),
  FIRST = qgisprocess:::qgis_default_value(),
  LAST = qgisprocess:::qgis_default_value(),
  TARGET = qgisprocess:::qgis_default_value(),
  TTARGET = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TMEAN}{\code{multilayer} - Mean Monthly Temperatures. .}

\item{TBASE}{\code{number} - Base Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NGDD}{\code{rasterDestination} - Number of Days above Base Temperature. Path for new raster layer.}

\item{TSUM}{\code{rasterDestination} - Growing Degree Days. Path for new raster layer.}

\item{FIRST}{\code{rasterDestination} - First Growing Degree Day. Path for new raster layer.}

\item{LAST}{\code{rasterDestination} - Last Growing Degree Day. Path for new raster layer.}

\item{TARGET}{\code{rasterDestination} - Degree Sum Target Day. Path for new raster layer.}

\item{TTARGET}{\code{number} - Degree Sum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Growing degree days (sagang:growingdegreedays). ---------------- Arguments ----------------  TMEAN: Mean Monthly Temperatures 	Argument type:	multilayer TBASE: Base Temperature 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NGDD: Number of Days above Base Temperature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TSUM: Growing Degree Days 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FIRST: First Growing Degree Day 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LAST: Last Growing Degree Day 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TARGET: Degree Sum Target Day 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TTARGET: Degree Sum 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FIRST - outputRaster - First Growing Degree Day
\item LAST - outputRaster - Last Growing Degree Day
\item NGDD - outputRaster - Number of Days above Base Temperature
\item TARGET - outputRaster - Degree Sum Target Day
\item TSUM - outputRaster - Growing Degree Days
}
}
}
