% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_histogrammatching.R
\name{sagang_histogrammatching}
\alias{sagang_histogrammatching}
\title{QGIS algorithm - Histogram matching}
\usage{
sagang_histogrammatching(
  GRID = qgisprocess:::qgis_default_value(),
  MATCHED = qgisprocess:::qgis_default_value(),
  REFERENCE = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  NCLASSES = qgisprocess:::qgis_default_value(),
  MAXSAMPLES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{MATCHED}{\code{rasterDestination} - Adjusted Grid. Path for new raster layer.}

\item{REFERENCE}{\code{raster} - Reference Grid. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) standard deviation", "(1) histogram")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NCLASSES}{\code{number} - Histogramm Classes. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAXSAMPLES}{\code{number} - Maximum Sample Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Histogram matching (sagang:histogrammatching). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MATCHED: Adjusted Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer REFERENCE: Reference Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer METHOD: Method 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) standard deviation 		- 1: (1) histogram 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NCLASSES: Histogramm Classes 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAXSAMPLES: Maximum Sample Size 	Default value:	1e+06 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item MATCHED - outputRaster - Adjusted Grid
}
}
}
