% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_hypsometry.R
\name{sagang_hypsometry}
\alias{sagang_hypsometry}
\title{QGIS algorithm - Hypsometry}
\usage{
sagang_hypsometry(
  ELEVATION = qgisprocess:::qgis_default_value(),
  TABLE = qgisprocess:::qgis_default_value(),
  COUNT = qgisprocess:::qgis_default_value(),
  SORTING = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  BZRANGE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{TABLE}{\code{vectorDestination} - Hypsometry. Path for new vector layer.}

\item{COUNT}{\code{number} - Number of Classes. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SORTING}{\code{enum}  of \verb{("(0) up", "(1) down")} - Sort. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) height", "(1) area")} - Classification Constant. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{BZRANGE}{\code{boolean} - Use Z-Range. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Hypsometry (sagang:hypsometry). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TABLE: Hypsometry 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer COUNT: Number of Classes 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SORTING: Sort 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) up 		- 1: (1) down 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Classification Constant 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) height 		- 1: (1) area 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' BZRANGE: Use Z-Range 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item TABLE - outputVector - Hypsometry
}
}
}
