% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_ihssharpening.R
\name{sagang_ihssharpening}
\alias{sagang_ihssharpening}
\title{QGIS algorithm - Ihs sharpening}
\usage{
sagang_ihssharpening(
  R = qgisprocess:::qgis_default_value(),
  G = qgisprocess:::qgis_default_value(),
  B = qgisprocess:::qgis_default_value(),
  PAN = qgisprocess:::qgis_default_value(),
  R_SHARP = qgisprocess:::qgis_default_value(),
  G_SHARP = qgisprocess:::qgis_default_value(),
  B_SHARP = qgisprocess:::qgis_default_value(),
  SHARPEN = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  PAN_MATCH = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{R}{\code{raster} - Red. Path to a raster layer.}

\item{G}{\code{raster} - Green. Path to a raster layer.}

\item{B}{\code{raster} - Blue. Path to a raster layer.}

\item{PAN}{\code{raster} - Panchromatic Channel. Path to a raster layer.}

\item{R_SHARP}{\code{rasterDestination} - Red. Path for new raster layer.}

\item{G_SHARP}{\code{rasterDestination} - Green. Path for new raster layer.}

\item{B_SHARP}{\code{rasterDestination} - Blue. Path for new raster layer.}

\item{SHARPEN}{\code{rasterDestination} - Sharpened Channels. Path for new raster layer.}

\item{RESAMPLING}{\code{enum}  of \verb{("(0) nearest neighbour", "(1) bilinear", "(2) cubic convolution")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PAN_MATCH}{\code{enum}  of \verb{("(0) normalized", "(1) standardized")} - Panchromatic Channel Matching. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Ihs sharpening (sagang:ihssharpening). ---------------- Arguments ----------------  R: Red 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer G: Green 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer B: Blue 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer PAN: Panchromatic Channel 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer R_SHARP: Red 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer G_SHARP: Green 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer B_SHARP: Blue 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SHARPEN: Sharpened Channels 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RESAMPLING: Resampling 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) nearest neighbour 		- 1: (1) bilinear 		- 2: (2) cubic convolution 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' PAN_MATCH: Panchromatic Channel Matching 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) normalized 		- 1: (1) standardized 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item B_SHARP - outputRaster - Blue
\item G_SHARP - outputRaster - Green
\item R_SHARP - outputRaster - Red
\item SHARPEN - outputRaster - Sharpened Channels
}
}
}
