% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_imcorrfeaturetracking.R
\name{sagang_imcorrfeaturetracking}
\alias{sagang_imcorrfeaturetracking}
\title{QGIS algorithm - Imcorr - feature tracking}
\usage{
sagang_imcorrfeaturetracking(
  GRID_1 = qgisprocess:::qgis_default_value(),
  GRID_2 = qgisprocess:::qgis_default_value(),
  DTM_1 = qgisprocess:::qgis_default_value(),
  DTM_2 = qgisprocess:::qgis_default_value(),
  CORRPOINTS = qgisprocess:::qgis_default_value(),
  CORRLINES = qgisprocess:::qgis_default_value(),
  SEARCH_CHIPSIZE = qgisprocess:::qgis_default_value(),
  REF_CHIPSIZE = qgisprocess:::qgis_default_value(),
  GRID_SPACING = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID_1}{\code{raster} - Grid 1. Path to a raster layer.}

\item{GRID_2}{\code{raster} - Grid 2. Path to a raster layer.}

\item{DTM_1}{\code{raster} - DTM 1. Path to a raster layer.}

\item{DTM_2}{\code{raster} - DTM 2. Path to a raster layer.}

\item{CORRPOINTS}{\code{vectorDestination} - Correlated Points. Path for new vector layer.}

\item{CORRLINES}{\code{vectorDestination} - Displacement Vector. Path for new vector layer.}

\item{SEARCH_CHIPSIZE}{\code{enum}  of \verb{("(0) 16x16", "(1) 32x32", "(2) 64x64", "(3) 128x128", "(4) 256x256")} - Search Chip Size (Cells). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{REF_CHIPSIZE}{\code{enum}  of \verb{("(0) 16x16", "(1) 32x32", "(2) 64x64", "(3) 128x128")} - Reference Chip Size (Cells). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRID_SPACING}{\code{number} - Grid Spacing (Map Units). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Imcorr - feature tracking (sagang:imcorrfeaturetracking). ---------------- Arguments ----------------  GRID_1: Grid 1 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GRID_2: Grid 2 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DTM_1: DTM 1 (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DTM_2: DTM 2 (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CORRPOINTS: Correlated Points 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer CORRLINES: Displacement Vector 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer SEARCH_CHIPSIZE: Search Chip Size (Cells) 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) 16x16 		- 1: (1) 32x32 		- 2: (2) 64x64 		- 3: (3) 128x128 		- 4: (4) 256x256 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' REF_CHIPSIZE: Reference Chip Size (Cells) 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) 16x16 		- 1: (1) 32x32 		- 2: (2) 64x64 		- 3: (3) 128x128 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' GRID_SPACING: Grid Spacing (Map Units) 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CORRLINES - outputVector - Displacement Vector
\item CORRPOINTS - outputVector - Correlated Points
}
}
}
