% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_importatlasboundaryfile.R
\name{sagang_importatlasboundaryfile}
\alias{sagang_importatlasboundaryfile}
\title{QGIS algorithm - Import atlas boundary file}
\usage{
sagang_importatlasboundaryfile(
  FILE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FILE}{\code{file} - File. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Import atlas boundary file (sagang:importatlasboundaryfile). ---------------- Arguments ----------------  FILE: File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
