% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_importbuildingsketchesfromcitygml.R
\name{sagang_importbuildingsketchesfromcitygml}
\alias{sagang_importbuildingsketchesfromcitygml}
\title{QGIS algorithm - Import building sketches from citygml}
\usage{
sagang_importbuildingsketchesfromcitygml(
  BUILDINGS = qgisprocess:::qgis_default_value(),
  FILES = qgisprocess:::qgis_default_value(),
  PARTS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{BUILDINGS}{\code{vectorDestination} - Buildings. Path for new vector layer.}

\item{FILES}{\code{file} - Files. Path to a file.}

\item{PARTS}{\code{boolean} - Check for Building Parts. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Import building sketches from citygml (sagang:importbuildingsketchesfromcitygml). ---------------- Arguments ----------------  BUILDINGS: Buildings 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer FILES: Files 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file PARTS: Check for Building Parts 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item BUILDINGS - outputVector - Buildings
}
}
}
