% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_importdxffiles.R
\name{sagang_importdxffiles}
\alias{sagang_importdxffiles}
\title{QGIS algorithm - Import dxf files}
\usage{
sagang_importdxffiles(
  SHAPES = qgisprocess:::qgis_default_value(),
  TABLES = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  FILTER = qgisprocess:::qgis_default_value(),
  DCIRCLE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{vectorDestination} - Shapes. Path for new vector layer.}

\item{TABLES}{\code{vectorDestination} - Tables. Path for new vector layer.}

\item{FILE}{\code{file} - File. Path to a file.}

\item{FILTER}{\code{enum}  of \verb{("(0) all entities", "(1) only entities with layer definition", "(2) only entities without layer definition")} - Import Filter. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DCIRCLE}{\code{number} - Circle Point Distance (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Import dxf files (sagang:importdxffiles). ---------------- Arguments ----------------  SHAPES: Shapes 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer TABLES: Tables 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer FILE: File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file FILTER: Import Filter 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) all entities 		- 1: (1) only entities with layer definition 		- 2: (2) only entities without layer definition 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DCIRCLE: Circle Point Distance (Degree) 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item SHAPES - outputVector - Shapes
\item TABLES - outputVector - Tables
}
}
}
