% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_importsentinel2scene.R
\name{sagang_importsentinel2scene}
\alias{sagang_importsentinel2scene}
\title{QGIS algorithm - Import sentinel-2 scene}
\usage{
sagang_importsentinel2scene(
  METAFILE = qgisprocess:::qgis_default_value(),
  BANDS = qgisprocess:::qgis_default_value(),
  MULTI2GRIDS = qgisprocess:::qgis_default_value(),
  SKIP_TCI = qgisprocess:::qgis_default_value(),
  SKIP_60M = qgisprocess:::qgis_default_value(),
  REFLECTANCE = qgisprocess:::qgis_default_value(),
  PROJECTION = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{METAFILE}{\code{file} - Metadata File. Path to a file.}

\item{BANDS}{\code{rasterDestination} - Bands. Path for new raster layer.}

\item{MULTI2GRIDS}{\code{boolean} - Multispectral Bands as Grid Collection. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SKIP_TCI}{\code{boolean} - Skip True Color Image. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SKIP_60M}{\code{boolean} - Skip Aerosol, Vapour, Cirrus. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{REFLECTANCE}{\code{enum}  of \verb{("(0) original", "(1) fraction")} - Reflectance Values. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PROJECTION}{\code{enum}  of \verb{("(0) UTM North", "(1) UTM South", "(2) Geographic Coordinates")} - Coordinate System. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESAMPLING}{\code{enum}  of \verb{("(0) Nearest Neighbour", "(1) Bilinear Interpolation", "(2) Bicubic Spline Interpolation", "(3) B-Spline Interpolation")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Import sentinel-2 scene (sagang:importsentinel2scene). ---------------- Arguments ----------------  METAFILE: Metadata File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file BANDS: Bands 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MULTI2GRIDS: Multispectral Bands as Grid Collection 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SKIP_TCI: Skip True Color Image 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SKIP_60M: Skip Aerosol, Vapour, Cirrus 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression REFLECTANCE: Reflectance Values 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) original 		- 1: (1) fraction 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' PROJECTION: Coordinate System 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) UTM North 		- 1: (1) UTM South 		- 2: (2) Geographic Coordinates 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RESAMPLING: Resampling 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) Nearest Neighbour 		- 1: (1) Bilinear Interpolation 		- 2: (2) Bicubic Spline Interpolation 		- 3: (3) B-Spline Interpolation 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item BANDS - outputRaster - Bands
}
}
}
