% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_importsimplefeaturesfromwellknowntext.R
\name{sagang_importsimplefeaturesfromwellknowntext}
\alias{sagang_importsimplefeaturesfromwellknowntext}
\title{QGIS algorithm - Import simple features from well known text}
\usage{
sagang_importsimplefeaturesfromwellknowntext(
  SHAPES = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  WKT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{vectorDestination} - WKT Import. Path for new vector layer.}

\item{FILE}{\code{file} - File. Path to a file.}

\item{WKT}{\code{string} - WKT String. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Import simple features from well known text (sagang:importsimplefeaturesfromwellknowntext). ---------------- Arguments ----------------  SHAPES: WKT Import 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer FILE: File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file WKT: WKT String 	Default value:	 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item SHAPES - outputVector - WKT Import
}
}
}
