% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_importusgssrtmgrid.R
\name{sagang_importusgssrtmgrid}
\alias{sagang_importusgssrtmgrid}
\title{QGIS algorithm - Import usgs srtm grid}
\usage{
sagang_importusgssrtmgrid(
  GRIDS = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  RESOLUTION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{rasterDestination} - Grids. Path for new raster layer.}

\item{FILE}{\code{file} - Files. Path to a file.}

\item{RESOLUTION}{\code{enum}  of \verb{("(0) 1 arc-second", "(1) 3 arc-second")} - Resolution. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Import usgs srtm grid (sagang:importusgssrtmgrid). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FILE: Files 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file RESOLUTION: Resolution 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) 1 arc-second 		- 1: (1) 3 arc-second 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRIDS - outputRaster - Grids
}
}
}
