% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_importwaspterrainmapfile.R
\name{sagang_importwaspterrainmapfile}
\alias{sagang_importwaspterrainmapfile}
\title{QGIS algorithm - Import wasp terrain map file}
\usage{
sagang_importwaspterrainmapfile(
  SHAPES = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{vectorDestination} - Contour Lines. Path for new vector layer.}

\item{FILE}{\code{file} - File Name. Path to a file.}

\item{METHOD}{\code{enum}  of \verb{("(0) elevation", "(1) roughness", "(2) elevation and roughness")} - Input Specification. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Import wasp terrain map file (sagang:importwaspterrainmapfile). ---------------- Arguments ----------------  SHAPES: Contour Lines 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer FILE: File Name 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file METHOD: Input Specification 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) elevation 		- 1: (1) roughness 		- 2: (2) elevation and roughness 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item SHAPES - outputVector - Contour Lines
}
}
}
