% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_importwrfgeogridbinaryformat.R
\name{sagang_importwrfgeogridbinaryformat}
\alias{sagang_importwrfgeogridbinaryformat}
\title{QGIS algorithm - Import wrf geogrid binary format}
\usage{
sagang_importwrfgeogridbinaryformat(
  GRIDS = qgisprocess:::qgis_default_value(),
  FILE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{rasterDestination} - Grids. Path for new raster layer.}

\item{FILE}{\code{file} - File. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Import wrf geogrid binary format (sagang:importwrfgeogridbinaryformat). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FILE: File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRIDS - outputRaster - Grids
}
}
}
