% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_inverseprincipalcomponentsrotation.R
\name{sagang_inverseprincipalcomponentsrotation}
\alias{sagang_inverseprincipalcomponentsrotation}
\title{QGIS algorithm - Inverse principal components rotation}
\usage{
sagang_inverseprincipalcomponentsrotation(
  PCA = qgisprocess:::qgis_default_value(),
  EIGEN = qgisprocess:::qgis_default_value(),
  GRIDS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{PCA}{\code{multilayer} - Principal Components. .}

\item{EIGEN}{\code{source} - Eigen Vectors. Path to a vector layer.}

\item{GRIDS}{\code{rasterDestination} - Grids. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Inverse principal components rotation (sagang:inverseprincipalcomponentsrotation). ---------------- Arguments ----------------  PCA: Principal Components 	Argument type:	multilayer EIGEN: Eigen Vectors 	Argument type:	source 	Acceptable values: 		- Path to a vector layer GRIDS: Grids 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRIDS - outputRaster - Grids
}
}
}
