% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_invertselectionofshapeslayer.R
\name{sagang_invertselectionofshapeslayer}
\alias{sagang_invertselectionofshapeslayer}
\title{QGIS algorithm - Invert selection of shapes layer}
\usage{
sagang_invertselectionofshapeslayer(
  INPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input. Path to a vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Invert selection of shapes layer (sagang:invertselectionofshapeslayer). ---------------- Arguments ----------------  INPUT: Input 	Argument type:	source 	Acceptable values: 		- Path to a vector layer
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
