% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_isochronesvariablespeed.R
\name{sagang_isochronesvariablespeed}
\alias{sagang_isochronesvariablespeed}
\title{QGIS algorithm - Isochrones variable speed}
\usage{
sagang_isochronesvariablespeed(
  DEM = qgisprocess:::qgis_default_value(),
  SLOPE = qgisprocess:::qgis_default_value(),
  FLOWACC = qgisprocess:::qgis_default_value(),
  CN = qgisprocess:::qgis_default_value(),
  MANNING = qgisprocess:::qgis_default_value(),
  TIME = qgisprocess:::qgis_default_value(),
  SPEED = qgisprocess:::qgis_default_value(),
  AVGMANNING = qgisprocess:::qgis_default_value(),
  AVGCN = qgisprocess:::qgis_default_value(),
  THRSMIXED = qgisprocess:::qgis_default_value(),
  THRSCHANNEL = qgisprocess:::qgis_default_value(),
  AVGRAINFALL = qgisprocess:::qgis_default_value(),
  CHANSLOPE = qgisprocess:::qgis_default_value(),
  MINSPEED = qgisprocess:::qgis_default_value(),
  TARGET_PT_X = qgisprocess:::qgis_default_value(),
  TARGET_PT_Y = qgisprocess:::qgis_default_value(),
  TARGET_PT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{SLOPE}{\code{raster} - Slope. Path to a raster layer.}

\item{FLOWACC}{\code{raster} - Catchment Area. Path to a raster layer.}

\item{CN}{\code{raster} - Curve Number. Path to a raster layer.}

\item{MANNING}{\code{raster} - Manning's N. Path to a raster layer.}

\item{TIME}{\code{rasterDestination} - Time Out(h). Path for new raster layer.}

\item{SPEED}{\code{rasterDestination} - Speed (m\code{/}s). Path for new raster layer.}

\item{AVGMANNING}{\code{number} - Avg. Manning's N. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{AVGCN}{\code{number} - Avg. Curve Number. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRSMIXED}{\code{number} - Mixed Flow Threshold (ha). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRSCHANNEL}{\code{number} - Channel Definition Threshold (ha). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{AVGRAINFALL}{\code{number} - Avg. Rainfall Intensity (mm\code{/}h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CHANSLOPE}{\code{number} - Channel side slope(m\code{/}m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MINSPEED}{\code{number} - Min. Flow Speed (m\code{/}s). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TARGET_PT_X}{\code{number} - Target X Coordinate. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TARGET_PT_Y}{\code{number} - Target Y Coordinate. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TARGET_PT}{\code{source} - Target Point. Path to a vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Isochrones variable speed (sagang:isochronesvariablespeed). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SLOPE: Slope 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FLOWACC: Catchment Area 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CN: Curve Number (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MANNING: Manning's N (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TIME: Time Out(h) 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SPEED: Speed (m/s) 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer AVGMANNING: Avg. Manning's N 	Default value:	0.15 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression AVGCN: Avg. Curve Number 	Default value:	75 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRSMIXED: Mixed Flow Threshold (ha) 	Default value:	18 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRSCHANNEL: Channel Definition Threshold (ha) 	Default value:	360 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression AVGRAINFALL: Avg. Rainfall Intensity (mm/h) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CHANSLOPE: Channel side slope(m/m) 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MINSPEED: Min. Flow Speed (m/s) 	Default value:	0.05 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TARGET_PT_X: Target X Coordinate 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TARGET_PT_Y: Target Y Coordinate 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TARGET_PT: Target Point (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item SPEED - outputRaster - Speed (m\code{/}s)
\item TIME - outputRaster - Time Out(h)
}
}
}
