% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_isodataclusteringforgrids.R
\name{sagang_isodataclusteringforgrids}
\alias{sagang_isodataclusteringforgrids}
\title{QGIS algorithm - Isodata clustering for grids}
\usage{
sagang_isodataclusteringforgrids(
  FEATURES = qgisprocess:::qgis_default_value(),
  CLUSTER = qgisprocess:::qgis_default_value(),
  STATISTICS = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  ITERATIONS = qgisprocess:::qgis_default_value(),
  CLUSTER_INI = qgisprocess:::qgis_default_value(),
  CLUSTER_MAX = qgisprocess:::qgis_default_value(),
  SAMPLES_MIN = qgisprocess:::qgis_default_value(),
  INITIALIZE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FEATURES}{\code{multilayer} - Features. .}

\item{CLUSTER}{\code{rasterDestination} - Clusters. Path for new raster layer.}

\item{STATISTICS}{\code{vectorDestination} - Statistics. Path for new vector layer.}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ITERATIONS}{\code{number} - Maximum Number of Iterations. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CLUSTER_INI}{\code{number} - Initial Number of Clusters. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CLUSTER_MAX}{\code{number} - Maximum Number of Clusters. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SAMPLES_MIN}{\code{number} - Minimum Number of Samples in a Cluster. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INITIALIZE}{\code{enum}  of \verb{("(0) random", "(1) periodical", "(2) keep values")} - Start Partition. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Isodata clustering for grids (sagang:isodataclusteringforgrids). ---------------- Arguments ----------------  FEATURES: Features 	Argument type:	multilayer CLUSTER: Clusters 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STATISTICS: Statistics 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer NORMALIZE: Normalize 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ITERATIONS: Maximum Number of Iterations 	Default value:	20 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CLUSTER_INI: Initial Number of Clusters 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CLUSTER_MAX: Maximum Number of Clusters 	Default value:	16 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SAMPLES_MIN: Minimum Number of Samples in a Cluster 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression INITIALIZE: Start Partition 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) random 		- 1: (1) periodical 		- 2: (2) keep values 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLUSTER - outputRaster - Clusters
\item STATISTICS - outputVector - Statistics
}
}
}
