% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_kmeansclusteringforgrids.R
\name{sagang_kmeansclusteringforgrids}
\alias{sagang_kmeansclusteringforgrids}
\title{QGIS algorithm - K-means clustering for grids}
\usage{
sagang_kmeansclusteringforgrids(
  GRIDS = qgisprocess:::qgis_default_value(),
  CLUSTER = qgisprocess:::qgis_default_value(),
  STATISTICS = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  NCLUSTER = qgisprocess:::qgis_default_value(),
  MAXITER = qgisprocess:::qgis_default_value(),
  NORMALISE = qgisprocess:::qgis_default_value(),
  INITIALIZE = qgisprocess:::qgis_default_value(),
  OLDVERSION = qgisprocess:::qgis_default_value(),
  UPDATEVIEW = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{CLUSTER}{\code{rasterDestination} - Clusters. Path for new raster layer.}

\item{STATISTICS}{\code{vectorDestination} - Statistics. Path for new vector layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) Iterative Minimum Distance (Forgy 1965)", "(1) Hill-Climbing (Rubin 1967)", "(2) Combined Minimum Distance / Hillclimbing")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NCLUSTER}{\code{number} - Clusters. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAXITER}{\code{number} - Maximum Iterations. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NORMALISE}{\code{boolean} - Normalise. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INITIALIZE}{\code{enum}  of \verb{("(0) random", "(1) periodical", "(2) keep values")} - Start Partition. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OLDVERSION}{\code{boolean} - Old Version. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{UPDATEVIEW}{\code{boolean} - Update View. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen K-means clustering for grids (sagang:kmeansclusteringforgrids). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	multilayer CLUSTER: Clusters 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STATISTICS: Statistics 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer METHOD: Method 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Iterative Minimum Distance (Forgy 1965) 		- 1: (1) Hill-Climbing (Rubin 1967) 		- 2: (2) Combined Minimum Distance / Hillclimbing 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NCLUSTER: Clusters 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAXITER: Maximum Iterations 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NORMALISE: Normalise 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression INITIALIZE: Start Partition 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) random 		- 1: (1) periodical 		- 2: (2) keep values 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' OLDVERSION: Old Version 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression UPDATEVIEW: Update View 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLUSTER - outputRaster - Clusters
\item STATISTICS - outputVector - Statistics
}
}
}
