% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_lakeflood.R
\name{sagang_lakeflood}
\alias{sagang_lakeflood}
\title{QGIS algorithm - Lake flood}
\usage{
sagang_lakeflood(
  ELEV = qgisprocess:::qgis_default_value(),
  SEEDS = qgisprocess:::qgis_default_value(),
  LEVEL = qgisprocess:::qgis_default_value(),
  OUTDEPTH = qgisprocess:::qgis_default_value(),
  OUTLEVEL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEV}{\code{raster} - DEM. Path to a raster layer.}

\item{SEEDS}{\code{raster} - Seeds. Path to a raster layer.}

\item{LEVEL}{\code{boolean} - Absolute Water Levels. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTDEPTH}{\code{rasterDestination} - Lake. Path for new raster layer.}

\item{OUTLEVEL}{\code{rasterDestination} - Surface. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Lake flood (sagang:lakeflood). ---------------- Arguments ----------------  ELEV: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SEEDS: Seeds 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LEVEL: Absolute Water Levels 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OUTDEPTH: Lake 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer OUTLEVEL: Surface 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTDEPTH - outputRaster - Lake
\item OUTLEVEL - outputRaster - Surface
}
}
}
