% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_landcoverscenariooffset.R
\name{sagang_landcoverscenariooffset}
\alias{sagang_landcoverscenariooffset}
\title{QGIS algorithm - Land cover scenario offset}
\usage{
sagang_landcoverscenariooffset(
  ROAD_LINES = qgisprocess:::qgis_default_value(),
  ROAD_CLASS = qgisprocess:::qgis_default_value(),
  CHANNEL_LINES = qgisprocess:::qgis_default_value(),
  CHANNEL_ORDER = qgisprocess:::qgis_default_value(),
  DEM = qgisprocess:::qgis_default_value(),
  VEGETATION = qgisprocess:::qgis_default_value(),
  LANDCOVER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ROAD_LINES}{\code{source} - Roads. Path to a vector layer.}

\item{ROAD_CLASS}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{CHANNEL_LINES}{\code{source} - Channel Network. Path to a vector layer.}

\item{CHANNEL_ORDER}{\code{field} - Stream Order. The name of an existing field. ; delimited list of existing field names.}

\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{VEGETATION}{\code{raster} - Vegetation. Path to a raster layer.}

\item{LANDCOVER}{\code{rasterDestination} - Land Cover. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Land cover scenario offset (sagang:landcoverscenariooffset). ---------------- Arguments ----------------  ROAD_LINES: Roads 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ROAD_CLASS: Attribute 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names CHANNEL_LINES: Channel Network (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer CHANNEL_ORDER: Stream Order 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer VEGETATION: Vegetation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LANDCOVER: Land Cover 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item LANDCOVER - outputRaster - Land Cover
}
}
}
