% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_landsatimportwithoptions.R
\name{sagang_landsatimportwithoptions}
\alias{sagang_landsatimportwithoptions}
\title{QGIS algorithm - Landsat import with options}
\usage{
sagang_landsatimportwithoptions(
  FILES = qgisprocess:::qgis_default_value(),
  BANDS = qgisprocess:::qgis_default_value(),
  PROJECTION = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  SHOW_RGB = qgisprocess:::qgis_default_value(),
  SHOW_R = qgisprocess:::qgis_default_value(),
  SHOW_G = qgisprocess:::qgis_default_value(),
  SHOW_B = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FILES}{\code{file} - Files. Path to a file.}

\item{BANDS}{\code{rasterDestination} - Bands. Path for new raster layer.}

\item{PROJECTION}{\code{enum}  of \verb{("(0) UTM North", "(1) UTM South", "(2) Geographic Coordinates")} - Coordinate System. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESAMPLING}{\code{enum}  of \verb{("(0) Nearest Neighbour", "(1) Bilinear Interpolation", "(2) Bicubic Spline Interpolation", "(3) B-Spline Interpolation")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SHOW_RGB}{\code{boolean} - Show a Composite. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SHOW_R}{\code{enum}  of \code{("(0) no choice available")} - Red. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SHOW_G}{\code{enum}  of \code{("(0) no choice available")} - Green. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SHOW_B}{\code{enum}  of \code{("(0) no choice available")} - Blue. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Landsat import with options (sagang:landsatimportwithoptions). ---------------- Arguments ----------------  FILES: Files 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file BANDS: Bands 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PROJECTION: Coordinate System 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) UTM North 		- 1: (1) UTM South 		- 2: (2) Geographic Coordinates 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RESAMPLING: Resampling 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) Nearest Neighbour 		- 1: (1) Bilinear Interpolation 		- 2: (2) Bicubic Spline Interpolation 		- 3: (3) B-Spline Interpolation 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SHOW_RGB: Show a Composite 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SHOW_R: Red 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) no choice available 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SHOW_G: Green 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) no choice available 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SHOW_B: Blue 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) no choice available 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item BANDS - outputRaster - Bands
}
}
}
