% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_landsurfacetemperature.R
\name{sagang_landsurfacetemperature}
\alias{sagang_landsurfacetemperature}
\title{QGIS algorithm - Land surface temperature}
\usage{
sagang_landsurfacetemperature(
  IRRADIANCE = qgisprocess:::qgis_default_value(),
  IRRADIANCE_DEFAULT = qgisprocess:::qgis_default_value(),
  ALBEDO = qgisprocess:::qgis_default_value(),
  ALBEDO_DEFAULT = qgisprocess:::qgis_default_value(),
  EMISSIVITY = qgisprocess:::qgis_default_value(),
  EMISSIVITY_DEFAULT = qgisprocess:::qgis_default_value(),
  CONVECTION = qgisprocess:::qgis_default_value(),
  CONVECTION_DEFAULT = qgisprocess:::qgis_default_value(),
  T_AIR = qgisprocess:::qgis_default_value(),
  T_AIR_DEFAULT = qgisprocess:::qgis_default_value(),
  T_SKY = qgisprocess:::qgis_default_value(),
  T_SKY_DEFAULT = qgisprocess:::qgis_default_value(),
  T_INITIAL = qgisprocess:::qgis_default_value(),
  T_INITIAL_DEFAULT = qgisprocess:::qgis_default_value(),
  LST = qgisprocess:::qgis_default_value(),
  UNIT = qgisprocess:::qgis_default_value(),
  ITERATIONS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{IRRADIANCE}{\code{raster} - Global Irradiance. Path to a raster layer.}

\item{IRRADIANCE_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ALBEDO}{\code{raster} - Albedo. Path to a raster layer.}

\item{ALBEDO_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EMISSIVITY}{\code{raster} - Emissivity. Path to a raster layer.}

\item{EMISSIVITY_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CONVECTION}{\code{raster} - Convection Coefficient. Path to a raster layer.}

\item{CONVECTION_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T_AIR}{\code{raster} - Ambient Air Temperature. Path to a raster layer.}

\item{T_AIR_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T_SKY}{\code{raster} - Radiant Sky Temperature. Path to a raster layer.}

\item{T_SKY_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T_INITIAL}{\code{raster} - Initial Temperature Estimation. Path to a raster layer.}

\item{T_INITIAL_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LST}{\code{rasterDestination} - Land Surface Temperature. Path for new raster layer.}

\item{UNIT}{\code{enum}  of \verb{("(0) Kelvin", "(1) Celsius")} - Unit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ITERATIONS}{\code{number} - Iterations. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Land surface temperature (sagang:landsurfacetemperature). ---------------- Arguments ----------------  IRRADIANCE: Global Irradiance (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer IRRADIANCE_DEFAULT: Default 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ALBEDO: Albedo (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer ALBEDO_DEFAULT: Default 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EMISSIVITY: Emissivity (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer EMISSIVITY_DEFAULT: Default 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CONVECTION: Convection Coefficient (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CONVECTION_DEFAULT: Default 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T_AIR: Ambient Air Temperature (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer T_AIR_DEFAULT: Default 	Default value:	300 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T_SKY: Radiant Sky Temperature (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer T_SKY_DEFAULT: Default 	Default value:	280 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T_INITIAL: Initial Temperature Estimation (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer T_INITIAL_DEFAULT: Default 	Default value:	300 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LST: Land Surface Temperature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer UNIT: Unit 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Kelvin 		- 1: (1) Celsius 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' ITERATIONS: Iterations 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item LST - outputRaster - Land Surface Temperature
}
}
}
