% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_landsurfacetemperaturelapserates.R
\name{sagang_landsurfacetemperaturelapserates}
\alias{sagang_landsurfacetemperaturelapserates}
\title{QGIS algorithm - Land surface temperature (lapse rates)}
\usage{
sagang_landsurfacetemperaturelapserates(
  DEM = qgisprocess:::qgis_default_value(),
  SWR = qgisprocess:::qgis_default_value(),
  LAI = qgisprocess:::qgis_default_value(),
  LAI_MAX = qgisprocess:::qgis_default_value(),
  LST = qgisprocess:::qgis_default_value(),
  Z_REFERENCE = qgisprocess:::qgis_default_value(),
  T_REFERENCE = qgisprocess:::qgis_default_value(),
  T_GRADIENT = qgisprocess:::qgis_default_value(),
  C_FACTOR = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{SWR}{\code{raster} - Short Wave Radiation Ratio. Path to a raster layer.}

\item{LAI}{\code{raster} - Leaf Area Index. Path to a raster layer.}

\item{LAI_MAX}{\code{number} - Maximum LAI. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LST}{\code{rasterDestination} - Land Surface Temperature. Path for new raster layer.}

\item{Z_REFERENCE}{\code{number} - Elevation. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T_REFERENCE}{\code{number} - Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T_GRADIENT}{\code{number} - Temperature Lapse Rate. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{C_FACTOR}{\code{number} - C Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Land surface temperature (lapse rates) (sagang:landsurfacetemperaturelapserates). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SWR: Short Wave Radiation Ratio (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LAI: Leaf Area Index (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LAI_MAX: Maximum LAI 	Default value:	8 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LST: Land Surface Temperature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer Z_REFERENCE: Elevation 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T_REFERENCE: Temperature 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T_GRADIENT: Temperature Lapse Rate 	Default value:	0.65 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression C_FACTOR: C Factor 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item LST - outputRaster - Land Surface Temperature
}
}
}
