% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_largestcirclesinpolygons.R
\name{sagang_largestcirclesinpolygons}
\alias{sagang_largestcirclesinpolygons}
\title{QGIS algorithm - Largest circles in polygons}
\usage{
sagang_largestcirclesinpolygons(
  POLYGONS = qgisprocess:::qgis_default_value(),
  CIRCLES = qgisprocess:::qgis_default_value(),
  RESOLUTION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{CIRCLES}{\code{vectorDestination} - Circles. Path for new vector layer.}

\item{RESOLUTION}{\code{number} - Resolution. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Largest circles in polygons (sagang:largestcirclesinpolygons). ---------------- Arguments ----------------  POLYGONS: Polygons 	Argument type:	source 	Acceptable values: 		- Path to a vector layer CIRCLES: Circles 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer RESOLUTION: Resolution 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CIRCLES - outputVector - Circles
}
}
}
