% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_latitudelongitudegraticule.R
\name{sagang_latitudelongitudegraticule}
\alias{sagang_latitudelongitudegraticule}
\title{QGIS algorithm - Latitude/longitude graticule}
\usage{
sagang_latitudelongitudegraticule(
  CRS_METHOD = qgisprocess:::qgis_default_value(),
  CRS_PROJ4 = qgisprocess:::qgis_default_value(),
  CRS_FILE = qgisprocess:::qgis_default_value(),
  CRS_EPSG = qgisprocess:::qgis_default_value(),
  CRS_EPSG_AUTH = qgisprocess:::qgis_default_value(),
  GRATICULE = qgisprocess:::qgis_default_value(),
  COORDS = qgisprocess:::qgis_default_value(),
  XMIN = qgisprocess:::qgis_default_value(),
  XMAX = qgisprocess:::qgis_default_value(),
  YMIN = qgisprocess:::qgis_default_value(),
  YMAX = qgisprocess:::qgis_default_value(),
  INTERVAL = qgisprocess:::qgis_default_value(),
  FIXED = qgisprocess:::qgis_default_value(),
  FITTED = qgisprocess:::qgis_default_value(),
  RESOLUTION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CRS_METHOD}{\code{enum}  of \verb{("(0) Proj4 Parameters", "(1) EPSG Code", "(2) Well Known Text File")} - Get CRS Definition from.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CRS_PROJ4}{\code{string} - Proj4 Parameters. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CRS_FILE}{\code{file} - Well Known Text File. Path to a file.}

\item{CRS_EPSG}{\code{number} - Authority Code. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CRS_EPSG_AUTH}{\code{string} - Authority. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRATICULE}{\code{vectorDestination} - Graticule. Path for new vector layer.}

\item{COORDS}{\code{vectorDestination} - Frame Coordinates. Path for new vector layer.}

\item{XMIN}{\code{number} - Minimum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{XMAX}{\code{number} - Maximum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YMIN}{\code{number} - Minimum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YMAX}{\code{number} - Maximum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INTERVAL}{\code{enum}  of \verb{("(0) fixed interval", "(1) fitted interval")} - Interval. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FIXED}{\code{number} - Fixed Interval (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FITTED}{\code{number} - Number of Intervals. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RESOLUTION}{\code{number} - Minimum Resolution (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Latitude/longitude graticule (sagang:latitudelongitudegraticule). ---------------- Arguments ----------------  CRS_METHOD: Get CRS Definition from... 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Proj4 Parameters 		- 1: (1) EPSG Code 		- 2: (2) Well Known Text File 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CRS_PROJ4: Proj4 Parameters 	Default value:	 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CRS_FILE: Well Known Text File 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file CRS_EPSG: Authority Code 	Default value:	-1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CRS_EPSG_AUTH: Authority 	Default value:	EPSG 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GRATICULE: Graticule 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer COORDS: Frame Coordinates 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer XMIN: Minimum 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression XMAX: Maximum 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YMIN: Minimum 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YMAX: Maximum 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression INTERVAL: Interval 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) fixed interval 		- 1: (1) fitted interval 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FIXED: Fixed Interval (Degree) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FITTED: Number of Intervals 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RESOLUTION: Minimum Resolution (Degree) 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item COORDS - outputVector - Frame Coordinates
\item GRATICULE - outputVector - Graticule
}
}
}
