% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_layerofextremevalue.R
\name{sagang_layerofextremevalue}
\alias{sagang_layerofextremevalue}
\title{QGIS algorithm - Layer of extreme value}
\usage{
sagang_layerofextremevalue(
  GRIDS = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  CRITERIA = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{RESULT}{\code{rasterDestination} - Result. Path for new raster layer.}

\item{CRITERIA}{\code{enum}  of \verb{("(0) Maximum", "(1) Minimum")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Layer of extreme value (sagang:layerofextremevalue). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	multilayer RESULT: Result 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CRITERIA: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Maximum 		- 1: (1) Minimum 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Result
}
}
}
