% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_leastcostpaths.R
\name{sagang_leastcostpaths}
\alias{sagang_leastcostpaths}
\title{QGIS algorithm - Least cost paths}
\usage{
sagang_leastcostpaths(
  SOURCE = qgisprocess:::qgis_default_value(),
  DEM = qgisprocess:::qgis_default_value(),
  VALUES = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  LINE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SOURCE}{\code{source} - Source Point(s). Path to a vector layer.}

\item{DEM}{\code{raster} - Accumulated Cost Surface. Path to a raster layer.}

\item{VALUES}{\code{multilayer} - Values. .}

\item{POINTS}{\code{vectorDestination} - Profile Points. Path for new vector layer.}

\item{LINE}{\code{vectorDestination} - Profile Lines. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Least cost paths (sagang:leastcostpaths). ---------------- Arguments ----------------  SOURCE: Source Point(s) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer DEM: Accumulated Cost Surface 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer VALUES: Values (optional) 	Argument type:	multilayer POINTS: Profile Points 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer LINE: Profile Lines 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item LINE - outputVector - Profile Lines
\item POINTS - outputVector - Profile Points
}
}
}
