% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_linecrossings.R
\name{sagang_linecrossings}
\alias{sagang_linecrossings}
\title{QGIS algorithm - Line crossings}
\usage{
sagang_linecrossings(
  LINES_A = qgisprocess:::qgis_default_value(),
  LINES_B = qgisprocess:::qgis_default_value(),
  CROSSINGS = qgisprocess:::qgis_default_value(),
  ATTRIBUTES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LINES_A}{\code{source} - 1st Lines Layer. Path to a vector layer.}

\item{LINES_B}{\code{source} - 2nd Lines Layer. Path to a vector layer.}

\item{CROSSINGS}{\code{vectorDestination} - Crossings. Path for new vector layer.}

\item{ATTRIBUTES}{\code{enum}  of \verb{("(0) index", "(1) attributes", "(2) index and attributes")} - Parent Attributes. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Line crossings (sagang:linecrossings). ---------------- Arguments ----------------  LINES_A: 1st Lines Layer 	Argument type:	source 	Acceptable values: 		- Path to a vector layer LINES_B: 2nd Lines Layer 	Argument type:	source 	Acceptable values: 		- Path to a vector layer CROSSINGS: Crossings 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer ATTRIBUTES: Parent Attributes 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) index 		- 1: (1) attributes 		- 2: (2) index and attributes 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item CROSSINGS - outputVector - Crossings
}
}
}
