% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_linepolygonintersection.R
\name{sagang_linepolygonintersection}
\alias{sagang_linepolygonintersection}
\title{QGIS algorithm - Line-polygon intersection}
\usage{
sagang_linepolygonintersection(
  LINES = qgisprocess:::qgis_default_value(),
  POLYGONS = qgisprocess:::qgis_default_value(),
  INTERSECT = qgisprocess:::qgis_default_value(),
  DIFFERENCE = qgisprocess:::qgis_default_value(),
  ATTRIBUTES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{INTERSECT}{\code{vectorDestination} - Intersection. Path for new vector layer.}

\item{DIFFERENCE}{\code{vectorDestination} - Difference. Path for new vector layer.}

\item{ATTRIBUTES}{\code{enum}  of \verb{("(0) polygon", "(1) line", "(2) line and polygon")} - Attributes. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Line-polygon intersection (sagang:linepolygonintersection). ---------------- Arguments ----------------  LINES: Lines 	Argument type:	source 	Acceptable values: 		- Path to a vector layer POLYGONS: Polygons 	Argument type:	source 	Acceptable values: 		- Path to a vector layer INTERSECT: Intersection 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DIFFERENCE: Difference 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer ATTRIBUTES: Attributes 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) polygon 		- 1: (1) line 		- 2: (2) line and polygon 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIFFERENCE - outputVector - Difference
\item INTERSECT - outputVector - Intersection
}
}
}
