% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_localminimaandmaxima.R
\name{sagang_localminimaandmaxima}
\alias{sagang_localminimaandmaxima}
\title{QGIS algorithm - Local minima and maxima}
\usage{
sagang_localminimaandmaxima(
  GRID = qgisprocess:::qgis_default_value(),
  MINIMA = qgisprocess:::qgis_default_value(),
  MAXIMA = qgisprocess:::qgis_default_value(),
  IDENTITY = qgisprocess:::qgis_default_value(),
  ABSOLUTE = qgisprocess:::qgis_default_value(),
  BOUNDARY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{MINIMA}{\code{vectorDestination} - Minima. Path for new vector layer.}

\item{MAXIMA}{\code{vectorDestination} - Maxima. Path for new vector layer.}

\item{IDENTITY}{\code{boolean} - Identical Values. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ABSOLUTE}{\code{boolean} - Absolute. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{BOUNDARY}{\code{boolean} - Boundary Cells. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Local minima and maxima (sagang:localminimaandmaxima). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MINIMA: Minima 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer MAXIMA: Maxima 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer IDENTITY: Identical Values 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ABSOLUTE: Absolute 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression BOUNDARY: Boundary Cells 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item MAXIMA - outputVector - Maxima
\item MINIMA - outputVector - Minima
}
}
}
