% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_localstatisticalmeasures.R
\name{sagang_localstatisticalmeasures}
\alias{sagang_localstatisticalmeasures}
\title{QGIS algorithm - Local statistical measures}
\usage{
sagang_localstatisticalmeasures(
  GRID = qgisprocess:::qgis_default_value(),
  CONTRAST = qgisprocess:::qgis_default_value(),
  ENERGY = qgisprocess:::qgis_default_value(),
  ENTROPY = qgisprocess:::qgis_default_value(),
  VARIANCE = qgisprocess:::qgis_default_value(),
  TYPE = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  NORM_MIN = qgisprocess:::qgis_default_value(),
  NORM_MAX = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{CONTRAST}{\code{rasterDestination} - Contrast. Path for new raster layer.}

\item{ENERGY}{\code{rasterDestination} - Energy. Path for new raster layer.}

\item{ENTROPY}{\code{rasterDestination} - Entropy. Path for new raster layer.}

\item{VARIANCE}{\code{rasterDestination} - Variance. Path for new raster layer.}

\item{TYPE}{\code{enum}  of \verb{("(0) square", "(1) circle")} - Kernel. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NORMALIZE}{\code{enum}  of \verb{("(0) no", "(1) scale to range")} - Normalization. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NORM_MIN}{\code{number} - Minimum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NORM_MAX}{\code{number} - Maximum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Local statistical measures (sagang:localstatisticalmeasures). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CONTRAST: Contrast 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ENERGY: Energy 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ENTROPY: Entropy 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VARIANCE: Variance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TYPE: Kernel 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) square 		- 1: (1) circle 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RADIUS: Radius 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NORMALIZE: Normalization 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) no 		- 1: (1) scale to range 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NORM_MIN: Minimum 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NORM_MAX: Maximum 	Default value:	255 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONTRAST - outputRaster - Contrast
\item ENERGY - outputRaster - Energy
\item ENTROPY - outputRaster - Entropy
\item VARIANCE - outputRaster - Variance
}
}
}
