% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_longitudinalgridstatistics.R
\name{sagang_longitudinalgridstatistics}
\alias{sagang_longitudinalgridstatistics}
\title{QGIS algorithm - Longitudinal grid statistics}
\usage{
sagang_longitudinalgridstatistics(
  GRID = qgisprocess:::qgis_default_value(),
  STATS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{STATS}{\code{vectorDestination} - Latitudinal Statistics. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Longitudinal grid statistics (sagang:longitudinalgridstatistics). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer STATS: Latitudinal Statistics 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item STATS - outputVector - Latitudinal Statistics
}
}
}
