% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_lsfactor.R
\name{sagang_lsfactor}
\alias{sagang_lsfactor}
\title{QGIS algorithm - Ls factor}
\usage{
sagang_lsfactor(
  SLOPE = qgisprocess:::qgis_default_value(),
  AREA = qgisprocess:::qgis_default_value(),
  LS = qgisprocess:::qgis_default_value(),
  CONV = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  EROSIVITY = qgisprocess:::qgis_default_value(),
  STABILITY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SLOPE}{\code{raster} - Slope. Path to a raster layer.}

\item{AREA}{\code{raster} - Catchment Area. Path to a raster layer.}

\item{LS}{\code{rasterDestination} - LS Factor. Path for new raster layer.}

\item{CONV}{\code{enum}  of \verb{("(0) no conversion (areas already given as specific catchment area)", "(1) 1 / cell size (specific catchment area)", "(2) square root (catchment length)")} - Area to Length Conversion. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) Moore et al. 1991", "(1) Desmet & Govers 1996", "(2) Boehner & Selige 2006")} - Method (LS). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EROSIVITY}{\code{number} - Rill\code{/}Interrill Erosivity. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{STABILITY}{\code{enum}  of \verb{("(0) stable", "(1) instable (thawing)")} - Stability. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Ls factor (sagang:lsfactor). ---------------- Arguments ----------------  SLOPE: Slope 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer AREA: Catchment Area 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LS: LS Factor 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONV: Area to Length Conversion 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) no conversion (areas already given as specific catchment area) 		- 1: (1) 1 / cell size (specific catchment area) 		- 2: (2) square root (catchment length) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Method (LS) 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Moore et al. 1991 		- 1: (1) Desmet & Govers 1996 		- 2: (2) Boehner & Selige 2006 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' EROSIVITY: Rill/Interrill Erosivity 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression STABILITY: Stability 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) stable 		- 1: (1) instable (thawing) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item LS - outputRaster - LS Factor
}
}
}
