% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_lsfactorfieldbased.R
\name{sagang_lsfactorfieldbased}
\alias{sagang_lsfactorfieldbased}
\title{QGIS algorithm - Ls-factor, field based}
\usage{
sagang_lsfactorfieldbased(
  DEM = qgisprocess:::qgis_default_value(),
  FIELDS = qgisprocess:::qgis_default_value(),
  STATISTICS = qgisprocess:::qgis_default_value(),
  UPSLOPE_AREA = qgisprocess:::qgis_default_value(),
  UPSLOPE_LENGTH = qgisprocess:::qgis_default_value(),
  UPSLOPE_SLOPE = qgisprocess:::qgis_default_value(),
  LS_FACTOR = qgisprocess:::qgis_default_value(),
  BALANCE = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  METHOD_SLOPE = qgisprocess:::qgis_default_value(),
  METHOD_AREA = qgisprocess:::qgis_default_value(),
  STOP_AT_EDGE = qgisprocess:::qgis_default_value(),
  EROSIVITY = qgisprocess:::qgis_default_value(),
  STABILITY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{FIELDS}{\code{source} - Fields. Path to a vector layer.}

\item{STATISTICS}{\code{vectorDestination} - Field Statistics. Path for new vector layer.}

\item{UPSLOPE_AREA}{\code{rasterDestination} - Upslope Length Factor. Path for new raster layer.}

\item{UPSLOPE_LENGTH}{\code{rasterDestination} - Effective Flow Length. Path for new raster layer.}

\item{UPSLOPE_SLOPE}{\code{rasterDestination} - Upslope Slope. Path for new raster layer.}

\item{LS_FACTOR}{\code{rasterDestination} - LS Factor. Path for new raster layer.}

\item{BALANCE}{\code{rasterDestination} - Sediment Balance. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) Moore & Nieber 1989", "(1) Desmet & Govers 1996", "(2) Wischmeier & Smith 1978")} - LS Calculation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD_SLOPE}{\code{enum}  of \verb{("(0) local slope", "(1) distance weighted average catchment slope")} - Type of Slope. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD_AREA}{\code{enum}  of \verb{("(0) specific catchment area (contour length simply as cell size)", "(1) specific catchment area (contour length dependent on aspect)", "(2) catchment length (square root of catchment area)", "(3) effective flow length")} - Specific Catchment Area. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{STOP_AT_EDGE}{\code{boolean} - Stop at Edge. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EROSIVITY}{\code{number} - Rill\code{/}Interrill Erosivity. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{STABILITY}{\code{enum}  of \verb{("(0) stable", "(1) instable (thawing)")} - Stability. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Ls-factor, field based (sagang:lsfactorfieldbased). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FIELDS: Fields (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer STATISTICS: Field Statistics 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer UPSLOPE_AREA: Upslope Length Factor 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer UPSLOPE_LENGTH: Effective Flow Length 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer UPSLOPE_SLOPE: Upslope Slope 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LS_FACTOR: LS Factor 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BALANCE: Sediment Balance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: LS Calculation 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Moore & Nieber 1989 		- 1: (1) Desmet & Govers 1996 		- 2: (2) Wischmeier & Smith 1978 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD_SLOPE: Type of Slope 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) local slope 		- 1: (1) distance weighted average catchment slope 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD_AREA: Specific Catchment Area 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) specific catchment area (contour length simply as cell size) 		- 1: (1) specific catchment area (contour length dependent on aspect) 		- 2: (2) catchment length (square root of catchment area) 		- 3: (3) effective flow length 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' STOP_AT_EDGE: Stop at Edge 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EROSIVITY: Rill/Interrill Erosivity 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression STABILITY: Stability 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) stable 		- 1: (1) instable (thawing) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item BALANCE - outputRaster - Sediment Balance
\item LS_FACTOR - outputRaster - LS Factor
\item STATISTICS - outputVector - Field Statistics
\item UPSLOPE_AREA - outputRaster - Upslope Length Factor
\item UPSLOPE_LENGTH - outputRaster - Effective Flow Length
\item UPSLOPE_SLOPE - outputRaster - Upslope Slope
}
}
}
