% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_lsfactoronestep.R
\name{sagang_lsfactoronestep}
\alias{sagang_lsfactoronestep}
\title{QGIS algorithm - Ls factor (one step)}
\usage{
sagang_lsfactoronestep(
  DEM = qgisprocess:::qgis_default_value(),
  LS_FACTOR = qgisprocess:::qgis_default_value(),
  LS_METHOD = qgisprocess:::qgis_default_value(),
  PREPROCESSING = qgisprocess:::qgis_default_value(),
  MINSLOPE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{LS_FACTOR}{\code{rasterDestination} - LS Factor. Path for new raster layer.}

\item{LS_METHOD}{\code{enum}  of \verb{("(0) Moore et al. 1991", "(1) Desmond & Govers 1996")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PREPROCESSING}{\code{enum}  of \verb{("(0) none", "(1) Fill Sinks (Wang & Liu)", "(2) Sink Removal")} - Preprocessing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MINSLOPE}{\code{number} - Minimum Slope. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Ls factor (one step) (sagang:lsfactoronestep). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LS_FACTOR: LS Factor 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LS_METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Moore et al. 1991 		- 1: (1) Desmond & Govers 1996 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' PREPROCESSING: Preprocessing 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) none 		- 1: (1) Fill Sinks (Wang & Liu) 		- 2: (2) Sink Removal 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' MINSLOPE: Minimum Slope 	Default value:	0.0001 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item LS_FACTOR - outputRaster - LS Factor
}
}
}
