% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_massbalanceindex.R
\name{sagang_massbalanceindex}
\alias{sagang_massbalanceindex}
\title{QGIS algorithm - Mass balance index}
\usage{
sagang_massbalanceindex(
  DEM = qgisprocess:::qgis_default_value(),
  HREL = qgisprocess:::qgis_default_value(),
  MBI = qgisprocess:::qgis_default_value(),
  TSLOPE = qgisprocess:::qgis_default_value(),
  TCURVE = qgisprocess:::qgis_default_value(),
  THREL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{HREL}{\code{raster} - Vertical Distance to Channel Network. Path to a raster layer.}

\item{MBI}{\code{rasterDestination} - Mass Balance Index. Path for new raster layer.}

\item{TSLOPE}{\code{number} - T Slope. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TCURVE}{\code{number} - T Curvature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THREL}{\code{number} - T Vertical Distance to Channel Network. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Mass balance index (sagang:massbalanceindex). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer HREL: Vertical Distance to Channel Network (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MBI: Mass Balance Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TSLOPE: T Slope 	Default value:	15 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TCURVE: T Curvature 	Default value:	0.01 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THREL: T Vertical Distance to Channel Network 	Default value:	15 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item MBI - outputRaster - Mass Balance Index
}
}
}
