% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_maximumentropyclassifcation.R
\name{sagang_maximumentropyclassifcation}
\alias{sagang_maximumentropyclassifcation}
\title{QGIS algorithm - Maximum entropy classifcation}
\usage{
sagang_maximumentropyclassifcation(
  TRAINING = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  FEATURES_NUM = qgisprocess:::qgis_default_value(),
  FEATURES_CAT = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  CLASSES_LUT = qgisprocess:::qgis_default_value(),
  PROB = qgisprocess:::qgis_default_value(),
  PROBS = qgisprocess:::qgis_default_value(),
  PROBS_CREATE = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  YT_FILE_LOAD = qgisprocess:::qgis_default_value(),
  YT_FILE_SAVE = qgisprocess:::qgis_default_value(),
  YT_REGUL = qgisprocess:::qgis_default_value(),
  YT_REGUL_VAL = qgisprocess:::qgis_default_value(),
  YT_NUMASREAL = qgisprocess:::qgis_default_value(),
  DL_ALPHA = qgisprocess:::qgis_default_value(),
  DL_THRESHOLD = qgisprocess:::qgis_default_value(),
  DL_ITERATIONS = qgisprocess:::qgis_default_value(),
  NUM_CLASSES = qgisprocess:::qgis_default_value(),
  PROB_MIN = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TRAINING}{\code{source} - Training Areas. Path to a vector layer.}

\item{FIELD}{\code{field} - Class Name. The name of an existing field. ; delimited list of existing field names.}

\item{FEATURES_NUM}{\code{multilayer} - Numerical Features. .}

\item{FEATURES_CAT}{\code{multilayer} - Categorical Features. .}

\item{CLASSES}{\code{rasterDestination} - Classes. Path for new raster layer.}

\item{CLASSES_LUT}{\code{vectorDestination} - Look-up Table. Path for new vector layer.}

\item{PROB}{\code{rasterDestination} - Probability. Path for new raster layer.}

\item{PROBS}{\code{rasterDestination} - Probabilities. Path for new raster layer.}

\item{PROBS_CREATE}{\code{boolean} - Create Propabilities. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) Yoshimasa Tsuruoka", "(1) Dekang Lin")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{YT_FILE_LOAD}{\code{file} - Load from File.... Path to a file.}

\item{YT_FILE_SAVE}{\code{file} - Save to File.... Path to a file.}

\item{YT_REGUL}{\code{enum}  of \verb{("(0) none", "(1) L1", "(2) L2")} - Regularization. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{YT_REGUL_VAL}{\code{number} - Regularization Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YT_NUMASREAL}{\code{boolean} - Real-valued Numerical Features. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DL_ALPHA}{\code{number} - Alpha. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DL_THRESHOLD}{\code{number} - Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DL_ITERATIONS}{\code{number} - Maximum Iterations. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NUM_CLASSES}{\code{number} - Number of Numeric Value Classes. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PROB_MIN}{\code{number} - Minimum Probability. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Maximum entropy classifcation (sagang:maximumentropyclassifcation). ---------------- Arguments ----------------  TRAINING: Training Areas 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Class Name 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FEATURES_NUM: Numerical Features (optional) 	Argument type:	multilayer FEATURES_CAT: Categorical Features (optional) 	Argument type:	multilayer CLASSES: Classes 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CLASSES_LUT: Look-up Table 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer PROB: Probability 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PROBS: Probabilities 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PROBS_CREATE: Create Propabilities 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Yoshimasa Tsuruoka 		- 1: (1) Dekang Lin 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' YT_FILE_LOAD: Load from File... 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file YT_FILE_SAVE: Save to File... 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file YT_REGUL: Regularization 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) none 		- 1: (1) L1 		- 2: (2) L2 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' YT_REGUL_VAL: Regularization Factor 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YT_NUMASREAL: Real-valued Numerical Features 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DL_ALPHA: Alpha 	Default value:	0.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DL_THRESHOLD: Threshold 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DL_ITERATIONS: Maximum Iterations 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NUM_CLASSES: Number of Numeric Value Classes 	Default value:	32 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PROB_MIN: Minimum Probability 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Classes
\item CLASSES_LUT - outputVector - Look-up Table
\item PROB - outputRaster - Probability
\item PROBS - outputRaster - Probabilities
}
}
}
