% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_maximumflowpathlength.R
\name{sagang_maximumflowpathlength}
\alias{sagang_maximumflowpathlength}
\title{QGIS algorithm - Maximum flow path length}
\usage{
sagang_maximumflowpathlength(
  ELEVATION = qgisprocess:::qgis_default_value(),
  WEIGHTS = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  DIRECTION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{WEIGHTS}{\code{raster} - Weights. Path to a raster layer.}

\item{DISTANCE}{\code{rasterDestination} - Flow Path Length. Path for new raster layer.}

\item{DIRECTION}{\code{enum}  of \verb{("(0) downstream", "(1) upstream")} - Direction of Measurement. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Maximum flow path length (sagang:maximumflowpathlength). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer WEIGHTS: Weights (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DISTANCE: Flow Path Length 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIRECTION: Direction of Measurement 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) downstream 		- 1: (1) upstream 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item DISTANCE - outputRaster - Flow Path Length
}
}
}
