% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_mergevectorlayers.R
\name{sagang_mergevectorlayers}
\alias{sagang_mergevectorlayers}
\title{QGIS algorithm - Merge vector layers}
\usage{
sagang_mergevectorlayers(
  INPUT = qgisprocess:::qgis_default_value(),
  MERGED = qgisprocess:::qgis_default_value(),
  SRCINFO = qgisprocess:::qgis_default_value(),
  MATCH = qgisprocess:::qgis_default_value(),
  DELETE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Layers. .}

\item{MERGED}{\code{vectorDestination} - Merged Layer. Path for new vector layer.}

\item{SRCINFO}{\code{boolean} - Add Source Information. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MATCH}{\code{boolean} - Match Fields by Name. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELETE}{\code{boolean} - Delete. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Merge vector layers (sagang:mergevectorlayers). ---------------- Arguments ----------------  INPUT: Layers 	Argument type:	multilayer MERGED: Merged Layer 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer SRCINFO: Add Source Information 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MATCH: Match Fields by Name 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DELETE: Delete 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item MERGED - outputVector - Merged Layer
}
}
}
