% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_metricconversions.R
\name{sagang_metricconversions}
\alias{sagang_metricconversions}
\title{QGIS algorithm - Metric conversions}
\usage{
sagang_metricconversions(
  GRID = qgisprocess:::qgis_default_value(),
  CONV = qgisprocess:::qgis_default_value(),
  CONVERSION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{CONV}{\code{rasterDestination} - Converted Grid. Path for new raster layer.}

\item{CONVERSION}{\code{enum}  of \verb{("(0) radians to degree", "(1) degree to radians", "(2) Celsius to Fahrenheit", "(3) Fahrenheit to Celsius")} - Conversion. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Metric conversions (sagang:metricconversions). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CONV: Converted Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONVERSION: Conversion 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) radians to degree 		- 1: (1) degree to radians 		- 2: (2) Celsius to Fahrenheit 		- 3: (3) Fahrenheit to Celsius 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONV - outputRaster - Converted Grid
}
}
}
