% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_mirrorgrid.R
\name{sagang_mirrorgrid}
\alias{sagang_mirrorgrid}
\title{QGIS algorithm - Mirror grid}
\usage{
sagang_mirrorgrid(
  GRID = qgisprocess:::qgis_default_value(),
  MIRROR = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{MIRROR}{\code{rasterDestination} - Mirror Grid. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) horizontally", "(1) vertically", "(2) both")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Mirror grid (sagang:mirrorgrid). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MIRROR: Mirror Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) horizontally 		- 1: (1) vertically 		- 2: (2) both 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item MIRROR - outputRaster - Mirror Grid
}
}
}
