% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_monthlyglobalbylatitude.R
\name{sagang_monthlyglobalbylatitude}
\alias{sagang_monthlyglobalbylatitude}
\title{QGIS algorithm - Monthly global by latitude}
\usage{
sagang_monthlyglobalbylatitude(
  SOLARRAD = qgisprocess:::qgis_default_value(),
  ALBEDO = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  YEAR = qgisprocess:::qgis_default_value(),
  DLAT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SOLARRAD}{\code{vectorDestination} - Solar Radiation. Path for new vector layer.}

\item{ALBEDO}{\code{source} - Albedo. Path to a vector layer.}

\item{FIELD}{\code{field} - Field. The name of an existing field. ; delimited list of existing field names.}

\item{YEAR}{\code{number} - Year (ka). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DLAT}{\code{number} - Latitude Increment (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Monthly global by latitude (sagang:monthlyglobalbylatitude). ---------------- Arguments ----------------  SOLARRAD: Solar Radiation 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer ALBEDO: Albedo (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Field 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names YEAR: Year (ka) 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DLAT: Latitude Increment (Degree) 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item SOLARRAD - outputVector - Solar Radiation
}
}
}
