% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_morphometricfeatures.R
\name{sagang_morphometricfeatures}
\alias{sagang_morphometricfeatures}
\title{QGIS algorithm - Morphometric features}
\usage{
sagang_morphometricfeatures(
  DEM = qgisprocess:::qgis_default_value(),
  FEATURES = qgisprocess:::qgis_default_value(),
  ELEVATION = qgisprocess:::qgis_default_value(),
  SLOPE = qgisprocess:::qgis_default_value(),
  ASPECT = qgisprocess:::qgis_default_value(),
  PROFC = qgisprocess:::qgis_default_value(),
  PLANC = qgisprocess:::qgis_default_value(),
  LONGC = qgisprocess:::qgis_default_value(),
  CROSC = qgisprocess:::qgis_default_value(),
  MAXIC = qgisprocess:::qgis_default_value(),
  MINIC = qgisprocess:::qgis_default_value(),
  SIZE = qgisprocess:::qgis_default_value(),
  TOL_SLOPE = qgisprocess:::qgis_default_value(),
  TOL_CURVE = qgisprocess:::qgis_default_value(),
  EXPONENT = qgisprocess:::qgis_default_value(),
  ZSCALE = qgisprocess:::qgis_default_value(),
  CONSTRAIN = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{FEATURES}{\code{rasterDestination} - Morphometric Features. Path for new raster layer.}

\item{ELEVATION}{\code{rasterDestination} - Generalized Surface. Path for new raster layer.}

\item{SLOPE}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{ASPECT}{\code{rasterDestination} - Aspect. Path for new raster layer.}

\item{PROFC}{\code{rasterDestination} - Profile Curvature. Path for new raster layer.}

\item{PLANC}{\code{rasterDestination} - Plan Curvature. Path for new raster layer.}

\item{LONGC}{\code{rasterDestination} - Longitudinal Curvature. Path for new raster layer.}

\item{CROSC}{\code{rasterDestination} - Cross-Sectional Curvature. Path for new raster layer.}

\item{MAXIC}{\code{rasterDestination} - Maximum Curvature. Path for new raster layer.}

\item{MINIC}{\code{rasterDestination} - Minimum Curvature. Path for new raster layer.}

\item{SIZE}{\code{number} - Scale Radius (Cells). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TOL_SLOPE}{\code{number} - Slope Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TOL_CURVE}{\code{number} - Curvature Tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EXPONENT}{\code{number} - Distance Weighting Exponent. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ZSCALE}{\code{number} - Vertical Scaling. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CONSTRAIN}{\code{boolean} - Constrain. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Morphometric features (sagang:morphometricfeatures). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FEATURES: Morphometric Features 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ELEVATION: Generalized Surface 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SLOPE: Slope 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ASPECT: Aspect 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PROFC: Profile Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PLANC: Plan Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LONGC: Longitudinal Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CROSC: Cross-Sectional Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MAXIC: Maximum Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MINIC: Minimum Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SIZE: Scale Radius (Cells) 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TOL_SLOPE: Slope Tolerance 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TOL_CURVE: Curvature Tolerance 	Default value:	0.0001 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EXPONENT: Distance Weighting Exponent 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ZSCALE: Vertical Scaling 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CONSTRAIN: Constrain 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item ASPECT - outputRaster - Aspect
\item CROSC - outputRaster - Cross-Sectional Curvature
\item ELEVATION - outputRaster - Generalized Surface
\item FEATURES - outputRaster - Morphometric Features
\item LONGC - outputRaster - Longitudinal Curvature
\item MAXIC - outputRaster - Maximum Curvature
\item MINIC - outputRaster - Minimum Curvature
\item PLANC - outputRaster - Plan Curvature
\item PROFC - outputRaster - Profile Curvature
\item SLOPE - outputRaster - Slope
}
}
}
