% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_morphometricprotectionindex.R
\name{sagang_morphometricprotectionindex}
\alias{sagang_morphometricprotectionindex}
\title{QGIS algorithm - Morphometric protection index}
\usage{
sagang_morphometricprotectionindex(
  DEM = qgisprocess:::qgis_default_value(),
  PROTECTION = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{PROTECTION}{\code{rasterDestination} - Protection Index. Path for new raster layer.}

\item{RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Morphometric protection index (sagang:morphometricprotectionindex). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer PROTECTION: Protection Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RADIUS: Radius 	Default value:	2000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item PROTECTION - outputRaster - Protection Index
}
}
}
