% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_multidirectionleefilter.R
\name{sagang_multidirectionleefilter}
\alias{sagang_multidirectionleefilter}
\title{QGIS algorithm - Multi direction lee filter}
\usage{
sagang_multidirectionleefilter(
  INPUT = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  STDDEV = qgisprocess:::qgis_default_value(),
  DIR = qgisprocess:::qgis_default_value(),
  NOISE_ABS = qgisprocess:::qgis_default_value(),
  NOISE_REL = qgisprocess:::qgis_default_value(),
  WEIGHTED = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - Filtered Grid. Path for new raster layer.}

\item{STDDEV}{\code{rasterDestination} - Minimum Standard Deviation. Path for new raster layer.}

\item{DIR}{\code{rasterDestination} - Direction of Minimum Standard Deviation. Path for new raster layer.}

\item{NOISE_ABS}{\code{number} - Estimated Noise (absolute). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NOISE_REL}{\code{number} - Estimated Noise (relative). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WEIGHTED}{\code{boolean} - Weighted. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) noise variance given as absolute value", "(1) noise variance given relative to mean standard deviation", "(2) original calculation (Ringeler)")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Multi direction lee filter (sagang:multidirectionleefilter). ---------------- Arguments ----------------  INPUT: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RESULT: Filtered Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STDDEV: Minimum Standard Deviation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIR: Direction of Minimum Standard Deviation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer NOISE_ABS: Estimated Noise (absolute) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NOISE_REL: Estimated Noise (relative) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WEIGHTED: Weighted 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Method 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) noise variance given as absolute value 		- 1: (1) noise variance given relative to mean standard deviation 		- 2: (2) original calculation (Ringeler) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIR - outputRaster - Direction of Minimum Standard Deviation
\item RESULT - outputRaster - Filtered Grid
\item STDDEV - outputRaster - Minimum Standard Deviation
}
}
}
