% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sagang_multiresolutionindexofvalleybottomflatnessmrvbf.R
\name{sagang_multiresolutionindexofvalleybottomflatnessmrvbf}
\alias{sagang_multiresolutionindexofvalleybottomflatnessmrvbf}
\title{QGIS algorithm - Multiresolution index of valley bottom flatness (mrvbf)}
\usage{
sagang_multiresolutionindexofvalleybottomflatnessmrvbf(
  DEM = qgisprocess:::qgis_default_value(),
  MRVBF = qgisprocess:::qgis_default_value(),
  MRRTF = qgisprocess:::qgis_default_value(),
  T_SLOPE = qgisprocess:::qgis_default_value(),
  T_PCTL_V = qgisprocess:::qgis_default_value(),
  T_PCTL_R = qgisprocess:::qgis_default_value(),
  P_SLOPE = qgisprocess:::qgis_default_value(),
  P_PCTL = qgisprocess:::qgis_default_value(),
  UPDATE = qgisprocess:::qgis_default_value(),
  CLASSIFY = qgisprocess:::qgis_default_value(),
  MAX_RES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{MRVBF}{\code{rasterDestination} - MRVBF. Path for new raster layer.}

\item{MRRTF}{\code{rasterDestination} - MRRTF. Path for new raster layer.}

\item{T_SLOPE}{\code{number} - Initial Threshold for Slope. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T_PCTL_V}{\code{number} - Threshold for Elevation Percentile (Lowness). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T_PCTL_R}{\code{number} - Threshold for Elevation Percentile (Upness). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_SLOPE}{\code{number} - Shape Parameter for Slope. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_PCTL}{\code{number} - Shape Parameter for Elevation Percentile. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{UPDATE}{\code{boolean} - Update Views. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CLASSIFY}{\code{boolean} - Classify. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAX_RES}{\code{number} - Maximum Resolution (Percentage). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Multiresolution index of valley bottom flatness (mrvbf) (sagang:multiresolutionindexofvalleybottomflatnessmrvbf). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MRVBF: MRVBF 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MRRTF: MRRTF 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer T_SLOPE: Initial Threshold for Slope 	Default value:	16 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T_PCTL_V: Threshold for Elevation Percentile (Lowness) 	Default value:	0.4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T_PCTL_R: Threshold for Elevation Percentile (Upness) 	Default value:	0.35 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_SLOPE: Shape Parameter for Slope 	Default value:	4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_PCTL: Shape Parameter for Elevation Percentile 	Default value:	3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression UPDATE: Update Views 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CLASSIFY: Classify 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAX_RES: Maximum Resolution (Percentage) 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item MRRTF - outputRaster - MRRTF
\item MRVBF - outputRaster - MRVBF
}
}
}
