% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_multiscaletopographicpositionindextpi.R
\name{sagang_multiscaletopographicpositionindextpi}
\alias{sagang_multiscaletopographicpositionindextpi}
\title{QGIS algorithm - Multi-scale topographic position index (tpi)}
\usage{
sagang_multiscaletopographicpositionindextpi(
  DEM = qgisprocess:::qgis_default_value(),
  TPI = qgisprocess:::qgis_default_value(),
  SCALE_MIN = qgisprocess:::qgis_default_value(),
  SCALE_MAX = qgisprocess:::qgis_default_value(),
  SCALE_NUM = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{TPI}{\code{rasterDestination} - Topographic Position Index. Path for new raster layer.}

\item{SCALE_MIN}{\code{number} - Minimum Scale. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SCALE_MAX}{\code{number} - Maximum Scale. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SCALE_NUM}{\code{number} - Number of Scales. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Multi-scale topographic position index (tpi) (sagang:multiscaletopographicpositionindextpi). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TPI: Topographic Position Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SCALE_MIN: Minimum Scale 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SCALE_MAX: Maximum Scale 	Default value:	8 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SCALE_NUM: Number of Scales 	Default value:	3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item TPI - outputRaster - Topographic Position Index
}
}
}
