% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_notchfilterforgrids.R
\name{sagang_notchfilterforgrids}
\alias{sagang_notchfilterforgrids}
\title{QGIS algorithm - Notch filter for grids}
\usage{
sagang_notchfilterforgrids(
  INPUT_GRID = qgisprocess:::qgis_default_value(),
  LOWPASS_UPPER = qgisprocess:::qgis_default_value(),
  HIPASS_LOWER = qgisprocess:::qgis_default_value(),
  FILTERED_GRID = qgisprocess:::qgis_default_value(),
  HIGH_PASS = qgisprocess:::qgis_default_value(),
  LOW_PASS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_GRID}{\code{raster} - Input Grid. Path to a raster layer.}

\item{LOWPASS_UPPER}{\code{rasterDestination} - Lowpass Upper. Path for new raster layer.}

\item{HIPASS_LOWER}{\code{rasterDestination} - Highpass Lower. Path for new raster layer.}

\item{FILTERED_GRID}{\code{rasterDestination} - Notch. Path for new raster layer.}

\item{HIGH_PASS}{\code{number} - Upper search distance . A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LOW_PASS}{\code{number} - Lower search distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Notch filter for grids (sagang:notchfilterforgrids). ---------------- Arguments ----------------  INPUT_GRID: Input Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LOWPASS_UPPER: Lowpass Upper 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer HIPASS_LOWER: Highpass Lower 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FILTERED_GRID: Notch 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer HIGH_PASS: Upper search distance  	Default value:	75 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LOW_PASS: Lower search distance 	Default value:	25 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FILTERED_GRID - outputRaster - Notch
\item HIPASS_LOWER - outputRaster - Highpass Lower
\item LOWPASS_UPPER - outputRaster - Lowpass Upper
}
}
}
