% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_objectbasedimagesegmentation.R
\name{sagang_objectbasedimagesegmentation}
\alias{sagang_objectbasedimagesegmentation}
\title{QGIS algorithm - Object based image segmentation}
\usage{
sagang_objectbasedimagesegmentation(
  FEATURES = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  OBJECTS = qgisprocess:::qgis_default_value(),
  SEEDS_BAND_WIDTH = qgisprocess:::qgis_default_value(),
  RGA_NEIGHBOUR = qgisprocess:::qgis_default_value(),
  RGA_METHOD = qgisprocess:::qgis_default_value(),
  RGA_SIG_1 = qgisprocess:::qgis_default_value(),
  RGA_SIG_2 = qgisprocess:::qgis_default_value(),
  RGA_SIMILARITY = qgisprocess:::qgis_default_value(),
  MAJORITY_RADIUS = qgisprocess:::qgis_default_value(),
  POSTPROCESSING = qgisprocess:::qgis_default_value(),
  NCLUSTER = qgisprocess:::qgis_default_value(),
  SPLIT_CLUSTERS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FEATURES}{\code{multilayer} - Features. .}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OBJECTS}{\code{vectorDestination} - Segments. Path for new vector layer.}

\item{SEEDS_BAND_WIDTH}{\code{number} - Band Width for Seed Point Generation. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RGA_NEIGHBOUR}{\code{enum}  of \verb{("(0) 4 (Neumann)", "(1) 8 (Moore)")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RGA_METHOD}{\code{enum}  of \verb{("(0) feature space and position", "(1) feature space")} - Distance. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RGA_SIG_1}{\code{number} - Variance in Feature Space. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RGA_SIG_2}{\code{number} - Variance in Position Space. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RGA_SIMILARITY}{\code{number} - Similarity Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAJORITY_RADIUS}{\code{number} - Generalization. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{POSTPROCESSING}{\code{enum}  of \verb{("(0) none", "(1) unsupervised classification")} - Post-Processing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NCLUSTER}{\code{number} - Number of Clusters. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SPLIT_CLUSTERS}{\code{boolean} - Split Clusters. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Object based image segmentation (sagang:objectbasedimagesegmentation). ---------------- Arguments ----------------  FEATURES: Features 	Argument type:	multilayer NORMALIZE: Normalize 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OBJECTS: Segments 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer SEEDS_BAND_WIDTH: Band Width for Seed Point Generation 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RGA_NEIGHBOUR: Neighbourhood 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) 4 (Neumann) 		- 1: (1) 8 (Moore) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RGA_METHOD: Distance 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) feature space and position 		- 1: (1) feature space 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RGA_SIG_1: Variance in Feature Space 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RGA_SIG_2: Variance in Position Space 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RGA_SIMILARITY: Similarity Threshold 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAJORITY_RADIUS: Generalization 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression POSTPROCESSING: Post-Processing 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) none 		- 1: (1) unsupervised classification 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NCLUSTER: Number of Clusters 	Default value:	12 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SPLIT_CLUSTERS: Split Clusters 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OBJECTS - outputVector - Segments
}
}
}
