% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_overlandflow.R
\name{sagang_overlandflow}
\alias{sagang_overlandflow}
\title{QGIS algorithm - Overland flow}
\usage{
sagang_overlandflow(
  DEM = qgisprocess:::qgis_default_value(),
  ROUGHNESS = qgisprocess:::qgis_default_value(),
  ROUGHNESS_DEFAULT = qgisprocess:::qgis_default_value(),
  INTER_MAX = qgisprocess:::qgis_default_value(),
  INTER_MAX_DEFAULT = qgisprocess:::qgis_default_value(),
  POND_MAX = qgisprocess:::qgis_default_value(),
  POND_MAX_DEFAULT = qgisprocess:::qgis_default_value(),
  INFIL_MAX = qgisprocess:::qgis_default_value(),
  INFIL_MAX_DEFAULT = qgisprocess:::qgis_default_value(),
  INTERCEPT = qgisprocess:::qgis_default_value(),
  PONDING = qgisprocess:::qgis_default_value(),
  INFILTRAT = qgisprocess:::qgis_default_value(),
  FLOW = qgisprocess:::qgis_default_value(),
  VELOCITY = qgisprocess:::qgis_default_value(),
  PRECIP = qgisprocess:::qgis_default_value(),
  PRECIP_DEFAULT = qgisprocess:::qgis_default_value(),
  ET_POT = qgisprocess:::qgis_default_value(),
  ET_POT_DEFAULT = qgisprocess:::qgis_default_value(),
  RESET = qgisprocess:::qgis_default_value(),
  TIME_STOP = qgisprocess:::qgis_default_value(),
  TIME_STEP = qgisprocess:::qgis_default_value(),
  TIME_UPDATE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{ROUGHNESS}{\code{raster} - Manning Roughness. Path to a raster layer.}

\item{ROUGHNESS_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INTER_MAX}{\code{raster} - Interception Capacity (mm). Path to a raster layer.}

\item{INTER_MAX_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{POND_MAX}{\code{raster} - Ponding Capacity (mm). Path to a raster layer.}

\item{POND_MAX_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INFIL_MAX}{\code{raster} - Infiltration Capacity (mm\code{/}h). Path to a raster layer.}

\item{INFIL_MAX_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INTERCEPT}{\code{rasterDestination} - Interception (mm). Path for new raster layer.}

\item{PONDING}{\code{rasterDestination} - Ponding (mm). Path for new raster layer.}

\item{INFILTRAT}{\code{rasterDestination} - Infiltration (mm). Path for new raster layer.}

\item{FLOW}{\code{rasterDestination} - Flow (mm). Path for new raster layer.}

\item{VELOCITY}{\code{rasterDestination} - Velocity (m\code{/}s). Path for new raster layer.}

\item{PRECIP}{\code{raster} - Precipitation (mm\code{/}h). Path to a raster layer.}

\item{PRECIP_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ET_POT}{\code{raster} - Potential Evapotranspiration (mm\code{/}h). Path to a raster layer.}

\item{ET_POT_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RESET}{\code{boolean} - Reset. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME_STOP}{\code{number} - Simulation Time (h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME_STEP}{\code{number} - Time Step Adjustment. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME_UPDATE}{\code{number} - Map Update Frequency (min). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Overland flow (sagang:overlandflow). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer ROUGHNESS: Manning Roughness (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer ROUGHNESS_DEFAULT: Default 	Default value:	0.03 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression INTER_MAX: Interception Capacity (mm) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer INTER_MAX_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression POND_MAX: Ponding Capacity (mm) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer POND_MAX_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression INFIL_MAX: Infiltration Capacity (mm/h) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer INFIL_MAX_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression INTERCEPT: Interception (mm) 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PONDING: Ponding (mm) 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer INFILTRAT: Infiltration (mm) 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FLOW: Flow (mm) 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VELOCITY: Velocity (m/s) 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PRECIP: Precipitation (mm/h) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer PRECIP_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ET_POT: Potential Evapotranspiration (mm/h) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer ET_POT_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RESET: Reset 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME_STOP: Simulation Time (h) 	Default value:	6 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME_STEP: Time Step Adjustment 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME_UPDATE: Map Update Frequency (min) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLOW - outputRaster - Flow (mm)
\item INFILTRAT - outputRaster - Infiltration (mm)
\item INTERCEPT - outputRaster - Interception (mm)
\item PONDING - outputRaster - Ponding (mm)
\item VELOCITY - outputRaster - Velocity (m\code{/}s)
}
}
}
