% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_overlandflowdistancetochannelnetwork.R
\name{sagang_overlandflowdistancetochannelnetwork}
\alias{sagang_overlandflowdistancetochannelnetwork}
\title{QGIS algorithm - Overland flow distance to channel network}
\usage{
sagang_overlandflowdistancetochannelnetwork(
  ELEVATION = qgisprocess:::qgis_default_value(),
  CHANNELS = qgisprocess:::qgis_default_value(),
  ROUTE = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  DISTVERT = qgisprocess:::qgis_default_value(),
  DISTHORZ = qgisprocess:::qgis_default_value(),
  TIME = qgisprocess:::qgis_default_value(),
  SDR = qgisprocess:::qgis_default_value(),
  FIELDS = qgisprocess:::qgis_default_value(),
  PASSES = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  BOUNDARY = qgisprocess:::qgis_default_value(),
  FLOW_B = qgisprocess:::qgis_default_value(),
  FLOW_K = qgisprocess:::qgis_default_value(),
  FLOW_K_DEFAULT = qgisprocess:::qgis_default_value(),
  FLOW_R = qgisprocess:::qgis_default_value(),
  FLOW_R_DEFAULT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{CHANNELS}{\code{raster} - Channel Network. Path to a raster layer.}

\item{ROUTE}{\code{raster} - Preferred Routing. Path to a raster layer.}

\item{DISTANCE}{\code{rasterDestination} - Overland Flow Distance. Path for new raster layer.}

\item{DISTVERT}{\code{rasterDestination} - Vertical Overland Flow Distance. Path for new raster layer.}

\item{DISTHORZ}{\code{rasterDestination} - Horizontal Overland Flow Distance. Path for new raster layer.}

\item{TIME}{\code{rasterDestination} - Flow Travel Time. Path for new raster layer.}

\item{SDR}{\code{rasterDestination} - Sediment Yield Delivery Ratio. Path for new raster layer.}

\item{FIELDS}{\code{raster} - Fields. Path to a raster layer.}

\item{PASSES}{\code{rasterDestination} - Fields Visited. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) D8", "(1) MFD")} - Flow Algorithm. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{BOUNDARY}{\code{boolean} - Boundary Cells. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FLOW_B}{\code{number} - Beta. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FLOW_K}{\code{raster} - Manning-Strickler Coefficient. Path to a raster layer.}

\item{FLOW_K_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FLOW_R}{\code{raster} - Flow Depth. Path to a raster layer.}

\item{FLOW_R_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Overland flow distance to channel network (sagang:overlandflowdistancetochannelnetwork). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CHANNELS: Channel Network 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer ROUTE: Preferred Routing (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DISTANCE: Overland Flow Distance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DISTVERT: Vertical Overland Flow Distance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DISTHORZ: Horizontal Overland Flow Distance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TIME: Flow Travel Time 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SDR: Sediment Yield Delivery Ratio 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FIELDS: Fields (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer PASSES: Fields Visited 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Flow Algorithm 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) D8 		- 1: (1) MFD 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' BOUNDARY: Boundary Cells 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FLOW_B: Beta 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FLOW_K: Manning-Strickler Coefficient (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FLOW_K_DEFAULT: Default 	Default value:	20 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FLOW_R: Flow Depth (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FLOW_R_DEFAULT: Default 	Default value:	0.05 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item DISTANCE - outputRaster - Overland Flow Distance
\item DISTHORZ - outputRaster - Horizontal Overland Flow Distance
\item DISTVERT - outputRaster - Vertical Overland Flow Distance
\item PASSES - outputRaster - Fields Visited
\item SDR - outputRaster - Sediment Yield Delivery Ratio
\item TIME - outputRaster - Flow Travel Time
}
}
}
