% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_patternanalysis.R
\name{sagang_patternanalysis}
\alias{sagang_patternanalysis}
\title{QGIS algorithm - Pattern analysis}
\usage{
sagang_patternanalysis(
  INPUT = qgisprocess:::qgis_default_value(),
  NDC = qgisprocess:::qgis_default_value(),
  RELATIVE = qgisprocess:::qgis_default_value(),
  FRAGMENTATION = qgisprocess:::qgis_default_value(),
  CVN = qgisprocess:::qgis_default_value(),
  DIVERSITY = qgisprocess:::qgis_default_value(),
  DOMINANCE = qgisprocess:::qgis_default_value(),
  MAXNUMCLASS = qgisprocess:::qgis_default_value(),
  KERNEL_TYPE = qgisprocess:::qgis_default_value(),
  KERNEL_RADIUS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input Grid. Path to a raster layer.}

\item{NDC}{\code{rasterDestination} - Number of Classes. Path for new raster layer.}

\item{RELATIVE}{\code{rasterDestination} - Relative Richness. Path for new raster layer.}

\item{FRAGMENTATION}{\code{rasterDestination} - Fragmentation. Path for new raster layer.}

\item{CVN}{\code{rasterDestination} - Center vs. Neighbours. Path for new raster layer.}

\item{DIVERSITY}{\code{rasterDestination} - Diversity. Path for new raster layer.}

\item{DOMINANCE}{\code{rasterDestination} - Dominance. Path for new raster layer.}

\item{MAXNUMCLASS}{\code{number} - Max. Number of Classes. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KERNEL_TYPE}{\code{enum}  of \verb{("(0) Square", "(1) Circle")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Pattern analysis (sagang:patternanalysis). ---------------- Arguments ----------------  INPUT: Input Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer NDC: Number of Classes 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RELATIVE: Relative Richness 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FRAGMENTATION: Fragmentation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CVN: Center vs. Neighbours 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIVERSITY: Diversity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DOMINANCE: Dominance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MAXNUMCLASS: Max. Number of Classes 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KERNEL_TYPE: Kernel Type 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Square 		- 1: (1) Circle 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' KERNEL_RADIUS: Radius 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CVN - outputRaster - Center vs. Neighbours
\item DIVERSITY - outputRaster - Diversity
\item DOMINANCE - outputRaster - Dominance
\item FRAGMENTATION - outputRaster - Fragmentation
\item NDC - outputRaster - Number of Classes
\item RELATIVE - outputRaster - Relative Richness
}
}
}
