% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_phenipsgridsannual.R
\name{sagang_phenipsgridsannual}
\alias{sagang_phenipsgridsannual}
\title{QGIS algorithm - Phenips (grids, annual)}
\usage{
sagang_phenipsgridsannual(
  ATmean = qgisprocess:::qgis_default_value(),
  ATmax = qgisprocess:::qgis_default_value(),
  SIrel = qgisprocess:::qgis_default_value(),
  GENERATIONS = qgisprocess:::qgis_default_value(),
  ONSET = qgisprocess:::qgis_default_value(),
  ONSET_FILIAL_1 = qgisprocess:::qgis_default_value(),
  ONSET_SISTER_1 = qgisprocess:::qgis_default_value(),
  ONSET_FILIAL_2 = qgisprocess:::qgis_default_value(),
  ONSET_SISTER_2 = qgisprocess:::qgis_default_value(),
  ONSET_FILIAL_3 = qgisprocess:::qgis_default_value(),
  ONSET_SISTER_3 = qgisprocess:::qgis_default_value(),
  BTSUM_FILIAL_1 = qgisprocess:::qgis_default_value(),
  BTSUM_SISTER_1 = qgisprocess:::qgis_default_value(),
  BTSUM_FILIAL_2 = qgisprocess:::qgis_default_value(),
  BTSUM_SISTER_2 = qgisprocess:::qgis_default_value(),
  BTSUM_FILIAL_3 = qgisprocess:::qgis_default_value(),
  BTSUM_SISTER_3 = qgisprocess:::qgis_default_value(),
  LAT_GRID = qgisprocess:::qgis_default_value(),
  LAT_CONST = qgisprocess:::qgis_default_value(),
  DToptimum = qgisprocess:::qgis_default_value(),
  DTminimum = qgisprocess:::qgis_default_value(),
  FAminimum = qgisprocess:::qgis_default_value(),
  DayLength = qgisprocess:::qgis_default_value(),
  DDminimum = qgisprocess:::qgis_default_value(),
  DDtotal = qgisprocess:::qgis_default_value(),
  Risk_DayMax = qgisprocess:::qgis_default_value(),
  Risk_Decay = qgisprocess:::qgis_default_value(),
  YD_Begin = qgisprocess:::qgis_default_value(),
  YD_End_Onset = qgisprocess:::qgis_default_value(),
  YD_End = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ATmean}{\code{multilayer} - Mean Temperature. .}

\item{ATmax}{\code{multilayer} - Maximum Temperature. .}

\item{SIrel}{\code{multilayer} - Solar Irradiation. .}

\item{GENERATIONS}{\code{rasterDestination} - Potential Number of Generations. Path for new raster layer.}

\item{ONSET}{\code{rasterDestination} - Onset Day of Infestation. Path for new raster layer.}

\item{ONSET_FILIAL_1}{\code{rasterDestination} - Onset Day, 1. Filial Generation. Path for new raster layer.}

\item{ONSET_SISTER_1}{\code{rasterDestination} - Onset Day, 1. Sister Generation. Path for new raster layer.}

\item{ONSET_FILIAL_2}{\code{rasterDestination} - Onset Day, 2. Filial Generation. Path for new raster layer.}

\item{ONSET_SISTER_2}{\code{rasterDestination} - Onset Day, 2. Sister Generation. Path for new raster layer.}

\item{ONSET_FILIAL_3}{\code{rasterDestination} - Onset Day, 3. Filial Generation. Path for new raster layer.}

\item{ONSET_SISTER_3}{\code{rasterDestination} - Onset Day, 3. Sister Generation. Path for new raster layer.}

\item{BTSUM_FILIAL_1}{\code{rasterDestination} - State, 1. Filial Generation. Path for new raster layer.}

\item{BTSUM_SISTER_1}{\code{rasterDestination} - State, 1. Sister Generation. Path for new raster layer.}

\item{BTSUM_FILIAL_2}{\code{rasterDestination} - State, 2. Filial Generation. Path for new raster layer.}

\item{BTSUM_SISTER_2}{\code{rasterDestination} - State, 2. Sister Generation. Path for new raster layer.}

\item{BTSUM_FILIAL_3}{\code{rasterDestination} - State, 3. Filial Generation. Path for new raster layer.}

\item{BTSUM_SISTER_3}{\code{rasterDestination} - State, 3. Sister Generation. Path for new raster layer.}

\item{LAT_GRID}{\code{raster} - Latitude. Path to a raster layer.}

\item{LAT_CONST}{\code{number} - Latitude. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DToptimum}{\code{number} - Developmental Optimum Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DTminimum}{\code{number} - Developmental Minimum Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FAminimum}{\code{number} - Minimum Temperature for Flight Activity. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DayLength}{\code{number} - Minimum Day Length for Flight Activity. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DDminimum}{\code{number} - Minimum Thermal Sum for Infestation. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DDtotal}{\code{number} - Thermal Sum for Total Development. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{Risk_DayMax}{\code{number} - Day of Maximum Risk after Onset. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{Risk_Decay}{\code{number} - Decay of Risk after Maximum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YD_Begin}{\code{string} - Begin of Parental Development. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YD_End_Onset}{\code{string} - End of Breeding. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YD_End}{\code{string} - End of Development. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Phenips (grids, annual) (sagang:phenipsgridsannual). ---------------- Arguments ----------------  ATmean: Mean Temperature 	Argument type:	multilayer ATmax: Maximum Temperature 	Argument type:	multilayer SIrel: Solar Irradiation 	Argument type:	multilayer GENERATIONS: Potential Number of Generations 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ONSET: Onset Day of Infestation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ONSET_FILIAL_1: Onset Day, 1. Filial Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ONSET_SISTER_1: Onset Day, 1. Sister Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ONSET_FILIAL_2: Onset Day, 2. Filial Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ONSET_SISTER_2: Onset Day, 2. Sister Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ONSET_FILIAL_3: Onset Day, 3. Filial Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ONSET_SISTER_3: Onset Day, 3. Sister Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BTSUM_FILIAL_1: State, 1. Filial Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BTSUM_SISTER_1: State, 1. Sister Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BTSUM_FILIAL_2: State, 2. Filial Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BTSUM_SISTER_2: State, 2. Sister Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BTSUM_FILIAL_3: State, 3. Filial Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BTSUM_SISTER_3: State, 3. Sister Generation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LAT_GRID: Latitude (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LAT_CONST: Latitude 	Default value:	50 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DToptimum: Developmental Optimum Temperature 	Default value:	30.4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DTminimum: Developmental Minimum Temperature 	Default value:	8.3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FAminimum: Minimum Temperature for Flight Activity 	Default value:	16.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DayLength: Minimum Day Length for Flight Activity 	Default value:	14.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DDminimum: Minimum Thermal Sum for Infestation 	Default value:	140 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DDtotal: Thermal Sum for Total Development 	Default value:	557 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Risk_DayMax: Day of Maximum Risk after Onset 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Risk_Decay: Decay of Risk after Maximum 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YD_Begin: Begin of Parental Development 	Default value:	2021-04-01 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YD_End_Onset: End of Breeding 	Default value:	2021-08-31 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YD_End: End of Development 	Default value:	2021-10-31 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item BTSUM_FILIAL_1 - outputRaster - State, 1. Filial Generation
\item BTSUM_FILIAL_2 - outputRaster - State, 2. Filial Generation
\item BTSUM_FILIAL_3 - outputRaster - State, 3. Filial Generation
\item BTSUM_SISTER_1 - outputRaster - State, 1. Sister Generation
\item BTSUM_SISTER_2 - outputRaster - State, 2. Sister Generation
\item BTSUM_SISTER_3 - outputRaster - State, 3. Sister Generation
\item GENERATIONS - outputRaster - Potential Number of Generations
\item ONSET - outputRaster - Onset Day of Infestation
\item ONSET_FILIAL_1 - outputRaster - Onset Day, 1. Filial Generation
\item ONSET_FILIAL_2 - outputRaster - Onset Day, 2. Filial Generation
\item ONSET_FILIAL_3 - outputRaster - Onset Day, 3. Filial Generation
\item ONSET_SISTER_1 - outputRaster - Onset Day, 1. Sister Generation
\item ONSET_SISTER_2 - outputRaster - Onset Day, 2. Sister Generation
\item ONSET_SISTER_3 - outputRaster - Onset Day, 3. Sister Generation
}
}
}
