% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_phenipstable.R
\name{sagang_phenipstable}
\alias{sagang_phenipstable}
\title{QGIS algorithm - Phenips (table)}
\usage{
sagang_phenipstable(
  WEATHER = qgisprocess:::qgis_default_value(),
  ATmean = qgisprocess:::qgis_default_value(),
  ATmax = qgisprocess:::qgis_default_value(),
  SIrel = qgisprocess:::qgis_default_value(),
  PHENOLOGY = qgisprocess:::qgis_default_value(),
  SUMMARY = qgisprocess:::qgis_default_value(),
  LATITUDE = qgisprocess:::qgis_default_value(),
  LIMIT = qgisprocess:::qgis_default_value(),
  DToptimum = qgisprocess:::qgis_default_value(),
  DTminimum = qgisprocess:::qgis_default_value(),
  FAminimum = qgisprocess:::qgis_default_value(),
  DayLength = qgisprocess:::qgis_default_value(),
  DDminimum = qgisprocess:::qgis_default_value(),
  DDtotal = qgisprocess:::qgis_default_value(),
  Risk_DayMax = qgisprocess:::qgis_default_value(),
  Risk_Decay = qgisprocess:::qgis_default_value(),
  YD_Begin = qgisprocess:::qgis_default_value(),
  YD_End_Onset = qgisprocess:::qgis_default_value(),
  YD_End = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{WEATHER}{\code{source} - Weather Data. Path to a vector layer.}

\item{ATmean}{\code{field} - Mean Temperature. The name of an existing field. ; delimited list of existing field names.}

\item{ATmax}{\code{field} - Maximum Temperature. The name of an existing field. ; delimited list of existing field names.}

\item{SIrel}{\code{field} - Solar Irradiation. The name of an existing field. ; delimited list of existing field names.}

\item{PHENOLOGY}{\code{vectorDestination} - Phenology. Path for new vector layer.}

\item{SUMMARY}{\code{vectorDestination} - Summary. Path for new vector layer.}

\item{LATITUDE}{\code{number} - Latitude. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LIMIT}{\code{boolean} - Limit. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DToptimum}{\code{number} - Developmental Optimum Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DTminimum}{\code{number} - Developmental Minimum Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FAminimum}{\code{number} - Minimum Temperature for Flight Activity. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DayLength}{\code{number} - Minimum Day Length for Flight Activity. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DDminimum}{\code{number} - Minimum Thermal Sum for Infestation. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DDtotal}{\code{number} - Thermal Sum for Total Development. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{Risk_DayMax}{\code{number} - Day of Maximum Risk after Onset. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{Risk_Decay}{\code{number} - Decay of Risk after Maximum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YD_Begin}{\code{string} - Begin of Parental Development. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YD_End_Onset}{\code{string} - End of Breeding. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{YD_End}{\code{string} - End of Development. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Phenips (table) (sagang:phenipstable). ---------------- Arguments ----------------  WEATHER: Weather Data 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ATmean: Mean Temperature 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names ATmax: Maximum Temperature 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names SIrel: Solar Irradiation 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names PHENOLOGY: Phenology 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer SUMMARY: Summary 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer LATITUDE: Latitude 	Default value:	50 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LIMIT: Limit 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DToptimum: Developmental Optimum Temperature 	Default value:	30.4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DTminimum: Developmental Minimum Temperature 	Default value:	8.3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FAminimum: Minimum Temperature for Flight Activity 	Default value:	16.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DayLength: Minimum Day Length for Flight Activity 	Default value:	14.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DDminimum: Minimum Thermal Sum for Infestation 	Default value:	140 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DDtotal: Thermal Sum for Total Development 	Default value:	557 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Risk_DayMax: Day of Maximum Risk after Onset 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Risk_Decay: Decay of Risk after Maximum 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YD_Begin: Begin of Parental Development 	Default value:	2021-04-01 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YD_End_Onset: End of Breeding 	Default value:	2021-08-31 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression YD_End: End of Development 	Default value:	2021-10-31 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item PHENOLOGY - outputVector - Phenology
\item SUMMARY - outputVector - Summary
}
}
}
