% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_pointtolinedistances.R
\name{sagang_pointtolinedistances}
\alias{sagang_pointtolinedistances}
\title{QGIS algorithm - Point to line distances}
\usage{
sagang_pointtolinedistances(
  POINTS = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  DISTANCES = qgisprocess:::qgis_default_value(),
  LINES = qgisprocess:::qgis_default_value(),
  LINE_ID = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{RESULT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{DISTANCES}{\code{vectorDestination} - Distances. Path for new vector layer.}

\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{LINE_ID}{\code{field} - Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Point to line distances (sagang:pointtolinedistances). ---------------- Arguments ----------------  POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer RESULT: Result 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DISTANCES: Distances 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer LINES: Lines 	Argument type:	source 	Acceptable values: 		- Path to a vector layer LINE_ID: Identifier 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names
}
\details{
\subsection{Outputs description}{
\itemize{
\item DISTANCES - outputVector - Distances
\item RESULT - outputVector - Result
}
}
}
