% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_pointtopointdistances.R
\name{sagang_pointtopointdistances}
\alias{sagang_pointtopointdistances}
\title{QGIS algorithm - Point to point distances}
\usage{
sagang_pointtopointdistances(
  POINTS = qgisprocess:::qgis_default_value(),
  ID_POINTS = qgisprocess:::qgis_default_value(),
  NEAR = qgisprocess:::qgis_default_value(),
  ID_NEAR = qgisprocess:::qgis_default_value(),
  DISTANCES = qgisprocess:::qgis_default_value(),
  LINES = qgisprocess:::qgis_default_value(),
  FORMAT = qgisprocess:::qgis_default_value(),
  MAX_DIST = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{ID_POINTS}{\code{field} - Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{NEAR}{\code{source} - Near Points. Path to a vector layer.}

\item{ID_NEAR}{\code{field} - Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{DISTANCES}{\code{vectorDestination} - Distances. Path for new vector layer.}

\item{LINES}{\code{vectorDestination} - Distances as Lines. Path for new vector layer.}

\item{FORMAT}{\code{enum}  of \verb{("(0) complete input times near points matrix", "(1) each pair with a single record", "(2) find only the nearest point for each input point")} - Output Format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MAX_DIST}{\code{number} - Maximum Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Point to point distances (sagang:pointtopointdistances). ---------------- Arguments ----------------  POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ID_POINTS: Identifier 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names NEAR: Near Points (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ID_NEAR: Identifier 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names DISTANCES: Distances 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer LINES: Distances as Lines 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer FORMAT: Output Format 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) complete input times near points matrix 		- 1: (1) each pair with a single record 		- 2: (2) find only the nearest point for each input point 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' MAX_DIST: Maximum Distance 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item DISTANCES - outputVector - Distances
\item LINES - outputVector - Distances as Lines
}
}
}
