% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_polartocartesiancoordinates.R
\name{sagang_polartocartesiancoordinates}
\alias{sagang_polartocartesiancoordinates}
\title{QGIS algorithm - Polar to cartesian coordinates}
\usage{
sagang_polartocartesiancoordinates(
  POLAR = qgisprocess:::qgis_default_value(),
  F_EXAGG = qgisprocess:::qgis_default_value(),
  D_EXAGG = qgisprocess:::qgis_default_value(),
  CARTES = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  DEGREE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POLAR}{\code{source} - Polar Coordinates. Path to a vector layer.}

\item{F_EXAGG}{\code{field} - Exaggeration. The name of an existing field. ; delimited list of existing field names.}

\item{D_EXAGG}{\code{number} - Exaggeration Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CARTES}{\code{vectorDestination} - Cartesion Coordinates. Path for new vector layer.}

\item{RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEGREE}{\code{boolean} - Degree. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Polar to cartesian coordinates (sagang:polartocartesiancoordinates). ---------------- Arguments ----------------  POLAR: Polar Coordinates 	Argument type:	source 	Acceptable values: 		- Path to a vector layer F_EXAGG: Exaggeration 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names D_EXAGG: Exaggeration Factor 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CARTES: Cartesion Coordinates 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer RADIUS: Radius 	Default value:	6371000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DEGREE: Degree 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CARTES - outputVector - Cartesion Coordinates
}
}
}
