% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_polygoncategoriestogrid.R
\name{sagang_polygoncategoriestogrid}
\alias{sagang_polygoncategoriestogrid}
\title{QGIS algorithm - Polygon categories to grid}
\usage{
sagang_polygoncategoriestogrid(
  POLYGONS = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  MULTIPLE = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX =
    qgisprocess:::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess:::qgis_default_value(),
  CATEGORY = qgisprocess:::qgis_default_value(),
  COVERAGE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{FIELD}{\code{field} - Category. The name of an existing field. ; delimited list of existing field names.}

\item{METHOD}{\code{enum}  of \verb{("(0) cell wise", "(1) polygon wise")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MULTIPLE}{\code{enum}  of \verb{("(0) minimum coverage", "(1) maximum coverage")} - Multiple Polygons. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CLASSES}{\code{vectorDestination} - Classification. Path for new vector layer.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CATEGORY}{\code{rasterDestination} - Category. Path for new raster layer.}

\item{COVERAGE}{\code{rasterDestination} - Coverage. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Polygon categories to grid (sagang:polygoncategoriestogrid). ---------------- Arguments ----------------  POLYGONS: Polygons 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Category 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names METHOD: Method 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) cell wise 		- 1: (1) polygon wise 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' MULTIPLE: Multiple Polygons 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) minimum coverage 		- 1: (1) maximum coverage 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CLASSES: Classification 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX: Output extent (optional) 	Argument type:	extent 	Acceptable values: 		- A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105' 		- Path to a layer. The extent of the layer is used. TARGET_USER_SIZE: Cellsize 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CATEGORY: Category 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer COVERAGE: Coverage 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item CATEGORY - outputRaster - Category
\item CLASSES - outputVector - Classification
\item COVERAGE - outputRaster - Coverage
}
}
}
