% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_polygonclipping.R
\name{sagang_polygonclipping}
\alias{sagang_polygonclipping}
\title{QGIS algorithm - Polygon clipping}
\usage{
sagang_polygonclipping(
  CLIP = qgisprocess:::qgis_default_value(),
  S_INPUT = qgisprocess:::qgis_default_value(),
  S_OUTPUT = qgisprocess:::qgis_default_value(),
  M_INPUT = qgisprocess:::qgis_default_value(),
  M_OUTPUT = qgisprocess:::qgis_default_value(),
  DISSOLVE = qgisprocess:::qgis_default_value(),
  MULTIPLE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CLIP}{\code{source} - Clip Features. Path to a vector layer.}

\item{S_INPUT}{\code{source} - Input Features. Path to a vector layer.}

\item{S_OUTPUT}{\code{vectorDestination} - Output Features. Path for new vector layer.}

\item{M_INPUT}{\code{multilayer} - Input Features. .}

\item{M_OUTPUT}{\code{vectorDestination} - Output Features. Path for new vector layer.}

\item{DISSOLVE}{\code{boolean} - Dissolve Clip Features. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MULTIPLE}{\code{boolean} - Multiple Input Features. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Polygon clipping (sagang:polygonclipping). ---------------- Arguments ----------------  CLIP: Clip Features 	Argument type:	source 	Acceptable values: 		- Path to a vector layer S_INPUT: Input Features 	Argument type:	source 	Acceptable values: 		- Path to a vector layer S_OUTPUT: Output Features 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer M_INPUT: Input Features 	Argument type:	multilayer M_OUTPUT: Output Features 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DISSOLVE: Dissolve Clip Features 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MULTIPLE: Multiple Input Features 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item M_OUTPUT - outputVector - Output Features
\item S_OUTPUT - outputVector - Output Features
}
}
}
