% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_polygongeneralization.R
\name{sagang_polygongeneralization}
\alias{sagang_polygongeneralization}
\title{QGIS algorithm - Polygon generalization}
\usage{
sagang_polygongeneralization(
  POLYGONS = qgisprocess:::qgis_default_value(),
  GENERALIZED = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POLYGONS}{\code{source} - Shapes. Path to a vector layer.}

\item{GENERALIZED}{\code{vectorDestination} - Shape Indices. Path for new vector layer.}

\item{THRESHOLD}{\code{number} - Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Polygon generalization (sagang:polygongeneralization). ---------------- Arguments ----------------  POLYGONS: Shapes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer GENERALIZED: Shape Indices 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer THRESHOLD: Threshold 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item GENERALIZED - outputVector - Shape Indices
}
}
}
