% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_polygonshapeindices.R
\name{sagang_polygonshapeindices}
\alias{sagang_polygonshapeindices}
\title{QGIS algorithm - Polygon shape indices}
\usage{
sagang_polygonshapeindices(
  SHAPES = qgisprocess:::qgis_default_value(),
  INDEX = qgisprocess:::qgis_default_value(),
  DMAX = qgisprocess:::qgis_default_value(),
  GYROS = qgisprocess:::qgis_default_value(),
  FERET = qgisprocess:::qgis_default_value(),
  FERET_DIRS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{INDEX}{\code{vectorDestination} - Shape Indices. Path for new vector layer.}

\item{DMAX}{\code{vectorDestination} - Maximum Diameter. Path for new vector layer.}

\item{GYROS}{\code{boolean} - Diameter of Gyration. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FERET}{\code{boolean} - Feret Diameters. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FERET_DIRS}{\code{number} - Number of Directions. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Polygon shape indices (sagang:polygonshapeindices). ---------------- Arguments ----------------  SHAPES: Shapes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer INDEX: Shape Indices 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DMAX: Maximum Diameter 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer GYROS: Diameter of Gyration 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FERET: Feret Diameters 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FERET_DIRS: Number of Directions 	Default value:	18 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item DMAX - outputVector - Maximum Diameter
\item INDEX - outputVector - Shape Indices
}
}
}
