% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_polygonupdate.R
\name{sagang_polygonupdate}
\alias{sagang_polygonupdate}
\title{QGIS algorithm - Polygon update}
\usage{
sagang_polygonupdate(
  A = qgisprocess:::qgis_default_value(),
  B = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  SPLIT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{A}{\code{source} - Layer A. Path to a vector layer.}

\item{B}{\code{source} - Layer B. Path to a vector layer.}

\item{RESULT}{\code{vectorDestination} - Update. Path for new vector layer.}

\item{SPLIT}{\code{boolean} - Split Parts. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Polygon update (sagang:polygonupdate). ---------------- Arguments ----------------  A: Layer A 	Argument type:	source 	Acceptable values: 		- Path to a vector layer B: Layer B 	Argument type:	source 	Acceptable values: 		- Path to a vector layer RESULT: Update 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer SPLIT: Split Parts 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Update
}
}
}
