% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_potentialincomingsolarradiation.R
\name{sagang_potentialincomingsolarradiation}
\alias{sagang_potentialincomingsolarradiation}
\title{QGIS algorithm - Potential incoming solar radiation}
\usage{
sagang_potentialincomingsolarradiation(
  GRD_DEM = qgisprocess:::qgis_default_value(),
  GRD_SVF = qgisprocess:::qgis_default_value(),
  GRD_VAPOUR = qgisprocess:::qgis_default_value(),
  GRD_VAPOUR_DEFAULT = qgisprocess:::qgis_default_value(),
  GRD_LINKE = qgisprocess:::qgis_default_value(),
  GRD_LINKE_DEFAULT = qgisprocess:::qgis_default_value(),
  GRD_DIRECT = qgisprocess:::qgis_default_value(),
  GRD_DIFFUS = qgisprocess:::qgis_default_value(),
  GRD_TOTAL = qgisprocess:::qgis_default_value(),
  GRD_RATIO = qgisprocess:::qgis_default_value(),
  GRD_FLAT = qgisprocess:::qgis_default_value(),
  GRD_DURATION = qgisprocess:::qgis_default_value(),
  GRD_SUNRISE = qgisprocess:::qgis_default_value(),
  GRD_SUNSET = qgisprocess:::qgis_default_value(),
  SOLARCONST = qgisprocess:::qgis_default_value(),
  LOCALSVF = qgisprocess:::qgis_default_value(),
  UNITS = qgisprocess:::qgis_default_value(),
  SHADOW = qgisprocess:::qgis_default_value(),
  LOCATION = qgisprocess:::qgis_default_value(),
  LATITUDE = qgisprocess:::qgis_default_value(),
  PERIOD = qgisprocess:::qgis_default_value(),
  DAY = qgisprocess:::qgis_default_value(),
  DAY_STOP = qgisprocess:::qgis_default_value(),
  DAYS_STEP = qgisprocess:::qgis_default_value(),
  MOMENT = qgisprocess:::qgis_default_value(),
  HOUR_STEP = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ATMOSPHERE = qgisprocess:::qgis_default_value(),
  PRESSURE = qgisprocess:::qgis_default_value(),
  WATER = qgisprocess:::qgis_default_value(),
  DUST = qgisprocess:::qgis_default_value(),
  LUMPED = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRD_DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{GRD_SVF}{\code{raster} - Sky View Factor. Path to a raster layer.}

\item{GRD_VAPOUR}{\code{raster} - Water Vapour Pressure (mbar). Path to a raster layer.}

\item{GRD_VAPOUR_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRD_LINKE}{\code{raster} - Linke Turbidity Coefficient. Path to a raster layer.}

\item{GRD_LINKE_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRD_DIRECT}{\code{rasterDestination} - Direct Insolation. Path for new raster layer.}

\item{GRD_DIFFUS}{\code{rasterDestination} - Diffuse Insolation. Path for new raster layer.}

\item{GRD_TOTAL}{\code{rasterDestination} - Total Insolation. Path for new raster layer.}

\item{GRD_RATIO}{\code{rasterDestination} - Direct to Diffuse Ratio. Path for new raster layer.}

\item{GRD_FLAT}{\code{rasterDestination} - Compare to Flat Terrain. Path for new raster layer.}

\item{GRD_DURATION}{\code{rasterDestination} - Duration of Insolation. Path for new raster layer.}

\item{GRD_SUNRISE}{\code{rasterDestination} - Sunrise. Path for new raster layer.}

\item{GRD_SUNSET}{\code{rasterDestination} - Sunset. Path for new raster layer.}

\item{SOLARCONST}{\code{number} - Solar Constant (W \code{/} m_). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LOCALSVF}{\code{boolean} - Local Sky View Factor. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{UNITS}{\code{enum}  of \verb{("(0) kWh / m_", "(1) kJ / m_", "(2) J / cm_")} - Units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SHADOW}{\code{enum}  of \verb{("(0) slim", "(1) fat", "(2) none")} - Shadow. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LOCATION}{\code{enum}  of \verb{("(0) constant latitude", "(1) calculate from grid system")} - Location. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LATITUDE}{\code{number} - Latitude. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PERIOD}{\code{enum}  of \verb{("(0) moment", "(1) day", "(2) range of days")} - Time Period. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DAY}{\code{string} - Day. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DAY_STOP}{\code{string} - Last Day. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DAYS_STEP}{\code{number} - Resolution (d). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MOMENT}{\code{number} - Moment (h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{HOUR_STEP}{\code{number} - Resolution (h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) Height of Atmosphere and Vapour Pressure", "(1) Air Pressure, Water and Dust Content", "(2) Lumped Atmospheric Transmittance", "(3) Hofierka and Suri")} - Atmospheric Effects. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ATMOSPHERE}{\code{number} - Height of Atmosphere (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PRESSURE}{\code{number} - Barometric Pressure (mbar). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WATER}{\code{number} - Water Content (cm). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DUST}{\code{number} - Dust (ppm). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LUMPED}{\code{number} - Lumped Atmospheric Transmittance (Percent). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Potential incoming solar radiation (sagang:potentialincomingsolarradiation). ---------------- Arguments ----------------  GRD_DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GRD_SVF: Sky View Factor (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GRD_VAPOUR: Water Vapour Pressure (mbar) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GRD_VAPOUR_DEFAULT: Default 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GRD_LINKE: Linke Turbidity Coefficient (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GRD_LINKE_DEFAULT: Default 	Default value:	3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GRD_DIRECT: Direct Insolation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer GRD_DIFFUS: Diffuse Insolation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer GRD_TOTAL: Total Insolation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer GRD_RATIO: Direct to Diffuse Ratio 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer GRD_FLAT: Compare to Flat Terrain 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer GRD_DURATION: Duration of Insolation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer GRD_SUNRISE: Sunrise 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer GRD_SUNSET: Sunset 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SOLARCONST: Solar Constant (W / m_) 	Default value:	1367 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LOCALSVF: Local Sky View Factor 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression UNITS: Units 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) kWh / m_ 		- 1: (1) kJ / m_ 		- 2: (2) J / cm_ 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SHADOW: Shadow 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) slim 		- 1: (1) fat 		- 2: (2) none 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' LOCATION: Location 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) constant latitude 		- 1: (1) calculate from grid system 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' LATITUDE: Latitude 	Default value:	53 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PERIOD: Time Period 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) moment 		- 1: (1) day 		- 2: (2) range of days 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DAY: Day 	Default value:	2021-03-24 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DAY_STOP: Last Day 	Default value:	2021-03-24 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DAYS_STEP: Resolution (d) 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MOMENT: Moment (h) 	Default value:	12 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression HOUR_STEP: Resolution (h) 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Atmospheric Effects 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) Height of Atmosphere and Vapour Pressure 		- 1: (1) Air Pressure, Water and Dust Content 		- 2: (2) Lumped Atmospheric Transmittance 		- 3: (3) Hofierka and Suri 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' ATMOSPHERE: Height of Atmosphere (m) 	Default value:	12000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PRESSURE: Barometric Pressure (mbar) 	Default value:	1013 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WATER: Water Content (cm) 	Default value:	1.68 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DUST: Dust (ppm) 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LUMPED: Lumped Atmospheric Transmittance (Percent) 	Default value:	70 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRD_DIFFUS - outputRaster - Diffuse Insolation
\item GRD_DIRECT - outputRaster - Direct Insolation
\item GRD_DURATION - outputRaster - Duration of Insolation
\item GRD_FLAT - outputRaster - Compare to Flat Terrain
\item GRD_RATIO - outputRaster - Direct to Diffuse Ratio
\item GRD_SUNRISE - outputRaster - Sunrise
\item GRD_SUNSET - outputRaster - Sunset
\item GRD_TOTAL - outputRaster - Total Insolation
}
}
}
