% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_principalcomponentanalysis.R
\name{sagang_principalcomponentanalysis}
\alias{sagang_principalcomponentanalysis}
\title{QGIS algorithm - Principal component analysis}
\usage{
sagang_principalcomponentanalysis(
  GRIDS = qgisprocess:::qgis_default_value(),
  PCA = qgisprocess:::qgis_default_value(),
  EIGEN_INPUT = qgisprocess:::qgis_default_value(),
  EIGEN = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  COMPONENTS = qgisprocess:::qgis_default_value(),
  OVERWRITE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{PCA}{\code{rasterDestination} - Principal Components. Path for new raster layer.}

\item{EIGEN_INPUT}{\code{source} - Eigen Vectors. Path to a vector layer.}

\item{EIGEN}{\code{vectorDestination} - Eigen Vectors. Path for new vector layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) correlation matrix", "(1) variance-covariance matrix", "(2) sums-of-squares-and-cross-products matrix")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{COMPONENTS}{\code{number} - Number of Components. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OVERWRITE}{\code{boolean} - Overwrite Previous Results. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Principal component analysis (sagang:principalcomponentanalysis). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	multilayer PCA: Principal Components 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer EIGEN_INPUT: Eigen Vectors (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer EIGEN: Eigen Vectors 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer METHOD: Method 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) correlation matrix 		- 1: (1) variance-covariance matrix 		- 2: (2) sums-of-squares-and-cross-products matrix 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' COMPONENTS: Number of Components 	Default value:	3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OVERWRITE: Overwrite Previous Results 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item EIGEN - outputVector - Eigen Vectors
\item PCA - outputRaster - Principal Components
}
}
}
