% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_principalcomponentbasedimagesharpening.R
\name{sagang_principalcomponentbasedimagesharpening}
\alias{sagang_principalcomponentbasedimagesharpening}
\title{QGIS algorithm - Principal component based image sharpening}
\usage{
sagang_principalcomponentbasedimagesharpening(
  GRIDS = qgisprocess:::qgis_default_value(),
  PAN = qgisprocess:::qgis_default_value(),
  SHARPEN = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  PAN_MATCH = qgisprocess:::qgis_default_value(),
  OVERWRITE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Original Channels. .}

\item{PAN}{\code{raster} - Panchromatic Channel. Path to a raster layer.}

\item{SHARPEN}{\code{rasterDestination} - Sharpened Channels. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) correlation matrix", "(1) variance-covariance matrix", "(2) sums-of-squares-and-cross-products matrix")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESAMPLING}{\code{enum}  of \verb{("(0) nearest neighbour", "(1) bilinear", "(2) cubic convolution")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PAN_MATCH}{\code{enum}  of \verb{("(0) normalized", "(1) standardized")} - Panchromatic Channel Matching. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OVERWRITE}{\code{boolean} - Overwrite. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Principal component based image sharpening (sagang:principalcomponentbasedimagesharpening). ---------------- Arguments ----------------  GRIDS: Original Channels 	Argument type:	multilayer PAN: Panchromatic Channel 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SHARPEN: Sharpened Channels 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Method 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) correlation matrix 		- 1: (1) variance-covariance matrix 		- 2: (2) sums-of-squares-and-cross-products matrix 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RESAMPLING: Resampling 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) nearest neighbour 		- 1: (1) bilinear 		- 2: (2) cubic convolution 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' PAN_MATCH: Panchromatic Channel Matching 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) normalized 		- 1: (1) standardized 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' OVERWRITE: Overwrite 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item SHARPEN - outputRaster - Sharpened Channels
}
}
}
