% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_profilesfromlines.R
\name{sagang_profilesfromlines}
\alias{sagang_profilesfromlines}
\title{QGIS algorithm - Profiles from lines}
\usage{
sagang_profilesfromlines(
  DEM = qgisprocess:::qgis_default_value(),
  VALUES = qgisprocess:::qgis_default_value(),
  LINES = qgisprocess:::qgis_default_value(),
  NAME = qgisprocess:::qgis_default_value(),
  PROFILE = qgisprocess:::qgis_default_value(),
  PROFILES = qgisprocess:::qgis_default_value(),
  SPLIT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{VALUES}{\code{multilayer} - Values. .}

\item{LINES}{\code{source} - Lines. Path to a vector layer.}

\item{NAME}{\code{field} - Name. The name of an existing field. ; delimited list of existing field names.}

\item{PROFILE}{\code{vectorDestination} - Profiles. Path for new vector layer.}

\item{PROFILES}{\code{vectorDestination} - Profiles. Path for new vector layer.}

\item{SPLIT}{\code{boolean} - Each Line as new Profile. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Profiles from lines (sagang:profilesfromlines). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer VALUES: Values (optional) 	Argument type:	multilayer LINES: Lines 	Argument type:	source 	Acceptable values: 		- Path to a vector layer NAME: Name 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names PROFILE: Profiles 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer PROFILES: Profiles 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer SPLIT: Each Line as new Profile 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item PROFILE - outputVector - Profiles
\item PROFILES - outputVector - Profiles
}
}
}
