% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_proximityraster.R
\name{sagang_proximityraster}
\alias{sagang_proximityraster}
\title{QGIS algorithm - Proximity raster}
\usage{
sagang_proximityraster(
  FEATURES = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  DIRECTION = qgisprocess:::qgis_default_value(),
  ALLOCATION = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FEATURES}{\code{raster} - Features. Path to a raster layer.}

\item{DISTANCE}{\code{rasterDestination} - Distance. Path for new raster layer.}

\item{DIRECTION}{\code{rasterDestination} - Direction. Path for new raster layer.}

\item{ALLOCATION}{\code{rasterDestination} - Allocation. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Proximity raster (sagang:proximityraster). ---------------- Arguments ----------------  FEATURES: Features 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DISTANCE: Distance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIRECTION: Direction 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ALLOCATION: Allocation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item ALLOCATION - outputRaster - Allocation
\item DIRECTION - outputRaster - Direction
\item DISTANCE - outputRaster - Distance
}
}
}
