% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_quasidynamicflowaccumulation.R
\name{sagang_quasidynamicflowaccumulation}
\alias{sagang_quasidynamicflowaccumulation}
\title{QGIS algorithm - Quasi-dynamic flow accumulation}
\usage{
sagang_quasidynamicflowaccumulation(
  DEM = qgisprocess:::qgis_default_value(),
  FLOW_K = qgisprocess:::qgis_default_value(),
  FLOW_K_DEFAULT = qgisprocess:::qgis_default_value(),
  FLOW_ACC = qgisprocess:::qgis_default_value(),
  FLOW = qgisprocess:::qgis_default_value(),
  TIME_MEAN = qgisprocess:::qgis_default_value(),
  TIME_CONC = qgisprocess:::qgis_default_value(),
  VELOCITY = qgisprocess:::qgis_default_value(),
  TIME = qgisprocess:::qgis_default_value(),
  ROUTING = qgisprocess:::qgis_default_value(),
  FLOW_DEPTH = qgisprocess:::qgis_default_value(),
  FLOW_CONST = qgisprocess:::qgis_default_value(),
  RESET = qgisprocess:::qgis_default_value(),
  RAIN = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{FLOW_K}{\code{raster} - Manning-Strickler Coefficient. Path to a raster layer.}

\item{FLOW_K_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FLOW_ACC}{\code{rasterDestination} - Flow Accumulation. Path for new raster layer.}

\item{FLOW}{\code{rasterDestination} - Flow Through. Path for new raster layer.}

\item{TIME_MEAN}{\code{rasterDestination} - Flow Travel Time. Path for new raster layer.}

\item{TIME_CONC}{\code{rasterDestination} - Flow Concentration. Path for new raster layer.}

\item{VELOCITY}{\code{rasterDestination} - Flow Velocity. Path for new raster layer.}

\item{TIME}{\code{number} - Time. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ROUTING}{\code{enum}  of \verb{("(0) D8", "(1) MFD")} - Flow Routing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FLOW_DEPTH}{\code{enum}  of \verb{("(0) constant", "(1) dynamic")} - Flow Depth. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FLOW_CONST}{\code{number} - Constant Flow Depth. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RESET}{\code{boolean} - Reset. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RAIN}{\code{number} - Rain. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Quasi-dynamic flow accumulation (sagang:quasidynamicflowaccumulation). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FLOW_K: Manning-Strickler Coefficient (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FLOW_K_DEFAULT: Default 	Default value:	20 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FLOW_ACC: Flow Accumulation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FLOW: Flow Through 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TIME_MEAN: Flow Travel Time 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TIME_CONC: Flow Concentration 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VELOCITY: Flow Velocity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TIME: Time 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ROUTING: Flow Routing 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) D8 		- 1: (1) MFD 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FLOW_DEPTH: Flow Depth 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) constant 		- 1: (1) dynamic 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FLOW_CONST: Constant Flow Depth 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RESET: Reset 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RAIN: Rain 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLOW - outputRaster - Flow Through
\item FLOW_ACC - outputRaster - Flow Accumulation
\item TIME_CONC - outputRaster - Flow Concentration
\item TIME_MEAN - outputRaster - Flow Travel Time
\item VELOCITY - outputRaster - Flow Velocity
}
}
}
