% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_raosqdiversityindex.R
\name{sagang_raosqdiversityindex}
\alias{sagang_raosqdiversityindex}
\title{QGIS algorithm - Rao's q diversity index}
\usage{
sagang_raosqdiversityindex(
  VALUES = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  INDEX = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  LAMBDA = qgisprocess:::qgis_default_value(),
  KERNEL_TYPE = qgisprocess:::qgis_default_value(),
  KERNEL_RADIUS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{VALUES}{\code{multilayer} - Values. .}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INDEX}{\code{rasterDestination} - Rao's Q. Path for new raster layer.}

\item{DISTANCE}{\code{enum}  of \verb{("(0) Euclidian", "(1) Manhatten", "(2) Canberra", "(3) Minkowski")} - Distance. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LAMBDA}{\code{number} - Lambda. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KERNEL_TYPE}{\code{enum}  of \verb{("(0) Square", "(1) Circle")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Rao's q diversity index (sagang:raosqdiversityindex). ---------------- Arguments ----------------  VALUES: Values 	Argument type:	multilayer NORMALIZE: Normalize 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression INDEX: Rao's Q 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DISTANCE: Distance 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Euclidian 		- 1: (1) Manhatten 		- 2: (2) Canberra 		- 3: (3) Minkowski 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' LAMBDA: Lambda 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KERNEL_TYPE: Kernel Type 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Square 		- 1: (1) Circle 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' KERNEL_RADIUS: Radius 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item INDEX - outputRaster - Rao's Q
}
}
}
