% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_rastermasking.R
\name{sagang_rastermasking}
\alias{sagang_rastermasking}
\title{QGIS algorithm - Raster masking}
\usage{
sagang_rastermasking(
  LIST = qgisprocess:::qgis_default_value(),
  GRID = qgisprocess:::qgis_default_value(),
  MASKED = qgisprocess:::qgis_default_value(),
  GRIDS = qgisprocess:::qgis_default_value(),
  GRIDS_CREATE = qgisprocess:::qgis_default_value(),
  GRIDS_MASKED = qgisprocess:::qgis_default_value(),
  MASK = qgisprocess:::qgis_default_value(),
  NODATA = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LIST}{\code{boolean} - List Processing. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{MASKED}{\code{rasterDestination} - Masked Grid. Path for new raster layer.}

\item{GRIDS}{\code{multilayer} - Grids. .}

\item{GRIDS_CREATE}{\code{boolean} - Create Copies. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRIDS_MASKED}{\code{rasterDestination} - Masked Grids. Path for new raster layer.}

\item{MASK}{\code{raster} - Mask. Path to a raster layer.}

\item{NODATA}{\code{enum}  of \verb{("(0) no-data cells", "(1) data cells")} - Mask Cells. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Raster masking (sagang:rastermasking). ---------------- Arguments ----------------  LIST: List Processing 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MASKED: Masked Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer GRIDS: Grids 	Argument type:	multilayer GRIDS_CREATE: Create Copies 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GRIDS_MASKED: Masked Grids 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MASK: Mask 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer NODATA: Mask Cells 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) no-data cells 		- 1: (1) data cells 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRIDS_MASKED - outputRaster - Masked Grids
\item MASKED - outputRaster - Masked Grid
}
}
}
