% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_rasterproximitybuffer.R
\name{sagang_rasterproximitybuffer}
\alias{sagang_rasterproximitybuffer}
\title{QGIS algorithm - Raster proximity buffer}
\usage{
sagang_rasterproximitybuffer(
  SOURCE = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  ALLOC = qgisprocess:::qgis_default_value(),
  BUFFER = qgisprocess:::qgis_default_value(),
  DIST = qgisprocess:::qgis_default_value(),
  IVAL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SOURCE}{\code{raster} - Source Grid. Path to a raster layer.}

\item{DISTANCE}{\code{rasterDestination} - Distance Grid. Path for new raster layer.}

\item{ALLOC}{\code{rasterDestination} - Allocation Grid. Path for new raster layer.}

\item{BUFFER}{\code{rasterDestination} - Buffer Grid. Path for new raster layer.}

\item{DIST}{\code{number} - Buffer distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{IVAL}{\code{number} - Equidistance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Raster proximity buffer (sagang:rasterproximitybuffer). ---------------- Arguments ----------------  SOURCE: Source Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DISTANCE: Distance Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ALLOC: Allocation Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer BUFFER: Buffer Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIST: Buffer distance 	Default value:	500 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression IVAL: Equidistance 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item ALLOC - outputRaster - Allocation Grid
\item BUFFER - outputRaster - Buffer Grid
\item DISTANCE - outputRaster - Distance Grid
}
}
}
