% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_rasterskeletonization.R
\name{sagang_rasterskeletonization}
\alias{sagang_rasterskeletonization}
\title{QGIS algorithm - Raster skeletonization}
\usage{
sagang_rasterskeletonization(
  INPUT = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  VECTOR = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  INIT_METHOD = qgisprocess:::qgis_default_value(),
  INIT_THRESHOLD = qgisprocess:::qgis_default_value(),
  CONVERGENCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - Skeleton. Path for new raster layer.}

\item{VECTOR}{\code{vectorDestination} - Skeleton. Path for new vector layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) Standard", "(1) Hilditch's Algorithm", "(2) Channel Skeleton")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{INIT_METHOD}{\code{enum}  of \verb{("(0) Less than", "(1) Greater than")} - Initialisation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{INIT_THRESHOLD}{\code{number} - Threshold (Init.). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CONVERGENCE}{\code{number} - Convergence. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Raster skeletonization (sagang:rasterskeletonization). ---------------- Arguments ----------------  INPUT: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RESULT: Skeleton 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VECTOR: Skeleton 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Standard 		- 1: (1) Hilditch's Algorithm 		- 2: (2) Channel Skeleton 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' INIT_METHOD: Initialisation 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Less than 		- 1: (1) Greater than 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' INIT_THRESHOLD: Threshold (Init.) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CONVERGENCE: Convergence 	Default value:	3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Skeleton
\item VECTOR - outputVector - Skeleton
}
}
}
