% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_rasterstatisticsforpolygons.R
\name{sagang_rasterstatisticsforpolygons}
\alias{sagang_rasterstatisticsforpolygons}
\title{QGIS algorithm - Raster statistics for polygons}
\usage{
sagang_rasterstatisticsforpolygons(
  GRIDS = qgisprocess:::qgis_default_value(),
  POLYGONS = qgisprocess:::qgis_default_value(),
  NAMING = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  PARALLELIZED = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  COUNT = qgisprocess:::qgis_default_value(),
  MIN = qgisprocess:::qgis_default_value(),
  MAX = qgisprocess:::qgis_default_value(),
  RANGE = qgisprocess:::qgis_default_value(),
  SUM = qgisprocess:::qgis_default_value(),
  MEAN = qgisprocess:::qgis_default_value(),
  VAR = qgisprocess:::qgis_default_value(),
  STDDEV = qgisprocess:::qgis_default_value(),
  GINI = qgisprocess:::qgis_default_value(),
  QUANTILES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{NAMING}{\code{enum}  of \verb{("(0) grid number", "(1) grid name")} - Field Naming. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) simple and fast", "(1) polygon wise (cell centers)", "(2) polygon wise (cell area)", "(3) polygon wise (cell area weighted)")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PARALLELIZED}{\code{boolean} - Use Multiple Cores. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RESULT}{\code{vectorDestination} - Statistics. Path for new vector layer.}

\item{COUNT}{\code{boolean} - Number of Cells. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MIN}{\code{boolean} - Minimum. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAX}{\code{boolean} - Maximum. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RANGE}{\code{boolean} - Range. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SUM}{\code{boolean} - Sum. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MEAN}{\code{boolean} - Mean. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{VAR}{\code{boolean} - Variance. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{STDDEV}{\code{boolean} - Standard Deviation. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GINI}{\code{boolean} - Gini. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{QUANTILES}{\code{string} - Percentiles. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Raster statistics for polygons (sagang:rasterstatisticsforpolygons). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	multilayer POLYGONS: Polygons 	Argument type:	source 	Acceptable values: 		- Path to a vector layer NAMING: Field Naming 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) grid number 		- 1: (1) grid name 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) simple and fast 		- 1: (1) polygon wise (cell centers) 		- 2: (2) polygon wise (cell area) 		- 3: (3) polygon wise (cell area weighted) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' PARALLELIZED: Use Multiple Cores 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RESULT: Statistics 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer COUNT: Number of Cells 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MIN: Minimum 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAX: Maximum 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RANGE: Range 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SUM: Sum 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MEAN: Mean 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression VAR: Variance 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression STDDEV: Standard Deviation 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GINI: Gini 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression QUANTILES: Percentiles 	Default value:	 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Statistics
}
}
}
