% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sagang_regressionanalysispointsandpredictorgrid.R
\name{sagang_regressionanalysispointsandpredictorgrid}
\alias{sagang_regressionanalysispointsandpredictorgrid}
\title{QGIS algorithm - Regression analysis (points and predictor grid)}
\usage{
sagang_regressionanalysispointsandpredictorgrid(
  PREDICTOR = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  ATTRIBUTE = qgisprocess:::qgis_default_value(),
  REGRESSION = qgisprocess:::qgis_default_value(),
  RESIDUAL = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{PREDICTOR}{\code{raster} - Predictor. Path to a raster layer.}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{ATTRIBUTE}{\code{field} - Dependent Variable. The name of an existing field. ; delimited list of existing field names.}

\item{REGRESSION}{\code{rasterDestination} - Regression. Path for new raster layer.}

\item{RESIDUAL}{\code{vectorDestination} - Residuals. Path for new vector layer.}

\item{RESAMPLING}{\code{enum}  of \verb{("(0) Nearest Neighbour", "(1) Bilinear Interpolation", "(2) Bicubic Spline Interpolation", "(3) B-Spline Interpolation")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) Y = a + b * X (linear)", "(1) Y = a + b / X", "(2) Y = a / (b - X)", "(3) Y = a * X^b (power)", "(4) Y = a e^(b * X) (exponential)", "(5) Y = a + b * ln(X) (logarithmic)")} - Regression Function. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Regression analysis (points and predictor grid) (sagang:regressionanalysispointsandpredictorgrid). ---------------- Arguments ----------------  PREDICTOR: Predictor 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ATTRIBUTE: Dependent Variable 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names REGRESSION: Regression 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RESIDUAL: Residuals 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer RESAMPLING: Resampling 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) Nearest Neighbour 		- 1: (1) Bilinear Interpolation 		- 2: (2) Bicubic Spline Interpolation 		- 3: (3) B-Spline Interpolation 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Regression Function 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Y = a + b * X (linear) 		- 1: (1) Y = a + b / X 		- 2: (2) Y = a / (b - X) 		- 3: (3) Y = a * X^b (power) 		- 4: (4) Y = a e^(b * X) (exponential) 		- 5: (5) Y = a + b * ln(X) (logarithmic) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item REGRESSION - outputRaster - Regression
\item RESIDUAL - outputVector - Residuals
}
}
}
