% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_relativeheightsandslopepositions.R
\name{sagang_relativeheightsandslopepositions}
\alias{sagang_relativeheightsandslopepositions}
\title{QGIS algorithm - Relative heights and slope positions}
\usage{
sagang_relativeheightsandslopepositions(
  DEM = qgisprocess:::qgis_default_value(),
  HO = qgisprocess:::qgis_default_value(),
  HU = qgisprocess:::qgis_default_value(),
  NH = qgisprocess:::qgis_default_value(),
  SH = qgisprocess:::qgis_default_value(),
  MS = qgisprocess:::qgis_default_value(),
  W = qgisprocess:::qgis_default_value(),
  T = qgisprocess:::qgis_default_value(),
  E = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{HO}{\code{rasterDestination} - Slope Height. Path for new raster layer.}

\item{HU}{\code{rasterDestination} - Valley Depth. Path for new raster layer.}

\item{NH}{\code{rasterDestination} - Normalized Height. Path for new raster layer.}

\item{SH}{\code{rasterDestination} - Standardized Height. Path for new raster layer.}

\item{MS}{\code{rasterDestination} - Mid-Slope Positon. Path for new raster layer.}

\item{W}{\code{number} - w. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{T}{\code{number} - t. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{E}{\code{number} - e. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Relative heights and slope positions (sagang:relativeheightsandslopepositions). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer HO: Slope Height 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer HU: Valley Depth 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer NH: Normalized Height 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SH: Standardized Height 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MS: Mid-Slope Positon 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer W: w 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression T: t 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression E: e 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item HO - outputRaster - Slope Height
\item HU - outputRaster - Valley Depth
\item MS - outputRaster - Mid-Slope Positon
\item NH - outputRaster - Normalized Height
\item SH - outputRaster - Standardized Height
}
}
}
