% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_removeduplicatepoints.R
\name{sagang_removeduplicatepoints}
\alias{sagang_removeduplicatepoints}
\title{QGIS algorithm - Remove duplicate points}
\usage{
sagang_removeduplicatepoints(
  POINTS = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  NUMERIC = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{RESULT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{NUMERIC}{\code{enum}  of \verb{("(0) take values from the point to be kept", "(1) minimum values of all duplicates", "(2) maximum values of all duplicates", "(3) mean values of all duplicates")} - Value Aggregation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) first point", "(1) last point", "(2) point with minimum attribute value", "(3) point with maximum attribute value")} - Point to Keep. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Remove duplicate points (sagang:removeduplicatepoints). ---------------- Arguments ----------------  POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Attribute 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names RESULT: Result 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer NUMERIC: Value Aggregation 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) take values from the point to be kept 		- 1: (1) minimum values of all duplicates 		- 2: (2) maximum values of all duplicates 		- 3: (3) mean values of all duplicates 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Point to Keep 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) first point 		- 1: (1) last point 		- 2: (2) point with minimum attribute value 		- 3: (3) point with maximum attribute value 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Result
}
}
}
