% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_removesmallpixelclumpstonodata.R
\name{sagang_removesmallpixelclumpstonodata}
\alias{sagang_removesmallpixelclumpstonodata}
\title{QGIS algorithm - Remove small pixel clumps (to no-data)}
\usage{
sagang_removesmallpixelclumpstonodata(
  GRID = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Input Grid. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Filtered Grid. Path for new raster layer.}

\item{THRESHOLD}{\code{number} - Min. Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Remove small pixel clumps (to no-data) (sagang:removesmallpixelclumpstonodata). ---------------- Arguments ----------------  GRID: Input Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OUTPUT: Filtered Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer THRESHOLD: Min. Size 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Filtered Grid
}
}
}
