% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_resamplingfilter.R
\name{sagang_resamplingfilter}
\alias{sagang_resamplingfilter}
\title{QGIS algorithm - Resampling filter}
\usage{
sagang_resamplingfilter(
  GRID = qgisprocess:::qgis_default_value(),
  LOPASS = qgisprocess:::qgis_default_value(),
  HIPASS = qgisprocess:::qgis_default_value(),
  SCALE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{LOPASS}{\code{rasterDestination} - Low Pass Filter. Path for new raster layer.}

\item{HIPASS}{\code{rasterDestination} - High Pass Filter. Path for new raster layer.}

\item{SCALE}{\code{number} - Scale Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Resampling filter (sagang:resamplingfilter). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LOPASS: Low Pass Filter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer HIPASS: High Pass Filter 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SCALE: Scale Factor 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item HIPASS - outputRaster - High Pass Filter
\item LOPASS - outputRaster - Low Pass Filter
}
}
}
