% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_rivergridgeneration.R
\name{sagang_rivergridgeneration}
\alias{sagang_rivergridgeneration}
\title{QGIS algorithm - Rivergridgeneration}
\usage{
sagang_rivergridgeneration(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  SX = qgisprocess:::qgis_default_value(),
  SY = qgisprocess:::qgis_default_value(),
  MX = qgisprocess:::qgis_default_value(),
  MY = qgisprocess:::qgis_default_value(),
  Owrite = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Gel_ndemodell (DTM). Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - HG Raster. Path for new raster layer.}

\item{SX}{\code{number} - Abflusspfad-Quelle, x-Wert. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SY}{\code{number} - Abflusspfad-Quelle, y-Wert. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MX}{\code{number} - Abflusspfad-M_ndung, x-Wert. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MY}{\code{number} - Abflusspfad-M_ndung, y-Wert. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{Owrite}{\code{boolean} - Overwrite RiverGridCells. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Rivergridgeneration (sagang:rivergridgeneration). ---------------- Arguments ----------------  INPUT: Gel_ndemodell (DTM) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OUTPUT: HG Raster 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SX: Abflusspfad-Quelle, x-Wert 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SY: Abflusspfad-Quelle, y-Wert 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MX: Abflusspfad-M_ndung, x-Wert 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MY: Abflusspfad-M_ndung, y-Wert 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression Owrite: Overwrite RiverGridCells 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - HG Raster
}
}
}
