% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_safetyfactor.R
\name{sagang_safetyfactor}
\alias{sagang_safetyfactor}
\title{QGIS algorithm - Safetyfactor}
\usage{
sagang_safetyfactor(
  A = qgisprocess:::qgis_default_value(),
  Bmin = qgisprocess:::qgis_default_value(),
  Bmax = qgisprocess:::qgis_default_value(),
  Cmin = qgisprocess:::qgis_default_value(),
  Cmax = qgisprocess:::qgis_default_value(),
  Dmin = qgisprocess:::qgis_default_value(),
  Dmax = qgisprocess:::qgis_default_value(),
  Emin = qgisprocess:::qgis_default_value(),
  Emax = qgisprocess:::qgis_default_value(),
  Fmin = qgisprocess:::qgis_default_value(),
  Fmax = qgisprocess:::qgis_default_value(),
  fBmin = qgisprocess:::qgis_default_value(),
  fBmax = qgisprocess:::qgis_default_value(),
  fCmin = qgisprocess:::qgis_default_value(),
  fCmax = qgisprocess:::qgis_default_value(),
  fDmin = qgisprocess:::qgis_default_value(),
  fDmax = qgisprocess:::qgis_default_value(),
  fEmin = qgisprocess:::qgis_default_value(),
  fEmax = qgisprocess:::qgis_default_value(),
  fFmin = qgisprocess:::qgis_default_value(),
  fFmax = qgisprocess:::qgis_default_value(),
  fI = qgisprocess:::qgis_default_value(),
  G = qgisprocess:::qgis_default_value(),
  H = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{A}{\code{raster} - Slope grid (rad). Path to a raster layer.}

\item{Bmin}{\code{raster} - Min thickness grid (m) . Path to a raster layer.}

\item{Bmax}{\code{raster} - Max thickness grid (m) . Path to a raster layer.}

\item{Cmin}{\code{raster} - Min saturation grid (-) . Path to a raster layer.}

\item{Cmax}{\code{raster} - Max saturation grid (-) . Path to a raster layer.}

\item{Dmin}{\code{raster} - Min friction grid (degree) . Path to a raster layer.}

\item{Dmax}{\code{raster} - Max friction grid (degree) . Path to a raster layer.}

\item{Emin}{\code{raster} - Min density grid (g\code{/}cm3). Path to a raster layer.}

\item{Emax}{\code{raster} - Max density grid (g\code{/}cm3). Path to a raster layer.}

\item{Fmin}{\code{raster} - Min cohesion grid (MPa) . Path to a raster layer.}

\item{Fmax}{\code{raster} - Max cohesion grid (MPa) . Path to a raster layer.}

\item{fBmin}{\code{number} - Min global thickness (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fBmax}{\code{number} - Max global thickness (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fCmin}{\code{number} - Min global saturation (-). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fCmax}{\code{number} - Max global saturation (-). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fDmin}{\code{number} - Min global friction (degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fDmax}{\code{number} - Max global friction (degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fEmin}{\code{number} - Min global density (g\code{/}cm3). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fEmax}{\code{number} - Max global density (g\code{/}cm3). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fFmin}{\code{number} - Min global cohesion (MPa). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fFmax}{\code{number} - Max global cohesion (MPa). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fI}{\code{number} - Parameter sampling runs. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{G}{\code{rasterDestination} - FS values. Path for new raster layer.}

\item{H}{\code{rasterDestination} - FS classes. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Safetyfactor (sagang:safetyfactor). ---------------- Arguments ----------------  A: Slope grid (rad) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Bmin: Min thickness grid (m)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Bmax: Max thickness grid (m)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Cmin: Min saturation grid (-)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Cmax: Max saturation grid (-)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Dmin: Min friction grid (degree)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Dmax: Max friction grid (degree)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Emin: Min density grid (g/cm3) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Emax: Max density grid (g/cm3) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Fmin: Min cohesion grid (MPa)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Fmax: Max cohesion grid (MPa)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer fBmin: Min global thickness (m) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fBmax: Max global thickness (m) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fCmin: Min global saturation (-) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fCmax: Max global saturation (-) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fDmin: Min global friction (degree) 	Default value:	33 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fDmax: Max global friction (degree) 	Default value:	33 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fEmin: Min global density (g/cm3) 	Default value:	1.6 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fEmax: Max global density (g/cm3) 	Default value:	1.6 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fFmin: Min global cohesion (MPa) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fFmax: Max global cohesion (MPa) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fI: Parameter sampling runs 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression G: FS values 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer H: FS classes 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item G - outputRaster - FS values
\item H - outputRaster - FS classes
}
}
}
