% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_sagawetnessindex.R
\name{sagang_sagawetnessindex}
\alias{sagang_sagawetnessindex}
\title{QGIS algorithm - Saga wetness index}
\usage{
sagang_sagawetnessindex(
  DEM = qgisprocess:::qgis_default_value(),
  WEIGHT = qgisprocess:::qgis_default_value(),
  AREA = qgisprocess:::qgis_default_value(),
  SLOPE = qgisprocess:::qgis_default_value(),
  AREA_MOD = qgisprocess:::qgis_default_value(),
  TWI = qgisprocess:::qgis_default_value(),
  SUCTION = qgisprocess:::qgis_default_value(),
  AREA_TYPE = qgisprocess:::qgis_default_value(),
  SLOPE_TYPE = qgisprocess:::qgis_default_value(),
  SLOPE_MIN = qgisprocess:::qgis_default_value(),
  SLOPE_OFF = qgisprocess:::qgis_default_value(),
  SLOPE_WEIGHT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{WEIGHT}{\code{raster} - Weights. Path to a raster layer.}

\item{AREA}{\code{rasterDestination} - Catchment Area. Path for new raster layer.}

\item{SLOPE}{\code{rasterDestination} - Catchment Slope. Path for new raster layer.}

\item{AREA_MOD}{\code{rasterDestination} - Modified Catchment Area. Path for new raster layer.}

\item{TWI}{\code{rasterDestination} - Topographic Wetness Index. Path for new raster layer.}

\item{SUCTION}{\code{number} - Suction. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{AREA_TYPE}{\code{enum}  of \verb{("(0) total catchment area", "(1) square root of catchment area", "(2) specific catchment area")} - Type of Area. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SLOPE_TYPE}{\code{enum}  of \verb{("(0) local slope", "(1) catchment slope")} - Type of Slope. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SLOPE_MIN}{\code{number} - Minimum Slope. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SLOPE_OFF}{\code{number} - Offset Slope. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SLOPE_WEIGHT}{\code{number} - Slope Weighting. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Saga wetness index (sagang:sagawetnessindex). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer WEIGHT: Weights (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer AREA: Catchment Area 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SLOPE: Catchment Slope 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer AREA_MOD: Modified Catchment Area 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TWI: Topographic Wetness Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SUCTION: Suction 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression AREA_TYPE: Type of Area 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) total catchment area 		- 1: (1) square root of catchment area 		- 2: (2) specific catchment area 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SLOPE_TYPE: Type of Slope 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) local slope 		- 1: (1) catchment slope 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SLOPE_MIN: Minimum Slope 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SLOPE_OFF: Offset Slope 	Default value:	0.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SLOPE_WEIGHT: Slope Weighting 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item AREA - outputRaster - Catchment Area
\item AREA_MOD - outputRaster - Modified Catchment Area
\item SLOPE - outputRaster - Catchment Slope
\item TWI - outputRaster - Topographic Wetness Index
}
}
}
