% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_salem.R
\name{sagang_salem}
\alias{sagang_salem}
\title{QGIS algorithm - Salem}
\usage{
sagang_salem(
  SURFACE_T0 = qgisprocess:::qgis_default_value(),
  REGOLITH_T0 = qgisprocess:::qgis_default_value(),
  REGOLITH_T0_DEFAULT = qgisprocess:::qgis_default_value(),
  ALLOCHTHONE = qgisprocess:::qgis_default_value(),
  ALLOCHTHONE_DEFAULT = qgisprocess:::qgis_default_value(),
  SURFACE = qgisprocess:::qgis_default_value(),
  REGOLITH = qgisprocess:::qgis_default_value(),
  DIFFERENCE = qgisprocess:::qgis_default_value(),
  TIME_START = qgisprocess:::qgis_default_value(),
  TIME_STOP = qgisprocess:::qgis_default_value(),
  TIME_STEP = qgisprocess:::qgis_default_value(),
  DIFFUSIVE_KD = qgisprocess:::qgis_default_value(),
  DIFFUSIVE_NEIGHBOURS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SURFACE_T0}{\code{raster} - Initial Surface Elevation. Path to a raster layer.}

\item{REGOLITH_T0}{\code{raster} - Initial Regolith Thickness. Path to a raster layer.}

\item{REGOLITH_T0_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ALLOCHTHONE}{\code{raster} - Allochthone Input. Path to a raster layer.}

\item{ALLOCHTHONE_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SURFACE}{\code{rasterDestination} - Surface Elevation. Path for new raster layer.}

\item{REGOLITH}{\code{rasterDestination} - Regolith Thickness. Path for new raster layer.}

\item{DIFFERENCE}{\code{rasterDestination} - Difference. Path for new raster layer.}

\item{TIME_START}{\code{number} - Start (Years BP). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME_STOP}{\code{number} - Stop (Years BP). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME_STEP}{\code{number} - Step (Years). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DIFFUSIVE_KD}{\code{number} - Diffusivity Coefficient Kd (m^2\code{/}a). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DIFFUSIVE_NEIGHBOURS}{\code{enum}  of \verb{("(0) Neumann", "(1) Moore")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Salem (sagang:salem). ---------------- Arguments ----------------  SURFACE_T0: Initial Surface Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer REGOLITH_T0: Initial Regolith Thickness (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer REGOLITH_T0_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ALLOCHTHONE: Allochthone Input (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer ALLOCHTHONE_DEFAULT: Default 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SURFACE: Surface Elevation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer REGOLITH: Regolith Thickness 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIFFERENCE: Difference 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TIME_START: Start (Years BP) 	Default value:	50000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME_STOP: Stop (Years BP) 	Default value:	10000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME_STEP: Step (Years) 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DIFFUSIVE_KD: Diffusivity Coefficient Kd (m^2/a) 	Default value:	0.01 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DIFFUSIVE_NEIGHBOURS: Neighbourhood 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Neumann 		- 1: (1) Moore 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item DIFFERENCE - outputRaster - Difference
\item REGOLITH - outputRaster - Regolith Thickness
\item SURFACE - outputRaster - Surface Elevation
}
}
}
