% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_seededregiongrowing.R
\name{sagang_seededregiongrowing}
\alias{sagang_seededregiongrowing}
\title{QGIS algorithm - Seeded region growing}
\usage{
sagang_seededregiongrowing(
  SEEDS = qgisprocess:::qgis_default_value(),
  FEATURES = qgisprocess:::qgis_default_value(),
  SEGMENTS = qgisprocess:::qgis_default_value(),
  SIMILARITY = qgisprocess:::qgis_default_value(),
  TABLE = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  NEIGHBOUR = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  SIG_1 = qgisprocess:::qgis_default_value(),
  SIG_2 = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  REFRESH = qgisprocess:::qgis_default_value(),
  LEAFSIZE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SEEDS}{\code{raster} - Seeds. Path to a raster layer.}

\item{FEATURES}{\code{multilayer} - Features. .}

\item{SEGMENTS}{\code{rasterDestination} - Segments. Path for new raster layer.}

\item{SIMILARITY}{\code{rasterDestination} - Similarity. Path for new raster layer.}

\item{TABLE}{\code{vectorDestination} - Seeds. Path for new vector layer.}

\item{NORMALIZE}{\code{boolean} - Normalize Features. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NEIGHBOUR}{\code{enum}  of \verb{("(0) 4 (von Neumann)", "(1) 8 (Moore)")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) feature space and position", "(1) feature space")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SIG_1}{\code{number} - Variance in Feature Space. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SIG_2}{\code{number} - Variance in Position Space. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRESHOLD}{\code{number} - Similarity Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{REFRESH}{\code{boolean} - Refresh. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LEAFSIZE}{\code{number} - Leaf Size (for Speed Optimisation). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Seeded region growing (sagang:seededregiongrowing). ---------------- Arguments ----------------  SEEDS: Seeds 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FEATURES: Features 	Argument type:	multilayer SEGMENTS: Segments 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SIMILARITY: Similarity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TABLE: Seeds 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer NORMALIZE: Normalize Features 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NEIGHBOUR: Neighbourhood 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) 4 (von Neumann) 		- 1: (1) 8 (Moore) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) feature space and position 		- 1: (1) feature space 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SIG_1: Variance in Feature Space 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SIG_2: Variance in Position Space 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRESHOLD: Similarity Threshold 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression REFRESH: Refresh 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LEAFSIZE: Leaf Size (for Speed Optimisation) 	Default value:	256 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item SEGMENTS - outputRaster - Segments
\item SIMILARITY - outputRaster - Similarity
\item TABLE - outputVector - Seeds
}
}
}
