% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_selectanddelete.R
\name{sagang_selectanddelete}
\alias{sagang_selectanddelete}
\title{QGIS algorithm - Select and delete}
\usage{
sagang_selectanddelete(
  SHAPES = qgisprocess:::qgis_default_value(),
  ATTRIBUTE = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  EXPRESSION = qgisprocess:::qgis_default_value(),
  COMPARE = qgisprocess:::qgis_default_value(),
  CASE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{ATTRIBUTE}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{METHOD}{\code{enum}  of \verb{("(0) string expression", "(1) numerical expression")} - Expression Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXPRESSION}{\code{string} - Expression. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{COMPARE}{\code{enum}  of \verb{("(0) attribute is identical with search expression", "(1) attribute contains search expression", "(2) attribute is contained in search expression")} - Select if.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CASE}{\code{boolean} - Case Sensitive. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Select and delete (sagang:selectanddelete). ---------------- Arguments ----------------  SHAPES: Shapes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer ATTRIBUTE: Attribute 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names METHOD: Expression Type 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) string expression 		- 1: (1) numerical expression 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' EXPRESSION: Expression 	Default value:	expression 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression COMPARE: Select if... 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) attribute is identical with search expression 		- 1: (1) attribute contains search expression 		- 2: (2) attribute is contained in search expression 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CASE: Case Sensitive 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{

No output description provided.
}
}
