% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_selectgridfromlist.R
\name{sagang_selectgridfromlist}
\alias{sagang_selectgridfromlist}
\title{QGIS algorithm - Select grid from list}
\usage{
sagang_selectgridfromlist(
  GRIDS = qgisprocess:::qgis_default_value(),
  GRID = qgisprocess:::qgis_default_value(),
  INDEX = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grid List. .}

\item{GRID}{\code{rasterDestination} - Grid. Path for new raster layer.}

\item{INDEX}{\code{number} - Index. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Select grid from list (sagang:selectgridfromlist). ---------------- Arguments ----------------  GRIDS: Grid List 	Argument type:	multilayer GRID: Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer INDEX: Index 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item GRID - outputRaster - Grid
}
}
}
