% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_separatepointsbydirection.R
\name{sagang_separatepointsbydirection}
\alias{sagang_separatepointsbydirection}
\title{QGIS algorithm - Separate points by direction}
\usage{
sagang_separatepointsbydirection(
  OUTPUT = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  DIRECTIONS = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{OUTPUT}{\code{vectorDestination} - Ouput. Path for new vector layer.}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{DIRECTIONS}{\code{number} - Number of Directions. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TOLERANCE}{\code{number} - Tolerance (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Separate points by direction (sagang:separatepointsbydirection). ---------------- Arguments ----------------  OUTPUT: Ouput 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer DIRECTIONS: Number of Directions 	Default value:	4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TOLERANCE: Tolerance (Degree) 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Ouput
}
}
}
