% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_shalstab.R
\name{sagang_shalstab}
\alias{sagang_shalstab}
\title{QGIS algorithm - Shalstab}
\usage{
sagang_shalstab(
  A = qgisprocess:::qgis_default_value(),
  B = qgisprocess:::qgis_default_value(),
  Cmin = qgisprocess:::qgis_default_value(),
  Cmax = qgisprocess:::qgis_default_value(),
  Dmin = qgisprocess:::qgis_default_value(),
  Dmax = qgisprocess:::qgis_default_value(),
  Emin = qgisprocess:::qgis_default_value(),
  Emax = qgisprocess:::qgis_default_value(),
  Fmin = qgisprocess:::qgis_default_value(),
  Fmax = qgisprocess:::qgis_default_value(),
  Jmin = qgisprocess:::qgis_default_value(),
  Jmax = qgisprocess:::qgis_default_value(),
  fCmin = qgisprocess:::qgis_default_value(),
  fCmax = qgisprocess:::qgis_default_value(),
  fDmin = qgisprocess:::qgis_default_value(),
  fDmax = qgisprocess:::qgis_default_value(),
  fEmin = qgisprocess:::qgis_default_value(),
  fEmax = qgisprocess:::qgis_default_value(),
  fFmin = qgisprocess:::qgis_default_value(),
  fFmax = qgisprocess:::qgis_default_value(),
  fJmin = qgisprocess:::qgis_default_value(),
  fJmax = qgisprocess:::qgis_default_value(),
  fK = qgisprocess:::qgis_default_value(),
  G = qgisprocess:::qgis_default_value(),
  H = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{A}{\code{raster} - Slope grid (rad). Path to a raster layer.}

\item{B}{\code{raster} - Catchment area grid (m2). Path to a raster layer.}

\item{Cmin}{\code{raster} - Min Density grid (g\code{/}cm3). Path to a raster layer.}

\item{Cmax}{\code{raster} - Max Density grid (g\code{/}cm3). Path to a raster layer.}

\item{Dmin}{\code{raster} - Min Hydraulic conductivity grid (m\code{/}hr) . Path to a raster layer.}

\item{Dmax}{\code{raster} - Max Hydraulic conductivity grid (m\code{/}hr) . Path to a raster layer.}

\item{Emin}{\code{raster} - Min Thickness grid (m). Path to a raster layer.}

\item{Emax}{\code{raster} - Max Thickness grid (m). Path to a raster layer.}

\item{Fmin}{\code{raster} - Min Friction angle grid (degree) . Path to a raster layer.}

\item{Fmax}{\code{raster} - Max Friction angle grid (degree) . Path to a raster layer.}

\item{Jmin}{\code{raster} - Min Bulk cohesion grid (MPa) . Path to a raster layer.}

\item{Jmax}{\code{raster} - Max Bulk cohesion grid (MPa) . Path to a raster layer.}

\item{fCmin}{\code{number} - Global minimum density (g\code{/}cm3). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fCmax}{\code{number} - Global maximum density (g\code{/}cm3). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fDmin}{\code{number} - Global minimum conductivity (m\code{/}hr). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fDmax}{\code{number} - Global maximum conductivity (m\code{/}hr). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fEmin}{\code{number} - Global minimum thickness (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fEmax}{\code{number} - Global maximum thickness (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fFmin}{\code{number} - Global minimum friction angle (degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fFmax}{\code{number} - Global maximum friction angle (degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fJmin}{\code{number} - Global minimum bulk cohesion (MPa). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fJmax}{\code{number} - Global maximum bulk cohesion (MPa). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fK}{\code{number} - Parameter sampling runs. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{G}{\code{rasterDestination} - CR values. Path for new raster layer.}

\item{H}{\code{rasterDestination} - CR classes. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Shalstab (sagang:shalstab). ---------------- Arguments ----------------  A: Slope grid (rad) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer B: Catchment area grid (m2) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Cmin: Min Density grid (g/cm3) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Cmax: Max Density grid (g/cm3) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Dmin: Min Hydraulic conductivity grid (m/hr)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Dmax: Max Hydraulic conductivity grid (m/hr)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Emin: Min Thickness grid (m) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Emax: Max Thickness grid (m) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Fmin: Min Friction angle grid (degree)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Fmax: Max Friction angle grid (degree)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Jmin: Min Bulk cohesion grid (MPa)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Jmax: Max Bulk cohesion grid (MPa)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer fCmin: Global minimum density (g/cm3) 	Default value:	1.6 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fCmax: Global maximum density (g/cm3) 	Default value:	1.6 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fDmin: Global minimum conductivity (m/hr) 	Default value:	2.7 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fDmax: Global maximum conductivity (m/hr) 	Default value:	2.7 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fEmin: Global minimum thickness (m) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fEmax: Global maximum thickness (m) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fFmin: Global minimum friction angle (degree) 	Default value:	33 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fFmax: Global maximum friction angle (degree) 	Default value:	33 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fJmin: Global minimum bulk cohesion (MPa) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fJmax: Global maximum bulk cohesion (MPa) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fK: Parameter sampling runs 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression G: CR values 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer H: CR classes 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item G - outputRaster - CR values
\item H - outputRaster - CR classes
}
}
}
