% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_shannonindex.R
\name{sagang_shannonindex}
\alias{sagang_shannonindex}
\title{QGIS algorithm - Shannon index}
\usage{
sagang_shannonindex(
  CATEGORIES = qgisprocess:::qgis_default_value(),
  COUNT = qgisprocess:::qgis_default_value(),
  INDEX = qgisprocess:::qgis_default_value(),
  EVENNESS = qgisprocess:::qgis_default_value(),
  KERNEL_TYPE = qgisprocess:::qgis_default_value(),
  KERNEL_RADIUS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CATEGORIES}{\code{raster} - Categories. Path to a raster layer.}

\item{COUNT}{\code{rasterDestination} - Number of Categories. Path for new raster layer.}

\item{INDEX}{\code{rasterDestination} - Shannon Index. Path for new raster layer.}

\item{EVENNESS}{\code{rasterDestination} - Evenness. Path for new raster layer.}

\item{KERNEL_TYPE}{\code{enum}  of \verb{("(0) Square", "(1) Circle")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Shannon index (sagang:shannonindex). ---------------- Arguments ----------------  CATEGORIES: Categories 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer COUNT: Number of Categories 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer INDEX: Shannon Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer EVENNESS: Evenness 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer KERNEL_TYPE: Kernel Type 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Square 		- 1: (1) Circle 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' KERNEL_RADIUS: Radius 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item COUNT - outputRaster - Number of Categories
\item EVENNESS - outputRaster - Evenness
\item INDEX - outputRaster - Shannon Index
}
}
}
