% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_shapesbuffer.R
\name{sagang_shapesbuffer}
\alias{sagang_shapesbuffer}
\title{QGIS algorithm - Shapes buffer}
\usage{
sagang_shapesbuffer(
  SHAPES = qgisprocess:::qgis_default_value(),
  BUFFER = qgisprocess:::qgis_default_value(),
  DIST_FIELD = qgisprocess:::qgis_default_value(),
  DIST_FIELD_DEFAULT = qgisprocess:::qgis_default_value(),
  DIST_SCALE = qgisprocess:::qgis_default_value(),
  DISSOLVE = qgisprocess:::qgis_default_value(),
  NZONES = qgisprocess:::qgis_default_value(),
  POLY_INNER = qgisprocess:::qgis_default_value(),
  DARC = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{BUFFER}{\code{vectorDestination} - Buffer. Path for new vector layer.}

\item{DIST_FIELD}{\code{field} - Buffer Distance. The name of an existing field. ; delimited list of existing field names.}

\item{DIST_FIELD_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DIST_SCALE}{\code{number} - Scaling Factor for Attribute Value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DISSOLVE}{\code{boolean} - Dissolve Buffers. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NZONES}{\code{number} - Number of Buffer Zones. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{POLY_INNER}{\code{boolean} - Inner Buffer. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DARC}{\code{number} - Arc Vertex Distance (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Shapes buffer (sagang:shapesbuffer). ---------------- Arguments ----------------  SHAPES: Shapes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer BUFFER: Buffer 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DIST_FIELD: Buffer Distance 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names DIST_FIELD_DEFAULT: Default 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DIST_SCALE: Scaling Factor for Attribute Value 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DISSOLVE: Dissolve Buffers 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NZONES: Number of Buffer Zones 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression POLY_INNER: Inner Buffer 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DARC: Arc Vertex Distance (Degree) 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item BUFFER - outputVector - Buffer
}
}
}
