% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_shrinkandexpand.R
\name{sagang_shrinkandexpand}
\alias{sagang_shrinkandexpand}
\title{QGIS algorithm - Shrink and expand}
\usage{
sagang_shrinkandexpand(
  INPUT = qgisprocess:::qgis_default_value(),
  RESULT = qgisprocess:::qgis_default_value(),
  OPERATION = qgisprocess:::qgis_default_value(),
  CIRCLE = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  EXPAND = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Grid. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - Result Grid. Path for new raster layer.}

\item{OPERATION}{\code{enum}  of \verb{("(0) shrink", "(1) expand", "(2) shrink and expand", "(3) expand and shrink")} - Operation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CIRCLE}{\code{enum}  of \verb{("(0) Square", "(1) Circle")} - Search Mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EXPAND}{\code{enum}  of \verb{("(0) minimum", "(1) maximum", "(2) mean", "(3) majority")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Shrink and expand (sagang:shrinkandexpand). ---------------- Arguments ----------------  INPUT: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RESULT: Result Grid 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer OPERATION: Operation 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) shrink 		- 1: (1) expand 		- 2: (2) shrink and expand 		- 3: (3) expand and shrink 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CIRCLE: Search Mode 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Square 		- 1: (1) Circle 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RADIUS: Radius 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EXPAND: Method 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) minimum 		- 1: (1) maximum 		- 2: (2) mean 		- 3: (3) majority 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Result Grid
}
}
}
