% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_sieveandclump.R
\name{sagang_sieveandclump}
\alias{sagang_sieveandclump}
\title{QGIS algorithm - Sieve and clump}
\usage{
sagang_sieveandclump(
  CLASSES = qgisprocess:::qgis_default_value(),
  FILTERED = qgisprocess:::qgis_default_value(),
  MODE = qgisprocess:::qgis_default_value(),
  SIEVE = qgisprocess:::qgis_default_value(),
  EXPAND = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{CLASSES}{\code{raster} - Classes. Path to a raster layer.}

\item{FILTERED}{\code{rasterDestination} - Sieve and Clump. Path for new raster layer.}

\item{MODE}{\code{enum}  of \verb{("(0) Neumann", "(1) Moore")} - Filter. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SIEVE}{\code{number} - Sieving Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EXPAND}{\code{number} - Maximum Expansion Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Sieve and clump (sagang:sieveandclump). ---------------- Arguments ----------------  CLASSES: Classes 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FILTERED: Sieve and Clump 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MODE: Filter 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Neumann 		- 1: (1) Moore 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SIEVE: Sieving Threshold 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EXPAND: Maximum Expansion Distance 	Default value:	4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FILTERED - outputRaster - Sieve and Clump
}
}
}
