% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_sieveclasses.R
\name{sagang_sieveclasses}
\alias{sagang_sieveclasses}
\title{QGIS algorithm - Sieve classes}
\usage{
sagang_sieveclasses(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  MODE = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  ALL = qgisprocess:::qgis_default_value(),
  CLASS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Classes. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Sieved Classes. Path for new raster layer.}

\item{MODE}{\code{enum}  of \verb{("(0) Neumann", "(1) Moore")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{THRESHOLD}{\code{number} - Minimum Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ALL}{\code{enum}  of \verb{("(0) single class", "(1) all classes")} - Class Selection. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CLASS}{\code{number} - Class Identifier. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Sieve classes (sagang:sieveclasses). ---------------- Arguments ----------------  INPUT: Classes 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OUTPUT: Sieved Classes 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MODE: Neighbourhood 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Neumann 		- 1: (1) Moore 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' THRESHOLD: Minimum Threshold 	Default value:	4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ALL: Class Selection 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) single class 		- 1: (1) all classes 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CLASS: Class Identifier 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Sieved Classes
}
}
}
