% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_simulation.R
\name{sagang_simulation}
\alias{sagang_simulation}
\title{QGIS algorithm - Simulation}
\usage{
sagang_simulation(
  DEM = qgisprocess:::qgis_default_value(),
  FUEL = qgisprocess:::qgis_default_value(),
  WINDSPD = qgisprocess:::qgis_default_value(),
  WINDDIR = qgisprocess:::qgis_default_value(),
  M1H = qgisprocess:::qgis_default_value(),
  M10H = qgisprocess:::qgis_default_value(),
  M100H = qgisprocess:::qgis_default_value(),
  MHERB = qgisprocess:::qgis_default_value(),
  MWOOD = qgisprocess:::qgis_default_value(),
  IGNITION = qgisprocess:::qgis_default_value(),
  TIME = qgisprocess:::qgis_default_value(),
  FLAME = qgisprocess:::qgis_default_value(),
  INTENSITY = qgisprocess:::qgis_default_value(),
  UPDATEVIEW = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{FUEL}{\code{raster} - Fuel Model. Path to a raster layer.}

\item{WINDSPD}{\code{raster} - Wind Speed. Path to a raster layer.}

\item{WINDDIR}{\code{raster} - Wind Direction. Path to a raster layer.}

\item{M1H}{\code{raster} - Dead Fuel Moisture 1H. Path to a raster layer.}

\item{M10H}{\code{raster} - Dead Fuel Moisture 10H. Path to a raster layer.}

\item{M100H}{\code{raster} - Dead Fuel Moisture 100H. Path to a raster layer.}

\item{MHERB}{\code{raster} - Herbaceous Fuel Moisture. Path to a raster layer.}

\item{MWOOD}{\code{raster} - Wood Fuel Moisture. Path to a raster layer.}

\item{IGNITION}{\code{raster} - Ignition Points. Path to a raster layer.}

\item{TIME}{\code{rasterDestination} - Time. Path for new raster layer.}

\item{FLAME}{\code{rasterDestination} - Flame Length. Path for new raster layer.}

\item{INTENSITY}{\code{rasterDestination} - Intensity. Path for new raster layer.}

\item{UPDATEVIEW}{\code{boolean} - Update View. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Simulation (sagang:simulation). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FUEL: Fuel Model 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer WINDSPD: Wind Speed 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer WINDDIR: Wind Direction 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer M1H: Dead Fuel Moisture 1H 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer M10H: Dead Fuel Moisture 10H 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer M100H: Dead Fuel Moisture 100H 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MHERB: Herbaceous Fuel Moisture 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MWOOD: Wood Fuel Moisture 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer IGNITION: Ignition Points 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TIME: Time 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FLAME: Flame Length 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer INTENSITY: Intensity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer UPDATEVIEW: Update View 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLAME - outputRaster - Flame Length
\item INTENSITY - outputRaster - Intensity
\item TIME - outputRaster - Time
}
}
}
