% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_sinkremoval.R
\name{sagang_sinkremoval}
\alias{sagang_sinkremoval}
\title{QGIS algorithm - Sink removal}
\usage{
sagang_sinkremoval(
  DEM = qgisprocess:::qgis_default_value(),
  SINKROUTE = qgisprocess:::qgis_default_value(),
  DEM_PREPROC = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  THRSHEIGHT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{SINKROUTE}{\code{raster} - Sink Route. Path to a raster layer.}

\item{DEM_PREPROC}{\code{rasterDestination} - Preprocessed DEM. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) Deepen Drainage Routes", "(1) Fill Sinks")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{THRESHOLD}{\code{boolean} - Threshold. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRSHEIGHT}{\code{number} - Threshold Height. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Sink removal (sagang:sinkremoval). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SINKROUTE: Sink Route (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DEM_PREPROC: Preprocessed DEM 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Method 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Deepen Drainage Routes 		- 1: (1) Fill Sinks 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' THRESHOLD: Threshold 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRSHEIGHT: Threshold Height 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item DEM_PREPROC - outputRaster - Preprocessed DEM
}
}
}
