% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_slopeaspectcurvature.R
\name{sagang_slopeaspectcurvature}
\alias{sagang_slopeaspectcurvature}
\title{QGIS algorithm - Slope, aspect, curvature}
\usage{
sagang_slopeaspectcurvature(
  ELEVATION = qgisprocess:::qgis_default_value(),
  SLOPE = qgisprocess:::qgis_default_value(),
  ASPECT = qgisprocess:::qgis_default_value(),
  C_GENE = qgisprocess:::qgis_default_value(),
  C_PROF = qgisprocess:::qgis_default_value(),
  C_PLAN = qgisprocess:::qgis_default_value(),
  C_TANG = qgisprocess:::qgis_default_value(),
  C_LONG = qgisprocess:::qgis_default_value(),
  C_CROS = qgisprocess:::qgis_default_value(),
  C_MINI = qgisprocess:::qgis_default_value(),
  C_MAXI = qgisprocess:::qgis_default_value(),
  C_TOTA = qgisprocess:::qgis_default_value(),
  C_ROTO = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  UNIT_SLOPE = qgisprocess:::qgis_default_value(),
  UNIT_ASPECT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{SLOPE}{\code{rasterDestination} - Slope. Path for new raster layer.}

\item{ASPECT}{\code{rasterDestination} - Aspect. Path for new raster layer.}

\item{C_GENE}{\code{rasterDestination} - General Curvature. Path for new raster layer.}

\item{C_PROF}{\code{rasterDestination} - Profile Curvature. Path for new raster layer.}

\item{C_PLAN}{\code{rasterDestination} - Plan Curvature. Path for new raster layer.}

\item{C_TANG}{\code{rasterDestination} - Tangential Curvature. Path for new raster layer.}

\item{C_LONG}{\code{rasterDestination} - Longitudinal Curvature. Path for new raster layer.}

\item{C_CROS}{\code{rasterDestination} - Cross-Sectional Curvature. Path for new raster layer.}

\item{C_MINI}{\code{rasterDestination} - Minimal Curvature. Path for new raster layer.}

\item{C_MAXI}{\code{rasterDestination} - Maximal Curvature. Path for new raster layer.}

\item{C_TOTA}{\code{rasterDestination} - Total Curvature. Path for new raster layer.}

\item{C_ROTO}{\code{rasterDestination} - Flow Line Curvature. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) maximum slope (Travis et al. 1975)", "(1) maximum triangle slope (Tarboton 1997)", "(2) least squares fitted plane (Horn 1981, Costa-Cabral & Burgess 1996)", "(3) 6 parameter 2nd order polynom (Evans 1979)", "(4) 6 parameter 2nd order polynom (Heerdegen & Beran 1982)", "(5) 6 parameter 2nd order polynom (Bauer, Rohdenburg, Bork 1985)", "(6) 9 parameter 2nd order polynom (Zevenbergen & Thorne 1987)", "(7) 10 parameter 3rd order polynom (Haralick 1983)", "(8) 10 parameter 3rd order polynom (Florinsky 2009)")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{UNIT_SLOPE}{\code{enum}  of \verb{("(0) radians", "(1) degree", "(2) percent rise")} - Unit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{UNIT_ASPECT}{\code{enum}  of \verb{("(0) radians", "(1) degree")} - Unit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Slope, aspect, curvature (sagang:slopeaspectcurvature). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SLOPE: Slope 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ASPECT: Aspect 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_GENE: General Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_PROF: Profile Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_PLAN: Plan Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_TANG: Tangential Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_LONG: Longitudinal Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_CROS: Cross-Sectional Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_MINI: Minimal Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_MAXI: Maximal Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_TOTA: Total Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_ROTO: Flow Line Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Method 	Default value:	6 	Argument type:	enum 	Available values: 		- 0: (0) maximum slope (Travis et al. 1975) 		- 1: (1) maximum triangle slope (Tarboton 1997) 		- 2: (2) least squares fitted plane (Horn 1981, Costa-Cabral & Burgess 1996) 		- 3: (3) 6 parameter 2nd order polynom (Evans 1979) 		- 4: (4) 6 parameter 2nd order polynom (Heerdegen & Beran 1982) 		- 5: (5) 6 parameter 2nd order polynom (Bauer, Rohdenburg, Bork 1985) 		- 6: (6) 9 parameter 2nd order polynom (Zevenbergen & Thorne 1987) 		- 7: (7) 10 parameter 3rd order polynom (Haralick 1983) 		- 8: (8) 10 parameter 3rd order polynom (Florinsky 2009) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' UNIT_SLOPE: Unit 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) radians 		- 1: (1) degree 		- 2: (2) percent rise 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' UNIT_ASPECT: Unit 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) radians 		- 1: (1) degree 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item ASPECT - outputRaster - Aspect
\item C_CROS - outputRaster - Cross-Sectional Curvature
\item C_GENE - outputRaster - General Curvature
\item C_LONG - outputRaster - Longitudinal Curvature
\item C_MAXI - outputRaster - Maximal Curvature
\item C_MINI - outputRaster - Minimal Curvature
\item C_PLAN - outputRaster - Plan Curvature
\item C_PROF - outputRaster - Profile Curvature
\item C_ROTO - outputRaster - Flow Line Curvature
\item C_TANG - outputRaster - Tangential Curvature
\item C_TOTA - outputRaster - Total Curvature
\item SLOPE - outputRaster - Slope
}
}
}
