% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_snappointstogrid.R
\name{sagang_snappointstogrid}
\alias{sagang_snappointstogrid}
\title{QGIS algorithm - Snap points to grid}
\usage{
sagang_snappointstogrid(
  INPUT = qgisprocess:::qgis_default_value(),
  GRID = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  MOVES = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  SHAPE = qgisprocess:::qgis_default_value(),
  EXTREME = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Points. Path to a vector layer.}

\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{OUTPUT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{MOVES}{\code{vectorDestination} - Moves. Path for new vector layer.}

\item{DISTANCE}{\code{number} - Search Distance (Map Units). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SHAPE}{\code{enum}  of \verb{("(0) circle", "(1) square")} - Search Shape. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXTREME}{\code{enum}  of \verb{("(0) minimum", "(1) maximum")} - Extreme. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Snap points to grid (sagang:snappointstogrid). ---------------- Arguments ----------------  INPUT: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OUTPUT: Result 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer MOVES: Moves 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DISTANCE: Search Distance (Map Units) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SHAPE: Search Shape 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) circle 		- 1: (1) square 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' EXTREME: Extreme 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) minimum 		- 1: (1) maximum 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item MOVES - outputVector - Moves
\item OUTPUT - outputVector - Result
}
}
}
