% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_snowcover.R
\name{sagang_snowcover}
\alias{sagang_snowcover}
\title{QGIS algorithm - Snow cover}
\usage{
sagang_snowcover(
  T = qgisprocess:::qgis_default_value(),
  P = qgisprocess:::qgis_default_value(),
  DAYS = qgisprocess:::qgis_default_value(),
  MEAN = qgisprocess:::qgis_default_value(),
  MAXIMUM = qgisprocess:::qgis_default_value(),
  QUANTILE = qgisprocess:::qgis_default_value(),
  QUANT_VAL = qgisprocess:::qgis_default_value(),
  TIME = qgisprocess:::qgis_default_value(),
  MONTH = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{T}{\code{multilayer} - Mean Temperature. .}

\item{P}{\code{multilayer} - Precipitation. .}

\item{DAYS}{\code{rasterDestination} - Snow Cover Days. Path for new raster layer.}

\item{MEAN}{\code{rasterDestination} - Average. Path for new raster layer.}

\item{MAXIMUM}{\code{rasterDestination} - Maximum. Path for new raster layer.}

\item{QUANTILE}{\code{rasterDestination} - Quantile. Path for new raster layer.}

\item{QUANT_VAL}{\code{number} - Value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME}{\code{enum}  of \verb{("(0) Year", "(1) January - March", "(2) April - June", "(3) July - September", "(4) October - December", "(5) Single Month")} - Time. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MONTH}{\code{enum}  of \verb{("(0) January", "(1) February", "(2) March", "(3) April", "(4) May", "(5) June", "(6) July", "(7) August", "(8) September", "(9) October", "(10) November", "(11) December")} - Month. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Snow cover (sagang:snowcover). ---------------- Arguments ----------------  T: Mean Temperature 	Argument type:	multilayer P: Precipitation 	Argument type:	multilayer DAYS: Snow Cover Days 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MEAN: Average 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MAXIMUM: Maximum 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer QUANTILE: Quantile 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer QUANT_VAL: Value 	Default value:	50 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME: Time 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Year 		- 1: (1) January - March 		- 2: (2) April - June 		- 3: (3) July - September 		- 4: (4) October - December 		- 5: (5) Single Month 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' MONTH: Month 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) January 		- 1: (1) February 		- 2: (2) March 		- 3: (3) April 		- 4: (4) May 		- 5: (5) June 		- 6: (6) July 		- 7: (7) August 		- 8: (8) September 		- 9: (9) October 		- 10: (10) November 		- 11: (11) December 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item DAYS - outputRaster - Snow Cover Days
\item MAXIMUM - outputRaster - Maximum
\item MEAN - outputRaster - Average
\item QUANTILE - outputRaster - Quantile
}
}
}
