% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_spatialpointpatternanalysis.R
\name{sagang_spatialpointpatternanalysis}
\alias{sagang_spatialpointpatternanalysis}
\title{QGIS algorithm - Spatial point pattern analysis}
\usage{
sagang_spatialpointpatternanalysis(
  POINTS = qgisprocess:::qgis_default_value(),
  WEIGHT = qgisprocess:::qgis_default_value(),
  CENTRE = qgisprocess:::qgis_default_value(),
  STDDIST = qgisprocess:::qgis_default_value(),
  STEP = qgisprocess:::qgis_default_value(),
  BBOX = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{WEIGHT}{\code{field} - Weight. The name of an existing field. ; delimited list of existing field names.}

\item{CENTRE}{\code{vectorDestination} - Mean Centre. Path for new vector layer.}

\item{STDDIST}{\code{vectorDestination} - Standard Distance. Path for new vector layer.}

\item{STEP}{\code{number} - Vertex Distance (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{BBOX}{\code{vectorDestination} - Bounding Box. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Spatial point pattern analysis (sagang:spatialpointpatternanalysis). ---------------- Arguments ----------------  POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer WEIGHT: Weight 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names CENTRE: Mean Centre 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer STDDIST: Standard Distance 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer STEP: Vertex Distance (Degree) 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression BBOX: Bounding Box 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item BBOX - outputVector - Bounding Box
\item CENTRE - outputVector - Mean Centre
\item STDDIST - outputVector - Standard Distance
}
}
}
