% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_splitshapeslayer.R
\name{sagang_splitshapeslayer}
\alias{sagang_splitshapeslayer}
\title{QGIS algorithm - Split shapes layer}
\usage{
sagang_splitshapeslayer(
  SHAPES = qgisprocess:::qgis_default_value(),
  CUTS = qgisprocess:::qgis_default_value(),
  EXTENT = qgisprocess:::qgis_default_value(),
  NX = qgisprocess:::qgis_default_value(),
  NY = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{CUTS}{\code{vectorDestination} - Tiles. Path for new vector layer.}

\item{EXTENT}{\code{vectorDestination} - Extent. Path for new vector layer.}

\item{NX}{\code{number} - Number of horizontal tiles. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NY}{\code{number} - Number of vertical tiles. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) completely contained", "(1) intersects", "(2) center")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Split shapes layer (sagang:splitshapeslayer). ---------------- Arguments ----------------  SHAPES: Shapes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer CUTS: Tiles 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer EXTENT: Extent 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer NX: Number of horizontal tiles 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NY: Number of vertical tiles 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Method 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) completely contained 		- 1: (1) intersects 		- 2: (2) center 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item CUTS - outputVector - Tiles
\item EXTENT - outputVector - Extent
}
}
}
