% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_splitshapeslayercompletely.R
\name{sagang_splitshapeslayercompletely}
\alias{sagang_splitshapeslayercompletely}
\title{QGIS algorithm - Split shapes layer completely}
\usage{
sagang_splitshapeslayercompletely(
  SHAPES = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  LIST = qgisprocess:::qgis_default_value(),
  NAMING = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Input. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{LIST}{\code{vectorDestination} - Output. Path for new vector layer.}

\item{NAMING}{\code{enum}  of \verb{("(0) number of order", "(1) attribute")} - Name by.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Split shapes layer completely (sagang:splitshapeslayercompletely). ---------------- Arguments ----------------  SHAPES: Input 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Attribute 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names LIST: Output 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer NAMING: Name by... 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) number of order 		- 1: (1) attribute 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item LIST - outputVector - Output
}
}
}
