% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_splittableshapesbyattribute.R
\name{sagang_splittableshapesbyattribute}
\alias{sagang_splittableshapesbyattribute}
\title{QGIS algorithm - Split table/shapes by attribute}
\usage{
sagang_splittableshapesbyattribute(
  TABLE = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  CUTS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TABLE}{\code{source} - Table \code{/} Shapes. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{CUTS}{\code{vectorDestination} - Cuts. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Split table/shapes by attribute (sagang:splittableshapesbyattribute). ---------------- Arguments ----------------  TABLE: Table / Shapes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Attribute 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names CUTS: Cuts 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item CUTS - outputVector - Cuts
}
}
}
