% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_statisticsforrasters.R
\name{sagang_statisticsforrasters}
\alias{sagang_statisticsforrasters}
\title{QGIS algorithm - Statistics for rasters}
\usage{
sagang_statisticsforrasters(
  GRIDS = qgisprocess:::qgis_default_value(),
  WEIGHTS = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  MEAN = qgisprocess:::qgis_default_value(),
  MIN = qgisprocess:::qgis_default_value(),
  MAX = qgisprocess:::qgis_default_value(),
  RANGE = qgisprocess:::qgis_default_value(),
  SUM = qgisprocess:::qgis_default_value(),
  SUM2 = qgisprocess:::qgis_default_value(),
  VAR = qgisprocess:::qgis_default_value(),
  STDDEV = qgisprocess:::qgis_default_value(),
  STDDEVLO = qgisprocess:::qgis_default_value(),
  STDDEVHI = qgisprocess:::qgis_default_value(),
  PCTL = qgisprocess:::qgis_default_value(),
  PCTL_VAL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Values. .}

\item{WEIGHTS}{\code{multilayer} - Weights. .}

\item{RESAMPLING}{\code{enum}  of \verb{("(0) Nearest Neighbour", "(1) Bilinear Interpolation", "(2) Bicubic Spline Interpolation", "(3) B-Spline Interpolation")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MEAN}{\code{rasterDestination} - Arithmetic Mean. Path for new raster layer.}

\item{MIN}{\code{rasterDestination} - Minimum. Path for new raster layer.}

\item{MAX}{\code{rasterDestination} - Maximum. Path for new raster layer.}

\item{RANGE}{\code{rasterDestination} - Range. Path for new raster layer.}

\item{SUM}{\code{rasterDestination} - Sum. Path for new raster layer.}

\item{SUM2}{\code{rasterDestination} - Sum2. Path for new raster layer.}

\item{VAR}{\code{rasterDestination} - Variance. Path for new raster layer.}

\item{STDDEV}{\code{rasterDestination} - Standard Deviation. Path for new raster layer.}

\item{STDDEVLO}{\code{rasterDestination} - Mean less Standard Deviation. Path for new raster layer.}

\item{STDDEVHI}{\code{rasterDestination} - Mean plus Standard Deviation. Path for new raster layer.}

\item{PCTL}{\code{rasterDestination} - Percentile. Path for new raster layer.}

\item{PCTL_VAL}{\code{number} - Percentile. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Statistics for rasters (sagang:statisticsforrasters). ---------------- Arguments ----------------  GRIDS: Values 	Argument type:	multilayer WEIGHTS: Weights (optional) 	Argument type:	multilayer RESAMPLING: Resampling 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Nearest Neighbour 		- 1: (1) Bilinear Interpolation 		- 2: (2) Bicubic Spline Interpolation 		- 3: (3) B-Spline Interpolation 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' MEAN: Arithmetic Mean 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MIN: Minimum 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MAX: Maximum 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RANGE: Range 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SUM: Sum 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SUM2: Sum2 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VAR: Variance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STDDEV: Standard Deviation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STDDEVLO: Mean less Standard Deviation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer STDDEVHI: Mean plus Standard Deviation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PCTL: Percentile 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer PCTL_VAL: Percentile 	Default value:	50 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item MAX - outputRaster - Maximum
\item MEAN - outputRaster - Arithmetic Mean
\item MIN - outputRaster - Minimum
\item PCTL - outputRaster - Percentile
\item RANGE - outputRaster - Range
\item STDDEV - outputRaster - Standard Deviation
\item STDDEVHI - outputRaster - Mean plus Standard Deviation
\item STDDEVLO - outputRaster - Mean less Standard Deviation
\item SUM - outputRaster - Sum
\item SUM2 - outputRaster - Sum2
\item VAR - outputRaster - Variance
}
}
}
