% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_summitextraction.R
\name{sagang_summitextraction}
\alias{sagang_summitextraction}
\title{QGIS algorithm - Summit extraction}
\usage{
sagang_summitextraction(
  SURFACE = qgisprocess:::qgis_default_value(),
  SUMMITS = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  SCALE_FILTER = qgisprocess:::qgis_default_value(),
  SCALE_TPI = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SURFACE}{\code{raster} - Surface. Path to a raster layer.}

\item{SUMMITS}{\code{vectorDestination} - Summits. Path for new vector layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) Resampling Filter", "(1) Topographic Position Index (TPI)")} - Trend Surface. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SCALE_FILTER}{\code{number} - Scale. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SCALE_TPI}{\code{number} - Scale. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRESHOLD}{\code{number} - Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Summit extraction (sagang:summitextraction). ---------------- Arguments ----------------  SURFACE: Surface 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SUMMITS: Summits 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer METHOD: Trend Surface 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Resampling Filter 		- 1: (1) Topographic Position Index (TPI) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SCALE_FILTER: Scale 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SCALE_TPI: Scale 	Default value:	300 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRESHOLD: Threshold 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item SUMMITS - outputVector - Summits
}
}
}
