% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_sunriseandsunset.R
\name{sagang_sunriseandsunset}
\alias{sagang_sunriseandsunset}
\title{QGIS algorithm - Sunrise and sunset}
\usage{
sagang_sunriseandsunset(
  TARGET = qgisprocess:::qgis_default_value(),
  SUNRISE = qgisprocess:::qgis_default_value(),
  SUNSET = qgisprocess:::qgis_default_value(),
  LENGTH = qgisprocess:::qgis_default_value(),
  DAY = qgisprocess:::qgis_default_value(),
  TIME = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TARGET}{\code{raster} - Target System. Path to a raster layer.}

\item{SUNRISE}{\code{rasterDestination} - Sunrise. Path for new raster layer.}

\item{SUNSET}{\code{rasterDestination} - Sunset. Path for new raster layer.}

\item{LENGTH}{\code{rasterDestination} - Day Length. Path for new raster layer.}

\item{DAY}{\code{string} - Day of Month. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TIME}{\code{enum}  of \verb{("(0) local", "(1) world")} - Time. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Sunrise and sunset (sagang:sunriseandsunset). ---------------- Arguments ----------------  TARGET: Target System 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SUNRISE: Sunrise 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SUNSET: Sunset 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LENGTH: Day Length 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DAY: Day of Month 	Default value:	2021-03-24 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TIME: Time 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) local 		- 1: (1) world 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item LENGTH - outputRaster - Day Length
\item SUNRISE - outputRaster - Sunrise
\item SUNSET - outputRaster - Sunset
}
}
}
