% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_superpixelsegmentation.R
\name{sagang_superpixelsegmentation}
\alias{sagang_superpixelsegmentation}
\title{QGIS algorithm - Superpixel segmentation}
\usage{
sagang_superpixelsegmentation(
  FEATURES = qgisprocess:::qgis_default_value(),
  NORMALIZE = qgisprocess:::qgis_default_value(),
  POLYGONS = qgisprocess:::qgis_default_value(),
  MAX_ITERATIONS = qgisprocess:::qgis_default_value(),
  REGULARIZATION = qgisprocess:::qgis_default_value(),
  SIZE = qgisprocess:::qgis_default_value(),
  SIZE_MIN = qgisprocess:::qgis_default_value(),
  SUPERPIXELS_DO = qgisprocess:::qgis_default_value(),
  SUPERPIXELS = qgisprocess:::qgis_default_value(),
  POSTPROCESSING = qgisprocess:::qgis_default_value(),
  NCLUSTER = qgisprocess:::qgis_default_value(),
  SPLIT_CLUSTERS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{FEATURES}{\code{multilayer} - Features. .}

\item{NORMALIZE}{\code{boolean} - Normalize. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{POLYGONS}{\code{vectorDestination} - Segments. Path for new vector layer.}

\item{MAX_ITERATIONS}{\code{number} - Maximum Iterations. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{REGULARIZATION}{\code{number} - Regularization. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SIZE}{\code{number} - Region Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SIZE_MIN}{\code{number} - Minimum Region Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SUPERPIXELS_DO}{\code{boolean} - Create Superpixel Grids. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SUPERPIXELS}{\code{rasterDestination} - Superpixels. Path for new raster layer.}

\item{POSTPROCESSING}{\code{enum}  of \verb{("(0) none", "(1) unsupervised classification")} - Post-Processing. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NCLUSTER}{\code{number} - Number of Clusters. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SPLIT_CLUSTERS}{\code{boolean} - Split Clusters. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Superpixel segmentation (sagang:superpixelsegmentation). ---------------- Arguments ----------------  FEATURES: Features 	Argument type:	multilayer NORMALIZE: Normalize 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression POLYGONS: Segments 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer MAX_ITERATIONS: Maximum Iterations 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression REGULARIZATION: Regularization 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SIZE: Region Size 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SIZE_MIN: Minimum Region Size 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SUPERPIXELS_DO: Create Superpixel Grids 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SUPERPIXELS: Superpixels 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer POSTPROCESSING: Post-Processing 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) none 		- 1: (1) unsupervised classification 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NCLUSTER: Number of Clusters 	Default value:	12 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SPLIT_CLUSTERS: Split Clusters 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item POLYGONS - outputVector - Segments
\item SUPERPIXELS - outputRaster - Superpixels
}
}
}
