% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_supervisedclassificationforgrids.R
\name{sagang_supervisedclassificationforgrids}
\alias{sagang_supervisedclassificationforgrids}
\title{QGIS algorithm - Supervised classification for grids}
\usage{
sagang_supervisedclassificationforgrids(
  GRIDS = qgisprocess:::qgis_default_value(),
  NORMALISE = qgisprocess:::qgis_default_value(),
  CLASSES = qgisprocess:::qgis_default_value(),
  CLASSES_LUT = qgisprocess:::qgis_default_value(),
  QUALITY = qgisprocess:::qgis_default_value(),
  TRAINING = qgisprocess:::qgis_default_value(),
  TRAINING_CLASS = qgisprocess:::qgis_default_value(),
  FILE_LOAD = qgisprocess:::qgis_default_value(),
  FILE_SAVE = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  THRESHOLD_DIST = qgisprocess:::qgis_default_value(),
  THRESHOLD_ANGLE = qgisprocess:::qgis_default_value(),
  THRESHOLD_PROB = qgisprocess:::qgis_default_value(),
  RELATIVE_PROB = qgisprocess:::qgis_default_value(),
  WTA_0 = qgisprocess:::qgis_default_value(),
  WTA_1 = qgisprocess:::qgis_default_value(),
  WTA_2 = qgisprocess:::qgis_default_value(),
  WTA_3 = qgisprocess:::qgis_default_value(),
  WTA_4 = qgisprocess:::qgis_default_value(),
  WTA_5 = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Features. .}

\item{NORMALISE}{\code{boolean} - Normalise. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CLASSES}{\code{rasterDestination} - Classification. Path for new raster layer.}

\item{CLASSES_LUT}{\code{vectorDestination} - Look-up Table. Path for new vector layer.}

\item{QUALITY}{\code{rasterDestination} - Quality. Path for new raster layer.}

\item{TRAINING}{\code{source} - Training Areas. Path to a vector layer.}

\item{TRAINING_CLASS}{\code{field} - Class Identifier. The name of an existing field. ; delimited list of existing field names.}

\item{FILE_LOAD}{\code{file} - Load Statistics from File.... Path to a file.}

\item{FILE_SAVE}{\code{file} - Save Statistics to File.... Path to a file.}

\item{METHOD}{\code{enum}  of \verb{("(0) Binary Encoding", "(1) Parallelepiped", "(2) Minimum Distance", "(3) Mahalanobis Distance", "(4) Maximum Likelihood", "(5) Spectral Angle Mapping", "(6) Winner Takes All")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{THRESHOLD_DIST}{\code{number} - Distance Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRESHOLD_ANGLE}{\code{number} - Spectral Angle Threshold (Degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRESHOLD_PROB}{\code{number} - Probability Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RELATIVE_PROB}{\code{enum}  of \verb{("(0) absolute", "(1) relative")} - Probability Reference. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{WTA_0}{\code{boolean} - Binary Encoding. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WTA_1}{\code{boolean} - Parallelepiped. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WTA_2}{\code{boolean} - Minimum Distance. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WTA_3}{\code{boolean} - Mahalanobis Distance. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WTA_4}{\code{boolean} - Maximum Likelihood. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{WTA_5}{\code{boolean} - Spectral Angle Mapping. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Supervised classification for grids (sagang:supervisedclassificationforgrids). ---------------- Arguments ----------------  GRIDS: Features 	Argument type:	multilayer NORMALISE: Normalise 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CLASSES: Classification 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CLASSES_LUT: Look-up Table 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer QUALITY: Quality 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TRAINING: Training Areas (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer TRAINING_CLASS: Class Identifier 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FILE_LOAD: Load Statistics from File... 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file FILE_SAVE: Save Statistics to File... 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file METHOD: Method 	Default value:	2 	Argument type:	enum 	Available values: 		- 0: (0) Binary Encoding 		- 1: (1) Parallelepiped 		- 2: (2) Minimum Distance 		- 3: (3) Mahalanobis Distance 		- 4: (4) Maximum Likelihood 		- 5: (5) Spectral Angle Mapping 		- 6: (6) Winner Takes All 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' THRESHOLD_DIST: Distance Threshold 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRESHOLD_ANGLE: Spectral Angle Threshold (Degree) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRESHOLD_PROB: Probability Threshold 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RELATIVE_PROB: Probability Reference 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) absolute 		- 1: (1) relative 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' WTA_0: Binary Encoding 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WTA_1: Parallelepiped 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WTA_2: Minimum Distance 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WTA_3: Mahalanobis Distance 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WTA_4: Maximum Likelihood 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression WTA_5: Spectral Angle Mapping 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CLASSES - outputRaster - Classification
\item CLASSES_LUT - outputVector - Look-up Table
\item QUALITY - outputRaster - Quality
}
}
}
