% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_surfacegradientandconcentration.R
\name{sagang_surfacegradientandconcentration}
\alias{sagang_surfacegradientandconcentration}
\title{QGIS algorithm - Surface, gradient and concentration}
\usage{
sagang_surfacegradientandconcentration(
  MASK = qgisprocess:::qgis_default_value(),
  SURF = qgisprocess:::qgis_default_value(),
  GRAD = qgisprocess:::qgis_default_value(),
  CONC = qgisprocess:::qgis_default_value(),
  SURF_E = qgisprocess:::qgis_default_value(),
  CONC_IN = qgisprocess:::qgis_default_value(),
  CONC_OUT = qgisprocess:::qgis_default_value(),
  CONC_E = qgisprocess:::qgis_default_value(),
  GRAD_MIN = qgisprocess:::qgis_default_value(),
  NEIGHBOURS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{MASK}{\code{raster} - Mask. Path to a raster layer.}

\item{SURF}{\code{rasterDestination} - Surface. Path for new raster layer.}

\item{GRAD}{\code{rasterDestination} - Gradient. Path for new raster layer.}

\item{CONC}{\code{rasterDestination} - Concentration. Path for new raster layer.}

\item{SURF_E}{\code{number} - Surface Approximation Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CONC_IN}{\code{number} - Inlet Concentration. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CONC_OUT}{\code{number} - Outlet Concentration. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CONC_E}{\code{number} - Concentration Approximation Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRAD_MIN}{\code{number} - Minimum Gradient. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NEIGHBOURS}{\code{enum}  of \verb{("(0) Moore (8)", "(1) Neumann (4)", "(2) Optimised")} - Neighbourhood. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Surface, gradient and concentration (sagang:surfacegradientandconcentration). ---------------- Arguments ----------------  MASK: Mask 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SURF: Surface 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer GRAD: Gradient 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONC: Concentration 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SURF_E: Surface Approximation Threshold 	Default value:	0.001 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CONC_IN: Inlet Concentration 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CONC_OUT: Outlet Concentration 	Default value:	3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CONC_E: Concentration Approximation Threshold 	Default value:	0.001 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GRAD_MIN: Minimum Gradient 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NEIGHBOURS: Neighbourhood 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Moore (8) 		- 1: (1) Neumann (4) 		- 2: (2) Optimised 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONC - outputRaster - Concentration
\item GRAD - outputRaster - Gradient
\item SURF - outputRaster - Surface
}
}
}
