% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_temperaturelapserates.R
\name{sagang_temperaturelapserates}
\alias{sagang_temperaturelapserates}
\title{QGIS algorithm - Temperature lapse rates}
\usage{
sagang_temperaturelapserates(
  TEMP = qgisprocess:::qgis_default_value(),
  TGROUND = qgisprocess:::qgis_default_value(),
  LAPSE = qgisprocess:::qgis_default_value(),
  TEXTREME = qgisprocess:::qgis_default_value(),
  TIME = qgisprocess:::qgis_default_value(),
  EXTREME = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TEMP}{\code{multilayer} - Atmospheric Lapse Rates. .}

\item{TGROUND}{\code{multilayer} - Surface Temperature. .}

\item{LAPSE}{\code{rasterDestination} - Temperature Lapse Rate at Extreme. Path for new raster layer.}

\item{TEXTREME}{\code{rasterDestination} - Daily Extreme Temperature. Path for new raster layer.}

\item{TIME}{\code{rasterDestination} - Hour of Daily Extreme Temperature. Path for new raster layer.}

\item{EXTREME}{\code{enum}  of \verb{("(0) minimum", "(1) maximum")} - Temperature extreme. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Temperature lapse rates (sagang:temperaturelapserates). ---------------- Arguments ----------------  TEMP: Atmospheric Lapse Rates 	Argument type:	multilayer TGROUND: Surface Temperature 	Argument type:	multilayer LAPSE: Temperature Lapse Rate at Extreme 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TEXTREME: Daily Extreme Temperature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TIME: Hour of Daily Extreme Temperature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer EXTREME: Temperature extreme 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) minimum 		- 1: (1) maximum 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item LAPSE - outputRaster - Temperature Lapse Rate at Extreme
\item TEXTREME - outputRaster - Daily Extreme Temperature
\item TIME - outputRaster - Hour of Daily Extreme Temperature
}
}
}
