% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_texturalfeatures.R
\name{sagang_texturalfeatures}
\alias{sagang_texturalfeatures}
\title{QGIS algorithm - Textural features}
\usage{
sagang_texturalfeatures(
  GRID = qgisprocess:::qgis_default_value(),
  ASM = qgisprocess:::qgis_default_value(),
  CONTRAST = qgisprocess:::qgis_default_value(),
  CORRELATION = qgisprocess:::qgis_default_value(),
  VARIANCE = qgisprocess:::qgis_default_value(),
  IDM = qgisprocess:::qgis_default_value(),
  SUM_AVERAGE = qgisprocess:::qgis_default_value(),
  SUM_ENTROPY = qgisprocess:::qgis_default_value(),
  SUM_VARIANCE = qgisprocess:::qgis_default_value(),
  ENTROPY = qgisprocess:::qgis_default_value(),
  DIF_VARIANCE = qgisprocess:::qgis_default_value(),
  DIF_ENTROPY = qgisprocess:::qgis_default_value(),
  MOC_1 = qgisprocess:::qgis_default_value(),
  MOC_2 = qgisprocess:::qgis_default_value(),
  DIRECTION = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  MAX_CATS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{ASM}{\code{rasterDestination} - Angular Second Moment. Path for new raster layer.}

\item{CONTRAST}{\code{rasterDestination} - Contrast. Path for new raster layer.}

\item{CORRELATION}{\code{rasterDestination} - Correlation. Path for new raster layer.}

\item{VARIANCE}{\code{rasterDestination} - Variance. Path for new raster layer.}

\item{IDM}{\code{rasterDestination} - Inverse Diff Moment. Path for new raster layer.}

\item{SUM_AVERAGE}{\code{rasterDestination} - Sum Average. Path for new raster layer.}

\item{SUM_ENTROPY}{\code{rasterDestination} - Sum Entropy. Path for new raster layer.}

\item{SUM_VARIANCE}{\code{rasterDestination} - Sum Variance. Path for new raster layer.}

\item{ENTROPY}{\code{rasterDestination} - Entropy. Path for new raster layer.}

\item{DIF_VARIANCE}{\code{rasterDestination} - Difference Variance. Path for new raster layer.}

\item{DIF_ENTROPY}{\code{rasterDestination} - Difference Entropy. Path for new raster layer.}

\item{MOC_1}{\code{rasterDestination} - Measure of Correlation-1. Path for new raster layer.}

\item{MOC_2}{\code{rasterDestination} - Measure of Correlation-2. Path for new raster layer.}

\item{DIRECTION}{\code{enum}  of \verb{("(0) all", "(1) N-S", "(2) NE-SW", "(3) E-W", "(4) SE-NW")} - Direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DISTANCE}{\code{number} - Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAX_CATS}{\code{number} - Maximum Number of Categories. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Textural features (sagang:texturalfeatures). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer ASM: Angular Second Moment 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONTRAST: Contrast 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CORRELATION: Correlation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VARIANCE: Variance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer IDM: Inverse Diff Moment 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SUM_AVERAGE: Sum Average 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SUM_ENTROPY: Sum Entropy 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SUM_VARIANCE: Sum Variance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ENTROPY: Entropy 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIF_VARIANCE: Difference Variance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIF_ENTROPY: Difference Entropy 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MOC_1: Measure of Correlation-1 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MOC_2: Measure of Correlation-2 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DIRECTION: Direction 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) all 		- 1: (1) N-S 		- 2: (2) NE-SW 		- 3: (3) E-W 		- 4: (4) SE-NW 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' RADIUS: Radius 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DISTANCE: Distance 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAX_CATS: Maximum Number of Categories 	Default value:	256 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item ASM - outputRaster - Angular Second Moment
\item CONTRAST - outputRaster - Contrast
\item CORRELATION - outputRaster - Correlation
\item DIF_ENTROPY - outputRaster - Difference Entropy
\item DIF_VARIANCE - outputRaster - Difference Variance
\item ENTROPY - outputRaster - Entropy
\item IDM - outputRaster - Inverse Diff Moment
\item MOC_1 - outputRaster - Measure of Correlation-1
\item MOC_2 - outputRaster - Measure of Correlation-2
\item SUM_AVERAGE - outputRaster - Sum Average
\item SUM_ENTROPY - outputRaster - Sum Entropy
\item SUM_VARIANCE - outputRaster - Sum Variance
\item VARIANCE - outputRaster - Variance
}
}
}
