% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_thermicbeltclassification.R
\name{sagang_thermicbeltclassification}
\alias{sagang_thermicbeltclassification}
\title{QGIS algorithm - Thermic belt classification}
\usage{
sagang_thermicbeltclassification(
  GSL = qgisprocess:::qgis_default_value(),
  GST = qgisprocess:::qgis_default_value(),
  FROST = qgisprocess:::qgis_default_value(),
  NIVAL_TEMP = qgisprocess:::qgis_default_value(),
  TREE_TEMP = qgisprocess:::qgis_default_value(),
  ATB = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GSL}{\code{raster} - Growing Season Length. Path to a raster layer.}

\item{GST}{\code{raster} - Mean Temperature. Path to a raster layer.}

\item{FROST}{\code{raster} - Frost occurence. Path to a raster layer.}

\item{NIVAL_TEMP}{\code{number} - Minimum Temperature Nival. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TREE_TEMP}{\code{number} - Minimum Temperature Treeline. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ATB}{\code{rasterDestination} - Thermal Belts. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Thermic belt classification (sagang:thermicbeltclassification). ---------------- Arguments ----------------  GSL: Growing Season Length 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GST: Mean Temperature 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer FROST: Frost occurence 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer NIVAL_TEMP: Minimum Temperature Nival 	Default value:	3.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TREE_TEMP: Minimum Temperature Treeline 	Default value:	6.4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ATB: Thermal Belts 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item ATB - outputRaster - Thermal Belts
}
}
}
