% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_tiling.R
\name{sagang_tiling}
\alias{sagang_tiling}
\title{QGIS algorithm - Tiling}
\usage{
sagang_tiling(
  GRID = qgisprocess:::qgis_default_value(),
  TILES = qgisprocess:::qgis_default_value(),
  TILES_SAVE = qgisprocess:::qgis_default_value(),
  TILES_PATH = qgisprocess:::qgis_default_value(),
  TILES_NAME = qgisprocess:::qgis_default_value(),
  OVERLAP = qgisprocess:::qgis_default_value(),
  OVERLAP_SYM = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  NX = qgisprocess:::qgis_default_value(),
  NY = qgisprocess:::qgis_default_value(),
  DCELL = qgisprocess:::qgis_default_value(),
  DX = qgisprocess:::qgis_default_value(),
  DY = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{TILES}{\code{rasterDestination} - Tiles. Path for new raster layer.}

\item{TILES_SAVE}{\code{boolean} - Save Tiles to Disk. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TILES_PATH}{\code{file} - Output Directory. Path to a file.}

\item{TILES_NAME}{\code{string} - Base Name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OVERLAP}{\code{number} - Overlapping Cells. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OVERLAP_SYM}{\code{enum}  of \verb{("(0) symmetric", "(1) bottom / left", "(2) top / right")} - Add Cells. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) number of grid cells per tile", "(1) coordinates (offset, range, cell size, tile size)")} - Tile Size Definition. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NX}{\code{number} - Number of Column Cells. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NY}{\code{number} - Number of Row Cells. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DCELL}{\code{number} - Cell Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DX}{\code{number} - Tile Size (X). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DY}{\code{number} - Tile Size (Y). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Tiling (sagang:tiling). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TILES: Tiles 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TILES_SAVE: Save Tiles to Disk 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TILES_PATH: Output Directory 	Default value:	False 	Argument type:	file 	Acceptable values: 		- Path to a file TILES_NAME: Base Name 	Default value:	tile 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OVERLAP: Overlapping Cells 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OVERLAP_SYM: Add Cells 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) symmetric 		- 1: (1) bottom / left 		- 2: (2) top / right 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Tile Size Definition 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) number of grid cells per tile 		- 1: (1) coordinates (offset, range, cell size, tile size) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NX: Number of Column Cells 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NY: Number of Row Cells 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DCELL: Cell Size 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DX: Tile Size (X) 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DY: Tile Size (Y) 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item TILES - outputRaster - Tiles
}
}
}
