% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_topmodel.R
\name{sagang_topmodel}
\alias{sagang_topmodel}
\title{QGIS algorithm - Topmodel}
\usage{
sagang_topmodel(
  ATANB = qgisprocess:::qgis_default_value(),
  MOIST = qgisprocess:::qgis_default_value(),
  WEATHER = qgisprocess:::qgis_default_value(),
  RECORD_P = qgisprocess:::qgis_default_value(),
  RECORD_ET = qgisprocess:::qgis_default_value(),
  RECORD_DATE = qgisprocess:::qgis_default_value(),
  TABLE = qgisprocess:::qgis_default_value(),
  DTIME = qgisprocess:::qgis_default_value(),
  NCLASSES = qgisprocess:::qgis_default_value(),
  P_QS0 = qgisprocess:::qgis_default_value(),
  P_LNTE = qgisprocess:::qgis_default_value(),
  P_MODEL = qgisprocess:::qgis_default_value(),
  P_SR0 = qgisprocess:::qgis_default_value(),
  P_SRZMAX = qgisprocess:::qgis_default_value(),
  P_SUZ_TD = qgisprocess:::qgis_default_value(),
  P_VCH = qgisprocess:::qgis_default_value(),
  P_VR = qgisprocess:::qgis_default_value(),
  P_K0 = qgisprocess:::qgis_default_value(),
  P_PSI = qgisprocess:::qgis_default_value(),
  P_DTHETA = qgisprocess:::qgis_default_value(),
  BINF = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ATANB}{\code{raster} - Topographic Wetness Index. Path to a raster layer.}

\item{MOIST}{\code{rasterDestination} - Soil Moisture Deficit. Path for new raster layer.}

\item{WEATHER}{\code{source} - Weather Records. Path to a vector layer.}

\item{RECORD_P}{\code{field} - Precipitation (m \code{/} dt). The name of an existing field. ; delimited list of existing field names.}

\item{RECORD_ET}{\code{field} - Evapotranspiration (m \code{/} dt). The name of an existing field. ; delimited list of existing field names.}

\item{RECORD_DATE}{\code{field} - Date\code{/}Time. The name of an existing field. ; delimited list of existing field names.}

\item{TABLE}{\code{vectorDestination} - Simulation Output. Path for new vector layer.}

\item{DTIME}{\code{number} - Time Step (h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NCLASSES}{\code{number} - Number of Classes. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_QS0}{\code{number} - Initial subsurface flow per unit area (m\code{/}h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_LNTE}{\code{number} - Areal average of ln(T0) = ln(Te) (ln(m^2\code{/}h)). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_MODEL}{\code{number} - Model parameter (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_SR0}{\code{number} - Initial root zone storage deficit (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_SRZMAX}{\code{number} - Maximum root zone storage deficit (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_SUZ_TD}{\code{number} - Unsaturated zone time delay per unit storage deficit (h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_VCH}{\code{number} - Main channel routing velocity (m\code{/}h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_VR}{\code{number} - Internal subcatchment routing velocity (m\code{/}h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_K0}{\code{number} - Surface hydraulic conductivity (m\code{/}h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_PSI}{\code{number} - Wetting front suction (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{P_DTHETA}{\code{number} - Water content change across the wetting front. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{BINF}{\code{boolean} - Green-Ampt Infiltration. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Topmodel (sagang:topmodel). ---------------- Arguments ----------------  ATANB: Topographic Wetness Index 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer MOIST: Soil Moisture Deficit 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer WEATHER: Weather Records 	Argument type:	source 	Acceptable values: 		- Path to a vector layer RECORD_P: Precipitation (m / dt) 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names RECORD_ET: Evapotranspiration (m / dt) 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names RECORD_DATE: Date/Time 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TABLE: Simulation Output 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DTIME: Time Step (h) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NCLASSES: Number of Classes 	Default value:	30 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_QS0: Initial subsurface flow per unit area (m/h) 	Default value:	3.3e-05 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_LNTE: Areal average of ln(T0) = ln(Te) (ln(m^2/h)) 	Default value:	5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_MODEL: Model parameter (m) 	Default value:	0.032 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_SR0: Initial root zone storage deficit (m) 	Default value:	0.002 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_SRZMAX: Maximum root zone storage deficit (m) 	Default value:	0.05 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_SUZ_TD: Unsaturated zone time delay per unit storage deficit (h) 	Default value:	50 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_VCH: Main channel routing velocity (m/h) 	Default value:	3600 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_VR: Internal subcatchment routing velocity (m/h) 	Default value:	3600 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_K0: Surface hydraulic conductivity (m/h) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_PSI: Wetting front suction (m) 	Default value:	0.02 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression P_DTHETA: Water content change across the wetting front 	Default value:	0.1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression BINF: Green-Ampt Infiltration 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item MOIST - outputRaster - Soil Moisture Deficit
\item TABLE - outputVector - Simulation Output
}
}
}
