% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_topographiccorrection.R
\name{sagang_topographiccorrection}
\alias{sagang_topographiccorrection}
\title{QGIS algorithm - Topographic correction}
\usage{
sagang_topographiccorrection(
  DEM = qgisprocess:::qgis_default_value(),
  ORIGINAL = qgisprocess:::qgis_default_value(),
  CORRECTED = qgisprocess:::qgis_default_value(),
  AZI = qgisprocess:::qgis_default_value(),
  HGT = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  MINNAERT = qgisprocess:::qgis_default_value(),
  MAXCELLS = qgisprocess:::qgis_default_value(),
  MAXVALUE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{ORIGINAL}{\code{raster} - Original Image. Path to a raster layer.}

\item{CORRECTED}{\code{rasterDestination} - Corrected Image. Path for new raster layer.}

\item{AZI}{\code{number} - Azimuth. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{HGT}{\code{number} - Height. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) Cosine Correction (Teillet et al. 1982)", "(1) Cosine Correction (Civco 1989)", "(2) Minnaert Correction", "(3) Minnaert Correction with Slope (Riano et al. 2003)", "(4) Minnaert Correction with Slope (Law & Nichol 2004)", "(5) C Correction", "(6) Normalization (after Civco, modified by Law & Nichol)")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MINNAERT}{\code{number} - Minnaert Correction. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAXCELLS}{\code{number} - Maximum Cells (C Correction Analysis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAXVALUE}{\code{enum}  of \verb{("(0) 1 byte (0-255)", "(1) 2 byte (0-65535)")} - Value Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Topographic correction (sagang:topographiccorrection). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer ORIGINAL: Original Image 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CORRECTED: Corrected Image 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer AZI: Azimuth 	Default value:	180 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression HGT: Height 	Default value:	45 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Method 	Default value:	4 	Argument type:	enum 	Available values: 		- 0: (0) Cosine Correction (Teillet et al. 1982) 		- 1: (1) Cosine Correction (Civco 1989) 		- 2: (2) Minnaert Correction 		- 3: (3) Minnaert Correction with Slope (Riano et al. 2003) 		- 4: (4) Minnaert Correction with Slope (Law & Nichol 2004) 		- 5: (5) C Correction 		- 6: (6) Normalization (after Civco, modified by Law & Nichol) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' MINNAERT: Minnaert Correction 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAXCELLS: Maximum Cells (C Correction Analysis) 	Default value:	1000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAXVALUE: Value Range 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) 1 byte (0-255) 		- 1: (1) 2 byte (0-65535) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item CORRECTED - outputRaster - Corrected Image
}
}
}
