% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_topographicopenness.R
\name{sagang_topographicopenness}
\alias{sagang_topographicopenness}
\title{QGIS algorithm - Topographic openness}
\usage{
sagang_topographicopenness(
  DEM = qgisprocess:::qgis_default_value(),
  POS = qgisprocess:::qgis_default_value(),
  NEG = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  DLEVEL = qgisprocess:::qgis_default_value(),
  NDIRS = qgisprocess:::qgis_default_value(),
  UNIT = qgisprocess:::qgis_default_value(),
  NADIR = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{POS}{\code{rasterDestination} - Positive Openness. Path for new raster layer.}

\item{NEG}{\code{rasterDestination} - Negative Openness. Path for new raster layer.}

\item{RADIUS}{\code{number} - Radial Limit. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) multi scale", "(1) line tracing")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DLEVEL}{\code{number} - Multi Scale Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NDIRS}{\code{number} - Number of Sectors. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{UNIT}{\code{enum}  of \verb{("(0) Radians", "(1) Degree")} - Unit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NADIR}{\code{boolean} - Difference from Nadir. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Topographic openness (sagang:topographicopenness). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer POS: Positive Openness 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer NEG: Negative Openness 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RADIUS: Radial Limit 	Default value:	10000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Method 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) multi scale 		- 1: (1) line tracing 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' DLEVEL: Multi Scale Factor 	Default value:	3 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NDIRS: Number of Sectors 	Default value:	8 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression UNIT: Unit 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Radians 		- 1: (1) Degree 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NADIR: Difference from Nadir 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item NEG - outputRaster - Negative Openness
\item POS - outputRaster - Positive Openness
}
}
}
