% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_topographicwetnessindextwi.R
\name{sagang_topographicwetnessindextwi}
\alias{sagang_topographicwetnessindextwi}
\title{QGIS algorithm - Topographic wetness index (twi)}
\usage{
sagang_topographicwetnessindextwi(
  SLOPE = qgisprocess:::qgis_default_value(),
  AREA = qgisprocess:::qgis_default_value(),
  TRANS = qgisprocess:::qgis_default_value(),
  TWI = qgisprocess:::qgis_default_value(),
  CONV = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SLOPE}{\code{raster} - Slope. Path to a raster layer.}

\item{AREA}{\code{raster} - Catchment Area. Path to a raster layer.}

\item{TRANS}{\code{raster} - Transmissivity. Path to a raster layer.}

\item{TWI}{\code{rasterDestination} - Topographic Wetness Index. Path for new raster layer.}

\item{CONV}{\code{enum}  of \verb{("(0) no conversion (areas already given as specific catchment area)", "(1) 1 / cell size (pseudo specific catchment area)")} - Area Conversion. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{METHOD}{\code{enum}  of \verb{("(0) Standard", "(1) TOPMODEL")} - Method (TWI). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Topographic wetness index (twi) (sagang:topographicwetnessindextwi). ---------------- Arguments ----------------  SLOPE: Slope 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer AREA: Catchment Area 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TRANS: Transmissivity (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer TWI: Topographic Wetness Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONV: Area Conversion 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) no conversion (areas already given as specific catchment area) 		- 1: (1) 1 / cell size (pseudo specific catchment area) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' METHOD: Method (TWI) 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Standard 		- 1: (1) TOPMODEL 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item TWI - outputRaster - Topographic Wetness Index
}
}
}
