% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_transectthroughpolygonshapefile.R
\name{sagang_transectthroughpolygonshapefile}
\alias{sagang_transectthroughpolygonshapefile}
\title{QGIS algorithm - Transect through polygon shapefile}
\usage{
sagang_transectthroughpolygonshapefile(
  TRANSECT = qgisprocess:::qgis_default_value(),
  THEME = qgisprocess:::qgis_default_value(),
  THEME_FIELD = qgisprocess:::qgis_default_value(),
  TRANSECT_RESULT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TRANSECT}{\code{source} - Line Transect(s). Path to a vector layer.}

\item{THEME}{\code{source} - Theme. Path to a vector layer.}

\item{THEME_FIELD}{\code{field} - Theme Field. The name of an existing field. ; delimited list of existing field names.}

\item{TRANSECT_RESULT}{\code{vectorDestination} - Result table. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Transect through polygon shapefile (sagang:transectthroughpolygonshapefile). ---------------- Arguments ----------------  TRANSECT: Line Transect(s) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer THEME: Theme 	Argument type:	source 	Acceptable values: 		- Path to a vector layer THEME_FIELD: Theme Field 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TRANSECT_RESULT: Result table 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item TRANSECT_RESULT - outputVector - Result table
}
}
}
