% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_transposerasterlayers.R
\name{sagang_transposerasterlayers}
\alias{sagang_transposerasterlayers}
\title{QGIS algorithm - Transpose raster layers}
\usage{
sagang_transposerasterlayers(
  GRIDS = qgisprocess:::qgis_default_value(),
  TRANSPOSED = qgisprocess:::qgis_default_value(),
  MIRROR_X = qgisprocess:::qgis_default_value(),
  MIRROR_Y = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{TRANSPOSED}{\code{rasterDestination} - Transposed Grids. Path for new raster layer.}

\item{MIRROR_X}{\code{boolean} - Mirror Horizontally. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MIRROR_Y}{\code{boolean} - Mirror Vertically. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Transpose raster layers (sagang:transposerasterlayers). ---------------- Arguments ----------------  GRIDS: Grids 	Argument type:	multilayer TRANSPOSED: Transposed Grids 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MIRROR_X: Mirror Horizontally 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MIRROR_Y: Mirror Vertically 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item TRANSPOSED - outputRaster - Transposed Grids
}
}
}
