% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_traveltimecalculation.R
\name{sagang_traveltimecalculation}
\alias{sagang_traveltimecalculation}
\title{QGIS algorithm - Travel time calculation}
\usage{
sagang_traveltimecalculation(
  DESTINATIONS = qgisprocess:::qgis_default_value(),
  LANDCOVER = qgisprocess:::qgis_default_value(),
  LANDCOVER_TO_TRAVELTIME = qgisprocess:::qgis_default_value(),
  FIELD_LANDCOVER = qgisprocess:::qgis_default_value(),
  FIELD_TRAVELTIME = qgisprocess:::qgis_default_value(),
  TRAVELTIME_MINUTES = qgisprocess:::qgis_default_value(),
  TRAVELTIME_LUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DESTINATIONS}{\code{source} - Destination Points. Path to a vector layer.}

\item{LANDCOVER}{\code{raster} - Land Cover. Path to a raster layer.}

\item{LANDCOVER_TO_TRAVELTIME}{\code{source} - Travel Times. Path to a vector layer.}

\item{FIELD_LANDCOVER}{\code{field} - Land Cover ID. The name of an existing field. ; delimited list of existing field names.}

\item{FIELD_TRAVELTIME}{\code{field} - Travel Time. The name of an existing field. ; delimited list of existing field names.}

\item{TRAVELTIME_MINUTES}{\code{rasterDestination} - Travel Time. Path for new raster layer.}

\item{TRAVELTIME_LUT}{\code{source} - Travel Time Zones Classification. Path to a vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Travel time calculation (sagang:traveltimecalculation). ---------------- Arguments ----------------  DESTINATIONS: Destination Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer LANDCOVER: Land Cover 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LANDCOVER_TO_TRAVELTIME: Travel Times 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD_LANDCOVER: Land Cover ID 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FIELD_TRAVELTIME: Travel Time 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TRAVELTIME_MINUTES: Travel Time 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TRAVELTIME_LUT: Travel Time Zones Classification (optional) 	Argument type:	source 	Acceptable values: 		- Path to a vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item TRAVELTIME_MINUTES - outputRaster - Travel Time
}
}
}
