% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_treegrowthseason.R
\name{sagang_treegrowthseason}
\alias{sagang_treegrowthseason}
\title{QGIS algorithm - Tree growth season}
\usage{
sagang_treegrowthseason(
  T = qgisprocess:::qgis_default_value(),
  TMIN = qgisprocess:::qgis_default_value(),
  TMAX = qgisprocess:::qgis_default_value(),
  P = qgisprocess:::qgis_default_value(),
  SWC = qgisprocess:::qgis_default_value(),
  SWC_DEFAULT = qgisprocess:::qgis_default_value(),
  SWC_SURFACE = qgisprocess:::qgis_default_value(),
  SW1_RESIST = qgisprocess:::qgis_default_value(),
  LAT_DEF = qgisprocess:::qgis_default_value(),
  SMT = qgisprocess:::qgis_default_value(),
  SMP = qgisprocess:::qgis_default_value(),
  LGS = qgisprocess:::qgis_default_value(),
  FIRST = qgisprocess:::qgis_default_value(),
  LAST = qgisprocess:::qgis_default_value(),
  TLH = qgisprocess:::qgis_default_value(),
  DT_MIN = qgisprocess:::qgis_default_value(),
  SW_MIN = qgisprocess:::qgis_default_value(),
  LGS_MIN = qgisprocess:::qgis_default_value(),
  SMT_MIN = qgisprocess:::qgis_default_value(),
  TLH_MAX_DIFF = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{T}{\code{multilayer} - Mean Temperature. .}

\item{TMIN}{\code{multilayer} - Minimum Temperature. .}

\item{TMAX}{\code{multilayer} - Maximum Temperature. .}

\item{P}{\code{multilayer} - Precipitation. .}

\item{SWC}{\code{raster} - Soil Water Capacity of Profile. Path to a raster layer.}

\item{SWC_DEFAULT}{\code{number} - Default. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SWC_SURFACE}{\code{number} - Top Soil Water Capacity. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SW1_RESIST}{\code{number} - Transpiration Resistance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LAT_DEF}{\code{number} - Default Latitude. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SMT}{\code{rasterDestination} - Mean Temperature. Path for new raster layer.}

\item{SMP}{\code{rasterDestination} - Precipitation Sum. Path for new raster layer.}

\item{LGS}{\code{rasterDestination} - Length. Path for new raster layer.}

\item{FIRST}{\code{rasterDestination} - First Growing Day. Path for new raster layer.}

\item{LAST}{\code{rasterDestination} - Last Growing Day. Path for new raster layer.}

\item{TLH}{\code{rasterDestination} - Tree Line Height. Path for new raster layer.}

\item{DT_MIN}{\code{number} - Threshold Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SW_MIN}{\code{number} - Minimum Soil Water Content (Percent). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LGS_MIN}{\code{number} - Minimum Length. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SMT_MIN}{\code{number} - Minimum Mean Temperature. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TLH_MAX_DIFF}{\code{number} - Maximum Tree Line Height Difference. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Tree growth season (sagang:treegrowthseason). ---------------- Arguments ----------------  T: Mean Temperature 	Argument type:	multilayer TMIN: Minimum Temperature 	Argument type:	multilayer TMAX: Maximum Temperature 	Argument type:	multilayer P: Precipitation 	Argument type:	multilayer SWC: Soil Water Capacity of Profile (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SWC_DEFAULT: Default 	Default value:	220 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SWC_SURFACE: Top Soil Water Capacity 	Default value:	30 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SW1_RESIST: Transpiration Resistance 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LAT_DEF: Default Latitude 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SMT: Mean Temperature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SMP: Precipitation Sum 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LGS: Length 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer FIRST: First Growing Day 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LAST: Last Growing Day 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TLH: Tree Line Height 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer DT_MIN: Threshold Temperature 	Default value:	0.9 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SW_MIN: Minimum Soil Water Content (Percent) 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LGS_MIN: Minimum Length 	Default value:	94 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SMT_MIN: Minimum Mean Temperature 	Default value:	6.4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TLH_MAX_DIFF: Maximum Tree Line Height Difference 	Default value:	3000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item FIRST - outputRaster - First Growing Day
\item LAST - outputRaster - Last Growing Day
\item LGS - outputRaster - Length
\item SMP - outputRaster - Precipitation Sum
\item SMT - outputRaster - Mean Temperature
\item TLH - outputRaster - Tree Line Height
}
}
}
