% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_trendanalysisshapes.R
\name{sagang_trendanalysisshapes}
\alias{sagang_trendanalysisshapes}
\title{QGIS algorithm - Trend analysis (shapes)}
\usage{
sagang_trendanalysisshapes(
  TABLE = qgisprocess:::qgis_default_value(),
  FIELD_X = qgisprocess:::qgis_default_value(),
  FIELD_Y = qgisprocess:::qgis_default_value(),
  FORMULA = qgisprocess:::qgis_default_value(),
  FORMULAS = qgisprocess:::qgis_default_value(),
  TREND = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TABLE}{\code{source} - Shapes. Path to a vector layer.}

\item{FIELD_X}{\code{field} - X Values. The name of an existing field. ; delimited list of existing field names.}

\item{FIELD_Y}{\code{field} - Y Values. The name of an existing field. ; delimited list of existing field names.}

\item{FORMULA}{\code{string} - Formula. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FORMULAS}{\code{enum}  of \verb{("(0) a + b * x (linear)", "(1) a + b * x + c * x^2 (quadric)", "(2) a + b * x + c * x^2 + d * x^3 (cubic)", "(3) a + b * ln(x) (logarithmic)", "(4) a + b * x^c (power)", "(5) a + b / x", "(6) a + b * (1 - exp(-x / c))", "(7) a + b * (1 - exp(-(x / c)^2))")} - Pre-defined Formulas. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{TREND}{\code{vectorDestination} - Table (with Trend). Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Trend analysis (shapes) (sagang:trendanalysisshapes). ---------------- Arguments ----------------  TABLE: Shapes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD_X: X Values 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FIELD_Y: Y Values 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names FORMULA: Formula 	Default value:	m * x + b 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression FORMULAS: Pre-defined Formulas 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) a + b * x (linear) 		- 1: (1) a + b * x + c * x^2 (quadric) 		- 2: (2) a + b * x + c * x^2 + d * x^3 (cubic) 		- 3: (3) a + b * ln(x) (logarithmic) 		- 4: (4) a + b * x^c (power) 		- 5: (5) a + b / x 		- 6: (6) a + b * (1 - exp(-x / c)) 		- 7: (7) a + b * (1 - exp(-(x / c)^2)) 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' TREND: Table (with Trend) 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item TREND - outputVector - Table (with Trend)
}
}
}
