% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_uniquevaluestatisticsforgrids.R
\name{sagang_uniquevaluestatisticsforgrids}
\alias{sagang_uniquevaluestatisticsforgrids}
\title{QGIS algorithm - Unique value statistics for grids}
\usage{
sagang_uniquevaluestatisticsforgrids(
  GRIDS = qgisprocess:::qgis_default_value(),
  MAJORITY = qgisprocess:::qgis_default_value(),
  MINORITY = qgisprocess:::qgis_default_value(),
  NUNIQUES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Values. .}

\item{MAJORITY}{\code{rasterDestination} - Majority. Path for new raster layer.}

\item{MINORITY}{\code{rasterDestination} - Minority. Path for new raster layer.}

\item{NUNIQUES}{\code{rasterDestination} - Number of Unique Values. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Unique value statistics for grids (sagang:uniquevaluestatisticsforgrids). ---------------- Arguments ----------------  GRIDS: Values 	Argument type:	multilayer MAJORITY: Majority 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MINORITY: Minority 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer NUNIQUES: Number of Unique Values 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item MAJORITY - outputRaster - Majority
\item MINORITY - outputRaster - Minority
\item NUNIQUES - outputRaster - Number of Unique Values
}
}
}
