% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_universalimagequalityindex.R
\name{sagang_universalimagequalityindex}
\alias{sagang_universalimagequalityindex}
\title{QGIS algorithm - Universal image quality index}
\usage{
sagang_universalimagequalityindex(
  GRID_A = qgisprocess:::qgis_default_value(),
  GRID_B = qgisprocess:::qgis_default_value(),
  QUALITY = qgisprocess:::qgis_default_value(),
  CORRELATION = qgisprocess:::qgis_default_value(),
  LUMINANCE = qgisprocess:::qgis_default_value(),
  CONTRAST = qgisprocess:::qgis_default_value(),
  K1 = qgisprocess:::qgis_default_value(),
  K2 = qgisprocess:::qgis_default_value(),
  L = qgisprocess:::qgis_default_value(),
  KERNEL_TYPE = qgisprocess:::qgis_default_value(),
  KERNEL_RADIUS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID_A}{\code{raster} - First Grid. Path to a raster layer.}

\item{GRID_B}{\code{raster} - Second Grid. Path to a raster layer.}

\item{QUALITY}{\code{rasterDestination} - Universal Image Quality Index. Path for new raster layer.}

\item{CORRELATION}{\code{rasterDestination} - Correlation. Path for new raster layer.}

\item{LUMINANCE}{\code{rasterDestination} - Luminance. Path for new raster layer.}

\item{CONTRAST}{\code{rasterDestination} - Contrast. Path for new raster layer.}

\item{K1}{\code{number} - k1. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{K2}{\code{number} - k2. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{L}{\code{number} - L. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KERNEL_TYPE}{\code{enum}  of \verb{("(0) Square", "(1) Circle")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_RADIUS}{\code{number} - Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Universal image quality index (sagang:universalimagequalityindex). ---------------- Arguments ----------------  GRID_A: First Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer GRID_B: Second Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer QUALITY: Universal Image Quality Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CORRELATION: Correlation 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer LUMINANCE: Luminance 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CONTRAST: Contrast 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer K1: k1 	Default value:	0.01 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression K2: k2 	Default value:	0.03 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression L: L 	Default value:	255 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KERNEL_TYPE: Kernel Type 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Square 		- 1: (1) Circle 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' KERNEL_RADIUS: Radius 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CONTRAST - outputRaster - Contrast
\item CORRELATION - outputRaster - Correlation
\item LUMINANCE - outputRaster - Luminance
\item QUALITY - outputRaster - Universal Image Quality Index
}
}
}
