% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_universalkriging.R
\name{sagang_universalkriging}
\alias{sagang_universalkriging}
\title{QGIS algorithm - Universal kriging}
\usage{
sagang_universalkriging(
  POINTS = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX =
    qgisprocess:::qgis_default_value(),
  TARGET_USER_SIZE = qgisprocess:::qgis_default_value(),
  PREDICTION = qgisprocess:::qgis_default_value(),
  VARIANCE = qgisprocess:::qgis_default_value(),
  TQUALITY = qgisprocess:::qgis_default_value(),
  VAR_MAXDIST = qgisprocess:::qgis_default_value(),
  VAR_NCLASSES = qgisprocess:::qgis_default_value(),
  VAR_NSKIP = qgisprocess:::qgis_default_value(),
  VAR_MODEL = qgisprocess:::qgis_default_value(),
  LOG = qgisprocess:::qgis_default_value(),
  BLOCK = qgisprocess:::qgis_default_value(),
  DBLOCK = qgisprocess:::qgis_default_value(),
  CV_METHOD = qgisprocess:::qgis_default_value(),
  CV_SUMMARY = qgisprocess:::qgis_default_value(),
  CV_RESIDUALS = qgisprocess:::qgis_default_value(),
  CV_SAMPLES = qgisprocess:::qgis_default_value(),
  SEARCH_RANGE = qgisprocess:::qgis_default_value(),
  SEARCH_RADIUS = qgisprocess:::qgis_default_value(),
  SEARCH_POINTS_ALL = qgisprocess:::qgis_default_value(),
  SEARCH_POINTS_MIN = qgisprocess:::qgis_default_value(),
  SEARCH_POINTS_MAX = qgisprocess:::qgis_default_value(),
  PREDICTORS = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  COORDS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{TARGET_USER_XMIN_TARGET_USER_XMAX_TARGET_USER_YMIN_TARGET_USER_YMAX}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used.. Original algorithm parameter name: TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX.}

\item{TARGET_USER_SIZE}{\code{number} - Cellsize. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PREDICTION}{\code{rasterDestination} - Prediction. Path for new raster layer.}

\item{VARIANCE}{\code{rasterDestination} - Prediction Error. Path for new raster layer.}

\item{TQUALITY}{\code{enum}  of \verb{("(0) Standard Deviation", "(1) Variance")} - Error Measure. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{VAR_MAXDIST}{\code{number} - Maximum Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{VAR_NCLASSES}{\code{number} - Lag Distance Classes. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{VAR_NSKIP}{\code{number} - Skip. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{VAR_MODEL}{\code{string} - Model. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{LOG}{\code{boolean} - Logarithmic Transformation. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{BLOCK}{\code{boolean} - Block Kriging. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DBLOCK}{\code{number} - Block Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CV_METHOD}{\code{enum}  of \verb{("(0) none", "(1) leave one out", "(2) 2-fold", "(3) k-fold")} - Cross Validation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CV_SUMMARY}{\code{vectorDestination} - Cross Validation Summary. Path for new vector layer.}

\item{CV_RESIDUALS}{\code{vectorDestination} - Cross Validation Residuals. Path for new vector layer.}

\item{CV_SAMPLES}{\code{number} - Cross Validation Subsamples. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SEARCH_RANGE}{\code{enum}  of \verb{("(0) local", "(1) global")} - Search Range. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_RADIUS}{\code{number} - Maximum Search Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SEARCH_POINTS_ALL}{\code{enum}  of \verb{("(0) maximum number of nearest points", "(1) all points within search distance")} - Number of Points. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEARCH_POINTS_MIN}{\code{number} - Minimum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SEARCH_POINTS_MAX}{\code{number} - Maximum. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PREDICTORS}{\code{multilayer} - Predictors. .}

\item{RESAMPLING}{\code{enum}  of \verb{("(0) Nearest Neighbour", "(1) Bilinear Interpolation", "(2) Bicubic Spline Interpolation", "(3) B-Spline Interpolation")} - Resampling. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{COORDS}{\code{boolean} - Coordinates. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Universal kriging (sagang:universalkriging). ---------------- Arguments ----------------  POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD: Attribute 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names TARGET_USER_XMIN TARGET_USER_XMAX TARGET_USER_YMIN TARGET_USER_YMAX: Output extent (optional) 	Argument type:	extent 	Acceptable values: 		- A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105' 		- Path to a layer. The extent of the layer is used. TARGET_USER_SIZE: Cellsize 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PREDICTION: Prediction 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VARIANCE: Prediction Error 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TQUALITY: Error Measure 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Standard Deviation 		- 1: (1) Variance 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' VAR_MAXDIST: Maximum Distance 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression VAR_NCLASSES: Lag Distance Classes 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression VAR_NSKIP: Skip 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression VAR_MODEL: Model 	Default value:	a + b * x 	Argument type:	string 	Acceptable values: 		- String value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression LOG: Logarithmic Transformation 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression BLOCK: Block Kriging 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DBLOCK: Block Size 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression CV_METHOD: Cross Validation 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) none 		- 1: (1) leave one out 		- 2: (2) 2-fold 		- 3: (3) k-fold 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CV_SUMMARY: Cross Validation Summary 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer CV_RESIDUALS: Cross Validation Residuals 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer CV_SAMPLES: Cross Validation Subsamples 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_RANGE: Search Range 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) local 		- 1: (1) global 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SEARCH_RADIUS: Maximum Search Distance 	Default value:	1000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_POINTS_ALL: Number of Points 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) maximum number of nearest points 		- 1: (1) all points within search distance 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SEARCH_POINTS_MIN: Minimum 	Default value:	16 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SEARCH_POINTS_MAX: Maximum 	Default value:	20 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PREDICTORS: Predictors (optional) 	Argument type:	multilayer RESAMPLING: Resampling 	Default value:	3 	Argument type:	enum 	Available values: 		- 0: (0) Nearest Neighbour 		- 1: (1) Bilinear Interpolation 		- 2: (2) Bicubic Spline Interpolation 		- 3: (3) B-Spline Interpolation 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' COORDS: Coordinates 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CV_RESIDUALS - outputVector - Cross Validation Residuals
\item CV_SUMMARY - outputVector - Cross Validation Summary
\item PREDICTION - outputRaster - Prediction
\item VARIANCE - outputRaster - Prediction Error
}
}
}
