% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_upslopeanddownslopecurvature.R
\name{sagang_upslopeanddownslopecurvature}
\alias{sagang_upslopeanddownslopecurvature}
\title{QGIS algorithm - Upslope and downslope curvature}
\usage{
sagang_upslopeanddownslopecurvature(
  DEM = qgisprocess:::qgis_default_value(),
  C_LOCAL = qgisprocess:::qgis_default_value(),
  C_UP = qgisprocess:::qgis_default_value(),
  C_UP_LOCAL = qgisprocess:::qgis_default_value(),
  C_DOWN = qgisprocess:::qgis_default_value(),
  C_DOWN_LOCAL = qgisprocess:::qgis_default_value(),
  WEIGHTING = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{C_LOCAL}{\code{rasterDestination} - Local Curvature. Path for new raster layer.}

\item{C_UP}{\code{rasterDestination} - Upslope Curvature. Path for new raster layer.}

\item{C_UP_LOCAL}{\code{rasterDestination} - Local Upslope Curvature. Path for new raster layer.}

\item{C_DOWN}{\code{rasterDestination} - Downslope Curvature. Path for new raster layer.}

\item{C_DOWN_LOCAL}{\code{rasterDestination} - Local Downslope Curvature. Path for new raster layer.}

\item{WEIGHTING}{\code{number} - Upslope Weighting. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Upslope and downslope curvature (sagang:upslopeanddownslopecurvature). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer C_LOCAL: Local Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_UP: Upslope Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_UP_LOCAL: Local Upslope Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_DOWN: Downslope Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer C_DOWN_LOCAL: Local Downslope Curvature 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer WEIGHTING: Upslope Weighting 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item C_DOWN - outputRaster - Downslope Curvature
\item C_DOWN_LOCAL - outputRaster - Local Downslope Curvature
\item C_LOCAL - outputRaster - Local Curvature
\item C_UP - outputRaster - Upslope Curvature
\item C_UP_LOCAL - outputRaster - Local Upslope Curvature
}
}
}
