% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_valleyandridgedetectiontophatapproach.R
\name{sagang_valleyandridgedetectiontophatapproach}
\alias{sagang_valleyandridgedetectiontophatapproach}
\title{QGIS algorithm - Valley and ridge detection (top hat approach)}
\usage{
sagang_valleyandridgedetectiontophatapproach(
  DEM = qgisprocess:::qgis_default_value(),
  VALLEY = qgisprocess:::qgis_default_value(),
  HILL = qgisprocess:::qgis_default_value(),
  VALLEY_IDX = qgisprocess:::qgis_default_value(),
  HILL_IDX = qgisprocess:::qgis_default_value(),
  SLOPE_IDX = qgisprocess:::qgis_default_value(),
  RADIUS_VALLEY = qgisprocess:::qgis_default_value(),
  RADIUS_HILL = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{VALLEY}{\code{rasterDestination} - Valley Depth. Path for new raster layer.}

\item{HILL}{\code{rasterDestination} - Hill Height. Path for new raster layer.}

\item{VALLEY_IDX}{\code{rasterDestination} - Valley Index. Path for new raster layer.}

\item{HILL_IDX}{\code{rasterDestination} - Hill Index. Path for new raster layer.}

\item{SLOPE_IDX}{\code{rasterDestination} - Hillslope Index. Path for new raster layer.}

\item{RADIUS_VALLEY}{\code{number} - Valley Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{RADIUS_HILL}{\code{number} - Hill Radius. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{THRESHOLD}{\code{number} - Elevation Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) default", "(1) alternative")} - Slope Index. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Valley and ridge detection (top hat approach) (sagang:valleyandridgedetectiontophatapproach). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer VALLEY: Valley Depth 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer HILL: Hill Height 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer VALLEY_IDX: Valley Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer HILL_IDX: Hill Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SLOPE_IDX: Hillslope Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RADIUS_VALLEY: Valley Radius 	Default value:	1000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression RADIUS_HILL: Hill Radius 	Default value:	1000 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression THRESHOLD: Elevation Threshold 	Default value:	100 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Slope Index 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) default 		- 1: (1) alternative 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item HILL - outputRaster - Hill Height
\item HILL_IDX - outputRaster - Hill Index
\item SLOPE_IDX - outputRaster - Hillslope Index
\item VALLEY - outputRaster - Valley Depth
\item VALLEY_IDX - outputRaster - Valley Index
}
}
}
