% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_valleydepth.R
\name{sagang_valleydepth}
\alias{sagang_valleydepth}
\title{QGIS algorithm - Valley depth}
\usage{
sagang_valleydepth(
  ELEVATION = qgisprocess:::qgis_default_value(),
  VALLEY_DEPTH = qgisprocess:::qgis_default_value(),
  RIDGE_LEVEL = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  MAXITER = qgisprocess:::qgis_default_value(),
  NOUNDERGROUND = qgisprocess:::qgis_default_value(),
  ORDER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{VALLEY_DEPTH}{\code{rasterDestination} - Valley Depth. Path for new raster layer.}

\item{RIDGE_LEVEL}{\code{rasterDestination} - Ridge Level. Path for new raster layer.}

\item{THRESHOLD}{\code{number} - Tension Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{MAXITER}{\code{number} - Maximum Iterations. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NOUNDERGROUND}{\code{boolean} - Keep Ridge Level above Surface. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ORDER}{\code{number} - Ridge Detection Threshold. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Valley depth (sagang:valleydepth). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer VALLEY_DEPTH: Valley Depth 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RIDGE_LEVEL: Ridge Level 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer THRESHOLD: Tension Threshold 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression MAXITER: Maximum Iterations 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression NOUNDERGROUND: Keep Ridge Level above Surface 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ORDER: Ridge Detection Threshold 	Default value:	4 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item RIDGE_LEVEL - outputRaster - Ridge Level
\item VALLEY_DEPTH - outputRaster - Valley Depth
}
}
}
