% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_vectorisinggridclasses.R
\name{sagang_vectorisinggridclasses}
\alias{sagang_vectorisinggridclasses}
\title{QGIS algorithm - Vectorising grid classes}
\usage{
sagang_vectorisinggridclasses(
  GRID = qgisprocess:::qgis_default_value(),
  POLYGONS = qgisprocess:::qgis_default_value(),
  CLASS_ALL = qgisprocess:::qgis_default_value(),
  CLASS_ID = qgisprocess:::qgis_default_value(),
  SPLIT = qgisprocess:::qgis_default_value(),
  ALLVERTICES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{POLYGONS}{\code{vectorDestination} - Polygons. Path for new vector layer.}

\item{CLASS_ALL}{\code{enum}  of \verb{("(0) one single class specified by class identifier", "(1) all classes")} - Class Selection. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CLASS_ID}{\code{number} - Class Identifier. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SPLIT}{\code{enum}  of \verb{("(0) one single (multi-)polygon object", "(1) each island as separated polygon")} - Vectorised class as.... Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{ALLVERTICES}{\code{boolean} - Keep Vertices on Straight Lines. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Vectorising grid classes (sagang:vectorisinggridclasses). ---------------- Arguments ----------------  GRID: Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer POLYGONS: Polygons 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer CLASS_ALL: Class Selection 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) one single class specified by class identifier 		- 1: (1) all classes 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' CLASS_ID: Class Identifier 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression SPLIT: Vectorised class as... 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) one single (multi-)polygon object 		- 1: (1) each island as separated polygon 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' ALLVERTICES: Keep Vertices on Straight Lines 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item POLYGONS - outputVector - Polygons
}
}
}
