% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_vegetationindexslopebased.R
\name{sagang_vegetationindexslopebased}
\alias{sagang_vegetationindexslopebased}
\title{QGIS algorithm - Vegetation index (slope based)}
\usage{
sagang_vegetationindexslopebased(
  RED = qgisprocess:::qgis_default_value(),
  NIR = qgisprocess:::qgis_default_value(),
  DVI = qgisprocess:::qgis_default_value(),
  NDVI = qgisprocess:::qgis_default_value(),
  RVI = qgisprocess:::qgis_default_value(),
  NRVI = qgisprocess:::qgis_default_value(),
  TVI = qgisprocess:::qgis_default_value(),
  CTVI = qgisprocess:::qgis_default_value(),
  TTVI = qgisprocess:::qgis_default_value(),
  SAVI = qgisprocess:::qgis_default_value(),
  SOIL = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{RED}{\code{raster} - Red Reflectance. Path to a raster layer.}

\item{NIR}{\code{raster} - Near Infrared Reflectance. Path to a raster layer.}

\item{DVI}{\code{rasterDestination} - Difference Vegetation Index. Path for new raster layer.}

\item{NDVI}{\code{rasterDestination} - Normalized Difference Vegetation Index. Path for new raster layer.}

\item{RVI}{\code{rasterDestination} - Ratio Vegetation Index. Path for new raster layer.}

\item{NRVI}{\code{rasterDestination} - Normalized Ratio Vegetation Index. Path for new raster layer.}

\item{TVI}{\code{rasterDestination} - Transformed Vegetation Index. Path for new raster layer.}

\item{CTVI}{\code{rasterDestination} - Corrected Transformed Vegetation Index. Path for new raster layer.}

\item{TTVI}{\code{rasterDestination} - Thiam's Transformed Vegetation Index. Path for new raster layer.}

\item{SAVI}{\code{rasterDestination} - Soil Adjusted Vegetation Index. Path for new raster layer.}

\item{SOIL}{\code{number} - Soil Adjustment Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Vegetation index (slope based) (sagang:vegetationindexslopebased). ---------------- Arguments ----------------  RED: Red Reflectance 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer NIR: Near Infrared Reflectance 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DVI: Difference Vegetation Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer NDVI: Normalized Difference Vegetation Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer RVI: Ratio Vegetation Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer NRVI: Normalized Ratio Vegetation Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TVI: Transformed Vegetation Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer CTVI: Corrected Transformed Vegetation Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer TTVI: Thiam's Transformed Vegetation Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SAVI: Soil Adjusted Vegetation Index 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SOIL: Soil Adjustment Factor 	Default value:	0.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item CTVI - outputRaster - Corrected Transformed Vegetation Index
\item DVI - outputRaster - Difference Vegetation Index
\item NDVI - outputRaster - Normalized Difference Vegetation Index
\item NRVI - outputRaster - Normalized Ratio Vegetation Index
\item RVI - outputRaster - Ratio Vegetation Index
\item SAVI - outputRaster - Soil Adjusted Vegetation Index
\item TTVI - outputRaster - Thiam's Transformed Vegetation Index
\item TVI - outputRaster - Transformed Vegetation Index
}
}
}
