% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_visibilitypoints.R
\name{sagang_visibilitypoints}
\alias{sagang_visibilitypoints}
\title{QGIS algorithm - Visibility (points)}
\usage{
sagang_visibilitypoints(
  ELEVATION = qgisprocess:::qgis_default_value(),
  VISIBILITY = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  FIELD_HEIGHT = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{VISIBILITY}{\code{rasterDestination} - Visibility. Path for new raster layer.}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{FIELD_HEIGHT}{\code{field} - Height. The name of an existing field. ; delimited list of existing field names.}

\item{METHOD}{\code{enum}  of \verb{("(0) Visibility", "(1) Shade", "(2) Distance", "(3) Size")} - Unit. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Visibility (points) (sagang:visibilitypoints). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer VISIBILITY: Visibility 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer POINTS: Points 	Argument type:	source 	Acceptable values: 		- Path to a vector layer FIELD_HEIGHT: Height 	Argument type:	field 	Acceptable values: 		- The name of an existing field 		- ; delimited list of existing field names METHOD: Unit 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Visibility 		- 1: (1) Shade 		- 2: (2) Distance 		- 3: (3) Size 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item VISIBILITY - outputRaster - Visibility
}
}
}
