% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_waterretentioncapacity.R
\name{sagang_waterretentioncapacity}
\alias{sagang_waterretentioncapacity}
\title{QGIS algorithm - Water retention capacity}
\usage{
sagang_waterretentioncapacity(
  SHAPES = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  DEM = qgisprocess:::qgis_default_value(),
  RETENTION = qgisprocess:::qgis_default_value(),
  INTERPOL = qgisprocess:::qgis_default_value(),
  SLOPECORR = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SHAPES}{\code{source} - Plot Holes. Path to a vector layer.}

\item{OUTPUT}{\code{vectorDestination} - Final Parameters. Path for new vector layer.}

\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{RETENTION}{\code{rasterDestination} - Water Retention Capacity. Path for new raster layer.}

\item{INTERPOL}{\code{enum}  of \verb{("(0) Multilevel B-Spline Interpolation", "(1) Inverse Distance Weighted")} - Interpolation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SLOPECORR}{\code{boolean} - Slope Correction. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Water retention capacity (sagang:waterretentioncapacity). ---------------- Arguments ----------------  SHAPES: Plot Holes 	Argument type:	source 	Acceptable values: 		- Path to a vector layer OUTPUT: Final Parameters 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer RETENTION: Water Retention Capacity 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer INTERPOL: Interpolation 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Multilevel B-Spline Interpolation 		- 1: (1) Inverse Distance Weighted 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' SLOPECORR: Slope Correction 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Final Parameters
\item RETENTION - outputRaster - Water Retention Capacity
}
}
}
