% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_watershedbasins.R
\name{sagang_watershedbasins}
\alias{sagang_watershedbasins}
\title{QGIS algorithm - Watershed basins}
\usage{
sagang_watershedbasins(
  ELEVATION = qgisprocess:::qgis_default_value(),
  CHANNELS = qgisprocess:::qgis_default_value(),
  SINKROUTE = qgisprocess:::qgis_default_value(),
  BASINS = qgisprocess:::qgis_default_value(),
  MINSIZE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{CHANNELS}{\code{raster} - Channel Network. Path to a raster layer.}

\item{SINKROUTE}{\code{raster} - Sink Route. Path to a raster layer.}

\item{BASINS}{\code{rasterDestination} - Watershed Basins. Path for new raster layer.}

\item{MINSIZE}{\code{number} - Min. Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Watershed basins (sagang:watershedbasins). ---------------- Arguments ----------------  ELEVATION: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CHANNELS: Channel Network 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer SINKROUTE: Sink Route (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer BASINS: Watershed Basins 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MINSIZE: Min. Size 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item BASINS - outputRaster - Watershed Basins
}
}
}
