% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_watershedbasinsextended.R
\name{sagang_watershedbasinsextended}
\alias{sagang_watershedbasinsextended}
\title{QGIS algorithm - Watershed basins (extended)}
\usage{
sagang_watershedbasinsextended(
  DEM = qgisprocess:::qgis_default_value(),
  CHANNELS = qgisprocess:::qgis_default_value(),
  BASINS = qgisprocess:::qgis_default_value(),
  SUBBASINS = qgisprocess:::qgis_default_value(),
  V_BASINS = qgisprocess:::qgis_default_value(),
  V_SUBBASINS = qgisprocess:::qgis_default_value(),
  HEADS = qgisprocess:::qgis_default_value(),
  MOUTHS = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{CHANNELS}{\code{raster} - Drainage Network. Path to a raster layer.}

\item{BASINS}{\code{rasterDestination} - Basins. Path for new raster layer.}

\item{SUBBASINS}{\code{rasterDestination} - Subbasins. Path for new raster layer.}

\item{V_BASINS}{\code{vectorDestination} - Basins. Path for new vector layer.}

\item{V_SUBBASINS}{\code{vectorDestination} - Subbasins. Path for new vector layer.}

\item{HEADS}{\code{vectorDestination} - River Heads. Path for new vector layer.}

\item{MOUTHS}{\code{vectorDestination} - River Mouths. Path for new vector layer.}

\item{DISTANCE}{\code{boolean} - Flow Distances. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Watershed basins (extended) (sagang:watershedbasinsextended). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CHANNELS: Drainage Network 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer BASINS: Basins 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer SUBBASINS: Subbasins 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer V_BASINS: Basins 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer V_SUBBASINS: Subbasins 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer HEADS: River Heads 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer MOUTHS: River Mouths 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer DISTANCE: Flow Distances 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item BASINS - outputRaster - Basins
\item HEADS - outputVector - River Heads
\item MOUTHS - outputVector - River Mouths
\item SUBBASINS - outputRaster - Subbasins
\item V_BASINS - outputVector - Basins
\item V_SUBBASINS - outputVector - Subbasins
}
}
}
