% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_wedgefail.R
\name{sagang_wedgefail}
\alias{sagang_wedgefail}
\title{QGIS algorithm - Wedgefail}
\usage{
sagang_wedgefail(
  DEM = qgisprocess:::qgis_default_value(),
  C = qgisprocess:::qgis_default_value(),
  D = qgisprocess:::qgis_default_value(),
  Emin = qgisprocess:::qgis_default_value(),
  Emax = qgisprocess:::qgis_default_value(),
  fC = qgisprocess:::qgis_default_value(),
  fD = qgisprocess:::qgis_default_value(),
  fEmin = qgisprocess:::qgis_default_value(),
  fEmax = qgisprocess:::qgis_default_value(),
  ff = qgisprocess:::qgis_default_value(),
  F = qgisprocess:::qgis_default_value(),
  fI = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{C}{\code{raster} - Dip\code{/}Plunge direction grid (degree) . Path to a raster layer.}

\item{D}{\code{raster} - Dip\code{/}Plunge grid (degree) . Path to a raster layer.}

\item{Emin}{\code{raster} - Min friction angle grid (degree) . Path to a raster layer.}

\item{Emax}{\code{raster} - Max friction angle grid (degree) . Path to a raster layer.}

\item{fC}{\code{number} - Global dip\code{/}plunge direction (degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fD}{\code{number} - Global dip\code{/}plunge (degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fEmin}{\code{number} - Min global friction angle (degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fEmax}{\code{number} - Max global friction angle (degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ff}{\code{number} - Cone radius (degree). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{F}{\code{rasterDestination} - Failures. Path for new raster layer.}

\item{fI}{\code{number} - Parameter sampling runs. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{METHOD}{\code{enum}  of \verb{("(0) Slide", "(1) Topple")} - Mode. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Wedgefail (sagang:wedgefail). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer C: Dip/Plunge direction grid (degree)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer D: Dip/Plunge grid (degree)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Emin: Min friction angle grid (degree)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Emax: Max friction angle grid (degree)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer fC: Global dip/plunge direction (degree) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fD: Global dip/plunge (degree) 	Default value:	35 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fEmin: Min global friction angle (degree) 	Default value:	35 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fEmax: Max global friction angle (degree) 	Default value:	35 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ff: Cone radius (degree) 	Default value:	0 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression F: Failures 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer fI: Parameter sampling runs 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression METHOD: Mode 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) Slide 		- 1: (1) Topple 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3'
}
\details{
\subsection{Outputs description}{
\itemize{
\item F - outputRaster - Failures
}
}
}
