% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_wetness.R
\name{sagang_wetness}
\alias{sagang_wetness}
\title{QGIS algorithm - Wetness}
\usage{
sagang_wetness(
  DEM = qgisprocess:::qgis_default_value(),
  Cmin = qgisprocess:::qgis_default_value(),
  Cmax = qgisprocess:::qgis_default_value(),
  Dmin = qgisprocess:::qgis_default_value(),
  Dmax = qgisprocess:::qgis_default_value(),
  Emin = qgisprocess:::qgis_default_value(),
  Emax = qgisprocess:::qgis_default_value(),
  fCmin = qgisprocess:::qgis_default_value(),
  fCmax = qgisprocess:::qgis_default_value(),
  fDmin = qgisprocess:::qgis_default_value(),
  fDmax = qgisprocess:::qgis_default_value(),
  fEmin = qgisprocess:::qgis_default_value(),
  fEmax = qgisprocess:::qgis_default_value(),
  fH = qgisprocess:::qgis_default_value(),
  F = qgisprocess:::qgis_default_value(),
  G = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  PREPROC = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{Cmin}{\code{raster} - Min hydraulic conductivity grid (m\code{/}hr) . Path to a raster layer.}

\item{Cmax}{\code{raster} - Max hydraulic conductivity grid (m\code{/}hr) . Path to a raster layer.}

\item{Dmin}{\code{raster} - Min groundwater recharge grid (m\code{/}hr) . Path to a raster layer.}

\item{Dmax}{\code{raster} - Max groundwater recharge grid (m\code{/}hr) . Path to a raster layer.}

\item{Emin}{\code{raster} - Min material depth grid (m). Path to a raster layer.}

\item{Emax}{\code{raster} - Max material depth grid (m). Path to a raster layer.}

\item{fCmin}{\code{number} - Min global material conductivity (m\code{/}hr). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fCmax}{\code{number} - Max global material conductivity (m\code{/}hr). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fDmin}{\code{number} - Min global groundwater recharge (m\code{/}hr). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fDmax}{\code{number} - Max global groundwater recharge (m\code{/}hr). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fEmin}{\code{number} - Min global material depth (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fEmax}{\code{number} - Max global material depth (m). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{fH}{\code{number} - Parameter sampling runs. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{F}{\code{rasterDestination} - WI values. Path for new raster layer.}

\item{G}{\code{rasterDestination} - WI classes. Path for new raster layer.}

\item{METHOD}{\code{enum}  of \verb{("(0) Deterministic 8", "(1) Rho 8", "(2) Braunschweiger Reliefmodell", "(3) Deterministic Infinity", "(4) Multiple Flow Direction", "(5) Multiple Triangular Flow Directon")} - Catchment Area Calculation. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PREPROC}{\code{boolean} - Preprocessing. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Wetness (sagang:wetness). ---------------- Arguments ----------------  DEM: DEM 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Cmin: Min hydraulic conductivity grid (m/hr)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Cmax: Max hydraulic conductivity grid (m/hr)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Dmin: Min groundwater recharge grid (m/hr)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Dmax: Max groundwater recharge grid (m/hr)  (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Emin: Min material depth grid (m) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer Emax: Max material depth grid (m) (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer fCmin: Min global material conductivity (m/hr) 	Default value:	2.7 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fCmax: Max global material conductivity (m/hr) 	Default value:	2.7 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fDmin: Min global groundwater recharge (m/hr) 	Default value:	0.001 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fDmax: Max global groundwater recharge (m/hr) 	Default value:	0.001 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fEmin: Min global material depth (m) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fEmax: Max global material depth (m) 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression fH: Parameter sampling runs 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression F: WI values 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer G: WI classes 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer METHOD: Catchment Area Calculation 	Default value:	4 	Argument type:	enum 	Available values: 		- 0: (0) Deterministic 8 		- 1: (1) Rho 8 		- 2: (2) Braunschweiger Reliefmodell 		- 3: (3) Deterministic Infinity 		- 4: (4) Multiple Flow Direction 		- 5: (5) Multiple Triangular Flow Directon 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' PREPROC: Preprocessing 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item F - outputRaster - WI values
\item G - outputRaster - WI classes
}
}
}
