% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_windeffectcorrection.R
\name{sagang_windeffectcorrection}
\alias{sagang_windeffectcorrection}
\title{QGIS algorithm - Wind effect correction}
\usage{
sagang_windeffectcorrection(
  BOUNDARY = qgisprocess:::qgis_default_value(),
  WIND = qgisprocess:::qgis_default_value(),
  OBSERVED = qgisprocess:::qgis_default_value(),
  B_GRID = qgisprocess:::qgis_default_value(),
  WINDCORR = qgisprocess:::qgis_default_value(),
  B_SOURCE = qgisprocess:::qgis_default_value(),
  B_CONST = qgisprocess:::qgis_default_value(),
  B_MAX = qgisprocess:::qgis_default_value(),
  B_STEPS = qgisprocess:::qgis_default_value(),
  KERNEL_TYPE = qgisprocess:::qgis_default_value(),
  KERNEL_SIZE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{BOUNDARY}{\code{raster} - Boundary Layer. Path to a raster layer.}

\item{WIND}{\code{raster} - Wind Effect. Path to a raster layer.}

\item{OBSERVED}{\code{raster} - Observations. Path to a raster layer.}

\item{B_GRID}{\code{rasterDestination} - Calibrated Scaling Factor. Path for new raster layer.}

\item{WINDCORR}{\code{rasterDestination} - Corrected Wind Effect. Path for new raster layer.}

\item{B_SOURCE}{\code{enum}  of \verb{("(0) constant", "(1) calibrate")} - Scaling Factor. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{B_CONST}{\code{number} - Constant Scaling Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{B_MAX}{\code{number} - Maximum Scaling Factor. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{B_STEPS}{\code{number} - Number of Steps. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{KERNEL_TYPE}{\code{enum}  of \verb{("(0) Square", "(1) Circle")} - Kernel Type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{KERNEL_SIZE}{\code{number} - Kernel Size. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Wind effect correction (sagang:windeffectcorrection). ---------------- Arguments ----------------  BOUNDARY: Boundary Layer 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer WIND: Wind Effect 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OBSERVED: Observations 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer B_GRID: Calibrated Scaling Factor 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer WINDCORR: Corrected Wind Effect 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer B_SOURCE: Scaling Factor 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) constant 		- 1: (1) calibrate 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' B_CONST: Constant Scaling Factor 	Default value:	0.01 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression B_MAX: Maximum Scaling Factor 	Default value:	0.05 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression B_STEPS: Number of Steps 	Default value:	10 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression KERNEL_TYPE: Kernel Type 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Square 		- 1: (1) Circle 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' KERNEL_SIZE: Kernel Size 	Default value:	2 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item B_GRID - outputRaster - Calibrated Scaling Factor
\item WINDCORR - outputRaster - Corrected Wind Effect
}
}
}
