% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_windeffectwindwardleewardindex.R
\name{sagang_windeffectwindwardleewardindex}
\alias{sagang_windeffectwindwardleewardindex}
\title{QGIS algorithm - Wind effect (windward / leeward index)}
\usage{
sagang_windeffectwindwardleewardindex(
  DEM = qgisprocess:::qgis_default_value(),
  DIR = qgisprocess:::qgis_default_value(),
  DIR_UNITS = qgisprocess:::qgis_default_value(),
  LEN = qgisprocess:::qgis_default_value(),
  LEN_SCALE = qgisprocess:::qgis_default_value(),
  EFFECT = qgisprocess:::qgis_default_value(),
  AFH = qgisprocess:::qgis_default_value(),
  MAXDIST = qgisprocess:::qgis_default_value(),
  DIR_CONST = qgisprocess:::qgis_default_value(),
  OLDVER = qgisprocess:::qgis_default_value(),
  ACCEL = qgisprocess:::qgis_default_value(),
  PYRAMIDS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{DIR}{\code{raster} - Wind Direction. Path to a raster layer.}

\item{DIR_UNITS}{\code{enum}  of \verb{("(0) radians", "(1) degree")} - Wind Direction Units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{LEN}{\code{raster} - Wind Speed. Path to a raster layer.}

\item{LEN_SCALE}{\code{number} - Scaling. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{EFFECT}{\code{rasterDestination} - Wind Effect. Path for new raster layer.}

\item{AFH}{\code{rasterDestination} - Effective Air Flow Heights. Path for new raster layer.}

\item{MAXDIST}{\code{number} - Search Distance (km). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DIR_CONST}{\code{number} - Constant Wind Direction. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OLDVER}{\code{boolean} - Old Version. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ACCEL}{\code{number} - Acceleration. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{PYRAMIDS}{\code{boolean} - Elevation Averaging. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Wind effect (windward / leeward index) (sagang:windeffectwindwardleewardindex). ---------------- Arguments ----------------  DEM: Elevation 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DIR: Wind Direction (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer DIR_UNITS: Wind Direction Units 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) radians 		- 1: (1) degree 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' LEN: Wind Speed (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer LEN_SCALE: Scaling 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression EFFECT: Wind Effect 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer AFH: Effective Air Flow Heights 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer MAXDIST: Search Distance (km) 	Default value:	300 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression DIR_CONST: Constant Wind Direction 	Default value:	135 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression OLDVER: Old Version 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ACCEL: Acceleration 	Default value:	1.5 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression PYRAMIDS: Elevation Averaging 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item AFH - outputRaster - Effective Air Flow Heights
\item EFFECT - outputRaster - Wind Effect
}
}
}
