% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_womblingedgedetection.R
\name{sagang_womblingedgedetection}
\alias{sagang_womblingedgedetection}
\title{QGIS algorithm - Wombling (edge detection)}
\usage{
sagang_womblingedgedetection(
  TMAGNITUDE = qgisprocess:::qgis_default_value(),
  TDIRECTION = qgisprocess:::qgis_default_value(),
  TNEIGHBOUR = qgisprocess:::qgis_default_value(),
  ALIGNMENT = qgisprocess:::qgis_default_value(),
  NEIGHBOUR = qgisprocess:::qgis_default_value(),
  FEATURE = qgisprocess:::qgis_default_value(),
  EDGE_POINTS = qgisprocess:::qgis_default_value(),
  EDGE_LINES = qgisprocess:::qgis_default_value(),
  GRADIENTS_OUT = qgisprocess:::qgis_default_value(),
  GRADIENTS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TMAGNITUDE}{\code{number} - Minimum Magnitude. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TDIRECTION}{\code{number} - Maximum Angle. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TNEIGHBOUR}{\code{number} - Minimum Neighbours. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ALIGNMENT}{\code{enum}  of \verb{("(0) between cells", "(1) on cell")} - Alignment. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NEIGHBOUR}{\code{enum}  of \verb{("(0) Rooke's case", "(1) Queen's case")} - Edge Connectivity. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FEATURE}{\code{raster} - Feature. Path to a raster layer.}

\item{EDGE_POINTS}{\code{vectorDestination} - Edge Points. Path for new vector layer.}

\item{EDGE_LINES}{\code{vectorDestination} - Edge Segments. Path for new vector layer.}

\item{GRADIENTS_OUT}{\code{boolean} - Output of Gradients. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{GRADIENTS}{\code{rasterDestination} - Gradients. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Wombling (edge detection) (sagang:womblingedgedetection). ---------------- Arguments ----------------  TMAGNITUDE: Minimum Magnitude 	Default value:	90 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TDIRECTION: Maximum Angle 	Default value:	30 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TNEIGHBOUR: Minimum Neighbours 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ALIGNMENT: Alignment 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) between cells 		- 1: (1) on cell 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NEIGHBOUR: Edge Connectivity 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Rooke's case 		- 1: (1) Queen's case 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FEATURE: Feature 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer EDGE_POINTS: Edge Points 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer EDGE_LINES: Edge Segments 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer GRADIENTS_OUT: Output of Gradients 	Default value:	false 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression GRADIENTS: Gradients 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer
}
\details{
\subsection{Outputs description}{
\itemize{
\item EDGE_LINES - outputVector - Edge Segments
\item EDGE_POINTS - outputVector - Edge Points
\item GRADIENTS - outputRaster - Gradients
}
}
}
