% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sagang_womblingformultiplefeaturesedgedetection.R
\name{sagang_womblingformultiplefeaturesedgedetection}
\alias{sagang_womblingformultiplefeaturesedgedetection}
\title{QGIS algorithm - Wombling for multiple features (edge detection)}
\usage{
sagang_womblingformultiplefeaturesedgedetection(
  TMAGNITUDE = qgisprocess:::qgis_default_value(),
  TDIRECTION = qgisprocess:::qgis_default_value(),
  TNEIGHBOUR = qgisprocess:::qgis_default_value(),
  ALIGNMENT = qgisprocess:::qgis_default_value(),
  NEIGHBOUR = qgisprocess:::qgis_default_value(),
  FEATURES = qgisprocess:::qgis_default_value(),
  EDGE_CELLS = qgisprocess:::qgis_default_value(),
  OUTPUT_ADD = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ZERO_AS_NODATA = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{TMAGNITUDE}{\code{number} - Minimum Magnitude. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TDIRECTION}{\code{number} - Maximum Angle. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TNEIGHBOUR}{\code{number} - Minimum Neighbours. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ALIGNMENT}{\code{enum}  of \verb{("(0) between cells", "(1) on cell")} - Alignment. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NEIGHBOUR}{\code{enum}  of \verb{("(0) Rooke's case", "(1) Queen's case")} - Edge Connectivity. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FEATURES}{\code{multilayer} - Features. .}

\item{EDGE_CELLS}{\code{rasterDestination} - Edges. Path for new raster layer.}

\item{OUTPUT_ADD}{\code{enum}  of \verb{("(0) none", "(1) gradients", "(2) edge cells")} - Additional Output. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{ZERO_AS_NODATA}{\code{boolean} - Zero as No-Data. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Wombling for multiple features (edge detection) (sagang:womblingformultiplefeaturesedgedetection). ---------------- Arguments ----------------  TMAGNITUDE: Minimum Magnitude 	Default value:	90 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TDIRECTION: Maximum Angle 	Default value:	30 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression TNEIGHBOUR: Minimum Neighbours 	Default value:	1 	Argument type:	number 	Acceptable values: 		- A numeric value 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression ALIGNMENT: Alignment 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) between cells 		- 1: (1) on cell 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' NEIGHBOUR: Edge Connectivity 	Default value:	1 	Argument type:	enum 	Available values: 		- 0: (0) Rooke's case 		- 1: (1) Queen's case 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' FEATURES: Features 	Argument type:	multilayer EDGE_CELLS: Edges 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer OUTPUT_ADD: Additional Output 	Default value:	0 	Argument type:	enum 	Available values: 		- 0: (0) none 		- 1: (1) gradients 		- 2: (2) edge cells 	Acceptable values: 		- Number of selected option, e.g. '1' 		- Comma separated list of options, e.g. '1,3' OUTPUT: Output 	Argument type:	rasterDestination 	Acceptable values: 		- Path for new raster layer ZERO_AS_NODATA: Zero as No-Data 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item EDGE_CELLS - outputRaster - Edges
\item OUTPUT - outputRaster - Output
}
}
}
