% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagang_zonalrasterstatistics.R
\name{sagang_zonalrasterstatistics}
\alias{sagang_zonalrasterstatistics}
\title{QGIS algorithm - Zonal raster statistics}
\usage{
sagang_zonalrasterstatistics(
  ZONES = qgisprocess:::qgis_default_value(),
  CATLIST = qgisprocess:::qgis_default_value(),
  STATLIST = qgisprocess:::qgis_default_value(),
  ASPECT = qgisprocess:::qgis_default_value(),
  OUTTAB = qgisprocess:::qgis_default_value(),
  SHORTNAMES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ZONES}{\code{raster} - Zone Grid. Path to a raster layer.}

\item{CATLIST}{\code{multilayer} - Categorical Grids. .}

\item{STATLIST}{\code{multilayer} - Grids to analyse. .}

\item{ASPECT}{\code{raster} - Aspect. Path to a raster layer.}

\item{OUTTAB}{\code{vectorDestination} - Zonal Statistics. Path for new vector layer.}

\item{SHORTNAMES}{\code{boolean} - Short Field Names. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Next Gen Zonal raster statistics (sagang:zonalrasterstatistics). ---------------- Arguments ----------------  ZONES: Zone Grid 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer CATLIST: Categorical Grids (optional) 	Argument type:	multilayer STATLIST: Grids to analyse (optional) 	Argument type:	multilayer ASPECT: Aspect (optional) 	Argument type:	raster 	Acceptable values: 		- Path to a raster layer OUTTAB: Zonal Statistics 	Argument type:	vectorDestination 	Acceptable values: 		- Path for new vector layer SHORTNAMES: Short Field Names 	Default value:	true 	Argument type:	boolean 	Acceptable values: 		- 1 for true/yes 		- 0 for false/no 		- field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field 		- expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTTAB - outputVector - Zonal Statistics
}
}
}
