% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_buildvirtualraster.R
\name{gdal_buildvirtualraster}
\alias{gdal_buildvirtualraster}
\title{QGIS algorithm Build virtual raster}
\usage{
gdal_buildvirtualraster(
  INPUT = qgisprocess::qgis_default_value(),
  RESOLUTION = qgisprocess::qgis_default_value(),
  SEPARATE = qgisprocess::qgis_default_value(),
  PROJ_DIFFERENCE = qgisprocess::qgis_default_value(),
  ADD_ALPHA = qgisprocess::qgis_default_value(),
  ASSIGN_CRS = qgisprocess::qgis_default_value(),
  RESAMPLING = qgisprocess::qgis_default_value(),
  SRC_NODATA = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input layers. .}

\item{RESOLUTION}{\code{enum}  of \verb{("Average", "Highest", "Lowest")} - Resolution. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEPARATE}{\code{boolean} - Place each input file into a separate band. 1 for true/yes. 0 for false/no.}

\item{PROJ_DIFFERENCE}{\code{boolean} - Allow projection difference. 1 for true/yes. 0 for false/no.}

\item{ADD_ALPHA}{\code{boolean} - Add alpha mask band to VRT when source raster has none. 1 for true/yes. 0 for false/no.}

\item{ASSIGN_CRS}{\code{crs} - Override projection for the output file. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{RESAMPLING}{\code{enum}  of \verb{("Nearest Neighbour", "Bilinear", "Cubic Convolution", "B-Spline Convolution", "Lanczos Windowed Sinc", "Average", "Mode")} - Resampling algorithm. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SRC_NODATA}{\code{string} - Nodata value(s) for input bands (space separated). String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Virtual. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Build virtual raster (gdal:buildvirtualraster)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Virtual
}
}
}
