% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_cliprasterbymasklayer.R
\name{gdal_cliprasterbymasklayer}
\alias{gdal_cliprasterbymasklayer}
\title{QGIS algorithm Clip raster by mask layer}
\usage{
gdal_cliprasterbymasklayer(
  INPUT = qgisprocess::qgis_default_value(),
  MASK = qgisprocess::qgis_default_value(),
  SOURCE_CRS = qgisprocess::qgis_default_value(),
  TARGET_CRS = qgisprocess::qgis_default_value(),
  TARGET_EXTENT = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  ALPHA_BAND = qgisprocess::qgis_default_value(),
  CROP_TO_CUTLINE = qgisprocess::qgis_default_value(),
  KEEP_RESOLUTION = qgisprocess::qgis_default_value(),
  SET_RESOLUTION = qgisprocess::qgis_default_value(),
  X_RESOLUTION = qgisprocess::qgis_default_value(),
  Y_RESOLUTION = qgisprocess::qgis_default_value(),
  MULTITHREADING = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{MASK}{\code{source} - Mask layer. Path to a vector layer.}

\item{SOURCE_CRS}{\code{crs} - Source CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{TARGET_CRS}{\code{crs} - Target CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{TARGET_EXTENT}{\code{extent} - Target extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{NODATA}{\code{number} - Assign a specified nodata value to output bands. A numeric value.}

\item{ALPHA_BAND}{\code{boolean} - Create an output alpha band. 1 for true/yes. 0 for false/no.}

\item{CROP_TO_CUTLINE}{\code{boolean} - Match the extent of the clipped raster to the extent of the mask layer. 1 for true/yes. 0 for false/no.}

\item{KEEP_RESOLUTION}{\code{boolean} - Keep resolution of input raster. 1 for true/yes. 0 for false/no.}

\item{SET_RESOLUTION}{\code{boolean} - Set output file resolution. 1 for true/yes. 0 for false/no.}

\item{X_RESOLUTION}{\code{number} - X Resolution to output bands. A numeric value.}

\item{Y_RESOLUTION}{\code{number} - Y Resolution to output bands. A numeric value.}

\item{MULTITHREADING}{\code{boolean} - Use multithreaded warping implementation. 1 for true/yes. 0 for false/no.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Use Input Layer Data Type", "Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Clipped (mask). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Clip raster by mask layer (gdal:cliprasterbymasklayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Clipped (mask)
}
}
}
