% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_clipvectorbypolygon.R
\name{gdal_clipvectorbypolygon}
\alias{gdal_clipvectorbypolygon}
\title{QGIS algorithm Clip vector by mask layer}
\usage{
gdal_clipvectorbypolygon(
  INPUT = qgisprocess::qgis_default_value(),
  MASK = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{MASK}{\code{source} - Mask layer. Path to a vector layer.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{OUTPUT}{\code{vectorDestination} - Clipped (mask). Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Clip vector by mask layer (gdal:clipvectorbypolygon)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Clipped
}
}
}
