% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_contour.R
\name{gdal_contour}
\alias{gdal_contour}
\title{QGIS algorithm Contour}
\usage{
gdal_contour(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  INTERVAL = qgisprocess::qgis_default_value(),
  FIELD_NAME = qgisprocess::qgis_default_value(),
  CREATE_3D = qgisprocess::qgis_default_value(),
  IGNORE_NODATA = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  OFFSET = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{INTERVAL}{\code{number} - Interval between contour lines. A numeric value.}

\item{FIELD_NAME}{\code{string} - Attribute name (if not set, no elevation attribute is attached). String value.}

\item{CREATE_3D}{\code{boolean} - Produce 3D vector. 1 for true/yes. 0 for false/no.}

\item{IGNORE_NODATA}{\code{boolean} - Treat all raster values as valid. 1 for true/yes. 0 for false/no.}

\item{NODATA}{\code{number} - Input pixel value to treat as "nodata". A numeric value.}

\item{OFFSET}{\code{number} - Offset from zero relative to which to interpret intervals. A numeric value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{vectorDestination} - Contours. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Contour (gdal:contour)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Contours
}
}
}
