% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_convertformat.R
\name{gdal_convertformat}
\alias{gdal_convertformat}
\title{QGIS algorithm Convert format}
\usage{
gdal_convertformat(
  INPUT = qgisprocess::qgis_default_value(),
  CONVERT_ALL_LAYERS = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{CONVERT_ALL_LAYERS}{\code{boolean} - Convert all layers from dataset. 1 for true/yes. 0 for false/no.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{OUTPUT}{\code{vectorDestination} - Converted. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Convert format (gdal:convertformat)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Converted
}
}
}
