% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_dissolve.R
\name{gdal_dissolve}
\alias{gdal_dissolve}
\title{QGIS algorithm Dissolve}
\usage{
gdal_dissolve(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  GEOMETRY = qgisprocess::qgis_default_value(),
  EXPLODE_COLLECTIONS = qgisprocess::qgis_default_value(),
  KEEP_ATTRIBUTES = qgisprocess::qgis_default_value(),
  COUNT_FEATURES = qgisprocess::qgis_default_value(),
  COMPUTE_AREA = qgisprocess::qgis_default_value(),
  COMPUTE_STATISTICS = qgisprocess::qgis_default_value(),
  STATISTICS_ATTRIBUTE = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Dissolve field. The name of an existing field. ; delimited list of existing field names.}

\item{GEOMETRY}{\code{string} - Geometry column name. String value.}

\item{EXPLODE_COLLECTIONS}{\code{boolean} - Produce one feature for each geometry in any kind of geometry collection in the source file. 1 for true/yes. 0 for false/no.}

\item{KEEP_ATTRIBUTES}{\code{boolean} - Keep input attributes. 1 for true/yes. 0 for false/no.}

\item{COUNT_FEATURES}{\code{boolean} - Count dissolved features. 1 for true/yes. 0 for false/no.}

\item{COMPUTE_AREA}{\code{boolean} - Compute area and perimeter of dissolved features. 1 for true/yes. 0 for false/no.}

\item{COMPUTE_STATISTICS}{\code{boolean} - Compute min/max/sum/mean for attribute. 1 for true/yes. 0 for false/no.}

\item{STATISTICS_ATTRIBUTE}{\code{field} - Numeric attribute to calculate statistics on. The name of an existing field. ; delimited list of existing field names.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{OUTPUT}{\code{vectorDestination} - Dissolved. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Dissolve (gdal:dissolve)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Dissolved
}
}
}
