% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_extractprojection.R
\name{gdal_extractprojection}
\alias{gdal_extractprojection}
\title{QGIS algorithm Extract projection}
\usage{
gdal_extractprojection(
  INPUT = qgisprocess::qgis_default_value(),
  PRJ_FILE_CREATE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input file. Path to a raster layer.}

\item{PRJ_FILE_CREATE}{\code{boolean} - Create also .prj file. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Extract projection (gdal:extractprojection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item WORLD_FILE - outputFile - World file
\item PRJ_FILE - outputFile - ESRI Shapefile prj file
}
}
}
