% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_fillnodata.R
\name{gdal_fillnodata}
\alias{gdal_fillnodata}
\title{QGIS algorithm Fill nodata}
\usage{
gdal_fillnodata(
  INPUT = qgisprocess:::qgis_default_value(),
  BAND = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  ITERATIONS = qgisprocess:::qgis_default_value(),
  NO_MASK = qgisprocess:::qgis_default_value(),
  MASK_LAYER = qgisprocess:::qgis_default_value(),
  OPTIONS = qgisprocess:::qgis_default_value(),
  EXTRA = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{DISTANCE}{\code{number} - Maximum distance (in pixels) to search out for values to interpolate. A numeric value.}

\item{ITERATIONS}{\code{number} - Number of smoothing iterations to run after the interpolation. A numeric value.}

\item{NO_MASK}{\code{boolean} - Do not use the default validity mask for the input band. 1 for true/yes. 0 for false/no.}

\item{MASK_LAYER}{\code{raster} - Validity mask. Path to a raster layer.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Filled. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Fill nodata (gdal:fillnodata)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Filled
}
}
}
