% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_gdal2xyz.R
\name{gdal_gdal2xyz}
\alias{gdal_gdal2xyz}
\title{QGIS algorithm gdal2xyz}
\usage{
gdal_gdal2xyz(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  CSV = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{CSV}{\code{boolean} - Output comma-separated values. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{fileDestination} - XYZ ASCII file. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL gdal2xyz (gdal:gdal2xyz)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - XYZ ASCII file
}
}
}
