% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_gridinversedistancenearestneighbor.R
\name{gdal_gridinversedistancenearestneighbor}
\alias{gdal_gridinversedistancenearestneighbor}
\title{QGIS algorithm Grid (IDW with nearest neighbor searching)}
\usage{
gdal_gridinversedistancenearestneighbor(
  INPUT = qgisprocess::qgis_default_value(),
  Z_FIELD = qgisprocess::qgis_default_value(),
  POWER = qgisprocess::qgis_default_value(),
  SMOOTHING = qgisprocess::qgis_default_value(),
  RADIUS = qgisprocess::qgis_default_value(),
  MAX_POINTS = qgisprocess::qgis_default_value(),
  MIN_POINTS = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Point layer. Path to a vector layer.}

\item{Z_FIELD}{\code{field} - Z value from field. The name of an existing field. ; delimited list of existing field names.}

\item{POWER}{\code{number} - Weighting power. A numeric value.}

\item{SMOOTHING}{\code{number} - Smoothing. A numeric value.}

\item{RADIUS}{\code{number} - The radius of the search circle. A numeric value.}

\item{MAX_POINTS}{\code{number} - Maximum number of data points to use. A numeric value.}

\item{MIN_POINTS}{\code{number} - Minimum number of data points to use. A numeric value.}

\item{NODATA}{\code{number} - NODATA marker to fill empty points. A numeric value.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{rasterDestination} - Interpolated (IDW with NN search). Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Grid (IDW with nearest neighbor searching) (gdal:gridinversedistancenearestneighbor)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Interpolated (IDW with NN search)
}
}
}
