% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_merge.R
\name{gdal_merge}
\alias{gdal_merge}
\title{QGIS algorithm Merge}
\usage{
gdal_merge(
  INPUT = qgisprocess:::qgis_default_value(),
  PCT = qgisprocess:::qgis_default_value(),
  SEPARATE = qgisprocess:::qgis_default_value(),
  NODATA_INPUT = qgisprocess:::qgis_default_value(),
  NODATA_OUTPUT = qgisprocess:::qgis_default_value(),
  OPTIONS = qgisprocess:::qgis_default_value(),
  EXTRA = qgisprocess:::qgis_default_value(),
  DATA_TYPE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input layers. .}

\item{PCT}{\code{boolean} - Grab pseudocolor table from first layer. 1 for true/yes. 0 for false/no.}

\item{SEPARATE}{\code{boolean} - Place each input file into a separate band. 1 for true/yes. 0 for false/no.}

\item{NODATA_INPUT}{\code{number} - Input pixel value to treat as "nodata". A numeric value.}

\item{NODATA_OUTPUT}{\code{number} - Assign specified "nodata" value to output. A numeric value.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{rasterDestination} - Merged. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Merge (gdal:merge)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Merged
}
}
}
