% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_offsetcurve.R
\name{gdal_offsetcurve}
\alias{gdal_offsetcurve}
\title{QGIS algorithm Offset curve}
\usage{
gdal_offsetcurve(
  INPUT = qgisprocess:::qgis_default_value(),
  GEOMETRY = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  OPTIONS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{GEOMETRY}{\code{string} - Geometry column name. String value.}

\item{DISTANCE}{\code{distance} - Offset distance (left-sided: positive, right-sided: negative). A numeric value.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{OUTPUT}{\code{vectorDestination} - Offset curve. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Offset curve (gdal:offsetcurve)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Offset curve
}
}
}
