% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_onesidebuffer.R
\name{gdal_onesidebuffer}
\alias{gdal_onesidebuffer}
\title{QGIS algorithm One side buffer}
\usage{
gdal_onesidebuffer(
  INPUT = qgisprocess:::qgis_default_value(),
  GEOMETRY = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  BUFFER_SIDE = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  DISSOLVE = qgisprocess:::qgis_default_value(),
  EXPLODE_COLLECTIONS = qgisprocess:::qgis_default_value(),
  OPTIONS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{GEOMETRY}{\code{string} - Geometry column name. String value.}

\item{DISTANCE}{\code{distance} - Buffer distance. A numeric value.}

\item{BUFFER_SIDE}{\code{enum}  of \verb{("Right", "Left")} - Buffer side. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FIELD}{\code{field} - Dissolve by attribute. The name of an existing field. ; delimited list of existing field names.}

\item{DISSOLVE}{\code{boolean} - Dissolve all results. 1 for true/yes. 0 for false/no.}

\item{EXPLODE_COLLECTIONS}{\code{boolean} - Produce one feature for each geometry in any kind of geometry collection in the source file. 1 for true/yes. 0 for false/no.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{OUTPUT}{\code{vectorDestination} - One-sided buffer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL One side buffer (gdal:onesidebuffer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - One-sided buffer
}
}
}
