% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_overviews.R
\name{gdal_overviews}
\alias{gdal_overviews}
\title{QGIS algorithm Build overviews (pyramids)}
\usage{
gdal_overviews(
  INPUT = qgisprocess:::qgis_default_value(),
  CLEAN = qgisprocess:::qgis_default_value(),
  LEVELS = qgisprocess:::qgis_default_value(),
  RESAMPLING = qgisprocess:::qgis_default_value(),
  FORMAT = qgisprocess:::qgis_default_value(),
  EXTRA = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{CLEAN}{\code{boolean} - Remove all existing overviews. 1 for true/yes. 0 for false/no.}

\item{LEVELS}{\code{string} - Overview levels (e.g. 2 4 8 16). String value.}

\item{RESAMPLING}{\code{enum}  of \verb{("Nearest Neighbour (default)", "Average", "Gaussian", "Cubic Convolution", "B-Spline Convolution", "Lanczos Windowed Sinc", "Average MP", "Average in Mag/Phase Space", "Mode")} - Resampling method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FORMAT}{\code{enum}  of \verb{("Internal (if possible)", "External (GTiff .ovr)", "External (ERDAS Imagine .aux)")} - Overviews format. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Build overviews (pyramids) (gdal:overviews)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Pyramidized
}
}
}
