% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_rastercalculator.R
\name{gdal_rastercalculator}
\alias{gdal_rastercalculator}
\title{QGIS algorithm Raster calculator}
\usage{
gdal_rastercalculator(
  INPUT_A = qgisprocess:::qgis_default_value(),
  BAND_A = qgisprocess:::qgis_default_value(),
  INPUT_B = qgisprocess:::qgis_default_value(),
  BAND_B = qgisprocess:::qgis_default_value(),
  INPUT_C = qgisprocess:::qgis_default_value(),
  BAND_C = qgisprocess:::qgis_default_value(),
  INPUT_D = qgisprocess:::qgis_default_value(),
  BAND_D = qgisprocess:::qgis_default_value(),
  INPUT_E = qgisprocess:::qgis_default_value(),
  BAND_E = qgisprocess:::qgis_default_value(),
  INPUT_F = qgisprocess:::qgis_default_value(),
  BAND_F = qgisprocess:::qgis_default_value(),
  FORMULA = qgisprocess:::qgis_default_value(),
  NO_DATA = qgisprocess:::qgis_default_value(),
  RTYPE = qgisprocess:::qgis_default_value(),
  OPTIONS = qgisprocess:::qgis_default_value(),
  EXTRA = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_A}{\code{raster} - Input layer A. Path to a raster layer.}

\item{BAND_A}{\code{band} - Number of raster band for A. Integer value representing an existing raster band number.}

\item{INPUT_B}{\code{raster} - Input layer B. Path to a raster layer.}

\item{BAND_B}{\code{band} - Number of raster band for B. Integer value representing an existing raster band number.}

\item{INPUT_C}{\code{raster} - Input layer C. Path to a raster layer.}

\item{BAND_C}{\code{band} - Number of raster band for C. Integer value representing an existing raster band number.}

\item{INPUT_D}{\code{raster} - Input layer D. Path to a raster layer.}

\item{BAND_D}{\code{band} - Number of raster band for D. Integer value representing an existing raster band number.}

\item{INPUT_E}{\code{raster} - Input layer E. Path to a raster layer.}

\item{BAND_E}{\code{band} - Number of raster band for E. Integer value representing an existing raster band number.}

\item{INPUT_F}{\code{raster} - Input layer F. Path to a raster layer.}

\item{BAND_F}{\code{band} - Number of raster band for F. Integer value representing an existing raster band number.}

\item{FORMULA}{\code{string} - Calculation in gdalnumeric syntax using +-/* or any numpy array functions (i.e. logical_and()). String value.}

\item{NO_DATA}{\code{number} - Set output nodata value. A numeric value.}

\item{RTYPE}{\code{enum}  of \verb{("Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64")} - Output raster type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Calculated. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Raster calculator (gdal:rastercalculator)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Calculated
}
}
}
