% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_rasterize.R
\name{gdal_rasterize}
\alias{gdal_rasterize}
\title{QGIS algorithm Rasterize (vector to raster)}
\usage{
gdal_rasterize(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  BURN = qgisprocess::qgis_default_value(),
  UNITS = qgisprocess::qgis_default_value(),
  WIDTH = qgisprocess::qgis_default_value(),
  HEIGHT = qgisprocess::qgis_default_value(),
  EXTENT = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  INIT = qgisprocess::qgis_default_value(),
  INVERT = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Field to use for a burn-in value. The name of an existing field. ; delimited list of existing field names.}

\item{BURN}{\code{number} - A fixed value to burn. A numeric value.}

\item{UNITS}{\code{enum}  of \verb{("Pixels", "Georeferenced units")} - Output raster size units. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{WIDTH}{\code{number} - Width/Horizontal resolution. A numeric value.}

\item{HEIGHT}{\code{number} - Height/Vertical resolution. A numeric value.}

\item{EXTENT}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{NODATA}{\code{number} - Assign a specified nodata value to output bands. A numeric value.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{INIT}{\code{number} - Pre-initialize the output image with value. A numeric value.}

\item{INVERT}{\code{boolean} - Invert rasterization. 1 for true/yes. 0 for false/no.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{OUTPUT}{\code{rasterDestination} - Rasterized. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Rasterize (vector to raster) (gdal:rasterize)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Rasterized
}
}
}
