% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_rgbtopct.R
\name{gdal_rgbtopct}
\alias{gdal_rgbtopct}
\title{QGIS algorithm RGB to PCT}
\usage{
gdal_rgbtopct(
  INPUT = qgisprocess:::qgis_default_value(),
  NCOLORS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{NCOLORS}{\code{number} - Number of colors. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - RGB to PCT. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL RGB to PCT (gdal:rgbtopct)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - RGB to PCT
}
}
}
