% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_tileindex.R
\name{gdal_tileindex}
\alias{gdal_tileindex}
\title{QGIS algorithm Tile index}
\usage{
gdal_tileindex(
  LAYERS = qgisprocess::qgis_default_value(),
  PATH_FIELD_NAME = qgisprocess::qgis_default_value(),
  ABSOLUTE_PATH = qgisprocess::qgis_default_value(),
  PROJ_DIFFERENCE = qgisprocess::qgis_default_value(),
  TARGET_CRS = qgisprocess::qgis_default_value(),
  CRS_FIELD_NAME = qgisprocess::qgis_default_value(),
  CRS_FORMAT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{LAYERS}{\code{multilayer} - Input files. .}

\item{PATH_FIELD_NAME}{\code{string} - Field name to hold the file path to the indexed rasters. String value.}

\item{ABSOLUTE_PATH}{\code{boolean} - Store absolute path to the indexed rasters. 1 for true/yes. 0 for false/no.}

\item{PROJ_DIFFERENCE}{\code{boolean} - Skip files with different projection reference. 1 for true/yes. 0 for false/no.}

\item{TARGET_CRS}{\code{crs} - Transform geometries to the given CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{CRS_FIELD_NAME}{\code{string} - The name of the field to store the SRS of each tile. String value.}

\item{CRS_FORMAT}{\code{enum}  of \verb{("Auto", "Well-known text (WKT)", "EPSG", "Proj.4")} - The format in which the CRS of each tile must be written. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{vectorDestination} - Tile index. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GDAL Tile index (gdal:tileindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Tile inde
}
}
}
