% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_translate.R
\name{gdal_translate}
\alias{gdal_translate}
\title{QGIS algorithm Translate (convert format)}
\usage{
gdal_translate(
  INPUT = qgisprocess::qgis_default_value(),
  TARGET_CRS = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  COPY_SUBDATASETS = qgisprocess::qgis_default_value(),
  OPTIONS = qgisprocess::qgis_default_value(),
  EXTRA = qgisprocess::qgis_default_value(),
  DATA_TYPE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{TARGET_CRS}{\code{crs} - Override the projection for the output file. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{NODATA}{\code{number} - Assign a specified nodata value to output bands. A numeric value.}

\item{COPY_SUBDATASETS}{\code{boolean} - Copy all subdatasets of this file to individual output files. 1 for true/yes. 0 for false/no.}

\item{OPTIONS}{\code{string} - Additional creation options. String value.}

\item{EXTRA}{\code{string} - Additional command-line parameters. String value.}

\item{DATA_TYPE}{\code{enum}  of \verb{("Use Input Layer Data Type", "Byte", "Int16", "UInt16", "UInt32", "Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32", "CFloat64")} - Output data type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{rasterDestination} - Converted. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by GDAL Translate (convert format) (gdal:translate)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Converted
}
}
}
