% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_circle.R
\name{grass7_r_circle}
\alias{grass7_r_circle}
\title{QGIS algorithm r.circle}
\usage{
grass7_r_circle(
  coordinates = qgisprocess::qgis_default_value(),
  min = qgisprocess::qgis_default_value(),
  max = qgisprocess::qgis_default_value(),
  multiplier = qgisprocess::qgis_default_value(),
  .b = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{coordinates}{\code{point} - The coordinate of the center (east,north). Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{min}{\code{number} - Minimum radius for ring/circle map (in meters). A numeric value.}

\item{max}{\code{number} - Maximum radius for ring/circle map (in meters). A numeric value.}

\item{multiplier}{\code{number} - Data value multiplier. A numeric value.}

\item{.b}{\code{boolean} - Generate binary raster map. 1 for true/yes. 0 for false/no. Original algorithm parameter name: -b.}

\item{output}{\code{rasterDestination} - Circles. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.circle (grass7:r.circle)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Circles
}
}
}
