% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_r_series_accumulate.R
\name{grass7_r_series_accumulate}
\alias{grass7_r_series_accumulate}
\title{QGIS algorithm r.series.accumulate}
\usage{
grass7_r_series_accumulate(
  input = qgisprocess::qgis_default_value(),
  lower = qgisprocess::qgis_default_value(),
  upper = qgisprocess::qgis_default_value(),
  method = qgisprocess::qgis_default_value(),
  scale = qgisprocess::qgis_default_value(),
  shift = qgisprocess::qgis_default_value(),
  range = qgisprocess::qgis_default_value(),
  limits = qgisprocess::qgis_default_value(),
  .n = qgisprocess::qgis_default_value(),
  .f = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_REGION_CELLSIZE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_OPT = qgisprocess::qgis_default_value(),
  GRASS_RASTER_FORMAT_META = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{multilayer} - Input raster layer(s). .}

\item{lower}{\code{raster} - Raster map specifying the lower accumulation limit. Path to a raster layer.}

\item{upper}{\code{raster} - Raster map specifying the upper accumulation limit. Path to a raster layer.}

\item{method}{\code{enum}  of \verb{("gdd", "bedd", "huglin", "mean")} - This method will be applied to compute the accumulative values from the input maps. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{scale}{\code{number} - Scale factor for input. A numeric value.}

\item{shift}{\code{number} - Shift factor for input. A numeric value.}

\item{range}{\code{range} - Ignore values outside this range (min,max). Two comma separated numeric values, e.g. '1,10'.}

\item{limits}{\code{range} - Lower and upper accumulation limits (lower,upper). Two comma separated numeric values, e.g. '1,10'.}

\item{.n}{\code{boolean} - Propagate NULLs. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.f}{\code{boolean} - Create a FCELL map (floating point single precision) as output. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{rasterDestination} - Accumulated. Path for new raster layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_REGION_CELLSIZE_PARAMETER}{\code{number} - GRASS GIS 7 region cellsize (leave 0 for default). A numeric value.}

\item{GRASS_RASTER_FORMAT_OPT}{\code{string} - Output Rasters format options (createopt). String value.}

\item{GRASS_RASTER_FORMAT_META}{\code{string} - Output Rasters format metadata options (metaopt). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS r.series.accumulate (grass7:r.series.accumulate)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputRaster - Accumulated
}
}
}
