% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7_v_buffer.R
\name{grass7_v_buffer}
\alias{grass7_v_buffer}
\title{QGIS algorithm v.buffer}
\usage{
grass7_v_buffer(
  input = qgisprocess::qgis_default_value(),
  cats = qgisprocess::qgis_default_value(),
  where = qgisprocess::qgis_default_value(),
  type = qgisprocess::qgis_default_value(),
  distance = qgisprocess::qgis_default_value(),
  minordistance = qgisprocess::qgis_default_value(),
  angle = qgisprocess::qgis_default_value(),
  layer = qgisprocess::qgis_default_value(),
  column = qgisprocess::qgis_default_value(),
  scale = qgisprocess::qgis_default_value(),
  tolerance = qgisprocess::qgis_default_value(),
  .s = qgisprocess::qgis_default_value(),
  .c = qgisprocess::qgis_default_value(),
  .t = qgisprocess::qgis_default_value(),
  output = qgisprocess::qgis_default_value(),
  GRASS_REGION_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_SNAP_TOLERANCE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_MIN_AREA_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_OUTPUT_TYPE_PARAMETER = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_DSCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_LCO = qgisprocess::qgis_default_value(),
  GRASS_VECTOR_EXPORT_NOCAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{input}{\code{source} - Input vector layer. Path to a vector layer.}

\item{cats}{\code{string} - Category values. String value.}

\item{where}{\code{string} - WHERE conditions of SQL statement without 'where' keyword. String value.}

\item{type}{\code{enum}  of \verb{("point", "line", "boundary", "centroid", "area")} - Input feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{distance}{\code{number} - Buffer distance in map units. A numeric value.}

\item{minordistance}{\code{number} - Buffer distance along minor axis in map units. A numeric value.}

\item{angle}{\code{number} - Angle of major axis in degrees. A numeric value.}

\item{layer}{\code{string} - Layer number or name ('-1' for all layers). String value.}

\item{column}{\code{field} - Name of column to use for buffer distances. The name of an existing field. ; delimited list of existing field names.}

\item{scale}{\code{number} - Scaling factor for attribute column values. A numeric value.}

\item{tolerance}{\code{number} - Maximum distance between theoretical arc and polygon segments as multiple of buffer. A numeric value.}

\item{.s}{\code{boolean} - Make outside corners straight. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.c}{\code{boolean} - Do not make caps at the ends of polylines. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{.t}{\code{boolean} - Transfer categories and attributes. 1 for true/yes. 0 for false/no. Original algorithm parameter name: {name}.}

\item{output}{\code{vectorDestination} - Buffer. Path for new vector layer.}

\item{GRASS_REGION_PARAMETER}{\code{extent} - GRASS GIS 7 region extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{GRASS_SNAP_TOLERANCE_PARAMETER}{\code{number} - v.in.ogr snap tolerance (-1 = no snap). A numeric value.}

\item{GRASS_MIN_AREA_PARAMETER}{\code{number} - v.in.ogr min area. A numeric value.}

\item{GRASS_OUTPUT_TYPE_PARAMETER}{\code{enum}  of \verb{("auto", "point", "line", "area")} - v.out.ogr output type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{GRASS_VECTOR_DSCO}{\code{string} - v.out.ogr output data source options (dsco). String value.}

\item{GRASS_VECTOR_LCO}{\code{string} - v.out.ogr output layer options (lco). String value.}

\item{GRASS_VECTOR_EXPORT_NOCAT}{\code{boolean} - Also export features without category (not labeled). Otherwise only features with category are exported. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by GRASS v.buffer (grass7:v.buffer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item output - outputVector - Buffer
}
}
}
