% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_affinetransform.R
\name{qgis_affinetransform}
\alias{qgis_affinetransform}
\title{QGIS algorithm - Affine transform}
\usage{
qgis_affinetransform(
  INPUT = qgisprocess:::qgis_default_value(),
  DELTA_X = qgisprocess:::qgis_default_value(),
  DELTA_Y = qgisprocess:::qgis_default_value(),
  DELTA_Z = qgisprocess:::qgis_default_value(),
  DELTA_M = qgisprocess:::qgis_default_value(),
  SCALE_X = qgisprocess:::qgis_default_value(),
  SCALE_Y = qgisprocess:::qgis_default_value(),
  SCALE_Z = qgisprocess:::qgis_default_value(),
  SCALE_M = qgisprocess:::qgis_default_value(),
  ROTATION_Z = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DELTA_X}{\code{distance} - Translation (x-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_Y}{\code{distance} - Translation (y-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_Z}{\code{number} - Translation (z-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_M}{\code{number} - Translation (m values). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SCALE_X}{\code{number} - Scale factor (x-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SCALE_Y}{\code{number} - Scale factor (y-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SCALE_Z}{\code{number} - Scale factor (z-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SCALE_M}{\code{number} - Scale factor (m values). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ROTATION_Z}{\code{number} - Rotation around z-axis (degrees counter-clockwise). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Transformed. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Affine transform (native:affinetransform). Applies an affine transformation to geometries. Applies an affine transformation to the geometries from a layer. Affine transformations can include translation, scaling and rotation. The operations are performed in a scale, rotation, translation order.  Z and M values present in the geometry can also be translated and scaled independently.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Transformed
}
}
}
