% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_angletonearest.R
\name{qgis_angletonearest}
\alias{qgis_angletonearest}
\title{QGIS algorithm - Align points to features}
\usage{
qgis_angletonearest(
  INPUT = qgisprocess:::qgis_default_value(),
  REFERENCE_LAYER = qgisprocess:::qgis_default_value(),
  MAX_DISTANCE = qgisprocess:::qgis_default_value(),
  FIELD_NAME = qgisprocess:::qgis_default_value(),
  APPLY_SYMBOLOGY = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{REFERENCE_LAYER}{\code{source} - Reference layer. Path to a vector layer.}

\item{MAX_DISTANCE}{\code{distance} - Maximum distance to consider. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FIELD_NAME}{\code{string} - Angle field name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{APPLY_SYMBOLOGY}{\code{boolean} - Automatically apply symbology. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Aligned layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Align points to features (native:angletonearest). Rotates point features to align them to nearby features. This algorithm calculates the rotation required to align point features with their nearest feature from another reference layer. A new field is added to the output layer which is filled with the angle (in degrees, clockwise) to the nearest reference feature.  Optionally, the output layer's symbology can be set to automatically use the calculated rotation field to rotate marker symbols.  If desired, a maximum distance to use when aligning points can be set, to avoid aligning isolated points to distant features.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Aligned layer
}
}
}
