% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_antimeridiansplit.R
\name{qgis_antimeridiansplit}
\alias{qgis_antimeridiansplit}
\title{QGIS algorithm - Geodesic line split at antimeridian}
\usage{
qgis_antimeridiansplit(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Split. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Geodesic line split at antimeridian (native:antimeridiansplit). Splits lines into multiple geodesic segments when the line crosses the antimeridian (±180 degrees longitude). This algorithm splits a line into multiple geodesic segments, whenever the line crosses the antimeridian (±180 degrees longitude).  Splitting at the antimeridian helps the visual display of the lines in some projections. The returned geometry will always be a multi-part geometry.  Whenever line segments in the input geometry cross the antimeridian, they will be split into two segments, with the latitude of the breakpoint being determined using a geodesic line connecting the points either side of this segment. The current project ellipsoid setting will be used when calculating this breakpoint.  If the input geometry contains M or Z values, these will be linearly interpolated for the new vertices created at the antimeridian.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Split
}
}
}
