% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_arraytranslatedfeatures.R
\name{qgis_arraytranslatedfeatures}
\alias{qgis_arraytranslatedfeatures}
\title{QGIS algorithm - Array of translated features}
\usage{
qgis_arraytranslatedfeatures(
  INPUT = qgisprocess:::qgis_default_value(),
  COUNT = qgisprocess:::qgis_default_value(),
  DELTA_X = qgisprocess:::qgis_default_value(),
  DELTA_Y = qgisprocess:::qgis_default_value(),
  DELTA_Z = qgisprocess:::qgis_default_value(),
  DELTA_M = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{COUNT}{\code{number} - Number of features to create. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_X}{\code{distance} - Step distance (x-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_Y}{\code{distance} - Step distance (y-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_Z}{\code{number} - Step distance (z-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_M}{\code{number} - Step distance (m values). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Translated. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Array of translated features (native:arraytranslatedfeatures). Creates multiple translated copies of features in a layer. This algorithm creates copies of features in a layer, by creating multiple translated versions of each feature. Each copy is incrementally displaced by a preset amount in the x/y/z/m axis.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Translated
}
}
}
