% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_barplot.R
\name{qgis_barplot}
\alias{qgis_barplot}
\title{QGIS algorithm Bar plot}
\usage{
qgis_barplot(
  INPUT = qgisprocess::qgis_default_value(),
  NAME_FIELD = qgisprocess::qgis_default_value(),
  VALUE_FIELD = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{NAME_FIELD}{\code{field} - Category name field. The name of an existing field. ; delimited list of existing field names.}

\item{VALUE_FIELD}{\code{field} - Value field. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{fileDestination} - Bar plot. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Bar plot (qgis:barplot)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputHtml - Bar plot
}
}
}
