% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_bufferbym.R
\name{qgis_bufferbym}
\alias{qgis_bufferbym}
\title{QGIS algorithm Variable width buffer (by M value)}
\usage{
qgis_bufferbym(
  INPUT = qgisprocess::qgis_default_value(),
  SEGMENTS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{SEGMENTS}{\code{number} - Segments. A numeric value.}

\item{OUTPUT}{\code{sink} - Buffered. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Variable width buffer (by M value) (native:bufferbym)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Buffered
}
}
}
