% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_cellstackpercentile.R
\name{qgis_cellstackpercentile}
\alias{qgis_cellstackpercentile}
\title{QGIS algorithm - Cell stack percentile}
\usage{
qgis_cellstackpercentile(
  INPUT = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  PERCENTILE = qgisprocess:::qgis_default_value(),
  IGNORE_NODATA = qgisprocess:::qgis_default_value(),
  REFERENCE_LAYER = qgisprocess:::qgis_default_value(),
  OUTPUT_NODATA_VALUE = qgisprocess:::qgis_default_value(),
  CREATION_OPTIONS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input layers. .}

\item{METHOD}{\code{enum}  of \verb{("Nearest rank", "Inclusive linear interpolation (PERCENTILE.INC)", "Exclusive linear interpolation (PERCENTILE.EXC)")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{PERCENTILE}{\code{number} - Percentile. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{IGNORE_NODATA}{\code{boolean} - Ignore NoData values. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{REFERENCE_LAYER}{\code{raster} - Reference layer. Path to a raster layer.}

\item{OUTPUT_NODATA_VALUE}{\code{number} - Output NoData value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{CREATION_OPTIONS}{\code{string} - Creation options. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{rasterDestination} - Output layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Cell stack percentile (native:cellstackpercentile). Generates a raster containing the cell-wise percentile value of a stack of input rasters. This algorithm generates a raster containing the cell-wise percentile value of a stack of input rasters. The percentile to return is determined by the percentile input value (ranges between 0 and 1). At each cell location, the specified percentile is obtained using the respective value from the stack of all overlaid and sorted cell values of the input rasters.  There are three methods for percentile calculation:ul    liNearest rank/li   liInclusive linear interpolation (PERCENTILE.INC)/li   liExclusive linear interpolation (PERCENTILE.EXC)/li/ul While the output value can stay the same for the nearest rank method (obtains the value that is nearest to the specified percentile), the linear interpolation method return unique values for different percentiles. Both interpolation methods follow their counterpart methods implemented by LibreOffice or Microsoft Excel.   The output raster's extent and resolution is defined by a reference raster. If the input raster layers that do not match the cell size of the reference raster layer will be resampled using nearest neighbor resampling. NoData values in any of the input layers will result in a NoData cell output if the Ignore NoData parameter is not set. The output raster data type will be set to the most complex data type present in the input datasets.
}
\details{
\subsection{Outputs description}{
\itemize{
\item CRS_AUTHID - outputString - CRS authority identifier
\item EXTENT - outputString - Extent
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item OUTPUT - outputRaster - Output layer
\item TOTAL_PIXEL_COUNT - outputNumber - Total pixel count
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
}
}
}
