% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_climbalongline.R
\name{qgis_climbalongline}
\alias{qgis_climbalongline}
\title{QGIS algorithm Climb Along Line}
\usage{
qgis_climbalongline(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Line layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Climb layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Climb Along Line (qgis:climbalongline)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Climb layer
\item TOTALCLIMB - outputNumber - Total climb
\item TOTALDESCENT - outputNumber - Total descent
\item MINELEVATION - outputNumber - Minimum elevation
\item MAXELEVATION - outputNumber - Maximum elevation
}
}
}
