% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_combinestyles.R
\name{qgis_combinestyles}
\alias{qgis_combinestyles}
\title{QGIS algorithm Combine style databases}
\usage{
qgis_combinestyles(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  OBJECTS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input databases. .}

\item{OUTPUT}{\code{fileDestination} - Output style database. Path for new file.}

\item{OBJECTS}{\code{enum}  of \verb{("Symbols", "Color ramps", "Text formats", "Label settings")} - Objects to combine. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Combine style databases (native:combinestyles)
}
\details{
\subsection{Outputs description}{
\itemize{
\item COLORRAMPS - outputNumber - Color ramp count
\item LABELSETTINGS - outputNumber - Label settings count
\item OUTPUT - outputFile - Output style database
\item SYMBOLS - outputNumber - Symbol count
\item TEXTFORMATS - outputNumber - Text format count
}
}
}
