% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_concavehull.R
\name{qgis_concavehull}
\alias{qgis_concavehull}
\title{QGIS algorithm Concave hull (alpha shapes)}
\usage{
qgis_concavehull(
  INPUT = qgisprocess::qgis_default_value(),
  ALPHA = qgisprocess::qgis_default_value(),
  HOLES = qgisprocess::qgis_default_value(),
  NO_MULTIGEOMETRY = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input point layer. Path to a vector layer.}

\item{ALPHA}{\code{number} - Threshold (0-1, where 1 is equivalent with Convex Hull). A numeric value.}

\item{HOLES}{\code{boolean} - Allow holes. 1 for true/yes. 0 for false/no.}

\item{NO_MULTIGEOMETRY}{\code{boolean} - Split multipart geometry into singleparts geometries. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Concave hull. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Concave hull (alpha shapes) (qgis:concavehull)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Concave hull
}
}
}
