% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_convertgpsdata.R
\name{qgis_convertgpsdata}
\alias{qgis_convertgpsdata}
\title{QGIS algorithm - Convert GPS data}
\usage{
qgis_convertgpsdata(
  INPUT = qgisprocess:::qgis_default_value(),
  FORMAT = qgisprocess:::qgis_default_value(),
  FEATURE_TYPE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{file} - Input file. Path to a file.}

\item{FORMAT}{\code{string} - Format. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FEATURE_TYPE}{\code{enum}  of \verb{("Waypoints", "Routes", "Tracks")} - Feature type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{fileDestination} - Output. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convert GPS data (native:convertgpsdata). This algorithm uses the GPSBabel tool to convert a GPS data file from a range of formats to the GPX standard format.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - Output
\item OUTPUT_LAYER - outputVector - Output layer
}
}
}
