% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_converttocurves.R
\name{qgis_converttocurves}
\alias{qgis_converttocurves}
\title{QGIS algorithm Convert to curved geometries}
\usage{
qgis_converttocurves(
  INPUT = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  ANGLE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DISTANCE}{\code{number} - Maximum distance tolerance. A numeric value.}

\item{ANGLE}{\code{number} - Maximum angle tolerance. A numeric value.}

\item{OUTPUT}{\code{sink} - Curves. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Convert to curved geometries (native:converttocurves)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Curves
}
}
}
