% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_countpointsinpolygon.R
\name{qgis_countpointsinpolygon}
\alias{qgis_countpointsinpolygon}
\title{QGIS algorithm Count points in polygon}
\usage{
qgis_countpointsinpolygon(
  POLYGONS = qgisprocess:::qgis_default_value(),
  POINTS = qgisprocess:::qgis_default_value(),
  WEIGHT = qgisprocess:::qgis_default_value(),
  CLASSFIELD = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{POLYGONS}{\code{source} - Polygons. Path to a vector layer.}

\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{WEIGHT}{\code{field} - Weight field. The name of an existing field. ; delimited list of existing field names.}

\item{CLASSFIELD}{\code{field} - Class field. The name of an existing field. ; delimited list of existing field names.}

\item{FIELD}{\code{string} - Count field name. String value.}

\item{OUTPUT}{\code{sink} - Count. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Count points in polygon (native:countpointsinpolygon)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Count
}
}
}
