% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_densifygeometriesgivenaninterval.R
\name{qgis_densifygeometriesgivenaninterval}
\alias{qgis_densifygeometriesgivenaninterval}
\title{QGIS algorithm - Densify by interval}
\usage{
qgis_densifygeometriesgivenaninterval(
  INPUT = qgisprocess:::qgis_default_value(),
  INTERVAL = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{INTERVAL}{\code{distance} - Interval between vertices to add. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Densified. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Densify by interval (native:densifygeometriesgivenaninterval). Creates a densified version of geometries by setting a maximum distance for segments. This algorithm takes a polygon or line layer and generates a new one in which the geometries have a larger number of vertices than the original one.  Geometries are densified by adding additional vertices on edges that have a maximum distance of the interval parameter in map units.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Densified
}
}
}
