% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_detectvectorchanges.R
\name{qgis_detectvectorchanges}
\alias{qgis_detectvectorchanges}
\title{QGIS algorithm Detect dataset changes}
\usage{
qgis_detectvectorchanges(
  ORIGINAL = qgisprocess::qgis_default_value(),
  REVISED = qgisprocess::qgis_default_value(),
  COMPARE_ATTRIBUTES = qgisprocess::qgis_default_value(),
  MATCH_TYPE = qgisprocess::qgis_default_value(),
  UNCHANGED = qgisprocess::qgis_default_value(),
  ADDED = qgisprocess::qgis_default_value(),
  DELETED = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ORIGINAL}{\code{source} - Original layer. Path to a vector layer.}

\item{REVISED}{\code{source} - Revised layer. Path to a vector layer.}

\item{COMPARE_ATTRIBUTES}{\code{field} - Attributes to consider for match (or none to compare geometry only). The name of an existing field. ; delimited list of existing field names.}

\item{MATCH_TYPE}{\code{enum}  of \verb{("Exact Match", "Tolerant Match (Topological Equality)")} - Geometry comparison behavior. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{UNCHANGED}{\code{sink} - Unchanged features. Path for new vector layer.}

\item{ADDED}{\code{sink} - Added features. Path for new vector layer.}

\item{DELETED}{\code{sink} - Deleted features. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Detect dataset changes (native:detectvectorchanges)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ADDED - outputVector - Added features
\item ADDED_COUNT - outputNumber - Count of features added in revised layer
\item DELETED - outputVector - Deleted features
\item DELETED_COUNT - outputNumber - Count of features deleted from original layer
\item UNCHANGED - outputVector - Unchanged features
\item UNCHANGED_COUNT - outputNumber - Count of unchanged features
}
}
}
