% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_dissolve.R
\name{qgis_dissolve}
\alias{qgis_dissolve}
\title{QGIS algorithm Dissolve}
\usage{
qgis_dissolve(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Dissolve field(s). The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Dissolved. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Dissolve (native:dissolve)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Dissolved
}
}
}
