% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_dropmzvalues.R
\name{qgis_dropmzvalues}
\alias{qgis_dropmzvalues}
\title{QGIS algorithm - Drop M/Z values}
\usage{
qgis_dropmzvalues(
  INPUT = qgisprocess:::qgis_default_value(),
  DROP_M_VALUES = qgisprocess:::qgis_default_value(),
  DROP_Z_VALUES = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DROP_M_VALUES}{\code{boolean} - Drop M Values. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DROP_Z_VALUES}{\code{boolean} - Drop Z Values. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Z\code{/}M Dropped. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Drop M/Z values (native:dropmzvalues). Removes any measure (M) or Z values from input geometries. This algorithm can remove any measure (M) or Z values from input geometries.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Z\code{/}M Dropped
}
}
}
