% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_executesql.R
\name{qgis_executesql}
\alias{qgis_executesql}
\title{QGIS algorithm - Execute SQL}
\usage{
qgis_executesql(
  INPUT_DATASOURCES = qgisprocess:::qgis_default_value(),
  INPUT_QUERY = qgisprocess:::qgis_default_value(),
  INPUT_UID_FIELD = qgisprocess:::qgis_default_value(),
  INPUT_GEOMETRY_FIELD = qgisprocess:::qgis_default_value(),
  INPUT_GEOMETRY_TYPE = qgisprocess:::qgis_default_value(),
  INPUT_GEOMETRY_CRS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_DATASOURCES}{\code{multilayer} - Additional input datasources (called input1, .., inputN in the query). .}

\item{INPUT_QUERY}{\code{execute_sql} - SQL query. .}

\item{INPUT_UID_FIELD}{\code{string} - Unique identifier field. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INPUT_GEOMETRY_FIELD}{\code{string} - Geometry field. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INPUT_GEOMETRY_TYPE}{\code{enum}  of \verb{("Autodetect", "No geometry", "Point", "LineString", "Polygon", "MultiPoint", "MultiLineString", "MultiPolygon")} - Geometry type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{INPUT_GEOMETRY_CRS}{\code{crs} - CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT}{\code{sink} - SQL Output. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Execute SQL (qgis:executesql). Runs a simple or complex query with SQL syntax on the source layer. Input datasources are identified with input1, input2, ..., inputN and a simple query will look like: SELECT * FROM input1 The result of the query will be added as a new layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - SQL Output
}
}
}
