% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_explodehstorefield.R
\name{qgis_explodehstorefield}
\alias{qgis_explodehstorefield}
\title{QGIS algorithm Explode HStore Field}
\usage{
qgis_explodehstorefield(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  EXPECTED_FIELDS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - HStore field. The name of an existing field. ; delimited list of existing field names.}

\item{EXPECTED_FIELDS}{\code{string} - Expected list of fields separated by a comma. String value.}

\item{OUTPUT}{\code{sink} - Exploded. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Explode HStore Field (native:explodehstorefield)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Exploded
}
}
}
