% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_exportaddgeometrycolumns.R
\name{qgis_exportaddgeometrycolumns}
\alias{qgis_exportaddgeometrycolumns}
\title{QGIS algorithm - Add geometry attributes}
\usage{
qgis_exportaddgeometrycolumns(
  INPUT = qgisprocess:::qgis_default_value(),
  CALC_METHOD = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{CALC_METHOD}{\code{enum}  of \verb{("Layer CRS", "Project CRS", "Ellipsoidal")} - Calculate using. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Added geom info. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Add geometry attributes (qgis:exportaddgeometrycolumns). This algorithm computes geometric properties of the features in a vector layer. It generates a new vector layer with the same content as the input one, but with additional attributes in its attributes table, containing geometric measurements. Depending on the geometry type of the vector layer, the attributes added to the table will be different.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Added geom info
}
}
}
