% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_exportaddgeometrycolumns.R
\name{qgis_exportaddgeometrycolumns}
\alias{qgis_exportaddgeometrycolumns}
\title{QGIS algorithm Add geometry attributes}
\usage{
qgis_exportaddgeometrycolumns(
  INPUT = qgisprocess::qgis_default_value(),
  CALC_METHOD = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{CALC_METHOD}{\code{enum}  of \verb{("Layer CRS", "Project CRS", "Ellipsoidal")} - Calculate using. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Added geom info. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Add geometry attributes (qgis:exportaddgeometrycolumns)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Added geom info
}
}
}
