% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_extenttolayer.R
\name{qgis_extenttolayer}
\alias{qgis_extenttolayer}
\title{QGIS algorithm Create layer from extent}
\usage{
qgis_extenttolayer(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{extent} - Extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{OUTPUT}{\code{sink} - Extent. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create layer from extent (native:extenttolayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Extent
}
}
}
