% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_extractvertices.R
\name{qgis_extractvertices}
\alias{qgis_extractvertices}
\title{QGIS algorithm Extract vertices}
\usage{
qgis_extractvertices(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Vertices. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Extract vertices (native:extractvertices)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Vertices
}
}
}
