% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_fieldcalculator.R
\name{qgis_fieldcalculator}
\alias{qgis_fieldcalculator}
\title{QGIS algorithm - Field calculator}
\usage{
qgis_fieldcalculator(
  INPUT = qgisprocess:::qgis_default_value(),
  FIELD_NAME = qgisprocess:::qgis_default_value(),
  FIELD_TYPE = qgisprocess:::qgis_default_value(),
  FIELD_LENGTH = qgisprocess:::qgis_default_value(),
  FIELD_PRECISION = qgisprocess:::qgis_default_value(),
  FORMULA = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD_NAME}{\code{string} - Field name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FIELD_TYPE}{\code{enum}  of \verb{("Decimal (double)", "Integer (32 bit)", "Text (string)", "Date", "Time", "Date & Time", "Boolean", "Binary Object (BLOB)", "String List", "Integer List", "Decimal (double) List")} - Result field type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FIELD_LENGTH}{\code{number} - Result field length. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FIELD_PRECISION}{\code{number} - Result field precision. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{FORMULA}{\code{expression} - Formula. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{OUTPUT}{\code{sink} - Calculated. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Field calculator (native:fieldcalculator). Computes a new vector layer with the same features of the input layer, but either overwriting an existing attribute or adding an additional attribute. This algorithm computes a new vector layer with the same features of the input layer, but either overwriting an existing attribute or adding an additional attribute. The values of this field are computed from each feature using an expression, based on the properties and attributes of the feature. Note that if "Field name" is an existing field in the layer then all the rest of the field settings are ignored.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Calculated
}
}
}
