% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_fieldcalculator.R
\name{qgis_fieldcalculator}
\alias{qgis_fieldcalculator}
\title{QGIS algorithm Field calculator}
\usage{
qgis_fieldcalculator(
  INPUT = qgisprocess::qgis_default_value(),
  FIELD_NAME = qgisprocess::qgis_default_value(),
  FIELD_TYPE = qgisprocess::qgis_default_value(),
  FIELD_LENGTH = qgisprocess::qgis_default_value(),
  FIELD_PRECISION = qgisprocess::qgis_default_value(),
  FORMULA = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD_NAME}{\code{string} - Field name. String value.}

\item{FIELD_TYPE}{\code{enum}  of \verb{("Float", "Integer", "String", "Date")} - Result field type. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{FIELD_LENGTH}{\code{number} - Result field length. A numeric value.}

\item{FIELD_PRECISION}{\code{number} - Result field precision. A numeric value.}

\item{FORMULA}{\code{expression} - Formula. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{OUTPUT}{\code{sink} - Calculated. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Field calculator (native:fieldcalculator)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Calculated
}
}
}
