% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_filedownloader.R
\name{qgis_filedownloader}
\alias{qgis_filedownloader}
\title{QGIS algorithm Download file}
\usage{
qgis_filedownloader(
  URL = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  DATA = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{URL}{\code{string} - URL. String value.}

\item{METHOD}{\code{enum}  of \verb{("GET", "POST")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DATA}{\code{string} - Data. String value.}

\item{OUTPUT}{\code{fileDestination} - File destination. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Download file (native:filedownloader)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - File destination
}
}
}
