% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_filterlayersbytype.R
\name{qgis_filterlayersbytype}
\alias{qgis_filterlayersbytype}
\title{QGIS algorithm - Filter layers by type}
\usage{
qgis_filterlayersbytype(
  INPUT = qgisprocess:::qgis_default_value(),
  VECTOR = qgisprocess:::qgis_default_value(),
  RASTER = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{layer} - Input layer. Path to a vector, raster or mesh layer.}

\item{VECTOR}{\code{vectorDestination} - Vector features. Path for new vector layer.}

\item{RASTER}{\code{rasterDestination} - Raster layer. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Filter layers by type (native:filterlayersbytype). Filters layers by type This algorithm filters layer by their type. Incoming layers will be directed to different outputs based on whether they are a vector or raster layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item RASTER - outputRaster - Raster layer
\item VECTOR - outputVector - Vector features
}
}
}
