% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_flattenrelationships.R
\name{qgis_flattenrelationships}
\alias{qgis_flattenrelationships}
\title{QGIS algorithm - Flatten relationship}
\usage{
qgis_flattenrelationships(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{vector} - Input layer. Path to a vector layer.}

\item{OUTPUT}{\code{sink} - Flattened layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Flatten relationship (native:flattenrelationships). Flatten a relationship for a vector layer. This algorithm flattens a relationship for a vector layer, exporting a single layer containing one master feature per related feature. This master feature contains all the attributes for the related features.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Flattened layer
}
}
}
