% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_fuzzifyrastersmallmembership.R
\name{qgis_fuzzifyrastersmallmembership}
\alias{qgis_fuzzifyrastersmallmembership}
\title{QGIS algorithm Fuzzify raster (small membership)}
\usage{
qgis_fuzzifyrastersmallmembership(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  FUZZYMIDPOINT = qgisprocess::qgis_default_value(),
  FUZZYSPREAD = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Input Raster. Path to a raster layer.}

\item{BAND}{\code{band} - Band Number. Integer value representing an existing raster band number.}

\item{FUZZYMIDPOINT}{\code{number} - Function midpoint. A numeric value.}

\item{FUZZYSPREAD}{\code{number} - Function spread. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Fuzzified raster. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Fuzzify raster (small membership) (native:fuzzifyrastersmallmembership)
}
\details{
\subsection{Outputs description}{
\itemize{
\item EXTENT - outputString - Extent
\item CRS_AUTHID - outputString - CRS authority identifier
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item TOTAL_PIXEL_COUNT - outputNumber - Total pixel count
\item OUTPUT - outputRaster - Fuzzified raster
}
}
}
