% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_importphotos.R
\name{qgis_importphotos}
\alias{qgis_importphotos}
\title{QGIS algorithm Import geotagged photos}
\usage{
qgis_importphotos(
  FOLDER = qgisprocess::qgis_default_value(),
  RECURSIVE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  INVALID = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{FOLDER}{\code{file} - Input folder. Path to a file.}

\item{RECURSIVE}{\code{boolean} - Scan recursively. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Photos. Path for new vector layer.}

\item{INVALID}{\code{sink} - Invalid photos table. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Import geotagged photos (native:importphotos)
}
\details{
\subsection{Outputs description}{
\itemize{
\item INVALID - outputVector - Invalid photos table
\item OUTPUT - outputVector - Photos
}
}
}
