% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_joinbylocationsummary.R
\name{qgis_joinbylocationsummary}
\alias{qgis_joinbylocationsummary}
\title{QGIS algorithm - Join attributes by location (summary)}
\usage{
qgis_joinbylocationsummary(
  INPUT = qgisprocess:::qgis_default_value(),
  PREDICATE = qgisprocess:::qgis_default_value(),
  JOIN = qgisprocess:::qgis_default_value(),
  JOIN_FIELDS = qgisprocess:::qgis_default_value(),
  SUMMARIES = qgisprocess:::qgis_default_value(),
  DISCARD_NONMATCHING = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Join to features in. Path to a vector layer.}

\item{PREDICATE}{\code{enum}  of \verb{("intersect", "contain", "equal", "touch", "overlap", "are within", "cross")} - Where the features. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{JOIN}{\code{source} - By comparing to. Path to a vector layer.}

\item{JOIN_FIELDS}{\code{field} - Fields to summarise (leave empty to use all fields). The name of an existing field. ; delimited list of existing field names.}

\item{SUMMARIES}{\code{enum}  of \verb{("count", "unique", "min", "max", "range", "sum", "mean", "median", "stddev", "minority", "majority", "q1", "q3", "iqr", "empty", "filled", "min_length", "max_length", "mean_length")} - Summaries to calculate (leave empty to use all available). Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DISCARD_NONMATCHING}{\code{boolean} - Discard records which could not be joined. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Joined layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Join attributes by location (summary) (native:joinbylocationsummary). Calculates summaries of attributes from one vector layer to another by location. This algorithm takes an input vector layer and creates a new vector layer that is an extended version of the input one, with additional attributes in its attribute table.  The additional attributes and their values are taken from a second vector layer. A spatial criteria is applied to select the values from the second layer that are added to each feature from the first layer in the resulting one.  The algorithm calculates a statistical summary for the values from matching features in the second layer( e.g. maximum value, mean value, etc ).
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Joined layer
}
}
}
