% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_joinbynearest.R
\name{qgis_joinbynearest}
\alias{qgis_joinbynearest}
\title{QGIS algorithm Join attributes by nearest}
\usage{
qgis_joinbynearest(
  INPUT = qgisprocess::qgis_default_value(),
  INPUT_2 = qgisprocess::qgis_default_value(),
  FIELDS_TO_COPY = qgisprocess::qgis_default_value(),
  DISCARD_NONMATCHING = qgisprocess::qgis_default_value(),
  PREFIX = qgisprocess::qgis_default_value(),
  NEIGHBORS = qgisprocess::qgis_default_value(),
  MAX_DISTANCE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  NON_MATCHING = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{INPUT_2}{\code{source} - Input layer 2. Path to a vector layer.}

\item{FIELDS_TO_COPY}{\code{field} - Layer 2 fields to copy (leave empty to copy all fields). The name of an existing field. ; delimited list of existing field names.}

\item{DISCARD_NONMATCHING}{\code{boolean} - Discard records which could not be joined. 1 for true/yes. 0 for false/no.}

\item{PREFIX}{\code{string} - Joined field prefix. String value.}

\item{NEIGHBORS}{\code{number} - Maximum nearest neighbors. A numeric value.}

\item{MAX_DISTANCE}{\code{distance} - Maximum distance. A numeric value.}

\item{OUTPUT}{\code{sink} - Joined layer. Path for new vector layer.}

\item{NON_MATCHING}{\code{sink} - Unjoinable features from first layer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Join attributes by nearest (native:joinbynearest)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Joined layer
\item NON_MATCHING - outputVector - Unjoinable features from first layer
\item JOINED_COUNT - outputNumber - Number of joined features from input table
\item UNJOINABLE_COUNT - outputNumber - Number of unjoinable features from input table
}
}
}
