% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_meshexporttimeseries.R
\name{qgis_meshexporttimeseries}
\alias{qgis_meshexporttimeseries}
\title{QGIS algorithm - Export time series values from points of a mesh dataset}
\usage{
qgis_meshexporttimeseries(
  INPUT = qgisprocess:::qgis_default_value(),
  DATASET_GROUPS = qgisprocess:::qgis_default_value(),
  STARTING_TIME = qgisprocess:::qgis_default_value(),
  FINISHING_TIME = qgisprocess:::qgis_default_value(),
  TIME_STEP = qgisprocess:::qgis_default_value(),
  INPUT_POINTS = qgisprocess:::qgis_default_value(),
  COORDINATES_DIGITS = qgisprocess:::qgis_default_value(),
  DATASET_DIGITS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{mesh} - Input mesh layer. Path to a mesh layer.}

\item{DATASET_GROUPS}{\code{meshdatasetgroups} - Dataset groups. .}

\item{STARTING_TIME}{\code{meshdatasettime} - Starting time. .}

\item{FINISHING_TIME}{\code{meshdatasettime} - Finishing time. .}

\item{TIME_STEP}{\code{number} - Time step (hours). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{INPUT_POINTS}{\code{source} - Points for data export. Path to a vector layer.}

\item{COORDINATES_DIGITS}{\code{number} - Digits count for coordinates. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DATASET_DIGITS}{\code{number} - Digits count for dataset value. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{fileDestination} - Exported data CSV file. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Export time series values from points of a mesh dataset (native:meshexporttimeseries). Extracts a mesh dataset's time series values from points contained in a vector layer. This algorithm extracts mesh's dataset time series values from points contained in a vector layer. If the time step is kept to its default value (0 hours), the time step used is the one of the two first datasets of the first selected dataset group.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - Exported data CSV file
}
}
}
