% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_nearestneighbouranalysis.R
\name{qgis_nearestneighbouranalysis}
\alias{qgis_nearestneighbouranalysis}
\title{QGIS algorithm Nearest neighbour analysis}
\usage{
qgis_nearestneighbouranalysis(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT_HTML_FILE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{OUTPUT_HTML_FILE}{\code{fileDestination} - Nearest neighbour. Path for new file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Nearest neighbour analysis (native:nearestneighbouranalysis)
}
\details{
\subsection{Outputs description}{
\itemize{
\item EXPECTED_MD - outputNumber - Expected mean distance
\item NN_INDEX - outputNumber - Nearest neighbour index
\item OBSERVED_MD - outputNumber - Observed mean distance
\item OUTPUT_HTML_FILE - outputHtml - Nearest neighbour
\item POINT_COUNT - outputNumber - Number of points
\item Z_SCORE - outputNumber - Z-score
}
}
}
