% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_orderbyexpression.R
\name{qgis_orderbyexpression}
\alias{qgis_orderbyexpression}
\title{QGIS algorithm - Order by expression}
\usage{
qgis_orderbyexpression(
  INPUT = qgisprocess:::qgis_default_value(),
  EXPRESSION = qgisprocess:::qgis_default_value(),
  ASCENDING = qgisprocess:::qgis_default_value(),
  NULLS_FIRST = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{EXPRESSION}{\code{expression} - Expression. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{ASCENDING}{\code{boolean} - Sort ascending. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{NULLS_FIRST}{\code{boolean} - Sort nulls first. 1 for true/yes. 0 for false/no. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Ordered. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Order by expression (native:orderbyexpression). This algorithm sorts a vector layer according to an expression. Be careful, it might not work as expected with some providers, the order might not be kept every time.  For help with QGIS expression functions, see the inbuilt help for specific functions which is available in the expression builder.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Ordered
}
}
}
