% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_package.R
\name{qgis_package}
\alias{qgis_package}
\title{QGIS algorithm Package layers}
\usage{
qgis_package(
  LAYERS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  OVERWRITE = qgisprocess::qgis_default_value(),
  SAVE_STYLES = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{LAYERS}{\code{multilayer} - Input layers. .}

\item{OUTPUT}{\code{fileDestination} - Destination GeoPackage. Path for new file.}

\item{OVERWRITE}{\code{boolean} - Overwrite existing GeoPackage. 1 for true/yes. 0 for false/no.}

\item{SAVE_STYLES}{\code{boolean} - Save layer styles into GeoPackage. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Package layers (native:package)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputFile - Destination GeoPackage
\item OUTPUT_LAYERS - outputMultilayer - Layers within new package
}
}
}
