% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_pointsalonglines.R
\name{qgis_pointsalonglines}
\alias{qgis_pointsalonglines}
\title{QGIS algorithm Points along geometry}
\usage{
qgis_pointsalonglines(
  INPUT = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  START_OFFSET = qgisprocess::qgis_default_value(),
  END_OFFSET = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DISTANCE}{\code{distance} - Distance. A numeric value.}

\item{START_OFFSET}{\code{distance} - Start offset. A numeric value.}

\item{END_OFFSET}{\code{distance} - End offset. A numeric value.}

\item{OUTPUT}{\code{sink} - Interpolated points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Points along geometry (native:pointsalonglines)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Interpolated points
}
}
}
