% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_pointstopath.R
\name{qgis_pointstopath}
\alias{qgis_pointstopath}
\title{QGIS algorithm Points to path}
\usage{
qgis_pointstopath(
  INPUT = qgisprocess::qgis_default_value(),
  CLOSE_PATH = qgisprocess::qgis_default_value(),
  ORDER_EXPRESSION = qgisprocess::qgis_default_value(),
  NATURAL_SORT = qgisprocess::qgis_default_value(),
  GROUP_EXPRESSION = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  OUTPUT_TEXT_DIR = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{CLOSE_PATH}{\code{boolean} - Create closed paths. 1 for true/yes. 0 for false/no.}

\item{ORDER_EXPRESSION}{\code{expression} - Order expression. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{NATURAL_SORT}{\code{boolean} - Sort text containing numbers naturally. 1 for true/yes. 0 for false/no.}

\item{GROUP_EXPRESSION}{\code{expression} - Path group expression. A valid QGIS expression string, e.g "road_name" = 'MAIN RD'.}

\item{OUTPUT}{\code{sink} - Paths. Path for new vector layer.}

\item{OUTPUT_TEXT_DIR}{\code{folderDestination} - Directory for text output. Path for an existing or new folder.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Points to path (native:pointstopath)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Paths
\item OUTPUT_TEXT_DIR - outputFolder - Directory for text output
\item NUM_PATHS - outputNumber - Number of paths
}
}
}
