% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_pointtolayer.R
\name{qgis_pointtolayer}
\alias{qgis_pointtolayer}
\title{QGIS algorithm Create layer from point}
\usage{
qgis_pointtolayer(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{point} - Point. Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{OUTPUT}{\code{sink} - Point. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create layer from point (native:pointtolayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Point
}
}
}
