% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_poleofinaccessibility.R
\name{qgis_poleofinaccessibility}
\alias{qgis_poleofinaccessibility}
\title{QGIS algorithm Pole of inaccessibility}
\usage{
qgis_poleofinaccessibility(
  INPUT = qgisprocess::qgis_default_value(),
  TOLERANCE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{TOLERANCE}{\code{distance} - Tolerance. A numeric value.}

\item{OUTPUT}{\code{sink} - Point. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Pole of inaccessibility (native:poleofinaccessibility)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Point
}
}
}
