% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_polygonize.R
\name{qgis_polygonize}
\alias{qgis_polygonize}
\title{QGIS algorithm Polygonize}
\usage{
qgis_polygonize(
  INPUT = qgisprocess::qgis_default_value(),
  KEEP_FIELDS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{KEEP_FIELDS}{\code{boolean} - Keep table structure of line layer. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Polygons. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Polygonize (native:polygonize)
}
\details{
\subsection{Outputs description}{
\itemize{
\item NUM_POLYGONS - outputNumber - Number of polygons
\item OUTPUT - outputVector - Polygons
}
}
}
