% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_projectpointcartesian.R
\name{qgis_projectpointcartesian}
\alias{qgis_projectpointcartesian}
\title{QGIS algorithm - Project points (Cartesian)}
\usage{
qgis_projectpointcartesian(
  INPUT = qgisprocess:::qgis_default_value(),
  BEARING = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{BEARING}{\code{number} - Bearing (degrees from North). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DISTANCE}{\code{distance} - Distance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Projected. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Project points (Cartesian) (native:projectpointcartesian). Creates a point layer with geometries projected by a specified distance and bearing (azimuth). This algorithm projects point geometries by a specified distance and bearing (azimuth), creating a new point layer with the projected points.  The distance is specified in layer units, and the bearing in degrees clockwise from North.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Projected
}
}
}
