% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_randomextract.R
\name{qgis_randomextract}
\alias{qgis_randomextract}
\title{QGIS algorithm Random extract}
\usage{
qgis_randomextract(
  INPUT = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  NUMBER = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("Number of features", "Percentage of features")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NUMBER}{\code{number} - Number/percentage of features. A numeric value.}

\item{OUTPUT}{\code{sink} - Extracted (random). Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Random extract (native:randomextract)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Extracted
}
}
}
