% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rastercalculator.R
\name{qgis_rastercalculator}
\alias{qgis_rastercalculator}
\title{QGIS algorithm Raster calculator}
\usage{
qgis_rastercalculator(
  EXPRESSION = qgisprocess::qgis_default_value(),
  LAYERS = qgisprocess::qgis_default_value(),
  CELLSIZE = qgisprocess::qgis_default_value(),
  EXTENT = qgisprocess::qgis_default_value(),
  CRS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{EXPRESSION}{\code{raster_calc_expression} - Expression. .}

\item{LAYERS}{\code{multilayer} - Reference layer(s) (used for automated extent, cellsize, and CRS). .}

\item{CELLSIZE}{\code{number} - Cell size (use 0 or empty to set it automatically). A numeric value.}

\item{EXTENT}{\code{extent} - Output extent. A comma delimited string of x min, x max, y min, y max. E.g. '4,10,101,105'. Path to a layer. The extent of the layer is used..}

\item{CRS}{\code{crs} - Output CRS. CRS as an auth ID (e.g. 'EPSG:3111'). CRS as a PROJ4 string (e.g. 'PROJ4:…'). CRS as a WKT string (e.g. 'WKT:…'). Path to a layer. The CRS of the layer is used..}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Raster calculator (qgis:rastercalculator)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
