% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rasterlayerproperties.R
\name{qgis_rasterlayerproperties}
\alias{qgis_rasterlayerproperties}
\title{QGIS algorithm Raster layer properties}
\usage{
qgis_rasterlayerproperties(
  INPUT = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Raster layer properties (native:rasterlayerproperties)
}
\details{
\subsection{Outputs description}{
\itemize{
\item BAND_COUNT - outputNumber - Number of bands in raster
\item CRS_AUTHID - outputString - CRS authority identifier
\item EXTENT - outputString - Extent
\item HAS_NODATA_VALUE - outputBoolean - Band has a nodata value set
\item HEIGHT_IN_PIXELS - outputNumber - Height in pixels
\item NODATA_VALUE - outputNumber - Band nodata value
\item PIXEL_HEIGHT - outputNumber - Pixel size (height) in map units
\item PIXEL_WIDTH - outputNumber - Pixel size (width) in map units
\item WIDTH_IN_PIXELS - outputNumber - Width in pixels
\item X_MAX - outputNumber - Maximum x-coordinate
\item X_MIN - outputNumber - Minimum x-coordinate
\item Y_MAX - outputNumber - Maximum y-coordinate
\item Y_MIN - outputNumber - Minimum y-coordinate
}
}
}
