% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rastersampling.R
\name{qgis_rastersampling}
\alias{qgis_rastersampling}
\title{QGIS algorithm - Sample raster values}
\usage{
qgis_rastersampling(
  INPUT = qgisprocess:::qgis_default_value(),
  RASTERCOPY = qgisprocess:::qgis_default_value(),
  COLUMN_PREFIX = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{RASTERCOPY}{\code{raster} - Raster layer. Path to a raster layer.}

\item{COLUMN_PREFIX}{\code{string} - Output column prefix. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Sampled. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Sample raster values (native:rastersampling). Samples raster values under a set of points. This algorithm creates a new vector layer with the same attributes of the input layer and the raster values corresponding on the point location.  If the raster layer has more than one band, all the band values are sampled.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Sampled
}
}
}
