% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rastersampling.R
\name{qgis_rastersampling}
\alias{qgis_rastersampling}
\title{QGIS algorithm Sample raster values}
\usage{
qgis_rastersampling(
  INPUT = qgisprocess::qgis_default_value(),
  RASTERCOPY = qgisprocess::qgis_default_value(),
  COLUMN_PREFIX = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input Point Layer. Path to a vector layer.}

\item{RASTERCOPY}{\code{raster} - Raster Layer to sample. Path to a raster layer.}

\item{COLUMN_PREFIX}{\code{string} - Output column prefix. String value.}

\item{OUTPUT}{\code{sink} - Sampled Points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS Sample raster values (qgis:rastersampling)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Sampled Points
}
}
}
