% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_rectanglesovalsdiamonds.R
\name{qgis_rectanglesovalsdiamonds}
\alias{qgis_rectanglesovalsdiamonds}
\title{QGIS algorithm - Rectangles, ovals, diamonds}
\usage{
qgis_rectanglesovalsdiamonds(
  INPUT = qgisprocess:::qgis_default_value(),
  SHAPE = qgisprocess:::qgis_default_value(),
  WIDTH = qgisprocess:::qgis_default_value(),
  HEIGHT = qgisprocess:::qgis_default_value(),
  ROTATION = qgisprocess:::qgis_default_value(),
  SEGMENTS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{SHAPE}{\code{enum}  of \verb{("Rectangle", "Diamond", "Oval")} - Shape. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{WIDTH}{\code{distance} - Width. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{HEIGHT}{\code{distance} - Height. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ROTATION}{\code{number} - Rotation. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SEGMENTS}{\code{number} - Segments. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Polygon. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Rectangles, ovals, diamonds (native:rectanglesovalsdiamonds). Creates rectangle, oval or diamond-shaped polygons from the input point layer using specified width, height and (optional) rotation values. Multipart inputs should be promoted to singleparts first.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Polygon
}
}
}
