% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_removeduplicatesbyattribute.R
\name{qgis_removeduplicatesbyattribute}
\alias{qgis_removeduplicatesbyattribute}
\title{QGIS algorithm - Delete duplicates by attribute}
\usage{
qgis_removeduplicatesbyattribute(
  INPUT = qgisprocess:::qgis_default_value(),
  FIELDS = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  DUPLICATES = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELDS}{\code{field} - Field to match duplicates by. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Filtered (no duplicates). Path for new vector layer.}

\item{DUPLICATES}{\code{sink} - Filtered (duplicates). Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Delete duplicates by attribute (native:removeduplicatesbyattribute). Removes duplicate rows by a field value (or multiple field values). This algorithm removes duplicate rows by a field value (or multiple field values). The first matching row will be retained, and duplicates will be discarded.  Optionally, these duplicate records can be saved to a separate output for analysis.
}
\details{
\subsection{Outputs description}{
\itemize{
\item DUPLICATES - outputVector - Filtered (duplicates)
\item DUPLICATE_COUNT - outputNumber - Count of discarded duplicate records
\item OUTPUT - outputVector - Filtered (no duplicates)
\item RETAINED_COUNT - outputNumber - Count of retained records
}
}
}
