% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_removenullgeometries.R
\name{qgis_removenullgeometries}
\alias{qgis_removenullgeometries}
\title{QGIS algorithm Remove null geometries}
\usage{
qgis_removenullgeometries(
  INPUT = qgisprocess::qgis_default_value(),
  REMOVE_EMPTY = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  NULL_OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{REMOVE_EMPTY}{\code{boolean} - Also remove empty geometries. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Non null geometries. Path for new vector layer.}

\item{NULL_OUTPUT}{\code{sink} - Null geometries. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Remove null geometries (native:removenullgeometries)
}
\details{
\subsection{Outputs description}{
\itemize{
\item NULL_OUTPUT - outputVector - Null geometries
\item OUTPUT - outputVector - Non null geometries
}
}
}
