% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_retainfields.R
\name{qgis_retainfields}
\alias{qgis_retainfields}
\title{QGIS algorithm Retain fields}
\usage{
qgis_retainfields(
  INPUT = qgisprocess::qgis_default_value(),
  FIELDS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELDS}{\code{field} - Fields to retain. The name of an existing field. ; delimited list of existing field names.}

\item{OUTPUT}{\code{sink} - Retained fields. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Retain fields (native:retainfields)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Retained fields
}
}
}
