% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_ruggednessindex.R
\name{qgis_ruggednessindex}
\alias{qgis_ruggednessindex}
\title{QGIS algorithm Ruggedness index}
\usage{
qgis_ruggednessindex(
  INPUT = qgisprocess::qgis_default_value(),
  Z_FACTOR = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{raster} - Elevation layer. Path to a raster layer.}

\item{Z_FACTOR}{\code{number} - Z factor. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Ruggedness. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Ruggedness index (native:ruggednessindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Ruggedness
}
}
}
