% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_savefeatures.R
\name{qgis_savefeatures}
\alias{qgis_savefeatures}
\title{QGIS algorithm Save vector features to file}
\usage{
qgis_savefeatures(
  INPUT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  LAYER_NAME = qgisprocess::qgis_default_value(),
  DATASOURCE_OPTIONS = qgisprocess::qgis_default_value(),
  LAYER_OPTIONS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Vector features. Path to a vector layer.}

\item{OUTPUT}{\code{fileDestination} - Saved features. Path for new file.}

\item{LAYER_NAME}{\code{string} - Layer name. String value.}

\item{DATASOURCE_OPTIONS}{\code{string} - GDAL dataset options (separate individual options with semicolons). String value.}

\item{LAYER_OPTIONS}{\code{string} - GDAL layer options (separate individual options with semicolons). String value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Save vector features to file (native:savefeatures)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FILE_PATH - outputString - File name and path
\item LAYER_NAME - outputString - Layer name
\item OUTPUT - outputFile - Saved features
}
}
}
