% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_savelog.R
\name{qgis_savelog}
\alias{qgis_savelog}
\title{QGIS algorithm Save log to file}
\usage{
qgis_savelog(
  OUTPUT = qgisprocess:::qgis_default_value(),
  USE_HTML = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{OUTPUT}{\code{fileDestination} - Log file. Path for new file.}

\item{USE_HTML}{\code{boolean} - Use HTML formatting. 1 for true/yes. 0 for false/no.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Save log to file (native:savelog)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputHtml - Log file
}
}
}
