% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_segmentizebymaxangle.R
\name{qgis_segmentizebymaxangle}
\alias{qgis_segmentizebymaxangle}
\title{QGIS algorithm Segmentize by maximum angle}
\usage{
qgis_segmentizebymaxangle(
  INPUT = qgisprocess::qgis_default_value(),
  ANGLE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{ANGLE}{\code{number} - Maximum angle between vertices (degrees). A numeric value.}

\item{OUTPUT}{\code{sink} - Segmentized. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Segmentize by maximum angle (native:segmentizebymaxangle)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Segmentized
}
}
}
