% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_setzfromraster.R
\name{qgis_setzfromraster}
\alias{qgis_setzfromraster}
\title{QGIS algorithm Drape (set Z value from raster)}
\usage{
qgis_setzfromraster(
  INPUT = qgisprocess::qgis_default_value(),
  RASTER = qgisprocess::qgis_default_value(),
  BAND = qgisprocess::qgis_default_value(),
  NODATA = qgisprocess::qgis_default_value(),
  SCALE = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{RASTER}{\code{raster} - Raster layer. Path to a raster layer.}

\item{BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{NODATA}{\code{number} - Value for nodata or non-intersecting vertices. A numeric value.}

\item{SCALE}{\code{number} - Scale factor. A numeric value.}

\item{OUTPUT}{\code{sink} - Draped. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Drape (set Z value from raster) (native:setzfromraster)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Draped
}
}
}
