% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_shortestline.R
\name{qgis_shortestline}
\alias{qgis_shortestline}
\title{QGIS algorithm Shortest line between features}
\usage{
qgis_shortestline(
  SOURCE = qgisprocess:::qgis_default_value(),
  DESTINATION = qgisprocess:::qgis_default_value(),
  METHOD = qgisprocess:::qgis_default_value(),
  NEIGHBORS = qgisprocess:::qgis_default_value(),
  DISTANCE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{SOURCE}{\code{source} - Source layer. Path to a vector layer.}

\item{DESTINATION}{\code{source} - Destination layer. Path to a vector layer.}

\item{METHOD}{\code{enum}  of \verb{("Distance to Nearest Point on feature", "Distance to Feature Centroid")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{NEIGHBORS}{\code{number} - Maximum number of neighbors. A numeric value.}

\item{DISTANCE}{\code{distance} - Maximum distance. A numeric value.}

\item{OUTPUT}{\code{sink} - Shortest lines. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Shortest line between features (native:shortestline)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Shortest lines
}
}
}
