% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_shortestpathlayertopoint.R
\name{qgis_shortestpathlayertopoint}
\alias{qgis_shortestpathlayertopoint}
\title{QGIS algorithm - Shortest path (layer to point)}
\usage{
qgis_shortestpathlayertopoint(
  INPUT = qgisprocess:::qgis_default_value(),
  STRATEGY = qgisprocess:::qgis_default_value(),
  DIRECTION_FIELD = qgisprocess:::qgis_default_value(),
  VALUE_FORWARD = qgisprocess:::qgis_default_value(),
  VALUE_BACKWARD = qgisprocess:::qgis_default_value(),
  VALUE_BOTH = qgisprocess:::qgis_default_value(),
  DEFAULT_DIRECTION = qgisprocess:::qgis_default_value(),
  SPEED_FIELD = qgisprocess:::qgis_default_value(),
  DEFAULT_SPEED = qgisprocess:::qgis_default_value(),
  TOLERANCE = qgisprocess:::qgis_default_value(),
  START_POINTS = qgisprocess:::qgis_default_value(),
  END_POINT = qgisprocess:::qgis_default_value(),
  POINT_TOLERANCE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  OUTPUT_NON_ROUTABLE = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Vector layer representing network. Path to a vector layer.}

\item{STRATEGY}{\code{enum}  of \verb{("Shortest", "Fastest")} - Path type to calculate. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{DIRECTION_FIELD}{\code{field} - Direction field. The name of an existing field. ; delimited list of existing field names.}

\item{VALUE_FORWARD}{\code{string} - Value for forward direction. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{VALUE_BACKWARD}{\code{string} - Value for backward direction. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{VALUE_BOTH}{\code{string} - Value for both directions. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DEFAULT_DIRECTION}{\code{enum}  of \verb{("Forward direction", "Backward direction", "Both directions")} - Default direction. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SPEED_FIELD}{\code{field} - Speed field. The name of an existing field. ; delimited list of existing field names.}

\item{DEFAULT_SPEED}{\code{number} - Default speed (km\code{/}h). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{TOLERANCE}{\code{distance} - Topology tolerance. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{START_POINTS}{\code{source} - Vector layer with start points. Path to a vector layer.}

\item{END_POINT}{\code{point} - End point. Point coordinate as an 'x,y' string, e.g. '1.5,10.1'.}

\item{POINT_TOLERANCE}{\code{distance} - Maximum point distance from network. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Shortest path. Path for new vector layer.}

\item{OUTPUT_NON_ROUTABLE}{\code{sink} - Non-routable features. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Shortest path (layer to point) (native:shortestpathlayertopoint). Computes optimal (shortest or fastest) routes from multiple start points defined by a vector layer and a given end point. This algorithm computes optimal (shortest or fastest) routes from multiple start points defined by a vector layer and a given end point.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Shortest path
\item OUTPUT_NON_ROUTABLE - outputVector - Non-routable features
}
}
}
