% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_shpencodinginfo.R
\name{qgis_shpencodinginfo}
\alias{qgis_shpencodinginfo}
\title{QGIS algorithm - Extract Shapefile encoding}
\usage{
qgis_shpencodinginfo(
  INPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{file} - Input layer. Path to a file.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Extract Shapefile encoding (native:shpencodinginfo). Extracts the attribute encoding information embedded in a Shapefile. This algorithm extracts the attribute encoding information embedded in a Shapefile.  Both the encoding specified by an optional .cpg file and any encoding details present in the .dbf LDID header block are considered.
}
\details{
\subsection{Outputs description}{
\itemize{
\item CPG_ENCODING - outputString - CPG Encoding
\item ENCODING - outputString - Shapefile Encoding
\item LDID_ENCODING - outputString - LDID Encoding
}
}
}
