% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_singlesidedbuffer.R
\name{qgis_singlesidedbuffer}
\alias{qgis_singlesidedbuffer}
\title{QGIS algorithm Single sided buffer}
\usage{
qgis_singlesidedbuffer(
  INPUT = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  SIDE = qgisprocess::qgis_default_value(),
  SEGMENTS = qgisprocess::qgis_default_value(),
  JOIN_STYLE = qgisprocess::qgis_default_value(),
  MITER_LIMIT = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DISTANCE}{\code{distance} - Distance. A numeric value.}

\item{SIDE}{\code{enum}  of \verb{("Left", "Right")} - Side. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{SEGMENTS}{\code{number} - Segments. A numeric value.}

\item{JOIN_STYLE}{\code{enum}  of \verb{("Round", "Miter", "Bevel")} - Join style. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{MITER_LIMIT}{\code{number} - Miter limit. A numeric value.}

\item{OUTPUT}{\code{sink} - Buffered. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Single sided buffer (native:singlesidedbuffer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Buffered
}
}
}
