% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_smoothgeometry.R
\name{qgis_smoothgeometry}
\alias{qgis_smoothgeometry}
\title{QGIS algorithm Smooth}
\usage{
qgis_smoothgeometry(
  INPUT = qgisprocess:::qgis_default_value(),
  ITERATIONS = qgisprocess:::qgis_default_value(),
  OFFSET = qgisprocess:::qgis_default_value(),
  MAX_ANGLE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{ITERATIONS}{\code{number} - Iterations. A numeric value.}

\item{OFFSET}{\code{number} - Offset. A numeric value.}

\item{MAX_ANGLE}{\code{number} - Maximum node angle to smooth. A numeric value.}

\item{OUTPUT}{\code{sink} - Smoothed. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Smooth (native:smoothgeometry)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Smoothed
}
}
}
