% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_splitfeaturesbycharacter.R
\name{qgis_splitfeaturesbycharacter}
\alias{qgis_splitfeaturesbycharacter}
\title{QGIS algorithm Split features by character}
\usage{
qgis_splitfeaturesbycharacter(
  INPUT = qgisprocess:::qgis_default_value(),
  FIELD = qgisprocess:::qgis_default_value(),
  CHAR = qgisprocess:::qgis_default_value(),
  REGEX = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{FIELD}{\code{field} - Split using values in field. The name of an existing field. ; delimited list of existing field names.}

\item{CHAR}{\code{string} - Split values using character. String value.}

\item{REGEX}{\code{boolean} - Use regular expression separator. 1 for true/yes. 0 for false/no.}

\item{OUTPUT}{\code{sink} - Split. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Split features by character (native:splitfeaturesbycharacter)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Split
}
}
}
