% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_splitlinesbylength.R
\name{qgis_splitlinesbylength}
\alias{qgis_splitlinesbylength}
\title{QGIS algorithm - Split lines by maximum length}
\usage{
qgis_splitlinesbylength(
  INPUT = qgisprocess:::qgis_default_value(),
  LENGTH = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{LENGTH}{\code{distance} - Maximum line length. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Split. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Split lines by maximum length (native:splitlinesbylength). Splits lines into parts which are not longer than a specified length. This algorithm takes a line (or curve) layer and splits each feature into multiple parts, where each part is of a specified maximum length.  Z and M values at the start and end of the new line substrings are linearly interpolated from existing values.  ---------------- Notes ----------------   - This algorithm drops existing primary keys or FID values and regenerates them in output layers.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Split
}
}
}
