% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_transect.R
\name{qgis_transect}
\alias{qgis_transect}
\title{QGIS algorithm - Transect}
\usage{
qgis_transect(
  INPUT = qgisprocess:::qgis_default_value(),
  LENGTH = qgisprocess:::qgis_default_value(),
  ANGLE = qgisprocess:::qgis_default_value(),
  SIDE = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{LENGTH}{\code{distance} - Length of the transect. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{ANGLE}{\code{number} - Angle in degrees from the original line at the vertices. A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{SIDE}{\code{enum}  of \verb{("Left", "Right", "Both")} - Side to create the transects. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{OUTPUT}{\code{sink} - Transect. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Transect (native:transect). Creates transects on vertices for (multi)linestrings. This algorithm creates transects on vertices for (multi)linestrings. A transect is a line oriented from an angle (by default perpendicular) to the input polylines (at vertices).  Field(s) from feature(s) are returned in the transect with these new fields: - TR_FID: ID of the original feature - TR_ID: ID of the transect. Each transect have an unique ID - TR_SEGMENT: ID of the segment of the linestring - TR_ANGLE: Angle in degrees from the original line at the vertex - TR_LENGTH: Total length of the transect returned - TR_ORIENT: Side of the transect (only on the left or right of the line, or both side)   ---------------- Notes ----------------   - This algorithm drops existing primary keys or FID values and regenerates them in output layers.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Transect
}
}
}
