% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_transferannotationsfrommain.R
\name{qgis_transferannotationsfrommain}
\alias{qgis_transferannotationsfrommain}
\title{QGIS algorithm - Transfer annotations from main layer}
\usage{
qgis_transferannotationsfrommain(
  LAYER_NAME = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{LAYER_NAME}{\code{string} - New layer name. String value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Transfer annotations from main layer (native:transferannotationsfrommain). Transfers all annotations from the main annotation layer in a project to a new annotation layer. This algorithm transfers all annotations from the main annotation layer in a project to a new annotation layer.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputLayer - New annotation layer
}
}
}
