% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_translategeometry.R
\name{qgis_translategeometry}
\alias{qgis_translategeometry}
\title{QGIS algorithm - Translate}
\usage{
qgis_translategeometry(
  INPUT = qgisprocess:::qgis_default_value(),
  DELTA_X = qgisprocess:::qgis_default_value(),
  DELTA_Y = qgisprocess:::qgis_default_value(),
  DELTA_Z = qgisprocess:::qgis_default_value(),
  DELTA_M = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{DELTA_X}{\code{distance} - Offset distance (x-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_Y}{\code{distance} - Offset distance (y-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_Z}{\code{number} - Offset distance (z-axis). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{DELTA_M}{\code{number} - Offset distance (m values). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Translated. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Translate (native:translategeometry). This algorithm moves the geometries within a layer, by offsetting them with a specified x and y displacement.  Z and M values present in the geometry can also be translated.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Translated
}
}
}
