% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_voronoipolygons.R
\name{qgis_voronoipolygons}
\alias{qgis_voronoipolygons}
\title{QGIS algorithm - Voronoi polygons}
\usage{
qgis_voronoipolygons(
  INPUT = qgisprocess:::qgis_default_value(),
  BUFFER = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{BUFFER}{\code{number} - Buffer region (\% of extent). A numeric value. field:FIELD_NAME to use a data defined value taken from the FIELD_NAME field. expression:SOME EXPRESSION to use a data defined value calculated using a custom QGIS expression.}

\item{OUTPUT}{\code{sink} - Voronoi polygons. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by QGIS Voronoi polygons (qgis:voronoipolygons). This algorithm takes a points layer and generates a polygon layer containing the voronoi polygons corresponding to those input points.
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Voronoi polygons
}
}
}
