% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_wedgebuffers.R
\name{qgis_wedgebuffers}
\alias{qgis_wedgebuffers}
\title{QGIS algorithm Create wedge buffers}
\usage{
qgis_wedgebuffers(
  INPUT = qgisprocess::qgis_default_value(),
  AZIMUTH = qgisprocess::qgis_default_value(),
  WIDTH = qgisprocess::qgis_default_value(),
  OUTER_RADIUS = qgisprocess::qgis_default_value(),
  INNER_RADIUS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT}{\code{source} - Input layer. Path to a vector layer.}

\item{AZIMUTH}{\code{number} - Azimuth (degrees from North). A numeric value.}

\item{WIDTH}{\code{number} - Wedge width (in degrees). A numeric value.}

\item{OUTER_RADIUS}{\code{number} - Outer radius. A numeric value.}

\item{INNER_RADIUS}{\code{number} - Inner radius. A numeric value.}

\item{OUTPUT}{\code{sink} - Buffers. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Create wedge buffers (native:wedgebuffers)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Buffers
}
}
}
