% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis_zonalhistogram.R
\name{qgis_zonalhistogram}
\alias{qgis_zonalhistogram}
\title{QGIS algorithm Zonal histogram}
\usage{
qgis_zonalhistogram(
  INPUT_RASTER = qgisprocess::qgis_default_value(),
  RASTER_BAND = qgisprocess::qgis_default_value(),
  INPUT_VECTOR = qgisprocess::qgis_default_value(),
  COLUMN_PREFIX = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{INPUT_RASTER}{\code{raster} - Raster layer. Path to a raster layer.}

\item{RASTER_BAND}{\code{band} - Band number. Integer value representing an existing raster band number.}

\item{INPUT_VECTOR}{\code{source} - Vector layer containing zones. Path to a vector layer.}

\item{COLUMN_PREFIX}{\code{string} - Output column prefix. String value.}

\item{OUTPUT}{\code{sink} - Output zones. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by QGIS (native c++) Zonal histogram (native:zonalhistogram)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputVector - Output zones
}
}
}
