% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_aggregationindex.R
\name{saga_aggregationindex}
\alias{saga_aggregationindex}
\title{QGIS algorithm Aggregation index}
\usage{
saga_aggregationindex(
  INPUT = qgisprocess::qgis_default_value(),
  MAXNUMCLASS = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input Grid. Path to a raster layer.}

\item{MAXNUMCLASS}{\code{number} - Max. Number of Classes. A numeric value.}

\item{RESULT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Aggregation index (saga:aggregationindex)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Result
}
}
}
