% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_binaryerosionreconstruction.R
\name{saga_binaryerosionreconstruction}
\alias{saga_binaryerosionreconstruction}
\title{QGIS algorithm Binary erosion-reconstruction}
\usage{
saga_binaryerosionreconstruction(
  INPUT_GRID = qgisprocess:::qgis_default_value(),
  OUTPUT_GRID = qgisprocess:::qgis_default_value(),
  RADIUS = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT_GRID}{\code{raster} - Input Grid. Path to a raster layer.}

\item{OUTPUT_GRID}{\code{rasterDestination} - Output Grid. Path for new raster layer.}

\item{RADIUS}{\code{number} - Filter Size (Radius). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Binary erosion-reconstruction (saga:binaryerosionreconstruction)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT_GRID - outputRaster - Output Grid
}
}
}
