% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_convertmultipointstopoints.R
\name{saga_convertmultipointstopoints}
\alias{saga_convertmultipointstopoints}
\title{QGIS algorithm Convert multipoints to points}
\usage{
saga_convertmultipointstopoints(
  MULTIPOINTS = qgisprocess::qgis_default_value(),
  POINTS = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{MULTIPOINTS}{\code{source} - Multipoints. Path to a vector layer.}

\item{POINTS}{\code{vectorDestination} - Points. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Convert multipoints to points (saga:convertmultipointstopoints)
}
\details{
\subsection{Outputs description}{
\itemize{
\item POINTS - outputVector - Points
}
}
}
