% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_createpointgrid.R
\name{saga_createpointgrid}
\alias{saga_createpointgrid}
\title{QGIS algorithm Create point grid}
\usage{
saga_createpointgrid(
  POINTS = qgisprocess::qgis_default_value(),
  X_EXTENT_MIN = qgisprocess::qgis_default_value(),
  X_EXTENT_MAX = qgisprocess::qgis_default_value(),
  Y_EXTENT_MIN = qgisprocess::qgis_default_value(),
  Y_EXTENT_MAX = qgisprocess::qgis_default_value(),
  DIST = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POINTS}{\code{vectorDestination} - Point Grid. Path for new vector layer.}

\item{X_EXTENT_MIN}{\code{number} - X-Extent (Min). A numeric value.}

\item{X_EXTENT_MAX}{\code{number} - X-Extent (Max). A numeric value.}

\item{Y_EXTENT_MIN}{\code{number} - Y-Extent (Min). A numeric value.}

\item{Y_EXTENT_MAX}{\code{number} - Y-Extent (Max). A numeric value.}

\item{DIST}{\code{number} - Distance. A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Create point grid (saga:createpointgrid)
}
\details{
\subsection{Outputs description}{
\itemize{
\item POINTS - outputVector - Point Grid
}
}
}
