% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_fastregiongrowingalgorithm.R
\name{saga_fastregiongrowingalgorithm}
\alias{saga_fastregiongrowingalgorithm}
\title{QGIS algorithm Fast region growing algorithm}
\usage{
saga_fastregiongrowingalgorithm(
  INPUT = qgisprocess::qgis_default_value(),
  START = qgisprocess::qgis_default_value(),
  REP = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  MEAN = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{multilayer} - Input Grids. .}

\item{START}{\code{raster} - Seeds Grid. Path to a raster layer.}

\item{REP}{\code{raster} - Smooth Rep. Path to a raster layer.}

\item{RESULT}{\code{rasterDestination} - Segmente. Path for new raster layer.}

\item{MEAN}{\code{rasterDestination} - Mean. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Fast region growing algorithm (saga:fastregiongrowingalgorithm)
}
\details{
\subsection{Outputs description}{
\itemize{
\item MEAN - outputRaster - Mean
\item RESULT - outputRaster - Segmente
}
}
}
