% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_flowpathlength.R
\name{saga_flowpathlength}
\alias{saga_flowpathlength}
\title{QGIS algorithm Flow path length}
\usage{
saga_flowpathlength(
  ELEVATION = qgisprocess::qgis_default_value(),
  SEED = qgisprocess::qgis_default_value(),
  SEEDS_ONLY = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  CONVERGENCE = qgisprocess::qgis_default_value(),
  LENGTH = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{SEED}{\code{raster} - Seeds. Path to a raster layer.}

\item{SEEDS_ONLY}{\code{boolean} - Seeds Only. 1 for true/yes. 0 for false/no.}

\item{METHOD}{\code{enum}  of \verb{("[0] Deterministic 8 (D8)", "[1] Multiple Flow Direction (FD8)")} - Flow Routing Algorithm. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{CONVERGENCE}{\code{number} - Convergence (FD8). A numeric value.}

\item{LENGTH}{\code{rasterDestination} - Flow Path Length. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Flow path length (saga:flowpathlength)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LENGTH - outputRaster - Flow Path Length
}
}
}
