% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_geographiccoordinategrids.R
\name{saga_geographiccoordinategrids}
\alias{saga_geographiccoordinategrids}
\title{QGIS algorithm Geographic coordinate grids}
\usage{
saga_geographiccoordinategrids(
  GRID = qgisprocess::qgis_default_value(),
  LON = qgisprocess::qgis_default_value(),
  LAT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{LON}{\code{rasterDestination} - Longitude. Path for new raster layer.}

\item{LAT}{\code{rasterDestination} - Latitude. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Geographic coordinate grids (saga:geographiccoordinategrids)
}
\details{
\subsection{Outputs description}{
\itemize{
\item LAT - outputRaster - Latitude
\item LON - outputRaster - Longitude
}
}
}
