% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_globalmoransiforrasterlayer.R
\name{saga_globalmoransiforrasterlayer}
\alias{saga_globalmoransiforrasterlayer}
\title{QGIS algorithm Global moran's i for raster layer}
\usage{
saga_globalmoransiforrasterlayer(
  GRID = qgisprocess::qgis_default_value(),
  CONTIGUITY = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRID}{\code{raster} - Grid. Path to a raster layer.}

\item{CONTIGUITY}{\code{enum}  of \verb{("[0] Rook", "[1] Queen")} - Case of contiguity. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{RESULT}{\code{vectorDestination} - Result. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Global moran's i for raster layer (saga:globalmoransiforrasterlayer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputVector - Result
}
}
}
