% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_lakeflood.R
\name{saga_lakeflood}
\alias{saga_lakeflood}
\title{QGIS algorithm Lake flood}
\usage{
saga_lakeflood(
  ELEV = qgisprocess::qgis_default_value(),
  SEEDS = qgisprocess::qgis_default_value(),
  LEVEL = qgisprocess::qgis_default_value(),
  OUTDEPTH = qgisprocess::qgis_default_value(),
  OUTLEVEL = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{ELEV}{\code{raster} - DEM. Path to a raster layer.}

\item{SEEDS}{\code{raster} - Seeds. Path to a raster layer.}

\item{LEVEL}{\code{boolean} - Absolute Water Levels. 1 for true/yes. 0 for false/no.}

\item{OUTDEPTH}{\code{rasterDestination} - Lake. Path for new raster layer.}

\item{OUTLEVEL}{\code{rasterDestination} - Surface. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Lake flood (saga:lakeflood)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTDEPTH - outputRaster - Lake
\item OUTLEVEL - outputRaster - Surface
}
}
}
