% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_orderedweightedaveraging.R
\name{saga_orderedweightedaveraging}
\alias{saga_orderedweightedaveraging}
\title{QGIS algorithm Ordered weighted averaging}
\usage{
saga_orderedweightedaveraging(
  GRIDS = qgisprocess::qgis_default_value(),
  WEIGHTS = qgisprocess::qgis_default_value(),
  OUTPUT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Input Grids. .}

\item{WEIGHTS}{\code{matrix} - Weights. A comma delimited list of values.}

\item{OUTPUT}{\code{rasterDestination} - Output Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Ordered weighted averaging (saga:orderedweightedaveraging)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output Grid
}
}
}
