% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_paramemeltonruggednessnumber.R
\name{saga_paramemeltonruggednessnumber}
\alias{saga_paramemeltonruggednessnumber}
\title{QGIS algorithm Paramemelton ruggedness number}
\usage{
saga_paramemeltonruggednessnumber(
  DEM = qgisprocess:::qgis_default_value(),
  AREA = qgisprocess:::qgis_default_value(),
  ZMAX = qgisprocess:::qgis_default_value(),
  MRN = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{AREA}{\code{rasterDestination} - Catchment Area. Path for new raster layer.}

\item{ZMAX}{\code{rasterDestination} - Maximum Height. Path for new raster layer.}

\item{MRN}{\code{rasterDestination} - Melton Ruggedness Number. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Paramemelton ruggedness number (saga:paramemeltonruggednessnumber)
}
\details{
\subsection{Outputs description}{
\itemize{
\item AREA - outputRaster - Catchment Area
\item MRN - outputRaster - Melton Ruggedness Number
\item ZMAX - outputRaster - Maximum Height
}
}
}
