% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rasterdivision.R
\name{saga_rasterdivision}
\alias{saga_rasterdivision}
\title{QGIS algorithm Raster division}
\usage{
saga_rasterdivision(
  A = qgisprocess::qgis_default_value(),
  B = qgisprocess::qgis_default_value(),
  C = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{A}{\code{raster} - Dividend. Path to a raster layer.}

\item{B}{\code{raster} - Divisor. Path to a raster layer.}

\item{C}{\code{rasterDestination} - Quotient. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster division (saga:rasterdivision)
}
\details{
\subsection{Outputs description}{
\itemize{
\item C - outputRaster - Quotient
}
}
}
