% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rasterproduct.R
\name{saga_rasterproduct}
\alias{saga_rasterproduct}
\title{QGIS algorithm Raster product}
\usage{
saga_rasterproduct(
  GRIDS = qgisprocess::qgis_default_value(),
  RESULT = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{GRIDS}{\code{multilayer} - Grids. .}

\item{RESULT}{\code{rasterDestination} - Product. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster product (saga:rasterproduct)
}
\details{
\subsection{Outputs description}{
\itemize{
\item RESULT - outputRaster - Product
}
}
}
