% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_rasterproximitybuffer.R
\name{saga_rasterproximitybuffer}
\alias{saga_rasterproximitybuffer}
\title{QGIS algorithm Raster proximity buffer}
\usage{
saga_rasterproximitybuffer(
  SOURCE = qgisprocess::qgis_default_value(),
  DIST = qgisprocess::qgis_default_value(),
  IVAL = qgisprocess::qgis_default_value(),
  DISTANCE = qgisprocess::qgis_default_value(),
  ALLOC = qgisprocess::qgis_default_value(),
  BUFFER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SOURCE}{\code{raster} - Source Grid. Path to a raster layer.}

\item{DIST}{\code{number} - Buffer distance. A numeric value.}

\item{IVAL}{\code{number} - Equidistance. A numeric value.}

\item{DISTANCE}{\code{rasterDestination} - Distance Grid. Path for new raster layer.}

\item{ALLOC}{\code{rasterDestination} - Allocation Grid. Path for new raster layer.}

\item{BUFFER}{\code{rasterDestination} - Buffer Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Raster proximity buffer (saga:rasterproximitybuffer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item ALLOC - outputRaster - Allocation Grid
\item BUFFER - outputRaster - Buffer Grid
\item DISTANCE - outputRaster - Distance Grid
}
}
}
