% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_removesmallpixelclumpstonodata.R
\name{saga_removesmallpixelclumpstonodata}
\alias{saga_removesmallpixelclumpstonodata}
\title{QGIS algorithm Remove small pixel clumps (to no-data)}
\usage{
saga_removesmallpixelclumpstonodata(
  GRID = qgisprocess:::qgis_default_value(),
  THRESHOLD = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{GRID}{\code{raster} - Input Grid. Path to a raster layer.}

\item{THRESHOLD}{\code{number} - Min. Size. A numeric value.}

\item{OUTPUT}{\code{rasterDestination} - Filtered Grid. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Remove small pixel clumps (to no-data) (saga:removesmallpixelclumpstonodata)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Filtered Grid
}
}
}
