% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_singlevaluedecompositionopencv.R
\name{saga_singlevaluedecompositionopencv}
\alias{saga_singlevaluedecompositionopencv}
\title{QGIS algorithm Single value decomposition (opencv)}
\usage{
saga_singlevaluedecompositionopencv(
  INPUT = qgisprocess:::qgis_default_value(),
  OUTPUT = qgisprocess:::qgis_default_value(),
  RANGE_MIN = qgisprocess:::qgis_default_value(),
  RANGE_MAX = qgisprocess:::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{INPUT}{\code{raster} - Input. Path to a raster layer.}

\item{OUTPUT}{\code{rasterDestination} - Output. Path for new raster layer.}

\item{RANGE_MIN}{\code{number} - Range (Min). A numeric value.}

\item{RANGE_MAX}{\code{number} - Range (Max). A numeric value.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Single value decomposition (opencv) (saga:singlevaluedecompositionopencv)
}
\details{
\subsection{Outputs description}{
\itemize{
\item OUTPUT - outputRaster - Output
}
}
}
