% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_sinkdrainageroutedetection.R
\name{saga_sinkdrainageroutedetection}
\alias{saga_sinkdrainageroutedetection}
\title{QGIS algorithm Sink drainage route detection}
\usage{
saga_sinkdrainageroutedetection(
  ELEVATION = qgisprocess::qgis_default_value(),
  THRSHEIGHT = qgisprocess::qgis_default_value(),
  SINKROUTE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{ELEVATION}{\code{raster} - Elevation. Path to a raster layer.}

\item{THRSHEIGHT}{\code{number} - Threshold Height. A numeric value.}

\item{SINKROUTE}{\code{rasterDestination} - Sink Route. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Sink drainage route detection (saga:sinkdrainageroutedetection)
}
\details{
\subsection{Outputs description}{
\itemize{
\item SINKROUTE - outputRaster - Sink Route
}
}
}
