% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_sinkremoval.R
\name{saga_sinkremoval}
\alias{saga_sinkremoval}
\title{QGIS algorithm Sink removal}
\usage{
saga_sinkremoval(
  DEM = qgisprocess::qgis_default_value(),
  SINKROUTE = qgisprocess::qgis_default_value(),
  METHOD = qgisprocess::qgis_default_value(),
  THRSHEIGHT = qgisprocess::qgis_default_value(),
  DEM_PREPROC = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{SINKROUTE}{\code{raster} - Sink Route. Path to a raster layer.}

\item{METHOD}{\code{enum}  of \verb{("[0] Deepen Drainage Routes", "[1] Fill Sinks")} - Method. Number of selected option, e.g. '1'. Comma separated list of options, e.g. '1,3'.}

\item{THRSHEIGHT}{\code{number} - Threshold Height. A numeric value.}

\item{DEM_PREPROC}{\code{rasterDestination} - Preprocessed DEM. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Sink removal (saga:sinkremoval)
}
\details{
\subsection{Outputs description}{
\itemize{
\item DEM_PREPROC - outputRaster - Preprocessed DEM
}
}
}
