% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_splitshapeslayerrandomly.R
\name{saga_splitshapeslayerrandomly}
\alias{saga_splitshapeslayerrandomly}
\title{QGIS algorithm Split shapes layer randomly}
\usage{
saga_splitshapeslayerrandomly(
  SHAPES = qgisprocess::qgis_default_value(),
  PERCENT = qgisprocess::qgis_default_value(),
  A = qgisprocess::qgis_default_value(),
  B = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{PERCENT}{\code{number} - Split ratio (\%). A numeric value.}

\item{A}{\code{vectorDestination} - Group A. Path for new vector layer.}

\item{B}{\code{vectorDestination} - Group B. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Split shapes layer randomly (saga:splitshapeslayerrandomly)
}
\details{
\subsection{Outputs description}{
\itemize{
\item A - outputVector - Group A
\item B - outputVector - Group B
}
}
}
