% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_strahlerorder.R
\name{saga_strahlerorder}
\alias{saga_strahlerorder}
\title{QGIS algorithm Strahler order}
\usage{
saga_strahlerorder(
  DEM = qgisprocess::qgis_default_value(),
  STRAHLER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - Elevation. Path to a raster layer.}

\item{STRAHLER}{\code{rasterDestination} - Strahler Order. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Strahler order (saga:strahlerorder)
}
\details{
\subsection{Outputs description}{
\itemize{
\item STRAHLER - outputRaster - Strahler Order
}
}
}
