% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_successiveflowrouting.R
\name{saga_successiveflowrouting}
\alias{saga_successiveflowrouting}
\title{QGIS algorithm Successive flow routing}
\usage{
saga_successiveflowrouting(
  DEM = qgisprocess::qgis_default_value(),
  ITERATIONS = qgisprocess::qgis_default_value(),
  RUNOFF = qgisprocess::qgis_default_value(),
  MANNING = qgisprocess::qgis_default_value(),
  FLOW = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = .complete_output_option(),
  .quiet = .quiet_option(),
  .messages = .message_option()
)
}
\arguments{
\item{DEM}{\code{raster} - DEM. Path to a raster layer.}

\item{ITERATIONS}{\code{number} - Iterations. A numeric value.}

\item{RUNOFF}{\code{number} - Runoff. A numeric value.}

\item{MANNING}{\code{number} - Manning's Roughness. A numeric value.}

\item{FLOW}{\code{rasterDestination} - Flow. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifying if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}

\item{.quiet}{logical specifying if parameter \code{.quiet} for \code{qgisprocess::qgis_run_algorithm()} Default value is \code{TRUE}.}

\item{.messages}{logical specifying if messages from \code{qgisprocess::qgis_run_algorithm()} should be printed (\code{TRUE}) or not (\code{FALSE}). Default value is \code{FALSE}.}
}
\description{
QGIS Algorithm provided by SAGA Successive flow routing (saga:successiveflowrouting)
}
\details{
\subsection{Outputs description}{
\itemize{
\item FLOW - outputRaster - Flow
}
}
}
