% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_tcilow.R
\name{saga_tcilow}
\alias{saga_tcilow}
\title{QGIS algorithm Tci low}
\usage{
saga_tcilow(
  DISTANCE = qgisprocess::qgis_default_value(),
  TWI = qgisprocess::qgis_default_value(),
  TCILOW = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{DISTANCE}{\code{raster} - Vertical Distance to Channel Network. Path to a raster layer.}

\item{TWI}{\code{raster} - Topographic Wetness Index. Path to a raster layer.}

\item{TCILOW}{\code{rasterDestination} - TCI Low. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Tci low (saga:tcilow)
}
\details{
\subsection{Outputs description}{
\itemize{
\item TCILOW - outputRaster - TCI Low
}
}
}
