% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_variabledistancebuffer.R
\name{saga_variabledistancebuffer}
\alias{saga_variabledistancebuffer}
\title{QGIS algorithm Variable distance buffer}
\usage{
saga_variabledistancebuffer(
  SHAPES = qgisprocess::qgis_default_value(),
  DIST_FIELD = qgisprocess::qgis_default_value(),
  DIST_SCALE = qgisprocess::qgis_default_value(),
  NZONES = qgisprocess::qgis_default_value(),
  DARC = qgisprocess::qgis_default_value(),
  DISSOLVE = qgisprocess::qgis_default_value(),
  POLY_INNER = qgisprocess::qgis_default_value(),
  BUFFER = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{SHAPES}{\code{source} - Shapes. Path to a vector layer.}

\item{DIST_FIELD}{\code{field} - Buffer Distance. The name of an existing field. ; delimited list of existing field names.}

\item{DIST_SCALE}{\code{number} - Scaling Factor for Attribute Value. A numeric value.}

\item{NZONES}{\code{number} - Number of Buffer Zones. A numeric value.}

\item{DARC}{\code{number} - Arc Vertex Distance (Degree). A numeric value.}

\item{DISSOLVE}{\code{boolean} - Dissolve Buffers. 1 for true/yes. 0 for false/no.}

\item{POLY_INNER}{\code{boolean} - Inner Buffer. 1 for true/yes. 0 for false/no.}

\item{BUFFER}{\code{vectorDestination} - Buffer. Path for new vector layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Variable distance buffer (saga:variabledistancebuffer)
}
\details{
\subsection{Outputs description}{
\itemize{
\item BUFFER - outputVector - Buffer
}
}
}
