% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_variogramsurface.R
\name{saga_variogramsurface}
\alias{saga_variogramsurface}
\title{QGIS algorithm Variogram surface}
\usage{
saga_variogramsurface(
  POINTS = qgisprocess::qgis_default_value(),
  FIELD = qgisprocess::qgis_default_value(),
  DISTCOUNT = qgisprocess::qgis_default_value(),
  NSKIP = qgisprocess::qgis_default_value(),
  COUNT = qgisprocess::qgis_default_value(),
  VARIANCE = qgisprocess::qgis_default_value(),
  COVARIANCE = qgisprocess::qgis_default_value(),
  ...,
  .complete_output = TRUE
)
}
\arguments{
\item{POINTS}{\code{source} - Points. Path to a vector layer.}

\item{FIELD}{\code{field} - Attribute. The name of an existing field. ; delimited list of existing field names.}

\item{DISTCOUNT}{\code{number} - Number of Distance Classes. A numeric value.}

\item{NSKIP}{\code{number} - Skip Number. A numeric value.}

\item{COUNT}{\code{rasterDestination} - Number of Pairs. Path for new raster layer.}

\item{VARIANCE}{\code{rasterDestination} - Variogram Surface. Path for new raster layer.}

\item{COVARIANCE}{\code{rasterDestination} - Covariance Surface. Path for new raster layer.}

\item{...}{further parameters passed to \code{qgisprocess::qgis_run_algorithm()}}

\item{.complete_output}{logical specifing if complete out of \code{qgisprocess::qgis_run_algorithm()} should be used (\code{TRUE}) or first output (most likely the main) should read (\code{FALSE}). Default value is \code{TRUE}.}
}
\description{
QGIS Algorithm provided by SAGA Variogram surface (saga:variogramsurface)
}
\details{
\subsection{Outputs description}{
\itemize{
\item COUNT - outputRaster - Number of Pairs
\item COVARIANCE - outputRaster - Covariance Surface
\item VARIANCE - outputRaster - Variogram Surface
}
}
}
